[DERIVED] Analysis Dataset of Adverse Events

Analysis dataset of adverse events derived dataset.

Usage

data(ADAE)

Format

A data frame with 3495 observations and 62 variables.

Source

For more details see the help vignette: vignette(topic = 'ADNIMERGE2-Analysis-Data', package = 'ADNIMERGE2')

Details

STUDYID character Study Identifier , with character options: ADNI

USUBJID character Unique Subject Identifier

SUBJID character Subject Identifier for the Study

SITEID character Study Site Identifier

TRTA character Actual Arm

TRTP character Planned Arm

AGE numeric Age (in Years) Range value: 50.5, ..., 90.3

AGEU character Age Units , with character options: Years

AGEGR1 character Pooled Age Group 1 , with character options: >=65, <65

AGEGR1N numeric Pooled Age Group 1 (N) Range value: 1, ..., 2

SEX character Sex , with character options: Female, Male

RACE factor Race , with factor levels: American Indian or Alaskan Native, Asian, Black or African American, Native Hawaiian or Other Pacific Islander, Other Pacific Islander, White, More than one race, Unknown

RACEN numeric Race (N) Range value: 1, ..., 999

ETHNIC character Ethnicity , with character options: Not Hispanic or Latino, Hispanic or Latino, Unknown

DX factor Baseline Diagnostics Status , with factor levels: CN, MCI, DEM

ENRLFL character Enrolled Population Flag , with character options: Yes, No

ENRLDT Date Date of Enrollment Date: YYY-MM-DD

EOSDT Date End of Study Date Date: YYY-MM-DD

EOSDTF character End of Study Date Imputation Flag , with character options: M

ASTDT Date Analysis Start Date Date: YYY-MM-DD

ASTDTF character Analysis Start Date Imputation Flag

ASTDY numeric Analysis Start Relative Day Range value: -4327, ..., 6929

AENDT Date Analysis End Date Date: YYY-MM-DD

AENDTF character Analysis End Date Imputation Flag

AENDY numeric Analysis End Relative Day Range value: -133, ..., 337328

ADURN numeric Adverse Events Duration (N) Range value: -363, ..., 332062

ADURU character Analysis Duration Unit , with character options: days

AETERM character Reported Term for the Adverse Event

AELLT character Lowest Level Term

AELLTCD character Lowest Level Term Code

AEDECOD character Dictionary-Derived Term

AEPTCD character Preferred Term Code

AEHLT character High Level Term

AEHLTCD character High Level Term Code

AEHLGT character High Level Group Term

AEHLGTCD character High Level Group Term Code

AESOC character Primary Sytem Organ Class

AESOCCD character Primary Sytem Organ Class Code

AESEV character Severity/Intensity , with character options: Moderate, Severe, Mild

AESER character Serious Event , with character options: No, Yes

AESCONG character Congenital Anomaly or Birth Defect , with character options: No

AESDISAB character Persist or Signif Disability/Incapacity , with character options: No, Yes

AESDTH character Results in Death , with character options: Yes, No

AESHOSP character Requires or Prologns Hospitalization , with character options: No, Yes

AESLIFE character Is Life Threating , with character options: Yes, No

AESMIE character Other Medically Important Serious Event , with character options: No, Yes

AERELCM character Related to Concomitant Therapy , with character options: Not related, Unlikely, Possibly, Definitely, Probably

AERELFLRBTBN character Related to Florbetaben Tracer , with character options: Not related, Unlikely, Definitely, Possibly

AERELFLRBPR character Related to Florbetapir Tracer , with character options: Not related, Unlikely, Possibly, Probably

AEHIMG character Related to PET/MRI Imaging Procedure , with character options: Not related, Possibly, Unlikely, Definitely, Probably

AERELTAU character Related to Flotaucipir Tracer , with character options: Not related, Possibly, Probably, Unlikely

AERELNAV character Related to NAV-4694 Tracer , with character options: Not related, Definitely, Possibly

AERELMK character Related to MK-6240 Tracer , with character options: Not related, Unlikely

AERELPI character Related to PI-2620 Tracer , with character options: Not related

AEHLUMB character Related to Lumbar Puncture , with character options: Not related, Definitely, Probably, Unlikely, Possibly

AERELCOVID character Related to COVID-19 Illness , with character options: Unlikely, Not related, Definitely (Confirmed COVID-19 test), Probably, Possibly

AERELPAN character Related to COVID-19 Pandemic Disruption , with character options: Not related, Definitely, Probably, Unlikely, Possibly

AERELAD character Related to Early Stage/Progression of AD , with character options: Not related, Unlikely, Possibly, Probably, Definitely

AERELATESP character Related to Other Study Procedure(s) , with character options: Not related, Unlikely, Probably, Definitely, Possibly

AECONTRT character Concomitant or Additional Trtmnt Given , with character options: Yes, No

AEOUT character Outcome of Adverse Event , with character options: Recovered/Resolved, Fatal, Not recovered/Not resolved, Recovering/Resolving, Recovered/Resolved with sequelae, Unknown

AESEQ integer Sequence Number Range value: 1, ..., 32

Author

adni-data@googlegroups.com