ADAS-Cognitive Behavior

ADAS-Cognitive Behavior data. More information is available at https://adni.loni.usc.edu/data-samples/adni-data/

Usage

data(ADAS_ADNI1)

Format

A data frame with 3873 observations and 87 variables.

Source

https://adni.loni.usc.edu/data-samples/adni-data/

Details

ORIGPROT character Original study protocol , with options: ADNI1

COLPROT factor Study protocol of data collection , with factor levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4

PTID character Participant ID

RID numeric Participant roster ID Range value: 2, ..., 1435

VISCODE character Visit code , with options: bl, m06, m24, m12, m48, m18, uns1, m36

VISCODE2 character , with options: bl, m06, m24, m12, m48, m18, uns1, m36

EXAMDATE Examination Date Date: YYY-MM-DD

COT1LIST character Trial 1 1=BUTTER; 2=ARM; 3=SHORE; 4=LETTER; 5=QUEEN; 6=CABIN; 7=POLE; 8=TICKET; 9=GRASS; 10=ENGINE Decoded Value: Yes

COT2LIST character Trial 2 1=POLE; 2=LETTER; 3=BUTTER; 4=QUEEN; 5=ARM; 6=SHORE; 7=GRASS; 8=CABIN; 9=TICKET; 10=ENGINE Decoded Value: Yes

COT3LIST character Trial 3 1=SHORE; 2=LETTER; 3=ARM; 4=CABIN; 5=POLE; 6=TICKET; 7=ENGINE; 8=GRASS; 9=BUTTER; 10=QUEEN Decoded Value: Yes

COT1SCOR numeric Total Trial 1 0..10

COT2SCOR numeric Total Trial 2 0..10

COT3SCOR numeric Total Trial 1 0..10

COP1COMP character If any trial not administered, check one: 1=Not done (for reasons other than physical/cognitive); 2=Participant refused; 4=Participant unable to complete 3 trials for cognitive reasons (NOTE: A score of '0' should be given in each of Trials 1, 2, and 3 if the task was administered and the participant was unable to complete the task for cognitive reasons.) Decoded Value: Yes

COCOMND character Check each command performed correctly. 1= Make a fist.; 2=Point to the ceiling, then to the floor.; 3=Put the pencil on top of the card, then put it back.; 4=Put the watch on the other side of the pencil and turn over the card.; 5=Tap each shoulder twice with two fingers, keeping your eyes shut.; 6=None completed correctly. Decoded Value: Yes

COP2COMP character If task not administered, check one: 1=Not done (for reasons other than physical/cognitive); 2=Participant refused; 3=Participant unable for physical reasons;4=Participant unable for cognitive reasons (NOTE: This should only be checked if the task was not administered due to cognitive reasons.) Decoded Value: Yes

COCONSTR character Check each figure drawn correctly. 1= None, attempted but drew no forms correctly.; 2=Circle; 3=Two overlapping rectangles; 4=Rhombus; 5=Cube Decoded Value: Yes

COP3COMP character If task not administered, check one: 1=Not done (for reasons other than physical/cognitive); 2=Participant refused; 3=Participant unable for physical reasons; 4=Participant unable for cognitive reasons (NOTE: This should only be checked if the task was not administered due to cognitive reasons.) Decoded Value: Yes

COT4LIST character Check each word correctly recalled. 1=BUTTER; 2=ARM; 3=SHORE; 4=LETTER; 5=QUEEN; 6=CABIN; 7=POLE; 8=TICKET; 9=GRASS; 10=ENGINE Decoded Value: Yes

COT4TOTL numeric Total 0..10

COP4COMP character If task not completed, check one: 1=Not done (for reasons other than physical/cognitive); 2=Participant refused; 4=Participant unable for cognitive reasons (NOTE: A score of '0' should be given if the participant was unable to complete the task for cognitive reasons.) Decoded Value: Yes

CONAME character Check each object and finger named correctly. 1=Flower; 2=Bed; 3=Whistle; 4=Pencil; 5=Rattle; 6=Mask; 7=Scissors; 8=Comb; 9=Wallet; 10=Harmonica; 11=Stethoscope; 12=Tongs; 13=Thumb; 14=Middle; 15=Ring; 16=Index; 17=Pinky; 18=None completed correctly Decoded Value: Yes

COP5COMP character If task not administered, check one: 1=Not done (for reasons other than physical/cognitive); 2=Participant refused; 4=Participant unable for cognitive reasons (NOTE: This should only be checked if the task was not administered due to cognitive reasons.) Decoded Value: Yes

COIDEA character Check each step completed correctly. 1=Fold a letter.; 2=Put letter in envelope.; 3=Seal envelope.; 4=Address envelope.; 5=Indicate where stamp goes.; 6=None completed correctly Decoded Value: Yes

COP6COMP character If task not administered, check one: 1=Not done (for reasons other than physical/cognitive); 2=Participant refused; 3=Participant unable for physical reasons; 4=Participant unable for cognitive reasons (NOTE: This should only be checked if the task was not administered due to cognitive reasons.) Decoded Value: Yes

COORIEN character Check each item answered correctly. 1=Full name; 2=Month; 3=Date; 4=Year; 5=Day; 6=Season; 7=Place; 8=Time of day; 9=None completed correctly Decoded Value: Yes

COP7COMP character If task not administered, check one: 1=Not done (for reasons other than physical/cognitive); 2=Participant refused; 4=Participant unable for cognitive reasons (NOTE: This should only be checked if the task was not administered due to cognitive reasons.) Decoded Value: Yes

CO8NURSE character NURSE - Participant Response 1=Yes; 0=No Decoded Value: Yes

CO8MAGAZ character MAGAZINE - Participant Response 1=Yes; 0=No Decoded Value: Yes

CO8WIZRD character WIZARD - Participant Response 1=Yes; 0=No Decoded Value: Yes

CO8WIZRR character Reminder given? 1=Check here if reminder given Decoded Value: Yes

CO8VAN character VAN - Participant Response 1=Yes; 0=No Decoded Value: Yes

CO8VANR character Reminder given? 1=Check here if reminder given Decoded Value: Yes

CO8LEPRD character LEOPARD - Participant Response 1=Yes; 0=No Decoded Value: Yes

CO8LEPRR character Reminder given? 1=Check here if reminder given Decoded Value: Yes

CO8SALE character SALE - Participant Response 1=Yes; 0=No Decoded Value: Yes

CO8SALER character Reminder given? 1=Check here if reminder given Decoded Value: Yes

CO8SEA character SEA - Participant Response 1=Yes; 0=No Decoded Value: Yes

CO8SEAR character Reminder given? 1=Check here if reminder given Decoded Value: Yes

CO8TRAIN character TRAIN - Participant Response 1=Yes; 0=No Decoded Value: Yes

CO8TRAIR character Reminder given? 1=Check here if reminder given Decoded Value: Yes

CO8COIN character COIN - Participant Response 1=Yes; 0=No Decoded Value: Yes

CO8COINR character Reminder given? 1=Check here if reminder given Decoded Value: Yes

CO8SHIP character SHIP - Participant Response 1=Yes; 0=No Decoded Value: Yes

CO8SHIPR character Reminder given? 1=Check here if reminder given Decoded Value: Yes

CO8INST character INSTITUTION - Participant Response 1=Yes; 0=No Decoded Value: Yes

CO8INSTR character Reminder given? 1=Check here if reminder given Decoded Value: Yes

CO8MAP character MAP - Participant Response 1=Yes; 0=No Decoded Value: Yes

CO8MAPR character Reminder given? 1=Check here if reminder given Decoded Value: Yes

CO8AXE character AXE - Participant Response 1=Yes; 0=No Decoded Value: Yes

CO8AXER character Reminder given? 1=Check here if reminder given Decoded Value: Yes

CO8BOARD character BOARD - Participant Response 1=Yes; 0=No Decoded Value: Yes

CO8BOARR character Reminder given? 1=Check here if reminder given Decoded Value: Yes

CO8CARRT character CARROT - Participant Response 1=Yes; 0=No Decoded Value: Yes

CO8CARRR character Reminder given? 1=Check here if reminder given Decoded Value: Yes

CO8MILK character MILK - Participant Response 1=Yes; 0=No Decoded Value: Yes

CO8MILKR character Reminder given? 1=Check here if reminder given Decoded Value: Yes

CO8VOL character VOLUME - Participant Response 1=Yes; 0=No Decoded Value: Yes

CO8VOLR character Reminder given? 1=Check here if reminder given Decoded Value: Yes

CO8FORST character FOREST - Participant Response 1=Yes; 0=No Decoded Value: Yes

CO8FORSR character Reminder given? 1=Check here if reminder given Decoded Value: Yes

CO8ANCHR character ANCHOR - Participant Response 1=Yes; 0=No Decoded Value: Yes

CO8ANCRR character Reminder given? 1=Check here if reminder given Decoded Value: Yes

CO8GEM character GEM - Participant Response 1=Yes; 0=No Decoded Value: Yes

CO8GEMR character Reminder given? 1=Check here if reminder given Decoded Value: Yes

CO8CAT character CAT - Participant Response 1=Yes; 0=No Decoded Value: Yes

CO8CATR character Reminder given? 1=Check here if reminder given Decoded Value: Yes

CO8FUND character FUND - Participant Response 1=Yes; 0=No Decoded Value: Yes

CO8FUNDR character Reminder given? 1=Check here if reminder given Decoded Value: Yes

CO8EDGE character EDGE - Participant Response 1=Yes; 0=No Decoded Value: Yes

CO8EDGER character Reminder given? 1=Check here if reminder given Decoded Value: Yes

CO8CAKE character CAKE - Participant Response 1=Yes; 0=No Decoded Value: Yes

CO8CAKER character Reminder given? 1=Check here if reminder given Decoded Value: Yes

COP8COMP character If task not administered, check one: 1=Not done (for reasons other than physical/cognitive); 2=Participant refused; 4=Participant unable for cognitive reasons Decoded Value: Yes

COINSTRC character Check level of impairment. Base your answer only on observations made during Word Recognition task (Question #8). 1=None; 2=Very mild: forgets once; 3=Mild: must be reminded two times; 4=Moderate: must be reminded 3-4 times; 5=Moderately severe: must be reminded 5-6 times; 6=Severe: must be reminded 7 or more times. Decoded Value: Yes

COCOMPRE character Check level of impairment. 1=None: participant understands; 2=Very mild: one or two instances of misunderstanding; 3=Mild 3-5 instances of misunderstanding; 4=Moderate: requires several repetitions and rephrasings; 5=Moderately severe: participant only occasionally responds correctly, i.e., yes-or-no questions; 6=Severe: participant rarely responds to questions appropriately, not due to poverty of speech Decoded Value: Yes

COWRDFND character Check one response 1=None; 2=Very Mild: one or two instances, not clinically significant; 3=Mild: noticeable circumlocution or synonym substitution; 4=Moderate: loss of words without compensation on occasion; 5=Moderately severe: frequent loss of words without compensation; 6=Severe: nearly total loss of content words, speech sounds empty, one-to-two-word utterances Decoded Value: Yes

COLANG character Check level of impairment. 1=None: participant speaks clearly and/or is understandable; 2=Very mild: one instance of lack of understandability; 3=Mild: participant has difficulty <25% of the time; 4=Moderate: participant has difficulty 25-50% of the time; 5=Moderately Severe: participant has difficulty more than 50% of the time; 6=Severe: one- or two-word utterances, fluent but empty speech, mute Decoded Value: Yes

CONMCXLA numeric 14a. Number of targets hit (range = 0-49) 0..49

CONMCXLB numeric 14b. Number of errors 0..99

CONMCXLC numeric 14c. Number of times reminded of task 0..99

COP14CMP character If task not administered, check one: 1=Not done (for reasons other than physical/cognitive); 2=Participant refused; 3=Participant unable for physical reasons; 4=Participant unable for cognitive reasons Decoded Value: Yes

ID numeric Range value: 4, ..., 7784

SITEID numeric Site ID Range value: 1, ..., 127

USERDATE Date record created Date: YYY-MM-DD

USERDATE2 Date record last updated Date: YYY-MM-DD

update_stamp POSIXct: YYY-MM-DD

Author

adni-data@googlegroups.com