ADAS-Cognitive Behavior

ADAS-Cognitive Behavior data. More information is available at https://adni.loni.usc.edu/data-samples/adni-data/

Usage

data(ADAS_ADNIGO23)

Format

A data frame with 7746 observations and 125 variables.

Source

https://adni.loni.usc.edu/data-samples/adni-data/

Details

ORIGPROT character Original study protocol , with options: ADNI1, ADNIGO, ADNI2, ADNI3

COLPROT factor Study protocol of data collection , with factor levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4

PTID character Participant ID

RID numeric Range value: 2, ..., 7125

VISCODE character

VISCODE2 character Translated visit code

VISDATE Assessment EXAMDATE when present; otherwise Registry EXAMDATE Date: YYY-MM-DD

WORDLIST character Word Recall and Recognition Tasks List Number 1=English Word List 1; 3=English Word List 2; 5=English Word List 3; 7=English Word List 4; 2=Spanish Word List 1; 4=Spanish Word List 2; 6=Spanish Word List 3; 8=Spanish Word List 4 Decoded Value: Yes

Q1UNABLE character Task 1. Word Recall

ADNIGO, ADNI2: 1=Not done (for reasons other than physical/cognitive); 2=Participant refused; 4=Participant unable to complete 3 trials for cognitive reasons

ADNI3: 0=Conducted;1=Not done (for reasons other than physical/cognitive);2=Participant refused;4=Participant unable to complete 3 trials for cognitive reasons

Decoded Value: Yes

Q1TR1 character Trial 1

ADNIGO, ADNI2: 0=Word 1;1=Word 2; 2=Word 3; 3=Word 4; 4=Word 5; 5=Word 6; 6=Word 7; 7=Word 8; 8=Word 9; 9=Word10;10=None recalled correctly

ADNI3: 0=Butter;1=Arm;2=Shore;3=Letter;4=Queen;5=Cabin;6=Pole;7=Ticket;8=Grass;9=Engine;10=None recalled correctly

Decoded Value: Yes

Q1TR2 character Trial 2

ADNIGO, ADNI2: 0=Word 1;1=Word 2; 2=Word 3; 3=Word 4; 4=Word 5; 5=Word 6; 6=Word 7; 7=Word 8; 8=Word 9; 9=Word10;10=None recalled correctly

ADNI3: 0=Pole;1=Letter;2=Butter;3=Queen;4=Arm;5=Shore;6=Grass;7=Cabin;8=Ticket;9=Engine;10=None recalled correctly

Decoded Value: Yes

Q1TR3 character Trial 3

ADNIGO, ADNI2: 0=Word 1;1=Word 2; 2=Word 3; 3=Word 4; 4=Word 5; 5=Word 6; 6=Word 7; 7=Word 8; 8=Word 9; 9=Word10;10=None recalled correctly

ADNI3: 0=Shore;1=Letter;2=Arm;3=Cabin;4=Pole;5=Ticket;6=Engine;7=Grass;8=Butter;9=Queen;10=None recalled correctly

Decoded Value: Yes

Q1TRIT numeric Trial 1 Total Range value: 0, ..., 10

Q1TR2T numeric Trial 2 Total Range value: 0, ..., 10

Q1TRT numeric Trial 3 Total Range value: 0, ..., 10

Q1SCORE numeric ADAS Score Component Q1 Word Recall 0..10

TIMEEND numeric Time Ended: Range value: -1, ..., 1819

Q2UNABLE character Task 2. Commands

ADNIGO, ADNI2: 1=Not done (for reasons other than physical/cognitive); 2=Participant refused; 3=Participant unable for physical reasons;4=Participant unable for cognitive reasons

ADNI3: 0=Conducted;1=Not done (for reasons other than physical/cognitive);2=Participant refused;3=Participant unable for physical reasons;4=Participant unable for cognitive reasons

Decoded Value: Yes

Q2TASK character 2. Commands

ADNIGO, ADNI2: 1=Make a fist; 2=Point to the ceiling, then to the floor; 3=Put the pencil on top of the card, then put it back; 4=Put the watch on the other side of the pencil and turn over the card; 5=Tap each shoulder twice with two fingers, keeping your eyes shut; 0=None performed correctly

ADNI3: 0=None performed correctly;1=a. Make a fist;2=b. Point to the ceiling, then to the floor;3=c. Put the pencil on top of the card, then put it back;4=d. Put the watch on the other side of the pencil and turn over the card;5=e. Tap each shoulder twice with two fingers, keeping your eyes shut

Decoded Value: Yes

Q2SCORE numeric ADAS Score Component Q2 Commands 0..5

Q3UNABLE character Task 3. Constructional Praxis

ADNIGO, ADNI2: 1=Not done (for reasons other than physical/cognitive); 2=Participant refused; 3=Participant unable for physical reasons; 4=Participant unable for cognitive reasons

ADNI3: 0=Conducted;1=Not done (for reasons other than physical/cognitive);2=Participant refused;3=Participant unable for physical reasons;4=Participant unable for cognitive reasons

Decoded Value: Yes

Q3TASK1 character a. Circle

ADNIGO, ADNI2: 1=Shape drawn correctly;2=Shape drawn incorrectly (at least one side/section of shape drawn);3=No recognizable attempt at drawing any side/section of shape

ADNI3: 1=Figure drawn correctly;2=Figure drawn incorrectly (at least one side/section of figure drawn);3=No recognizable attempt at drawing any side/section of figure

Decoded Value: Yes

Q3TASK2 character b. Two overlapping rectangles

ADNIGO, ADNI2: 1=Shape drawn correctly;2=Shape drawn incorrectly (at least one side/section of shape drawn);3=No recognizable attempt at drawing any side/section of shape

ADNI3: 1=Figure drawn correctly;2=Figure drawn incorrectly (at least one side/section of figure drawn);3=No recognizable attempt at drawing any side/section of figure

Decoded Value: Yes

Q3TASK3 character c. Rhombus (Diamond)

ADNIGO, ADNI2: 1=Shape drawn correctly;2=Shape drawn incorrectly (at least one side/section of shape drawn);3=No recognizable attempt at drawing any side/section of shape

ADNI3: 1=Figure drawn correctly;2=Figure drawn incorrectly (at least one side/section of figure drawn);3=No recognizable attempt at drawing any side/section of figure

Decoded Value: Yes

Q3TASK4 character d. Cube

ADNIGO, ADNI2: 1=Shape drawn correctly;2=Shape drawn incorrectly (at least one side/section of shape drawn);3=No recognizable attempt at drawing any side/section of shape

ADNI3: 1=Figure drawn correctly;2=Figure drawn incorrectly (at least one side/section of figure drawn);3=No recognizable attempt at drawing any side/section of figure

Decoded Value: Yes

Q3SCORE numeric ADAS Score Component Q3 Constructional Praxis 0..5

Q4UNABLE character Task 4. Delayed Word Recall

ADNIGO, ADNI2: 1=Not done (for reasons other than physical/cognitive); 2=Participant refused; 4=Participant unable for cognitive reasons

ADNI3: 0=Conducted;1=Not done (for reasons other than physical/cognitive);2=Participant refused;4=Participant unable for cognitive reasons

Decoded Value: Yes

TIMEBEGAN numeric Time Began: Range value: -1, ..., 1826

Q4TASK character 4. Delayed Word Recall

ADNIGO, ADNI2: 0=Word 1;1=Word 2; 2=Word 3; 3=Word 4; 4=Word 5; 5=Word 6; 6=Word 7; 7=Word 8; 8=Word 9; 9=Word10;10=None recalled correctly

ADNI3: 0=Butter;1=Arm;2=Shore;3=Letter;4=Queen;5=Cabin;6=Pole;7=Ticket;8=Grass;9=Engine;10=None recalled correctly

Decoded Value: Yes

Q4SCORE numeric ADAS Score Component Q4 Delayed Word Recall 0..10

Q5UNABLE character Task 5. Naming Objects and Fingers

ADNIGO, ADNI2: 1=Not done (for reasons other than physical/cognitive); 2=Participant refused; 4=Participant unable for cognitive reasons

ADNI3: 0=Conducted;1=Not done (for reasons other than physical/cognitive);2=Participant refused;4=Participant unable for cognitive reasons

Decoded Value: Yes

Q5TASK character 5. Naming Objects and Fingers

ADNIGO: 1=Flower;2=Bed;3=Whistle;4=Pencil;5=Rattle;6=Mask;7=Scissors;8=Comb;9=Wallet;10=Harmonica;11=Stethoscope;12=Tongs;13=Thumb;14=Middle;15=Ring;16=Index;17=Pinky;0=None completed correctly

ADNI2: 1=Flower;2=Bed;3=Whistle;4=Pencil;5=Rattle;6=Mask;7=Scissors;8=Comb;9=Wallet;10=Harmonica;11=Stethoscope;12=Tongs;13=Thumb;16=Index/Pointer;14=Middle;15=Ring;17=Pinky;0=None completed correctly

Decoded Value: Yes

Q5SCORE numeric ADAS Score Component Q5 Naming Objects and Fingers (without cue) 0..5

Q5NAME1 character Flower 0=Incorrect;1=Correct - no cue;2=Correct with semantic cue Decoded Value: Yes

Q5NAME2 character Bed 0=Incorrect;1=Correct - no cue;2=Correct with semantic cue Decoded Value: Yes

Q5NAME3 character Whistle 0=Incorrect;1=Correct - no cue;2=Correct with semantic cue Decoded Value: Yes

Q5NAME4 character Pencil 0=Incorrect;1=Correct - no cue;2=Correct with semantic cue Decoded Value: Yes

Q5NAME5 character Rattle 0=Incorrect;1=Correct - no cue;2=Correct with semantic cue Decoded Value: Yes

Q5NAME6 character Mask 0=Incorrect;1=Correct - no cue;2=Correct with semantic cue Decoded Value: Yes

Q5NAME7 character Scissors 0=Incorrect;1=Correct - no cue;2=Correct with semantic cue Decoded Value: Yes

Q5NAME8 character Comb 0=Incorrect;1=Correct - no cue;2=Correct with semantic cue Decoded Value: Yes

Q5NAME9 character Wallet 0=Incorrect;1=Correct - no cue;2=Correct with semantic cue Decoded Value: Yes

Q5NAME10 character Harmonica 0=Incorrect;1=Correct - no cue;2=Correct with semantic cue Decoded Value: Yes

Q5NAME11 character Stethoscope 0=Incorrect;1=Correct - no cue;2=Correct with semantic cue Decoded Value: Yes

Q5NAME12 character Tongs 0=Incorrect;1=Correct - no cue;2=Correct with semantic cue Decoded Value: Yes

Q5FINGER character Naming Fingers 0=None recalled correctly;1=Thumb;2=Middle;3=Ring;4=Index/Pointer/Forefinger;5=Pinky/Little Decoded Value: Yes

Q5SCORE_CUE numeric ADAS Score Component Q5 Naming Objects and Fingers (with cue when necessary) 0..5

Q6UNABLE character Task 6. Ideational Praxis

ADNIGO, ADNI2: 1=Not done (for reasons other than physical/cognitive); 2=Participant refused; 3=Participant unable for physical reasons; 4=Participant unable for cognitive reasons

ADNI3: 0=Conducted;1=Not done (for reasons other than physical/cognitive);2=Participant refused;3=Participant unable for physical reasons;4=Participant unable for cognitive reasons

Decoded Value: Yes

Q6TASK character 6. Ideational Praxis

ADNIGO: 1=Fold a letter; 2=Put letter in envelope; 3=Seal envelope; 4=Address envelope; 5=Indicate where stamp goes; 0=None completed correctly

ADNI2: 1=Fold the paper; 2=Put the paper in the envelope; 3=Seal envelope; 4=Address envelope; 5=Indicate where stamp goes; 0=None completed correctly

ADNI3: 0=None completed correctly;1=a. Fold the paper;2=b. Put the paper in the envelope;3=c. Seal envelope;4=d. Address envelope;5=e. Show me where the stamp goes

Decoded Value: Yes

Q6SCORE numeric ADAS Score Component Q6 Ideational Praxis 0..5

Q7UNABLE character Task 7. Orientation

ADNIGO, ADNI2: 1=Not done (for reasons other than physical/cognitive); 2=Participant refused; 4=Participant unable for cognitive reasons

ADNI3: 0=Conducted;1=Not done (for reasons other than physical/cognitive);2=Participant refused;4=Participant unable for cognitive reasons

Decoded Value: Yes

Q7TASK character 7. Orientation

ADNIGO: 1=Full name;2=Month;3=Date;4=Year;5=Day;6=Season;7=Place;8=Time of day;0=None completed correctly

ADNI2: 1=Full name;2=Month;3=Date;4=Year;5=Day;6=Season;7=Place;8=Clock Time (+/- one hour);0=None completed correctly

ADNI3: 0=None completed correctly;1=a. Full name (exact);2=b. Month (exact);3=c. Date (± 1 day);4=d. Year (exact);5=e. Day of the week (exact);6=f. Season (within 1 week of upcoming or within 2 weeks of previous season);7=g. Place (Refer to your site-specific scoring rules) ;8=h. Clock Time (+/- one hour)

Decoded Value: Yes

Q7SCORE numeric ADAS Score Component Q7 Orientation 0..8

Q8UNABLE character Task 8. Word Recognition

ADNIGO, ADNI2: 1=Not done (for reasons other than physical/cognitive); 2=Participant refused; 4=Participant unable for cognitive reasons

ADNI3: 0=Conducted;1=Not done (for reasons other than physical/cognitive);2=Participant refused;4=Participant unable for cognitive reasons

Decoded Value: Yes

Q8WORD1 character Nurse

ADNIGO, ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

Decoded Value: Yes

Q8WORD1R character Reminder given? 1=Check here if reminder given Decoded Value: Yes

Q8WORD2 character Magazine

ADNIGO, ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

Decoded Value: Yes

Q8WORD2R character Reminder given? 1=Check here if reminder given Decoded Value: Yes

Q8WORD3 character Wizard

ADNIGO, ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

Decoded Value: Yes

Q8WORD3R character Reminder given? [Word 3] 1=Check here if reminder given Decoded Value: Yes

Q8WORD4 character Van

ADNIGO, ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

Decoded Value: Yes

Q8WORD4R character Reminder given? [Word 4] 1=Check here if reminder given Decoded Value: Yes

Q8WORD5 character Leopard

ADNIGO, ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

Decoded Value: Yes

Q8WORD5R character Reminder given? [Word 5] 1=Check here if reminder given Decoded Value: Yes

Q8WORD6 character Sale

ADNIGO, ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

Decoded Value: Yes

Q8WORD6R character Reminder given? [Word 6] 1=Check here if reminder given Decoded Value: Yes

Q8WORD7 character Sea

ADNIGO, ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

Decoded Value: Yes

Q8WORD7R character Reminder given? [Word 7] 1=Check here if reminder given Decoded Value: Yes

Q8WORD8 character Train

ADNIGO, ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

Decoded Value: Yes

Q8WORD8R character Reminder given? [Word 8] 1=Check here if reminder given Decoded Value: Yes

Q8WORD9 character Coin

ADNIGO, ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

Decoded Value: Yes

Q8WORD9R character Reminder given? [Word 9] 1=Check here if reminder given Decoded Value: Yes

Q8WORD10 character Ship

ADNIGO, ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

Decoded Value: Yes

Q8WORD10R character Reminder given? [Word 10] 1=Check here if reminder given Decoded Value: Yes

Q8WORD11 character Institution

ADNIGO, ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

Decoded Value: Yes

Q8WORD11R character Reminder given? [Word 11] 1=Check here if reminder given Decoded Value: Yes

Q8WORD12 character Map

ADNIGO, ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

Decoded Value: Yes

Q8WORD12R character Reminder given? [Word 12] 1=Check here if reminder given Decoded Value: Yes

Q8WORD13 character Axe

ADNIGO, ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

Decoded Value: Yes

Q8WORD13R character Reminder given? [Word 13] 1=Check here if reminder given Decoded Value: Yes

Q8WORD14 character Board

ADNIGO, ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

Decoded Value: Yes

Q8WORD14R character Reminder given? [Word 14] 1=Check here if reminder given Decoded Value: Yes

Q8WORD15 character Carrot

ADNIGO, ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

Decoded Value: Yes

Q8WORD15R character Reminder given? [Word 15] 1=Check here if reminder given Decoded Value: Yes

Q8WORD16 character Milk

ADNIGO, ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

Decoded Value: Yes

Q8WORD16R character Reminder given? [Word 16] 1=Check here if reminder given Decoded Value: Yes

Q8WORD17 character Volume

ADNIGO, ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

Decoded Value: Yes

Q8WORD17R character Reminder given? [Word 17] 1=Check here if reminder given Decoded Value: Yes

Q8WORD18 character Forest

ADNIGO, ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

Decoded Value: Yes

Q8WORD18R character Reminder given? [Word 18] 1=Check here if reminder given Decoded Value: Yes

Q8WORD19 character Anchor

ADNIGO, ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

Decoded Value: Yes

Q8WORD19R character Reminder given? [Word 19] 1=Check here if reminder given Decoded Value: Yes

Q8WORD20 character Gem

ADNIGO, ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

Decoded Value: Yes

Q8WORD20R character Reminder given? [Word 20] 1=Check here if reminder given Decoded Value: Yes

Q8WORD21 character Cat

ADNIGO, ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

Decoded Value: Yes

Q8WORD21R character Reminder given? [Word 21] 1=Check here if reminder given Decoded Value: Yes

Q8WORD22 character Fund

ADNIGO, ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

Decoded Value: Yes

Q8WORD22R character Reminder given? [Word 22] 1=Check here if reminder given Decoded Value: Yes

Q8WORD23 character Edge

ADNIGO, ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

Decoded Value: Yes

Q8WORD23R character Reminder given? [Word 23] 1=Check here if reminder given Decoded Value: Yes

Q8WORD24 character Cake

ADNIGO, ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

Decoded Value: Yes

Q8WORD24R character Reminder given? [Word 24] 1=Check here if reminder given Decoded Value: Yes

Q8SCORE numeric ADAS Score Component Q8 Word Recognition 0..12

Q9TASK character Task 9. Remembering Test Instructions (score this item based on responses recorded in Item 8.)

ADNIGO, ADNI2: 0=None (0); 1=Very mild: forgets once (1); 2=Mild: must be reminded two times (2); 3=Moderate: must be reminded 3-4 times; 4=Moderately severe: must be reminded 5-6 times; 5=Severe: must be reminded 7 or more times.

ADNI3: 0=None: participant never needs extra reminders of instructions;1=Very mild: forgets once;2=Mild: must be reminded two times;3=Moderate: must be reminded 3-4 times;4=Moderately severe: must be reminded 5-6 times;5=Severe: must be reminded 7 or more times.

Decoded Value: Yes

Q9SCORE numeric ADAS Score Component Q9 Remembering Test Instructions 0..5

Q10TASK character 10. Comprehension of Spoken Language

ADNIGO, ADNI2: 0=None: participant understands; 1=Very mild: one or two instances of misunderstanding; 2=Mild 3-5 instances of misunderstanding; 3=Moderate: requires several repetitions and rephrasings; 4=Moderately severe: participant only occasionally responds correctly. i.e., yes-or-no questions; 5=Severe: participant rarely responds to questions appropriately, not due to poverty of speech

ADNI3: 0=None: no evidence of poor comprehension;1=Very mild: one or two instances of misunderstanding;2=Mild: 3-5 instances of misunderstanding;3=Moderate: requires several repetitions and rephrasings;4=Moderately severe: participant only occasionally responds correctly. i.e., yes-or-no questions;5=Severe: participant rarely responds to questions appropriately, not due to poverty of speech

Decoded Value: Yes

Q10SCORE numeric ADAS Score Component Q10 Comprehension of Spoken Language 0..5

Q11TASK character 11. Word-finding Difficulty

ADNIGO, ADNI2: 0=None; 1=Very Mild: one or two instances, not clinically significant; 2=Mild: noticeable circumlocution or synonym substitution; 3=Moderate: loss of words without compensation on occasion; 4=Moderately severe: frequent loss of words without compensation; 5=Severe: nearly total loss of content words, speech sounds empty, one-to-two-word utterances

ADNI3: 0=None: no evidence of word finding difficulty in spontaneous speech;1=Very Mild: one or two instances, not clinically significant;2=Mild: noticeable circumlocution or synonym substitution;3=Moderate: loss of words without compensation on occasion;4=Moderately severe: frequent loss of words without compensation;5=Severe: nearly total loss of content words, speech sounds empty, one-to-two-word utterances

Decoded Value: Yes

Q11SCORE numeric ADAS Score Component Q11 Word-finding Difficulty 0..5

Q12TASK character 12. Spoken Language Ability

ADNIGO, ADNI2: 0=None: participant speaks clearly and/or is understandable; 1=Very mild: one instance of lack of understandability; 2=Mild: participant has difficulty <25% of the time; 3=Moderate: participant has difficulty 25-50% of the time; 4=Moderately Severe: participant has difficulty more than 50% of the time; 5=Severe: one- or two-word utterances, fluent but empty speech, mute

ADNI3: 0=None: participant speaks clearly and/or is understandable;1=Very mild: one instance of lack of understandability;2=Mild: participant has difficulty <25% of the time;3=Moderate: participant has difficulty 25-50% of the time;4=Moderately Severe: participant has difficulty more than 50% of the time;5=Severe: one- or two-word utterances, fluent but empty speech, mute

Decoded Value: Yes

Q12SCORE numeric ADAS Score Component Q12 Language 0..5

TOTSCORE numeric Total Score (ADAS 11)

ADNIGO, ADNI2

ADNI3: 0..70

Q13UNABLE character Task 13. Number Cancellation

ADNIGO, ADNI2: 1=Not done (for reasons other than physical/cognitive); 2=Participant refused; 3=Participant unable for physical reasons; 4=Participant unable for cognitive reasons

ADNI3: 0=Conducted;1=Not done (for reasons other than physical/cognitive);2=Participant refused;3=Participant unable for physical reasons;4=Participant unable for cognitive reasons

Decoded Value: Yes

Q13TASKA numeric 13a. Number of targets hit

ADNIGO, ADNI2: 0..49

ADNI3

Q13TASKB numeric 13b. Number of errors

ADNIGO, ADNI2: 0..99

ADNI3

Q13TASKC numeric 13c. Number of times reminded of task

ADNIGO, ADNI2: 0..99

ADNI3

Q13SCORE numeric ADAS Score Component Q13 Number Cancellation 0..5

TOTAL13 numeric Total Score (ADAS 13)

ADNIGO, ADNI2

ADNI3: 0..85

DATE Date Collected Date: YYY-MM-DD

SOURCE character Information Source 1=Participant Visit;2=Telephone Call/Video Call Decoded Value: Yes

ID numeric Range value: 4, ..., 186200

SITEID numeric Range value: 1, ..., 63

USERDATE Date: YYY-MM-DD

USERDATE2 Date: YYY-MM-DD

update_stamp POSIXct: YYY-MM-DD

Author

adni-data@googlegroups.com