Tau AV-1451 PET Eligibility

Tau AV-1451 PET Eligibility data. More information is available at https://adni.loni.usc.edu/data-samples/adni-data/

Usage

data(ADNI2_OTELGTAU)

Format

A data frame with 148 observations and 22 variables.

Source

https://adni.loni.usc.edu/data-samples/adni-data/

Details

ORIGPROT character Original study protocol , with options: ADNI2, ADNI1, ADNIGO

ID numeric Range value: 2, ..., 308

RID numeric Participant roster ID Range value: 55, ..., 5277

SITEID numeric Site ID Range value: 1, ..., 60

VISCODE character Visit code , with options: v31, v21, v41, v51, tau

VISCODE2 character Translated visit code , with options: m36, m108, m24, m96, m48, m60, m120, m72

USERDATE Date record created Date: YYY-MM-DD

USERDATE2 Date record last updated Date: YYY-MM-DD

RECNO numeric Range value: 1, ..., 2

TAUELG numeric Please confirm if participant meets eligibility criteria: 1=Participant is eligible;0=Participant should be excluded

EXCLUSION numeric If participant should be excluded, select reason: 1=Exclusion Criteria

NVROT numeric 1=Other (specify)

NVRSPECIFY character If Other, specify: , with options: Participant did not consent to scan, Exluded from Past EKG results

EXC1 logical If Exclusion Criteria, which criteria does the participant meet: 1=[1] Has any condition that, in the investigator's opinion, could increase risk to the participant, limit the participant's ability to tolerate the experimental procedures, or interfere with collection/analysis of the data (for example, participants with severe chronic back pain might not be able to lie still during the scanning procedures).

EXC2 logical 1=[2] Has abnormal findings on physical examination, or laboratory screening tests that suggest the patient might have a condition that could, in the opinion of the investigator, affect his or her response to the radiopharmaceutical and related testing procedures.

EXC3 logical 1=[3] Is deemed likely to be unable to perform all of the addendum imaging procedures for any reason.

EXC4 logical 1=[4] Female that is pregnant, lactating, or of childbearing potential

EXC5 numeric 1=[5] Has a history of risk factors for Torsades de Pointes or is taking medications known to prolong QT interval (a list of prohibited and discouraged medications is provided by the Sponsor).

EXC6 numeric 1=[6] Has ECG obtained prior to the initial 18F-AV-1451 scan that in the opinion of the investigator is clinically significant with regard to the subject's participation in the study. Bazett's corrected QT (QTcB) interval must be evaluated and must not exceed 458 msec in males or 474 msec in females.

EXC7 logical 1=[7] Has hypersensitivity to 18F-AV-1451 or any of its excipients.

TAUCOMM character Comments: , with options: QT interval 502, Pt originally scheduled 1/26/16 but avid had transportation delays that would'vd had scan ending @ 1830 and pt didn't want end that late. Soonest could reschedule is 2/23/16., Abnormal EKG, but no evidence of conduct delay. Communicated with ADNI medical director and Bill Jasvist who felt subject okay if electrically stable, which he is., Patient taking cyclobenzaprine, which can cause prolongation of QT interval, Participant eligible to continue, Since consent obtained and record/EKG's initially reviewed, subject has been started on flecainide, so is no longer eligible for tau scan.

update_stamp POSIXct: YYY-MM-DD

Author

adni-data@googlegroups.com