[DERIVED] Response Analysis Dataset

Response analysis dataset derived dataset.

Usage

data(ADRS)

Format

A data frame with 15689 observations and 34 variables.

Source

For more details see the help vignette: vignette(topic = 'ADNIMERGE2-Analysis-Data', package = 'ADNIMERGE2')

Details

STUDYID character Study Identifier , with character options: ADNI

USUBJID character Unique Subject Identifier

SUBJID character Subject Identifier for the Study

COLPROT factor Study Protocol of Data Collection , with factor levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4

PARAMCD character Parameter Code , with character options: DX, DEATH

PARAM character Parameter , with character options: Clinical Diagnostics Status, Death

PARAMN numeric Parameter (N) Range value: 1, ..., 2

AVISIT character Analysis Visit

AVISITN numeric Analysis Visit (N) Range value: -1849, ..., 7085

VISIT character Visit Name

VISITNUM numeric Visit Number Range value: -2, ..., 999

ADY numeric Analysis Relative Day Range value: -1849, ..., 7085

ADT Date Analysis Date Date: YYY-MM-DD

AVAL numeric Analysis Value Range value: 1, ..., 4

AVALC character Analysis Value (C) , with character options: DEM, MCI, CN, DEATH

BASE character Baseline Value , with character options: DEM, MCI, CN

ABLFL character Baseline Record Flag , with character options: Y

ENRLFL character Enrolled Population Flag , with character options: Y

ENRLDT Date Date of Enrollment Date: YYY-MM-DD

EOSDT Date End of Study Date Date: YYY-MM-DD

EOSDTF character End of Study Date Imputation Flag , with character options: M

BLBFL character Baseline Records Population Flag , with character options: Y

FOLLOWPFL character Follow-up Records Population Flag , with character options: Y, M

ANL01FL character Analysis Flag , with character options: Y

AGE numeric Age (in Years) Range value: 50.5, ..., 91.4

AGEU character Age Units , with character options: Years

AGEGR1 character Pooled Age Group 1 , with character options: >=65, <65

AGEGR1N numeric Pooled Age Group 1 (N) Range value: 1, ..., 2

SEX character Sex , with character options: Male, Female

RACE factor Race , with factor levels: American Indian or Alaskan Native, Asian, Black or African American, Native Hawaiian or Other Pacific Islander, Other Pacific Islander, White, More than one race, Unknown

BMI numeric Body Mass Index Range value: 18.64, ..., 47.87

BMIBLU character Body Mass Index Unit , with character options: kg/m^2

DX factor Baseline Diagnostics Status , with factor levels: CN, MCI, DEM

AMYSTAT character Baseline Amyloid Status

Author

adni-data@googlegroups.com