ADSP Phenotype Harmonization Consortium (PHC) - Composite FLAIR Scores

ADSP Phenotype Harmonization Consortium (PHC) - Composite FLAIR Scores data. More information is available at https://adni.loni.usc.edu/data-samples/adni-data/

Usage

data(ADSP_PHC_FLAIR)

Format

A data frame with 7230 observations and 27 variables.

Source

https://adni.loni.usc.edu/data-samples/adni-data/

Details

ORIGPROT character Original study protocol , with options: ADNI4, ADNI1, ADNIGO, ADNI2, ADNI3

COLPROT character Study protocol of data collection , with options: ADNI4, ADNI2, ADNI3, ADNI1, ADNIGO

RID numeric Participant roster ID

PTID character Participant ID

SUBJID character Subject ID used in NIAGADS for the whole genome sequencing data files generated by the AD Sequencing Project.

VISCODE character Visit code

VISCODE2 character Translated visit code

PHC_LONIUID numeric LONI Image ID Range value: 35254, ..., 11433158

PHC_SCANDATE Date of scan Date: YYYY-MM-DD

PHC_Visit numeric Harmonized visit number Range value: 1, ..., 20

Imaging_Site numeric Location Site of Imaging Scan Range value: 2, ..., 941

PHC_Age_FLAIR numeric Age at each FLAIR visit Range value: 50.5325119780972, ..., 97.45106091718

FieldStrength numeric Strength of the magnetic field used in the MRI scanner, measured in Tesla. Range value: 1.5, ..., 3

Manufacturer character Manufacturer of the MRI scanner , with options: Siemens Healthineers, SIEMENS, Philips, GE MEDICAL SYSTEMS, Philips Medical Systems, Philips Healthcare, Siemens

Model character The specific model name of the MRI scanner, usually provided by the manufacturer.

Image_Dimensions_vox character Dimensions of the acquired FLAIR image

SliceThickness numeric Thickness of each individual slice or section in the FLAIR image, measured in millimeters Range value: 0.999145, ..., 15.000004

EchoTime numeric Time between the center of the excitation pulse and the center of the echo, measured in milliseconds. Range value: 0.000101923, ..., 0.445

RepetitionTime numeric Time between successive MRI pulse sequences, measured in milliseconds. Range value: 0, ..., 12

FlipAngle numeric Angle between the net magnetization vector and the magnetic field, usually determined by the MRI pulse sequence. Range value: 9, ..., 180

CNR_WM numeric Contrast-to-noise ratio for total white matter Range value: 2.762, ..., 4.706

SNR_WM numeric Signal-to-noise ratio for total white matter Range value: 2.848, ..., 4.964

CNR_GM numeric Contrast-to-noise ratio for total gray matter Range value: 2.663, ..., 4.522

SNR_GM numeric Signal-to-noise ratio for total gray matter Range value: 2.998, ..., 4.814

Contrast_WM_GM numeric Measure of the contrast difference between white matter and gray matter in the MRI scan. Range value: 0, ..., 1.623

PHC_WMHVOL_Harmonized numeric Total volume of white matter hyperintensities (WMH) in the brain, measured in cubic centimeters harmonized across studies using standard Combat method Range value: 0.00095041499869275, ..., 137.042170176941

update_stamp POSIXct: YYYY-MM-DD HH:MM:SS