ADSP Phenotype Harmonization Consortium (PHC) - Composite PET Amyloid Detailed Scores

ADSP Phenotype Harmonization Consortium (PHC) - Composite PET Amyloid Detailed Scores data. More information is available at https://adni.loni.usc.edu/data-samples/adni-data/

Usage

data(ADSP_PHC_PET_Amyloid_Detailed)

Format

A data frame with 4756 observations and 193 variables.

Source

https://adni.loni.usc.edu/data-samples/adni-data/

Details

ORIGPROT character Original study protocol , with options: ADNI4, ADNI1, ADNIGO, ADNI2, ADNI3

COLPROT character Study protocol of data collection , with options: ADNI4, ADNI1, ADNI2, ADNIGO, ADNI3

RID numeric Participant roster ID

PTID character Participant ID

VISCODE character Visit code

VISCODE2 character Translated visit code

PHC_LONIUID numeric LONI Image ID Range value: 75199, ..., 11394049

SUBJID character Subject ID used in NIAGADS for the whole genome sequencing data files generated by the AD Sequencing Project.

PHC_SCANDATE Date of scan Date: YYYY-MM-DD

PHC_Visit numeric Harmonized Visit Number Range value: 1, ..., 20

PHC_Age_PET numeric Participant's Age at PET Data Collection Range value: 50.7186858316222, ..., 98.0752908966461

PHC_TRACER character PET Radiotracer , with options: FBB, FBP, PIB, NAV

PHC_TRACER_SUVR_WARNING character Reminder of the incomparability of SUVRs from different tracers. Refer to ADSP PET Methods Document. , with options: USE CENTILOIDS (NOT SUVRs) TO COMPARE ACROSS TRACERS

PHC_ACQUISITION_START numeric Acquisition start-stop time post-injection (minutes) Range value: 50, ..., 90

PHC_ACQUISITION_END numeric Acquisition start-stop time post-injection (minutes) Range value: 70, ..., 110

PHC_ACQUISITION_TIME numeric Acquisition start-stop time post-injection 1 = 30-60 ; 2 = 40-60 ; 3 = 40-70 ; 4 = 45-75 ; 5 = 45-90 ; 6 = 50-70 ; 7 = 60-90 ; 8 = 70-90 ; 9 = 70-110 ; 10 = 80-100 ; 11 = 90-110 ; 12 = 30-45 ; 99 = outside of acceptable range

PHC_QC_IMAGE numeric Determination if scan passed or failed quality control– image quality 0 = Fail ; 1 = Pass ; 2 = Pending

PHC_QC_TIMING numeric Determination if scan passed or failed quality control– acquisition time 0 = Fail ; 1 = Pass ; 2 = Pending

PHC_QC_NOTES numeric Notes related to failed quality control. 1 = Abnormal image ; 2 = Technical problem ; 3 = Anatomic abnormality ; 4 = Bottom of brain cut ; 5 = Early finish ; 6 = Early start ; 7 = Front of brain cut ; 8 = High glucose ; 9 = Late start ; 10 = Motion-induced attenuation artifact ; 11 = No attenuation or scatter correction ; 12 = Scan aborted early ; 13 = Severe motion ; 14 = Top of brain cut ; 15 = Wrong reconstruction parameters (16) Wrong voxel size ; 88 = Pass/Pending ; 99 = Other

PHC_COMPLIANT_ACQUISITION_START numeric Tracer compliant acquisition start time post-injection (minutes) Range value: 50, ..., 90

PHC_COMPLIANT_ACQUISITION_END numeric Tracer compliant acquisition stop time post-injection (minutes) Range value: 70, ..., 110

PHC_INJECTED_DOSE numeric Measure of injected radioligand dose (millicurie) Range value: 8.1, ..., 15

PHC_DYNAMIC numeric Whether dynamic (multiple frames available) or static 1 = static ; 2 = dynamic

GAAIN_SUMMARY_SUVR numeric GAAIN summary cortical SUVR normalized by GAAIN whole cerebellum Range value: 0.752, ..., 2.698

GAAIN_WHOLECEREBELLUM_SUVR numeric Reference region - SUVR of GAAIN whole cerebellum normalized by GAAIN whole cerebellum Range value: 1, ..., 1

GAAIN_COMPOSITE_REF_SUVR numeric Reference region - SUVR of composite ref region (volume-weighted mean of GAAIN whole cerebellum, GAAIN brainstem and NPDKA eroded WM) normalized by GAAIN whole cerebellum Range value: 1.068, ..., 1.847

GAAIN_CEREBELLUM_CORTEX_SUVR numeric Reference region - SUVR of GAAIN cerebellum grey matter normalized by GAAIN whole cerebellum Range value: 0.543, ..., 0.972

NPDKA_SUMMARY_SUVR numeric NPDKA summary cortical SUVR (volume-weighted mean of frontal, cingulate, parietal and temporal regions) normalized by NPDKA whole cerebellum; See SCAN MRI-Free Amyloid PET Methods document on LONI Range value: 0.777, ..., 2.33

NPDKA_WHOLECEREBELLUM_SUVR numeric Reference region - SUVR of NPDKA whole cerebellum normalized by NPDKA whole cerebellum Range value: 1, ..., 1

NPDKA_COMPOSITE_REF_SUVR numeric Reference region - SUVR of composite ref region (volume-weighted mean of NPDKA whole cerebellum, NPDKA brainstem and NPDKA eroded WM) normalized by NPDKA whole cerebellum Range value: 1.06, ..., 1.796

NPDKA_CEREBELLUM_CORTEX_SUVR numeric Reference region - SUVR of NPDKA cerebellum grey matter normalized by NPDKA whole cerebellum Range value: 0.801, ..., 0.976

NPDKA_ERODED_SUBCORTICALWM_GAAINWC_SUVR numeric Reference region - SUVR of NPDKA eroded subcortical white matter normalized by GAAIN whole cerebellum; See SCAN MRI-Free Amyloid PET Methods document on LONI Range value: 1.114, ..., 2.955

NPDKA_ERODED_SUBCORTICALWM_NPDKAWC_SUVR numeric Reference region - SUVR of NPDKA eroded subcortical white matter normalized by NPDKA whole cerebellum; See SCAN MRI-Free Amyloid PET Methods document on LONI Range value: 1.121, ..., 2.971

BRAINSTEM_SUVR numeric brain-stem SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 1.021, ..., 1.978

CC_ANTERIOR_SUVR numeric cc-anterior SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.6, ..., 2.243

CC_CENTRAL_SUVR numeric cc-central SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.448, ..., 2.025

CC_MID_ANTERIOR_SUVR numeric cc-mid-anterior SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.458, ..., 2.177

CC_MID_POSTERIOR_SUVR numeric cc-mid-posterior SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.39, ..., 1.893

CC_POSTERIOR_SUVR numeric cc-posterior SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.776, ..., 2.821

CSF_SUVR numeric csf SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.455, ..., 1.402

VENTRICLE_3RD_SUVR numeric 3rd-ventricle SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.467, ..., 1.334

VENTRICLE_4TH_SUVR numeric 4th-ventricle SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.584, ..., 1.357

VENTRICLE_5TH_SUVR numeric 5th-ventricle SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.345, ..., 1.711

WM_HYPOINTENSITIES_SUVR numeric wm-hypointensities SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.465, ..., 2.669

NON_WM_HYPOINTENSITIES_SUVR numeric non-wm-hypointensities SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.556, ..., 2.548

CTX_BANKSSTS_SUVR numeric Volume-weighted mean SUVR of ctx-lh-bankssts and ctx-rh-bankssts normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.883, ..., 2.777

CTX_CAUDALANTERIORCINGULATE_SUVR numeric Volume-weighted mean SUVR of ctx-lh-caudalanteriorcingulate and ctx-rh-caudalanteriorcingulate normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.654, ..., 2.755

CTX_CAUDALMIDDLEFRONTAL_SUVR numeric Volume-weighted mean SUVR of ctx-lh-caudalmiddlefrontal and ctx-rh-caudalmiddlefrontal normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.736, ..., 2.358

CTX_CUNEUS_SUVR numeric Volume-weighted mean SUVR of ctx-lh-cuneus and ctx-rh-cuneus normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.732, ..., 2.181

CTX_ENTORHINAL_SUVR numeric Volume-weighted mean SUVR of ctx-lh-entorhinal and ctx-rh-entorhinal normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.597, ..., 1.785

CTX_FRONTALPOLE_SUVR numeric Volume-weighted mean SUVR of ctx-lh-frontalpole and ctx-rh-frontalpole normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.519, ..., 2.171

CTX_FUSIFORM_SUVR numeric Volume-weighted mean SUVR of ctx-lh-fusiform and ctx-rh-fusiform normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.813, ..., 2.159

CTX_INFERIORPARIETAL_SUVR numeric Volume-weighted mean SUVR of ctx-lh-inferiorparietal and ctx-rh-inferiorparietal normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.726, ..., 2.399

CTX_INFERIORTEMPORAL_SUVR numeric Volume-weighted mean SUVR of ctx-lh-inferiortemporal and ctx-rh-inferiortemporal normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.807, ..., 2.302

CTX_INSULA_SUVR numeric Volume-weighted mean SUVR of ctx-lh-insula and ctx-rh-insula normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.743, ..., 2.218

CTX_ISTHMUSCINGULATE_SUVR numeric Volume-weighted mean SUVR of ctx-lh-isthmuscingulate and ctx-rh-isthmuscingulate normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.807, ..., 2.6

CTX_LATERALOCCIPITAL_SUVR numeric Volume-weighted mean SUVR of ctx-lh-lateraloccipital and ctx-rh-lateraloccipital normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.636, ..., 2.022

CTX_LATERALORBITOFRONTAL_SUVR numeric Volume-weighted mean SUVR of ctx-lh-lateralorbitofrontal and ctx-rh-lateralorbitofrontal normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.819, ..., 2.396

CTX_LINGUAL_SUVR numeric Volume-weighted mean SUVR of ctx-lh-lingual and ctx-rh-lingual normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.68, ..., 2.116

CTX_MEDIALORBITOFRONTAL_SUVR numeric Volume-weighted mean SUVR of ctx-lh-medialorbitofrontal and ctx-rh-medialorbitofrontal normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.706, ..., 2.736

CTX_MIDDLETEMPORAL_SUVR numeric Volume-weighted mean SUVR of ctx-lh-middletemporal and ctx-rh-middletemporal normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.647, ..., 2.341

CTX_PARACENTRAL_SUVR numeric Volume-weighted mean SUVR of ctx-lh-paracentral and ctx-rh-paracentral normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.724, ..., 2.406

CTX_PARAHIPPOCAMPAL_SUVR numeric Volume-weighted mean SUVR of ctx-lh-parahippocampal and ctx-rh-parahippocampal normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.648, ..., 1.909

CTX_PARSOPERCULARIS_SUVR numeric Volume-weighted mean SUVR of ctx-lh-parsopercularis and ctx-rh-parsopercularis normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.767, ..., 2.266

CTX_PARSORBITALIS_SUVR numeric Volume-weighted mean SUVR of ctx-lh-parsorbitalis and ctx-rh-parsorbitalis normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.692, ..., 2.27

CTX_PARSTRIANGULARIS_SUVR numeric Volume-weighted mean SUVR of ctx-lh-parstriangularis and ctx-rh-parstriangularis normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.753, ..., 2.514

CTX_PERICALCARINE_SUVR numeric Volume-weighted mean SUVR of ctx-lh-pericalcarine and ctx-rh-pericalcarine normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.851, ..., 2.389

CTX_POSTCENTRAL_SUVR numeric Volume-weighted mean SUVR of ctx-lh-postcentral and ctx-rh-postcentral normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.762, ..., 1.91

CTX_POSTERIORCINGULATE_SUVR numeric Volume-weighted mean SUVR of ctx-lh-posteriorcingulate and ctx-rh-posteriorcingulate normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.651, ..., 2.563

CTX_PRECENTRAL_SUVR numeric Volume-weighted mean SUVR of ctx-lh-precentral and ctx-rh-precentral normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.813, ..., 1.938

CTX_PRECUNEUS_SUVR numeric Volume-weighted mean SUVR of ctx-lh-precuneus and ctx-rh-precuneus normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.768, ..., 2.988

CTX_ROSTRALANTERIORCINGULATE_SUVR numeric Volume-weighted mean SUVR of ctx-lh-rostralanteriorcingulate and ctx-rh-rostralanteriorcingulate normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.682, ..., 2.876

CTX_ROSTRALMIDDLEFRONTAL_SUVR numeric Volume-weighted mean SUVR of ctx-lh-rostralmiddlefrontal and ctx-rh-rostralmiddlefrontal normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.722, ..., 2.775

CTX_SUPERIORFRONTAL_SUVR numeric Volume-weighted mean SUVR of ctx-lh-superiorfrontal and ctx-rh-superiorfrontal normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.731, ..., 2.389

CTX_SUPERIORPARIETAL_SUVR numeric Volume-weighted mean SUVR of ctx-lh-superiorparietal and ctx-rh-superiorparietal normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.742, ..., 2.482

CTX_SUPERIORTEMPORAL_SUVR numeric Volume-weighted mean SUVR of ctx-lh-superiortemporal and ctx-rh-superiortemporal normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.711, ..., 2.303

CTX_SUPRAMARGINAL_SUVR numeric Volume-weighted mean SUVR of ctx-lh-supramarginal and ctx-rh-supramarginal normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.75, ..., 2.262

CTX_TEMPORALPOLE_SUVR numeric Volume-weighted mean SUVR of ctx-lh-temporalpole and ctx-rh-temporalpole normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.558, ..., 2.077

CTX_TRANSVERSETEMPORAL_SUVR numeric Volume-weighted mean SUVR of ctx-lh-transversetemporal and ctx-rh-transversetemporal normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.619, ..., 2.328

ACCUMBENS_AREA_SUVR numeric Volume-weighted mean SUVR of left-accumbens-area and right-accumbens-area normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.62, ..., 2.635

AMYGDALA_SUVR numeric Volume-weighted mean SUVR of left-amygdala and right-amygdala normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.749, ..., 2.095

CAUDATE_SUVR numeric Volume-weighted mean SUVR of left-caudate and right-caudate normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.51, ..., 1.954

CEREBELLUM_WHITE_MATTER_SUVR numeric Volume-weighted mean SUVR of left-cerebellum-white-matter and right-cerebellum-white-matter normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 1.122, ..., 1.924

CEREBRAL_WHITE_MATTER_SUVR numeric Volume-weighted mean SUVR of left-cerebral-white-matter and right-cerebral-white-matter normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 1.051, ..., 2.393

CHOROID_PLEXUS_SUVR numeric Volume-weighted mean SUVR of left-choroid-plexus and right-choroid-plexus normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.497, ..., 1.642

HIPPOCAMPUS_SUVR numeric Volume-weighted mean SUVR of left-hippocampus and right-hippocampus normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.699, ..., 1.619

OPTIC_CHIASM_SUVR numeric optic-chiasm SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.424, ..., 1.288

INF_LAT_VENT_SUVR numeric Volume-weighted mean SUVR of left-inf-lat-vent and right-inf-lat-vent normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.623, ..., 1.996

LATERAL_VENTRICLE_SUVR numeric Volume-weighted mean SUVR of left-lateral-ventricle and right-lateral-ventricle normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.467, ..., 1.489

PALLIDUM_SUVR numeric Volume-weighted mean SUVR of left-pallidum and right-pallidum normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 1.048, ..., 2.298

PUTAMEN_SUVR numeric Volume-weighted mean SUVR of left-putamen and right-putamen normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.899, ..., 2.595

THALAMUS_PROPER_SUVR numeric Volume-weighted mean SUVR of left-thalamus-proper and right-thalamus-proper normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.784, ..., 1.878

VENTRALDC_SUVR numeric Volume-weighted mean SUVR of left-ventraldc and right-ventraldc normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.943, ..., 1.779

VESSEL_SUVR numeric Volume-weighted mean SUVR of left-vessel and right-vessel normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.652, ..., 2.377

CTX_LH_BANKSSTS_SUVR numeric ctx-lh-bankssts SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.848, ..., 2.802

CTX_LH_CAUDALANTERIORCINGULATE_SUVR numeric ctx-lh-caudalanteriorcingulate SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.661, ..., 2.806

CTX_LH_CAUDALMIDDLEFRONTAL_SUVR numeric ctx-lh-caudalmiddlefrontal SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.751, ..., 2.288

CTX_LH_CUNEUS_SUVR numeric ctx-lh-cuneus SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.711, ..., 2.185

CTX_LH_ENTORHINAL_SUVR numeric ctx-lh-entorhinal SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.558, ..., 1.898

CTX_LH_FRONTALPOLE_SUVR numeric ctx-lh-frontalpole SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.496, ..., 2.041

CTX_LH_FUSIFORM_SUVR numeric ctx-lh-fusiform SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.787, ..., 2.173

CTX_LH_INFERIORPARIETAL_SUVR numeric ctx-lh-inferiorparietal SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.744, ..., 2.418

CTX_LH_INFERIORTEMPORAL_SUVR numeric ctx-lh-inferiortemporal SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.79, ..., 2.293

CTX_LH_INSULA_SUVR numeric ctx-lh-insula SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.733, ..., 2.187

CTX_LH_ISTHMUSCINGULATE_SUVR numeric ctx-lh-isthmuscingulate SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.792, ..., 2.699

CTX_LH_LATERALOCCIPITAL_SUVR numeric ctx-lh-lateraloccipital SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.62, ..., 2.131

CTX_LH_LATERALORBITOFRONTAL_SUVR numeric ctx-lh-lateralorbitofrontal SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.823, ..., 2.394

CTX_LH_LINGUAL_SUVR numeric ctx-lh-lingual SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.675, ..., 2.05

CTX_LH_MEDIALORBITOFRONTAL_SUVR numeric ctx-lh-medialorbitofrontal SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.693, ..., 2.691

CTX_LH_MIDDLETEMPORAL_SUVR numeric ctx-lh-middletemporal SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.648, ..., 2.291

CTX_LH_PARACENTRAL_SUVR numeric ctx-lh-paracentral SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.71, ..., 2.36

CTX_LH_PARAHIPPOCAMPAL_SUVR numeric ctx-lh-parahippocampal SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.659, ..., 1.853

CTX_LH_PARSOPERCULARIS_SUVR numeric ctx-lh-parsopercularis SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.738, ..., 2.252

CTX_LH_PARSORBITALIS_SUVR numeric ctx-lh-parsorbitalis SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.662, ..., 2.348

CTX_LH_PARSTRIANGULARIS_SUVR numeric ctx-lh-parstriangularis SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.733, ..., 2.423

CTX_LH_PERICALCARINE_SUVR numeric ctx-lh-pericalcarine SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.79, ..., 2.477

CTX_LH_POSTCENTRAL_SUVR numeric ctx-lh-postcentral SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.762, ..., 1.944

CTX_LH_POSTERIORCINGULATE_SUVR numeric ctx-lh-posteriorcingulate SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.657, ..., 2.595

CTX_LH_PRECENTRAL_SUVR numeric ctx-lh-precentral SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.808, ..., 1.898

CTX_LH_PRECUNEUS_SUVR numeric ctx-lh-precuneus SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.777, ..., 2.907

CTX_LH_ROSTRALANTERIORCINGULATE_SUVR numeric ctx-lh-rostralanteriorcingulate SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.649, ..., 2.846

CTX_LH_ROSTRALMIDDLEFRONTAL_SUVR numeric ctx-lh-rostralmiddlefrontal SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.738, ..., 2.767

CTX_LH_SUPERIORFRONTAL_SUVR numeric ctx-lh-superiorfrontal SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.728, ..., 2.403

CTX_LH_SUPERIORPARIETAL_SUVR numeric ctx-lh-superiorparietal SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.718, ..., 2.357

CTX_LH_SUPERIORTEMPORAL_SUVR numeric ctx-lh-superiortemporal SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.709, ..., 2.346

CTX_LH_SUPRAMARGINAL_SUVR numeric ctx-lh-supramarginal SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.743, ..., 2.239

CTX_LH_TEMPORALPOLE_SUVR numeric ctx-lh-temporalpole SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.506, ..., 2.176

CTX_LH_TRANSVERSETEMPORAL_SUVR numeric ctx-lh-transversetemporal SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.639, ..., 2.366

CTX_RH_BANKSSTS_SUVR numeric ctx-rh-bankssts SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.831, ..., 2.858

CTX_RH_CAUDALANTERIORCINGULATE_SUVR numeric ctx-rh-caudalanteriorcingulate SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.649, ..., 2.713

CTX_RH_CAUDALMIDDLEFRONTAL_SUVR numeric ctx-rh-caudalmiddlefrontal SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.715, ..., 2.434

CTX_RH_CUNEUS_SUVR numeric ctx-rh-cuneus SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.737, ..., 2.207

CTX_RH_ENTORHINAL_SUVR numeric ctx-rh-entorhinal SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.591, ..., 1.663

CTX_RH_FRONTALPOLE_SUVR numeric ctx-rh-frontalpole SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.518, ..., 2.28

CTX_RH_FUSIFORM_SUVR numeric ctx-rh-fusiform SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.803, ..., 2.206

CTX_RH_INFERIORPARIETAL_SUVR numeric ctx-rh-inferiorparietal SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.709, ..., 2.408

CTX_RH_INFERIORTEMPORAL_SUVR numeric ctx-rh-inferiortemporal SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.776, ..., 2.31

CTX_RH_INSULA_SUVR numeric ctx-rh-insula SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.722, ..., 2.249

CTX_RH_ISTHMUSCINGULATE_SUVR numeric ctx-rh-isthmuscingulate SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.767, ..., 2.529

CTX_RH_LATERALOCCIPITAL_SUVR numeric ctx-rh-lateraloccipital SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.653, ..., 2.116

CTX_RH_LATERALORBITOFRONTAL_SUVR numeric ctx-rh-lateralorbitofrontal SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.815, ..., 2.497

CTX_RH_LINGUAL_SUVR numeric ctx-rh-lingual SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.663, ..., 2.182

CTX_RH_MEDIALORBITOFRONTAL_SUVR numeric ctx-rh-medialorbitofrontal SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.708, ..., 2.776

CTX_RH_MIDDLETEMPORAL_SUVR numeric ctx-rh-middletemporal SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.645, ..., 2.388

CTX_RH_PARACENTRAL_SUVR numeric ctx-rh-paracentral SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.698, ..., 2.461

CTX_RH_PARAHIPPOCAMPAL_SUVR numeric ctx-rh-parahippocampal SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.63, ..., 1.972

CTX_RH_PARSOPERCULARIS_SUVR numeric ctx-rh-parsopercularis SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.761, ..., 2.322

CTX_RH_PARSORBITALIS_SUVR numeric ctx-rh-parsorbitalis SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.687, ..., 2.454

CTX_RH_PARSTRIANGULARIS_SUVR numeric ctx-rh-parstriangularis SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.766, ..., 2.592

CTX_RH_PERICALCARINE_SUVR numeric ctx-rh-pericalcarine SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.818, ..., 2.627

CTX_RH_POSTCENTRAL_SUVR numeric ctx-rh-postcentral SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.664, ..., 2.005

CTX_RH_POSTERIORCINGULATE_SUVR numeric ctx-rh-posteriorcingulate SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.646, ..., 2.569

CTX_RH_PRECENTRAL_SUVR numeric ctx-rh-precentral SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.786, ..., 2.032

CTX_RH_PRECUNEUS_SUVR numeric ctx-rh-precuneus SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.727, ..., 3.073

CTX_RH_ROSTRALANTERIORCINGULATE_SUVR numeric ctx-rh-rostralanteriorcingulate SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.684, ..., 2.914

CTX_RH_ROSTRALMIDDLEFRONTAL_SUVR numeric ctx-rh-rostralmiddlefrontal SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.705, ..., 2.782

CTX_RH_SUPERIORFRONTAL_SUVR numeric ctx-rh-superiorfrontal SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.723, ..., 2.372

CTX_RH_SUPERIORPARIETAL_SUVR numeric ctx-rh-superiorparietal SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.74, ..., 2.615

CTX_RH_SUPERIORTEMPORAL_SUVR numeric ctx-rh-superiortemporal SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.714, ..., 2.257

CTX_RH_SUPRAMARGINAL_SUVR numeric ctx-rh-supramarginal SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.728, ..., 2.286

CTX_RH_TEMPORALPOLE_SUVR numeric ctx-rh-temporalpole SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.505, ..., 1.976

CTX_RH_TRANSVERSETEMPORAL_SUVR numeric ctx-rh-transversetemporal SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.578, ..., 2.301

LEFT_ACCUMBENS_AREA_SUVR numeric left-accumbens-area SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.605, ..., 2.843

LEFT_AMYGDALA_SUVR numeric left-amygdala SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.601, ..., 1.998

LEFT_CAUDATE_SUVR numeric left-caudate SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.515, ..., 1.911

LEFT_CEREBELLUM_CORTEX_SUVR numeric left-cerebellum-cortex SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.772, ..., 1.011

LEFT_CEREBELLUM_WHITE_MATTER_SUVR numeric left-cerebellum-white-matter SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 1.085, ..., 1.96

LEFT_CEREBRAL_WHITE_MATTER_SUVR numeric left-cerebral-white-matter SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 1.053, ..., 2.393

LEFT_CHOROID_PLEXUS_SUVR numeric left-choroid-plexus SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.477, ..., 1.673

LEFT_HIPPOCAMPUS_SUVR numeric left-hippocampus SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.635, ..., 1.574

LEFT_INF_LAT_VENT_SUVR numeric left-inf-lat-vent SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.618, ..., 2.029

LEFT_LATERAL_VENTRICLE_SUVR numeric left-lateral-ventricle SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.618, ..., 2.029

LEFT_PALLIDUM_SUVR numeric left-pallidum SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 1.019, ..., 2.316

LEFT_PUTAMEN_SUVR numeric left-putamen SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.872, ..., 2.546

LEFT_THALAMUS_PROPER_SUVR numeric left-thalamus-proper SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.782, ..., 1.866

LEFT_VENTRALDC_SUVR numeric left-ventraldc SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.917, ..., 1.894

LEFT_VESSEL_SUVR numeric left-vessel SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.564, ..., 2.416

RIGHT_ACCUMBENS_AREA_SUVR numeric right-accumbens-area SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.614, ..., 2.556

RIGHT_AMYGDALA_SUVR numeric right-amygdala SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.667, ..., 2.009

RIGHT_CAUDATE_SUVR numeric right-caudate SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.482, ..., 2.042

RIGHT_CEREBELLUM_CORTEX_SUVR numeric right-cerebellum-cortex SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.789, ..., 1.024

RIGHT_CEREBELLUM_WHITE_MATTER_SUVR numeric right-cerebellum-white-matter SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 1.089, ..., 1.934

RIGHT_CEREBRAL_WHITE_MATTER_SUVR numeric right-cerebral-white-matter SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 1.048, ..., 2.392

RIGHT_CHOROID_PLEXUS_SUVR numeric right-choroid-plexus SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.499, ..., 1.645

RIGHT_HIPPOCAMPUS_SUVR numeric right-hippocampus SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.683, ..., 1.662

RIGHT_INF_LAT_VENT_SUVR numeric right-inf-lat-vent SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.62, ..., 1.96

RIGHT_LATERAL_VENTRICLE_SUVR numeric right-lateral-ventricle SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.455, ..., 1.49

RIGHT_PALLIDUM_SUVR numeric right-pallidum SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 1.022, ..., 2.281

RIGHT_PUTAMEN_SUVR numeric right-putamen SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.892, ..., 2.643

RIGHT_THALAMUS_PROPER_SUVR numeric right-thalamus-proper SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.787, ..., 1.892

RIGHT_VENTRALDC_SUVR numeric right-ventraldc SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.941, ..., 1.78

RIGHT_VESSEL_SUVR numeric right-vessel SUVR normalized by whole cerebellum; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.704, ..., 2.332

update_stamp POSIXct: YYYY-MM-DD HH:MM:SS