ADSP Phenotype Harmonization Consortium (PHC) - Composite PET Amyloid Simple Scores

ADSP Phenotype Harmonization Consortium (PHC) - Composite PET Amyloid Simple Scores data. More information is available at https://adni.loni.usc.edu/data-samples/adni-data/

Usage

data(ADSP_PHC_PET_Amyloid_Simple)

Format

A data frame with 4756 observations and 26 variables.

Source

https://adni.loni.usc.edu/data-samples/adni-data/

Details

ORIGPROT character Original study protocol , with options: ADNI4, ADNI1, ADNIGO, ADNI2, ADNI3

COLPROT character Study protocol of data collection , with options: ADNI4, ADNI1, ADNI2, ADNIGO, ADNI3

RID numeric Participant roster ID

PTID character Participant ID

SUBJID character Subject ID used in NIAGADS for the whole genome sequencing data files generated by the AD Sequencing Project.

VISCODE character Visit code

VISCODE2 character Translated visit code

PHC_LONIUID numeric LONI Image ID Range value: 75199, ..., 11394049

PHC_SCANDATE Date of scan Date: YYYY-MM-DD

PHC_Visit numeric Harmonized visit number Range value: 1, ..., 20

PHC_Age_PET numeric Participant's Age at PET Data Collection Range value: 50.7186858316222, ..., 98.0752908966461

PHC_TRACER character PET Radiotracer , with options: FBB, FBP, PIB, NAV

PHC_ACQUISITION_START numeric Acquisition start-stop time post-injection (minutes) Range value: 50, ..., 90

PHC_ACQUISITION_END numeric Acquisition start-stop time post-injection (minutes) Range value: 70, ..., 110

PHC_ACQUISITION_TIME numeric Acquisition start-stop time post-injection 1 = 30-60 ; 2 = 40-60 ; 3 = 40-70 ; 4 = 45-75 ; 5 = 45-90 ; 6 = 50-70 ; 7 = 60-90 ; 8 = 70-90 ; 9 = 70-110 ; 10 = 80-100 ; 11 = 90-110 ; 12 = 30-45 ; 99 = outside of acceptable range

PHC_QC_IMAGE numeric Determination if scan passed or failed quality control– image quality 0 = Fail ; 1 = Pass ; 2 = Pending

PHC_QC_TIMING numeric Determination if scan passed or failed quality control– acquisition time 0 = Fail ; 1 = Pass ; 2 = Pending

PHC_QC_NOTES numeric Notes related to failed quality control. 1 = Abnormal image ; 2 = Technical problem ; 3 = Anatomic abnormality ; 4 = Bottom of brain cut ; 5 = Early finish ; 6 = Early start ; 7 = Front of brain cut ; 8 = High glucose ; 9 = Late start ; 10 = Motion-induced attenuation artifact ; 11 = No attenuation or scatter correction ; 12 = Scan aborted early ; 13 = Severe motion ; 14 = Top of brain cut ; 15 = Wrong reconstruction parameters (16) Wrong voxel size ; 88 = Pass/Pending ; 99 = Other

PHC_COMPLIANT_ACQUISITION_START numeric Tracer compliant acquisition start time post-injection (minutes) Range value: 50, ..., 90

PHC_COMPLIANT_ACQUISITION_END numeric Tracer compliant acquisition stop time post-injection (minutes) Range value: 70, ..., 110

PHC_INJECTED_DOSE numeric Measure of injected radioligand dose (millicurie) Range value: 8.1, ..., 15

PHC_DYNAMIC numeric Whether dynamic (multiple frames available) or static 1 = static ; 2 = dynamic

PHC_CENTILOIDS numeric Centiloids (CLs) values converted from GAAIN summary cortical SUVR normalized by GAAIN whole cerebellum; See ADSP MRI-Free PET Methods document Range value: -63, ..., 194

PHC_AMYLOID_STATUS numeric Amyloid positivity determined by applying thresholds to the GAAIN summary cortical SUVR normalized by GAAIN whole cerebellum; See ADSP MRI-Free PET Methods document 0 = negative ; 1 = positive ; 9 = unknown

PHC_CL_FAIL logical Projected centiloid values for images that failed QC_TIMING A Boolean value: TRUE/ FALSE

update_stamp POSIXct: YYYY-MM-DD HH:MM:SS