ADSP Phenotype Harmonization Consortium (PHC) - Composite PET Tau Detailed Scores

ADSP Phenotype Harmonization Consortium (PHC) - Composite PET Tau Detailed Scores data. More information is available at https://adni.loni.usc.edu/data-samples/adni-data/

Usage

data(ADSP_PHC_PET_Tau_Detailed)

Format

A data frame with 2278 observations and 192 variables.

Source

https://adni.loni.usc.edu/data-samples/adni-data/

Details

ORIGPROT character Original study protocol , with options: ADNI4, ADNI1, ADNIGO, ADNI2, ADNI3

COLPROT character Study protocol of data collection , with options: ADNI4, ADNI3, ADNI2

RID numeric Participant roster ID

PTID character Participant ID

SUBJID character Subject ID used in NIAGADS for the whole genome sequencing data files generated by the AD Sequencing Project.

VISCODE character Visit code

VISCODE2 character Translated visit code

PHC_LONIUID numeric LONI Image ID Range value: 513671, ..., 11394000

PHC_SCANDATE Date of scan Date: YYYY-MM-DD

PHC_Visit numeric Harmonized visit number Range value: 1, ..., 18

PHC_Age_PET numeric Participant's Age at PET Data Collection Range value: 50.7843942505133, ..., 96.0602327173169

PHC_TRACER character PET Radiotracer , with options: FTP, MK, PI

PHC_TRACER_SUVR_WARNING character Reminder of the incomparability of SUVRs from different tracers. Refer to ADSP PET Methods Document , with options: DO NOT COMBINE TRACERS

PHC_ACQUISITION_START numeric Acquisition start-stop time post-injection (minutes) Range value: 45, ..., 90

PHC_ACQUISITION_END numeric Acquisition start-stop time post-injection (minutes) Range value: 75, ..., 110

PHC_ACQUISITION_TIME numeric Acquisition start-stop time post-injection 1 = 30-60 ; 2 = 40-60 ; 3 = 40-70 ; 4 = 45-75 ; 5 = 45-90 ; 6 = 50-70 ; 7 = 60-90 ; 8 = 70-90 ; 9 = 70-110 ; 10 = 80-100 ; 11 = 90-110 ; 12 = 30-45 ; 99 = outside of acceptable range

PHC_QC_IMAGE numeric Determination if scan passed or failed quality control– image quality 0 = Fail ; 1 = Pass ; 2 = Pending

PHC_QC_TIMING numeric Determination if scan passed or failed quality control– acquisition time 0 = Fail ; 1 = Pass ; 2 = Pending

PHC_QC_NOTES numeric Notes related to failed quality control. 1 = Abnormal image ; 2 = Technical problem ; 3 = Anatomic abnormality ; 4 = Bottom of brain cut ; 5 = Early finish ; 6 = Early start ; 7 = Front of brain cut ; 8 = High glucose ; 9 = Late start ; 10 = Motion-induced attenuation artifact ; 11 = No attenuation or scatter correction ; 12 = Scan aborted early ; 13 = Severe motion ; 14 = Top of brain cut ; 15 = Wrong reconstruction parameters (16) Wrong voxel size ; 88 = Pass/Pending ; 99 = Other

PHC_COMPLIANT_ACQUISITION_START numeric Tracer-specific compliant acquisition start time post-injection (minutes) Range value: 45, ..., 90

PHC_COMPLIANT_ACQUISITION_END numeric Tracer-specific compliant acquisition stop time post-injection (minutes) Range value: 75, ..., 110

PHC_INJECTED_DOSE numeric Measure of injected radioligand dose (millicurie) Range value: 10, ..., 10

PHC_DYNAMIC numeric Whether dynamic (multiple frames available) or static 1 = static ; 2 = dynamic

B12_SUVR numeric SUVR of B1/B2 normalized by NPDKA inferior cerebellum Range value: 0.783, ..., 3.254

B3_SUVR numeric SUVR of B3 normalized by NPDKA inferior cerebellum Range value: 0.821, ..., 3.025

B4_SUVR numeric SUVR of B4 normalized by NPDKA inferior cerebellum Range value: 0.772, ..., 3.391

B5_SUVR numeric SUVR of B5 normalized by NPDKA inferior cerebellum Range value: 0.751, ..., 3.415

B6_SUVR numeric SUVR of B6 normalized by NPDKA inferior cerebellum Range value: 0.652, ..., 2.31

META_TEMPORAL_SUVR numeric Meta temporal SUVR normalized by inferior cerebellar grey matter. Range value: 0.837, ..., 3.641

CTX_ENTORHINAL_SUVR numeric Volume-weighted mean SUVR of ctx-lh-entorhinal and ctx-rh-entorhinal normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.783, ..., 3.254

INFERIORCEREBELLUM_SUVR numeric Reference region - SUVR of NPDKA inferior cerebellar grey matter; See SCAN MRI-Free Tau PET methods document on LONI Range value: 1, ..., 1

ERODED_SUBCORTICALWM_SUVR numeric Reference region - SUVR of NPDKA eroded subcortical white matter normalized by NPDKA inferior cerebellar grey matter; See SCAN MRI-Free Tau PET Methods document on LONI Range value: 0.453, ..., 1.886

BRAINSTEM_SUVR numeric brain-stem SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.568, ..., 1.36

CC_ANTERIOR_SUVR numeric cc-anterior SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.501, ..., 1.385

CC_CENTRAL_SUVR numeric cc-central SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.503, ..., 1.374

CC_MID_ANTERIOR_SUVR numeric cc-mid-anterior SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.493, ..., 1.389

CC_MID_POSTERIOR_SUVR numeric cc-mid-posterior SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.509, ..., 1.642

CC_POSTERIOR_SUVR numeric cc-posterior SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.598, ..., 1.563

CSF_SUVR numeric csf SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.588, ..., 1.568

VENTRICLE_3RD_SUVR numeric 3rd-ventricle SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.531, ..., 1.426

VENTRICLE_4TH_SUVR numeric 4th-ventricle SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.588, ..., 1.669

VENTRICLE_5TH_SUVR numeric 5th-ventricle SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.428, ..., 1.292

WM_HYPOINTENSITIES_SUVR numeric wm-hypointensities SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.529, ..., 1.397

NON_WM_HYPOINTENSITIES_SUVR numeric non-wm-hypointensities SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.511, ..., 1.633

CTX_BANKSSTS_SUVR numeric Volume-weighted mean SUVR of ctx-lh-bankssts and ctx-rh-bankssts normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.647, ..., 4.478

CTX_CAUDALANTERIORCINGULATE_SUVR numeric Volume-weighted mean SUVR of ctx-lh-caudalanteriorcingulate and ctx-rh-caudalanteriorcingulate normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.582, ..., 2.479

CTX_CAUDALMIDDLEFRONTAL_SUVR numeric Volume-weighted mean SUVR of ctx-lh-caudalmiddlefrontal and ctx-rh-caudalmiddlefrontal normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.599, ..., 3.88

CTX_CUNEUS_SUVR numeric Volume-weighted mean SUVR of ctx-lh-cuneus and ctx-rh-cuneus normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.713, ..., 3.546

CTX_FRONTALPOLE_SUVR numeric Volume-weighted mean SUVR of ctx-lh-frontalpole and ctx-rh-frontalpole normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.658, ..., 2.567

CTX_FUSIFORM_SUVR numeric Volume-weighted mean SUVR of ctx-lh-fusiform and ctx-rh-fusiform normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.818, ..., 3.562

CTX_INFERIORPARIETAL_SUVR numeric Volume-weighted mean SUVR of ctx-lh-inferiorparietal and ctx-rh-inferiorparietal normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.841, ..., 4.383

CTX_INFERIORTEMPORAL_SUVR numeric Volume-weighted mean SUVR of ctx-lh-inferiortemporal and ctx-rh-inferiortemporal normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.832, ..., 4.219

CTX_INSULA_SUVR numeric Volume-weighted mean SUVR of ctx-lh-insula and ctx-rh-insula normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.604, ..., 2.903

CTX_ISTHMUSCINGULATE_SUVR numeric Volume-weighted mean SUVR of ctx-lh-isthmuscingulate and ctx-rh-isthmuscingulate normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.628, ..., 3.595

CTX_LATERALOCCIPITAL_SUVR numeric Volume-weighted mean SUVR of ctx-lh-lateraloccipital and ctx-rh-lateraloccipital normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.848, ..., 3.72

CTX_LATERALORBITOFRONTAL_SUVR numeric Volume-weighted mean SUVR of ctx-lh-lateralorbitofrontal and ctx-rh-lateralorbitofrontal normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.738, ..., 3.296

CTX_LINGUAL_SUVR numeric Volume-weighted mean SUVR of ctx-lh-lingual and ctx-rh-lingual normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.828, ..., 2.888

CTX_MEDIALORBITOFRONTAL_SUVR numeric Volume-weighted mean SUVR of ctx-lh-medialorbitofrontal and ctx-rh-medialorbitofrontal normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.668, ..., 3.044

CTX_MIDDLETEMPORAL_SUVR numeric Volume-weighted mean SUVR of ctx-lh-middletemporal and ctx-rh-middletemporal normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.813, ..., 3.7

CTX_PARACENTRAL_SUVR numeric Volume-weighted mean SUVR of ctx-lh-paracentral and ctx-rh-paracentral normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.619, ..., 2.051

CTX_PARAHIPPOCAMPAL_SUVR numeric Volume-weighted mean SUVR of ctx-lh-parahippocampal and ctx-rh-parahippocampal normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.728, ..., 2.825

CTX_PARSOPERCULARIS_SUVR numeric Volume-weighted mean SUVR of ctx-lh-parsopercularis and ctx-rh-parsopercularis normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.593, ..., 3.293

CTX_PARSORBITALIS_SUVR numeric Volume-weighted mean SUVR of ctx-lh-parsorbitalis and ctx-rh-parsorbitalis normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.711, ..., 3.352

CTX_PARSTRIANGULARIS_SUVR numeric Volume-weighted mean SUVR of ctx-lh-parstriangularis and ctx-rh-parstriangularis normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.732, ..., 3.163

CTX_PERICALCARINE_SUVR numeric Volume-weighted mean SUVR of ctx-lh-pericalcarine and ctx-rh-pericalcarine normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.67, ..., 2.932

CTX_POSTCENTRAL_SUVR numeric Volume-weighted mean SUVR of ctx-lh-postcentral and ctx-rh-postcentral normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.681, ..., 2.129

CTX_POSTERIORCINGULATE_SUVR numeric Volume-weighted mean SUVR of ctx-lh-posteriorcingulate and ctx-rh-posteriorcingulate normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.624, ..., 3.482

CTX_PRECENTRAL_SUVR numeric Volume-weighted mean SUVR of ctx-lh-precentral and ctx-rh-precentral normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.623, ..., 2.101

CTX_PRECUNEUS_SUVR numeric Volume-weighted mean SUVR of ctx-lh-precuneus and ctx-rh-precuneus normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.666, ..., 4.071

CTX_ROSTRALANTERIORCINGULATE_SUVR numeric Volume-weighted mean SUVR of ctx-lh-rostralanteriorcingulate and ctx-rh-rostralanteriorcingulate normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.554, ..., 2.986

CTX_ROSTRALMIDDLEFRONTAL_SUVR numeric Volume-weighted mean SUVR of ctx-lh-rostralmiddlefrontal and ctx-rh-rostralmiddlefrontal normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.717, ..., 4.212

CTX_SUPERIORFRONTAL_SUVR numeric Volume-weighted mean SUVR of ctx-lh-superiorfrontal and ctx-rh-superiorfrontal normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.63, ..., 2.744

CTX_SUPERIORPARIETAL_SUVR numeric Volume-weighted mean SUVR of ctx-lh-superiorparietal and ctx-rh-superiorparietal normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.725, ..., 3.794

CTX_SUPERIORTEMPORAL_SUVR numeric Volume-weighted mean SUVR of ctx-lh-superiortemporal and ctx-rh-superiortemporal normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.712, ..., 2.773

CTX_SUPRAMARGINAL_SUVR numeric Volume-weighted mean SUVR of ctx-lh-supramarginal and ctx-rh-supramarginal normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.742, ..., 3.174

CTX_TEMPORALPOLE_SUVR numeric Volume-weighted mean SUVR of ctx-lh-temporalpole and ctx-rh-temporalpole normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.785, ..., 2.589

CTX_TRANSVERSETEMPORAL_SUVR numeric Volume-weighted mean SUVR of ctx-lh-transversetemporal and ctx-rh-transversetemporal normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.578, ..., 2.425

ACCUMBENS_AREA_SUVR numeric Volume-weighted mean SUVR of left-accumbens-area and right-accumbens-area normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.51, ..., 2.433

AMYGDALA_SUVR numeric Volume-weighted mean SUVR of left-amygdala and right-amygdala normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.526, ..., 3.115

CAUDATE_SUVR numeric Volume-weighted mean SUVR of left-caudate and right-caudate normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.52, ..., 1.7

CEREBELLUM_CORTEX_SUVR numeric Volume-weighted mean SUVR of left-cerebellum-cortex and right-cerebellum-cortex normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.967, ..., 1.216

CEREBELLUM_WHITE_MATTER_SUVR numeric Volume-weighted mean SUVR of left-cerebellum-white-matter and right-cerebellum-white-matter normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.634, ..., 1.481

CEREBRAL_WHITE_MATTER_SUVR numeric Volume-weighted mean SUVR of left-cerebral-white-matter and right-cerebral-white-matter normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.568, ..., 2.461

CHOROID_PLEXUS_SUVR numeric Volume-weighted mean SUVR of left-choroid-plexus and right-choroid-plexus normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.559, ..., 3.419

HIPPOCAMPUS_SUVR numeric Volume-weighted mean SUVR of left-hippocampus and right-hippocampus normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.552, ..., 2.375

OPTIC_CHIASM_SUVR numeric optic-chiasm SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.523, ..., 1.709

INF_LAT_VENT_SUVR numeric Volume-weighted mean SUVR of left-inf-lat-vent and right-inf-lat-vent normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.55, ..., 2.305

LATERAL_VENTRICLE_SUVR numeric Volume-weighted mean SUVR of left-lateral-ventricle and right-lateral-ventricle normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.553, ..., 1.398

PALLIDUM_SUVR numeric Volume-weighted mean SUVR of left-pallidum and right-pallidum normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.555, ..., 2.632

PUTAMEN_SUVR numeric Volume-weighted mean SUVR of left-putamen and right-putamen normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.566, ..., 2.456

THALAMUS_PROPER_SUVR numeric Volume-weighted mean SUVR of left-thalamus-proper and right-thalamus-proper normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.549, ..., 2.164

VENTRALDC_SUVR numeric Volume-weighted mean SUVR of left-ventraldc and right-ventraldc normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.598, ..., 1.728

VESSEL_SUVR numeric Volume-weighted mean SUVR of left-vessel and right-vessel normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.585, ..., 2.429

CTX_LH_BANKSSTS_SUVR numeric ctx-lh-bankssts SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.65, ..., 4.478

CTX_LH_CAUDALANTERIORCINGULATE_SUVR numeric ctx-lh-caudalanteriorcingulate SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.586, ..., 2.463

CTX_LH_CAUDALMIDDLEFRONTAL_SUVR numeric ctx-lh-caudalmiddlefrontal SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.596, ..., 3.851

CTX_LH_CUNEUS_SUVR numeric ctx-lh-cuneus SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.699, ..., 3.723

CTX_LH_ENTORHINAL_SUVR numeric ctx-lh-entorhinal SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.752, ..., 3.147

CTX_LH_FRONTALPOLE_SUVR numeric ctx-lh-frontalpole SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.646, ..., 2.368

CTX_LH_FUSIFORM_SUVR numeric ctx-lh-fusiform SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.805, ..., 3.631

CTX_LH_INFERIORPARIETAL_SUVR numeric ctx-lh-inferiorparietal SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.843, ..., 4.506

CTX_LH_INFERIORTEMPORAL_SUVR numeric ctx-lh-inferiortemporal SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.834, ..., 4.17

CTX_LH_INSULA_SUVR numeric ctx-lh-insula SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.599, ..., 2.85

CTX_LH_ISTHMUSCINGULATE_SUVR numeric ctx-lh-isthmuscingulate SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.654, ..., 3.678

CTX_LH_LATERALOCCIPITAL_SUVR numeric ctx-lh-lateraloccipital SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.83, ..., 3.807

CTX_LH_LATERALORBITOFRONTAL_SUVR numeric ctx-lh-lateralorbitofrontal SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.715, ..., 3.477

CTX_LH_LINGUAL_SUVR numeric ctx-lh-lingual SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.827, ..., 3.077

CTX_LH_MEDIALORBITOFRONTAL_SUVR numeric ctx-lh-medialorbitofrontal SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.654, ..., 3.031

CTX_LH_MIDDLETEMPORAL_SUVR numeric ctx-lh-middletemporal SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.827, ..., 3.769

CTX_LH_PARACENTRAL_SUVR numeric ctx-lh-paracentral SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.603, ..., 2.04

CTX_LH_PARAHIPPOCAMPAL_SUVR numeric ctx-lh-parahippocampal SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.694, ..., 2.757

CTX_LH_PARSOPERCULARIS_SUVR numeric ctx-lh-parsopercularis SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.585, ..., 3.193

CTX_LH_PARSORBITALIS_SUVR numeric ctx-lh-parsorbitalis SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.685, ..., 3.664

CTX_LH_PARSTRIANGULARIS_SUVR numeric ctx-lh-parstriangularis SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.711, ..., 3.153

CTX_LH_PERICALCARINE_SUVR numeric ctx-lh-pericalcarine SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.658, ..., 2.728

CTX_LH_POSTCENTRAL_SUVR numeric ctx-lh-postcentral SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.662, ..., 2.126

CTX_LH_POSTERIORCINGULATE_SUVR numeric ctx-lh-posteriorcingulate SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.626, ..., 3.796

CTX_LH_PRECENTRAL_SUVR numeric ctx-lh-precentral SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.61, ..., 2.226

CTX_LH_PRECUNEUS_SUVR numeric ctx-lh-precuneus SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.672, ..., 4.184

CTX_LH_ROSTRALANTERIORCINGULATE_SUVR numeric ctx-lh-rostralanteriorcingulate SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.553, ..., 3.007

CTX_LH_ROSTRALMIDDLEFRONTAL_SUVR numeric ctx-lh-rostralmiddlefrontal SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.696, ..., 4.201

CTX_LH_SUPERIORFRONTAL_SUVR numeric ctx-lh-superiorfrontal SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.63, ..., 2.85

CTX_LH_SUPERIORPARIETAL_SUVR numeric ctx-lh-superiorparietal SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.719, ..., 3.955

CTX_LH_SUPERIORTEMPORAL_SUVR numeric ctx-lh-superiortemporal SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.72, ..., 2.918

CTX_LH_SUPRAMARGINAL_SUVR numeric ctx-lh-supramarginal SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.728, ..., 3.14

CTX_LH_TEMPORALPOLE_SUVR numeric ctx-lh-temporalpole SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.759, ..., 2.838

CTX_LH_TRANSVERSETEMPORAL_SUVR numeric ctx-lh-transversetemporal SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.584, ..., 2.468

CTX_RH_BANKSSTS_SUVR numeric ctx-rh-bankssts SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.643, ..., 4.479

CTX_RH_CAUDALANTERIORCINGULATE_SUVR numeric ctx-rh-caudalanteriorcingulate SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.579, ..., 2.492

CTX_RH_CAUDALMIDDLEFRONTAL_SUVR numeric ctx-rh-caudalmiddlefrontal SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.601, ..., 3.912

CTX_RH_CUNEUS_SUVR numeric ctx-rh-cuneus SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.724, ..., 3.628

CTX_RH_ENTORHINAL_SUVR numeric ctx-rh-entorhinal SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.773, ..., 3.406

CTX_RH_FRONTALPOLE_SUVR numeric ctx-rh-frontalpole SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.633, ..., 2.724

CTX_RH_FUSIFORM_SUVR numeric ctx-rh-fusiform SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.828, ..., 3.814

CTX_RH_INFERIORPARIETAL_SUVR numeric ctx-rh-inferiorparietal SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.839, ..., 4.279

CTX_RH_INFERIORTEMPORAL_SUVR numeric ctx-rh-inferiortemporal SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.821, ..., 4.493

CTX_RH_INSULA_SUVR numeric ctx-rh-insula SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.609, ..., 2.958

CTX_RH_ISTHMUSCINGULATE_SUVR numeric ctx-rh-isthmuscingulate SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.6, ..., 3.506

CTX_RH_LATERALOCCIPITAL_SUVR numeric ctx-rh-lateraloccipital SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.864, ..., 3.906

CTX_RH_LATERALORBITOFRONTAL_SUVR numeric ctx-rh-lateralorbitofrontal SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.75, ..., 3.112

CTX_RH_LINGUAL_SUVR numeric ctx-rh-lingual SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.83, ..., 2.877

CTX_RH_MEDIALORBITOFRONTAL_SUVR numeric ctx-rh-medialorbitofrontal SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.68, ..., 3.056

CTX_RH_MIDDLETEMPORAL_SUVR numeric ctx-rh-middletemporal SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.8, ..., 3.834

CTX_RH_PARACENTRAL_SUVR numeric ctx-rh-paracentral SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.633, ..., 2.105

CTX_RH_PARAHIPPOCAMPAL_SUVR numeric ctx-rh-parahippocampal SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.683, ..., 3.008

CTX_RH_PARSOPERCULARIS_SUVR numeric ctx-rh-parsopercularis SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.602, ..., 3.417

CTX_RH_PARSORBITALIS_SUVR numeric ctx-rh-parsorbitalis SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.733, ..., 3.091

CTX_RH_PARSTRIANGULARIS_SUVR numeric ctx-rh-parstriangularis SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.75, ..., 3.172

CTX_RH_PERICALCARINE_SUVR numeric ctx-rh-pericalcarine SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.683, ..., 3.211

CTX_RH_POSTCENTRAL_SUVR numeric ctx-rh-postcentral SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.701, ..., 2.131

CTX_RH_POSTERIORCINGULATE_SUVR numeric ctx-rh-posteriorcingulate SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.621, ..., 3.176

CTX_RH_PRECENTRAL_SUVR numeric ctx-rh-precentral SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.635, ..., 2.179

CTX_RH_PRECUNEUS_SUVR numeric ctx-rh-precuneus SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.66, ..., 3.952

CTX_RH_ROSTRALANTERIORCINGULATE_SUVR numeric ctx-rh-rostralanteriorcingulate SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.554, ..., 2.961

CTX_RH_ROSTRALMIDDLEFRONTAL_SUVR numeric ctx-rh-rostralmiddlefrontal SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.696, ..., 4.223

CTX_RH_SUPERIORFRONTAL_SUVR numeric ctx-rh-superiorfrontal SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.631, ..., 2.632

CTX_RH_SUPERIORPARIETAL_SUVR numeric ctx-rh-superiorparietal SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.731, ..., 3.623

CTX_RH_SUPERIORTEMPORAL_SUVR numeric ctx-rh-superiortemporal SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.704, ..., 2.622

CTX_RH_SUPRAMARGINAL_SUVR numeric ctx-rh-supramarginal SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.757, ..., 3.248

CTX_RH_TEMPORALPOLE_SUVR numeric ctx-rh-temporalpole SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.761, ..., 2.827

CTX_RH_TRANSVERSETEMPORAL_SUVR numeric ctx-rh-transversetemporal SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.568, ..., 2.364

LEFT_ACCUMBENS_AREA_SUVR numeric left-accumbens-area SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.512, ..., 2.595

LEFT_AMYGDALA_SUVR numeric left-amygdala SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.525, ..., 3.073

LEFT_CAUDATE_SUVR numeric left-caudate SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.538, ..., 1.678

LEFT_CEREBELLUM_CORTEX_SUVR numeric left-cerebellum-cortex SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.952, ..., 1.225

LEFT_CEREBELLUM_WHITE_MATTER_SUVR numeric left-cerebellum-white-matter SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.647, ..., 1.511

LEFT_CEREBRAL_WHITE_MATTER_SUVR numeric left-cerebral-white-matter SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.567, ..., 2.58

LEFT_CHOROID_PLEXUS_SUVR numeric left-choroid-plexus SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.56, ..., 3.631

LEFT_HIPPOCAMPUS_SUVR numeric left-hippocampus SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.561, ..., 2.405

LEFT_INF_LAT_VENT_SUVR numeric left-inf-lat-vent SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.569, ..., 2.281

LEFT_LATERAL_VENTRICLE_SUVR numeric left-lateral-ventricle SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.563, ..., 1.462

LEFT_PALLIDUM_SUVR numeric left-pallidum SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.553, ..., 2.689

LEFT_PUTAMEN_SUVR numeric left-putamen SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.571, ..., 2.434

LEFT_THALAMUS_PROPER_SUVR numeric left-thalamus-proper SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.55, ..., 2.212

LEFT_VENTRALDC_SUVR numeric left-ventraldc SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.598, ..., 2.092

LEFT_VESSEL_SUVR numeric left-vessel SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.523, ..., 2.53

RIGHT_ACCUMBENS_AREA_SUVR numeric right-accumbens-area SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.509, ..., 2.345

RIGHT_AMYGDALA_SUVR numeric right-amygdala SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.526, ..., 3.17

RIGHT_CAUDATE_SUVR numeric right-caudate SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.502, ..., 1.722

RIGHT_CEREBELLUM_CORTEX_SUVR numeric right-cerebellum-cortex SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.957, ..., 1.208

RIGHT_CEREBELLUM_WHITE_MATTER_SUVR numeric right-cerebellum-white-matter SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.621, ..., 1.496

RIGHT_CEREBRAL_WHITE_MATTER_SUVR numeric right-cerebral-white-matter SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.568, ..., 2.341

RIGHT_CHOROID_PLEXUS_SUVR numeric right-choroid-plexus SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.558, ..., 3.225

RIGHT_HIPPOCAMPUS_SUVR numeric right-hippocampus SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.543, ..., 2.346

RIGHT_INF_LAT_VENT_SUVR numeric right-inf-lat-vent SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.53, ..., 2.375

RIGHT_LATERAL_VENTRICLE_SUVR numeric right-lateral-ventricle SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.543, ..., 1.503

RIGHT_PALLIDUM_SUVR numeric right-pallidum SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.555, ..., 2.573

RIGHT_PUTAMEN_SUVR numeric right-putamen SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.562, ..., 2.598

RIGHT_THALAMUS_PROPER_SUVR numeric right-thalamus-proper SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.548, ..., 2.115

RIGHT_VENTRALDC_SUVR numeric right-ventraldc SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.598, ..., 1.737

RIGHT_VESSEL_SUVR numeric right-vessel SUVR normalized by inferior cerebellar grey matter; ROI volume is provided in MRI-Free NPDKA Appendix CSV on LONI Range value: 0.54, ..., 2.408

update_stamp POSIXct: YYYY-MM-DD HH:MM:SS