ADSP Phenotype Harmonization Consortium (PHC) - Composite Biomarker Plasma Scores

ADSP Phenotype Harmonization Consortium (PHC) - Composite Biomarker Plasma Scores data. More information is available at https://adni.loni.usc.edu/data-samples/adni-data/

Usage

data(ADSP_PHC_PLASMA)

Format

A data frame with 1720 observations and 23 variables.

Source

https://adni.loni.usc.edu/data-samples/adni-data/

Details

ORIGPROT character Original study protocol , with options: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4

COLPROT character Study protocol of data collection , with options: ADNI1, ADNI2, ADNI4, ADNI3, ADNIGO

RID numeric Participant roster ID

PTID character Participant ID

SUBJID character Subject ID used in NIAGADS for the whole genome sequencing data files generated by the AD Sequencing Project.

VISCODE character Visit code

VISCODE2 character Translated visit code

DRAWDATE Biomarker Draw Date Date: YYYY-MM-DD

PHC_Visit numeric Harmonized visit number Range value: 1, ..., 21

PHC_Age_Biomarker numeric Age at each biomarker visit Range value: 50.64, ..., 98.08

Tissue character Name of tissue extracted for biomarker measurement , with options: Plasma

Platform character Immunoassay platform used for biomarker measurement , with options: Fujirebio

AB40_RAW numeric Raw AB40 biomarker levels Range value: 188.78, ..., 492.01

PHC_AB40 numeric Harmonized Z-Score for AB40 biomarker levels Range value: -2.68821011257117, ..., 2.65305315296266

AB42_RAW numeric Raw AB42 biomarker levels Range value: 14.22, ..., 46.03

PHC_AB42 numeric Harmonized Z-Score for AB42 biomarker levels Range value: -2.90102651055181, ..., 2.81033524295847

Tau_RAW logical Raw Tau biomarker levels A Boolean value: TRUE/ FALSE

PHC_Tau logical Harmonized Z-Score for Tau biomarker levels A Boolean value: TRUE/ FALSE

pTau181_RAW logical Raw pTau181 biomarker levels A Boolean value: TRUE/ FALSE

PHC_pTau181 logical Harmonized Z-Score for pTau181 biomarker levels A Boolean value: TRUE/ FALSE

pTau217_RAW numeric Raw pTau217 biomarker levels Range value: 0.03, ..., 3.207

PHC_pTau217 numeric Harmonized Z-Score for pTau217 biomarker levels Range value: -2.12026575427977, ..., 2.99330563467903

update_stamp POSIXct: YYYY-MM-DD HH:MM:SS