ADSP Phenotype Harmonization Consortium (PHC) - Composite T1 FreeSurfer Scores

ADSP Phenotype Harmonization Consortium (PHC) - Composite T1 FreeSurfer Scores data. More information is available at https://adni.loni.usc.edu/data-samples/adni-data/

Usage

data(ADSP_PHC_T1_FS)

Format

A data frame with 10526 observations and 412 variables.

Source

https://adni.loni.usc.edu/data-samples/adni-data/

Details

ORIGPROT character Original study protocol , with options: ADNI1, ADNI4, ADNIGO, ADNI2, ADNI3

COLPROT character Study protocol of data collection , with options: ADNI1, ADNI4, ADNI2, ADNIGO, ADNI3

RID numeric Participant roster ID

PTID character Participant ID

VISCODE character Visit code

VISCODE2 character Translated visit code

PHC_LONIUID numeric LONI Image ID Range value: 25898, ..., 11433156

SUBJID character Subject ID used in NIAGADS for the whole genome sequencing data files generated by the AD Sequencing Project.

PHC_SCANDATE Date of scan Date: YYYY-MM-DD

PHC_Visit numeric Harmonized Visit Number Range value: 1, ..., 18

PHC_Age_T1 numeric Age at each T1 visit Range value: 50.53251198, ..., 97.45106092

ScannerBatchName character MRI scanner model/batch, usually provided by the manufacturer.

ScannerManufacturer character Scanner Manufacturer , with options: SIEMENS, Siemens Healthineers, Philips, GE MEDICAL SYSTEMS, Philips Medical Systems, SIEMENS|PixelMed, Philips Healthcare, Siemens

FieldSgth numeric Strength of the magnetic field used in the MRI scanner, measured in Tesla. Range value: 1.5, ..., 3

EstimatedTotalIntraCranialVol_combat numeric Total intracranial volume harmonized with combat Range value: 833279.0411, ..., 2162780.553

Left_Lateral_Ventricle_combat numeric Left lateral ventricle volume harmonized with combat Range value: 1904.260411, ..., 83762.17127

Left_Inf_Lat_Vent_combat numeric Left inferior lateral ventricle volume harmonized with combat Range value: 8.473863954, ..., 8366.112655

Left_Cerebellum_White_Matter_combat numeric Left cerebellar white matter volume harmonized with combat Range value: 7087.704657, ..., 28525.17401

Left_Cerebellum_Cortex_combat numeric Left cerebellar cortex volume harmonized with combat Range value: 30852.76834, ..., 72326.4463

Left_Thalamus_Proper_combat numeric Left thalamus volume harmonized with combat Range value: 3516.693269, ..., 10060.57603

Left_Caudate_combat numeric Left caudate volume harmonized with combat Range value: 1772.707417, ..., 6586.713587

Left_Putamen_combat numeric Left putamen volume harmonized with combat Range value: 1685.78718, ..., 8557.177616

Left_Pallidum_combat numeric Left pallidum volume harmonized with combat Range value: 831.6289527, ..., 3099.874204

X3rd_Ventricle_combat numeric 3rd ventricle volume harmonized with combat Range value: 420.2795864, ..., 4997.988348

X4th_Ventricle_combat numeric 4th ventricle volume harmonized with combat Range value: 637.4369802, ..., 5329.039527

Brain_Stem_combat numeric Brainstem volume harmonized with combat Range value: 11649.15022, ..., 34277.47824

Left_Hippocampus_combat numeric Left hippocampus volume harmonized with combat Range value: 1405.408689, ..., 5650.492077

Left_Amygdala_combat numeric Left amygdala volume harmonized with combat Range value: 292.7719036, ..., 2377.730057

CSF_combat numeric CSF volume harmonized with combat Range value: 505.0992808, ..., 4651.47172

Left_Accumbens_area_combat numeric Left accumbens volume harmonized with combat Range value: 52.76407245, ..., 841.0499421

Left_VentralDC_combat numeric Left ventral DC volume harmonized with combat Range value: 2251.161089, ..., 6641.244372

Left_choroid_plexus_combat numeric Left choroid plexus volume harmonized with combat Range value: 104.6329027, ..., 2141.18817

Right_Lateral_Ventricle_combat numeric Right lateral ventricle volume harmonized with combat Range value: 1324.741145, ..., 69461.4782

Right_Inf_Lat_Vent_combat numeric Right inferior lateral ventricle volume harmonized with combat Range value: 3.822563576, ..., 7157.127827

Right_Cerebellum_White_Matter_combat numeric Right cerebellar white matter volume harmonized with combat Range value: 6094.875152, ..., 29649.00314

Right_Cerebellum_Cortex_combat numeric Right cerebellar cortex volume harmonized with combat Range value: 28460.96692, ..., 73383.03386

Right_Thalamus_Proper_combat numeric Right thalamus volume harmonized with combat Range value: 3747.468108, ..., 9703.199399

Right_Caudate_combat numeric Right caudate volume harmonized with combat Range value: 1752.286937, ..., 6391.438268

Right_Putamen_combat numeric Right putamen volume harmonized with combat Range value: 1713.091987, ..., 8122.636963

Right_Pallidum_combat numeric Right pallidum volume harmonized with combat Range value: 870.0396595, ..., 2924.842249

Right_Hippocampus_combat numeric Right hippocampus volume harmonized with combat Range value: 1546.429228, ..., 5765.922114

Right_Amygdala_combat numeric Right amygdala volume harmonized with combat Range value: 435.0880262, ..., 2634.302049

Right_Accumbens_area_combat numeric Right accumbens volume harmonized with combat Range value: 38.55465584, ..., 857.9493923

Right_VentralDC_combat numeric Right ventral DC volume harmonized with combat Range value: 2254.810565, ..., 5795.028549

Right_choroid_plexus_combat numeric Right choroid plexus volume harmonized with combat Range value: 135.3854019, ..., 2166.42333

WM_hypointensities_combat numeric White matter hypointensities harmonized with combat Range value: 188.1529921, ..., 69664.42894

Optic_Chiasm_combat numeric Optic Chiasm volume harmonized with combat Range value: 17.43288798, ..., 442.0566069

CC_Posterior_combat numeric Posterior corpus collosum volume harmonized with combat Range value: 69.9861955, ..., 2099.296865

CC_Mid_Posterior_combat numeric Mid-posterior corpus collosum volume harmonized with combat Range value: 131.5909596, ..., 1417.105016

CC_Central_combat numeric Central corpus collosum volume harmonized with combat Range value: 90.53394045, ..., 1052.81641

CC_Mid_Anterior_combat numeric Mid-anterior corpus collosum volume harmonized with combat Range value: 90.34588784, ..., 1209.303152

CC_Anterior_combat numeric Anterior corpus collosum volume harmonized with combat Range value: 50.72910099, ..., 1581.423965

BrainSegVol_combat numeric Segmented brain volume harmonized with combat Range value: 702755.4073, ..., 1495064.113

BrainSegVolNotVent_combat numeric Segmented brain volume excluding ventricles harmonized with combat Range value: 679039.118, ..., 1465639.108

BrainSegVolNotVentSurf_combat numeric Surface size of the segmented brain volume excluding ventricles harmonized with combat Range value: 678451.9285, ..., 1464915.496

lhCortexVol_combat numeric Left cerebral cortex grey matter volume harmonized with combat Range value: 125280.3632, ..., 287094.5877

rhCortexVol_combat numeric Right cerebral cortex grey matter volume harmonized with combat Range value: 130125.6728, ..., 291360.9749

CortexVol_combat numeric Total cerebral cortex grey matter volume harmonized with combat Range value: 263308.928, ..., 576468.4842

lhCerebralWhiteMatterVol_combat numeric Left cerebral cortex white matter volume harmonized with combat Range value: 117361.6158, ..., 338232.59

rhCerebralWhiteMatterVol_combat numeric Right cerebral cortex white matter volume harmonized with combat Range value: 121955.7706, ..., 334929.1234

CerebralWhiteMatterVol_combat numeric Total cerebral cortex white matter volume harmonized with combat Range value: 239818.6239, ..., 673256.8337

SubCortGrayVol_combat numeric Total volume of the subcortical grey matter harmonized with combat Range value: 31237.97075, ..., 70213.5354

TotalGrayVol_combat numeric Total volume of all grey matter (cortical and subcortical) harmonized with combat Range value: 365872.8744, ..., 768892.4656

SupraTentorialVol_combat numeric Total volume of brain (minus the cerebellum) harmonized with combat Range value: 597666.4584, ..., 1335871.251

SupraTentorialVolNotVent_combat numeric Total volume of brain without ventricles (minus the cerebellum) harmonized with combat Range value: 578151.4375, ..., 1302698.373

SupraTentorialVolNotVentVox_combat numeric Number of voxels of brain without ventricles (minus the cerebellum) harmonized with combat Range value: 576052.9642, ..., 1299653.135

MaskVol_combat numeric Mask volume harmonized with combat Range value: 1044928.165, ..., 2161527.323

BrainSegVol_to_eTIV_combat numeric Ratio of total brain segmentation volume to ICV harmonized with combat Range value: 0.475517577, ..., 1.04343535

MaskVol_to_eTIV_combat numeric Ratio of mask segmentation volume to ICV harmonized with combat Range value: 0.680746293, ..., 1.394867376

lh_bankssts_thickness_combat numeric Left banks of the superior temporal sulcus thickness harmonized with combat Range value: 1.557046503, ..., 2.989331691

lh_caudalanteriorcingulate_thickness_combat numeric Left caudal anterior cingulate thickness harmonized with combat Range value: 1.398604011, ..., 3.79833567

lh_caudalmiddlefrontal_thickness_combat numeric Left caudal middle frontal gyri thickness harmonized with combat Range value: 1.59983754, ..., 2.963382264

lh_cuneus_thickness_combat numeric Left cuneus thickness harmonized with combat Range value: 1.30216346, ..., 2.50384103

lh_entorhinal_thickness_combat numeric Left entorhinal cortex thickness harmonized with combat Range value: 1.189832184, ..., 4.440961634

lh_fusiform_thickness_combat numeric Left fusiform gyri thickness harmonized with combat Range value: 1.738162335, ..., 3.276701567

lh_inferiorparietal_thickness_combat numeric Left inferior parietal gyri thickness harmonized with combat Range value: 1.558648505, ..., 2.884142245

lh_inferiortemporal_thickness_combat numeric Left inferior temporal gyri thickness harmonized with combat Range value: 1.830317902, ..., 3.41269075

lh_isthmuscingulate_thickness_combat numeric Left isthmus cingulate thickness harmonized with combat Range value: 1.509686853, ..., 3.097287301

lh_lateraloccipital_thickness_combat numeric Left lateral occipital gyri thickness harmonized with combat Range value: 1.307546746, ..., 2.635921008

lh_lateralorbitofrontal_thickness_combat numeric Left lateral orbitofrontal gyri thickness harmonized with combat Range value: 1.865510934, ..., 3.159727865

lh_lingual_thickness_combat numeric Left lingual gyri thickness harmonized with combat Range value: 1.369628283, ..., 2.559939437

lh_medialorbitofrontal_thickness_combat numeric Left medial orbitofrontal thickness harmonized with combat Range value: 1.687760224, ..., 3.159683724

lh_middletemporal_thickness_combat numeric Left middle temporal gyri thickness harmonized with combat Range value: 1.691477899, ..., 3.209231739

lh_parahippocampal_thickness_combat numeric Left parahippocampal gyri thickness harmonized with combat Range value: 1.418002512, ..., 3.87237816

lh_paracentral_thickness_combat numeric Left paracentral lobule thickness harmonized with combat Range value: 1.677587399, ..., 2.937276775

lh_parsopercularis_thickness_combat numeric Left pars opercularis thickness harmonized with combat Range value: 1.77180672, ..., 2.960496417

lh_parsorbitalis_thickness_combat numeric Left pars orbitalis thickness harmonized with combat Range value: 1.729945786, ..., 3.518990776

lh_parstriangularis_thickness_combat numeric Left pars triangularis thickness harmonized with combat Range value: 1.676687374, ..., 2.903305044

lh_pericalcarine_thickness_combat numeric Left pericalcarine gyri thickness harmonized with combat Range value: 1.145041673, ..., 2.425175599

lh_postcentral_thickness_combat numeric Left postcentral gyri thickness harmonized with combat Range value: 1.419877003, ..., 2.441478784

lh_posteriorcingulate_thickness_combat numeric Left posterior cingulate thickness harmonized with combat Range value: 1.627199539, ..., 3.065150217

lh_precentral_thickness_combat numeric Left precentral gyri thickness harmonized with combat Range value: 1.618749596, ..., 2.978417072

lh_precuneus_thickness_combat numeric Left precuneus thickness harmonized with combat Range value: 1.495567529, ..., 2.783370705

lh_rostralanteriorcingulate_thickness_combat numeric Left rostral anterior cingulate thickness harmonized with combat Range value: 1.757210423, ..., 3.815038929

lh_rostralmiddlefrontal_thickness_combat numeric Left rostral middle frontal gyri thickness harmonized with combat Range value: 1.619787611, ..., 2.850781282

lh_superiorfrontal_thickness_combat numeric Left superior frontal gyri thickness harmonized with combat Range value: 1.712287219, ..., 3.087436621

lh_superiorparietal_thickness_combat numeric Left superior parietal gyri thickness harmonized with combat Range value: 1.384307853, ..., 2.774102843

lh_superiortemporal_thickness_combat numeric Left superior temporal gyri thickness harmonized with combat Range value: 1.746565432, ..., 3.170250338

lh_supramarginal_thickness_combat numeric Left supramarginal gyri thickness harmonized with combat Range value: 1.575440318, ..., 3.056615134

lh_frontalpole_thickness_combat numeric Left frontal pole thickness harmonized with combat Range value: 1.564075648, ..., 3.841295283

lh_temporalpole_thickness_combat numeric Left temporal pole thickness harmonized with combat Range value: 1.776636706, ..., 4.539989631

lh_transversetemporal_thickness_combat numeric Left transverse temporal pole thickness harmonized with combat Range value: 1.183116193, ..., 3.020364274

lh_insula_thickness_combat numeric Left insula thickness harmonized with combat Range value: 2.050275239, ..., 3.481829981

lh_MeanThickness_thickness_combat numeric Left mean lobar thickness harmonized with combat Range value: 1.84833659, ..., 2.687148137

rh_bankssts_thickness_combat numeric Right banks of the superior temporal sulcus thickness harmonized with combat Range value: 1.589991537, ..., 3.082328115

rh_caudalanteriorcingulate_thickness_combat numeric Right caudal anterior cingulate thickness harmonized with combat Range value: 1.241409892, ..., 3.61132044

rh_caudalmiddlefrontal_thickness_combat numeric Right caudal middle frontal gyri thickness harmonized with combat Range value: 1.670730951, ..., 2.952829199

rh_cuneus_thickness_combat numeric Right cuneus thickness harmonized with combat Range value: 1.342695331, ..., 2.471749216

rh_entorhinal_thickness_combat numeric Right entorhinal cortex thickness harmonized with combat Range value: 1.294281603, ..., 4.698978147

rh_fusiform_thickness_combat numeric Right fusiform gyri thickness harmonized with combat Range value: 1.788338068, ..., 3.242436805

rh_inferiorparietal_thickness_combat numeric Right inferior parietal gyri thickness harmonized with combat Range value: 1.544872172, ..., 2.77253457

rh_inferiortemporal_thickness_combat numeric Right inferior temporal gyri thickness harmonized with combat Range value: 1.777239451, ..., 3.380365134

rh_isthmuscingulate_thickness_combat numeric Right isthmus cingulate thickness harmonized with combat Range value: 1.464829809, ..., 3.220089559

rh_lateraloccipital_thickness_combat numeric Right lateral occipital gyri thickness harmonized with combat Range value: 1.437018917, ..., 2.747453078

rh_lateralorbitofrontal_thickness_combat numeric Right lateral orbitofrontal gyri thickness harmonized with combat Range value: 1.866371064, ..., 3.285677371

rh_lingual_thickness_combat numeric Right lingual gyri thickness harmonized with combat Range value: 1.501084637, ..., 2.494190669

rh_medialorbitofrontal_thickness_combat numeric Right medial orbitofrontal thickness harmonized with combat Range value: 1.653785651, ..., 3.236446548

rh_middletemporal_thickness_combat numeric Right middle temporal gyri thickness harmonized with combat Range value: 1.874226313, ..., 3.240649124

rh_parahippocampal_thickness_combat numeric Right parahippocampal gyri thickness harmonized with combat Range value: 1.514890788, ..., 3.58300546

rh_paracentral_thickness_combat numeric Right paracentral lobule thickness harmonized with combat Range value: 1.619006415, ..., 3.092381664

rh_parsopercularis_thickness_combat numeric Right pars opercularis thickness harmonized with combat Range value: 1.619197987, ..., 2.897015317

rh_parsorbitalis_thickness_combat numeric Right pars orbitalis thickness harmonized with combat Range value: 1.760493937, ..., 3.354198359

rh_parstriangularis_thickness_combat numeric Right pars triangularis thickness harmonized with combat Range value: 1.70823596, ..., 2.936628469

rh_pericalcarine_thickness_combat numeric Right pericalcarine gyri thickness harmonized with combat Range value: 1.192240536, ..., 2.527883345

rh_postcentral_thickness_combat numeric Right postcentral gyri thickness harmonized with combat Range value: 1.459315966, ..., 2.545650866

rh_posteriorcingulate_thickness_combat numeric Right posterior cingulate thickness harmonized with combat Range value: 1.458540939, ..., 3.117970356

rh_precentral_thickness_combat numeric Right precentral gyri thickness harmonized with combat Range value: 1.517828489, ..., 2.894737801

rh_precuneus_thickness_combat numeric Right precuneus thickness harmonized with combat Range value: 1.518698281, ..., 2.735206663

rh_rostralanteriorcingulate_thickness_combat numeric Right rostral anterior cingulate thickness harmonized with combat Range value: 1.877903777, ..., 4.049077466

rh_rostralmiddlefrontal_thickness_combat numeric Right rostral middle frontal gyri thickness harmonized with combat Range value: 1.654201722, ..., 2.78859322

rh_superiorfrontal_thickness_combat numeric Right superior frontal gyri thickness harmonized with combat Range value: 1.80536102, ..., 3.085045498

rh_superiorparietal_thickness_combat numeric Right superior parietal gyri thickness harmonized with combat Range value: 1.312916138, ..., 2.807267968

rh_superiortemporal_thickness_combat numeric Right superior temporal gyri thickness harmonized with combat Range value: 1.80635558, ..., 3.169270747

rh_supramarginal_thickness_combat numeric Right supramarginal gyri thickness harmonized with combat Range value: 1.682600851, ..., 2.892363024

rh_frontalpole_thickness_combat numeric Right frontal pole thickness harmonized with combat Range value: 1.704547874, ..., 3.782706102

rh_temporalpole_thickness_combat numeric Right temporal pole thickness harmonized with combat Range value: 1.722079884, ..., 4.546333001

rh_transversetemporal_thickness_combat numeric Right transverse temporal pole thickness harmonized with combat Range value: 1.31983873, ..., 3.115342004

rh_insula_thickness_combat numeric Right insula thickness harmonized with combat Range value: 2.025438897, ..., 3.526824365

rh_MeanThickness_thickness_combat numeric Right mean lobar thickness harmonized with combat Range value: 1.849212272, ..., 2.734159325

lh_bankssts_volume_combat numeric Left banks of the superior temporal sulcus grey matter volume harmonized with combat Range value: 684.0426251, ..., 5133.839598

lh_caudalanteriorcingulate_volume_combat numeric Left caudal anterior cingulate grey matter volume harmonized with combat Range value: 37.12619328, ..., 3667.888136

lh_caudalmiddlefrontal_volume_combat numeric Left caudal middle frontal gyri grey matter volume harmonized with combat Range value: 1910.089596, ..., 9642.840659

lh_cuneus_volume_combat numeric Left cuneus grey matter volume harmonized with combat Range value: 1094.387927, ..., 5230.206076

lh_entorhinal_volume_combat numeric Left entorhinal cortex grey matter volume harmonized with combat Range value: 325.2580379, ..., 4729.274162

lh_fusiform_volume_combat numeric Left fusiform gyri grey matter volume harmonized with combat Range value: 3767.543994, ..., 13946.77316

lh_inferiorparietal_volume_combat numeric Left inferior parietal gyri grey matter volume harmonized with combat Range value: 4275.935276, ..., 17330.23298

lh_inferiortemporal_volume_combat numeric Left inferior temporal gyri grey matter volume harmonized with combat Range value: 3606.784424, ..., 16611.48667

lh_isthmuscingulate_volume_combat numeric Left isthmus cingulate grey matter volume harmonized with combat Range value: 793.8067468, ..., 4610.543822

lh_lateraloccipital_volume_combat numeric Left lateral occipital gyri grey matter volume harmonized with combat Range value: 3874.699179, ..., 18453.65446

lh_lateralorbitofrontal_volume_combat numeric Left lateral orbitofrontal gyri grey matter volume harmonized with combat Range value: 3487.928699, ..., 10497.94254

lh_lingual_volume_combat numeric Left lingual gyri grey matter volume harmonized with combat Range value: 2502.444299, ..., 10880.80573

lh_medialorbitofrontal_volume_combat numeric Left medial orbitofrontal grey matter volume harmonized with combat Range value: 2524.439211, ..., 7659.395629

lh_middletemporal_volume_combat numeric Left middle temporal gyri grey matter volume harmonized with combat Range value: 3757.415747, ..., 15947.16965

lh_parahippocampal_volume_combat numeric Left parahippocampal gyri grey matter volume harmonized with combat Range value: 810.7786724, ..., 3434.195779

lh_paracentral_volume_combat numeric Left paracentral lobule grey matter volume harmonized with combat Range value: 1510.562753, ..., 5951.771673

lh_parsopercularis_volume_combat numeric Left pars opercularis grey matter volume harmonized with combat Range value: 1808.593661, ..., 9691.107149

lh_parsorbitalis_volume_combat numeric Left pars orbitalis grey matter volume harmonized with combat Range value: 633.8574081, ..., 3890.014339

lh_parstriangularis_volume_combat numeric Left pars triangularis grey matter volume harmonized with combat Range value: 1348.188295, ..., 6271.159132

lh_pericalcarine_volume_combat numeric Left pericalcarine gyri grey matter volume harmonized with combat Range value: 742.4168723, ..., 5206.02104

lh_postcentral_volume_combat numeric Left postcentral gyri grey matter volume harmonized with combat Range value: 3769.733189, ..., 14353.17957

lh_posteriorcingulate_volume_combat numeric Left posterior cingulate grey matter volume harmonized with combat Range value: 397.4652731, ..., 4650.510028

lh_precentral_volume_combat numeric Left precentral gyri grey matter volume harmonized with combat Range value: 5897.134382, ..., 19561.23895

lh_precuneus_volume_combat numeric Left precuneus grey matter volume harmonized with combat Range value: 3515.089805, ..., 13082.19028

lh_rostralanteriorcingulate_volume_combat numeric Left rostral anterior cingulate grey matter volume harmonized with combat Range value: 773.6357655, ..., 4888.482987

lh_rostralmiddlefrontal_volume_combat numeric Left rostral middle frontal gyri grey matter volume harmonized with combat Range value: 5867.067797, ..., 22585.51395

lh_superiorfrontal_volume_combat numeric Left superior frontal gyri grey matter volume harmonized with combat Range value: 9209.879819, ..., 32413.86031

lh_superiorparietal_volume_combat numeric Left superior parietal gyri grey matter volume harmonized with combat Range value: 4593.048136, ..., 19678.45425

lh_superiortemporal_volume_combat numeric Left superior temporal gyri grey matter volume harmonized with combat Range value: 4122.24055, ..., 17797.11191

lh_supramarginal_volume_combat numeric Left supramarginal gyri grey matter volume harmonized with combat Range value: 4655.235153, ..., 17808.073

lh_frontalpole_volume_combat numeric Left frontal pole grey matter volume harmonized with combat Range value: 380.4947566, ..., 1843.039273

lh_temporalpole_volume_combat numeric Left temporal pole grey matter volume harmonized with combat Range value: 493.3348025, ..., 4245.358373

lh_transversetemporal_volume_combat numeric Left transverse temporal pole grey matter volume harmonized with combat Range value: 283.0775006, ..., 2355.349988

lh_insula_volume_combat numeric Left insula grey matter volume harmonized with combat Range value: 3510.41113, ..., 10067.7353

rh_bankssts_volume_combat numeric Right banks of the superior temporal sulcus grey matter volume harmonized with combat Range value: 719.571058, ..., 3689.037496

rh_caudalanteriorcingulate_volume_combat numeric Right caudal anterior cingulate grey matter volume harmonized with combat Range value: -135.3737814, ..., 3829.812414

rh_caudalmiddlefrontal_volume_combat numeric Right caudal middle frontal gyri grey matter volume harmonized with combat Range value: 1793.804535, ..., 10186.23469

rh_cuneus_volume_combat numeric Right cuneus grey matter volume harmonized with combat Range value: 1105.420634, ..., 6616.865043

rh_entorhinal_volume_combat numeric Right entorhinal cortex grey matter volume harmonized with combat Range value: 382.1271583, ..., 4084.794368

rh_fusiform_volume_combat numeric Right fusiform gyri grey matter volume harmonized with combat Range value: 3751.327303, ..., 14060.44005

rh_inferiorparietal_volume_combat numeric Right inferior parietal gyri grey matter volume harmonized with combat Range value: 4477.064393, ..., 21820.55665

rh_inferiortemporal_volume_combat numeric Right inferior temporal gyri grey matter volume harmonized with combat Range value: 3216.047218, ..., 16327.61961

rh_isthmuscingulate_volume_combat numeric Right isthmus cingulate grey matter volume harmonized with combat Range value: 209.1161591, ..., 3834.809819

rh_lateraloccipital_volume_combat numeric Right lateral occipital gyri grey matter volume harmonized with combat Range value: 4843.285942, ..., 20034.88073

rh_lateralorbitofrontal_volume_combat numeric Right lateral orbitofrontal gyri grey matter volume harmonized with combat Range value: 2906.323422, ..., 10079.90941

rh_lingual_volume_combat numeric Right lingual gyri grey matter volume harmonized with combat Range value: 3234.690199, ..., 10786.4873

rh_medialorbitofrontal_volume_combat numeric Right medial orbitofrontal grey matter volume harmonized with combat Range value: 2671.237232, ..., 10824.92533

rh_middletemporal_volume_combat numeric Right middle temporal gyri grey matter volume harmonized with combat Range value: 3819.59084, ..., 17009.73677

rh_parahippocampal_volume_combat numeric Right parahippocampal gyri grey matter volume harmonized with combat Range value: 789.60161, ..., 4386.074449

rh_paracentral_volume_combat numeric Right paracentral lobule grey matter volume harmonized with combat Range value: 1813.742895, ..., 6372.198199

rh_parsopercularis_volume_combat numeric Right pars opercularis grey matter volume harmonized with combat Range value: 1487.518395, ..., 6373.976055

rh_parsorbitalis_volume_combat numeric Right pars orbitalis grey matter volume harmonized with combat Range value: 440.9897344, ..., 3889.862538

rh_parstriangularis_volume_combat numeric Right pars triangularis grey matter volume harmonized with combat Range value: 1833.211177, ..., 7977.678728

rh_pericalcarine_volume_combat numeric Right pericalcarine gyri grey matter volume harmonized with combat Range value: 808.560845, ..., 4805.154352

rh_postcentral_volume_combat numeric Right postcentral gyri grey matter volume harmonized with combat Range value: 3563.954448, ..., 14497.09762

rh_posteriorcingulate_volume_combat numeric Right posterior cingulate grey matter volume harmonized with combat Range value: 375.7366496, ..., 4779.561827

rh_precentral_volume_combat numeric Right precentral gyri grey matter volume harmonized with combat Range value: 4520.407215, ..., 19096.9561

rh_precuneus_volume_combat numeric Right precuneus grey matter volume harmonized with combat Range value: 3541.873364, ..., 14549.9908

rh_rostralanteriorcingulate_volume_combat numeric Right rostral anterior cingulate grey matter volume harmonized with combat Range value: 274.0845188, ..., 3666.827354

rh_rostralmiddlefrontal_volume_combat numeric Right rostral middle frontal gyri grey matter volume harmonized with combat Range value: 6234.779366, ..., 29213.82144

rh_superiorfrontal_volume_combat numeric Right superior frontal gyri grey matter volume harmonized with combat Range value: 9047.638904, ..., 28259.75151

rh_superiorparietal_volume_combat numeric Right superior parietal gyri grey matter volume harmonized with combat Range value: 4429.741469, ..., 23504.11619

rh_superiortemporal_volume_combat numeric Right superior temporal gyri grey matter volume harmonized with combat Range value: 4909.45942, ..., 15644.18838

rh_supramarginal_volume_combat numeric Right supramarginal gyri grey matter volume harmonized with combat Range value: 3233.720921, ..., 14812.72412

rh_frontalpole_volume_combat numeric Right frontal pole grey matter volume harmonized with combat Range value: 443.0783677, ..., 2198.014763

rh_temporalpole_volume_combat numeric Right temporal pole grey matter volume harmonized with combat Range value: 641.9196442, ..., 4293.895679

rh_transversetemporal_volume_combat numeric Right transverse temporal pole grey matter volume harmonized with combat Range value: 315.893635, ..., 1524.37567

rh_insula_volume_combat numeric Right insula grey matter volume harmonized with combat Range value: 3370.448742, ..., 10842.33181

EstimatedTotalIntraCranialVol numeric Total intracranial volume raw metric Range value: 889326.5117, ..., 2121140.929

Left_Lateral_Ventricle numeric Left lateral ventricle volume raw metric Range value: 3265.2, ..., 89134.3

Left_Inf_Lat_Vent numeric Left inferior lateral ventricle volume raw metric Range value: 53.5, ..., 8618.2

Left_Cerebellum_White_Matter numeric Left cerebellar white matter volume raw metric Range value: 7046.7, ..., 24822.6

Left_Cerebellum_Cortex numeric Left cerebellar cortex volume raw metric Range value: 33959.5, ..., 71521.8

Left_Thalamus_Proper numeric Left thalamus volume raw metric Range value: 3992.6, ..., 9727.1

Left_Caudate numeric Left caudate volume raw metric Range value: 1878.2, ..., 6241.4

Left_Putamen numeric Left putamen volume raw metric Range value: 1909.4, ..., 8108.3

Left_Pallidum numeric Left pallidum volume raw metric Range value: 821.6, ..., 3176

X3rd_Ventricle numeric 3rd ventricle volume raw metric Range value: 509.6, ..., 5006

X4th_Ventricle numeric 4th ventricle volume raw metric Range value: 708.2, ..., 5333.8

Brain_Stem numeric Brainstem volume raw metric Range value: 12080.6, ..., 34591.5

Left_Hippocampus numeric Left hippocampus volume raw metric Range value: 1429.4, ..., 5542.2

Left_Amygdala numeric Left amygdala volume raw metric Range value: 345.4, ..., 2326.3

CSF numeric CSF volume raw metric Range value: 522.8, ..., 4830.9

Left_Accumbens_area numeric Left accumbens volume raw metric Range value: 72.1, ..., 834.7

Left_VentralDC numeric Left ventral DC volume raw metric Range value: 2095.4, ..., 6518.7

Left_vessel numeric Left vessel volume raw metric Range value: 0, ..., 580

Left_choroid_plexus numeric Left choroid plexus volume raw metric Range value: 93.3, ..., 2036.5

Right_Lateral_Ventricle numeric Right lateral ventricle volume raw metric Range value: 2540.7, ..., 70240.7

Right_Inf_Lat_Vent numeric Right inferior lateral ventricle volume raw metric Range value: 39.4, ..., 7059.2

Right_Cerebellum_White_Matter numeric Right cerebellar white matter volume raw metric Range value: 6870, ..., 27553.3

Right_Cerebellum_Cortex numeric Right cerebellar cortex volume raw metric Range value: 31610.6, ..., 73642

Right_Thalamus_Proper numeric Right thalamus volume raw metric Range value: 3898.3, ..., 9654.5

Right_Caudate numeric Right caudate volume raw metric Range value: 1807.6, ..., 6235.4

Right_Putamen numeric Right putamen volume raw metric Range value: 1832.8, ..., 7813.8

Right_Pallidum numeric Right pallidum volume raw metric Range value: 964.3, ..., 2911.2

Right_Hippocampus numeric Right hippocampus volume raw metric Range value: 1464.4, ..., 5575.7

Right_Amygdala numeric Right amygdala volume raw metric Range value: 452.8, ..., 2580.6

Right_Accumbens_area numeric Right accumbens volume raw metric Range value: 85.6, ..., 851.7

Right_VentralDC numeric Right ventral DC volume raw metric Range value: 2237.1, ..., 5714.6

Right_vessel numeric Right vessel volume raw metric Range value: 0, ..., 335.9

Right_choroid_plexus numeric Right choroid plexus volume raw metric Range value: 178, ..., 2087.1

5th_Ventricle numeric 5th ventricle volume raw metric Range value: 0, ..., 31.8

WM_hypointensities numeric White matter hypointensities raw metric Range value: 240.3, ..., 67617.6

Left_WM_hypointensities numeric Left white matter hypointensities raw metric Range value: 0, ..., 0

Right_WM_hypointensities numeric Right white matter hypointensities raw metric Range value: 0, ..., 0

non_WM_hypointensities numeric non-White matter hypointensities raw metric Range value: 0, ..., 480.2

Left_non_WM_hypointensities numeric Left non-White matter hypointensities raw metric Range value: 0, ..., 0

Right_non_WM_hypointensities numeric Right non-White matter hypointensities raw metric Range value: 0, ..., 0

Optic_Chiasm numeric Optic Chiasm volume raw metric Range value: 2.7, ..., 465.4

CC_Posterior numeric Posterior corpus collosum volume raw metric Range value: 93, ..., 2143.8

CC_Mid_Posterior numeric Mid-posterior corpus collosum volume raw metric Range value: 153.1, ..., 1379.3

CC_Central numeric Central corpus collosum volume raw metric Range value: 164.7, ..., 1056.6

CC_Mid_Anterior numeric Mid-anterior corpus collosum volume raw metric Range value: 140.9, ..., 1058

CC_Anterior numeric Anterior corpus collosum volume raw metric Range value: 156.8, ..., 1575.2

BrainSegVol numeric Segmented brain volume raw metric Range value: 711006, ..., 1497555

BrainSegVolNotVent numeric Segmented brain volume excluding ventricles raw metric Range value: 684476, ..., 1469088

BrainSegVolNotVentSurf numeric Surface size of the segmented brain volume excluding ventricles raw metric Range value: 683626.8863, ..., 1468779.629

lhCortexVol numeric Left cerebral cortex grey matter volume raw metric Range value: 113305.7058, ..., 295169.5683

rhCortexVol numeric Right cerebral cortex grey matter volume raw metric Range value: 126887.3505, ..., 297733.369

CortexVol numeric Total cerebral cortex grey matter volume raw metric Range value: 240193.0563, ..., 592902.9374

lhCerebralWhiteMatterVol numeric Left cerebral cortex white matter volume raw metric Range value: 111121.8718, ..., 339282.441

rhCerebralWhiteMatterVol numeric Right cerebral cortex white matter volume raw metric Range value: 115194.8984, ..., 344639.3376

CerebralWhiteMatterVol numeric Total cerebral cortex white matter volume raw metric Range value: 226316.7702, ..., 676194.021

SubCortGrayVol numeric Total volume of the subcortical grey matter raw metric Range value: 31162, ..., 70983

TotalGrayVol numeric Total volume of all grey matter (cortical and subcortical) raw metric Range value: 356617.0563, ..., 778674.9374

SupraTentorialVol numeric Total volume of brain (minus the cerebellum) raw metric Range value: 603044.8863, ..., 1337640.66

SupraTentorialVolNotVent numeric Total volume of brain without ventricles (minus the cerebellum) raw metric Range value: 580663.8863, ..., 1305712.629

SupraTentorialVolNotVentVox numeric Number of voxels of brain without ventricles (minus the cerebellum) raw metric Range value: 578764, ..., 1302471

MaskVol numeric Mask volume raw metric Range value: 1058574, ..., 2121724

BrainSegVol_to_eTIV numeric Ratio of total brain segmentation volume to ICV raw metric Range value: 0.542987, ..., 1.008595

MaskVol_to_eTIV numeric Ratio of mask segmentation volume to ICV raw metric Range value: 0.750969, ..., 1.367949

lhSurfaceHoles numeric Left surface hole volume raw metric Range value: 2, ..., 351

rhSurfaceHoles numeric Right surface hole volume raw metric Range value: 0, ..., 358

SurfaceHoles numeric Total surface hole volume raw metric Range value: 6, ..., 709

lh_bankssts_thickness numeric Left banks of the superior temporal sulcus thickness raw metric Range value: 1.607, ..., 2.881

lh_caudalanteriorcingulate_thickness numeric Left caudal anterior cingulate thickness raw metric Range value: 1.488, ..., 3.694

lh_caudalmiddlefrontal_thickness numeric Left caudal middle frontal gyri thickness raw metric Range value: 1.645, ..., 2.941

lh_cuneus_thickness numeric Left cuneus thickness raw metric Range value: 1.367, ..., 2.418

lh_entorhinal_thickness numeric Left entorhinal cortex thickness raw metric Range value: 1.485, ..., 4.465

lh_fusiform_thickness numeric Left fusiform gyri thickness raw metric Range value: 1.861, ..., 3.227

lh_inferiorparietal_thickness numeric Left inferior parietal gyri thickness raw metric Range value: 1.586, ..., 2.771

lh_inferiortemporal_thickness numeric Left inferior temporal gyri thickness raw metric Range value: 1.914, ..., 3.318

lh_isthmuscingulate_thickness numeric Left isthmus cingulate thickness raw metric Range value: 1.48, ..., 3.062

lh_lateraloccipital_thickness numeric Left lateral occipital gyri thickness raw metric Range value: 1.483, ..., 2.597

lh_lateralorbitofrontal_thickness numeric Left lateral orbitofrontal gyri thickness raw metric Range value: 1.899, ..., 3.121

lh_lingual_thickness numeric Left lingual gyri thickness raw metric Range value: 1.385, ..., 2.522

lh_medialorbitofrontal_thickness numeric Left medial orbitofrontal thickness raw metric Range value: 1.688, ..., 3.037

lh_middletemporal_thickness numeric Left middle temporal gyri thickness raw metric Range value: 1.853, ..., 3.234

lh_parahippocampal_thickness numeric Left parahippocampal gyri thickness raw metric Range value: 1.36, ..., 3.719

lh_paracentral_thickness numeric Left paracentral lobule thickness raw metric Range value: 1.504, ..., 2.935

lh_parsopercularis_thickness numeric Left pars opercularis thickness raw metric Range value: 1.791, ..., 2.934

lh_parsorbitalis_thickness numeric Left pars orbitalis thickness raw metric Range value: 1.778, ..., 3.418

lh_parstriangularis_thickness numeric Left pars triangularis thickness raw metric Range value: 1.622, ..., 2.844

lh_pericalcarine_thickness numeric Left pericalcarine gyri thickness raw metric Range value: 1.14, ..., 2.167

lh_postcentral_thickness numeric Left postcentral gyri thickness raw metric Range value: 1.427, ..., 2.453

lh_posteriorcingulate_thickness numeric Left posterior cingulate thickness raw metric Range value: 1.6, ..., 3.08

lh_precentral_thickness numeric Left precentral gyri thickness raw metric Range value: 1.557, ..., 2.901

lh_precuneus_thickness numeric Left precuneus thickness raw metric Range value: 1.546, ..., 2.79

lh_rostralanteriorcingulate_thickness numeric Left rostral anterior cingulate thickness raw metric Range value: 1.716, ..., 3.693

lh_rostralmiddlefrontal_thickness numeric Left rostral middle frontal gyri thickness raw metric Range value: 1.638, ..., 2.726

lh_superiorfrontal_thickness numeric Left superior frontal gyri thickness raw metric Range value: 1.844, ..., 3.025

lh_superiorparietal_thickness numeric Left superior parietal gyri thickness raw metric Range value: 1.448, ..., 2.68

lh_superiortemporal_thickness numeric Left superior temporal gyri thickness raw metric Range value: 1.792, ..., 3.165

lh_supramarginal_thickness numeric Left supramarginal gyri thickness raw metric Range value: 1.661, ..., 2.949

lh_frontalpole_thickness numeric Left frontal pole thickness raw metric Range value: 1.51, ..., 3.68

lh_temporalpole_thickness numeric Left temporal pole thickness raw metric Range value: 1.921, ..., 4.498

lh_transversetemporal_thickness numeric Left transverse temporal pole thickness raw metric Range value: 1.362, ..., 2.921

lh_insula_thickness numeric Left insula thickness raw metric Range value: 2.085, ..., 3.453

lh_MeanThickness_thickness numeric Left mean lobar thickness raw metric Range value: 1.84272, ..., 2.70079

rh_bankssts_thickness numeric Right banks of the superior temporal sulcus thickness raw metric Range value: 1.628, ..., 3.06

rh_caudalanteriorcingulate_thickness numeric Right caudal anterior cingulate thickness raw metric Range value: 1.505, ..., 3.478

rh_caudalmiddlefrontal_thickness numeric Right caudal middle frontal gyri thickness raw metric Range value: 1.629, ..., 2.935

rh_cuneus_thickness numeric Right cuneus thickness raw metric Range value: 1.4, ..., 2.439

rh_entorhinal_thickness numeric Right entorhinal cortex thickness raw metric Range value: 1.44, ..., 4.641

rh_fusiform_thickness numeric Right fusiform gyri thickness raw metric Range value: 1.864, ..., 3.235

rh_inferiorparietal_thickness numeric Right inferior parietal gyri thickness raw metric Range value: 1.596, ..., 2.729

rh_inferiortemporal_thickness numeric Right inferior temporal gyri thickness raw metric Range value: 1.912, ..., 3.354

rh_isthmuscingulate_thickness numeric Right isthmus cingulate thickness raw metric Range value: 1.43, ..., 3.172

rh_lateraloccipital_thickness numeric Right lateral occipital gyri thickness raw metric Range value: 1.533, ..., 2.706

rh_lateralorbitofrontal_thickness numeric Right lateral orbitofrontal gyri thickness raw metric Range value: 1.898, ..., 3.155

rh_lingual_thickness numeric Right lingual gyri thickness raw metric Range value: 1.488, ..., 2.454

rh_medialorbitofrontal_thickness numeric Right medial orbitofrontal thickness raw metric Range value: 1.653, ..., 3.107

rh_middletemporal_thickness numeric Right middle temporal gyri thickness raw metric Range value: 1.892, ..., 3.255

rh_parahippocampal_thickness numeric Right parahippocampal gyri thickness raw metric Range value: 1.439, ..., 3.511

rh_paracentral_thickness numeric Right paracentral lobule thickness raw metric Range value: 1.567, ..., 3.043

rh_parsopercularis_thickness numeric Right pars opercularis thickness raw metric Range value: 1.747, ..., 2.904

rh_parsorbitalis_thickness numeric Right pars orbitalis thickness raw metric Range value: 1.76, ..., 3.354

rh_parstriangularis_thickness numeric Right pars triangularis thickness raw metric Range value: 1.64, ..., 2.913

rh_pericalcarine_thickness numeric Right pericalcarine gyri thickness raw metric Range value: 1.179, ..., 2.205

rh_postcentral_thickness numeric Right postcentral gyri thickness raw metric Range value: 1.374, ..., 2.62

rh_posteriorcingulate_thickness numeric Right posterior cingulate thickness raw metric Range value: 1.555, ..., 3.012

rh_precentral_thickness numeric Right precentral gyri thickness raw metric Range value: 1.52, ..., 2.863

rh_precuneus_thickness numeric Right precuneus thickness raw metric Range value: 1.583, ..., 2.763

rh_rostralanteriorcingulate_thickness numeric Right rostral anterior cingulate thickness raw metric Range value: 1.951, ..., 3.867

rh_rostralmiddlefrontal_thickness numeric Right rostral middle frontal gyri thickness raw metric Range value: 1.583, ..., 2.77

rh_superiorfrontal_thickness numeric Right superior frontal gyri thickness raw metric Range value: 1.879, ..., 3.02

rh_superiorparietal_thickness numeric Right superior parietal gyri thickness raw metric Range value: 1.381, ..., 2.74

rh_superiortemporal_thickness numeric Right superior temporal gyri thickness raw metric Range value: 1.8, ..., 3.158

rh_supramarginal_thickness numeric Right supramarginal gyri thickness raw metric Range value: 1.65, ..., 2.886

rh_frontalpole_thickness numeric Right frontal pole thickness raw metric Range value: 1.621, ..., 3.669

rh_temporalpole_thickness numeric Right temporal pole thickness raw metric Range value: 1.914, ..., 4.514

rh_transversetemporal_thickness numeric Right transverse temporal pole thickness raw metric Range value: 1.291, ..., 3.006

rh_insula_thickness numeric Right insula thickness raw metric Range value: 2.082, ..., 3.438

rh_MeanThickness_thickness numeric Right mean lobar thickness raw metric Range value: 1.80663, ..., 2.72506

lh_bankssts_volume numeric Left banks of the superior temporal sulcus grey matter volume raw metric Range value: 839, ..., 4865

lh_caudalanteriorcingulate_volume numeric Left caudal anterior cingulate grey matter volume raw metric Range value: 324, ..., 3511

lh_caudalmiddlefrontal_volume numeric Left caudal middle frontal gyri grey matter volume raw metric Range value: 2609, ..., 9304

lh_cuneus_volume numeric Left cuneus grey matter volume raw metric Range value: 1306, ..., 5331

lh_entorhinal_volume numeric Left entorhinal cortex grey matter volume raw metric Range value: 361, ..., 4503

lh_fusiform_volume numeric Left fusiform gyri grey matter volume raw metric Range value: 4444, ..., 14014

lh_inferiorparietal_volume numeric Left inferior parietal gyri grey matter volume raw metric Range value: 4615, ..., 17442

lh_inferiortemporal_volume numeric Left inferior temporal gyri grey matter volume raw metric Range value: 4150, ..., 16678

lh_isthmuscingulate_volume numeric Left isthmus cingulate grey matter volume raw metric Range value: 947, ..., 4452

lh_lateraloccipital_volume numeric Left lateral occipital gyri grey matter volume raw metric Range value: 4821, ..., 18189

lh_lateralorbitofrontal_volume numeric Left lateral orbitofrontal gyri grey matter volume raw metric Range value: 3419, ..., 10377

lh_lingual_volume numeric Left lingual gyri grey matter volume raw metric Range value: 2692, ..., 10824

lh_medialorbitofrontal_volume numeric Left medial orbitofrontal grey matter volume raw metric Range value: 2572, ..., 7368

lh_middletemporal_volume numeric Left middle temporal gyri grey matter volume raw metric Range value: 4065, ..., 15715

lh_parahippocampal_volume numeric Left parahippocampal gyri grey matter volume raw metric Range value: 813, ..., 3367

lh_paracentral_volume numeric Left paracentral lobule grey matter volume raw metric Range value: 1766, ..., 5803

lh_parsopercularis_volume numeric Left pars opercularis grey matter volume raw metric Range value: 2043, ..., 8957

lh_parsorbitalis_volume numeric Left pars orbitalis grey matter volume raw metric Range value: 848, ..., 3431

lh_parstriangularis_volume numeric Left pars triangularis grey matter volume raw metric Range value: 1196, ..., 6039

lh_pericalcarine_volume numeric Left pericalcarine gyri grey matter volume raw metric Range value: 905, ..., 4215

lh_postcentral_volume numeric Left postcentral gyri grey matter volume raw metric Range value: 4396, ..., 14390

lh_posteriorcingulate_volume numeric Left posterior cingulate grey matter volume raw metric Range value: 968, ..., 4664

lh_precentral_volume numeric Left precentral gyri grey matter volume raw metric Range value: 5857, ..., 18847

lh_precuneus_volume numeric Left precuneus grey matter volume raw metric Range value: 4147, ..., 12844

lh_rostralanteriorcingulate_volume numeric Left rostral anterior cingulate grey matter volume raw metric Range value: 797, ..., 4640

lh_rostralmiddlefrontal_volume numeric Left rostral middle frontal gyri grey matter volume raw metric Range value: 5142, ..., 21847

lh_superiorfrontal_volume numeric Left superior frontal gyri grey matter volume raw metric Range value: 8647, ..., 29097

lh_superiorparietal_volume numeric Left superior parietal gyri grey matter volume raw metric Range value: 5267, ..., 19854

lh_superiortemporal_volume numeric Left superior temporal gyri grey matter volume raw metric Range value: 4846, ..., 16808

lh_supramarginal_volume numeric Left supramarginal gyri grey matter volume raw metric Range value: 4895, ..., 18033

lh_frontalpole_volume numeric Left frontal pole grey matter volume raw metric Range value: 336, ..., 1599

lh_temporalpole_volume numeric Left temporal pole grey matter volume raw metric Range value: 622, ..., 4254

lh_transversetemporal_volume numeric Left transverse temporal pole grey matter volume raw metric Range value: 401, ..., 2269

lh_insula_volume numeric Left insula grey matter volume raw metric Range value: 3921, ..., 9985

rh_bankssts_volume numeric Right banks of the superior temporal sulcus grey matter volume raw metric Range value: 877, ..., 3579

rh_caudalanteriorcingulate_volume numeric Right caudal anterior cingulate grey matter volume raw metric Range value: 346, ..., 3760

rh_caudalmiddlefrontal_volume numeric Right caudal middle frontal gyri grey matter volume raw metric Range value: 2405, ..., 9318

rh_cuneus_volume numeric Right cuneus grey matter volume raw metric Range value: 1571, ..., 6486

rh_entorhinal_volume numeric Right entorhinal cortex grey matter volume raw metric Range value: 455, ..., 3944

rh_fusiform_volume numeric Right fusiform gyri grey matter volume raw metric Range value: 3956, ..., 13849

rh_inferiorparietal_volume numeric Right inferior parietal gyri grey matter volume raw metric Range value: 5606, ..., 21876

rh_inferiortemporal_volume numeric Right inferior temporal gyri grey matter volume raw metric Range value: 4131, ..., 15982

rh_isthmuscingulate_volume numeric Right isthmus cingulate grey matter volume raw metric Range value: 453, ..., 3823

rh_lateraloccipital_volume numeric Right lateral occipital gyri grey matter volume raw metric Range value: 5604, ..., 19426

rh_lateralorbitofrontal_volume numeric Right lateral orbitofrontal gyri grey matter volume raw metric Range value: 2994, ..., 9968

rh_lingual_volume numeric Right lingual gyri grey matter volume raw metric Range value: 3265, ..., 10941

rh_medialorbitofrontal_volume numeric Right medial orbitofrontal grey matter volume raw metric Range value: 2728, ..., 10692

rh_middletemporal_volume numeric Right middle temporal gyri grey matter volume raw metric Range value: 4959, ..., 17204

rh_parahippocampal_volume numeric Right parahippocampal gyri grey matter volume raw metric Range value: 788, ..., 4084

rh_paracentral_volume numeric Right paracentral lobule grey matter volume raw metric Range value: 1850, ..., 6151

rh_parsopercularis_volume numeric Right pars opercularis grey matter volume raw metric Range value: 1707, ..., 6055

rh_parsorbitalis_volume numeric Right pars orbitalis grey matter volume raw metric Range value: 870, ..., 3935

rh_parstriangularis_volume numeric Right pars triangularis grey matter volume raw metric Range value: 1820, ..., 7069

rh_pericalcarine_volume numeric Right pericalcarine gyri grey matter volume raw metric Range value: 987, ..., 4870

rh_postcentral_volume numeric Right postcentral gyri grey matter volume raw metric Range value: 4594, ..., 13713

rh_posteriorcingulate_volume numeric Right posterior cingulate grey matter volume raw metric Range value: 615, ..., 4794

rh_precentral_volume numeric Right precentral gyri grey matter volume raw metric Range value: 5978, ..., 18711

rh_precuneus_volume numeric Right precuneus grey matter volume raw metric Range value: 4393, ..., 14963

rh_rostralanteriorcingulate_volume numeric Right rostral anterior cingulate grey matter volume raw metric Range value: 532, ..., 3591

rh_rostralmiddlefrontal_volume numeric Right rostral middle frontal gyri grey matter volume raw metric Range value: 6168, ..., 28382

rh_superiorfrontal_volume numeric Right superior frontal gyri grey matter volume raw metric Range value: 8546, ..., 28952

rh_superiorparietal_volume numeric Right superior parietal gyri grey matter volume raw metric Range value: 5905, ..., 22287

rh_superiortemporal_volume numeric Right superior temporal gyri grey matter volume raw metric Range value: 5509, ..., 15533

rh_supramarginal_volume numeric Right supramarginal gyri grey matter volume raw metric Range value: 4604, ..., 14651

rh_frontalpole_volume numeric Right frontal pole grey matter volume raw metric Range value: 454, ..., 1985

rh_temporalpole_volume numeric Right temporal pole grey matter volume raw metric Range value: 655, ..., 4241

rh_transversetemporal_volume numeric Right transverse temporal pole grey matter volume raw metric Range value: 357, ..., 1524

rh_insula_volume numeric Right insula grey matter volume raw metric Range value: 3934, ..., 10302

update_stamp POSIXct: YYYY-MM-DD HH:MM:SS