Adverse Events Log

Adverse Events Log data. More information is available at https://adni.loni.usc.edu/data-samples/adni-data/

Usage

data(ADVERSE)

Format

A data frame with 3675 observations and 89 variables.

Source

https://adni.loni.usc.edu/data-samples/adni-data/

Details

ORIGPROT character Original study protocol , with options: ADNIGO, ADNI3, ADNI1, ADNI2, ADNI4

COLPROT factor Study protocol of data collection , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4

PTID character Participant ID

RID numeric Participant roster ID Range value: 2, ..., 10806

VISCODE character Visit code

VISCODE2 character Translated visit code

AENUMBER numeric Adverse event number Range value: 1, ..., 34

AECHANGE character Is this event a: 1=New adverse event ; 2=Worsening condition or symptom present at screening/initial visit Decoded Value: Yes

AEPENUM numeric Initial Health Assessment Condition Number: Range value: 0, ..., 59

AECOMPOUND character Is this event part of a compound event? 0=No ; 1=Yes Decoded Value: Yes

AEHONSDT Event onset date Date: YYY-MM-DD

AEOUTCOME character Event outcome 1=Recovered/Resolved ; 2=Recovering/Resolving ; 3=Not recovered/Not resolved ; 4=Recovered/Resolved with sequelae ; 5=Fatal ; 6=Unknown Decoded Value: Yes

AEHCSDT Event cease date Date: YYY-MM-DD

AEHDTHDT numeric Date of death Range value: 2017, ..., 2025

AEHDBRN character Was a brain autopsy performed? 0=No ; 1=Yes Decoded Value: Yes

AEHCHRON character Initial Chronicity 1=Single occurrence ; 2=Intermittent - daily ; 3=Intermittent - weekly ; 4=Intermittent - monthly or less frequently ; 5=Persistent Decoded Value: Yes

AEHSEVR character Initial Severity 1=Mild ; 2=Moderate ; 3=Severe Decoded Value: Yes

AERELAD character Related to symptoms associated with the development of early stage AD or progression of AD 1=Definitely ; 2=Probably ; 3=Possibly ; 4=Unlikely ; 5=Not related Decoded Value: Yes

AERELCM character Related to concomitant therapy 1=Definitely ; 2=Probably ; 3=Possibly ; 4=Unlikely ; 5=Not related Decoded Value: Yes

AERELFLRBTBN character Related to Florbetaben tracer 1=Definitely ; 2=Probably ; 3=Possibly ; 4=Unlikely ; 5=Not related Decoded Value: Yes

AERELFLRBPR character Related to Florbetapir tracer 1=Definitely ; 2=Probably ; 3=Possibly ; 4=Unlikely ; 5=Not related Decoded Value: Yes

AEHIMG character Related to PET/MRI imaging procedure 1=Definitely ; 2=Probably ; 3=Possibly ; 4=Unlikely ; 5=Not related Decoded Value: Yes

AERELTAU character Related to Flotaucipir tracer 1=Definitely ; 2=Probably ; 3=Possibly ; 4=Unlikely ; 5=Not related Decoded Value: Yes

AERELNAV character Related to NAV-4694 tracer 1=Definitely ; 2=Probably ; 3=Possibly ; 4=Unlikely ; 5=Not related Decoded Value: Yes

AERELMK character Related to MK-6240 tracer 1=Definitely ; 2=Probably ; 3=Possibly ; 4=Unlikely ; 5=Not related Decoded Value: Yes

AERELPI character Related to PI-2620 tracer 1=Definitely ; 2=Probably ; 3=Possibly ; 4=Unlikely ; 5=Not related Decoded Value: Yes

AEHLUMB character Related to lumbar puncture 1=Definitely ; 2=Probably ; 3=Possibly ; 4=Unlikely ; 5=Not related Decoded Value: Yes

AERELCOVID character Related to COVID-19 illness 1=Definitely (Confirmed COVID-19 test) ; 2=Probably ; 3=Possibly ; 4=Unlikely ; 5=Not related Decoded Value: Yes

AERELPAN character Related to COVID-19 pandemic disruption 1=Definitely ; 2=Probably ; 3=Possibly ; 4=Unlikely ; 5=Not related Decoded Value: Yes

AERELATESP character Related to other study procedure(s) 1=Definitely ; 2=Probably ; 3=Possibly ; 4=Unlikely ; 5=Not related Decoded Value: Yes

AEHCMEDS character Was treatment (medication) for AE required? 0=No ; 1=Yes Decoded Value: Yes

AESERIOUS character Was event serious at any time during the trial? 0=No ; 1=Yes Decoded Value: Yes

AESERDATE Date event became serious: Date: YYY-MM-DD

AEDTSITE Date site became aware that the event is serious: Date: YYY-MM-DD

SAELIFE character Is this an SAE because event was life-threatening? 0=No ; 1=Yes Decoded Value: Yes

SAEHOSPIT character Is this an SAE because it lead to hospitalization? 0=No ; 1=Yes Decoded Value: Yes

SAEPROLONG character Is this an SAE because it lead to prolongation of hospitalization? 0=No ; 1=Yes Decoded Value: Yes

SAEDEATH character Is this an SAE because it lead to death? 0=No ; 1=Yes Decoded Value: Yes

SAECONGEN character Is this an SAE because it lead to a congenital anomaly or birth defect? 0=No ; 1=Yes Decoded Value: Yes

SAEDISAB character Is this an SAE because it caused persistent or significant disability/incapacity? 0=No ; 1=Yes Decoded Value: Yes

SAEOTHER character Is this an SAE because it lead to another medically important condition? 0=No ; 1=Yes Decoded Value: Yes

AECHANGE1 character Is there a change to report? 0=No ; 1=Yes Decoded Value: Yes

AEDATE1 Change 1. Date Date: YYY-MM-DD

AESEV1 character Change 1: Severity 1=Mild ; 2=Moderate ; 3=Severe Decoded Value: Yes

AECHRON1 character Change 1: Chronicity 1=Single occurrence ; 2=Intermittent - daily ; 3=Intermittent - weekly ; 4=Intermittent - monthly or less frequently ; 5=Persistent Decoded Value: Yes

AECHANGE2 character Is there a 2nd change to report? 0=No ; 1=Yes Decoded Value: Yes

AEDATE2 Change 2. Date Date: YYY-MM-DD

AESEV2 character Change 2: Severity 1=Mild ; 2=Moderate ; 3=Severe Decoded Value: Yes

AECHRON2 character Change 2: Chronicity 1=Single occurrence ; 2=Intermittent - daily ; 3=Intermittent - weekly ; 4=Intermittent - monthly or less frequently ; 5=Persistent Decoded Value: Yes

AECHANGE3 character Is there a 3rd change to report? 0=No ; 1=Yes Decoded Value: Yes

AEDATE3 Change 3. Date Date: YYY-MM-DD

AESEV3 character Change 3: Severity 1=Mild ; 2=Moderate ; 3=Severe Decoded Value: Yes

AECHRON3 character Change 3: Chronicity 1=Single occurrence ; 2=Intermittent - daily ; 3=Intermittent - weekly ; 4=Intermittent - monthly or less frequently ; 5=Persistent Decoded Value: Yes

AECHANGE4 character Is there a 4th change to report? 0=No ; 1=Yes Decoded Value: Yes

AEDATE4 Change 4. Date Date: YYY-MM-DD

AESEV4 character Change 4: Severity 1=Mild ; 2=Moderate ; 3=Severe Decoded Value: Yes

AECHRON4 character Change 4: Chronicity 1=Single occurrence ; 2=Intermittent - daily ; 3=Intermittent - weekly ; 4=Intermittent - monthly or less frequently ; 5=Persistent Decoded Value: Yes

AECHANGE5 character Is there a 5th change to report? 0=No ; 1=Yes Decoded Value: Yes

AEDATE5 logical Change 5. Date A Boolean value: TRUE/ FALSE

AESEV5 character Change 5: Severity 1=Mild ; 2=Moderate ; 3=Severe Decoded Value: Yes

AECHRON5 character Change 5: Chronicity 1=Single occurrence ; 2=Intermittent - daily ; 3=Intermittent - weekly ; 4=Intermittent - monthly or less frequently ; 5=Persistent Decoded Value: Yes

AECHANGE6 character Is there a 6th change to report? 0=No ; 1=Yes Decoded Value: Yes

AEDATE6 logical Change 6. Date A Boolean value: TRUE/ FALSE

AESEV6 character Change 6: Severity 1=Mild ; 2=Moderate ; 3=Severe Decoded Value: Yes

AECHRON6 character Change 6: Chronicity 1=Single occurrence ; 2=Intermittent - daily ; 3=Intermittent - weekly ; 4=Intermittent - monthly or less frequently ; 5=Persistent Decoded Value: Yes

AECHANGE7 character Is there a 7th change to report? 0=No ; 1=Yes Decoded Value: Yes

AEDATE7 logical Change 7. Date A Boolean value: TRUE/ FALSE

AESEV7 character Change 7: Severity 1=Mild ; 2=Moderate ; 3=Severe Decoded Value: Yes

AECHRON7 character Change 7: Chronicity 1=Single occurrence ; 2=Intermittent - daily ; 3=Intermittent - weekly ; 4=Intermittent - monthly or less frequently ; 5=Persistent Decoded Value: Yes

AECHANGE8 character Is there a 8th change to report? 0=No ; 1=Yes Decoded Value: Yes

AEDATE8 logical Change 8. Date A Boolean value: TRUE/ FALSE

AESEV8 character Change 8: Severity 1=Mild ; 2=Moderate ; 3=Severe Decoded Value: Yes

AECHRON8 character Change 8: Chronicity 1=Single occurrence ; 2=Intermittent - daily ; 3=Intermittent - weekly ; 4=Intermittent - monthly or less frequently ; 5=Persistent Decoded Value: Yes

AECHANGE9 character Is there a 9th change to report? 0=No ; 1=Yes Decoded Value: Yes

AEDATE9 logical Change 9. Date A Boolean value: TRUE/ FALSE

AESEV9 character Change 9: Severity 1=Mild ; 2=Moderate ; 3=Severe Decoded Value: Yes

AECHRON9 character Change 9: Chronicity 1=Single occurrence ; 2=Intermittent - daily ; 3=Intermittent - weekly ; 4=Intermittent - monthly or less frequently ; 5=Persistent Decoded Value: Yes

AECHANGE10 character Is there a 10th change to report? 0=No ; 1=Yes Decoded Value: Yes

AEDATE10 logical Change 10. Date A Boolean value: TRUE/ FALSE

AESEV10 character Change 10: Severity 1=Mild ; 2=Moderate ; 3=Severe Decoded Value: Yes

AECHRON10 character Change 10: Chronicity 1=Single occurrence ; 2=Intermittent - daily ; 3=Intermittent - weekly ; 4=Intermittent - monthly or less frequently ; 5=Persistent Decoded Value: Yes

ID numeric Record ID Range value: 563, ..., 186890

SITEID numeric Site ID Range value: 1, ..., 941

USERDATE User Create Date Date: YYY-MM-DD

USERDATE2 User Last Modify Date Date: YYY-MM-DD

DD_CRF_VERSION_LABEL character CRF Version Label , with options: v1

LANGUAGE_CODE character Language Code e=English Decoded Value: Yes

HAS_QC_ERROR character Has Quality Check Error 0=Does not have QC error or QC error has been approved ; 1=Has QC error Decoded Value: Yes

update_stamp POSIXct: YYY-MM-DD

Author

adni-data@googlegroups.com