Abbreviated Multidimensional Acculturation Scale

Abbreviated Multidimensional Acculturation Scale data. More information is available at https://adni.loni.usc.edu/data-samples/adni-data/

Usage

data(AMAS)

Format

A data frame with 904 observations and 59 variables.

Source

https://adni.loni.usc.edu/data-samples/adni-data/

Details

ORIGPROT character Original study protocol , with options: ADNI3, ADNI4, ADNI2, ADNIGO, ADNI1

COLPROT factor Study protocol of data collection , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4

PTID character Participant ID

RID numeric Participant roster ID Range value: 21, ..., 10819

VISCODE character Visit code , with options: 4_init, 4_bl, 4_m12

VISCODE2 character Translated visit code

VISDATE Registry EXAMDATE on matching VISCODE Date: YYY-MM-DD

DONE character Was assessment/procedure done? 0=No ; 1=Yes Decoded Value: Yes

NDREASON character If No, reason not done 1=Participant Unable for Cognitive Reasons ; 10=Not called for by protocol ; 11=Participant in a non-USA Site ; 12=Participant is not Latinx ; 2=Participant Unable for Other Reasons (Physical, auditory, etc.) ; 3=Participant/Study Partner Refused ; 7=Site Error ; 8=Other Reason ; 9=Unknown Decoded Value: Yes

AMAS1 character 1. I think of myself as being U.S. American. 1=Strongly disagree ; 2=Disagree somewhat ; 3=Agree somewhat ; 4=Strongly agree Decoded Value: Yes

AMAS2 character 2. I feel good about being U.S. American. 1=Strongly disagree ; 2=Disagree somewhat ; 3=Agree somewhat ; 4=Strongly agree Decoded Value: Yes

AMAS3 character 3. Being U.S. American plays an important part in my life. 1=Strongly disagree ; 2=Disagree somewhat ; 3=Agree somewhat ; 4=Strongly agree Decoded Value: Yes

AMAS4 character 4. I feel that I am part of the U.S. American culture. 1=Strongly disagree ; 2=Disagree somewhat ; 3=Agree somewhat ; 4=Strongly agree Decoded Value: Yes

AMAS5 character 5. I have a strong sense of being U.S. American. 1=Strongly disagree ; 2=Disagree somewhat ; 3=Agree somewhat ; 4=Strongly agree Decoded Value: Yes

AMAS6 character 6. I am proud of being U.S. American. 1=Strongly disagree ; 2=Disagree somewhat ; 3=Agree somewhat ; 4=Strongly agree Decoded Value: Yes

AMAS7 character 7. I think of myself as being___(a member of my culture of origin). 1=Strongly disagree ; 2=Disagree somewhat ; 3=Agree somewhat ; 4=Strongly agree Decoded Value: Yes

AMAS8 character 8. I feel good about being____(a member of my culture of origin). 1=Strongly disagree ; 2=Disagree somewhat ; 3=Agree somewhat ; 4=Strongly agree Decoded Value: Yes

AMAS9 character 9. Being___(a member of my culture of origin) plays an important part in my life. 1=Strongly disagree ; 2=Disagree somewhat ; 3=Agree somewhat ; 4=Strongly agree Decoded Value: Yes

AMAS10 character 10. I feel that I am part of___culture (culture of origin). 1=Strongly disagree ; 2=Disagree somewhat ; 3=Agree somewhat ; 4=Strongly agree Decoded Value: Yes

AMAS11 character 11. I have a strong sense of being___(culture of origin). 1=Strongly disagree ; 2=Disagree somewhat ; 3=Agree somewhat ; 4=Strongly agree Decoded Value: Yes

AMAS12 character 12. I am proud of being___(culture of origin). 1=Strongly disagree ; 2=Disagree somewhat ; 3=Agree somewhat ; 4=Strongly agree Decoded Value: Yes

AMAS13 character 13. at school or work 1=Not at all ; 2=A little ; 3=Pretty well ; 4=Extremely well Decoded Value: Yes

AMAS14 character 14. with American friends 1=Not at all ; 2=A little ; 3=Pretty well ; 4=Extremely well Decoded Value: Yes

AMAS15 character 15. on the phone 1=Not at all ; 2=A little ; 3=Pretty well ; 4=Extremely well Decoded Value: Yes

AMAS16 character 16. with strangers 1=Not at all ; 2=A little ; 3=Pretty well ; 4=Extremely well Decoded Value: Yes

AMAS17 character 17. in general 1=Not at all ; 2=A little ; 3=Pretty well ; 4=Extremely well Decoded Value: Yes

AMAS18 character 18. on television or in movies 1=Not at all ; 2=A little ; 3=Pretty well ; 4=Extremely well Decoded Value: Yes

AMAS19 character 19. in newspapers and magazines 1=Not at all ; 2=A little ; 3=Pretty well ; 4=Extremely well Decoded Value: Yes

AMAS20 character 20. words in songs 1=Not at all ; 2=A little ; 3=Pretty well ; 4=Extremely well Decoded Value: Yes

AMAS21 character 21. in general 1=Not at all ; 2=A little ; 3=Pretty well ; 4=Extremely well Decoded Value: Yes

AMAS22 character 22. with family 1=Not at all ; 2=A little ; 3=Pretty well ; 4=Extremely well Decoded Value: Yes

AMAS23 character 23. with friends from the same country as you 1=Not at all ; 2=A little ; 3=Pretty well ; 4=Extremely well Decoded Value: Yes

AMAS24 character 24. on the phone 1=Not at all ; 2=A little ; 3=Pretty well ; 4=Extremely well Decoded Value: Yes

AMAS25 character 25. with strangers 1=Not at all ; 2=A little ; 3=Pretty well ; 4=Extremely well Decoded Value: Yes

AMAS26 character 26. in general 1=Not at all ; 2=A little ; 3=Pretty well ; 4=Extremely well Decoded Value: Yes

AMAS27 character 27. on television or in movies 1=Not at all ; 2=A little ; 3=Pretty well ; 4=Extremely well Decoded Value: Yes

AMAS28 character 28. in newspapers and magazines 1=Not at all ; 2=A little ; 3=Pretty well ; 4=Extremely well Decoded Value: Yes

AMAS29 character 29. words in songs 1=Not at all ; 2=A little ; 3=Pretty well ; 4=Extremely well Decoded Value: Yes

AMAS30 character 30. in general 1=Not at all ; 2=A little ; 3=Pretty well ; 4=Extremely well Decoded Value: Yes

AMAS31 character 31. American national heroes 1=Not at all ; 2=A little ; 3=Pretty well ; 4=Extremely well Decoded Value: Yes

AMAS32 character 32. popular American television 1=Not at all ; 2=A little ; 3=Pretty well ; 4=Extremely well Decoded Value: Yes

AMAS33 character 33. popular American newspapers 1=Not at all ; 2=A little ; 3=Pretty well ; 4=Extremely well Decoded Value: Yes

AMAS34 character 34. popular American actors and actresses 1=Not at all ; 2=A little ; 3=Pretty well ; 4=Extremely well Decoded Value: Yes

AMAS35 character 35. American history 1=Not at all ; 2=A little ; 3=Pretty well ; 4=Extremely well Decoded Value: Yes

AMAS36 character 36. American political leaders 1=Not at all ; 2=A little ; 3=Pretty well ; 4=Extremely well Decoded Value: Yes

AMAS37 character 37. national heroes from your native culture 1=Not at all ; 2=A little ; 3=Pretty well ; 4=Extremely well Decoded Value: Yes

AMAS38 character 38. popular television shows in your native language 1=Not at all ; 2=A little ; 3=Pretty well ; 4=Extremely well Decoded Value: Yes

AMAS39 character 39. popular newspapers and magazines in your native language 1=Not at all ; 2=A little ; 3=Pretty well ; 4=Extremely well Decoded Value: Yes

AMAS40 character 40. popular actors and actresses from your native culture 1=Not at all ; 2=A little ; 3=Pretty well ; 4=Extremely well Decoded Value: Yes

AMAS41 character 41. history of your native culture 1=Not at all ; 2=A little ; 3=Pretty well ; 4=Extremely well Decoded Value: Yes

AMAS42 character 42. political leaders from your native culture 1=Not at all ; 2=A little ; 3=Pretty well ; 4=Extremely well Decoded Value: Yes

ID numeric Record ID Range value: 71, ..., 106604

SITEID numeric Site ID Range value: 2, ..., 941

USERDATE User Create Date Date: YYY-MM-DD

USERDATE2 User Last Modify Date Date: YYY-MM-DD

DD_CRF_VERSION_LABEL character CRF Version Label , with options: v1

LANGUAGE_CODE character Language Code e=English Decoded Value: Yes

HAS_QC_ERROR character Has Quality Check Error 0=Does not have QC error or QC error has been approved ; 1=Has QC error Decoded Value: Yes

update_stamp POSIXct: YYY-MM-DD

Author

adni-data@googlegroups.com