Amyloid Disclosure Visit

Amyloid Disclosure Visit data. More information is available at https://adni.loni.usc.edu/data-samples/adni-data/

Usage

data(AMYDISC)

Format

A data frame with 301 observations and 34 variables.

Source

https://adni.loni.usc.edu/data-samples/adni-data/

Details

ORIGPROT character Original study protocol , with options: ADNI3, ADNI4, ADNIGO, ADNI1, ADNI2

COLPROT factor Study protocol of data collection , with factor levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4

PTID character Participant ID

RID numeric Participant roster ID Range value: 21, ..., 10471

VISCODE character Visit code , with options: 4_init, 4_bl, 4_m12

VISCODE2 character Translated visit code

VISDATE Assessment EXAMDATE when present; otherwise Registry EXAMDATE Date: YYY-MM-DD

DONE character Was an amyloid PET disclosure event conducted? 0=No; 1=Yes Decoded Value: Yes

DISC character Was amyloid PET result disclosed to the participant? 0=No; 1=Yes Decoded Value: Yes

DISCND character If no, please explain why: 1=Participant not psychologically well; 2=Participant's study partner not available (Study partners are required to attend for MCI and DEM participants); 3=Other Decoded Value: Yes

OUTCOME character What was the amyloid PET scan result? 0=Non-Elevated; 1=Elevated Decoded Value: Yes

SOURCE character In what setting was the disclosure conducted? 1=Remotely; 2=In-Person Decoded Value: Yes

PTAGE numeric How long did the disclosure visit last? Range value: 5, ..., 90

JOINDISC character Did someone join the participant at the visit? 0=No; 1=Yes Decoded Value: Yes

JOINNUM numeric If yes, how many others joined the participant at the visit? Range value: 1, ..., 2

JOINSPCY character If yes, was their relationship to the participant? 1=Partner/spouse; 2=Child/Child-In-Law; 3=Sibling; 4=Other relative; 5=Friend; 6=Other Decoded Value: Yes

PTGRASP character Please rate this participant's understanding of their amyloid PET scan result. 1=Inadequate understanding; 2=Marginal understanding; 3=Adequate understanding Decoded Value: Yes

SPGRASP character If there was an ADNI study partner present during this visit, please rate their understanding of the amyloid result. 1=Inadequate understanding; 2=Marginal understanding; 3=Adequate understanding; 4=Not applicable – no study partner was present Decoded Value: Yes

PTREACT character Which of the following terms best describes this participant's overall emotional reaction to this visit? 1=Overall positive; 2=Overall neutral; 3=Overall mixed (both positive and negative); 4=Overall negative Decoded Value: Yes

SPREACT character If there was an ADNI study partner present, which of the following terms best describes this study partner's overall emotional reaction to this visit? 1=Overall positive; 2=Overall neutral; 3=Overall mixed (both positive and negative); 4=Overall negative; 5=Not applicable – no study partner was present Decoded Value: Yes

TOPICS character Which of the following topics did you discuss with this participant? 1=Other risk factors (e.g., age, sex, family history); 10=Other; 2=Meaning of the result; 3=PET scan image (i.e., showed participant their image); 4=Mean cortical SUVR; 5=PET read CRF; 6=Emotional impact of learning this information; 7=Potential stigma when sharing this result with others; 8=Potential discrimination (e.g., employment, insurance, housing); 9=Novel therapies for AD Decoded Value: Yes

CHAL character Did you experience any challenges during this visit, including in your discussion of any of the above? 0=No; 1=Yes Decoded Value: Yes

PTREAD character How confident did you feel assessing this participant's readiness to receive their result? 1=Not at all confident; 2=A little confident; 3=Somewhat confident; 4=Confident; 5=Very confident Decoded Value: Yes

PTDISCL character How confident did you feel disclosing to this participant? 1=Not at all confident; 2=A little confident; 3=Somewhat confident; 4=Confident; 5=Very confident Decoded Value: Yes

PTQSTNS character How confident did you feel answering this participant's questions? 1=Not at all confident; 2=A little confident; 3=Somewhat confident; 4=Confident; 5=Very confident Decoded Value: Yes

VSTRESULT character Overall, how did this visit go? 1=Poor; 2=Fair; 3=Good; 4=Very good; 5=Excellent Decoded Value: Yes

ID numeric Record ID Range value: 3900, ..., 82122

SITEID numeric Site ID Range value: 2, ..., 941

USERDATE User Create Date Date: YYY-MM-DD

USERDATE2 User Last Modify Date Date: YYY-MM-DD

DD_CRF_VERSION_LABEL character CRF Version Label , with options: v1

LANGUAGE_CODE character Language Code e=English Decoded Value: Yes

HAS_QC_ERROR character Has Quality Check Error 0=Does not have QC error or QC error has been approved; 1=Has QC error Decoded Value: Yes

update_stamp POSIXct: YYY-MM-DD

Author

adni-data@googlegroups.com