AV-45 PET Scan Information

AV-45 PET Scan Information data. More information is available at https://adni.loni.usc.edu/data-samples/adni-data/

Usage

data(AV45META)

Format

A data frame with 3400 observations and 52 variables.

Source

https://adni.loni.usc.edu/data-samples/adni-data/

Details

ORIGPROT character Original study protocol , with options: ADNI1, ADNIGO, ADNI2

COLPROT factor Study protocol of data collection , with factor levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4

ID numeric Range value: 8, ..., 6410

PTID character Participant ID

RID numeric Participant roster ID Range value: 2, ..., 5296

SITEID numeric Site ID Range value: 1, ..., 61

VISCODE character Visit code

VISCODE2 character Translated visit code

USERDATE Date Date record created Date: YYY-MM-DD

USERDATE2 Date Date record last updated Date: YYY-MM-DD

RECNO numeric Range value: 1, ..., 3

PMCONDCT character Was the scan conducted? 1=Yes;0=No Decoded Value: Yes

PMREASON character Reason why the scan was not conducted:

ADNIGO: 1=Illness;2=Participant unavailable;3=Participant unwilling;4=Administrative problems;5=Withdrawn consent;6=Other (specify)

ADNI2: 1=Illness;2=Participant unavailable;3=Participant unwilling;4=Administrative problems;8=Software/scanner error; 5=Withdrawn consent;7=Not called for by the protocol;6=Other (specify)

Decoded Value: Yes

RADTRACER character If no, was radiotracer administered? 1=Yes;0=No Decoded Value: Yes

EXAMDATE Date Examination Date Date: YYY-MM-DD

PMSCANNER character GE 1=check here Decoded Value: Yes

PMGEMODEL character If GE, Scanner Model:

ADNIGO: 1=Advance;2=Discovery LS;3=Discovery ST;4=Discovery RX;5=Discovery STE/VCT

ADNI2: 1=Advance;2=Discovery LS;3=Discovery ST;4=Discovery RX;5=Discovery STE/VCT; 6=Discovery 600; 7=Discovery 690

Decoded Value: Yes

PMSIEMENS character Siemens 1=check here Decoded Value: Yes

PMSIEMODEL character If Siemens, Scanner Model: 1=ACCEL/EXACT;2=Biograph (Model 1023/1024);7=Biograph HiRes (Model 1080);3=BioGraph TruePoint (Model 1093/1094);4=BioGraph mCT;5=HR+;6=HRRT Decoded Value: Yes

PMPHILLIPS character Phillips 1=check here Decoded Value: Yes

PMPHMODEL character If Phillips, Scanner Model: 1=Allegro;2=Gemini;3=Gemini-GXL;4=Gemini-TF Decoded Value: Yes

PMQCTIME character Time of today's Scanner QC

PMFDGTIME character Time of Florbetapir F18 dose assay

PMFDGDOS numeric Florbetapir F18 dose assay Range value: 0.4, ..., 98

PMFDGVOL numeric Florbetapir F18 Volume Range value: 0, ..., 97

PMINJTIME character Time of Florbetapir F18 injection

PMSCTIME character Standard Emission Scan Start Time

PMVARIAT character Any variations from protocol during Florbetapir F18 uptake? 1=Yes; 0=No Decoded Value: Yes

PMPROTID character Predefined Acquisition Protocol ID

PMFRAME numeric For the standard late acquisition, which framing rate was used?

ADNIGO: 1=4 frames, 5 min/frame (4x300s);2=2 scans, 10 min each (2x600s) (only for BioGraph scanners without list-mode)

ADNI2: 1=4 frames, 5 min/frame (4x300s);2=2 scans, 10 min each (2x600s) (only for BioGraph scanners without list-mode);3=1 frame, 20 min (1x1200s) (only for the oldest BioGraph scanners: Models 1023 or 1024)

PMDEVIAT character If any deviations, describe:

PMMOTION character Subject motion problems: 1=Yes;0=No Decoded Value: Yes

PMMALFUN character Scanner malfunction 1=Yes;0=No Decoded Value: Yes

PMOTHER character Other protocol variations: 1=Yes;0=No Decoded Value: Yes

PMRECON character Check which of the following reconstructions was used:

ADNIGO: 2=FORE/2D-OSEM (Siemens);4=OSEM3D (Siemens) (If HHRT scanners using OP, please select OSEM3D);5=3D Iterative (GE);1=3D-Ramla (Philips);3=3D Back-projection (GE)

ADNI2: 2=FORE/2D-OSEM (Siemens);4=OSEM3D (Siemens) (If HHRT scanners using OP, please select OSEM3D);5=3D Iterative (GE);6=FORE Iterative (GE); 1=3D or LOR Ramla (Philips);3=3D Back-projection (GE)

Decoded Value: Yes

PMSUBSET character # subsets:

ADNIGO: 14=14;16=16;20=20;99=Other

ADNI2: 12=12;14=14;16=16;20=20;21=21;24=24;32=32;99=Other

Decoded Value: Yes

PMSUBSPE numeric If Other, specify: Range value: 12, ..., 32

PMITERAT character # iterations: 4=4;6=6;99=Other Decoded Value: Yes

PMITERSPE logical If Other, specify: Boolean value: TRUE/ FALSE

PMRAMLA numeric If 3D or LOR Ramla, Lambda= Range value: 0.008, ..., 1

PMSHARP character Was 'Smooth' parameter set to 'Sharp'? 1=Check here to confirm Decoded Value: Yes

PMFILTER character If 3D Back-Projection, Ramp filter? 1=Check here to confirm Decoded Value: Yes

PMMODEON character If FORE/2D-OSEM select one of the following:

ADNIGO: 1=Brain Mode "ON" for PET-only Siemens scanners;2=TRIM "ON" for PET/CT Siemens scanners (older software versions); 3=TRIM not available for PET/CT scanners (new software versions)

ADNI2: 1=Brain Mode 'ON' for PET-only Siemens scanners;2=TRIM 'ON' for PET/CT Siemens scanners (older software versions); 3=TRIM not available for PET/CT scanners (new software versions)

Decoded Value: Yes

PMSMOOTH character No post-process smoothing: 1=Check here to confirm Decoded Value: Yes

PMATTEN character Attenuation Correction: 1=CT; 2=Ge-68+Segmentation;3=Cs-137+Segmentation Decoded Value: Yes

PMTRNSFR character Was data transferred to LONI within 24 hours of scan? 1=Yes;0=No Decoded Value: Yes

PMTRNDATE character Transfer Date

PMTRNDATEYR_DRVD numeric Transfer Year (Derived from PMTRNDATE) Range value: 2001, ..., 2017

PMARCHIVE character Was all raw PET data archived locally to be able to do complete reconstruction of PET Scan if needed? 1=Yes;0=No Decoded Value: Yes

PMLPDONE character Was a Lumbar Puncture completed prior to the Florbetapir F18 scan? 1=Yes;0=No Decoded Value: Yes

PMLPINTER character If Yes, What was the interval between LP and Florbetapir F18? 1=less than 6 hours;2=6-12 hours;3=13-24 hours;4=25-48 hours;5=49-72 hours;6=more than 72 hours Decoded Value: Yes

update_stamp POSIXct POSIXct: YYY-MM-DD

Author

adni-data@googlegroups.com