Banner Alzheimer's Institute NMRC Summaries

Banner Alzheimer's Institute NMRC Summaries data. More information is available at https://adni.loni.usc.edu/data-samples/adni-data/

Usage

data(BAINMRC)

Format

A data frame with 2026 observations and 304 variables.

Source

https://adni.loni.usc.edu/data-samples/adni-data/

Details

ORIGPROT character Original study protocol , with options: ADNI1

RID numeric Range value: 2, ..., 1430

VISCODE character , with options: bl, m06, m12, m24, m36, m48, m60, m18

EXAMDATE Date: YYY-MM-DD

IMAGEUID numeric Range value: 25392, ..., 208245

HIPPL01 numeric Globally normalized CMRgl from Hippocampus_L Range value: 0.845947, ..., 1.57295

FRTSUPL01 numeric Globally normalized CMRgl from Frontal_sup_L Range value: 1.15639, ..., 2.1298

FRTMIDL01 numeric Globally normalized CMRgl from Frontal_mid_L Range value: 1.11277, ..., 2.12475

PARAHIPL01 numeric Globally normalized CMRgl from ParaHippocampal_L Range value: 0.912239, ..., 1.5574

FUSFRML01 numeric Globally normalized CMRgl from Fusiform_L Range value: 0.833843, ..., 1.61013

OCCMIDL01 numeric Globally normalized CMRgl from Occipital_mid_L Range value: 1.08684, ..., 1.97188

ANGULL01 numeric Globally normalized CMRgl from Angular_L Range value: 0.815963, ..., 2.18265

PARIINFL01 numeric Globally normalized CMRgl from Parietal_Inf_L Range value: 0.802001, ..., 1.88925

SUPMRGL01 numeric Globally normalized CMRgl from SupraMarginal_L Range value: 1.2772, ..., 1.92524

TMPMIDL01 numeric Globally normalized CMRgl from Temporal _mid_L Range value: 0.995531, ..., 1.7407

PRECUNL01 numeric Globally normalized CMRgl from Precuneus_L Range value: 1.29607, ..., 2.24714

CINGPSTL01 numeric Globally normalized CMRgl from Cingulum_Post_L Range value: 0.972679, ..., 2.25141

PARAHIPR01 numeric Globally normalized CMRgl from ParaHippocampal_R Range value: 0.806904, ..., 1.61033

FRTSUPR01 numeric Globally normalized CMRgl from Frontal_Sup_R Range value: 1.34728, ..., 2.17756

OCCMIDR01 numeric Globally normalized CMRgl from Occipital_mid_R Range value: 1.05336, ..., 1.97827

FUSFRMR01 numeric Globally normalized CMRgl from Fusiform_R Range value: 0.829141, ..., 1.71375

FRTMIDR01 numeric Globally normalized CMRgl from Frontal_mid_R Range value: 1.11686, ..., 1.79081

ANGULR01 numeric Globally normalized CMRgl from Angular_R Range value: 1.15124, ..., 2.26394

PARIINFR01 numeric Globally normalized CMRgl from Parietal_Inf_R Range value: 1.13354, ..., 2.0973

TMPMIDR01 numeric Globally normalized CMRgl from Temporal_mid_R Range value: 1.18349, ..., 2.17798

SUPMRGR01 numeric Globally normalized CMRgl from SupraMarginal_R Range value: 1.01858, ..., 1.62387

PRCUNSR01 numeric Globally normalized CMRgl from Prcuneus_R Range value: 1.32524, ..., 2.38231

HIPPR01 logical Globally normalized CMRgl from Hippocampus_R Boolean value: TRUE/ FALSE

LINGUALL01 logical Globally normalized CMRgl from Lingual_L Boolean value: TRUE/ FALSE

LINGUALR01 logical Globally normalized CMRgl from Lingual_R Boolean value: TRUE/ FALSE

CINGPSTR01 logical Globally normalized CMRgl from Cingulum_Post_R Boolean value: TRUE/ FALSE

FRTINFL01 logical Globally normalized CMRgl from Frontal_Inf_L Boolean value: TRUE/ FALSE

FRTINFR01 logical Globally normalized CMRgl from Frontal_Inf_R Boolean value: TRUE/ FALSE

PARISUPL01 logical Globally normalized CMRgl from Parietal_Sup_L Boolean value: TRUE/ FALSE

PARISUPR01 logical Globally normalized CMRgl from Parietal_Sup_R Boolean value: TRUE/ FALSE

INSULAL01 logical Globally normalized CMRgl from Insula_L Boolean value: TRUE/ FALSE

INSULAR01 logical Globally normalized CMRgl from Insula_R Boolean value: TRUE/ FALSE

CINGANTL01 logical Globally normalized CMRgl from Cingulum_Ant_L Boolean value: TRUE/ FALSE

CINGANTR01 logical Globally normalized CMRgl from Cingulum_Ant_R Boolean value: TRUE/ FALSE

CINGMIDL01 logical Globally normalized CMRgl from Cingulum_Mid_L Boolean value: TRUE/ FALSE

CINGMIDR01 logical Globally normalized CMRgl from Cingulum_Mid_R Boolean value: TRUE/ FALSE

TMPSUPL01 logical Globally normalized CMRgl from Temporal _Sup_L Boolean value: TRUE/ FALSE

TMPSUPR01 logical Globally normalized CMRgl from Temporal _Sup_R Boolean value: TRUE/ FALSE

TMPINFL01 logical Globally normalized CMRgl from Temporal_Inf_L Boolean value: TRUE/ FALSE

TMPINFR01 numeric Globally normalized CMRgl from Temporal_Inf_R Range value: 1.07602, ..., 1.93136

FRTSUPL02 numeric Globally normalized CMRgl from Frontal_Sup_L Range value: 1.04989, ..., 1.88073

FRTMIDL02 numeric Globally normalized CMRgl from Frontal_Mid_L Range value: 1.09853, ..., 2.05721

CINGPSTL02 numeric Globally normalized CMRgl from Cingulum_Post_L Range value: 1.44932, ..., 2.52221

CINGMIDL02 numeric Globally normalized CMRgl from Cingulum_Mid_L Range value: 1.46622, ..., 2.48963

FRTSMEDL02 numeric Globally normalized CMRgl from Frontal_Sup_Medial_L Range value: 1.09455, ..., 2.19089

FUSFRML02 numeric Globally normalized CMRgl from Fusiform_L Range value: 0.808137, ..., 1.655

TMPMIDL02 numeric Globally normalized CMRgl from Temporal_Mid_L Range value: 0.953942, ..., 1.80762

OCCMIDL02 numeric Globally normalized CMRgl from Occipital_Mid_L Range value: 1.09092, ..., 2.03661

TMPINFL02 numeric Globally normalized CMRgl from Temporal_Inf_L Range value: 0.821633, ..., 1.55902

PRECUNL02 numeric Globally normalized CMRgl from Precuneus_L Range value: 1.36488, ..., 2.25017

CINGANTL02 numeric Globally normalized CMRgl from Cingulum_Ant_L Range value: 1.12117, ..., 2.20225

FRTMIDOR02 numeric Globally normalized CMRgl from Frontal_Mid_Orb_R Range value: 1.21638, ..., 2.2145

FRTMIDR02 numeric Globally normalized CMRgl from Frontal_Mid_R Range value: 1.27595, ..., 2.08287

FRTSUPR02 numeric Globally normalized CMRgl from Frontal_Sup_R Range value: 1.32622, ..., 2.19736

ANGULR02 numeric Globally normalized CMRgl from Angular_R Range value: 1.12105, ..., 2.09255

ANGULL02 numeric Globally normalized CMRgl from Angular_L Range value: 0.874483, ..., 2.12218

PARIINFR02 numeric Globally normalized CMRgl from Parietal_Inf_R Range value: 1.17451, ..., 2.25359

TMPMIDR02 numeric Globally normalized CMRgl from Temporal_Mid_R Range value: 1.11721, ..., 2.02806

INSULAR02 numeric Globally normalized CMRgl from Insula_R Range value: 0.908476, ..., 1.75238

TMPPOSR02 numeric Globally normalized CMRgl from Temporal_Pole_Sup_R Range value: 0.831152, ..., 1.67344

CINGANTR02 numeric Globally normalized CMRgl from Cingulum_Ant_R Range value: 1.07998, ..., 1.99467

CINGMIDR02 numeric Globally normalized CMRgl from Cingulum_Mid_R Range value: 1.36282, ..., 2.47688

PRECUNR02 numeric Globally normalized CMRgl from Precuneus_R Range value: 1.362, ..., 2.48783

RECTUSR02 numeric Globally normalized CMRgl from Rectus_R Range value: 0.990824, ..., 1.71868

CINGPSTR02 logical Globally normalized CMRgl from Cingulum_Post_R Boolean value: TRUE/ FALSE

PARIINFL02 logical Globally normalized CMRgl from Parietal_Inf_L Boolean value: TRUE/ FALSE

INSULAL02 logical Globally normalized CMRgl from Insula_L Boolean value: TRUE/ FALSE

FRTINFL02 logical Globally normalized CMRgl from Frontal_Inf_L Boolean value: TRUE/ FALSE

FRTINFR02 logical Globally normalized CMRgl from Frontal_Inf_R Boolean value: TRUE/ FALSE

OCCMIDR02 logical Globally normalized CMRgl from Occipital_Mid_R Boolean value: TRUE/ FALSE

FRTMIDOL02 logical Globally normalized CMRgl from Frontal_Mid_Orb_Right Boolean value: TRUE/ FALSE

FRTSMEDR02 logical Globally normalized CMRgl from Frontal_Sup_Medial_R Boolean value: TRUE/ FALSE

FUSFMR02 logical Globally normalized CMRgl from Fusiform_R Boolean value: TRUE/ FALSE

RECTUSL02 logical Globally normalized CMRgl from Rectus_L Boolean value: TRUE/ FALSE

TMPPOSL02 logical Globally normalized CMRgl from Temporal_Pole_Sup_L Boolean value: TRUE/ FALSE

PARISUPL02 logical Globally normalized CMRgl from Parietal_Sup_L Boolean value: TRUE/ FALSE

PARISUPR02 logical Globally normalized CMRgl from Parietal_Sup_L Boolean value: TRUE/ FALSE

SUPMRGL02 logical Globally normalized CMRgl from SupraMarginal_L Boolean value: TRUE/ FALSE

SUPMRGR02 logical Globally normalized CMRgl from SupraMarginal_R Boolean value: TRUE/ FALSE

TMPINFR02 logical Globally normalized CMRgl from Temporal_Inf_L Boolean value: TRUE/ FALSE

FRTMIDL03 logical Frontal_Mid_L Boolean value: TRUE/ FALSE

FRTMIDR03 logical Frontal_Mid_R Boolean value: TRUE/ FALSE

FRTSUPL03 logical Frontal_Sup_L Boolean value: TRUE/ FALSE

FRTSUPR03 logical Frontal_Sup_R Boolean value: TRUE/ FALSE

FRTSMEDL03 logical Frontal_Sup_Medial_L Boolean value: TRUE/ FALSE

FRTSMEDR03 logical Frontal_Sup_Medial_R Boolean value: TRUE/ FALSE

TMPSUPL03 logical Temporal_Sup_L Boolean value: TRUE/ FALSE

TMPSUPR03 numeric Temporal_Sup_R Range value: 0.906624, ..., 1.63651

TMPMIDL03 numeric Temporal_Mid_L Range value: 0.908312, ..., 1.45835

TMPMIDR03 numeric Temporal_Mid_R Range value: 0.870199, ..., 1.58523

TMPINFL03 numeric Temporal_Inf_L Range value: 0.868289, ..., 1.42746

TMPINFR03 numeric Temporal_Inf_R Range value: 0.826913, ..., 1.52987

FUSFRML03 logical Fusiform_L Boolean value: TRUE/ FALSE

FUSFRMR03 logical Fusiform_R Boolean value: TRUE/ FALSE

LINGUALL03 logical Lingual_L Boolean value: TRUE/ FALSE

LINGUALR03 logical Lingual_R Boolean value: TRUE/ FALSE

PARAHIPL03 logical ParaHippocampal_L Boolean value: TRUE/ FALSE

PARAHIPR03 logical ParaHippocampal_R Boolean value: TRUE/ FALSE

PARISUPL03 logical Parietal_Sup_L Boolean value: TRUE/ FALSE

PARISUPR03 logical Parietal_Sup_R Boolean value: TRUE/ FALSE

PARIINFL03 numeric Parietal_Inf_L Range value: 1.14395, ..., 1.82124

PARIINFR03 numeric Parietal_Inf_R Range value: 1.11841, ..., 1.94766

PRECUNL03 numeric Precuneus_L Range value: 1.30438, ..., 2.46674

PRECUNR03 numeric Precuneus_R Range value: 1.5906, ..., 2.42236

CINGANT03 numeric Cingulum_Ant Range value: 0.967814, ..., 1.73057

CINGMID03 numeric Cingulum_Mid Range value: 1.16246, ..., 2.08253

CINGPST03 numeric Cingulum_Post Range value: 0.832101, ..., 2.00208

ANGULR03 numeric Angular_R Range value: 0.86501, ..., 1.59786

OCCMIDL03 logical Occipital_Mid_L Boolean value: TRUE/ FALSE

OCCMIDR03 logical Occipital_Mid_R Boolean value: TRUE/ FALSE

OCCSUPL03 logical Occipital_Sup_L Boolean value: TRUE/ FALSE

OCCSUPR03 logical Occipital_Sup_R Boolean value: TRUE/ FALSE

OCCINFL03 logical Occipital_Inf_L Boolean value: TRUE/ FALSE

OCCINFR03 logical Occipital_Inf_R Boolean value: TRUE/ FALSE

THALAML03 logical Thalamus_L Boolean value: TRUE/ FALSE

THALAMR03 logical Thalamus_R Boolean value: TRUE/ FALSE

FRTMEDL03 logical Frontal_Medial_Left Boolean value: TRUE/ FALSE

FRTMEDR03 logical Frontal_Medial_Right Boolean value: TRUE/ FALSE

SUPMRGL03 numeric SupraMarginal_L Range value: 0.672312, ..., 1.30923

SUPMRGR03 logical SupraMarginal_R Boolean value: TRUE/ FALSE

FRTMIDL04 logical Frontal_Mid_L Boolean value: TRUE/ FALSE

FRTMIDR04 logical Frontal_Mid_R Boolean value: TRUE/ FALSE

FRTSUPL04 logical Frontal_Sup_L Boolean value: TRUE/ FALSE

FRTSUPR04 logical Frontal_Sup_R Boolean value: TRUE/ FALSE

FRTSMEDL04 logical Frontal_Sup_Medial_L Boolean value: TRUE/ FALSE

FRTSMEDR04 logical Frontal_Sup_Medial_R Boolean value: TRUE/ FALSE

TMPSUPL04 logical Temporal_Sup_L Boolean value: TRUE/ FALSE

TMPSUPR04 numeric Temporal_Sup_R Range value: 1.03792, ..., 1.76477

TMPMIDL04 numeric Temporal_Mid_L Range value: 0.835183, ..., 1.68137

TMPMIDR04 numeric Temporal_Mid_R Range value: 0.891837, ..., 1.68217

TMPINFL04 numeric Temporal_Inf_L Range value: 0.80388, ..., 1.84787

TMPINFR04 numeric Temporal_Inf_R Range value: 0.927987, ..., 1.68568

FUSFRML04 logical Fusiform_L Boolean value: TRUE/ FALSE

FUSFRMR04 logical Fusiform_R Boolean value: TRUE/ FALSE

LINGUALL04 logical Lingual_L Boolean value: TRUE/ FALSE

LINGUALR04 logical Lingual_R Boolean value: TRUE/ FALSE

PARAHIPL04 logical ParaHippocampal_L Boolean value: TRUE/ FALSE

PARAHIPR04 logical ParaHippocampal_R Boolean value: TRUE/ FALSE

PARISUPL04 logical Parietal_Sup_L Boolean value: TRUE/ FALSE

PARISUPR04 logical Parietal_Sup_R Boolean value: TRUE/ FALSE

PARIINFL04 logical Parietal_Inf_L Boolean value: TRUE/ FALSE

PARIINFR04 logical Parietal_Inf_R Boolean value: TRUE/ FALSE

PRECUNL04 logical Precuneus_L Boolean value: TRUE/ FALSE

PRECUNR04 logical Precuneus_R Boolean value: TRUE/ FALSE

CINGANT04 logical Cingulum_Ant Boolean value: TRUE/ FALSE

CINGMID04 logical Cingulum_Mid Boolean value: TRUE/ FALSE

CINGPST04 numeric Cingulum_Post Range value: 1.06925, ..., 2.12651

OCCMIDL04 logical Occipital_Mid_L Boolean value: TRUE/ FALSE

OCCMIDR04 numeric Occipital_Mid_R Range value: 0.897243, ..., 1.78806

OCCSUPL04 logical Occipital_Sup_L Boolean value: TRUE/ FALSE

OCCSUPR04 logical Occipital_Sup_R Boolean value: TRUE/ FALSE

OCCINFL04 logical Occipital_Inf_L Boolean value: TRUE/ FALSE

OCCINFR04 logical Occipital_Inf_R Boolean value: TRUE/ FALSE

THALAML04 logical Thalamus_L Boolean value: TRUE/ FALSE

THALAMR04 logical Thalamus_R Boolean value: TRUE/ FALSE

FRTMEDL04 logical Frontal_Medial_Left Boolean value: TRUE/ FALSE

FRTMEDR04 logical Frontal_Medial_Right Boolean value: TRUE/ FALSE

SUPMRGL04 numeric SupraMarginal_L Range value: 1.12521, ..., 1.92945

SUPMRGR04 logical SupraMarginal_R Boolean value: TRUE/ FALSE

ANGULL04 logical Angular_L Boolean value: TRUE/ FALSE

ANGULR04 numeric Angular_R Range value: 0.843005, ..., 1.73865

FRTMIDL05 logical Frontal_Mid_L Boolean value: TRUE/ FALSE

FRTMIDR05 logical Frontal_Mid_R Boolean value: TRUE/ FALSE

FRTSUPL05 logical Frontal_Sup_L Boolean value: TRUE/ FALSE

FRTSUPR05 logical Frontal_Sup_R Boolean value: TRUE/ FALSE

FRTSMEDL05 logical Frontal_Sup_Medial_L Boolean value: TRUE/ FALSE

FRTSMEDR05 logical Frontal_Sup_Medial_R Boolean value: TRUE/ FALSE

TMPSUPL05 logical Temporal_Sup_L Boolean value: TRUE/ FALSE

TMPSUPR05 numeric Temporal_Sup_R Range value: 0.87772, ..., 1.34958

TMPMIDL05 numeric Temporal_Mid_L Range value: 1.23457, ..., 1.96356

TMPMIDR05 numeric Temporal_Mid_R Range value: 1.03621, ..., 1.64083

TMPINFL05 logical Temporal_Inf_L Boolean value: TRUE/ FALSE

TMPINFR05 numeric Temporal_Inf_R Range value: 1.05647, ..., 1.76632

FUSFRML05 logical Fusiform_L Boolean value: TRUE/ FALSE

FUSFRMR05 logical Fusiform_R Boolean value: TRUE/ FALSE

PARAHIPL05 logical ParaHippocampal_L Boolean value: TRUE/ FALSE

PARAHIPR05 logical ParaHippocampal_R Boolean value: TRUE/ FALSE

PARISUPL05 logical Parietal_Sup_L Boolean value: TRUE/ FALSE

PARISUPR05 logical Parietal_Sup_R Boolean value: TRUE/ FALSE

PARIINFL05 numeric Parietal_Inf_L Range value: 0.859358, ..., 1.60546

PARIINFR05 numeric Parietal_Inf_R Range value: 0.884295, ..., 1.61886

PRECUNL05 logical Precuneus_L Boolean value: TRUE/ FALSE

PRECUNR05 numeric Precuneus_R Range value: 1.13373, ..., 2.44158

CINGANT05 logical Cingulum_Ant Boolean value: TRUE/ FALSE

CINGMID05 logical Cingulum_Mid Boolean value: TRUE/ FALSE

CINGPST05 numeric Cingulum_Post Range value: 1.05241, ..., 2.07561

OCCMIDL05 logical Occipital_Mid_L Boolean value: TRUE/ FALSE

OCCMIDR05 logical Occipital_Mid_R Boolean value: TRUE/ FALSE

OCCSUPL05 logical Occipital_Sup_L Boolean value: TRUE/ FALSE

OCCSUPR05 logical Occipital_Sup_R Boolean value: TRUE/ FALSE

OCCINFL05 logical Occipital_Inf_L Boolean value: TRUE/ FALSE

OCCINFR05 logical Occipital_Inf_R Boolean value: TRUE/ FALSE

THALAML05 logical Thalamus_L Boolean value: TRUE/ FALSE

THALAMR05 logical Thalamus_R Boolean value: TRUE/ FALSE

FRTMEDL05 logical Frontal_Medial_Left Boolean value: TRUE/ FALSE

FRTMEDR05 logical Frontal_Medial_Right Boolean value: TRUE/ FALSE

ANGULR05 numeric Angular_Right Range value: 1.04344, ..., 1.56019

ANGULL05 logical Angular_Left Boolean value: TRUE/ FALSE

CALCARL05 numeric Calcarine_L Range value: 0.935195, ..., 1.8764

CALCARR05 logical Calcarine_R Boolean value: TRUE/ FALSE

LINGUALL05 numeric Lingual_L Range value: 0.945546, ..., 1.81036

LINGUALR05 logical Lingual_R Boolean value: TRUE/ FALSE

CEREB8L05 numeric Cerebelum_8_L Range value: 1.02268, ..., 1.69108

CEREB8R05 logical Cerebelum_8_R Boolean value: TRUE/ FALSE

CERBCR2L05 numeric Cerebelum_Crus2_L Range value: 1.02973, ..., 1.68651

CERBCR2R05 logical Cerebelum_Crus2_R Boolean value: TRUE/ FALSE

SUPMRGL05 numeric SupraMarginal_L Range value: 0.862619, ..., 1.29025

SUPMRGR05 logical SupraMarginal_R Boolean value: TRUE/ FALSE

FRTMIDL06 logical Frontal_Mid_L Boolean value: TRUE/ FALSE

FRTMIDR06 logical Frontal_Mid_R Boolean value: TRUE/ FALSE

FRTSUPL06 logical Frontal_Sup_L Boolean value: TRUE/ FALSE

FRTSUPR06 logical Frontal_Sup_R Boolean value: TRUE/ FALSE

FRTSMEDL06 logical Frontal_Sup_Medial_L Boolean value: TRUE/ FALSE

FRTSMEDR06 logical Frontal_Sup_Medial_R Boolean value: TRUE/ FALSE

TMPSUPL06 logical Temporal_Sup_L Boolean value: TRUE/ FALSE

TMPSUPR06 numeric Temporal_Sup_R Range value: 1.18306, ..., 1.75266

TMPMIDL06 numeric Temporal_Mid_L Range value: 0.969484, ..., 1.54322

TMPMIDR06 numeric Temporal_Mid_R Range value: 1.1327, ..., 1.65291

TMPINFL06 numeric Temporal_Inf_L Range value: 1.00196, ..., 1.86324

TMPINFR06 numeric Temporal_Inf_R Range value: 1.19813, ..., 1.97752

FUSFRML06 numeric Fusiform_L Range value: 1.07881, ..., 1.71197

FUSFRMR06 numeric Fusiform_R Range value: 0.88559, ..., 1.60234

LINGUALL06 logical Lingual_L Boolean value: TRUE/ FALSE

LINGUALR06 logical Lingual_R Boolean value: TRUE/ FALSE

PARAHIPL06 logical ParaHippocampal_L Boolean value: TRUE/ FALSE

PARAHIPR06 logical ParaHippocampal_R Boolean value: TRUE/ FALSE

PARISUPL06 logical Parietal_Sup_L Boolean value: TRUE/ FALSE

PARISUPR06 logical Parietal_Sup_R Boolean value: TRUE/ FALSE

PARIINFL06 logical Parietal_Inf_L Boolean value: TRUE/ FALSE

PARIINFR06 logical Parietal_Inf_R Boolean value: TRUE/ FALSE

PRECUNL06 logical Precuneus_L Boolean value: TRUE/ FALSE

PRECUNR06 logical Precuneus_R Boolean value: TRUE/ FALSE

CINGANT06 logical Cingulum_Ant Boolean value: TRUE/ FALSE

CINGMID06 logical Cingulum_Mid Boolean value: TRUE/ FALSE

CINGPST06 logical Cingulum_Post Boolean value: TRUE/ FALSE

OCCMIDL06 logical Occipital_Mid_L Boolean value: TRUE/ FALSE

OCCMIDR06 numeric Occipital_Mid_R Range value: 0.873268, ..., 1.60524

OCCSUPL06 logical Occipital_Sup_L Boolean value: TRUE/ FALSE

OCCSUPR06 logical Occipital_Sup_R Boolean value: TRUE/ FALSE

OCCINFL06 logical Occipital_Inf_L Boolean value: TRUE/ FALSE

OCCINFR06 logical Occipital_Inf_R Boolean value: TRUE/ FALSE

THALAML06 logical Thalamus_L Boolean value: TRUE/ FALSE

THALAMR06 logical Thalamus_R Boolean value: TRUE/ FALSE

FRTMEDL06 logical Frontal_Medial_Left Boolean value: TRUE/ FALSE

FRTMEDR06 logical Frontal_Medial_Right Boolean value: TRUE/ FALSE

HIPPL06 logical Hippocampus_L Boolean value: TRUE/ FALSE

HIPPR06 numeric Hippocampus_R Range value: 1.01451, ..., 1.59095

INDGLBNDX logical Individualized Global Index Boolean value: TRUE/ FALSE

INDTEPNDX logical Individualized Temporal index Boolean value: TRUE/ FALSE

PRECUNL07 numeric Precuneus_L Range value: 1.15795, ..., 2.14675

CINGPST07 numeric Cingulum_Post Range value: 0.912872, ..., 1.86358

PARIINFR07 numeric Parietal_Inf_R Range value: 0.86859, ..., 1.58379

TMPMIDL07 numeric Temporal_Mid_L Range value: 0.798215, ..., 1.408

CUNEUSL08 numeric Cuneus_L Range value: 1.17065, ..., 2.03573

TMPMIDR08 numeric Temporal_Mid_R Range value: 0.958192, ..., 1.76537

TMPINFL09 numeric Temporal_Inf_L Range value: 0.92968, ..., 1.66584

CINGPST09 numeric Cingulum_Post Range value: 1.03065, ..., 2.02593

PRECUNL09 numeric Precuneus_L Range value: 1.54287, ..., 2.68315

PARIINFR09 numeric Parietal_Inf_R Range value: 1.0283, ..., 1.90475

TMPINFL10 numeric Temporal_Inf_L Range value: 0.908405, ..., 1.65062

PRECUNL10 numeric Precuneus_L Range value: 1.31553, ..., 2.24931

TMPINFL11 numeric Temporal_Inf_L Range value: 1.1063, ..., 1.55387

OCCMIDL11 numeric Occipital_Mid_L Range value: 0.973173, ..., 1.55247

CINGPST11 numeric Cingulum_Post_L Range value: 0.857508, ..., 1.87147

CINGMID11 numeric Cingulum_Post_R/Cingulum_Mid_R Range value: 1.03255, ..., 1.87679

TMPPOSR12 numeric Temporal_Pole_Sup_R Range value: 0.45133, ..., 0.936241

CINGPSTR12 numeric Cingulum_Post_R/Cingulum_Mid_R Range value: 0.424733, ..., 0.949642

FROITRN001 numeric Training group, all subjects Range value: 0.74025, ..., 1.24215

FROITRN002 numeric Training group, exclude HRRT and BioGraph HiRez Range value: 0.796567, ..., 1.23335

FROITST001 numeric Testing group, all scans Range value: 0.636083, ..., 1.14834

FROITST002 numeric Testing group, exclude HRRT and BioGraph HiRez Range value: 0.813588, ..., 1.10135

FROIWHL001 numeric All subjects Range value: 0.700387, ..., 1.80342

FROIWHL002 numeric Exclude HRRT and BioGraph HiRez Range value: 0.751776, ..., 1.63939

FRPMSK001 numeric Exclude HRRT and BioGraph HiRez Range value: 0.788669, ..., 1.40005

FRPMSK002 numeric All subjects Range value: 0.788669, ..., 1.40442

FRTMSK003 numeric Exclude HRRT and BioGraph HiRez Range value: 0.848437, ..., 1.46291

FRTMSK004 numeric All subjects Range value: 0.816065, ..., 1.51455

FRTMSK005 numeric All subjects Range value: 0.781659, ..., 1.30399

FRTMSK006 numeric All subjects Range value: 0.626235, ..., 1.11443

FRTMSK007 numeric All subjects Range value: 0.74963, ..., 1.43373

FRTMSK008 numeric Exclude HRRT and BioGraph HiRez Range value: 0.715832, ..., 1.39802

FRTMSK009 numeric All subjects Range value: 0.597849, ..., 1.17415

FRTMSK010 numeric Exclude HRRT and BioGraph HiRez Range value: 0.558626, ..., 1.13881

FRTMSK011 numeric All subjects Range value: 0.868071, ..., 1.43818

IPCA numeric percent annualized whole brain atrophy detected by Iterative Principal component Analysis (IPCA) Range value: -0.114104, ..., 0.869081

HCISUB1 numeric hypometabolic convergence index for 74 AD subjects Range value: 4.03, ..., 28.16

HCI99 numeric hypometabolic convergence index for normal controls, stable MCI, MCI converter and ADs Range value: 2.59, ..., 24.09

HCI numeric hypometabolic convergence index for ADNI1 subjects Range value: 0.45, ..., 34.6

NL1HIPP numeric difference of globally normalized CMRgl from Baseline to 12 months and 12 months to 24 months from Hippocampus for 58 AD patients Range value: -0.149295, ..., 0.201397

NL1PTEMP numeric difference of globally normalized CMRgl from Baseline to 12 months and 12 months to 24 months from the parietotemporal region for 58 AD patients Range value: -0.148037, ..., 0.24458

NL1PSTCG numeric difference of globally normalized CMRgl from Baseline to 12 months and 12 months to 24 months from the posterior cingulate for for 58 AD patients Range value: -0.172309, ..., 0.339278

NL2PSTCG numeric difference of globally normalized CMRgl from Baseline to 12 months and 12 months to 24 months from Posterior Cingulate in 129 MCIs. Range value: -0.208048, ..., 0.243036

NL2PARI numeric difference of globally normalized CMRgl from 12 months to 24 months from the inferior parietal region in 129 MCI subjects Range value: -0.118032, ..., 0.438594

NL3FRONT numeric difference of globally normalized CMRgl from Baseline to 12 months and 12 months to 24 months from superior frontal region of 74 normal controls. Range value: -0.169811, ..., 0.228805

DECLADFR numeric difference of globally normalized CMRgl from Baseline to 12 months from the superior frontal region of 74 AD patients Range value: -0.132305, ..., 0.216164

DECLADTMP numeric CMRgl from Baseline from the inferior temporal region of 74 AD patients Range value: 0.905927, ..., 1.77742

LONGHCI logical Extension of HCI to longitudinal data Boolean value: TRUE/ FALSE

BSLVBM logical VBManalysis of structural MRI data Boolean value: TRUE/ FALSE

update_stamp POSIXct: YYY-MM-DD

Author

adni-data@googlegroups.com