Brain Health Registry (BHR) SP CAREGIVER BURDEN

Brain Health Registry (BHR) SP CAREGIVER BURDEN data. More information is available at https://adni.loni.usc.edu/data-samples/adni-data/

Usage

data(BHR_SP_CAREGIVER_BURDEN)

Format

A data frame with 37 observations and 133 variables.

Source

https://adni.loni.usc.edu/data-samples/adni-data/

Details

AboutRID numeric Range value: 2201, ..., 6644

StudyPartnerID numeric Range value: 2509, ..., 6051

AboutPTID character

Timepoint character , with options: sp-m00, sp-m12, sp-m18, sp-m06

CollectedDateTime POSIXct: YYY-MM-DD

CollectedDate_DRVD Date: YYY-MM-DD

latest numeric Range value: 0, ..., 1

QID13 numeric Are you the primary caregiver for YOUR STUDY PARTNER, who needs extra assistance due to a disability, condition, or illness? We define a primary caregiver as the person who is primarily responsible for helping make medical care decisions, is emotionally invested in the person's care, and provides help with daily activities (almost daily). Daily caregiving activities may include helping with medical or therapeutic treatment, preparing meal, and/or helping with dressing and/or bathing. Caregivers may be caring for a loved with dementia, or be the parent to a child with development delays or burdensome health conditions, for example. 1=Yes, I am a caregiver ; 2=No, I am not a caregiver ; 3=I am not a caregiver based on this definition, but I do help someone with age-related issues

QID53 numeric Are you the primary caregiver for YOUR STUDY PARTNER, who needs extra assistance due to a disability, condition, or illness? We define a primary caregiver as the person who is primarily responsible for helping make medical care decisions, is emotionally invested in the person's care, and provides help with daily activities (or almost daily). Daily caregiving activities may include helping with medical or therapeutic treatments, preparing meals, and/or helping with dressing and/or bathing. Caregivers may be caring for a loved one with dementia, or be the parents to a child with developmental delays or burdensome health conditions, for example. 1=Yes, I am a caregiver for MY STUDY PARTNER ; 2=No, I am not a caregiver for MY STUDY PARTNER ; 3=I am not a caregiver based on this definition, but I do help MY STUDY PARTNER with some health issues

QID31 numeric Because you do not identify as a primary caregiver for YOUR STUDY PARTNER, we have no more questions for you. If you become a primary caregiver at a later time, you will have the opportunity to let us know at a future visit. Range value: 1, ..., 1

QID32 numeric Are you the primary caregiver for more than one person? 1=Yes ; 2=No

QID55 numeric Are you caregiving for more than one person? 1=Yes ; 3=No

QID24_13 numeric What is the primary condition that requires YOUR STUDY PARTNER to need your care? If YOUR STUDY PARTNER has multiple conditions, please select all that apply.... Alzheimer's disease 1=Yes ; =No

QID24_14 numeric What is the primary condition that requires YOUR STUDY PARTNER to need your care? If YOUR STUDY PARTNER has multiple conditions, please select all that apply.... Mild Cognitive Impairment (MCI) 1=Yes ; =No

QID24_15 logical What is the primary condition that requires YOUR STUDY PARTNER to need your care? If YOUR STUDY PARTNER has multiple conditions, please select all that apply.... Frontotemporal Dementia 1=Yes ; =No

QID24_16 logical What is the primary condition that requires YOUR STUDY PARTNER to need your care? If YOUR STUDY PARTNER has multiple conditions, please select all that apply.... Other dementia (e.g. Vascular dementia, Dementia with Lewy Bodies) 1=Yes ; =No

QID24_7 logical What is the primary condition that requires YOUR STUDY PARTNER to need your care? If YOUR STUDY PARTNER has multiple conditions, please select all that apply.... Dementia 1=Yes ; =No

QID24_17 logical What is the primary condition that requires YOUR STUDY PARTNER to need your care? If YOUR STUDY PARTNER has multiple conditions, please select all that apply.... Brain disease that affects movement (e.g. Parkinson's, Huntington's, ALS, Motor Neuron Disease, Multiple Sclerosis) 1=Yes ; =No

QID24_18 numeric What is the primary condition that requires YOUR STUDY PARTNER to need your care? If YOUR STUDY PARTNER has multiple conditions, please select all that apply.... Epilepsy 1=Yes ; =No

QID24_19 logical What is the primary condition that requires YOUR STUDY PARTNER to need your care? If YOUR STUDY PARTNER has multiple conditions, please select all that apply.... Autism 1=Yes ; =No

QID24_20 logical What is the primary condition that requires YOUR STUDY PARTNER to need your care? If YOUR STUDY PARTNER has multiple conditions, please select all that apply.... Psychiatric illness (e.g. schizophrenia, depression, bipolar disorder, anxiety disorder) 1=Yes ; =No

QID24_9 logical What is the primary condition that requires YOUR STUDY PARTNER to need your care? If YOUR STUDY PARTNER has multiple conditions, please select all that apply.... Mental or Psychiatric Illness 1=Yes ; =No

QID24_21 logical What is the primary condition that requires YOUR STUDY PARTNER to need your care? If YOUR STUDY PARTNER has multiple conditions, please select all that apply.... Intellectual Disability 1=Yes ; =No

QID24_6 logical What is the primary condition that requires YOUR STUDY PARTNER to need your care? If YOUR STUDY PARTNER has multiple conditions, please select all that apply.... Mental Disability or Impairment 1=Yes ; =No

QID24_8 logical What is the primary condition that requires YOUR STUDY PARTNER to need your care? If YOUR STUDY PARTNER has multiple conditions, please select all that apply.... Brain-related Injury (e.g. tumor, stroke, traumatic brain injury) 1=Yes ; =No

QID24_1 numeric What is the primary condition that requires YOUR STUDY PARTNER to need your care? If YOUR STUDY PARTNER has multiple conditions, please select all that apply.... Cancer 1=Yes ; =No

QID24_2 logical What is the primary condition that requires YOUR STUDY PARTNER to need your care? If YOUR STUDY PARTNER has multiple conditions, please select all that apply.... Diabetes 1=Yes ; =No

QID24_3 logical What is the primary condition that requires YOUR STUDY PARTNER to need your care? If YOUR STUDY PARTNER has multiple conditions, please select all that apply.... Heart Disease 1=Yes ; =No

QID24_4 logical What is the primary condition that requires YOUR STUDY PARTNER to need your care? If YOUR STUDY PARTNER has multiple conditions, please select all that apply.... Other Chronic Illness 1=Yes ; =No

QID24_5 numeric What is the primary condition that requires YOUR STUDY PARTNER to need your care? If YOUR STUDY PARTNER has multiple conditions, please select all that apply.... Physical Disability or Injury 1=Yes ; =No

QID24_11 numeric What is the primary condition that requires YOUR STUDY PARTNER to need your care? If YOUR STUDY PARTNER has multiple conditions, please select all that apply.... Age-Related Issues 1=Yes ; =No

QID24_10 logical What is the primary condition that requires YOUR STUDY PARTNER to need your care? If YOUR STUDY PARTNER has multiple conditions, please select all that apply.... Other, please specify: 1=Yes ; =No

QID24_10_TEXT logical What is the primary condition that requires YOUR STUDY PARTNER to need your care? If YOUR STUDY PARTNER has multiple conditions, please select all that apply.... Other, please specify: 1=Yes ; =No

QID25_1_TEXT numeric For how long have you been caregiving? For example: 1 year, 3 months, 0 years, 7 months, 9 years, 0 months... Years Range value: 0, ..., 10

QID25_2_TEXT numeric For how long have you been caregiving? For example: 1 year, 3 months, 0 years, 7 months, 9 years, 0 months... Months Range value: 0, ..., 8

QID33 numeric Do you live with YOUR STUDY PARTNER? 1=Yes ; 2=No

QID34 numeric What is your employment status? 1=Employed full-time ; 2=Employed part-time ; 3=Full-time student ; 4=Part-time student ; 5=Unemployed ; 6=Retired

QID35 numeric Did you have to leave your paid job, cut down on your hours, or drop out of school due to your caregiving responsibilities? 1=Yes ; 2=No

QID26_TEXT numeric In addition to your role as caregiver, how many hours per week do you work at a paid job or attend school? If you do not have a job or go to school, please write 0. Range value: 0, ..., 50

QID27_TEXT numeric How many hours per week do you spend caring for YOUR STUDY PARTNER? We know this may be difficult to estimate, just give your best guess. If this question is not relevant to you, please enter 0. Range value: 0, ..., 8

QID38_1_1 numeric Has it occurred in the past week?... Asking the same question over and over 2=No ; 1=Yes

QID38_1_2 numeric Has it occurred in the past week?... Trouble remembering recent events (i.e. items in newspaper or TV) 2=No ; 1=Yes

QID38_1_3 numeric Has it occurred in the past week?... Trouble remembering significant past events 2=No ; 1=Yes

QID38_1_4 numeric Has it occurred in the past week?... Losing or misplacing things 2=No ; 1=Yes

QID38_1_5 numeric Has it occurred in the past week?... Forgetting what day it is 2=No ; 1=Yes

QID38_1_6 numeric Has it occurred in the past week?... Starting, but not finishing, things 2=No ; 1=Yes

QID38_2_1 numeric How much has this bothered or upset you when it happened?... Asking the same question over and over 0=Not At All (0) ; 1=A Little (1) ; 2=Moderately (2) ; 3=Very Much (3) ; 4=Extremely (4) ; 5=Has NOT occurred in the past week

QID38_2_2 numeric How much has this bothered or upset you when it happened?... Trouble remembering recent events (i.e. items in newspaper or TV) 0=Not At All (0) ; 1=A Little (1) ; 2=Moderately (2) ; 3=Very Much (3) ; 4=Extremely (4) ; 5=Has NOT occurred in the past week

QID38_2_3 numeric How much has this bothered or upset you when it happened?... Trouble remembering significant past events 0=Not At All (0) ; 1=A Little (1) ; 2=Moderately (2) ; 3=Very Much (3) ; 4=Extremely (4) ; 5=Has NOT occurred in the past week

QID38_2_4 numeric How much has this bothered or upset you when it happened?... Losing or misplacing things 0=Not At All (0) ; 1=A Little (1) ; 2=Moderately (2) ; 3=Very Much (3) ; 4=Extremely (4) ; 5=Has NOT occurred in the past week

QID38_2_5 numeric How much has this bothered or upset you when it happened?... Forgetting what day it is 0=Not At All (0) ; 1=A Little (1) ; 2=Moderately (2) ; 3=Very Much (3) ; 4=Extremely (4) ; 5=Has NOT occurred in the past week

QID38_2_6 numeric How much has this bothered or upset you when it happened?... Starting, but not finishing, things 0=Not At All (0) ; 1=A Little (1) ; 2=Moderately (2) ; 3=Very Much (3) ; 4=Extremely (4) ; 5=Has NOT occurred in the past week

QID42_1_1 numeric Has it occurred in the past week?... Difficulty concentrating on a task 2=No ; 1=Yes

QID42_1_2 numeric Has it occurred in the past week?... Destroying property 2=No ; 1=Yes

QID42_1_3 numeric Has it occurred in the past week?... Doing things that embarrass you 2=No ; 1=Yes

QID42_1_4 numeric Has it occurred in the past week?... Waking you or other family members up at night 2=No ; 1=Yes

QID42_1_5 numeric Has it occurred in the past week?... Talking loudly and rapidly 2=No ; 1=Yes

QID42_1_25 numeric Has it occurred in the past week?... Appeared anxious or worried 2=No ; 1=Yes

QID42_2_1 numeric How much has this bothered or upset you when it happened?... Difficulty concentrating on a task 0=Not At All (0) ; 1=A Little (1) ; 2=Moderately (2) ; 3=Very Much (3) ; 4=Extremely (4) ; 5=Has NOT occurred in the past week

QID42_2_2 numeric How much has this bothered or upset you when it happened?... Destroying property 0=Not At All (0) ; 1=A Little (1) ; 2=Moderately (2) ; 3=Very Much (3) ; 4=Extremely (4) ; 5=Has NOT occurred in the past week

QID42_2_3 numeric How much has this bothered or upset you when it happened?... Doing things that embarrass you 0=Not At All (0) ; 1=A Little (1) ; 2=Moderately (2) ; 3=Very Much (3) ; 4=Extremely (4) ; 5=Has NOT occurred in the past week

QID42_2_4 numeric How much has this bothered or upset you when it happened?... Waking you or other family members up at night 0=Not At All (0) ; 1=A Little (1) ; 2=Moderately (2) ; 3=Very Much (3) ; 4=Extremely (4) ; 5=Has NOT occurred in the past week

QID42_2_5 numeric How much has this bothered or upset you when it happened?... Talking loudly and rapidly 0=Not At All (0) ; 1=A Little (1) ; 2=Moderately (2) ; 3=Very Much (3) ; 4=Extremely (4) ; 5=Has NOT occurred in the past week

QID42_2_25 numeric How much has this bothered or upset you when it happened?... Appeared anxious or worried 0=Not At All (0) ; 1=A Little (1) ; 2=Moderately (2) ; 3=Very Much (3) ; 4=Extremely (4) ; 5=Has NOT occurred in the past week

QID43_1_1 numeric Has it occurred in the past week?... Engaging in behavior that is potentially dangerous to self or others 2=No ; 1=Yes

QID43_1_2 numeric Has it occurred in the past week?... Threatened to hurt oneself 2=No ; 1=Yes

QID43_1_3 numeric Has it occurred in the past week?... Threatened to hurt others 2=No ; 1=Yes

QID43_1_4 numeric Has it occurred in the past week?... Aggressive to others verbally 2=No ; 1=Yes

QID43_1_5 numeric Has it occurred in the past week?... Appeared sad or depressed 2=No ; 1=Yes

QID43_1_6 numeric Has it occurred in the past week?... Expressing feelings of hopelessness or sadness about the future 2=No ; 1=Yes

QID43_2_1 numeric How much has this bothered or upset you when it happened?... Engaging in behavior that is potentially dangerous to self or others 0=Not At All (0) ; 1=A Little (1) ; 2=Moderately (2) ; 3=Very Much (3) ; 4=Extremely (4) ; 5=Has NOT occurred in the past week

QID43_2_2 numeric How much has this bothered or upset you when it happened?... Threatened to hurt oneself 0=Not At All (0) ; 1=A Little (1) ; 2=Moderately (2) ; 3=Very Much (3) ; 4=Extremely (4) ; 5=Has NOT occurred in the past week

QID43_2_3 numeric How much has this bothered or upset you when it happened?... Threatened to hurt others 0=Not At All (0) ; 1=A Little (1) ; 2=Moderately (2) ; 3=Very Much (3) ; 4=Extremely (4) ; 5=Has NOT occurred in the past week

QID43_2_4 numeric How much has this bothered or upset you when it happened?... Aggressive to others verbally 0=Not At All (0) ; 1=A Little (1) ; 2=Moderately (2) ; 3=Very Much (3) ; 4=Extremely (4) ; 5=Has NOT occurred in the past week

QID43_2_5 numeric How much has this bothered or upset you when it happened?... Appeared sad or depressed 0=Not At All (0) ; 1=A Little (1) ; 2=Moderately (2) ; 3=Very Much (3) ; 4=Extremely (4) ; 5=Has NOT occurred in the past week

QID43_2_6 numeric How much has this bothered or upset you when it happened?... Expressing feelings of hopelessness or sadness about the future 0=Not At All (0) ; 1=A Little (1) ; 2=Moderately (2) ; 3=Very Much (3) ; 4=Extremely (4) ; 5=Has NOT occurred in the past week

QID44_1_1 numeric Has it occurred in the past week?... Crying and tearfulness 2=No ; 1=Yes

QID44_1_2 numeric Has it occurred in the past week?... Commenting about death of self or others 2=No ; 1=Yes

QID44_1_3 numeric Has it occurred in the past week?... Talking about feeling lonely 2=No ; 1=Yes

QID44_1_4 numeric Has it occurred in the past week?... Comments about feeling worthless or being a burden to others 2=No ; 1=Yes

QID44_1_5 numeric Has it occurred in the past week?... Comments about feeling like a failure, or about not having any worthwhile accomplishments in life 2=No ; 1=Yes

QID44_1_6 numeric Has it occurred in the past week?... Arguing, irritability, and/or complaining 2=No ; 1=Yes

QID44_2_1 numeric How much has this bothered or upset you when it happened?... Crying and tearfulness 0=Not At All (0) ; 1=A Little (1) ; 2=Moderately (2) ; 3=Very Much (3) ; 4=Extremely (4) ; 5=Has NOT occurred in the past week

QID44_2_2 numeric How much has this bothered or upset you when it happened?... Commenting about death of self or others 0=Not At All (0) ; 1=A Little (1) ; 2=Moderately (2) ; 3=Very Much (3) ; 4=Extremely (4) ; 5=Has NOT occurred in the past week

QID44_2_3 numeric How much has this bothered or upset you when it happened?... Talking about feeling lonely 0=Not At All (0) ; 1=A Little (1) ; 2=Moderately (2) ; 3=Very Much (3) ; 4=Extremely (4) ; 5=Has NOT occurred in the past week

QID44_2_4 numeric How much has this bothered or upset you when it happened?... Comments about feeling worthless or being a burden to others 0=Not At All (0) ; 1=A Little (1) ; 2=Moderately (2) ; 3=Very Much (3) ; 4=Extremely (4) ; 5=Has NOT occurred in the past week

QID44_2_5 numeric How much has this bothered or upset you when it happened?... Comments about feeling like a failure, or about not having any worthwhile accomplishments in life 0=Not At All (0) ; 1=A Little (1) ; 2=Moderately (2) ; 3=Very Much (3) ; 4=Extremely (4) ; 5=Has NOT occurred in the past week

QID44_2_6 numeric How much has this bothered or upset you when it happened?... Arguing, irritability, and/or complaining 0=Not At All (0) ; 1=A Little (1) ; 2=Moderately (2) ; 3=Very Much (3) ; 4=Extremely (4) ; 5=Has NOT occurred in the past week

QID19_1 numeric As a caregiver do YOU feel...... That because of the time you spend caregiving that you don't have enough time for yourself? 0=Never (0) ; 1=Rarely (1) ; 2=Sometimes (2) ; 3=Quite Frequently (3) ; 4=Nearly Always (4)

QID19_2 numeric As a caregiver do YOU feel...... Stressed between caregiving and trying to meet other responsibilities (work/ family)? 0=Never (0) ; 1=Rarely (1) ; 2=Sometimes (2) ; 3=Quite Frequently (3) ; 4=Nearly Always (4)

QID19_3 numeric As a caregiver do YOU feel...... Angry when you are around YOUR STUDY PARTNER? 0=Never (0) ; 1=Rarely (1) ; 2=Sometimes (2) ; 3=Quite Frequently (3) ; 4=Nearly Always (4)

QID19_4 numeric As a caregiver do YOU feel...... That YOUR STUDY PARTNER currently affects your relationship with family members in a negative way? 0=Never (0) ; 1=Rarely (1) ; 2=Sometimes (2) ; 3=Quite Frequently (3) ; 4=Nearly Always (4)

QID19_5 numeric As a caregiver do YOU feel...... Strained when you are around YOUR STUDY PARTNER? 0=Never (0) ; 1=Rarely (1) ; 2=Sometimes (2) ; 3=Quite Frequently (3) ; 4=Nearly Always (4)

QID19_6 numeric As a caregiver do YOU feel...... That your health has suffered because of caregiving? 0=Never (0) ; 1=Rarely (1) ; 2=Sometimes (2) ; 3=Quite Frequently (3) ; 4=Nearly Always (4)

QID45_1 numeric As a caregiver do YOU feel...... That you don't have as much privacy as you would like because of caregiving? 0=Never (0) ; 1=Rarely (1) ; 2=Sometimes (2) ; 3=Quite Frequently (3) ; 4=Nearly Always (4)

QID45_2 numeric As a caregiver do YOU feel...... That your social life has suffered because you are caring for YOUR STUDY PARTNER? 0=Never (0) ; 1=Rarely (1) ; 2=Sometimes (2) ; 3=Quite Frequently (3) ; 4=Nearly Always (4)

QID45_3 numeric As a caregiver do YOU feel...... That you have lost control of your life since you began providing care for YOUR STUDY PARTNER? 0=Never (0) ; 1=Rarely (1) ; 2=Sometimes (2) ; 3=Quite Frequently (3) ; 4=Nearly Always (4)

QID45_4 numeric As a caregiver do YOU feel...... Uncertain about what to do about YOUR STUDY PARTNER? 0=Never (0) ; 1=Rarely (1) ; 2=Sometimes (2) ; 3=Quite Frequently (3) ; 4=Nearly Always (4)

QID45_5 numeric As a caregiver do YOU feel...... You should be doing more for YOUR STUDY PARTNER? 0=Never (0) ; 1=Rarely (1) ; 2=Sometimes (2) ; 3=Quite Frequently (3) ; 4=Nearly Always (4)

QID45_6 numeric As a caregiver do YOU feel...... You could do a better job caring for YOUR STUDY PARTNER? 0=Never (0) ; 1=Rarely (1) ; 2=Sometimes (2) ; 3=Quite Frequently (3) ; 4=Nearly Always (4)

QID20_1 numeric As a caregiver, do YOU feel...... I'm not getting enough sleep 0=Never (0) ; 1=Rarely (1) ; 2=Sometimes (2) ; 3=Quite Frequently (3) ; 4=Nearly Always (4)

QID20_2 numeric As a caregiver, do YOU feel...... My health has suffered 0=Never (0) ; 1=Rarely (1) ; 2=Sometimes (2) ; 3=Quite Frequently (3) ; 4=Nearly Always (4)

QID20_3 numeric As a caregiver, do YOU feel...... Caregiving has made me physically sick 0=Never (0) ; 1=Rarely (1) ; 2=Sometimes (2) ; 3=Quite Frequently (3) ; 4=Nearly Always (4)

QID20_4 numeric As a caregiver, do YOU feel...... I'm physically tired 0=Never (0) ; 1=Rarely (1) ; 2=Sometimes (2) ; 3=Quite Frequently (3) ; 4=Nearly Always (4)

QID20_5 numeric As a caregiver, do YOU feel...... I have headaches or other pains 0=Never (0) ; 1=Rarely (1) ; 2=Sometimes (2) ; 3=Quite Frequently (3) ; 4=Nearly Always (4)

QID21_1 numeric Select the response that best describes YOUR life.... I wish I were free to lead a life of my own 1=Strongly disagree (1) ; 2=Disagree (2) ; 3=Neither agree nor disagree (3) ; 4=Agree (4) ; 5=Strongly Agree (5)

QID21_2 numeric Select the response that best describes YOUR life.... I feel trapped by MY STUDY PARTNER'S condition 1=Strongly disagree (1) ; 2=Disagree (2) ; 3=Neither agree nor disagree (3) ; 4=Agree (4) ; 5=Strongly Agree (5)

QID21_3 numeric Select the response that best describes YOUR life.... I wish I could just run away 1=Strongly disagree (1) ; 2=Disagree (2) ; 3=Neither agree nor disagree (3) ; 4=Agree (4) ; 5=Strongly Agree (5)

QID29_1 numeric Select the response that best describes YOUR life.... Caring for MY STUDY PARTNER is important to me 1=Strongly disagree (1) ; 2=Disagree (2) ; 3=Neither agree nor disagree (3) ; 4=Agree (4) ; 5=Strongly agree (5)

QID29_2 numeric Select the response that best describes YOUR life.... I enjoy caring for MY STUDY PARTNER 1=Strongly disagree (1) ; 2=Disagree (2) ; 3=Neither agree nor disagree (3) ; 4=Agree (4) ; 5=Strongly agree (5)

QID29_3 numeric Select the response that best describes YOUR life.... Caring for MY STUDY PARTNER makes me feel good 1=Strongly disagree (1) ; 2=Disagree (2) ; 3=Neither agree nor disagree (3) ; 4=Agree (4) ; 5=Strongly agree (5)

QID29_4 numeric Select the response that best describes YOUR life.... I feel privileged to care for MY STUDY PARTNER 1=Strongly disagree (1) ; 2=Disagree (2) ; 3=Neither agree nor disagree (3) ; 4=Agree (4) ; 5=Strongly agree (5)

QID29_5 numeric Select the response that best describes YOUR life.... I resent having to care for MY STUDY PARTNER 1=Strongly disagree (1) ; 2=Disagree (2) ; 3=Neither agree nor disagree (3) ; 4=Agree (4) ; 5=Strongly agree (5)

QID29_6 numeric Select the response that best describes YOUR life.... I really want to care for MY STUDY PARTNER 1=Strongly disagree (1) ; 2=Disagree (2) ; 3=Neither agree nor disagree (3) ; 4=Agree (4) ; 5=Strongly agree (5)

QID29_7 logical Select the response that best describes YOUR life.... MY STUDY PARTNER shows real appreciation of my care 1=Strongly disagree (1) ; 2=Disagree (2) ; 3=Neither agree nor disagree (3) ; 4=Agree (4) ; 5=Strongly agree (5)

QID29_8 logical Select the response that best describes YOUR life.... Nothing I do seems to please MY STUDY PARTNER 1=Strongly disagree (1) ; 2=Disagree (2) ; 3=Neither agree nor disagree (3) ; 4=Agree (4) ; 5=Strongly agree (5)

QID29_9 numeric Select the response that best describes YOUR life.... MY STUDY PARTNER'S pleasure over little things gives me pleasure 1=Strongly disagree (1) ; 2=Disagree (2) ; 3=Neither agree nor disagree (3) ; 4=Agree (4) ; 5=Strongly agree (5)

QID47_1 numeric Select the response that best describes your life.... The things I do to take care of MY STUDY PARTNER are worthwhile 11=Strongly disagree (1) ; 12=Disagree (2) ; 13=Neither agree nor disagree (3) ; 14=Agree (4) ; 15=Strongly agree (5)

QID47_2 numeric Select the response that best describes your life.... Most of my caregiving activities seem trivial and unimportant 11=Strongly disagree (1) ; 12=Disagree (2) ; 13=Neither agree nor disagree (3) ; 14=Agree (4) ; 15=Strongly agree (5)

QID47_3 numeric Select the response that best describes your life.... There is great value in what I do everyday for MY STUDY PARTNER 11=Strongly disagree (1) ; 12=Disagree (2) ; 13=Neither agree nor disagree (3) ; 14=Agree (4) ; 15=Strongly agree (5)

QID47_4 numeric Select the response that best describes your life.... Being a caregiver is an important role to me 11=Strongly disagree (1) ; 12=Disagree (2) ; 13=Neither agree nor disagree (3) ; 14=Agree (4) ; 15=Strongly agree (5)

QID47_5 numeric Select the response that best describes your life.... I criticize myself for having negative emotions about having to be a caregiver 11=Strongly disagree (1) ; 12=Disagree (2) ; 13=Neither agree nor disagree (3) ; 14=Agree (4) ; 15=Strongly agree (5)

QID48_10 numeric Select the response that best describes your life.... I tell myself I shouldn't be thinking or feeling the way I am about being a caregiver 11=Strongly disagree (1) ; 12=Disagree (2) ; 13=Neither agree nor disagree (3) ; 14=Agree (4) ; 15=Strongly agree (5)

QID48_6 numeric Select the response that best describes your life.... I accept my negative feelings about caregiving, instead of pushing them away 11=Strongly disagree (1) ; 12=Disagree (2) ; 13=Neither agree nor disagree (3) ; 14=Agree (4) ; 15=Strongly agree (5)

QID48_7 numeric Select the response that best describes your life.... Being a caregiver prevents me from having a fulfilling life 11=Strongly disagree (1) ; 12=Disagree (2) ; 13=Neither agree nor disagree (3) ; 14=Agree (4) ; 15=Strongly agree (5)

QID48_8 numeric Select the response that best describes your life.... I am able to be happy and satisfied with my life, and be a caregiver 11=Strongly disagree (1) ; 12=Disagree (2) ; 13=Neither agree nor disagree (3) ; 14=Agree (4) ; 15=Strongly agree (5)

QID48_9 numeric Select the response that best describes your life.... I am at peace about leading a life as a caregiver 11=Strongly disagree (1) ; 12=Disagree (2) ; 13=Neither agree nor disagree (3) ; 14=Agree (4) ; 15=Strongly agree (5)

QID49_10 numeric As a result of being a caregiver, I have experienced the following changes in my life:... I changed my priorities about what is important in life 11=Strongly disagree (1) ; 12=Disagree (2) ; 13=Neither agree nor disagree (3) ; 14=Agree (4) ; 15=Strongly agree (5)

QID49_6 numeric As a result of being a caregiver, I have experienced the following changes in my life:... I have a greater sense of closeness with others 11=Strongly disagree (1) ; 12=Disagree (2) ; 13=Neither agree nor disagree (3) ; 14=Agree (4) ; 15=Strongly agree (5)

QID49_7 numeric As a result of being a caregiver, I have experienced the following changes in my life:... I established a new path for my life 11=Strongly disagree (1) ; 12=Disagree (2) ; 13=Neither agree nor disagree (3) ; 14=Agree (4) ; 15=Strongly agree (5)

QID49_8 numeric As a result of being a caregiver, I have experienced the following changes in my life:... I discovered that I was stronger than I thought I was 11=Strongly disagree (1) ; 12=Disagree (2) ; 13=Neither agree nor disagree (3) ; 14=Agree (4) ; 15=Strongly agree (5)

QID49_9 numeric As a result of being a caregiver, I have experienced the following changes in my life:... I have a stronger religious faith 11=Strongly disagree (1) ; 12=Disagree (2) ; 13=Neither agree nor disagree (3) ; 14=Agree (4) ; 15=Strongly agree (5)

update_stamp POSIXct: YYY-MM-DD