Brain Health Registry (BHR) SP EVERYDAY COGNITION

Brain Health Registry (BHR) SP EVERYDAY COGNITION data. More information is available at https://adni.loni.usc.edu/data-samples/adni-data/

Usage

data(BHR_SP_EVERYDAY_COGNITION)

Format

A data frame with 41 observations and 47 variables.

Source

https://adni.loni.usc.edu/data-samples/adni-data/

Details

StudyPartnerID numeric Range value: 2509, ..., 6051

AboutRID numeric Range value: 2201, ..., 6644

AboutPTID character

Timepoint character , with options: sp-m00, sp-m12, sp-m18, sp-m06

CollectedDateTime POSIXct: YYY-MM-DD

CollectedDate_DRVD Date: YYY-MM-DD

latest numeric Range value: 0, ..., 1

QID49_1 numeric Compared to 10 years ago, has there been any change in... YOUR STUDY PARTNER'S memory... Remembering a few shopping items without a list. 1=Better or no change ; 2=Questionable/ occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 5=I don't know

QID49_2 numeric Compared to 10 years ago, has there been any change in... YOUR STUDY PARTNER'S memory... Remembering things that happened recently (such as recent outings, events in the news). 1=Better or no change ; 2=Questionable/ occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 5=I don't know

QID49_3 numeric Compared to 10 years ago, has there been any change in... YOUR STUDY PARTNER'S memory... Recalling conversations a few days later. 1=Better or no change ; 2=Questionable/ occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 5=I don't know

QID49_4 numeric Compared to 10 years ago, has there been any change in... YOUR STUDY PARTNER'S memory... Remembering where he/she has placed objects. 1=Better or no change ; 2=Questionable/ occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 5=I don't know

QID49_5 numeric Compared to 10 years ago, has there been any change in... YOUR STUDY PARTNER'S memory... Repeating stories and/or questions. 1=Better or no change ; 2=Questionable/ occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 5=I don't know

QID49_6 numeric Compared to 10 years ago, has there been any change in... YOUR STUDY PARTNER'S memory... Remembering the current date or day of the week. 1=Better or no change ; 2=Questionable/ occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 5=I don't know

QID49_7 numeric Compared to 10 years ago, has there been any change in... YOUR STUDY PARTNER'S memory... Remembering he/she has already told someone something. 1=Better or no change ; 2=Questionable/ occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 5=I don't know

QID49_8 numeric Compared to 10 years ago, has there been any change in... YOUR STUDY PARTNER'S memory... Remembering appointments, meetings, or engagements. 1=Better or no change ; 2=Questionable/ occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 5=I don't know

QID50_1 numeric Compared to 10 years ago, has there been any change in... YOUR STUDY PARTNER'S language... Forgetting the names of objects. 8=Better or no change ; 2=Questionable/ occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 5=I don't know

QID50_2 numeric Compared to 10 years ago, has there been any change in... YOUR STUDY PARTNER'S language... Verbally giving instructions to others. 8=Better or no change ; 2=Questionable/ occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 5=I don't know

QID50_3 numeric Compared to 10 years ago, has there been any change in... YOUR STUDY PARTNER'S language... Finding the right words to use in a conversation. 8=Better or no change ; 2=Questionable/ occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 5=I don't know

QID50_4 numeric Compared to 10 years ago, has there been any change in... YOUR STUDY PARTNER'S language... Communicating thoughts in a conversation. 8=Better or no change ; 2=Questionable/ occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 5=I don't know

QID50_5 numeric Compared to 10 years ago, has there been any change in... YOUR STUDY PARTNER'S language... Following a story in a book or on TV. 8=Better or no change ; 2=Questionable/ occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 5=I don't know

QID50_6 numeric Compared to 10 years ago, has there been any change in... YOUR STUDY PARTNER'S language... Understanding the point of what other people are trying to say. 8=Better or no change ; 2=Questionable/ occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 5=I don't know

QID50_7 numeric Compared to 10 years ago, has there been any change in... YOUR STUDY PARTNER'S language... Remembering the meaning of common words. 8=Better or no change ; 2=Questionable/ occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 5=I don't know

QID50_8 numeric Compared to 10 years ago, has there been any change in... YOUR STUDY PARTNER'S language... Describing a program he/she has watched on TV. 8=Better or no change ; 2=Questionable/ occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 5=I don't know

QID50_9 numeric Compared to 10 years ago, has there been any change in... YOUR STUDY PARTNER'S language... Understanding spoken directions or instructions. 8=Better or no change ; 2=Questionable/ occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 5=I don't know

QID51_1 numeric Compared to 10 years ago, has there been any change in... YOUR STUDY PARTNER'S visual-spatial and perceptual abilities... Following a map to find a new location. 1=Better or no change ; 2=Questionable/ occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 5=I don't know

QID51_2 numeric Compared to 10 years ago, has there been any change in... YOUR STUDY PARTNER'S visual-spatial and perceptual abilities... Reading a map and helping with directions when someone else is driving. 1=Better or no change ; 2=Questionable/ occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 5=I don't know

QID51_3 numeric Compared to 10 years ago, has there been any change in... YOUR STUDY PARTNER'S visual-spatial and perceptual abilities... Finding his/her car in a parking lot. 1=Better or no change ; 2=Questionable/ occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 5=I don't know

QID51_4 numeric Compared to 10 years ago, has there been any change in... YOUR STUDY PARTNER'S visual-spatial and perceptual abilities... Finding his/her way back to a meeting spot in the mall or other location. 1=Better or no change ; 2=Questionable/ occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 5=I don't know

QID51_5 numeric Compared to 10 years ago, has there been any change in... YOUR STUDY PARTNER'S visual-spatial and perceptual abilities... Finding his/her way around a familiar neighborhood. 1=Better or no change ; 2=Questionable/ occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 5=I don't know

QID51_6 numeric Compared to 10 years ago, has there been any change in... YOUR STUDY PARTNER'S visual-spatial and perceptual abilities... Finding his/her way around a familiar store. 1=Better or no change ; 2=Questionable/ occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 5=I don't know

QID51_7 numeric Compared to 10 years ago, has there been any change in... YOUR STUDY PARTNER'S visual-spatial and perceptual abilities... Finding his/her way around a house visited many times. 1=Better or no change ; 2=Questionable/ occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 5=I don't know

QID52_1 numeric Compared to 10 years ago, has there been any change in... YOUR STUDY PARTNER'S executive functioning: planning... Planning a sequence of stops on a shopping trip. 1=Better or no change ; 2=Questionable/ occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 5=I don't know

QID52_2 numeric Compared to 10 years ago, has there been any change in... YOUR STUDY PARTNER'S executive functioning: planning... The ability to anticipate weather changes and plan accordingly (i.e., bring a coat or umbrella). 1=Better or no change ; 2=Questionable/ occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 5=I don't know

QID52_3 numeric Compared to 10 years ago, has there been any change in... YOUR STUDY PARTNER'S executive functioning: planning... Developing a schedule in advance of anticipated events. 1=Better or no change ; 2=Questionable/ occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 5=I don't know

QID52_4 numeric Compared to 10 years ago, has there been any change in... YOUR STUDY PARTNER'S executive functioning: planning... Thinking things through before acting. 1=Better or no change ; 2=Questionable/ occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 5=I don't know

QID52_5 numeric Compared to 10 years ago, has there been any change in... YOUR STUDY PARTNER'S executive functioning: planning... Thinking ahead. 1=Better or no change ; 2=Questionable/ occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 5=I don't know

QID53_1 numeric Compared to 10 years ago, has there been any change in... YOUR STUDY PARTNER'S executive functioning: organization... Keeping living and work space organized. 1=Better or no change ; 2=Questionable/ occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 5=I don't know

QID53_2 numeric Compared to 10 years ago, has there been any change in... YOUR STUDY PARTNER'S executive functioning: organization... Balancing the checkbook without error. 1=Better or no change ; 2=Questionable/ occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 5=I don't know

QID53_3 numeric Compared to 10 years ago, has there been any change in... YOUR STUDY PARTNER'S executive functioning: organization... Keeping financial records organized. 1=Better or no change ; 2=Questionable/ occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 5=I don't know

QID53_4 numeric Compared to 10 years ago, has there been any change in... YOUR STUDY PARTNER'S executive functioning: organization... Prioritizing tasks by importance. 1=Better or no change ; 2=Questionable/ occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 5=I don't know

QID53_5 numeric Compared to 10 years ago, has there been any change in... YOUR STUDY PARTNER'S executive functioning: organization... Keeping mail and papers organized. 1=Better or no change ; 2=Questionable/ occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 5=I don't know

QID53_6 numeric Compared to 10 years ago, has there been any change in... YOUR STUDY PARTNER'S executive functioning: organization... Using an organized strategy to manage a medication schedule involving multiple medications. 1=Better or no change ; 2=Questionable/ occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 5=I don't know

QID54_1 numeric Compared to 10 years ago, has there been any change in... YOUR STUDY PARTNER'S executive functioning: divided attention... The ability to do two things at once. 1=Better or no change ; 2=Questionable/ occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 5=I don't know

QID54_2 numeric Compared to 10 years ago, has there been any change in... YOUR STUDY PARTNER'S executive functioning: divided attention... Returning to a task after being interrupted. 1=Better or no change ; 2=Questionable/ occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 5=I don't know

QID54_3 numeric Compared to 10 years ago, has there been any change in... YOUR STUDY PARTNER'S executive functioning: divided attention... The ability to concentrate on a task without being distracted by external things in the environment. 1=Better or no change ; 2=Questionable/ occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 5=I don't know

QID54_4 numeric Compared to 10 years ago, has there been any change in... YOUR STUDY PARTNER'S executive functioning: divided attention... Cooking or working and talking at the same time. 1=Better or no change ; 2=Questionable/ occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 5=I don't know

update_stamp POSIXct: YYY-MM-DD