Brain Health Registry SP INITIAL

Brain Health Registry SP INITIAL data. More information is available at https://adni.loni.usc.edu/data-samples/adni-data/

Usage

data(BHR_SP_INITIAL)

Format

A data frame with 46 observations and 38 variables.

Source

https://adni.loni.usc.edu/data-samples/adni-data/

Details

AboutRID numeric Range value: 2201, ..., 6644

StudyPartnerID numeric Range value: 2509, ..., 6051

AboutPTID character

Timepoint character , with options: sp-m00, sp-m12, sp-m18, sp-m06

CollectedDateTime POSIXct POSIXct: YYY-MM-DD

CollectedDate_DRVD Date Date: YYY-MM-DD

latest numeric Range value: 0, ..., 1

QID2 numeric Are you concerned that YOUR STUDY PARTNER has a memory problem? 1=Yes;2=No;3=I don't know

QID3 numeric Has there been a change in YOUR STUDY PARTNER'S memory in the past 10 years? 1=Yes;2=No;3=I don't know

QID4 numeric Does YOUR STUDY PARTNER have trouble remembering whether he/she told someone something? 1=Yes;2=No;3=I don't know

QID22 numeric Range value: 1, ..., 3

QID23 numeric Range value: 1, ..., 3

QID24 numeric Range value: 1, ..., 3

QID5 numeric Compared to one year ago, does it seem to you that lately YOUR STUDY PARTNER is relying more on written reminders (e.g., shopping lists, calendars)? 1=Yes;2=Maybe;3=No

QID6 numeric Compared to one year ago, does YOUR STUDY PARTNER need more help from others to remember appointments, family occasions or holidays? 1=Yes;2=Maybe;3=No

QID27 numeric Range value: 1, ..., 3

QID28 numeric Range value: 1, ..., 3

QID7 numeric Compared to one year ago, is YOUR STUDY PARTNER having more difficulty managing money (e.g., paying bills, calculating change, completing tax forms)? 1=Yes;2=Maybe;3=No

QID30 numeric Range value: 2, ..., 3

QID31 numeric Range value: 2, ..., 3

QID32 numeric Range value: 1, ..., 3

QID33 numeric Range value: 2, ..., 3

QID34 numeric Range value: 1, ..., 3

QID35 numeric Range value: 2, ..., 3

QID8 numeric Compared to one year ago, does YOUR STUDY PARTNER show less interest in hobbies and social activities? 1=Yes;2=Maybe;3=No

QID9 numeric Has YOUR STUDY PARTNER ever been diagnosed with Mild Cognitive Impairment (MCI)? 1=Yes;2=No;3=I don't know

QID10 numeric Has YOUR STUDY PARTNER ever been diagnosed with Alzheimer's disease? 1=Yes;2=No;3=I don't know

QID11 numeric Are you concerned that YOU have a memory problem? 1=Yes;2=No

QID12 numeric Do YOU have trouble remembering if YOU have already told someone something? 1=Yes;2=No

QID13 numeric Compared to 10 years ago, would you say that there has been a change in YOUR memory? 1=Yes;2=No

QID14 numeric In general, would you say YOUR health is: 1=Excellent;2=Very Good;3=Good;4=Fair;5=Poor

QID15 numeric Compared to one year ago, how would you rate YOUR health in general now? 1=Much Better;2=Better;3=Same;4=Worse;5=Much Worse

QID16 numeric Over the last two months, how often have YOU been bothered by having little interest or pleasure in doing things? 1=Not at all;2=Several days;3=More than half the days;4=Nearly every day

QID17 numeric Over the last two months, how often have YOU been bothered by feeling down, depressed, or hopeless? 1=Not at all;2=Several days;3=More than half the days;4=Nearly every day

QID18 numeric Stress means a situation in which a person feels tense, restless, nervous or anxious or is unable to sleep at night because his/her mind is troubled all the time. During the past 12 months, would you say YOU experienced: 1=A lot of stress;2=A moderate amount of stress;3=Relatively little stress;4=Almost not stress at all

QID19 numeric Have YOU ever been diagnosed with Mild Cognitive Impairment (MCI). 1=Yes;2=No;3=I don't know

QID20 numeric Have YOU ever been diagnosed with Alzheimer's disease? 1=Yes;2=No;3=I don't know

update_stamp POSIXct POSIXct: YYY-MM-DD

Author

adni-data@googlegroups.com