Biomarker Samples

Biomarker Samples data. More information is available at https://adni.loni.usc.edu/data-samples/adni-data/

Usage

data(BIOMARK)

Format

A data frame with 13450 observations and 66 variables.

Source

https://adni.loni.usc.edu/data-samples/adni-data/

Details

ORIGPROT character Original study protocol , with options: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4

COLPROT factor Study protocol of data collection , with factor levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4

PTID character Participant ID

RID numeric Participant roster ID Range value: 2, ..., 10618

VISCODE character Visit code

VISCODE2 character Translated visit code

EXAMDATE Date of Collection Date: YYY-MM-DD

RECNO numeric Range value: 1, ..., 3

BIBLOOD character On this date, was blood for biomarkers collected?

ADNI1, ADNIGO: 1=Blood

ADNI1, ADNIGO: 1=Blood

ADNI2: 1=Yes;0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

BIURINE character 1=Urine Decoded Value: Yes

BLREASON character If no, please provide reason why blood was not collected

ADNI2: 1=Illness;2=Participant unavailable;3=Participant Unwilling;7=Administrative Problems;4=Withdrawn consent;5=Collection done on a different date for this visit;6=Other (specify)

ADNI3: 1=Illness;2=Participant unavailable;3=Participant Unwilling;4=Withdrawn consent;5=Collection done on a different date for this visit;6=Other (specify);7=Administrative Problems

ADNI4: 1=Illness; 2=Participant unavailable; 3=Participant unwilling; 4=Withdrawn consent; 5=Collection done on a different date for this visit; 6=Other (specify); 7=Administrative Problems

Decoded Value: Yes

BIFAST character Was participant fasting for at least 6 hours?

ADNI1, ADNIGO, ADNI2: 1=Yes;0=No

ADNI1, ADNIGO, ADNI2: 1=Yes;0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

BITIME character Time of Collection

BIREDTIME character Serum: Time Collected

BIREDAMT numeric Serum: Amount Collected

ADNI1, ADNIGO, ADNI2: 0..20

ADNI3, ADNI4

BIREDCENT character Serum: Centrifuged Time

BIREDTRNS character Serum: Transfer Time

BIREDVOL numeric Serum: Volume of Serum Transferred

ADNI1: 0..13

ADNIGO, ADNI2: 0..20

ADNI3, ADNI4

ADNI3, ADNI4

BIREDFROZ character Serum: Time Frozen

BILAVTIME character Plasma: Time Collected

BINEEDLE character Needle Used: 1=Sprotte;2=Sharp Decoded Value: Yes

BILAVAMT numeric Plasma: Amount Collected

ADNI1, ADNIGO, ADNI2: 0..20

ADNI3, ADNI4

ADNI3, ADNI4

BILAVCENT character Plasma: Centrifuged Time

BILAVTRNS character Plasma: Transfer Time

BILAVVOL numeric Plasma: Volume of Plasma Transferred

ADNI1: 0..13

ADNIGO, ADNI2: 0..20

ADNI3, ADNI4

ADNI3, ADNI4

BILAVFROZ character Plasma: Time Frozen

BIURITIME character Time Collected Urine

BIURIAMT numeric Amount Collected Urine 1..100

BIURITRNS character Transfer Time Urine

BIURIVOL numeric Volume of Urine Transfered Urine 1..13

BIURIFROZ numeric Time Frozen Urine Range value: 110, ..., 1940

BICSF character On this date, was LP conducted?

ADNI1, ADNIGO: 1=CSF

ADNI1, ADNIGO: 1=CSF

ADNI2: 1=Yes;0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

BINONE character comment

ADNI1: 1=None

ADNIGO: 1=No samples were collected this visit

Decoded Value: Yes

REASON character If no, please provide reason why the LP was not conducted

ADNI2: 1=Illness;2=Participant unavailable;3=Participant Unwilling;8=Administrative Problems;4=Withdrawn consent;6=Not called for by protocol at this visit;7=Collection done on a different date for this visit;5=Other (specify)

ADNI3: 1=Illness;2=Participant unavailable;3=Participant Unwilling;4=Withdrawn consent;5=Other (specify);6=Not called for by protocol at this visit;7=Collection done on a different date for this visit;8=Administrative Problems

ADNI4: 1=Illness; 2=Participant unavailable; 3=Participant unwilling; 4=Withdrawn consent; 5=Other (specify); 7=Collection done on a different date for this visit; 8=Administrative Problems

Decoded Value: Yes

BICSFFAST character Was participant fasting for at least 6 hours?

ADNI2: 1=Yes;0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

BICSFTIME character Time of LP

BIMETHOD character Method of Collection

ADNI1, ADNIGO, ADNI2, ADNI3: 1=gravity;2=syringe suction

ADNI1, ADNIGO, ADNI2, ADNI3: 1=gravity;2=syringe suction

ADNI4: 1=gravity; 2=syringe suction

Decoded Value: Yes

NEEDLESIZE character Needle Gauge

ADNI2: 15=18g Quincke (sharp beveled) needle; 7=19g Quincke (sharp beveled) needle;1=20g Quincke (sharp beveled) needle;8=21g Quincke (sharp beveled) needle;2=22g Quincke (sharp beveled) needle;9=23g Quincke (sharp beveled) needle;10=24g Quincke (sharp beveled) needle;3=25g Quincke (sharp beveled) needle;11=18g Sprotte (atraumatic) needle;12=19g Sprotte (atraumatic) needle;13=20g Sprotte (atraumatic) needle;14=21g Sprotte (atraumatic) needle;4=22g Sprotte (atraumatic) needle;16=23g Sprotte (atraumatic) needle; 5=24g Sprotte (atraumatic) needle; 17=25g Sprotte (atraumatic) needle;18=Other (specify)

ADNI3: 1=20g Quincke (sharp beveled) needle;2=22g Quincke (sharp beveled) needle;3=25g Quincke (sharp beveled) needle;4=22g Sprotte (atraumatic) needle;5=24g Sprotte (atraumatic) needle;7=19g Quincke (sharp beveled) needle;8=21g Quincke (sharp beveled) needle;9=23g Quincke (sharp beveled) needle;10=24g Quincke (sharp beveled) needle;11=18g Sprotte (atraumatic) needle;12=19g Sprotte (atraumatic) needle;13=20g Sprotte (atraumatic) needle;14=21g Sprotte (atraumatic) needle;15=18g Quincke (sharp beveled) needle;16=23g Sprotte (atraumatic) needle;17=25g Sprotte (atraumatic) needle;18=Other (specify)

ADNI4: 1=20g Quincke (sharp beveled) needle; 10=24g Quincke (sharp beveled) needle; 11=18g Sprotte (atraumatic) needle; 12=19g Sprotte (atraumatic) needle; 13=20g Sprotte (atraumatic) needle; 14=21g Sprotte (atraumatic) needle; 15=18g Quincke (sharp beveled) needle; 16=23g Sprotte (atraumatic) needle; 17=25g Sprotte (atraumatic) needle; 18=Other (specify); 2=22g Quincke (sharp beveled) needle; 3=25g Quincke (sharp beveled) needle; 4=22g Sprotte (atraumatic) needle; 5=24g Sprotte (atraumatic) needle; 7=19g Quincke (sharp beveled) needle; 8=21g Quincke (sharp beveled) needle; 9=23g Quincke (sharp beveled) needle

Decoded Value: Yes

INTERSPACE character LP performed at the:

ADNI2: 4=L1-L2 inter space; 2=L2-L3 inter space;1=L3-L4 inter space; 5=L4-L5 inter space

ADNI3: 1=L3-L4 inter space;2=L2-L3 inter space;4=L1-L2 inter space;5=L4-L5 inter space

ADNI4: 1=L3-L4 inter space; 2=L2-L3 inter space; 4=L1-L2 inter space; 5=L4-L5 inter space

Decoded Value: Yes

PTPOSITION character Patient Position

ADNI2: 1=Sitting, leaned over (preferred);2=Lying, curled up on side; 3=Other (specify)

ADNI3: 1=Sitting, leaned over (preferred);2=Lying, curled up on side;3=Other (specify)

ADNI4: 1=Sitting, leaned over (preferred); 2=Lying, curled up on side; 3=Other (specify)

Decoded Value: Yes

COLTUBETYP character Type of collection tube used

ADNI2, ADNI4: 1=Polypropylene; 2=Polystyrene

ADNI3: 1=Polypropylene;2=Polystyrene

Decoded Value: Yes

SHPTUBETYP character Type of tube used for shipping

ADNI2, ADNI4: 1=Polypropylene; 2=Polystyrene

ADNI3: 1=Polypropylene;2=Polystyrene

Decoded Value: Yes

TUBEMIN numeric If collected in polystyrene and shipped in polypropylene, please provide estimated amount of time CSF remained in polystyrene collection tube

ADNI2: 0..1440

ADNI3, ADNI4

BICSFAMT numeric CSF: Amount Collected

ADNI1, ADNIGO, ADNI2: 0..40

ADNI3, ADNI4

BICSFTRNS character CSF: Transfer Time

BICSFVOL numeric CSF: Volume of CSF Transferred

ADNI1, ADNIGO, ADNI2: 0..40

ADNI3, ADNI4

BICSFFROZ character CSF: Time Frozen

BILPPATCH character Check if any of the following was performed 1=Lumbar Puncture Blood Patch Decoded Value: Yes

BILPFLURO character

ADNI1: 1=Fluroscopy

ADNIGO, ADNI2: 1=Fluoroscopy

Decoded Value: Yes

BILPSPFILM character 1=Lumbar Spine Film Decoded Value: Yes

BILPPADATE character Date of Blood Patch

BILPFLDATE character Date of Fluoroscopy

BILPSPDATE character Date of Spine Film , with options: 10/10/2006, 10/30/2012, 03/01/2013, 2022-xx-xx, 2020-10-01, 2019-10-29, 2018-xx-xx

BILPPADATEYR_DRVD numeric Year of Blood Patch (Derived from BILPPADATE) Range value: 2006, ..., 2025

BILPFLDATEYR_DRVD numeric Year of Fluoroscopy (Derived from BILPFLDATE) Range value: 2005, ..., 2025

BILPSPDATEYR_DRVD numeric Year of Spine Film (Derived from BILPSPDATE) Range value: 2006, ..., 2022

BILPOTPROC character Check if any of the following was performed

ADNI3: 1=Lumbar Puncture Blood Patch;2=Fluoroscopy;3=Lumbar Spine Film

ADNI4: 1=Lumbar Puncture Blood Patch; 2=Fluoroscopy; 3=Lumbar Spine Film

Decoded Value: Yes

BIFEDDATE Date FedExed Date: YYY-MM-DD

ID numeric Record ID Range value: 3, ..., 186225

SITEID numeric Site ID Range value: 1, ..., 941

USERDATE User Create Date Date: YYY-MM-DD

USERDATE2 User Last Modify Date Date: YYY-MM-DD

DD_CRF_VERSION_LABEL character CRF Version Label , with options: v1

LANGUAGE_CODE character Language Code e=English Decoded Value: Yes

HAS_QC_ERROR character Has Quality Check Error 0=Does not have QC error or QC error has been approved; 1=Has QC error Decoded Value: Yes

update_stamp POSIXct: YYY-MM-DD

Author

adni-data@googlegroups.com