Diagnostic Summary - Baseline Changes

Diagnostic Summary - Baseline Changes data. More information is available at https://adni.loni.usc.edu/data-samples/adni-data/

Usage

data(BLCHANGE)

Format

A data frame with 15436 observations and 30 variables.

Source

https://adni.loni.usc.edu/data-samples/adni-data/

Details

ORIGPROT character Original study protocol , with options: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4

COLPROT factor Study protocol of data collection , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4

PTID character Participant ID

RID numeric Participant roster ID Range value: 2, ..., 10894

VISCODE character Visit code

VISCODE2 character Translated visit code

EXAMDATE Date Form Completed Date: YYY-MM-DD

BCPREDX character Pre-visit Diagnosis 1=NL ; 2=MCI ; 3=AD Decoded Value: Yes

BCADAS character 1. Clinically relevant worsening on ADAS?

ADNI1, ADNIGO, ADNI2: 1=Yes ; 0=No

ADNI3, ADNI4: 0=No ; 1=Yes

Decoded Value: Yes

BCMMSE character 2. Clinically relevant worsening on MMSE?

ADNI1, ADNIGO, ADNI2: 1=Yes ; 0=No

ADNI3, ADNI4: 0=No ; 1=Yes

Decoded Value: Yes

BCMMSREC character 3. Clinically relevant worsening on MMSE recall?

ADNI1, ADNIGO, ADNI2: 1=Yes ; 0=No

ADNI3, ADNI4: 0=No ; 1=Yes

Decoded Value: Yes

BCNMMMS character 4. Clinically relevant worsening on non-memory MMSE items?

ADNI1, ADNIGO, ADNI2: 1=Yes ; 0=No

ADNI3, ADNI4: 0=No ; 1=Yes

Decoded Value: Yes

BCNEUPSY character 5. Clinically relevant worsening in memory on neuropsych testing?

ADNI1, ADNIGO, ADNI2: 1=Yes ; 0=No

ADNI3, ADNI4: 0=No ; 1=Yes

Decoded Value: Yes

BCNONMEM character 6. Clinically relevant impairment/worsening in non-memory cognitive domains on neuropsych testing?

ADNI1, ADNIGO, ADNI2: 1=Yes ; 0=No

ADNI3, ADNI4: 0=No ; 1=Yes

Decoded Value: Yes

BCFAQ character 7. Clinically relevant worsening in activities of daily living (FAQ)?

ADNI1, ADNIGO, ADNI2: 1=Yes ; 0=No

ADNI3, ADNI4: 0=No ; 1=Yes

Decoded Value: Yes

BCCDR character 8. Clinically relevant deterioration on CDR Sum of Boxes or Overall CDR rating?

ADNI1, ADNIGO, ADNI2: 1=Yes ; 0=No

ADNI3, ADNI4: 0=No ; 1=Yes

Decoded Value: Yes

BCDEPRES character 9. Clinically relevant depression based on clinical judgement or GDS?

ADNI1, ADNIGO, ADNI2: 1=Yes ; 0=No

ADNI3, ADNI4: 0=No ; 1=Yes

Decoded Value: Yes

BCSTROKE character 10. Did subject have a stroke?

ADNI1, ADNIGO, ADNI2: 1=Yes ; 0=No

ADNI3, ADNI4: 0=No ; 1=Yes

Decoded Value: Yes

BCDELIR character 11. Is there evidence of a delirium (medication effect, toxic or metabolic encephalopathy)?

ADNI1, ADNIGO, ADNI2: 1=Yes ; 0=No

ADNI3, ADNI4: 0=No ; 1=Yes

Decoded Value: Yes

BCEXTCIR character 12. Has extenuating circumstance (such as a physical health problem, change in residence, charge in support network, death of a family member, etc.) contributed to a change in the subject's cognitive or functional performance?

ADNI1, ADNIGO, ADNI2: 1=Yes ; 0=No

ADNI3, ADNI4: 0=No ; 1=Yes

Decoded Value: Yes

BCCORADL character 13. Is the change in clinical status corroborated by informant report of changes in ADL?

ADNI1, ADNIGO, ADNI2: 1=Yes ; 0=No ; -1=NA/No change in clinical status

ADNI3, ADNI4: -1=NA/No change in clinical status ; 0=No ; 1=Yes

Decoded Value: Yes

BCCORCOG character 14. Is the change in clinical status corroborated by informant report of changes in cognition?

ADNI1, ADNIGO, ADNI2: 1=Yes ; 0=No ; -1=NA/No change in clinical status

ADNI3, ADNI4: -1=NA/No change in clinical status ; 0=No ; 1=Yes

Decoded Value: Yes

ID numeric Record ID Range value: 2, ..., 187119

SITEID numeric Site ID Range value: 1, ..., 941

USERDATE User Create Date Date: YYY-MM-DD

USERDATE2 User Last Modify Date Date: YYY-MM-DD

DD_CRF_VERSION_LABEL character CRF Version Label , with options: annual, sc, bl

LANGUAGE_CODE character Language Code e=English Decoded Value: Yes

HAS_QC_ERROR character Has Quality Check Error 0=Does not have QC error or QC error has been approved ; 1=Has QC error Decoded Value: Yes

update_stamp POSIXct: YYY-MM-DD

Author

adni-data@googlegroups.com