Boundary Shift Integral Summaries (Nick Fox's lab)

Boundary Shift Integral Summaries (Nick Fox's lab) data. More information is available at https://adni.loni.usc.edu/data-samples/adni-data/

Usage

data(BSI)

Format

A data frame with 3594 observations and 21 variables.

Source

https://adni.loni.usc.edu/data-samples/adni-data/

Details

ORIGPROT character Original study protocol , with options: ADNI1

ID numeric Range value: 1, ..., 3602

RID numeric Range value: 2, ..., 1435

SITEID numeric Range value: 1, ..., 127

VISCODE character , with options: sc, m06, m12, m24, m18, m36, bl, f

RECNO numeric Range value: 1, ..., 2

EXAMDATE Examination Date Date: YYY-MM-DD

IMAGEUID numeric imageUID Range value: 30968, ..., 143906

FLDSTRENG numeric Field Strength Range value: 1.5, ..., 3

BBSI numeric Boundary Shift Integral from baseline Range value: -108.14, ..., 97.315

DBCBBSI numeric Differential Bias Corrected Boundary Shift Integral from baseline Range value: -50.143, ..., 82.892

VBSI numeric Ventricular BSI Range value: -18.088, ..., 29.725

BRAINVOL numeric Segmented brain volume. Not to be used as a measure of intracranial volume. Range value: 680.31, ..., 1400.91

VENTVOL numeric Segmented ventricular volume Range value: 6.715, ..., 179.123

REGCOMMENT character

REGRATING numeric 1=OK ; 2=less reliable, still a reasonable measure ; 3=Borderline / unreliable, measure affected to some degree, still of some value ; 4=Rejected <ef><bf><bd> measure not reliable at all. Suggest is not used for analysis ; 5=Non disease related pathology affecting measure, suggest is removed from analysis

VENTACCEPT numeric A boolean to indicate whether or not we have confidence in the VBSI value. Similar purpose to the REGRATING. 0=Vent not accepted ; 1=Vent accepted

KMNDBCBBSI numeric An improved version of the Differential Bias Corrected Boundary Shift Integral from baseline, which: a) Has a new intensity normalisation step that normalises scan intensity by matching the intensities by tissue type (cerebrospinal fluid, grey matter, white mater) instead of being just based on the average whole brain intensity. This helps reduce artefactual change in the BSI that might be caused by change in contrasts between scans (e.g scanner changes) b) Automatically calculates the BSI window based on k-means clustering of tissue intensities for each scan pair Range value: -24.785, ..., 83.473

KMNREGCOMMENT character

KMNREGRATING numeric Range value: 1, ..., 8

update_stamp POSIXct: YYY-MM-DD

Author

adni-data@googlegroups.com