Cogstate Battery Results

Cogstate Battery Results data. More information is available at https://adni.loni.usc.edu/data-samples/adni-data/

Usage

data(CBBRESULTS)

Format

A data frame with 22205 observations and 44 variables.

Source

https://adni.loni.usc.edu/data-samples/adni-data/

Details

ORIGPROT character Original study protocol , with options: ADNI1, ADNIGO, ADNI2, ADNI3

COLPROT character Study protocol of data collection , with options: ADNI2, ADNI3

RID numeric Participant roster ID

SessionID numeric Identification number for the session. Range value: 2694, ..., 613669

EXAMDATE Examination Date Date: YYYY-MM-DD

TestTime Time at the start of the test session. hms: HH:MM:SS

Visit character Visit label. The visit options are:In Clinic = indicates the session was completed in the clinic.Remote = indicates the session was completed remotely (i.e., at-home)Unknown = it is not known whether the session was completed in the clinic or remotely (Unknown visit labels may change to In Clinic/ Remote over time as data is queried and clarified by sites). , with options: Remote, In Clinic, Unknown

SessionAttempt numeric The cumulative total of attempts of the Cogstate battery (including the current assessment). Range value: 1, ..., 51

SessionDuration numeric The total duration of the current session. Range value: 203, ..., 11907138

SessionCompletionPass character Session completion criteria.(i) Yes = all tests in session met completion criteria. If a test is attempted multiple times within the session, TestCompletionPass must = Yes for at least one of the attempts.(ii) No = at least one test in the session did not meet completion criteria. If a test is attempted multiple times within the session, TestCompletionPass = No for all attempts.(iii) If one or more tests in the battery do not have any data for the real test (and therefore, the test does not appear in the extract), SessionCompletionPass = No. , with options: Yes, No

SessionPerformancePass character Session performance criteria.(i) Yes = all tests in the session met the performance check. If a test is attempted multiple times within the session, TestPerformancePass must = Yes for the most recent attempt that passed completion.(ii) No = at least one test in the session did not meet performance check. If a test is attempted multiple times within the session, TestPerformancePass = No for the most recent attempt that passed completion.(iii) If TestCompletionPass = No for all attempts for a particular test, SessionPerformancePass = No.(iv) If one or more tests in the battery do not have any data for the real test (and therefore, the test does not appear in the extract), SessionPerformancePass = No. , with options: Yes, No

SessionIntegrityPass character Session integrity criteria.(i) Yes = all tests in the session met the integrity check. If a test is attempted multiple times within the session, TestIntegrityPass must = Yes for the most recent attempt that passed completion.(ii) No = at least one test in the session did not meet the integrity check. If a test is attempted multiple times within the session, TestIntegrityPass = No for the most recent attempt that passed completion.(iii) If TestCompletionPass = No for all attempts of a test that contributes to the integrity checks, SessionIntegrityPass = No.(iv) If one or more tests in the battery that contribute to the integrity checks do not have any data for the real test (and therefore, the test does not appear in the extract), SessionIntegrityPass = No. , with options: Yes, No

TestCode character Code name for the test. , with options: OneBack, Detection, Identification, OneCardLearning

TestAttempt numeric Tests in the Cogstate battery have been configured to repeat if the test completion and/or performance criteria are not met. TestAttempt indicates the attempted number of the test in the current session. Tests can repeat a maximum of 3 times. Range value: 1, ..., 3

TestCompletionScore numeric The completion score for the test [(TotalTrials/required correct responses) x 100; required correct responses is 35 for Detection, 30 for Identification, 80 for One Card Learning and 31 for One Back]. Range value: 72.5, ..., 300

TestPerformanceScore numeric The performance score for the test [reflects raw accuracy of performance; (TotalCorrect/TotalResponse) x 100] Range value: 0, ..., 100

TestCompletionPass character Test completion criteria (Yes = met completion check; No = did not meet completion check). , with options: Yes, No

TestPerformancePass character Test performance criteria (Yes = met performance check; No = did not meet performance check). , with options: Yes, No

TestIntegrityPass character Test integrity criteria (Yes = met integrity check; No = did not meet integrity check). , with options: No, Yes

TestDuration numeric The total time (ms) from test start until the end of the test. Range value: 31256, ..., 3782229

ReactionTime numeric Mean of log10 transformed reaction times of correct responses. Range value: 2.2601, ..., 3.9147

RawReactionTime numeric Mean of reaction times of correct responses. Range value: 182, ..., 23016.2308

RTVariability numeric Standard deviation of log10 transformed reaction times of correct responses. Range value: 0.0194, ..., 0.7633

RawRTVariability numeric Standard deviation of reaction times of correct responses. Range value: 12.7089, ..., 16932.8234

Accuracy numeric Accuracy of performance (arcsine square root proportion correct) Range value: 0, ..., 1.5708

RawAccuracy numeric Accuracy of performance (proportion of correct responses). Range value: 0, ..., 1

TotalCorrect numeric Number of correct responses. Range value: 0, ..., 80

TotalCorrectExclPant numeric Number of correct responses after Pant (post-anticipatory) responses have been excluded. Range value: 0, ..., 80

TotalErrors numeric Number of errors. Range value: 0, ..., 110

LegalErrors numeric Number of legal errors (errors that are not breaking rules of test). Range value: 0, ..., 89

RuleBreakErrors numeric Number of errors that break the rules of the tests. Range value: 0, ..., 100

TotalAnticipatory numeric Number of Anticipatory responses (i.e., responding before the card turns face-up). Range value: 0, ..., 80

TotalPost numeric Number of Post responses (i.e., responding during the response feedback period and before the beginning of the next trial). Range value: 0, ..., 35

TotalMaxOut numeric Number of MaxOut responses (i.e., when a response is not provided within the maximum trial duration). Range value: 0, ..., 36

TotalResponses numeric Number of responses provided. Range value: 0, ..., 167

TotalTrials numeric Number of trials completed. Range value: 0, ..., 91

StandardScoreZ numeric Standardized z-score for the primary outcome measure for each test. For the One Back test, Accuracy is used for this outcome. Range value: -7.8207, ..., 4.8991

StandardScoreT numeric Standardized T-score (mean = 100, SD = 10) for the primary outcome measure for each test. For the One Back test, Accuracy is used for this outcome. Range value: 21.793, ..., 148.9908

AltStandardScoreZ numeric Standardized z-score for One Back using Reaction Time as the outcome measure. Range value: -4.6389, ..., 2.7656

AltStandardScoreT numeric Standardized T-score (mean = 100, SD = 10) for One Back using Reaction Time as the outcome measure. Range value: 53.6111, ..., 127.6564

Psyattstdscr numeric The score on the Psychomotor Function/Attention composite, incorporating ReactionTime as the outcome measure for the Detection and Identification tests, relative to age-matched normative data. Range value: 37.6477, ..., 120.9856

LearnWMStdScr numeric The score on the Learning/Working Memory composite, incorporating Accuracy as the outcome measures for the One Card Learning and One Back tests, relative to age-matched normative data. Range value: 62.4325, ..., 137.4209

AltLearnWMStdScr numeric The score on the Learning/Working Memory composite, incorporating Accuracy as the outcome measure for the One Card Learning test and ReactionTime as the outcome measure for the One Back test, relative to age-matched normative data. Range value: 46.8986, ..., 136.5113

update_stamp POSIXct: YYYY-MM-DD HH:MM:SS