USC - DTI ROI Summary Measures v2 + v3 (Mean)

USC - DTI ROI Summary Measures v2 + v3 (Mean) data. More information is available at https://adni.loni.usc.edu/data-samples/adni-data/

Usage

data(DTIROI_MEAN)

Format

A data frame with 1750 observations and 310 variables.

Source

https://adni.loni.usc.edu/data-samples/adni-data/

Details

ORIGPROT character Original study protocol , with options: ADNI1, ADNI2, ADNI3, ADNIGO, ADNI4

COLPROT factor Study protocol of data collection , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4

PTID character Participant ID

RID numeric Participant roster ID Range value: 11, ..., 10793

VISCODE character Visit code , with options: init, y2, y1, sc, y4, y3, 4_sc, 4_init, 4_m12

VISCODE2 character Translated visit code

EXAMDATE Examination Date Date: YYY-MM-DD

VERSION character Preprocessing pipeline version 2 = preprocessing pipeline as outlined in methods PDF: (USC - DTI ROI Summary Measures Methods v2 [ADNI3], date: 12/01/2021) ; 3 = preprocessing pipeline as outlined in methods PDF: (USC - DTI ROI Summary Measures Methods v3 [ADNI4], date: 04/02/2025) ; -4 = image was not processed as noted in QC column. Decoded Value: Yes

MANUFACTURER character Scanner Manufacturer , with options: Siemens, GE, Philips, SIEMENS, GE MEDICAL SYSTEMS

VOLUMES numeric Number of volumes prescribed by acquisition protocol Range value: 31, ..., 127

IMAGEUID numeric IMAGE ID of the raw diffusion scan Range value: 808826, ..., 11303224

DISTORTION_CORRECTION character In V3, EPI susceptibility distortion correction was run using a synthetic b0 (Synb0), b0s acquired with opposite phase encoding directions (APPA), or was completed in scanner (scanner), in addition to warping DWIs to T1w. In V2, susceptibility distortion was only corrected by warping DWIs to T1w (empty cells). Synb0=synthetic b0 ; APPA=b0s acquired with opposite phase encoding directions ; scanner=completed in scanner.

VOLUMES_AP numeric Number of volumes in the opposite phase encoding direction raw scan. Empty cell indicates there was no opposite phase encoding image provided to users on the IDA. Range value: 7, ..., 13

IMAGEUID_AP numeric IMAGE ID of the raw diffusion scan with an opposite phase encoding direction. Empty cell indicates there was no opposite phase encoding image provided to users on the IDA. Range value: 10248884, ..., 11297753

RUNDATE Processing Date Date: YYY-MM-DD

STATUS character Processing Completed , with options: complete, not processed

QC character Preprocessed DWI and scalar map quality check notes include: (1) a note if raw DWI were not processed due to incomplete acquisition (e.g., fewer volumes acquired than protocol), failed DICOM conversion, poor field of view (i.e., image cropped), or severe artifacts (e.g., motion, distortion); or (2) notes if preprocessed DWI had any remaining artifacts (e.g., motion, distortion, ghosting, hyperintensities in FA, etc.), pathology or anatomical anomalies, atlas registration failure, or if the voxel size differed from the prescribed acquisition protocol. These images should be used at own discretion.

AD_MCP numeric Middle cerebellar peduncle (mean) Range value: 0.000864373, ..., 0.00140574

AD_PCT numeric Pontine crossing tract (mean) Range value: 0.000635765, ..., 0.00120459

AD_GCC numeric Genu of corpus callosum (mean) Range value: 0.00108743, ..., 0.00174352

AD_BCC numeric Body of corpus callosum (mean) Range value: 0.00117643, ..., 0.00183443

AD_SCC numeric Splenium of corpus callosum (mean) Range value: 0.000952219, ..., 0.00187696

AD_FX numeric Fornix (mean) Range value: 0.00136849, ..., 0.00336679

AD_CST_R numeric Corticospinal tract right (mean) Range value: 0.000697005, ..., 0.00124421

AD_CST_L numeric Corticospinal tract left (mean) Range value: 0.000743352, ..., 0.00128856

AD_ML_R numeric Medial lemniscus right (mean) Range value: 0.000853256, ..., 0.00153265

AD_ML_L numeric Medial lemniscus left (mean) Range value: 0.00091114, ..., 0.00144778

AD_ICP_R numeric Inferior cerebellar peduncle right (mean) Range value: 0.000773077, ..., 0.00137231

AD_ICP_L numeric Inferior cerebellar peduncle left (mean) Range value: 0.000799225, ..., 0.00135735

AD_SCP_R numeric Superior cerebellar peduncle right (mean) Range value: 0.00109985, ..., 0.00219184

AD_SCP_L numeric Superior cerebellar peduncle left (mean) Range value: 0.000948561, ..., 0.00211851

AD_CP_R numeric Cerebral peduncle right (mean) Range value: 0.000725959, ..., 0.00176159

AD_CP_L numeric Cerebral peduncle left (mean) Range value: 0.000786702, ..., 0.0016027

AD_ALIC_R numeric Anterior limb of internal capsule right (mean) Range value: 0.00087333, ..., 0.00160752

AD_ALIC_L numeric Anterior limb of internal capsule left (mean) Range value: 0.000899174, ..., 0.00162506

AD_PLIC_R numeric Posterior limb of internal capsule right (mean) Range value: 0.000959815, ..., 0.00153946

AD_PLIC_L numeric Posterior limb of internal capsule left (mean) Range value: 0.000917365, ..., 0.00151523

AD_RLIC_R numeric Retrolenticular part of internal capsule right (mean) Range value: 0.000959267, ..., 0.00154893

AD_RLIC_L numeric Retrolenticular part of internal capsule left (mean) Range value: 0.000914247, ..., 0.0015178

AD_ACR_R numeric Anterior corona radiata right (mean) Range value: 0.000875194, ..., 0.00150636

AD_ACR_L numeric Anterior corona radiata left (mean) Range value: 0.000922371, ..., 0.00151156

AD_SCR_R numeric Superior corona radiata right (mean) Range value: 0.00090299, ..., 0.00159924

AD_SCR_L numeric Superior corona radiata left (mean) Range value: 0.000925883, ..., 0.0015824

AD_PCR_R numeric Posterior corona radiata right (mean) Range value: 0.00100159, ..., 0.00176965

AD_PCR_L numeric Posterior corona radiata left (mean) Range value: 0.000922813, ..., 0.00169638

AD_PTR_R numeric Posterior thalamic radiation right (mean) Range value: 0.00105155, ..., 0.00156374

AD_PTR_L numeric Posterior thalamic radiation left (mean) Range value: 0.00106455, ..., 0.00167979

AD_SS_R numeric Sagittal stratum right (mean) Range value: 0.00102514, ..., 0.00157003

AD_SS_L numeric Sagittal stratum left (mean) Range value: 0.000949998, ..., 0.00172715

AD_EC_R numeric External capsule right (mean) Range value: 0.000869933, ..., 0.00139585

AD_EC_L numeric External capsule left (mean) Range value: 0.000911379, ..., 0.00139066

AD_CGC_R numeric Cingulum right (mean) Range value: 0.000803926, ..., 0.00125218

AD_CGC_L numeric Cingulum left (mean) Range value: 0.000773778, ..., 0.00126339

AD_CGH_R numeric Cingulum right (mean) Range value: 0.000789796, ..., 0.00136406

AD_CGH_L numeric Cingulum left (mean) Range value: 0.000674585, ..., 0.00148286

AD_FXST_R numeric Fornix (cres) / Stria terminalis right (mean) Range value: 0.000902871, ..., 0.00181092

AD_FXST_L numeric Fornix (cres) / Stria terminalis left (mean) Range value: 0.000844339, ..., 0.00146839

AD_SLF_R numeric Superior longitudinal fasciculus right (mean) Range value: 0.000917611, ..., 0.00136946

AD_SLF_L numeric Superior longitudinal fasciculus left (mean) Range value: 0.000934935, ..., 0.00134069

AD_SFO_R numeric Superior fronto-occipital fasciculus right (mean) Range value: 0.000864122, ..., 0.00185215

AD_SFO_L numeric Superior fronto-occipital fasciculus left (mean) Range value: 0.000872557, ..., 0.00200136

AD_UNC_R numeric Uncinate fasciculus right (mean) Range value: 0.000828013, ..., 0.00164766

AD_UNC_L numeric Uncinate fasciculus left (mean) Range value: 0.000783775, ..., 0.00156054

AD_TAP_R numeric Tapetum right (mean) Range value: 0.00118081, ..., 0.00264749

AD_TAP_L numeric Tapetum left (mean) Range value: 0.00124293, ..., 0.00283334

AD_CST numeric Corticospinal tract Bilateral (mean) Range value: 0.000746304, ..., 0.0012439

AD_ML numeric Medial lemniscus Bilateral (mean) Range value: 0.000892504, ..., 0.00147262

AD_ICP numeric Inferior cerebellar peduncle Bilateral (mean) Range value: 0.000810984, ..., 0.00136443

AD_SCP numeric Superior cerebellar peduncle Bilateral (mean) Range value: 0.00104909, ..., 0.00209897

AD_CP numeric Cerebral peduncle Bilateral (mean) Range value: 0.000789979, ..., 0.00168134

AD_ALIC numeric Anterior limb of internal capsule Bilateral (mean) Range value: 0.000916326, ..., 0.00161495

AD_PLIC numeric Posterior limb of internal capsule Bilateral (mean) Range value: 0.000939453, ..., 0.00152027

AD_RLIC numeric Retrolenticular part of internal capsule Bilateral (mean) Range value: 0.0010485, ..., 0.00150684

AD_ACR numeric Anterior corona radiata Bilateral (mean) Range value: 0.000911011, ..., 0.00148156

AD_SCR numeric Superior corona radiata Bilateral (mean) Range value: 0.000926379, ..., 0.00159025

AD_PCR numeric Posterior corona radiata Bilateral (mean) Range value: 0.00101048, ..., 0.00173321

AD_PTR numeric Posterior thalamic radiation Bilateral (mean) Range value: 0.0010662, ..., 0.00155257

AD_SS numeric Sagittal stratum Bilateral (mean) Range value: 0.00102692, ..., 0.0015771

AD_EC numeric External capsule Bilateral (mean) Range value: 0.000901249, ..., 0.00137195

AD_CGC numeric Cingulum Bilateral (mean) Range value: 0.000787912, ..., 0.00125077

AD_CGH numeric Cingulum Bilateral (mean) Range value: 0.00073583, ..., 0.00137865

AD_FXST numeric Fornix (cres) / Stria terminalis Bilateral (mean) Range value: 0.000897928, ..., 0.00164401

AD_SLF numeric Superior longitudinal fasciculus Bilateral (mean) Range value: 0.000926227, ..., 0.00135512

AD_SFO numeric Superior fronto-occipital fasciculus Bilateral (mean) Range value: 0.000884636, ..., 0.00186439

AD_UNC numeric Uncinate fasciculus Bilateral (mean) Range value: 0.000812439, ..., 0.00161605

AD_TAP numeric Tapetum Bilateral (mean) Range value: 0.00131883, ..., 0.00267621

AD_IC numeric Full internal capsule (mean) Range value: 0.000962232, ..., 0.00150003

AD_CC numeric Full corpus callosum (mean) Range value: 0.00109255, ..., 0.00180409

AD_CR numeric Full corona radiata (mean) Range value: 0.000973624, ..., 0.00153466

AD_FULLWM numeric Full white matter (mean) Range value: 0.00101086, ..., 0.00149077

FA_MCP numeric Middle cerebellar peduncle (mean) Range value: 0.252345, ..., 0.550161

FA_PCT numeric Pontine crossing tract (mean) Range value: 0.210801, ..., 0.516748

FA_GCC numeric Genu of corpus callosum (mean) Range value: 0.294705, ..., 0.647485

FA_BCC numeric Body of corpus callosum (mean) Range value: 0.275639, ..., 0.672372

FA_SCC numeric Splenium of corpus callosum (mean) Range value: 0.311704, ..., 0.744461

FA_FX numeric Fornix (mean) Range value: 0.0611964, ..., 0.526047

FA_CST_R numeric Corticospinal tract right (mean) Range value: 0.229988, ..., 0.65296

FA_CST_L numeric Corticospinal tract left (mean) Range value: 0.22369, ..., 0.64481

FA_ML_R numeric Medial lemniscus right (mean) Range value: 0.281006, ..., 0.728448

FA_ML_L numeric Medial lemniscus left (mean) Range value: 0.286499, ..., 0.780731

FA_ICP_R numeric Inferior cerebellar peduncle right (mean) Range value: 0.199086, ..., 0.592228

FA_ICP_L numeric Inferior cerebellar peduncle left (mean) Range value: 0.202831, ..., 0.611582

FA_SCP_R numeric Superior cerebellar peduncle right (mean) Range value: 0.261885, ..., 0.718028

FA_SCP_L numeric Superior cerebellar peduncle left (mean) Range value: 0.270221, ..., 0.776648

FA_CP_R numeric Cerebral peduncle right (mean) Range value: 0.357331, ..., 0.786855

FA_CP_L numeric Cerebral peduncle left (mean) Range value: 0.364529, ..., 0.78557

FA_ALIC_R numeric Anterior limb of internal capsule right (mean) Range value: 0.317535, ..., 0.615578

FA_ALIC_L numeric Anterior limb of internal capsule left (mean) Range value: 0.256225, ..., 0.607676

FA_PLIC_R numeric Posterior limb of internal capsule right (mean) Range value: 0.504614, ..., 0.71658

FA_PLIC_L numeric Posterior limb of internal capsule left (mean) Range value: 0.420518, ..., 0.742459

FA_RLIC_R numeric Retrolenticular part of internal capsule right (mean) Range value: 0.367615, ..., 0.750679

FA_RLIC_L numeric Retrolenticular part of internal capsule left (mean) Range value: 0.296874, ..., 0.665738

FA_ACR_R numeric Anterior corona radiata right (mean) Range value: 0.21618, ..., 0.475171

FA_ACR_L numeric Anterior corona radiata left (mean) Range value: 0.204649, ..., 0.460778

FA_SCR_R numeric Superior corona radiata right (mean) Range value: 0.281018, ..., 0.561446

FA_SCR_L numeric Superior corona radiata left (mean) Range value: 0.293459, ..., 0.581761

FA_PCR_R numeric Posterior corona radiata right (mean) Range value: 0.264524, ..., 0.518704

FA_PCR_L numeric Posterior corona radiata left (mean) Range value: 0.247135, ..., 0.522745

FA_PTR_R numeric Posterior thalamic radiation right (mean) Range value: 0.291212, ..., 0.628348

FA_PTR_L numeric Posterior thalamic radiation left (mean) Range value: 0.294805, ..., 0.589033

FA_SS_R numeric Sagittal stratum right (mean) Range value: 0.258145, ..., 0.609787

FA_SS_L numeric Sagittal stratum left (mean) Range value: 0.293284, ..., 0.615108

FA_EC_R numeric External capsule right (mean) Range value: 0.214441, ..., 0.465525

FA_EC_L numeric External capsule left (mean) Range value: 0.202631, ..., 0.486719

FA_CGC_R numeric Cingulum right (mean) Range value: 0.199259, ..., 0.536063

FA_CGC_L numeric Cingulum left (mean) Range value: 0.184967, ..., 0.547625

FA_CGH_R numeric Cingulum right (mean) Range value: 0.118283, ..., 0.481906

FA_CGH_L numeric Cingulum left (mean) Range value: 0.10786, ..., 0.466886

FA_FXST_R numeric Fornix (cres) / Stria terminalis right (mean) Range value: 0.210284, ..., 0.544872

FA_FXST_L numeric Fornix (cres) / Stria terminalis left (mean) Range value: 0.235789, ..., 0.592987

FA_SLF_R numeric Superior longitudinal fasciculus right (mean) Range value: 0.301397, ..., 0.538221

FA_SLF_L numeric Superior longitudinal fasciculus left (mean) Range value: 0.274124, ..., 0.525935

FA_SFO_R numeric Superior fronto-occipital fasciculus right (mean) Range value: 0.258072, ..., 0.505397

FA_SFO_L numeric Superior fronto-occipital fasciculus left (mean) Range value: 0.213813, ..., 0.480173

FA_UNC_R numeric Uncinate fasciculus right (mean) Range value: 0.142727, ..., 0.558439

FA_UNC_L numeric Uncinate fasciculus left (mean) Range value: 0.162396, ..., 0.563396

FA_TAP_R numeric Tapetum right (mean) Range value: 0.197727, ..., 0.573132

FA_TAP_L numeric Tapetum left (mean) Range value: 0.134764, ..., 0.503668

FA_CST numeric Corticospinal tract Bilateral (mean) Range value: 0.226778, ..., 0.648897

FA_ML numeric Medial lemniscus Bilateral (mean) Range value: 0.283881, ..., 0.745101

FA_ICP numeric Inferior cerebellar peduncle Bilateral (mean) Range value: 0.201763, ..., 0.585665

FA_SCP numeric Superior cerebellar peduncle Bilateral (mean) Range value: 0.268691, ..., 0.718652

FA_CP numeric Cerebral peduncle Bilateral (mean) Range value: 0.366346, ..., 0.786166

FA_ALIC numeric Anterior limb of internal capsule Bilateral (mean) Range value: 0.28777, ..., 0.606564

FA_PLIC numeric Posterior limb of internal capsule Bilateral (mean) Range value: 0.473246, ..., 0.708232

FA_RLIC numeric Retrolenticular part of internal capsule Bilateral (mean) Range value: 0.362365, ..., 0.658326

FA_ACR numeric Anterior corona radiata Bilateral (mean) Range value: 0.216347, ..., 0.454851

FA_SCR numeric Superior corona radiata Bilateral (mean) Range value: 0.287231, ..., 0.571456

FA_PCR numeric Posterior corona radiata Bilateral (mean) Range value: 0.25717, ..., 0.516042

FA_PTR numeric Posterior thalamic radiation Bilateral (mean) Range value: 0.329049, ..., 0.609791

FA_SS numeric Sagittal stratum Bilateral (mean) Range value: 0.277118, ..., 0.594567

FA_EC numeric External capsule Bilateral (mean) Range value: 0.210151, ..., 0.464564

FA_CGC numeric Cingulum Bilateral (mean) Range value: 0.191668, ..., 0.536745

FA_CGH numeric Cingulum Bilateral (mean) Range value: 0.113071, ..., 0.460723

FA_FXST numeric Fornix (cres) / Stria terminalis Bilateral (mean) Range value: 0.222603, ..., 0.557311

FA_SLF numeric Superior longitudinal fasciculus Bilateral (mean) Range value: 0.297422, ..., 0.530435

FA_SFO numeric Superior fronto-occipital fasciculus Bilateral (mean) Range value: 0.240942, ..., 0.472103

FA_UNC numeric Uncinate fasciculus Bilateral (mean) Range value: 0.152682, ..., 0.549497

FA_TAP numeric Tapetum Bilateral (mean) Range value: 0.187994, ..., 0.531757

FA_IC numeric Full internal capsule (mean) Range value: 0.418186, ..., 0.646093

FA_CC numeric Full corpus callosum (mean) Range value: 0.303971, ..., 0.68485

FA_CR numeric Full corona radiata (mean) Range value: 0.25626, ..., 0.508249

FA_FULLWM numeric Full white matter (mean) Range value: 0.339006, ..., 0.552456

MD_MCP numeric Middle cerebellar peduncle (mean) Range value: 0.000538402, ..., 0.000922075

MD_PCT numeric Pontine crossing tract (mean) Range value: 0.000438231, ..., 0.000876774

MD_GCC numeric Genu of corpus callosum (mean) Range value: 0.000698504, ..., 0.00117093

MD_BCC numeric Body of corpus callosum (mean) Range value: 0.000736046, ..., 0.00124782

MD_SCC numeric Splenium of corpus callosum (mean) Range value: 0.00062599, ..., 0.00125812

MD_FX numeric Fornix (mean) Range value: 0.00104399, ..., 0.00298852

MD_CST_R numeric Corticospinal tract right (mean) Range value: 0.000408575, ..., 0.000918554

MD_CST_L numeric Corticospinal tract left (mean) Range value: 0.000444655, ..., 0.000899429

MD_ML_R numeric Medial lemniscus right (mean) Range value: 0.000532252, ..., 0.00100212

MD_ML_L numeric Medial lemniscus left (mean) Range value: 0.000508878, ..., 0.0010474

MD_ICP_R numeric Inferior cerebellar peduncle right (mean) Range value: 0.000494412, ..., 0.000925809

MD_ICP_L numeric Inferior cerebellar peduncle left (mean) Range value: 0.000524807, ..., 0.000944346

MD_SCP_R numeric Superior cerebellar peduncle right (mean) Range value: 0.000700557, ..., 0.00161914

MD_SCP_L numeric Superior cerebellar peduncle left (mean) Range value: 0.000546938, ..., 0.0014295

MD_CP_R numeric Cerebral peduncle right (mean) Range value: 0.000368035, ..., 0.00106217

MD_CP_L numeric Cerebral peduncle left (mean) Range value: 0.000410472, ..., 0.00104219

MD_ALIC_R numeric Anterior limb of internal capsule right (mean) Range value: 0.000578151, ..., 0.00117585

MD_ALIC_L numeric Anterior limb of internal capsule left (mean) Range value: 0.000583517, ..., 0.00116933

MD_PLIC_R numeric Posterior limb of internal capsule right (mean) Range value: 0.000541061, ..., 0.000899276

MD_PLIC_L numeric Posterior limb of internal capsule left (mean) Range value: 0.000542291, ..., 0.000896734

MD_RLIC_R numeric Retrolenticular part of internal capsule right (mean) Range value: 0.00050776, ..., 0.000950347

MD_RLIC_L numeric Retrolenticular part of internal capsule left (mean) Range value: 0.000636994, ..., 0.000962304

MD_ACR_R numeric Anterior corona radiata right (mean) Range value: 0.000651486, ..., 0.0012336

MD_ACR_L numeric Anterior corona radiata left (mean) Range value: 0.000654708, ..., 0.00123377

MD_SCR_R numeric Superior corona radiata right (mean) Range value: 0.000611519, ..., 0.00113424

MD_SCR_L numeric Superior corona radiata left (mean) Range value: 0.000609465, ..., 0.00110716

MD_PCR_R numeric Posterior corona radiata right (mean) Range value: 0.000666802, ..., 0.00136957

MD_PCR_L numeric Posterior corona radiata left (mean) Range value: 0.000669869, ..., 0.00134033

MD_PTR_R numeric Posterior thalamic radiation right (mean) Range value: 0.000642956, ..., 0.00110866

MD_PTR_L numeric Posterior thalamic radiation left (mean) Range value: 0.000670468, ..., 0.00124786

MD_SS_R numeric Sagittal stratum right (mean) Range value: 0.00064897, ..., 0.00106108

MD_SS_L numeric Sagittal stratum left (mean) Range value: 0.000650695, ..., 0.00128965

MD_EC_R numeric External capsule right (mean) Range value: 0.000634019, ..., 0.00106651

MD_EC_L numeric External capsule left (mean) Range value: 0.000631123, ..., 0.0010586

MD_CGC_R numeric Cingulum right (mean) Range value: 0.00061873, ..., 0.000888201

MD_CGC_L numeric Cingulum left (mean) Range value: 0.000642707, ..., 0.000916112

MD_CGH_R numeric Cingulum right (mean) Range value: 0.000611251, ..., 0.00105117

MD_CGH_L numeric Cingulum left (mean) Range value: 0.000513676, ..., 0.00120918

MD_FXST_R numeric Fornix (cres) / Stria terminalis right (mean) Range value: 0.000658871, ..., 0.00127149

MD_FXST_L numeric Fornix (cres) / Stria terminalis left (mean) Range value: 0.000579136, ..., 0.00105693

MD_SLF_R numeric Superior longitudinal fasciculus right (mean) Range value: 0.000623397, ..., 0.00103359

MD_SLF_L numeric Superior longitudinal fasciculus left (mean) Range value: 0.000621985, ..., 0.0010204

MD_SFO_R numeric Superior fronto-occipital fasciculus right (mean) Range value: 0.000607355, ..., 0.00144193

MD_SFO_L numeric Superior fronto-occipital fasciculus left (mean) Range value: 0.000609021, ..., 0.00164983

MD_UNC_R numeric Uncinate fasciculus right (mean) Range value: 0.000560042, ..., 0.00130304

MD_UNC_L numeric Uncinate fasciculus left (mean) Range value: 0.000497959, ..., 0.00120965

MD_TAP_R numeric Tapetum right (mean) Range value: 0.000890469, ..., 0.00210118

MD_TAP_L numeric Tapetum left (mean) Range value: 0.000879502, ..., 0.00237742

MD_CST numeric Corticospinal tract Bilateral (mean) Range value: 0.000425304, ..., 0.000902124

MD_ML numeric Medial lemniscus Bilateral (mean) Range value: 0.000531991, ..., 0.00102379

MD_ICP numeric Inferior cerebellar peduncle Bilateral (mean) Range value: 0.000511224, ..., 0.00088883

MD_SCP numeric Superior cerebellar peduncle Bilateral (mean) Range value: 0.000625826, ..., 0.00151651

MD_CP numeric Cerebral peduncle Bilateral (mean) Range value: 0.000391221, ..., 0.00104791

MD_ALIC numeric Anterior limb of internal capsule Bilateral (mean) Range value: 0.000583594, ..., 0.00117308

MD_PLIC numeric Posterior limb of internal capsule Bilateral (mean) Range value: 0.000548241, ..., 0.000890865

MD_RLIC numeric Retrolenticular part of internal capsule Bilateral (mean) Range value: 0.000607829, ..., 0.00095053

MD_ACR numeric Anterior corona radiata Bilateral (mean) Range value: 0.000653176, ..., 0.00118316

MD_SCR numeric Superior corona radiata Bilateral (mean) Range value: 0.000614531, ..., 0.00111121

MD_PCR numeric Posterior corona radiata Bilateral (mean) Range value: 0.000674202, ..., 0.00135503

MD_PTR numeric Posterior thalamic radiation Bilateral (mean) Range value: 0.000657147, ..., 0.00112437

MD_SS numeric Sagittal stratum Bilateral (mean) Range value: 0.000680688, ..., 0.00111788

MD_EC numeric External capsule Bilateral (mean) Range value: 0.000632461, ..., 0.00105231

MD_CGC numeric Cingulum Bilateral (mean) Range value: 0.000635599, ..., 0.000887752

MD_CGH numeric Cingulum Bilateral (mean) Range value: 0.000558496, ..., 0.00111457

MD_FXST numeric Fornix (cres) / Stria terminalis Bilateral (mean) Range value: 0.000658591, ..., 0.0011443

MD_SLF numeric Superior longitudinal fasciculus Bilateral (mean) Range value: 0.000622668, ..., 0.00102702

MD_SFO numeric Superior fronto-occipital fasciculus Bilateral (mean) Range value: 0.000612006, ..., 0.00154059

MD_UNC numeric Uncinate fasciculus Bilateral (mean) Range value: 0.000575353, ..., 0.00126271

MD_TAP numeric Tapetum Bilateral (mean) Range value: 0.000931168, ..., 0.00224993

MD_IC numeric Full internal capsule (mean) Range value: 0.000581847, ..., 0.000970623

MD_CC numeric Full corpus callosum (mean) Range value: 0.000702356, ..., 0.00120624

MD_CR numeric Full corona radiata (mean) Range value: 0.000658942, ..., 0.00117606

MD_FULLWM numeric Full white matter (mean) Range value: 0.000641979, ..., 0.0010227

RD_MCP numeric Middle cerebellar peduncle (mean) Range value: 0.000360952, ..., 0.000707273

RD_PCT numeric Pontine crossing tract (mean) Range value: 0.000339464, ..., 0.000738328

RD_GCC numeric Genu of corpus callosum (mean) Range value: 0.000407506, ..., 0.000937897

RD_BCC numeric Body of corpus callosum (mean) Range value: 0.000417446, ..., 0.00100016

RD_SCC numeric Splenium of corpus callosum (mean) Range value: 0.000299174, ..., 0.00101665

RD_FX numeric Fornix (mean) Range value: 0.000747619, ..., 0.00279938

RD_CST_R numeric Corticospinal tract right (mean) Range value: 0.000261961, ..., 0.000764234

RD_CST_L numeric Corticospinal tract left (mean) Range value: 0.000273211, ..., 0.000717467

RD_ML_R numeric Medial lemniscus right (mean) Range value: 0.000248836, ..., 0.000786482

RD_ML_L numeric Medial lemniscus left (mean) Range value: 0.000205269, ..., 0.000847218

RD_ICP_R numeric Inferior cerebellar peduncle right (mean) Range value: 0.000326104, ..., 0.00073094

RD_ICP_L numeric Inferior cerebellar peduncle left (mean) Range value: 0.000339415, ..., 0.000771752

RD_SCP_R numeric Superior cerebellar peduncle right (mean) Range value: 0.000400842, ..., 0.00133279

RD_SCP_L numeric Superior cerebellar peduncle left (mean) Range value: 0.00036689, ..., 0.00113412

RD_CP_R numeric Cerebral peduncle right (mean) Range value: 0.000240145, ..., 0.000780001

RD_CP_L numeric Cerebral peduncle left (mean) Range value: 0.000225811, ..., 0.000797323

RD_ALIC_R numeric Anterior limb of internal capsule right (mean) Range value: 0.000387114, ..., 0.000960008

RD_ALIC_L numeric Anterior limb of internal capsule left (mean) Range value: 0.000346503, ..., 0.000941463

RD_PLIC_R numeric Posterior limb of internal capsule right (mean) Range value: 0.00027324, ..., 0.00058138

RD_PLIC_L numeric Posterior limb of internal capsule left (mean) Range value: 0.000294144, ..., 0.00063928

RD_RLIC_R numeric Retrolenticular part of internal capsule right (mean) Range value: 0.00023902, ..., 0.000717964

RD_RLIC_L numeric Retrolenticular part of internal capsule left (mean) Range value: 0.00036453, ..., 0.000717901

RD_ACR_R numeric Anterior corona radiata right (mean) Range value: 0.000508525, ..., 0.0010987

RD_ACR_L numeric Anterior corona radiata left (mean) Range value: 0.000519783, ..., 0.00109487

RD_SCR_R numeric Superior corona radiata right (mean) Range value: 0.000435616, ..., 0.000965589

RD_SCR_L numeric Superior corona radiata left (mean) Range value: 0.000435086, ..., 0.000919441

RD_PCR_R numeric Posterior corona radiata right (mean) Range value: 0.000481327, ..., 0.00116953

RD_PCR_L numeric Posterior corona radiata left (mean) Range value: 0.000497635, ..., 0.00116231

RD_PTR_R numeric Posterior thalamic radiation right (mean) Range value: 0.000391309, ..., 0.000911663

RD_PTR_L numeric Posterior thalamic radiation left (mean) Range value: 0.000457335, ..., 0.0010319

RD_SS_R numeric Sagittal stratum right (mean) Range value: 0.000434705, ..., 0.000844208

RD_SS_L numeric Sagittal stratum left (mean) Range value: 0.000416646, ..., 0.0010709

RD_EC_R numeric External capsule right (mean) Range value: 0.000495253, ..., 0.000945899

RD_EC_L numeric External capsule left (mean) Range value: 0.000464835, ..., 0.000929433

RD_CGC_R numeric Cingulum right (mean) Range value: 0.00045234, ..., 0.000766282

RD_CGC_L numeric Cingulum left (mean) Range value: 0.000438164, ..., 0.000765136

RD_CGH_R numeric Cingulum right (mean) Range value: 0.000473599, ..., 0.000932208

RD_CGH_L numeric Cingulum left (mean) Range value: 0.000433814, ..., 0.00107233

RD_FXST_R numeric Fornix (cres) / Stria terminalis right (mean) Range value: 0.000504701, ..., 0.00107425

RD_FXST_L numeric Fornix (cres) / Stria terminalis left (mean) Range value: 0.000379256, ..., 0.000892421

RD_SLF_R numeric Superior longitudinal fasciculus right (mean) Range value: 0.0004248, ..., 0.00086565

RD_SLF_L numeric Superior longitudinal fasciculus left (mean) Range value: 0.000451681, ..., 0.000860261

RD_SFO_R numeric Superior fronto-occipital fasciculus right (mean) Range value: 0.000462817, ..., 0.00126797

RD_SFO_L numeric Superior fronto-occipital fasciculus left (mean) Range value: 0.000457079, ..., 0.00149614

RD_UNC_R numeric Uncinate fasciculus right (mean) Range value: 0.000417975, ..., 0.00113286

RD_UNC_L numeric Uncinate fasciculus left (mean) Range value: 0.000345537, ..., 0.00103479

RD_TAP_R numeric Tapetum right (mean) Range value: 0.000600145, ..., 0.00185332

RD_TAP_L numeric Tapetum left (mean) Range value: 0.000697786, ..., 0.00219405

RD_CST numeric Corticospinal tract Bilateral (mean) Range value: 0.000271487, ..., 0.00073592

RD_ML numeric Medial lemniscus Bilateral (mean) Range value: 0.000242604, ..., 0.000815548

RD_ICP numeric Inferior cerebellar peduncle Bilateral (mean) Range value: 0.000339417, ..., 0.000727756

RD_SCP numeric Superior cerebellar peduncle Bilateral (mean) Range value: 0.000399202, ..., 0.00122528

RD_CP numeric Cerebral peduncle Bilateral (mean) Range value: 0.000232078, ..., 0.000787083

RD_ALIC numeric Anterior limb of internal capsule Bilateral (mean) Range value: 0.000393487, ..., 0.000952149

RD_PLIC numeric Posterior limb of internal capsule Bilateral (mean) Range value: 0.000291553, ..., 0.000586358

RD_RLIC numeric Retrolenticular part of internal capsule Bilateral (mean) Range value: 0.000360929, ..., 0.000708689

RD_ACR numeric Anterior corona radiata Bilateral (mean) Range value: 0.000515003, ..., 0.00104115

RD_SCR numeric Superior corona radiata Bilateral (mean) Range value: 0.000446576, ..., 0.000942543

RD_PCR numeric Posterior corona radiata Bilateral (mean) Range value: 0.000488885, ..., 0.00116594

RD_PTR numeric Posterior thalamic radiation Bilateral (mean) Range value: 0.000438526, ..., 0.000914865

RD_SS numeric Sagittal stratum Bilateral (mean) Range value: 0.000425496, ..., 0.000888264

RD_EC numeric External capsule Bilateral (mean) Range value: 0.000492962, ..., 0.000928367

RD_CGC numeric Cingulum Bilateral (mean) Range value: 0.000450064, ..., 0.00076197

RD_CGH numeric Cingulum Bilateral (mean) Range value: 0.000470176, ..., 0.000982526

RD_FXST numeric Fornix (cres) / Stria terminalis Bilateral (mean) Range value: 0.000454056, ..., 0.000935865

RD_SLF numeric Superior longitudinal fasciculus Bilateral (mean) Range value: 0.000444042, ..., 0.000862964

RD_SFO numeric Superior fronto-occipital fasciculus Bilateral (mean) Range value: 0.000471453, ..., 0.00137869

RD_UNC numeric Uncinate fasciculus Bilateral (mean) Range value: 0.000434886, ..., 0.00110189

RD_TAP numeric Tapetum Bilateral (mean) Range value: 0.000710458, ..., 0.00203679

RD_IC numeric Full internal capsule (mean) Range value: 0.000353446, ..., 0.000705921

RD_CC numeric Full corpus callosum (mean) Range value: 0.000387306, ..., 0.000926626

RD_CR numeric Full corona radiata (mean) Range value: 0.000500982, ..., 0.001016

RD_FULLWM numeric Full white matter (mean) Range value: 0.000434768, ..., 0.000821996

update_stamp POSIXct: YYY-MM-DD