USC - DTI ROI Summary Measures v2 + v3 (Robust Mean)

USC - DTI ROI Summary Measures v2 + v3 (Robust Mean) data. More information is available at https://adni.loni.usc.edu/data-samples/adni-data/

Usage

data(DTIROI_ROBUSTMEAN)

Format

A data frame with 1750 observations and 310 variables.

Source

https://adni.loni.usc.edu/data-samples/adni-data/

Details

ORIGPROT character Original study protocol , with options: ADNI1, ADNI2, ADNI3, ADNIGO, ADNI4

COLPROT factor Study protocol of data collection , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4

PTID character Participant ID

RID numeric Participant roster ID Range value: 11, ..., 10793

VISCODE character Visit code , with options: init, y2, y1, sc, y4, y3, 4_sc, 4_init, 4_m12

VISCODE2 character Translated visit code

EXAMDATE Examination Date Date: YYY-MM-DD

VERSION character Preprocessing pipeline version 2 = preprocessing pipeline as outlined in methods PDF: (USC - DTI ROI Summary Measures Methods v2 [ADNI3], date: 12/01/2021) ; 3 = preprocessing pipeline as outlined in methods PDF: (USC - DTI ROI Summary Measures Methods v3 [ADNI4], date: 04/02/2025) ; -4 = image was not processed as noted in QC column. Decoded Value: Yes

MANUFACTURER character Scanner Manufacturer , with options: Siemens, GE, Philips, SIEMENS, GE MEDICAL SYSTEMS

VOLUMES numeric Number of volumes prescribed by acquisition protocol Range value: 31, ..., 127

IMAGEUID numeric IMAGE ID of the raw diffusion scan Range value: 808826, ..., 11303224

DISTORTION_CORRECTION character In V3, EPI susceptibility distortion correction was run using a synthetic b0 (Synb0), b0s acquired with opposite phase encoding directions (APPA), or was completed in scanner (scanner), in addition to warping DWIs to T1w. In V2, susceptibility distortion was only corrected by warping DWIs to T1w (empty cells). Synb0=synthetic b0 ; APPA=b0s acquired with opposite phase encoding directions ; scanner=completed in scanner.

VOLUMES_AP numeric Number of volumes in the opposite phase encoding direction raw scan. Empty cell indicates there was no opposite phase encoding image provided to users on the IDA. Range value: 7, ..., 13

IMAGEUID_AP numeric IMAGE ID of the raw diffusion scan with an opposite phase encoding direction. Empty cell indicates there was no opposite phase encoding image provided to users on the IDA. Range value: 10248884, ..., 11297753

RUNDATE Processing Date Date: YYY-MM-DD

STATUS character Processing Completed , with options: complete, not processed

QC character Preprocessed DWI and scalar map quality check notes include: (1) a note if raw DWI were not processed due to incomplete acquisition (e.g., fewer volumes acquired than protocol), failed DICOM conversion, poor field of view (i.e., image cropped), or severe artifacts (e.g., motion, distortion); or (2) notes if preprocessed DWI had any remaining artifacts (e.g., motion, distortion, ghosting, hyperintensities in FA, etc.), pathology or anatomical anomalies, atlas registration failure, or if the voxel size differed from the prescribed acquisition protocol. These images should be used at own discretion.

AD_MCP numeric Middle cerebellar peduncle (robust mean) Range value: 0.000829046, ..., 0.001317

AD_PCT numeric Pontine crossing tract (robust mean) Range value: 0.000635654, ..., 0.0012143

AD_GCC numeric Genu of corpus callosum (robust mean) Range value: 0.00107923, ..., 0.0017148

AD_BCC numeric Body of corpus callosum (robust mean) Range value: 0.00114831, ..., 0.00179691

AD_SCC numeric Splenium of corpus callosum (robust mean) Range value: 0.000931843, ..., 0.00181219

AD_FX numeric Fornix (robust mean) Range value: 0.00136773, ..., 0.00343132

AD_CST_R numeric Corticospinal tract right (robust mean) Range value: 0.000700309, ..., 0.00123863

AD_CST_L numeric Corticospinal tract left (robust mean) Range value: 0.000727272, ..., 0.00127933

AD_ML_R numeric Medial lemniscus right (robust mean) Range value: 0.000871556, ..., 0.00148612

AD_ML_L numeric Medial lemniscus left (robust mean) Range value: 0.000904022, ..., 0.00145631

AD_ICP_R numeric Inferior cerebellar peduncle right (robust mean) Range value: 0.000769043, ..., 0.00134749

AD_ICP_L numeric Inferior cerebellar peduncle left (robust mean) Range value: 0.000790416, ..., 0.00132972

AD_SCP_R numeric Superior cerebellar peduncle right (robust mean) Range value: 0.00107794, ..., 0.00216635

AD_SCP_L numeric Superior cerebellar peduncle left (robust mean) Range value: 0.000933815, ..., 0.00206616

AD_CP_R numeric Cerebral peduncle right (robust mean) Range value: 0.000734282, ..., 0.00172317

AD_CP_L numeric Cerebral peduncle left (robust mean) Range value: 0.000790256, ..., 0.00156891

AD_ALIC_R numeric Anterior limb of internal capsule right (robust mean) Range value: 0.000877232, ..., 0.00160427

AD_ALIC_L numeric Anterior limb of internal capsule left (robust mean) Range value: 0.000899512, ..., 0.00163145

AD_PLIC_R numeric Posterior limb of internal capsule right (robust mean) Range value: 0.000964611, ..., 0.0015394

AD_PLIC_L numeric Posterior limb of internal capsule left (robust mean) Range value: 0.000920428, ..., 0.00150803

AD_RLIC_R numeric Retrolenticular part of internal capsule right (robust mean) Range value: 0.000967808, ..., 0.00151841

AD_RLIC_L numeric Retrolenticular part of internal capsule left (robust mean) Range value: 0.000913532, ..., 0.0015243

AD_ACR_R numeric Anterior corona radiata right (robust mean) Range value: 0.000870891, ..., 0.00149696

AD_ACR_L numeric Anterior corona radiata left (robust mean) Range value: 0.000917782, ..., 0.00150297

AD_SCR_R numeric Superior corona radiata right (robust mean) Range value: 0.000897053, ..., 0.00159583

AD_SCR_L numeric Superior corona radiata left (robust mean) Range value: 0.000922094, ..., 0.00156574

AD_PCR_R numeric Posterior corona radiata right (robust mean) Range value: 0.000991979, ..., 0.00178614

AD_PCR_L numeric Posterior corona radiata left (robust mean) Range value: 0.000910448, ..., 0.00172159

AD_PTR_R numeric Posterior thalamic radiation right (robust mean) Range value: 0.00101088, ..., 0.00155547

AD_PTR_L numeric Posterior thalamic radiation left (robust mean) Range value: 0.00105047, ..., 0.00163995

AD_SS_R numeric Sagittal stratum right (robust mean) Range value: 0.00102135, ..., 0.00155689

AD_SS_L numeric Sagittal stratum left (robust mean) Range value: 0.000930755, ..., 0.00173607

AD_EC_R numeric External capsule right (robust mean) Range value: 0.000870701, ..., 0.00138704

AD_EC_L numeric External capsule left (robust mean) Range value: 0.00090765, ..., 0.00139369

AD_CGC_R numeric Cingulum right (robust mean) Range value: 0.00079713, ..., 0.00124808

AD_CGC_L numeric Cingulum left (robust mean) Range value: 0.00077632, ..., 0.00125187

AD_CGH_R numeric Cingulum right (robust mean) Range value: 0.000783423, ..., 0.00129411

AD_CGH_L numeric Cingulum left (robust mean) Range value: 0.000656679, ..., 0.0014759

AD_FXST_R numeric Fornix (cres) / Stria terminalis right (robust mean) Range value: 0.000898062, ..., 0.00179612

AD_FXST_L numeric Fornix (cres) / Stria terminalis left (robust mean) Range value: 0.000832679, ..., 0.00145359

AD_SLF_R numeric Superior longitudinal fasciculus right (robust mean) Range value: 0.00091249, ..., 0.00136921

AD_SLF_L numeric Superior longitudinal fasciculus left (robust mean) Range value: 0.000929409, ..., 0.00134061

AD_SFO_R numeric Superior fronto-occipital fasciculus right (robust mean) Range value: 0.000863616, ..., 0.00183968

AD_SFO_L numeric Superior fronto-occipital fasciculus left (robust mean) Range value: 0.000868013, ..., 0.00190102

AD_UNC_R numeric Uncinate fasciculus right (robust mean) Range value: 0.00082834, ..., 0.00157334

AD_UNC_L numeric Uncinate fasciculus left (robust mean) Range value: 0.000779406, ..., 0.00151033

AD_TAP_R numeric Tapetum right (robust mean) Range value: 0.00116991, ..., 0.00270459

AD_TAP_L numeric Tapetum left (robust mean) Range value: 0.00120224, ..., 0.0028861

AD_CST numeric Corticospinal tract Bilateral (robust mean) Range value: 0.000747601, ..., 0.00123127

AD_ML numeric Medial lemniscus Bilateral (robust mean) Range value: 0.000906207, ..., 0.00145263

AD_ICP numeric Inferior cerebellar peduncle Bilateral (robust mean) Range value: 0.000804694, ..., 0.00133767

AD_SCP numeric Superior cerebellar peduncle Bilateral (robust mean) Range value: 0.00102481, ..., 0.00207276

AD_CP numeric Cerebral peduncle Bilateral (robust mean) Range value: 0.000802573, ..., 0.00164113

AD_ALIC numeric Anterior limb of internal capsule Bilateral (robust mean) Range value: 0.00091851, ..., 0.00161555

AD_PLIC numeric Posterior limb of internal capsule Bilateral (robust mean) Range value: 0.000942687, ..., 0.00151642

AD_RLIC numeric Retrolenticular part of internal capsule Bilateral (robust mean) Range value: 0.00104101, ..., 0.00150999

AD_ACR numeric Anterior corona radiata Bilateral (robust mean) Range value: 0.000905175, ..., 0.00147056

AD_SCR numeric Superior corona radiata Bilateral (robust mean) Range value: 0.000922737, ..., 0.0015798

AD_PCR numeric Posterior corona radiata Bilateral (robust mean) Range value: 0.00100198, ..., 0.00175337

AD_PTR numeric Posterior thalamic radiation Bilateral (robust mean) Range value: 0.00105728, ..., 0.00152739

AD_SS numeric Sagittal stratum Bilateral (robust mean) Range value: 0.00101621, ..., 0.00155588

AD_EC numeric External capsule Bilateral (robust mean) Range value: 0.000898161, ..., 0.00136933

AD_CGC numeric Cingulum Bilateral (robust mean) Range value: 0.00078856, ..., 0.0012405

AD_CGH numeric Cingulum Bilateral (robust mean) Range value: 0.000731739, ..., 0.00135534

AD_FXST numeric Fornix (cres) / Stria terminalis Bilateral (robust mean) Range value: 0.000869624, ..., 0.00162273

AD_SLF numeric Superior longitudinal fasciculus Bilateral (robust mean) Range value: 0.000920892, ..., 0.00135425

AD_SFO numeric Superior fronto-occipital fasciculus Bilateral (robust mean) Range value: 0.000869218, ..., 0.00179869

AD_UNC numeric Uncinate fasciculus Bilateral (robust mean) Range value: 0.000802693, ..., 0.00154848

AD_TAP numeric Tapetum Bilateral (robust mean) Range value: 0.00124986, ..., 0.00280773

AD_IC numeric Full internal capsule (robust mean) Range value: 0.000965563, ..., 0.00149749

AD_CC numeric Full corpus callosum (robust mean) Range value: 0.00106926, ..., 0.00175088

AD_CR numeric Full corona radiata (robust mean) Range value: 0.000969116, ..., 0.00152303

AD_FULLWM numeric Full white matter (robust mean) Range value: 0.000996245, ..., 0.00142226

FA_MCP numeric Middle cerebellar peduncle (robust mean) Range value: 0.248813, ..., 0.549816

FA_PCT numeric Pontine crossing tract (robust mean) Range value: 0.200079, ..., 0.518204

FA_GCC numeric Genu of corpus callosum (robust mean) Range value: 0.294355, ..., 0.652468

FA_BCC numeric Body of corpus callosum (robust mean) Range value: 0.266139, ..., 0.676548

FA_SCC numeric Splenium of corpus callosum (robust mean) Range value: 0.302237, ..., 0.755416

FA_FX numeric Fornix (robust mean) Range value: 0.0586734, ..., 0.532278

FA_CST_R numeric Corticospinal tract right (robust mean) Range value: 0.230267, ..., 0.656646

FA_CST_L numeric Corticospinal tract left (robust mean) Range value: 0.226838, ..., 0.650365

FA_ML_R numeric Medial lemniscus right (robust mean) Range value: 0.279958, ..., 0.7301

FA_ML_L numeric Medial lemniscus left (robust mean) Range value: 0.287144, ..., 0.787983

FA_ICP_R numeric Inferior cerebellar peduncle right (robust mean) Range value: 0.193219, ..., 0.596592

FA_ICP_L numeric Inferior cerebellar peduncle left (robust mean) Range value: 0.205008, ..., 0.595435

FA_SCP_R numeric Superior cerebellar peduncle right (robust mean) Range value: 0.257702, ..., 0.731428

FA_SCP_L numeric Superior cerebellar peduncle left (robust mean) Range value: 0.268533, ..., 0.786693

FA_CP_R numeric Cerebral peduncle right (robust mean) Range value: 0.32235, ..., 0.788253

FA_CP_L numeric Cerebral peduncle left (robust mean) Range value: 0.350584, ..., 0.783827

FA_ALIC_R numeric Anterior limb of internal capsule right (robust mean) Range value: 0.305473, ..., 0.628841

FA_ALIC_L numeric Anterior limb of internal capsule left (robust mean) Range value: 0.239185, ..., 0.617397

FA_PLIC_R numeric Posterior limb of internal capsule right (robust mean) Range value: 0.498038, ..., 0.724186

FA_PLIC_L numeric Posterior limb of internal capsule left (robust mean) Range value: 0.411263, ..., 0.751875

FA_RLIC_R numeric Retrolenticular part of internal capsule right (robust mean) Range value: 0.36305, ..., 0.738754

FA_RLIC_L numeric Retrolenticular part of internal capsule left (robust mean) Range value: 0.295967, ..., 0.665867

FA_ACR_R numeric Anterior corona radiata right (robust mean) Range value: 0.206771, ..., 0.475391

FA_ACR_L numeric Anterior corona radiata left (robust mean) Range value: 0.203699, ..., 0.464372

FA_SCR_R numeric Superior corona radiata right (robust mean) Range value: 0.259513, ..., 0.559231

FA_SCR_L numeric Superior corona radiata left (robust mean) Range value: 0.282371, ..., 0.588552

FA_PCR_R numeric Posterior corona radiata right (robust mean) Range value: 0.249689, ..., 0.519696

FA_PCR_L numeric Posterior corona radiata left (robust mean) Range value: 0.232454, ..., 0.519685

FA_PTR_R numeric Posterior thalamic radiation right (robust mean) Range value: 0.281566, ..., 0.627603

FA_PTR_L numeric Posterior thalamic radiation left (robust mean) Range value: 0.285254, ..., 0.595256

FA_SS_R numeric Sagittal stratum right (robust mean) Range value: 0.25673, ..., 0.610652

FA_SS_L numeric Sagittal stratum left (robust mean) Range value: 0.290514, ..., 0.618269

FA_EC_R numeric External capsule right (robust mean) Range value: 0.206798, ..., 0.461742

FA_EC_L numeric External capsule left (robust mean) Range value: 0.190149, ..., 0.481229

FA_CGC_R numeric Cingulum right (robust mean) Range value: 0.195094, ..., 0.539067

FA_CGC_L numeric Cingulum left (robust mean) Range value: 0.182076, ..., 0.544045

FA_CGH_R numeric Cingulum right (robust mean) Range value: 0.113434, ..., 0.47185

FA_CGH_L numeric Cingulum left (robust mean) Range value: 0.10481, ..., 0.47028

FA_FXST_R numeric Fornix (cres) / Stria terminalis right (robust mean) Range value: 0.207085, ..., 0.550102

FA_FXST_L numeric Fornix (cres) / Stria terminalis left (robust mean) Range value: 0.229574, ..., 0.594808

FA_SLF_R numeric Superior longitudinal fasciculus right (robust mean) Range value: 0.290606, ..., 0.537435

FA_SLF_L numeric Superior longitudinal fasciculus left (robust mean) Range value: 0.269563, ..., 0.527995

FA_SFO_R numeric Superior fronto-occipital fasciculus right (robust mean) Range value: 0.221473, ..., 0.514343

FA_SFO_L numeric Superior fronto-occipital fasciculus left (robust mean) Range value: 0.195763, ..., 0.490461

FA_UNC_R numeric Uncinate fasciculus right (robust mean) Range value: 0.142319, ..., 0.560214

FA_UNC_L numeric Uncinate fasciculus left (robust mean) Range value: 0.158052, ..., 0.567147

FA_TAP_R numeric Tapetum right (robust mean) Range value: 0.18739, ..., 0.5784

FA_TAP_L numeric Tapetum left (robust mean) Range value: 0.133857, ..., 0.510728

FA_CST numeric Corticospinal tract Bilateral (robust mean) Range value: 0.227816, ..., 0.653541

FA_ML numeric Medial lemniscus Bilateral (robust mean) Range value: 0.283328, ..., 0.753179

FA_ICP numeric Inferior cerebellar peduncle Bilateral (robust mean) Range value: 0.204613, ..., 0.574223

FA_SCP numeric Superior cerebellar peduncle Bilateral (robust mean) Range value: 0.263951, ..., 0.730009

FA_CP numeric Cerebral peduncle Bilateral (robust mean) Range value: 0.35189, ..., 0.784614

FA_ALIC numeric Anterior limb of internal capsule Bilateral (robust mean) Range value: 0.279054, ..., 0.616763

FA_PLIC numeric Posterior limb of internal capsule Bilateral (robust mean) Range value: 0.461686, ..., 0.719953

FA_RLIC numeric Retrolenticular part of internal capsule Bilateral (robust mean) Range value: 0.357634, ..., 0.658915

FA_ACR numeric Anterior corona radiata Bilateral (robust mean) Range value: 0.21453, ..., 0.454608

FA_SCR numeric Superior corona radiata Bilateral (robust mean) Range value: 0.271405, ..., 0.57357

FA_PCR numeric Posterior corona radiata Bilateral (robust mean) Range value: 0.244007, ..., 0.517341

FA_PTR numeric Posterior thalamic radiation Bilateral (robust mean) Range value: 0.322733, ..., 0.611

FA_SS numeric Sagittal stratum Bilateral (robust mean) Range value: 0.272583, ..., 0.596217

FA_EC numeric External capsule Bilateral (robust mean) Range value: 0.198309, ..., 0.462175

FA_CGC numeric Cingulum Bilateral (robust mean) Range value: 0.188919, ..., 0.536274

FA_CGH numeric Cingulum Bilateral (robust mean) Range value: 0.109208, ..., 0.461466

FA_FXST numeric Fornix (cres) / Stria terminalis Bilateral (robust mean) Range value: 0.217253, ..., 0.561457

FA_SLF numeric Superior longitudinal fasciculus Bilateral (robust mean) Range value: 0.288276, ..., 0.531831

FA_SFO numeric Superior fronto-occipital fasciculus Bilateral (robust mean) Range value: 0.224116, ..., 0.48538

FA_UNC numeric Uncinate fasciculus Bilateral (robust mean) Range value: 0.151583, ..., 0.55104

FA_TAP numeric Tapetum Bilateral (robust mean) Range value: 0.184844, ..., 0.539506

FA_IC numeric Full internal capsule (robust mean) Range value: 0.420054, ..., 0.661411

FA_CC numeric Full corpus callosum (robust mean) Range value: 0.293799, ..., 0.691702

FA_CR numeric Full corona radiata (robust mean) Range value: 0.242039, ..., 0.505409

FA_FULLWM numeric Full white matter (robust mean) Range value: 0.328698, ..., 0.549097

MD_MCP numeric Middle cerebellar peduncle (robust mean) Range value: 0.000520727, ..., 0.000821757

MD_PCT numeric Pontine crossing tract (robust mean) Range value: 0.000442809, ..., 0.000871865

MD_GCC numeric Genu of corpus callosum (robust mean) Range value: 0.000683782, ..., 0.00108986

MD_BCC numeric Body of corpus callosum (robust mean) Range value: 0.000673059, ..., 0.00117509

MD_SCC numeric Splenium of corpus callosum (robust mean) Range value: 0.000619261, ..., 0.00119253

MD_FX numeric Fornix (robust mean) Range value: 0.0010403, ..., 0.00308202

MD_CST_R numeric Corticospinal tract right (robust mean) Range value: 0.000408661, ..., 0.000911616

MD_CST_L numeric Corticospinal tract left (robust mean) Range value: 0.000442175, ..., 0.000880286

MD_ML_R numeric Medial lemniscus right (robust mean) Range value: 0.000533072, ..., 0.00100121

MD_ML_L numeric Medial lemniscus left (robust mean) Range value: 0.000501333, ..., 0.00104549

MD_ICP_R numeric Inferior cerebellar peduncle right (robust mean) Range value: 0.000493465, ..., 0.000881453

MD_ICP_L numeric Inferior cerebellar peduncle left (robust mean) Range value: 0.000516165, ..., 0.000926643

MD_SCP_R numeric Superior cerebellar peduncle right (robust mean) Range value: 0.000656447, ..., 0.0015982

MD_SCP_L numeric Superior cerebellar peduncle left (robust mean) Range value: 0.00052145, ..., 0.00141151

MD_CP_R numeric Cerebral peduncle right (robust mean) Range value: 0.000363368, ..., 0.00100136

MD_CP_L numeric Cerebral peduncle left (robust mean) Range value: 0.00040876, ..., 0.000943709

MD_ALIC_R numeric Anterior limb of internal capsule right (robust mean) Range value: 0.000573952, ..., 0.00117478

MD_ALIC_L numeric Anterior limb of internal capsule left (robust mean) Range value: 0.000577341, ..., 0.00118213

MD_PLIC_R numeric Posterior limb of internal capsule right (robust mean) Range value: 0.000541839, ..., 0.000884915

MD_PLIC_L numeric Posterior limb of internal capsule left (robust mean) Range value: 0.000552822, ..., 0.000894506

MD_RLIC_R numeric Retrolenticular part of internal capsule right (robust mean) Range value: 0.000535372, ..., 0.000943997

MD_RLIC_L numeric Retrolenticular part of internal capsule left (robust mean) Range value: 0.000640883, ..., 0.000949934

MD_ACR_R numeric Anterior corona radiata right (robust mean) Range value: 0.00064264, ..., 0.00121927

MD_ACR_L numeric Anterior corona radiata left (robust mean) Range value: 0.000650969, ..., 0.00122298

MD_SCR_R numeric Superior corona radiata right (robust mean) Range value: 0.000611339, ..., 0.00114902

MD_SCR_L numeric Superior corona radiata left (robust mean) Range value: 0.000611529, ..., 0.00109089

MD_PCR_R numeric Posterior corona radiata right (robust mean) Range value: 0.000667923, ..., 0.00138183

MD_PCR_L numeric Posterior corona radiata left (robust mean) Range value: 0.000669616, ..., 0.0013664

MD_PTR_R numeric Posterior thalamic radiation right (robust mean) Range value: 0.000643095, ..., 0.00109604

MD_PTR_L numeric Posterior thalamic radiation left (robust mean) Range value: 0.000661464, ..., 0.00118057

MD_SS_R numeric Sagittal stratum right (robust mean) Range value: 0.00063461, ..., 0.00104085

MD_SS_L numeric Sagittal stratum left (robust mean) Range value: 0.000637773, ..., 0.00126493

MD_EC_R numeric External capsule right (robust mean) Range value: 0.000627293, ..., 0.00105366

MD_EC_L numeric External capsule left (robust mean) Range value: 0.000628879, ..., 0.00104725

MD_CGC_R numeric Cingulum right (robust mean) Range value: 0.000624982, ..., 0.000878465

MD_CGC_L numeric Cingulum left (robust mean) Range value: 0.000639246, ..., 0.000897592

MD_CGH_R numeric Cingulum right (robust mean) Range value: 0.000608728, ..., 0.00100507

MD_CGH_L numeric Cingulum left (robust mean) Range value: 0.000496251, ..., 0.00119783

MD_FXST_R numeric Fornix (cres) / Stria terminalis right (robust mean) Range value: 0.000630743, ..., 0.00123439

MD_FXST_L numeric Fornix (cres) / Stria terminalis left (robust mean) Range value: 0.000570996, ..., 0.00101879

MD_SLF_R numeric Superior longitudinal fasciculus right (robust mean) Range value: 0.000623097, ..., 0.00102493

MD_SLF_L numeric Superior longitudinal fasciculus left (robust mean) Range value: 0.00062094, ..., 0.00101563

MD_SFO_R numeric Superior fronto-occipital fasciculus right (robust mean) Range value: 0.00060823, ..., 0.00133526

MD_SFO_L numeric Superior fronto-occipital fasciculus left (robust mean) Range value: 0.00060715, ..., 0.00162845

MD_UNC_R numeric Uncinate fasciculus right (robust mean) Range value: 0.000555553, ..., 0.00123852

MD_UNC_L numeric Uncinate fasciculus left (robust mean) Range value: 0.000500539, ..., 0.00105048

MD_TAP_R numeric Tapetum right (robust mean) Range value: 0.000806239, ..., 0.00208755

MD_TAP_L numeric Tapetum left (robust mean) Range value: 0.00083136, ..., 0.00247057

MD_CST numeric Corticospinal tract Bilateral (robust mean) Range value: 0.000424935, ..., 0.00089689

MD_ML numeric Medial lemniscus Bilateral (robust mean) Range value: 0.000531306, ..., 0.00102067

MD_ICP numeric Inferior cerebellar peduncle Bilateral (robust mean) Range value: 0.000505265, ..., 0.000874894

MD_SCP numeric Superior cerebellar peduncle Bilateral (robust mean) Range value: 0.000599782, ..., 0.00149563

MD_CP numeric Cerebral peduncle Bilateral (robust mean) Range value: 0.000388826, ..., 0.000963922

MD_ALIC numeric Anterior limb of internal capsule Bilateral (robust mean) Range value: 0.000584356, ..., 0.0011772

MD_PLIC numeric Posterior limb of internal capsule Bilateral (robust mean) Range value: 0.000548557, ..., 0.000872902

MD_RLIC numeric Retrolenticular part of internal capsule Bilateral (robust mean) Range value: 0.000612252, ..., 0.000933875

MD_ACR numeric Anterior corona radiata Bilateral (robust mean) Range value: 0.000647226, ..., 0.00116482

MD_SCR numeric Superior corona radiata Bilateral (robust mean) Range value: 0.000615078, ..., 0.00111782

MD_PCR numeric Posterior corona radiata Bilateral (robust mean) Range value: 0.000674359, ..., 0.00137404

MD_PTR numeric Posterior thalamic radiation Bilateral (robust mean) Range value: 0.000652886, ..., 0.00109555

MD_SS numeric Sagittal stratum Bilateral (robust mean) Range value: 0.000654656, ..., 0.00107304

MD_EC numeric External capsule Bilateral (robust mean) Range value: 0.000628181, ..., 0.00103973

MD_CGC numeric Cingulum Bilateral (robust mean) Range value: 0.000633, ..., 0.000874815

MD_CGH numeric Cingulum Bilateral (robust mean) Range value: 0.000553957, ..., 0.00109127

MD_FXST numeric Fornix (cres) / Stria terminalis Bilateral (robust mean) Range value: 0.000637934, ..., 0.00107901

MD_SLF numeric Superior longitudinal fasciculus Bilateral (robust mean) Range value: 0.000621994, ..., 0.00102013

MD_SFO numeric Superior fronto-occipital fasciculus Bilateral (robust mean) Range value: 0.000609919, ..., 0.00147151

MD_UNC numeric Uncinate fasciculus Bilateral (robust mean) Range value: 0.000571964, ..., 0.00113194

MD_TAP numeric Tapetum Bilateral (robust mean) Range value: 0.000852474, ..., 0.00225774

MD_IC numeric Full internal capsule (robust mean) Range value: 0.000576563, ..., 0.000953014

MD_CC numeric Full corpus callosum (robust mean) Range value: 0.000658579, ..., 0.00114916

MD_CR numeric Full corona radiata (robust mean) Range value: 0.000656279, ..., 0.00117547

MD_FULLWM numeric Full white matter (robust mean) Range value: 0.000626379, ..., 0.000953609

RD_MCP numeric Middle cerebellar peduncle (robust mean) Range value: 0.000351503, ..., 0.000644613

RD_PCT numeric Pontine crossing tract (robust mean) Range value: 0.000344147, ..., 0.000735081

RD_GCC numeric Genu of corpus callosum (robust mean) Range value: 0.000397507, ..., 0.00086999

RD_BCC numeric Body of corpus callosum (robust mean) Range value: 0.000389497, ..., 0.000888864

RD_SCC numeric Splenium of corpus callosum (robust mean) Range value: 0.000278346, ..., 0.000951786

RD_FX numeric Fornix (robust mean) Range value: 0.000726104, ..., 0.00289282

RD_CST_R numeric Corticospinal tract right (robust mean) Range value: 0.000260598, ..., 0.000768236

RD_CST_L numeric Corticospinal tract left (robust mean) Range value: 0.000270151, ..., 0.00071023

RD_ML_R numeric Medial lemniscus right (robust mean) Range value: 0.0002496, ..., 0.000782451

RD_ML_L numeric Medial lemniscus left (robust mean) Range value: 0.000196513, ..., 0.00085232

RD_ICP_R numeric Inferior cerebellar peduncle right (robust mean) Range value: 0.000333829, ..., 0.00069156

RD_ICP_L numeric Inferior cerebellar peduncle left (robust mean) Range value: 0.000327272, ..., 0.00076003

RD_SCP_R numeric Superior cerebellar peduncle right (robust mean) Range value: 0.000335188, ..., 0.00131033

RD_SCP_L numeric Superior cerebellar peduncle left (robust mean) Range value: 0.000326724, ..., 0.00111467

RD_CP_R numeric Cerebral peduncle right (robust mean) Range value: 0.000227326, ..., 0.00070077

RD_CP_L numeric Cerebral peduncle left (robust mean) Range value: 0.000208343, ..., 0.000675923

RD_ALIC_R numeric Anterior limb of internal capsule right (robust mean) Range value: 0.000377413, ..., 0.000963219

RD_ALIC_L numeric Anterior limb of internal capsule left (robust mean) Range value: 0.000336252, ..., 0.000951971

RD_PLIC_R numeric Posterior limb of internal capsule right (robust mean) Range value: 0.000275269, ..., 0.000567746

RD_PLIC_L numeric Posterior limb of internal capsule left (robust mean) Range value: 0.000291784, ..., 0.000632814

RD_RLIC_R numeric Retrolenticular part of internal capsule right (robust mean) Range value: 0.000267253, ..., 0.000709532

RD_RLIC_L numeric Retrolenticular part of internal capsule left (robust mean) Range value: 0.00036554, ..., 0.000698399

RD_ACR_R numeric Anterior corona radiata right (robust mean) Range value: 0.000509126, ..., 0.00107978

RD_ACR_L numeric Anterior corona radiata left (robust mean) Range value: 0.000520726, ..., 0.0010841

RD_SCR_R numeric Superior corona radiata right (robust mean) Range value: 0.000435126, ..., 0.000981095

RD_SCR_L numeric Superior corona radiata left (robust mean) Range value: 0.000418909, ..., 0.000922873

RD_PCR_R numeric Posterior corona radiata right (robust mean) Range value: 0.000484331, ..., 0.00118131

RD_PCR_L numeric Posterior corona radiata left (robust mean) Range value: 0.000499739, ..., 0.00118477

RD_PTR_R numeric Posterior thalamic radiation right (robust mean) Range value: 0.000406437, ..., 0.000872955

RD_PTR_L numeric Posterior thalamic radiation left (robust mean) Range value: 0.000445894, ..., 0.000963322

RD_SS_R numeric Sagittal stratum right (robust mean) Range value: 0.000428182, ..., 0.000827868

RD_SS_L numeric Sagittal stratum left (robust mean) Range value: 0.000395118, ..., 0.00104558

RD_EC_R numeric External capsule right (robust mean) Range value: 0.000495243, ..., 0.000939434

RD_EC_L numeric External capsule left (robust mean) Range value: 0.000463876, ..., 0.000916377

RD_CGC_R numeric Cingulum right (robust mean) Range value: 0.000446773, ..., 0.0007558

RD_CGC_L numeric Cingulum left (robust mean) Range value: 0.00043719, ..., 0.000746585

RD_CGH_R numeric Cingulum right (robust mean) Range value: 0.00047238, ..., 0.000876399

RD_CGH_L numeric Cingulum left (robust mean) Range value: 0.000415405, ..., 0.0010558

RD_FXST_R numeric Fornix (cres) / Stria terminalis right (robust mean) Range value: 0.000486854, ..., 0.00103208

RD_FXST_L numeric Fornix (cres) / Stria terminalis left (robust mean) Range value: 0.000372604, ..., 0.000820852

RD_SLF_R numeric Superior longitudinal fasciculus right (robust mean) Range value: 0.000423997, ..., 0.000855702

RD_SLF_L numeric Superior longitudinal fasciculus left (robust mean) Range value: 0.000452811, ..., 0.000853233

RD_SFO_R numeric Superior fronto-occipital fasciculus right (robust mean) Range value: 0.000457733, ..., 0.00116335

RD_SFO_L numeric Superior fronto-occipital fasciculus left (robust mean) Range value: 0.000437973, ..., 0.00149681

RD_UNC_R numeric Uncinate fasciculus right (robust mean) Range value: 0.000418055, ..., 0.00107663

RD_UNC_L numeric Uncinate fasciculus left (robust mean) Range value: 0.000354935, ..., 0.000881413

RD_TAP_R numeric Tapetum right (robust mean) Range value: 0.000534338, ..., 0.00185332

RD_TAP_L numeric Tapetum left (robust mean) Range value: 0.000638001, ..., 0.00226463

RD_CST numeric Corticospinal tract Bilateral (robust mean) Range value: 0.000267712, ..., 0.000738526

RD_ML numeric Medial lemniscus Bilateral (robust mean) Range value: 0.000235951, ..., 0.000813956

RD_ICP numeric Inferior cerebellar peduncle Bilateral (robust mean) Range value: 0.000331192, ..., 0.000717804

RD_SCP numeric Superior cerebellar peduncle Bilateral (robust mean) Range value: 0.000341295, ..., 0.00120448

RD_CP numeric Cerebral peduncle Bilateral (robust mean) Range value: 0.000216174, ..., 0.000669671

RD_ALIC numeric Anterior limb of internal capsule Bilateral (robust mean) Range value: 0.000390841, ..., 0.000958499

RD_PLIC numeric Posterior limb of internal capsule Bilateral (robust mean) Range value: 0.000289528, ..., 0.000589173

RD_RLIC numeric Retrolenticular part of internal capsule Bilateral (robust mean) Range value: 0.000362817, ..., 0.000687459

RD_ACR numeric Anterior corona radiata Bilateral (robust mean) Range value: 0.000515998, ..., 0.00101013

RD_SCR numeric Superior corona radiata Bilateral (robust mean) Range value: 0.00044216, ..., 0.000950696

RD_PCR numeric Posterior corona radiata Bilateral (robust mean) Range value: 0.00049104, ..., 0.00118257

RD_PTR numeric Posterior thalamic radiation Bilateral (robust mean) Range value: 0.000432238, ..., 0.000883545

RD_SS numeric Sagittal stratum Bilateral (robust mean) Range value: 0.000412026, ..., 0.000844668

RD_EC numeric External capsule Bilateral (robust mean) Range value: 0.000494361, ..., 0.000923274

RD_CGC numeric Cingulum Bilateral (robust mean) Range value: 0.000446508, ..., 0.00075128

RD_CGH numeric Cingulum Bilateral (robust mean) Range value: 0.000463489, ..., 0.000947194

RD_FXST numeric Fornix (cres) / Stria terminalis Bilateral (robust mean) Range value: 0.00044311, ..., 0.000884391

RD_SLF numeric Superior longitudinal fasciculus Bilateral (robust mean) Range value: 0.000442808, ..., 0.00085473

RD_SFO numeric Superior fronto-occipital fasciculus Bilateral (robust mean) Range value: 0.000459779, ..., 0.00132708

RD_UNC numeric Uncinate fasciculus Bilateral (robust mean) Range value: 0.000437758, ..., 0.000960378

RD_TAP numeric Tapetum Bilateral (robust mean) Range value: 0.000594297, ..., 0.00203638

RD_IC numeric Full internal capsule (robust mean) Range value: 0.000349624, ..., 0.000680266

RD_CC numeric Full corpus callosum (robust mean) Range value: 0.000372826, ..., 0.000870901

RD_CR numeric Full corona radiata (robust mean) Range value: 0.000499836, ..., 0.00101819

RD_FULLWM numeric Full white matter (robust mean) Range value: 0.000429675, ..., 0.000771754

update_stamp POSIXct: YYY-MM-DD