USC - DTI ROI Summary Measures v2 (Robust Mean)

USC - DTI ROI Summary Measures v2 (Robust Mean) data. More information is available at https://adni.loni.usc.edu/data-samples/adni-data/

Usage

data(DTIROI_ROBUSTMEAN_V2)

Format

A data frame with 972 observations and 306 variables.

Source

https://adni.loni.usc.edu/data-samples/adni-data/

Details

ORIGPROT character Original study protocol , with options: ADNI1, ADNIGO, ADNI2, ADNI3

COLPROT character Study protocol of data collection , with options: ADNI3

RID numeric Participant roster ID Range value: 21, ..., 6832

VISCODE character Visit code , with options: init, y1, y2, v41, v52, v51, v42, sc, nv

VISCODE2 character Translated visit code

EXAMDATE Examination Date Date: YYY-MM-DD

VERSION numeric Range value: 2, ..., 2

MANUFACTURER character Scanner Manufacturer , with options: Siemens, Philips, GE

VOLUMES numeric Number of volumes prescribed by acquisition protocol Range value: 31, ..., 127

IMAGEUID numeric IMAGE ID of the raw diffusion scan Range value: 808826, ..., 1332415

RUNDATE Date: YYY-MM-DD

STATUS character , with options: complete, not processed

QC character Quality check notes

AD_MCP numeric Middle cerebellar peduncle (robust mean) Range value: 0.000829046, ..., 0.00126646

AD_PCT numeric Pontine crossing tract (robust mean) Range value: 0.000678512, ..., 0.0012143

AD_GCC numeric Genu of corpus callosum (robust mean) Range value: 0.00121025, ..., 0.0016655

AD_BCC numeric Body of corpus callosum (robust mean) Range value: 0.00121387, ..., 0.00178963

AD_SCC numeric Splenium of corpus callosum (robust mean) Range value: 0.00123433, ..., 0.00180714

AD_FX numeric Fornix (robust mean) Range value: 0.00162864, ..., 0.00342397

AD_CST_R numeric Corticospinal tract right (robust mean) Range value: 0.000778074, ..., 0.00120754

AD_CST_L numeric Corticospinal tract left (robust mean) Range value: 0.00080593, ..., 0.00127833

AD_ML_R numeric Medial lemniscus right (robust mean) Range value: 0.000897082, ..., 0.00144868

AD_ML_L numeric Medial lemniscus left (robust mean) Range value: 0.000904022, ..., 0.00145631

AD_ICP_R numeric Inferior cerebellar peduncle right (robust mean) Range value: 0.000769043, ..., 0.00133886

AD_ICP_L numeric Inferior cerebellar peduncle left (robust mean) Range value: 0.00085011, ..., 0.00126012

AD_SCP_R numeric Superior cerebellar peduncle right (robust mean) Range value: 0.00118762, ..., 0.00216635

AD_SCP_L numeric Superior cerebellar peduncle left (robust mean) Range value: 0.00122373, ..., 0.00202224

AD_CP_R numeric Cerebral peduncle right (robust mean) Range value: 0.000995372, ..., 0.00172317

AD_CP_L numeric Cerebral peduncle left (robust mean) Range value: 0.00100764, ..., 0.00156891

AD_ALIC_R numeric Anterior limb of internal capsule right (robust mean) Range value: 0.000877232, ..., 0.00154782

AD_ALIC_L numeric Anterior limb of internal capsule left (robust mean) Range value: 0.000899512, ..., 0.00152884

AD_PLIC_R numeric Posterior limb of internal capsule right (robust mean) Range value: 0.00105146, ..., 0.00148386

AD_PLIC_L numeric Posterior limb of internal capsule left (robust mean) Range value: 0.0010599, ..., 0.00148493

AD_RLIC_R numeric Retrolenticular part of internal capsule right (robust mean) Range value: 0.00101477, ..., 0.00151841

AD_RLIC_L numeric Retrolenticular part of internal capsule left (robust mean) Range value: 0.00107977, ..., 0.0015243

AD_ACR_R numeric Anterior corona radiata right (robust mean) Range value: 0.000928974, ..., 0.00148136

AD_ACR_L numeric Anterior corona radiata left (robust mean) Range value: 0.000925331, ..., 0.00145864

AD_SCR_R numeric Superior corona radiata right (robust mean) Range value: 0.00090106, ..., 0.00159583

AD_SCR_L numeric Superior corona radiata left (robust mean) Range value: 0.000949049, ..., 0.00156574

AD_PCR_R numeric Posterior corona radiata right (robust mean) Range value: 0.000991979, ..., 0.00178614

AD_PCR_L numeric Posterior corona radiata left (robust mean) Range value: 0.000985566, ..., 0.00172159

AD_PTR_R numeric Posterior thalamic radiation right (robust mean) Range value: 0.00106679, ..., 0.00152649

AD_PTR_L numeric Posterior thalamic radiation left (robust mean) Range value: 0.00105047, ..., 0.00154731

AD_SS_R numeric Sagittal stratum right (robust mean) Range value: 0.00102135, ..., 0.00155689

AD_SS_L numeric Sagittal stratum left (robust mean) Range value: 0.0010158, ..., 0.00158366

AD_EC_R numeric External capsule right (robust mean) Range value: 0.000908008, ..., 0.00138704

AD_EC_L numeric External capsule left (robust mean) Range value: 0.00090765, ..., 0.00139369

AD_CGC_R numeric Cingulum right (robust mean) Range value: 0.000911211, ..., 0.00123791

AD_CGC_L numeric Cingulum left (robust mean) Range value: 0.000934741, ..., 0.00125187

AD_CGH_R numeric Cingulum right (robust mean) Range value: 0.000824341, ..., 0.00128279

AD_CGH_L numeric Cingulum left (robust mean) Range value: 0.000826855, ..., 0.00143047

AD_FXST_R numeric Fornix (cres) / Stria terminalis right (robust mean) Range value: 0.000898062, ..., 0.00179612

AD_FXST_L numeric Fornix (cres) / Stria terminalis left (robust mean) Range value: 0.00083766, ..., 0.00145359

AD_SLF_R numeric Superior longitudinal fasciculus right (robust mean) Range value: 0.000937772, ..., 0.00136921

AD_SLF_L numeric Superior longitudinal fasciculus left (robust mean) Range value: 0.000946865, ..., 0.00134061

AD_SFO_R numeric Superior fronto-occipital fasciculus right (robust mean) Range value: 0.000863616, ..., 0.00165804

AD_SFO_L numeric Superior fronto-occipital fasciculus left (robust mean) Range value: 0.000868013, ..., 0.00190102

AD_UNC_R numeric Uncinate fasciculus right (robust mean) Range value: 0.000830819, ..., 0.00149804

AD_UNC_L numeric Uncinate fasciculus left (robust mean) Range value: 0.000779406, ..., 0.00139073

AD_TAP_R numeric Tapetum right (robust mean) Range value: 0.0011764, ..., 0.00258272

AD_TAP_L numeric Tapetum left (robust mean) Range value: 0.00134851, ..., 0.0028861

AD_CST numeric Corticospinal tract Bilateral (robust mean) Range value: 0.000809323, ..., 0.00123102

AD_ML numeric Medial lemniscus Bilateral (robust mean) Range value: 0.000906207, ..., 0.00145263

AD_ICP numeric Inferior cerebellar peduncle Bilateral (robust mean) Range value: 0.000828097, ..., 0.00125567

AD_SCP numeric Superior cerebellar peduncle Bilateral (robust mean) Range value: 0.00122828, ..., 0.00207276

AD_CP numeric Cerebral peduncle Bilateral (robust mean) Range value: 0.00100187, ..., 0.00164113

AD_ALIC numeric Anterior limb of internal capsule Bilateral (robust mean) Range value: 0.00091851, ..., 0.00152083

AD_PLIC numeric Posterior limb of internal capsule Bilateral (robust mean) Range value: 0.0010661, ..., 0.00146677

AD_RLIC numeric Retrolenticular part of internal capsule Bilateral (robust mean) Range value: 0.00104742, ..., 0.00150999

AD_ACR numeric Anterior corona radiata Bilateral (robust mean) Range value: 0.000943332, ..., 0.00147056

AD_SCR numeric Superior corona radiata Bilateral (robust mean) Range value: 0.000937763, ..., 0.0015798

AD_PCR numeric Posterior corona radiata Bilateral (robust mean) Range value: 0.00100198, ..., 0.00175337

AD_PTR numeric Posterior thalamic radiation Bilateral (robust mean) Range value: 0.00105728, ..., 0.00152512

AD_SS numeric Sagittal stratum Bilateral (robust mean) Range value: 0.00101865, ..., 0.00151413

AD_EC numeric External capsule Bilateral (robust mean) Range value: 0.0009066, ..., 0.00136913

AD_CGC numeric Cingulum Bilateral (robust mean) Range value: 0.000934579, ..., 0.0012405

AD_CGH numeric Cingulum Bilateral (robust mean) Range value: 0.000845959, ..., 0.00131073

AD_FXST numeric Fornix (cres) / Stria terminalis Bilateral (robust mean) Range value: 0.000887586, ..., 0.00162273

AD_SLF numeric Superior longitudinal fasciculus Bilateral (robust mean) Range value: 0.000949245, ..., 0.00135425

AD_SFO numeric Superior fronto-occipital fasciculus Bilateral (robust mean) Range value: 0.000869218, ..., 0.00163133

AD_UNC numeric Uncinate fasciculus Bilateral (robust mean) Range value: 0.000802693, ..., 0.00139499

AD_TAP numeric Tapetum Bilateral (robust mean) Range value: 0.00127621, ..., 0.00264496

AD_IC numeric Full internal capsule (robust mean) Range value: 0.00102249, ..., 0.00147243

AD_CC numeric Full corpus callosum (robust mean) Range value: 0.0012568, ..., 0.00170049

AD_CR numeric Full corona radiata (robust mean) Range value: 0.000969116, ..., 0.00152303

AD_FULLWM numeric Full white matter (robust mean) Range value: 0.00102102, ..., 0.00142226

FA_MCP numeric Middle cerebellar peduncle (robust mean) Range value: 0.349887, ..., 0.547229

FA_PCT numeric Pontine crossing tract (robust mean) Range value: 0.249829, ..., 0.508054

FA_GCC numeric Genu of corpus callosum (robust mean) Range value: 0.318655, ..., 0.638687

FA_BCC numeric Body of corpus callosum (robust mean) Range value: 0.335393, ..., 0.649552

FA_SCC numeric Splenium of corpus callosum (robust mean) Range value: 0.382194, ..., 0.755416

FA_FX numeric Fornix (robust mean) Range value: 0.0586734, ..., 0.532278

FA_CST_R numeric Corticospinal tract right (robust mean) Range value: 0.325432, ..., 0.656646

FA_CST_L numeric Corticospinal tract left (robust mean) Range value: 0.322094, ..., 0.650365

FA_ML_R numeric Medial lemniscus right (robust mean) Range value: 0.357925, ..., 0.7301

FA_ML_L numeric Medial lemniscus left (robust mean) Range value: 0.333271, ..., 0.787983

FA_ICP_R numeric Inferior cerebellar peduncle right (robust mean) Range value: 0.265653, ..., 0.596592

FA_ICP_L numeric Inferior cerebellar peduncle left (robust mean) Range value: 0.280105, ..., 0.581042

FA_SCP_R numeric Superior cerebellar peduncle right (robust mean) Range value: 0.316322, ..., 0.731428

FA_SCP_L numeric Superior cerebellar peduncle left (robust mean) Range value: 0.377064, ..., 0.734093

FA_CP_R numeric Cerebral peduncle right (robust mean) Range value: 0.403532, ..., 0.752446

FA_CP_L numeric Cerebral peduncle left (robust mean) Range value: 0.427072, ..., 0.717686

FA_ALIC_R numeric Anterior limb of internal capsule right (robust mean) Range value: 0.324158, ..., 0.595764

FA_ALIC_L numeric Anterior limb of internal capsule left (robust mean) Range value: 0.239185, ..., 0.591678

FA_PLIC_R numeric Posterior limb of internal capsule right (robust mean) Range value: 0.498038, ..., 0.724186

FA_PLIC_L numeric Posterior limb of internal capsule left (robust mean) Range value: 0.466243, ..., 0.751875

FA_RLIC_R numeric Retrolenticular part of internal capsule right (robust mean) Range value: 0.380149, ..., 0.647626

FA_RLIC_L numeric Retrolenticular part of internal capsule left (robust mean) Range value: 0.38032, ..., 0.654021

FA_ACR_R numeric Anterior corona radiata right (robust mean) Range value: 0.210213, ..., 0.452849

FA_ACR_L numeric Anterior corona radiata left (robust mean) Range value: 0.203699, ..., 0.446926

FA_SCR_R numeric Superior corona radiata right (robust mean) Range value: 0.259513, ..., 0.545445

FA_SCR_L numeric Superior corona radiata left (robust mean) Range value: 0.282371, ..., 0.576903

FA_PCR_R numeric Posterior corona radiata right (robust mean) Range value: 0.257175, ..., 0.511573

FA_PCR_L numeric Posterior corona radiata left (robust mean) Range value: 0.232454, ..., 0.501072

FA_PTR_R numeric Posterior thalamic radiation right (robust mean) Range value: 0.336562, ..., 0.624485

FA_PTR_L numeric Posterior thalamic radiation left (robust mean) Range value: 0.285254, ..., 0.594092

FA_SS_R numeric Sagittal stratum right (robust mean) Range value: 0.311118, ..., 0.596426

FA_SS_L numeric Sagittal stratum left (robust mean) Range value: 0.307789, ..., 0.618269

FA_EC_R numeric External capsule right (robust mean) Range value: 0.206798, ..., 0.456075

FA_EC_L numeric External capsule left (robust mean) Range value: 0.190149, ..., 0.446271

FA_CGC_R numeric Cingulum right (robust mean) Range value: 0.238133, ..., 0.539067

FA_CGC_L numeric Cingulum left (robust mean) Range value: 0.247566, ..., 0.544045

FA_CGH_R numeric Cingulum right (robust mean) Range value: 0.1449, ..., 0.47185

FA_CGH_L numeric Cingulum left (robust mean) Range value: 0.12837, ..., 0.46172

FA_FXST_R numeric Fornix (cres) / Stria terminalis right (robust mean) Range value: 0.207085, ..., 0.536887

FA_FXST_L numeric Fornix (cres) / Stria terminalis left (robust mean) Range value: 0.229574, ..., 0.556767

FA_SLF_R numeric Superior longitudinal fasciculus right (robust mean) Range value: 0.290606, ..., 0.531961

FA_SLF_L numeric Superior longitudinal fasciculus left (robust mean) Range value: 0.285672, ..., 0.516587

FA_SFO_R numeric Superior fronto-occipital fasciculus right (robust mean) Range value: 0.23982, ..., 0.495335

FA_SFO_L numeric Superior fronto-occipital fasciculus left (robust mean) Range value: 0.195763, ..., 0.490461

FA_UNC_R numeric Uncinate fasciculus right (robust mean) Range value: 0.167692, ..., 0.524398

FA_UNC_L numeric Uncinate fasciculus left (robust mean) Range value: 0.17369, ..., 0.567147

FA_TAP_R numeric Tapetum right (robust mean) Range value: 0.18739, ..., 0.5784

FA_TAP_L numeric Tapetum left (robust mean) Range value: 0.133857, ..., 0.50957

FA_CST numeric Corticospinal tract Bilateral (robust mean) Range value: 0.326369, ..., 0.653541

FA_ML numeric Medial lemniscus Bilateral (robust mean) Range value: 0.34848, ..., 0.753179

FA_ICP numeric Inferior cerebellar peduncle Bilateral (robust mean) Range value: 0.279274, ..., 0.55775

FA_SCP numeric Superior cerebellar peduncle Bilateral (robust mean) Range value: 0.342002, ..., 0.730009

FA_CP numeric Cerebral peduncle Bilateral (robust mean) Range value: 0.415727, ..., 0.735261

FA_ALIC numeric Anterior limb of internal capsule Bilateral (robust mean) Range value: 0.279054, ..., 0.580949

FA_PLIC numeric Posterior limb of internal capsule Bilateral (robust mean) Range value: 0.480936, ..., 0.719953

FA_RLIC numeric Retrolenticular part of internal capsule Bilateral (robust mean) Range value: 0.397368, ..., 0.640174

FA_ACR numeric Anterior corona radiata Bilateral (robust mean) Range value: 0.21453, ..., 0.439109

FA_SCR numeric Superior corona radiata Bilateral (robust mean) Range value: 0.271405, ..., 0.554291

FA_PCR numeric Posterior corona radiata Bilateral (robust mean) Range value: 0.244007, ..., 0.497257

FA_PTR numeric Posterior thalamic radiation Bilateral (robust mean) Range value: 0.336106, ..., 0.608435

FA_SS numeric Sagittal stratum Bilateral (robust mean) Range value: 0.323742, ..., 0.596217

FA_EC numeric External capsule Bilateral (robust mean) Range value: 0.198309, ..., 0.450035

FA_CGC numeric Cingulum Bilateral (robust mean) Range value: 0.245587, ..., 0.536274

FA_CGH numeric Cingulum Bilateral (robust mean) Range value: 0.145372, ..., 0.461466

FA_FXST numeric Fornix (cres) / Stria terminalis Bilateral (robust mean) Range value: 0.217253, ..., 0.547766

FA_SLF numeric Superior longitudinal fasciculus Bilateral (robust mean) Range value: 0.288276, ..., 0.516713

FA_SFO numeric Superior fronto-occipital fasciculus Bilateral (robust mean) Range value: 0.224116, ..., 0.48538

FA_UNC numeric Uncinate fasciculus Bilateral (robust mean) Range value: 0.182484, ..., 0.545063

FA_TAP numeric Tapetum Bilateral (robust mean) Range value: 0.184844, ..., 0.539506

FA_IC numeric Full internal capsule (robust mean) Range value: 0.421612, ..., 0.646405

FA_CC numeric Full corpus callosum (robust mean) Range value: 0.376672, ..., 0.665237

FA_CR numeric Full corona radiata (robust mean) Range value: 0.242039, ..., 0.494819

FA_FULLWM numeric Full white matter (robust mean) Range value: 0.328698, ..., 0.532133

MD_MCP numeric Middle cerebellar peduncle (robust mean) Range value: 0.000520727, ..., 0.000769015

MD_PCT numeric Pontine crossing tract (robust mean) Range value: 0.00048992, ..., 0.000871865

MD_GCC numeric Genu of corpus callosum (robust mean) Range value: 0.000683782, ..., 0.00108986

MD_BCC numeric Body of corpus callosum (robust mean) Range value: 0.000673059, ..., 0.00113621

MD_SCC numeric Splenium of corpus callosum (robust mean) Range value: 0.000619261, ..., 0.00116704

MD_FX numeric Fornix (robust mean) Range value: 0.00110096, ..., 0.00294833

MD_CST_R numeric Corticospinal tract right (robust mean) Range value: 0.000467919, ..., 0.000854261

MD_CST_L numeric Corticospinal tract left (robust mean) Range value: 0.000499735, ..., 0.000862058

MD_ML_R numeric Medial lemniscus right (robust mean) Range value: 0.000533072, ..., 0.00100121

MD_ML_L numeric Medial lemniscus left (robust mean) Range value: 0.000501333, ..., 0.00104549

MD_ICP_R numeric Inferior cerebellar peduncle right (robust mean) Range value: 0.000493465, ..., 0.00086922

MD_ICP_L numeric Inferior cerebellar peduncle left (robust mean) Range value: 0.000516165, ..., 0.000818127

MD_SCP_R numeric Superior cerebellar peduncle right (robust mean) Range value: 0.000656447, ..., 0.0015982

MD_SCP_L numeric Superior cerebellar peduncle left (robust mean) Range value: 0.000694245, ..., 0.00141151

MD_CP_R numeric Cerebral peduncle right (robust mean) Range value: 0.000580593, ..., 0.00100136

MD_CP_L numeric Cerebral peduncle left (robust mean) Range value: 0.000561946, ..., 0.000943709

MD_ALIC_R numeric Anterior limb of internal capsule right (robust mean) Range value: 0.000573952, ..., 0.00108972

MD_ALIC_L numeric Anterior limb of internal capsule left (robust mean) Range value: 0.000580531, ..., 0.0011078

MD_PLIC_R numeric Posterior limb of internal capsule right (robust mean) Range value: 0.000541839, ..., 0.000821331

MD_PLIC_L numeric Posterior limb of internal capsule left (robust mean) Range value: 0.000554884, ..., 0.000894506

MD_RLIC_R numeric Retrolenticular part of internal capsule right (robust mean) Range value: 0.000610515, ..., 0.000943997

MD_RLIC_L numeric Retrolenticular part of internal capsule left (robust mean) Range value: 0.000644779, ..., 0.000949934

MD_ACR_R numeric Anterior corona radiata right (robust mean) Range value: 0.000693001, ..., 0.00117912

MD_ACR_L numeric Anterior corona radiata left (robust mean) Range value: 0.000685837, ..., 0.00114912

MD_SCR_R numeric Superior corona radiata right (robust mean) Range value: 0.000611339, ..., 0.00114902

MD_SCR_L numeric Superior corona radiata left (robust mean) Range value: 0.000619309, ..., 0.00109089

MD_PCR_R numeric Posterior corona radiata right (robust mean) Range value: 0.000667923, ..., 0.00138183

MD_PCR_L numeric Posterior corona radiata left (robust mean) Range value: 0.000669616, ..., 0.0013664

MD_PTR_R numeric Posterior thalamic radiation right (robust mean) Range value: 0.000643095, ..., 0.0010607

MD_PTR_L numeric Posterior thalamic radiation left (robust mean) Range value: 0.000661464, ..., 0.0011001

MD_SS_R numeric Sagittal stratum right (robust mean) Range value: 0.00063461, ..., 0.00103462

MD_SS_L numeric Sagittal stratum left (robust mean) Range value: 0.000640282, ..., 0.00109523

MD_EC_R numeric External capsule right (robust mean) Range value: 0.000627293, ..., 0.00105366

MD_EC_L numeric External capsule left (robust mean) Range value: 0.000628879, ..., 0.00104725

MD_CGC_R numeric Cingulum right (robust mean) Range value: 0.000624982, ..., 0.000874957

MD_CGC_L numeric Cingulum left (robust mean) Range value: 0.000639246, ..., 0.000880757

MD_CGH_R numeric Cingulum right (robust mean) Range value: 0.000619561, ..., 0.00100507

MD_CGH_L numeric Cingulum left (robust mean) Range value: 0.000617536, ..., 0.00111211

MD_FXST_R numeric Fornix (cres) / Stria terminalis right (robust mean) Range value: 0.000630743, ..., 0.00123439

MD_FXST_L numeric Fornix (cres) / Stria terminalis left (robust mean) Range value: 0.000608308, ..., 0.00101879

MD_SLF_R numeric Superior longitudinal fasciculus right (robust mean) Range value: 0.000623097, ..., 0.00102493

MD_SLF_L numeric Superior longitudinal fasciculus left (robust mean) Range value: 0.00062094, ..., 0.00101563

MD_SFO_R numeric Superior fronto-occipital fasciculus right (robust mean) Range value: 0.000615898, ..., 0.00130718

MD_SFO_L numeric Superior fronto-occipital fasciculus left (robust mean) Range value: 0.00060715, ..., 0.00154649

MD_UNC_R numeric Uncinate fasciculus right (robust mean) Range value: 0.000555553, ..., 0.00115847

MD_UNC_L numeric Uncinate fasciculus left (robust mean) Range value: 0.000553293, ..., 0.000982423

MD_TAP_R numeric Tapetum right (robust mean) Range value: 0.000806239, ..., 0.00208544

MD_TAP_L numeric Tapetum left (robust mean) Range value: 0.00088037, ..., 0.00227876

MD_CST numeric Corticospinal tract Bilateral (robust mean) Range value: 0.000487136, ..., 0.00085755

MD_ML numeric Medial lemniscus Bilateral (robust mean) Range value: 0.000531306, ..., 0.00102067

MD_ICP numeric Inferior cerebellar peduncle Bilateral (robust mean) Range value: 0.000505265, ..., 0.000828961

MD_SCP numeric Superior cerebellar peduncle Bilateral (robust mean) Range value: 0.000684876, ..., 0.00149563

MD_CP numeric Cerebral peduncle Bilateral (robust mean) Range value: 0.000592267, ..., 0.000963922

MD_ALIC numeric Anterior limb of internal capsule Bilateral (robust mean) Range value: 0.000584356, ..., 0.00109857

MD_PLIC numeric Posterior limb of internal capsule Bilateral (robust mean) Range value: 0.000548557, ..., 0.000837115

MD_RLIC numeric Retrolenticular part of internal capsule Bilateral (robust mean) Range value: 0.000625854, ..., 0.000933875

MD_ACR numeric Anterior corona radiata Bilateral (robust mean) Range value: 0.000698982, ..., 0.00116482

MD_SCR numeric Superior corona radiata Bilateral (robust mean) Range value: 0.000615078, ..., 0.00111782

MD_PCR numeric Posterior corona radiata Bilateral (robust mean) Range value: 0.000674359, ..., 0.00137404

MD_PTR numeric Posterior thalamic radiation Bilateral (robust mean) Range value: 0.000652886, ..., 0.001072

MD_SS numeric Sagittal stratum Bilateral (robust mean) Range value: 0.000654656, ..., 0.00102015

MD_EC numeric External capsule Bilateral (robust mean) Range value: 0.000628181, ..., 0.00103718

MD_CGC numeric Cingulum Bilateral (robust mean) Range value: 0.000633, ..., 0.000874815

MD_CGH numeric Cingulum Bilateral (robust mean) Range value: 0.000623255, ..., 0.00102765

MD_FXST numeric Fornix (cres) / Stria terminalis Bilateral (robust mean) Range value: 0.000637934, ..., 0.00107901

MD_SLF numeric Superior longitudinal fasciculus Bilateral (robust mean) Range value: 0.000621994, ..., 0.00102013

MD_SFO numeric Superior fronto-occipital fasciculus Bilateral (robust mean) Range value: 0.000611878, ..., 0.00130076

MD_UNC numeric Uncinate fasciculus Bilateral (robust mean) Range value: 0.000571964, ..., 0.00105823

MD_TAP numeric Tapetum Bilateral (robust mean) Range value: 0.000852474, ..., 0.00213624

MD_IC numeric Full internal capsule (robust mean) Range value: 0.000576563, ..., 0.000896818

MD_CC numeric Full corpus callosum (robust mean) Range value: 0.000658579, ..., 0.00110348

MD_CR numeric Full corona radiata (robust mean) Range value: 0.000656279, ..., 0.00117547

MD_FULLWM numeric Full white matter (robust mean) Range value: 0.000626379, ..., 0.000953609

RD_MCP numeric Middle cerebellar peduncle (robust mean) Range value: 0.000351503, ..., 0.000568794

RD_PCT numeric Pontine crossing tract (robust mean) Range value: 0.000371288, ..., 0.000731805

RD_GCC numeric Genu of corpus callosum (robust mean) Range value: 0.000406252, ..., 0.00086999

RD_BCC numeric Body of corpus callosum (robust mean) Range value: 0.00041021, ..., 0.00086736

RD_SCC numeric Splenium of corpus callosum (robust mean) Range value: 0.000278346, ..., 0.000893362

RD_FX numeric Fornix (robust mean) Range value: 0.000738356, ..., 0.00279355

RD_CST_R numeric Corticospinal tract right (robust mean) Range value: 0.000260598, ..., 0.000694924

RD_CST_L numeric Corticospinal tract left (robust mean) Range value: 0.000290163, ..., 0.000692882

RD_ML_R numeric Medial lemniscus right (robust mean) Range value: 0.0002496, ..., 0.000782451

RD_ML_L numeric Medial lemniscus left (robust mean) Range value: 0.000196513, ..., 0.00085232

RD_ICP_R numeric Inferior cerebellar peduncle right (robust mean) Range value: 0.000333829, ..., 0.000677194

RD_ICP_L numeric Inferior cerebellar peduncle left (robust mean) Range value: 0.000327272, ..., 0.000664314

RD_SCP_R numeric Superior cerebellar peduncle right (robust mean) Range value: 0.000335188, ..., 0.00131033

RD_SCP_L numeric Superior cerebellar peduncle left (robust mean) Range value: 0.000326724, ..., 0.00111467

RD_CP_R numeric Cerebral peduncle right (robust mean) Range value: 0.000277332, ..., 0.000696407

RD_CP_L numeric Cerebral peduncle left (robust mean) Range value: 0.0002881, ..., 0.000675923

RD_ALIC_R numeric Anterior limb of internal capsule right (robust mean) Range value: 0.000391163, ..., 0.000875392

RD_ALIC_L numeric Anterior limb of internal capsule left (robust mean) Range value: 0.000393668, ..., 0.000911956

RD_PLIC_R numeric Posterior limb of internal capsule right (robust mean) Range value: 0.000275269, ..., 0.000547229

RD_PLIC_L numeric Posterior limb of internal capsule left (robust mean) Range value: 0.000291854, ..., 0.000632814

RD_RLIC_R numeric Retrolenticular part of internal capsule right (robust mean) Range value: 0.000365152, ..., 0.000706202

RD_RLIC_L numeric Retrolenticular part of internal capsule left (robust mean) Range value: 0.000380854, ..., 0.000698399

RD_ACR_R numeric Anterior corona radiata right (robust mean) Range value: 0.000534629, ..., 0.00104274

RD_ACR_L numeric Anterior corona radiata left (robust mean) Range value: 0.000530167, ..., 0.000997638

RD_SCR_R numeric Superior corona radiata right (robust mean) Range value: 0.000438664, ..., 0.000981095

RD_SCR_L numeric Superior corona radiata left (robust mean) Range value: 0.000445099, ..., 0.000922873

RD_PCR_R numeric Posterior corona radiata right (robust mean) Range value: 0.000484331, ..., 0.00118131

RD_PCR_L numeric Posterior corona radiata left (robust mean) Range value: 0.000499739, ..., 0.00118477

RD_PTR_R numeric Posterior thalamic radiation right (robust mean) Range value: 0.000406437, ..., 0.000833142

RD_PTR_L numeric Posterior thalamic radiation left (robust mean) Range value: 0.000445894, ..., 0.000896019

RD_SS_R numeric Sagittal stratum right (robust mean) Range value: 0.000428182, ..., 0.000795549

RD_SS_L numeric Sagittal stratum left (robust mean) Range value: 0.000395118, ..., 0.000853168

RD_EC_R numeric External capsule right (robust mean) Range value: 0.000495243, ..., 0.000939434

RD_EC_L numeric External capsule left (robust mean) Range value: 0.000494264, ..., 0.000916377

RD_CGC_R numeric Cingulum right (robust mean) Range value: 0.000446773, ..., 0.0007558

RD_CGC_L numeric Cingulum left (robust mean) Range value: 0.000457109, ..., 0.000746585

RD_CGH_R numeric Cingulum right (robust mean) Range value: 0.00047238, ..., 0.000876399

RD_CGH_L numeric Cingulum left (robust mean) Range value: 0.000494102, ..., 0.00095321

RD_FXST_R numeric Fornix (cres) / Stria terminalis right (robust mean) Range value: 0.000486854, ..., 0.00103208

RD_FXST_L numeric Fornix (cres) / Stria terminalis left (robust mean) Range value: 0.000443667, ..., 0.000808008

RD_SLF_R numeric Superior longitudinal fasciculus right (robust mean) Range value: 0.000423997, ..., 0.000855702

RD_SLF_L numeric Superior longitudinal fasciculus left (robust mean) Range value: 0.000452811, ..., 0.000853233

RD_SFO_R numeric Superior fronto-occipital fasciculus right (robust mean) Range value: 0.000457733, ..., 0.0011278

RD_SFO_L numeric Superior fronto-occipital fasciculus left (robust mean) Range value: 0.000452398, ..., 0.00138187

RD_UNC_R numeric Uncinate fasciculus right (robust mean) Range value: 0.000418055, ..., 0.000984445

RD_UNC_L numeric Uncinate fasciculus left (robust mean) Range value: 0.000412259, ..., 0.000868889

RD_TAP_R numeric Tapetum right (robust mean) Range value: 0.000536914, ..., 0.00184458

RD_TAP_L numeric Tapetum left (robust mean) Range value: 0.000663217, ..., 0.00207476

RD_CST numeric Corticospinal tract Bilateral (robust mean) Range value: 0.000274615, ..., 0.000698706

RD_ML numeric Medial lemniscus Bilateral (robust mean) Range value: 0.000235951, ..., 0.000813956

RD_ICP numeric Inferior cerebellar peduncle Bilateral (robust mean) Range value: 0.000331192, ..., 0.000647061

RD_SCP numeric Superior cerebellar peduncle Bilateral (robust mean) Range value: 0.000341295, ..., 0.00120448

RD_CP numeric Cerebral peduncle Bilateral (robust mean) Range value: 0.0002936, ..., 0.000669671

RD_ALIC numeric Anterior limb of internal capsule Bilateral (robust mean) Range value: 0.000410768, ..., 0.000893911

RD_PLIC numeric Posterior limb of internal capsule Bilateral (robust mean) Range value: 0.000289528, ..., 0.000589173

RD_RLIC numeric Retrolenticular part of internal capsule Bilateral (robust mean) Range value: 0.000381354, ..., 0.000687459

RD_ACR numeric Anterior corona radiata Bilateral (robust mean) Range value: 0.000539997, ..., 0.00100967

RD_SCR numeric Superior corona radiata Bilateral (robust mean) Range value: 0.000445777, ..., 0.000950696

RD_PCR numeric Posterior corona radiata Bilateral (robust mean) Range value: 0.00049104, ..., 0.00118257

RD_PTR numeric Posterior thalamic radiation Bilateral (robust mean) Range value: 0.000432238, ..., 0.000844766

RD_SS numeric Sagittal stratum Bilateral (robust mean) Range value: 0.000412026, ..., 0.000812457

RD_EC numeric External capsule Bilateral (robust mean) Range value: 0.000494361, ..., 0.000923274

RD_CGC numeric Cingulum Bilateral (robust mean) Range value: 0.000459025, ..., 0.00075128

RD_CGH numeric Cingulum Bilateral (robust mean) Range value: 0.000503915, ..., 0.000889422

RD_FXST numeric Fornix (cres) / Stria terminalis Bilateral (robust mean) Range value: 0.000460076, ..., 0.000884391

RD_SLF numeric Superior longitudinal fasciculus Bilateral (robust mean) Range value: 0.000442808, ..., 0.00085473

RD_SFO numeric Superior fronto-occipital fasciculus Bilateral (robust mean) Range value: 0.000459779, ..., 0.00114048

RD_UNC numeric Uncinate fasciculus Bilateral (robust mean) Range value: 0.000441314, ..., 0.000892348

RD_TAP numeric Tapetum Bilateral (robust mean) Range value: 0.000625285, ..., 0.00191287

RD_IC numeric Full internal capsule (robust mean) Range value: 0.000349624, ..., 0.000658557

RD_CC numeric Full corpus callosum (robust mean) Range value: 0.000372826, ..., 0.000836384

RD_CR numeric Full corona radiata (robust mean) Range value: 0.000501843, ..., 0.00101819

RD_FULLWM numeric Full white matter (robust mean) Range value: 0.000429675, ..., 0.000771754

update_stamp POSIXct: YYY-MM-DD

Author

adni-data@googlegroups.com