USC - DTI ROI Summary Measures v1

USC - DTI ROI Summary Measures v1 data. More information is available at https://adni.loni.usc.edu/data-samples/adni-data/

Usage

data(DTIROI_V1)

Format

A data frame with 976 observations and 243 variables.

Source

https://adni.loni.usc.edu/data-samples/adni-data/

Details

ORIGPROT character Original study protocol , with options: ADNI1, ADNIGO, ADNI2

RID numeric Participant roster ID Range value: 123, ..., 5295

VISCODE character Visit code

VISCODE2 character Translated visit code

EXAMDATE Examination Date Date: YYY-MM-DD

VERSION numeric Version Range value: 1, ..., 1

VISNAME character

VISITNO numeric Visit Range value: 3, ..., 50

LONIUID_1 numeric LONI Image ID for FA Range value: 299595, ..., 848440

LONIUID_2 numeric LONI Image ID for MD Range value: 299814, ..., 848443

LONIUID_3 numeric LONI Image ID for RD Range value: 359669, ..., 848449

LONIUID_4 numeric LONI Image ID for AD Range value: 297755, ..., 848446

RUNDATE Date Date: YYY-MM-DD

STATUS character Status , with options: Complete

AD_CST_L numeric Corticospinal tract left Range value: 0.000993997, ..., 0.0023359

AD_CST_R numeric Corticospinal tract right Range value: 0.000963274, ..., 0.00228733

AD_ICP_L numeric Inferior cerebellar peduncle left Range value: 0.000976535, ..., 0.0022723

AD_ICP_R numeric Inferior cerebellar peduncle right Range value: 0.000859119, ..., 0.00228807

AD_ML_L numeric Medial lemniscus left Range value: 0.00100278, ..., 0.00183072

AD_ML_R numeric Medial lemniscus right Range value: 0.0010317, ..., 0.00158208

AD_SCP_L numeric Superior cerebellar peduncle left Range value: 0.00115878, ..., 0.00207185

AD_SCP_R numeric Superior cerebellar peduncle right Range value: 0.00120185, ..., 0.00205651

AD_CP_L numeric Cerebral peduncle left Range value: 0.00129964, ..., 0.00184099

AD_CP_R numeric Cerebral peduncle right Range value: 0.00132518, ..., 0.00185971

AD_ALIC_L numeric Anterior limb of internal capsule left Range value: 0.00100718, ..., 0.00164792

AD_ALIC_R numeric Anterior limb of internal capsule right Range value: 0.00101425, ..., 0.00156695

AD_PLIC_L numeric Posterior limb of internal capsule left Range value: 0.0010437, ..., 0.00142589

AD_PLIC_R numeric Posterior limb of internal capsule right Range value: 0.00105059, ..., 0.00145844

AD_PTR_L numeric Posterior thalamic radiation left Range value: 0.00114353, ..., 0.0021751

AD_PTR_R numeric Posterior thalamic radiation right Range value: 0.00110727, ..., 0.00209163

AD_ACR_L numeric Anterior corona radiata left Range value: 0.000980971, ..., 0.00177484

AD_ACR_R numeric Anterior corona radiata right Range value: 0.00100739, ..., 0.00171728

AD_SCR_L numeric Superior corona radiata left Range value: 0.000923883, ..., 0.00178072

AD_SCR_R numeric Superior corona radiata right Range value: 0.000920291, ..., 0.00164695

AD_PCR_L numeric Posterior corona radiata left Range value: 0.000997919, ..., 0.00256759

AD_PCR_R numeric Posterior corona radiata right Range value: 0.000970576, ..., 0.00233851

AD_CGC_L numeric Cingulum left Range value: 0.000973597, ..., 0.00165317

AD_CGC_R numeric Cingulum right Range value: 0.000938041, ..., 0.00181551

AD_CGH_L numeric Cingulum (hippocampus) left Range value: 0.000940835, ..., 0.00171008

AD_CGH_R numeric Cingulum (hippocampus) right Range value: 0.000964048, ..., 0.00169439

AD_FX_ST_L numeric Fornix (cres) / Stria terminalis left Range value: 0.00117625, ..., 0.00258322

AD_FX_ST_R numeric Fornix (cres) / Stria terminalis right Range value: 0.00116545, ..., 0.00268004

AD_SLF_L numeric Superior longitudinal fasciculus left Range value: 0.000921446, ..., 0.00132317

AD_SLF_R numeric Superior longitudinal fasciculus right Range value: 0.000909681, ..., 0.00145328

AD_SFO_L numeric Superior fronto-occipital fasciculus left Range value: 0.000879117, ..., 0.00295549

AD_SFO_R numeric Superior fronto-occipital fasciculus right Range value: 0.00090537, ..., 0.00264967

AD_IFO_L numeric Inferior fronto-occipital fasciculus left Range value: 0.00103934, ..., 0.00174342

AD_IFO_R numeric Inferior fronto-occipital fasciculus right Range value: 0.00101582, ..., 0.00175759

AD_SS_L numeric Sagittal stratum left Range value: 0.00100127, ..., 0.00177597

AD_SS_R numeric Sagittal stratum right Range value: 0.00101029, ..., 0.00206854

AD_EC_L numeric External capsule left Range value: 0.000979624, ..., 0.00140867

AD_EC_R numeric External capsule right Range value: 0.000962904, ..., 0.00140934

AD_UNC_L numeric Uncinate fasciculus left Range value: 0.000914262, ..., 0.00240358

AD_UNC_R numeric Uncinate fasciculus right Range value: 0.000837411, ..., 0.00212123

AD_FX_L numeric Fornix left Range value: 0.00158885, ..., 0.00422541

AD_FX_R numeric Fornix right Range value: 0.00174225, ..., 0.00367729

AD_GCC_L numeric Genu of corpus callosum left Range value: 0.00138464, ..., 0.00235767

AD_GCC_R numeric Genu of corpus callosum right Range value: 0.00134543, ..., 0.00213844

AD_BCC_L numeric Body of corpus callosum left Range value: 0.00148177, ..., 0.00312935

AD_BCC_R numeric Body of corpus callosum right Range value: 0.0014899, ..., 0.0029777

AD_SCC_L numeric Splenium of corpus callosum left Range value: 0.00142734, ..., 0.00292163

AD_SCC_R numeric Splenium of corpus callosum right Range value: 0.00139539, ..., 0.00281836

AD_RLIC_L numeric Retrolenticular part of internal capsule left Range value: 0.00106141, ..., 0.00152381

AD_RLIC_R numeric Retrolenticular part of internal capsule right Range value: 0.00102322, ..., 0.00152645

AD_TAP_L numeric Tapatum left Range value: 0.00125543, ..., 0.00315392

AD_TAP_R numeric Tapatum right Range value: 0.0011004, ..., 0.00296885

AD_SUMGCC numeric Bilateral genu of the corpus callosum Range value: 0.00136834, ..., 0.00222859

AD_SUMBCC numeric Bilateral body of the corpus callosum Range value: 0.00153413, ..., 0.003054

AD_SUMSCC numeric Bilateral splenium of the corpus callosum Range value: 0.00143118, ..., 0.00277783

AD_SUMCC numeric Bilateral full corpus callosum Range value: 0.00148458, ..., 0.00261532

AD_SUMFX numeric Bilateral fornix Range value: 0.00166384, ..., 0.00377617

FA_CST_L numeric Corticospinal tract left Range value: 0.21429, ..., 0.478769

FA_CST_R numeric Corticospinal tract right Range value: 0.219521, ..., 0.50649

FA_ICP_L numeric Inferior cerebellar peduncle left Range value: 0.200001, ..., 0.540077

FA_ICP_R numeric Inferior cerebellar peduncle right Range value: 0.190979, ..., 0.543427

FA_ML_L numeric Medial lemniscus left Range value: 0.312138, ..., 0.600448

FA_ML_R numeric Medial lemniscus right Range value: 0.286779, ..., 0.58931

FA_SCP_L numeric Superior cerebellar peduncle left Range value: 0.29042, ..., 0.493868

FA_SCP_R numeric Superior cerebellar peduncle right Range value: 0.313576, ..., 0.528718

FA_CP_L numeric Cerebral peduncle left Range value: 0.348713, ..., 0.556213

FA_CP_R numeric Cerebral peduncle right Range value: 0.356891, ..., 0.548463

FA_ALIC_L numeric Anterior limb of internal capsule left Range value: 0.195564, ..., 0.467866

FA_ALIC_R numeric Anterior limb of internal capsule right Range value: 0.237606, ..., 0.492184

FA_PLIC_L numeric Posterior limb of internal capsule left Range value: 0.410757, ..., 0.634323

FA_PLIC_R numeric Posterior limb of internal capsule right Range value: 0.42251, ..., 0.619324

FA_PTR_L numeric Posterior thalamic radiation left Range value: 0.218698, ..., 0.464181

FA_PTR_R numeric Posterior thalamic radiation right Range value: 0.199405, ..., 0.479922

FA_ACR_L numeric Anterior corona radiata left Range value: 0.159149, ..., 0.401258

FA_ACR_R numeric Anterior corona radiata right Range value: 0.168502, ..., 0.422831

FA_SCR_L numeric Superior corona radiata left Range value: 0.243831, ..., 0.460448

FA_SCR_R numeric Superior corona radiata right Range value: 0.249534, ..., 0.417498

FA_PCR_L numeric Posterior corona radiata left Range value: 0.224325, ..., 0.462149

FA_PCR_R numeric Posterior corona radiata right Range value: 0.226174, ..., 0.476579

FA_CGC_L numeric Cingulum left Range value: 0.179103, ..., 0.375623

FA_CGC_R numeric Cingulum right Range value: 0.148849, ..., 0.35148

FA_CGH_L numeric Cingulum (hippocampus) left Range value: 0.141236, ..., 0.314793

FA_CGH_R numeric Cingulum (hippocampus) right Range value: 0.161659, ..., 0.362394

FA_FX_ST_L numeric Fornix (cres) / Stria terminalis left Range value: 0.14481, ..., 0.402883

FA_FX_ST_R numeric Fornix (cres) / Stria terminalis right Range value: 0.123952, ..., 0.378265

FA_SLF_L numeric Superior longitudinal fasciculus left Range value: 0.225005, ..., 0.413362

FA_SLF_R numeric Superior longitudinal fasciculus right Range value: 0.257282, ..., 0.42343

FA_SFO_L numeric Superior fronto-occipital fasciculus left Range value: 0.0821821, ..., 0.363496

FA_SFO_R numeric Superior fronto-occipital fasciculus right Range value: 0.0922206, ..., 0.398694

FA_IFO_L numeric Inferior fronto-occipital fasciculus left Range value: 0.181494, ..., 0.397715

FA_IFO_R numeric Inferior fronto-occipital fasciculus right Range value: 0.153711, ..., 0.396123

FA_SS_L numeric Sagittal stratum left Range value: 0.220469, ..., 0.396485

FA_SS_R numeric Sagittal stratum right Range value: 0.175595, ..., 0.404134

FA_EC_L numeric External capsule left Range value: 0.158739, ..., 0.339948

FA_EC_R numeric External capsule right Range value: 0.166009, ..., 0.346483

FA_UNC_L numeric Uncinate fasciculus left Range value: 0.0929704, ..., 0.310998

FA_UNC_R numeric Uncinate fasciculus right Range value: 0.111383, ..., 0.31548

FA_FX_L numeric Fornix left Range value: 0.039708, ..., 0.335833

FA_FX_R numeric Fornix right Range value: 0.0640862, ..., 0.352706

FA_GCC_L numeric Genu of corpus callosum left Range value: 0.118239, ..., 0.585558

FA_GCC_R numeric Genu of corpus callosum right Range value: 0.128932, ..., 0.602886

FA_BCC_L numeric Body of corpus callosum left Range value: 0.110401, ..., 0.530601

FA_BCC_R numeric Body of corpus callosum right Range value: 0.118927, ..., 0.530068

FA_SCC_L numeric Splenium of corpus callosum left Range value: 0.146024, ..., 0.603467

FA_SCC_R numeric Splenium of corpus callosum right Range value: 0.126243, ..., 0.604542

FA_RLIC_L numeric Retrolenticular part of internal capsule left Range value: 0.313875, ..., 0.501661

FA_RLIC_R numeric Retrolenticular part of internal capsule right Range value: 0.334446, ..., 0.520078

FA_TAP_L numeric Tapatum left Range value: 0.0289957, ..., 0.409767

FA_TAP_R numeric Tapatum right Range value: 0.0587773, ..., 0.47323

FA_SUMGCC numeric Bilateral genu of the corpus callosum Range value: 0.12678, ..., 0.587242

FA_SUMBCC numeric Bilateral body of the corpus callosum Range value: 0.11513, ..., 0.53032

FA_SUMSCC numeric Bilateral splenium of the corpus callosum Range value: 0.148758, ..., 0.602003

FA_SUMCC numeric Bilateral full corpus callosum Range value: 0.19629, ..., 0.570519

FA_SUMFX numeric Bilateral fornix Range value: 0.065717, ..., 0.329724

MD_CST_L numeric Corticospinal tract left Range value: 0.000680956, ..., 0.00204066

MD_CST_R numeric Corticospinal tract right Range value: 0.000626507, ..., 0.00197098

MD_ICP_L numeric Inferior cerebellar peduncle left Range value: 0.000700378, ..., 0.00187846

MD_ICP_R numeric Inferior cerebellar peduncle right Range value: 0.000662808, ..., 0.00193251

MD_ML_L numeric Medial lemniscus left Range value: 0.000666465, ..., 0.00145528

MD_ML_R numeric Medial lemniscus right Range value: 0.000642953, ..., 0.00119683

MD_SCP_L numeric Superior cerebellar peduncle left Range value: 0.000793968, ..., 0.00146032

MD_SCP_R numeric Superior cerebellar peduncle right Range value: 0.000758536, ..., 0.00147208

MD_CP_L numeric Cerebral peduncle left Range value: 0.000794933, ..., 0.00127232

MD_CP_R numeric Cerebral peduncle right Range value: 0.000848849, ..., 0.00128908

MD_ALIC_L numeric Anterior limb of internal capsule left Range value: 0.000684958, ..., 0.0014214

MD_ALIC_R numeric Anterior limb of internal capsule right Range value: 0.000683233, ..., 0.00117625

MD_PLIC_L numeric Posterior limb of internal capsule left Range value: 0.000592406, ..., 0.000932953

MD_PLIC_R numeric Posterior limb of internal capsule right Range value: 0.000592934, ..., 0.000897941

MD_PTR_L numeric Posterior thalamic radiation left Range value: 0.000776136, ..., 0.00182813

MD_PTR_R numeric Posterior thalamic radiation right Range value: 0.000746882, ..., 0.00180862

MD_ACR_L numeric Anterior corona radiata left Range value: 0.000724209, ..., 0.00152839

MD_ACR_R numeric Anterior corona radiata right Range value: 0.000742919, ..., 0.0014598

MD_SCR_L numeric Superior corona radiata left Range value: 0.000667851, ..., 0.00138808

MD_SCR_R numeric Superior corona radiata right Range value: 0.000650919, ..., 0.00123191

MD_PCR_L numeric Posterior corona radiata left Range value: 0.000701259, ..., 0.00220364

MD_PCR_R numeric Posterior corona radiata right Range value: 0.000698319, ..., 0.00187262

MD_CGC_L numeric Cingulum left Range value: 0.000733628, ..., 0.00129201

MD_CGC_R numeric Cingulum right Range value: 0.000719108, ..., 0.00150079

MD_CGH_L numeric Cingulum (hippocampus) left Range value: 0.000741311, ..., 0.00148218

MD_CGH_R numeric Cingulum (hippocampus) right Range value: 0.000752325, ..., 0.0014515

MD_FX_ST_L numeric Fornix (cres) / Stria terminalis left Range value: 0.000859886, ..., 0.00229299

MD_FX_ST_R numeric Fornix (cres) / Stria terminalis right Range value: 0.000864476, ..., 0.00240514

MD_SLF_L numeric Superior longitudinal fasciculus left Range value: 0.000669879, ..., 0.00101601

MD_SLF_R numeric Superior longitudinal fasciculus right Range value: 0.000640319, ..., 0.00102618

MD_SFO_L numeric Superior fronto-occipital fasciculus left Range value: 0.000639704, ..., 0.00257162

MD_SFO_R numeric Superior fronto-occipital fasciculus right Range value: 0.000626051, ..., 0.00234641

MD_IFO_L numeric Inferior fronto-occipital fasciculus left Range value: 0.000766934, ..., 0.00147858

MD_IFO_R numeric Inferior fronto-occipital fasciculus right Range value: 0.000752035, ..., 0.00151579

MD_SS_L numeric Sagittal stratum left Range value: 0.000743559, ..., 0.00140108

MD_SS_R numeric Sagittal stratum right Range value: 0.000746493, ..., 0.00182045

MD_EC_L numeric External capsule left Range value: 0.000745804, ..., 0.00114929

MD_EC_R numeric External capsule right Range value: 0.000718573, ..., 0.00113225

MD_UNC_L numeric Uncinate fasciculus left Range value: 0.000748275, ..., 0.00222833

MD_UNC_R numeric Uncinate fasciculus right Range value: 0.000672916, ..., 0.00188058

MD_FX_L numeric Fornix left Range value: 0.00122076, ..., 0.00367397

MD_FX_R numeric Fornix right Range value: 0.00127621, ..., 0.00315862

MD_GCC_L numeric Genu of corpus callosum left Range value: 0.000841326, ..., 0.00201157

MD_GCC_R numeric Genu of corpus callosum right Range value: 0.000833173, ..., 0.00174139

MD_BCC_L numeric Body of corpus callosum left Range value: 0.00094121, ..., 0.00280511

MD_BCC_R numeric Body of corpus callosum right Range value: 0.000917638, ..., 0.00266126

MD_SCC_L numeric Splenium of corpus callosum left Range value: 0.000850511, ..., 0.00260368

MD_SCC_R numeric Splenium of corpus callosum right Range value: 0.000798887, ..., 0.00257251

MD_RLIC_L numeric Retrolenticular part of internal capsule left Range value: 0.000683271, ..., 0.0010464

MD_RLIC_R numeric Retrolenticular part of internal capsule right Range value: 0.000641464, ..., 0.00112352

MD_TAP_L numeric Tapatum left Range value: 0.00089972, ..., 0.00293113

MD_TAP_R numeric Tapatum right Range value: 0.000732027, ..., 0.00277181

MD_SUMGCC numeric Bilateral genu of the corpus callosum Range value: 0.000846759, ..., 0.00186883

MD_SUMBCC numeric Bilateral body of the corpus callosum Range value: 0.000943302, ..., 0.00271543

MD_SUMSCC numeric Bilateral splenium of the corpus callosum Range value: 0.000830337, ..., 0.00249503

MD_SUMCC numeric Bilateral full corpus callosum Range value: 0.000884068, ..., 0.00226628

MD_SUMFX numeric Bilateral fornix Range value: 0.00126744, ..., 0.00325246

RD_CST_L numeric Corticospinal tract left Range value: 0.000513187, ..., 0.00189304

RD_CST_R numeric Corticospinal tract right Range value: 0.000458123, ..., 0.0018128

RD_ICP_L numeric Inferior cerebellar peduncle left Range value: 0.0005623, ..., 0.00168154

RD_ICP_R numeric Inferior cerebellar peduncle right Range value: 0.000502488, ..., 0.00175473

RD_ML_L numeric Medial lemniscus left Range value: 0.000407372, ..., 0.00126756

RD_ML_R numeric Medial lemniscus right Range value: 0.000423634, ..., 0.0010048

RD_SCP_L numeric Superior cerebellar peduncle left Range value: 0.000606437, ..., 0.00122119

RD_SCP_R numeric Superior cerebellar peduncle right Range value: 0.000536879, ..., 0.00120509

RD_CP_L numeric Cerebral peduncle left Range value: 0.000542581, ..., 0.00104121

RD_CP_R numeric Cerebral peduncle right Range value: 0.000587045, ..., 0.00105528

RD_ALIC_L numeric Anterior limb of internal capsule left Range value: 0.000508444, ..., 0.00130813

RD_ALIC_R numeric Anterior limb of internal capsule right Range value: 0.000483792, ..., 0.00104695

RD_PLIC_L numeric Posterior limb of internal capsule left Range value: 0.000366757, ..., 0.00069468

RD_PLIC_R numeric Posterior limb of internal capsule right Range value: 0.000352689, ..., 0.000649413

RD_PTR_L numeric Posterior thalamic radiation left Range value: 0.000573759, ..., 0.00165465

RD_PTR_R numeric Posterior thalamic radiation right Range value: 0.000539348, ..., 0.00166712

RD_ACR_L numeric Anterior corona radiata left Range value: 0.00058294, ..., 0.00140517

RD_ACR_R numeric Anterior corona radiata right Range value: 0.000586911, ..., 0.00133106

RD_SCR_L numeric Superior corona radiata left Range value: 0.000524569, ..., 0.00119176

RD_SCR_R numeric Superior corona radiata right Range value: 0.000509481, ..., 0.00102439

RD_PCR_L numeric Posterior corona radiata left Range value: 0.000536286, ..., 0.00202166

RD_PCR_R numeric Posterior corona radiata right Range value: 0.000536925, ..., 0.00163967

RD_CGC_L numeric Cingulum left Range value: 0.000597966, ..., 0.00111142

RD_CGC_R numeric Cingulum right Range value: 0.000598527, ..., 0.00134343

RD_CGH_L numeric Cingulum (hippocampus) left Range value: 0.00063963, ..., 0.00136824

RD_CGH_R numeric Cingulum (hippocampus) right Range value: 0.000625422, ..., 0.00133005

RD_FX_ST_L numeric Fornix (cres) / Stria terminalis left Range value: 0.000701229, ..., 0.00214788

RD_FX_ST_R numeric Fornix (cres) / Stria terminalis right Range value: 0.000711395, ..., 0.00226769

RD_SLF_L numeric Superior longitudinal fasciculus left Range value: 0.000529939, ..., 0.000892315

RD_SLF_R numeric Superior longitudinal fasciculus right Range value: 0.000503521, ..., 0.000872657

RD_SFO_L numeric Superior fronto-occipital fasciculus left Range value: 0.000519998, ..., 0.002456

RD_SFO_R numeric Superior fronto-occipital fasciculus right Range value: 0.000486392, ..., 0.00219478

RD_IFO_L numeric Inferior fronto-occipital fasciculus left Range value: 0.000605482, ..., 0.00134616

RD_IFO_R numeric Inferior fronto-occipital fasciculus right Range value: 0.000599599, ..., 0.00139489

RD_SS_L numeric Sagittal stratum left Range value: 0.000580973, ..., 0.00126481

RD_SS_R numeric Sagittal stratum right Range value: 0.000593587, ..., 0.00169641

RD_EC_L numeric External capsule left Range value: 0.000628894, ..., 0.00103389

RD_EC_R numeric External capsule right Range value: 0.000592118, ..., 0.000993699

RD_UNC_L numeric Uncinate fasciculus left Range value: 0.000642883, ..., 0.0021407

RD_UNC_R numeric Uncinate fasciculus right Range value: 0.000590668, ..., 0.00176026

RD_FX_L numeric Fornix left Range value: 0.00100233, ..., 0.00339825

RD_FX_R numeric Fornix right Range value: 0.00104319, ..., 0.00296152

RD_GCC_L numeric Genu of corpus callosum left Range value: 0.000518407, ..., 0.00183852

RD_GCC_R numeric Genu of corpus callosum right Range value: 0.000520721, ..., 0.00158473

RD_BCC_L numeric Body of corpus callosum left Range value: 0.000638754, ..., 0.00264299

RD_BCC_R numeric Body of corpus callosum right Range value: 0.000610442, ..., 0.00252815

RD_SCC_L numeric Splenium of corpus callosum left Range value: 0.000524191, ..., 0.0024447

RD_SCC_R numeric Splenium of corpus callosum right Range value: 0.000500637, ..., 0.00244959

RD_RLIC_L numeric Retrolenticular part of internal capsule left Range value: 0.000491442, ..., 0.000837773

RD_RLIC_R numeric Retrolenticular part of internal capsule right Range value: 0.000450588, ..., 0.000922053

RD_TAP_L numeric Tapatum left Range value: 0.000721867, ..., 0.00288428

RD_TAP_R numeric Tapatum right Range value: 0.00054784, ..., 0.00269725

RD_SUMGCC numeric Bilateral genu of the corpus callosum Range value: 0.00053028, ..., 0.00168895

RD_SUMBCC numeric Bilateral body of the corpus callosum Range value: 0.000637537, ..., 0.00257343

RD_SUMSCC numeric Bilateral splenium of the corpus callosum Range value: 0.000523085, ..., 0.00235363

RD_SUMCC numeric Bilateral full corpus callosum Range value: 0.000569854, ..., 0.00209175

RD_SUMFX numeric Bilateral fornix Range value: 0.00105877, ..., 0.00299716

update_stamp POSIXct: YYY-MM-DD

Author

adni-data@googlegroups.com