Diagnostic Summary

Diagnostic Summary data. More information is available at https://adni.loni.usc.edu/data-samples/adni-data/

Usage

data(DXSUM)

Format

A data frame with 14686 observations and 42 variables.

Source

https://adni.loni.usc.edu/data-samples/adni-data/

Details

ORIGPROT character Original study protocol , with options: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4

COLPROT factor Study protocol of data collection , with factor levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4

PTID character Participant ID

RID numeric Participant roster ID Range value: 2, ..., 10501

VISCODE character Visit code

VISCODE2 character Translated visit code

EXAMDATE Date Date Form Completed Date: YYY-MM-DD

DIAGNOSIS character 1. Which best describes the participant's current diagnosis?

ADNI3: 1=CN;2=MCI;3=Dementia

ADNI4: 1=CN; 2=MCI; 3=Dementia

ADNI1: 1=CN;2=MCI;3=Dementia

ADNI2: 1=CN;2=MCI;3=Dementia

ADNIGO: 1=CN;2=MCI;3=Dementia

Decoded Value: Yes

DXNORM character Normal 1=Yes Decoded Value: Yes

DXNODEP character Mild Depression 1=Yes Decoded Value: Yes

DXMCI character Mild Cognitive Impairment 1=Yes Decoded Value: Yes

DXMDES character 2a. MCI features (select all that apply):

ADNI1: 1=MCI (Memory features); 2=MCI (Non-memory features)

ADNIGO: 1=MCI - Memory features (amnestic); 2=MCI - Non-memory features (non-amnestic)

ADNI2: 1=MCI - Memory features (amnestic); 2=MCI - Non-memory features (non-amnestic)

ADNI3: 1=MCI - Memory features (amnestic);2=MCI - Non-memory features (non-amnestic)

ADNI4: 1=MCI - Memory features (amnestic); 2=MCI - Non-memory features (non-amnestic)

DXMPTR1 character i. Subjective memory complaint

ADNI1: 1=Yes; 0=No

ADNIGO: 1=Yes; 0=No

ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

DXMPTR2 character ii. Informant memory complaint

ADNI1: 1=Yes; 0=No

ADNIGO: 1=Yes; 0=No

ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

DXMPTR3 character iii. Normal general cognitive function

ADNI1: 1=Yes; 0=No; 2=Marginal

ADNIGO: 1=Yes; 0=No; 2=Marginal

ADNI2: 1=Yes; 0=No; 2=Marginal

ADNI3: 0=No;1=Yes;2=Marginal

ADNI4: 0=No; 1=Yes; 2=Marginal

Decoded Value: Yes

DXMPTR4 character iv. Normal activities of daily living

ADNI1: 1=Yes; 0=No; 2=Marginal

ADNIGO: 1=Yes; 0=No; 2=Marginal

ADNI2: 1=Yes; 0=No; 2=Marginal

ADNI3: 0=No;1=Yes;2=Marginal

ADNI4: 0=No; 1=Yes; 2=Marginal

Decoded Value: Yes

DXMPTR5 character v. Objective memory impairment for age and education

ADNI1: 1=Yes; 0=No

ADNIGO: 1=Yes; 0=No

ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

DXMPTR6 character vi. Not demented by diagnostic criteria

ADNI1: 1=Yes;0=No

ADNIGO: 1=Yes;0=No

ADNI2: 1=Yes;0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

DXMDUE character 2b. Suspected cause of MCI:

ADNI1: 1=MCI due to Alzheimer's Disease; 2=MCI due to other etiology

ADNIGO: 1=MCI due to Alzheimer's Disease; 2=MCI due to other etiology

ADNI2: 1=MCI due to Alzheimer's Disease; 2=MCI due to other etiology

ADNI3: 1=MCI due to Alzheimer's Disease;2=MCI due to other etiology

ADNI4: 1=MCI due to Alzheimer's Disease; 2=MCI due to other etiology

Decoded Value: Yes

DXMOTHET character If MCI due to other etiology, select box(es) to indicate reason:

ADNI1: 1=Fronto-temporal Dementia; 2=Parkinson's Disease; 3=Huntington's Disease; 4=Progressive Supranuclear Palsy; 8=Corticobasal Degeneration; 9=Vascular Dementia; 10=Prion-Associated Dementia; 14=Other (specify)

ADNIGO: 1=Fronto-temporal Dementia; 2=Parkinson's Disease; 3=Huntington's Disease; 4=Progressive Supranuclear Palsy; 5=Alcohol-related Dementia; 6=NPH; 7=Major Depression; 8=Corticobasal Degeneration; 9=Vascular Dementia; 10=Prion-Associated Dementia; 11=HIV; 12=Primary Progressive Aphasia; 13=Posterior Cortical Dysfunction; 14=Other (specify)

ADNI2: 1=Fronto-temporal Dementia; 2=Parkinson's Disease; 3=Huntington's Disease; 4=Progressive Supranuclear Palsy; 5=Alcohol-related Dementia; 6=NPH; 7=Major Depression; 8=Corticobasal Degeneration; 9=Vascular Dementia; 10=Prion-Associated Dementia; 11=HIV; 12=Primary Progressive Aphasia; 13=Posterior Cortical Dysfunction; 14=Other (specify)

ADNI3: 1=Fronto-temporal Dementia;2=Parkinson's Disease;3=Huntington's Disease;4=Progressive Supranuclear Palsy;5=Alcohol-related Dementia;6=NPH;7=Major Depression;8=Corticobasal Degeneration;9=Vascular Dementia;10=Prion-Associated Dementia;11=HIV;12=Primary Progressive Aphasia;13=Posterior Cortical Dysfunction;14=Other (specify)

ADNI4: 1=Fronto-temporal Dementia; 10=Prion-Associated Dementia; 11=HIV; 12=Primary Progressive Aphasia; 13=Posterior Cortical Dysfunction; 14=Other (specify); 2=Parkinson's Disease; 3=Huntington's Disease; 4=Progressive Supranuclear Palsy; 5=Alcohol-related Dementia; 6=NPH; 7=Major Depression; 8=Corticobasal Degeneration; 9=Vascular Dementia

Decoded Value: Yes

DXDSEV character 3a. Dementia Severity - Clinician's Impression

ADNIGO: 1=Mild; 2=Moderate; 3=Severe

ADNI2: 1=Mild; 2=Moderate; 3=Severe

ADNI3: 1=Mild;2=Moderate;3=Severe

ADNI4: 1=Mild; 2=Moderate; 3=Severe

Decoded Value: Yes

DXDDUE character 3b. Suspected cause of dementia

ADNIGO: 1=Dementia due to Alzheimer's Disease; 2=Dementia due to other etiology

ADNI2: 1=Dementia due to Alzheimer's Disease; 2=Dementia due to other etiology

ADNI3: 1=Dementia due to Alzheimer's Disease;2=Dementia due to other etiology

ADNI4: 1=Dementia due to Alzheimer's Disease; 2=Dementia due to other etiology

Decoded Value: Yes

DXAD character Alzheimer's Disease 1=Yes Decoded Value: Yes

DXAPP character If Dementia due to Alzheimer's Disease, indicate likelihood:

ADNI1: 1=Probable; 2=Possible

ADNIGO: 1=Probable; 2=Possible

ADNI2: 1=Probable; 2=Possible

ADNI3: 1=Probable;2=Possible

ADNI4: 1=Probable; 2=Possible

Decoded Value: Yes

DXAPROB character If Probable AD, select box(es) for other symptoms present: 1=None;2=Stroke(s);3=Depression;4=Delirium;5=Parkinsonism;6=Metabolic/Toxic Disorder (specify);7=Other (specify) Decoded Value: Yes

DXAPOSS character If Possible AD, select box(es) to indicate reason:

ADNI1: 1=Atypical clinical course or features (specify); 2=Stroke(s); 3=Depression; 4=Delirium; 5=Parkinsonism; 6=Metabolic / Toxic Disorder (specify); 7=Other (specify)

ADNIGO: 1=Atypical clinical course or features (specify); 2=Stroke(s); 3=Depression; 4=Delirium; 5=Parkinsonism; 6=Metabolic / Toxic Disorder (specify); 7=Other (specify)

ADNI2: 1=Atypical clinical course or features (specify); 2=Stroke(s); 3=Depression; 4=Delirium; 5=Parkinsonism; 6=Metabolic / Toxic Disorder (specify); 7=Other (specify)

ADNI3: 1=Atypical clinical course or features (specify);2=Stroke(s);3=Depression;4=Delirium;5=Parkinsonism;6=Metabolic / Toxic Disorder (specify);7=Other (specify)

ADNI4: 1=Atypical clinical course or features (specify); 2=Stroke(s); 3=Depression; 4=Delirium; 5=Parkinsonism; 6=Metabolic / Toxic Disorder (specify); 7=Other (specify)

DXPARK character 4b. Parkinsonism symptoms present?

ADNI1: 1=Yes

ADNIGO: 1=Yes; 0=No

ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

DXPDES character If Parkinsonism, select box which indicates best diagnosis: 1=Parkinsonism without cognitive impairment; 2=Parkinsonism with cognitive impairment, not demented; 3=Parkinsonism with cognitive impairment, demented; 4=Atypical Parkinsonism Decoded Value: Yes

DXPCOG character If Parkinsonism with cognitive impairment, demented 1=PD;2=PDD;3=DLB;4=PDAD Decoded Value: Yes

DXPATYP character If Atypical Parkinsonism 1=PSP;2=CBGD;3=OPCA;4=SND;5=Shy Drager;6=Vascular;7=Other (specify) Decoded Value: Yes

DXDEP character 4a. Depressive symptoms present?

ADNIGO: 1=Yes; 0=No

ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

DXOTHDEM character Other Dementia (not Alzheimer's Disease) 1=Yes Decoded Value: Yes

DXODES character If dementia due to other etiology, select best diagnosis:

ADNI1: 1=Frontal; 3=Huntington's Disease; 5=Alcohol-related Dementia; 6=NPH; 7=Major Depression; 9=Vascular Dementia; 10=Prion-Associated Dementia; 11=HIV; 12=Primary Progressive Aphasia; 13=Posterior Cortical Dysfunction; 14=Other (specify)

ADNIGO: 1=Fronto-temporal Dementia; 2=Parkinson's Disease; 3=Huntington's Disease; 4=Progressive Supranuclear Palsy; 5=Alcohol-related Dementia; 6=NPH; 7=Major Depression; 8=Corticobasal Degeneration; 9=Vascular Dementia; 10=Prion-Associated Dementia; 11=HIV; 12=Primary Progressive Aphasia; 13=Posterior Cortical Dysfunction; 14=Other (specify)

ADNI2: 1=Fronto-temporal Dementia; 2=Parkinson's Disease; 3=Huntington's Disease; 4=Progressive Supranuclear Palsy; 5=Alcohol-related Dementia; 6=NPH; 7=Major Depression; 8=Corticobasal Degeneration; 9=Vascular Dementia; 10=Prion-Associated Dementia; 11=HIV; 12=Primary Progressive Aphasia; 13=Posterior Cortical Dysfunction; 14=Other (specify)

ADNI3: 1=Fronto-temporal Dementia;2=Parkinson's Disease;3=Huntington's Disease;4=Progressive Supranuclear Palsy;5=Alcohol-related Dementia;6=NPH;7=Major Depression;8=Corticobasal Degeneration;9=Vascular Dementia;10=Prion-Associated Dementia;11=HIV;12=Primary Progressive Aphasia;13=Posterior Cortical Dysfunction;14=Other (specify)

ADNI4: 1=Fronto-temporal Dementia; 10=Prion-Associated Dementia; 11=HIV; 12=Primary Progressive Aphasia; 13=Posterior Cortical Dysfunction; 14=Other (specify); 2=Parkinson's Disease; 3=Huntington's Disease; 4=Progressive Supranuclear Palsy; 5=Alcohol-related Dementia; 6=NPH; 7=Major Depression; 8=Corticobasal Degeneration; 9=Vascular Dementia

Decoded Value: Yes

DXCONFID character Physician Confidence in Diagnosis: 1=Uncertain; 2=Mildly Confident; 3=Moderately Confident; 4=Highly Confident Decoded Value: Yes

ID numeric Record ID Range value: 2, ..., 185684

SITEID numeric Site ID Range value: 1, ..., 941

USERDATE Date User Create Date Date: YYY-MM-DD

USERDATE2 Date User Last Modify Date Date: YYY-MM-DD

DD_CRF_VERSION_LABEL character CRF Version Label , with options: v1

LANGUAGE_CODE character Language Code e=English Decoded Value: Yes

HAS_QC_ERROR character Has Quality Check Error 0=Does not have QC error or QC error has been approved; 1=Has QC error Decoded Value: Yes

update_stamp POSIXct POSIXct: YYY-MM-DD

Author

adni-data@googlegroups.com