Everyday Cognition 12 Item - Participant Self-Report

Everyday Cognition 12 Item - Participant Self-Report data. More information is available at https://adni.loni.usc.edu/data-samples/adni-data/

Usage

data(ECOG12PT)

Format

A data frame with 1161 observations and 32 variables.

Source

https://adni.loni.usc.edu/data-samples/adni-data/

Details

ORIGPROT character Original study protocol , with options: ADNI4, ADNI2, ADNI3, ADNIGO, ADNI1

COLPROT factor Study protocol of data collection , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4

PTID character Participant ID

RID numeric Participant roster ID Range value: 21, ..., 10898

VISCODE character Visit code , with options: 4_bl, 4_init, 4_m12, 4_m24

VISCODE2 character Translated visit code

COLDATE Date collected Date: YYY-MM-DD

STAFFASST character 1. How much assistance was needed from the research staff to complete this form? 1=None ; 2=A little ; 3=A lot ; 4=Participant too cognitively impaired to complete Decoded Value: Yes

VALIDITY character 2. Validity of the information collected from the participant (based on research assistant's observations and interactions with the participant while they were completing the ECog): 1=1- Appeared Valid–The participant did not express any confusion or ask for any clarification in completing the ECog. ; 2=2- Questionable Validity–The participant needed to have the instructions repeated, and/or needed help in understanding a few of the items. ; 3=3- Appeared Invalid–The participant was clearly confused about the content of many of the questions or how to complete the form in general. ; 4=No information collected from participant Decoded Value: Yes

SOURCE character Information Source: 1=Participant Visit ; 2=Telephone Call/Video Call ; 3=Mail Decoded Value: Yes

CONCERN character Are you concerned that you have a memory or other thinking problem? 0=No ; 1=Yes Decoded Value: Yes

MEMREF character Compared to one year ago, do you feel that your memory has declined substantially? 0=No ; 1=Yes Decoded Value: Yes

ECOGPT1 character 1. Remembering where you have placed personal items or objects: 1=1- Better or no change ; 2=2- Questionable or occasionally worse ; 3=3- Consistently a little worse ; 4=4- Consistently much worse ; 9=9- I don’t know or not applicable Decoded Value: Yes

ECOGPT2 character 2. Remembering the current date or day of the week 1=1- Better or no change ; 2=2- Questionable or occasionally worse ; 3=3- Consistently a little worse ; 4=4- Consistently much worse ; 9=9- I don’t know or not applicable Decoded Value: Yes

ECOGPT3 character 3. Communicating or expressing ideas in a conversation 1=1- Better or no change ; 2=2- Questionable or occasionally worse ; 3=3- Consistently a little worse ; 4=4- Consistently much worse ; 9=9- I don’t know or not applicable Decoded Value: Yes

ECOGPT4 character 4. Understanding spoken directions or instructions 1=1- Better or no change ; 2=2- Questionable or occasionally worse ; 3=3- Consistently a little worse ; 4=4- Consistently much worse ; 9=9- I don’t know or not applicable Decoded Value: Yes

ECOGPT5 character 5. Difficulty calculating distances when driving or driving too close to another vehicle 1=1- Better or no change ; 2=2- Questionable or occasionally worse ; 3=3- Consistently a little worse ; 4=4- Consistently much worse ; 9=9- I don’t know or not applicable Decoded Value: Yes

ECOGPT6 character 6. Finding your way around a familiar home 1=1- Better or no change ; 2=2- Questionable or occasionally worse ; 3=3- Consistently a little worse ; 4=4- Consistently much worse ; 9=9- I don’t know or not applicable Decoded Value: Yes

ECOGPT7 character 7. Anticipating weather changes and planning accordingly 1=1- Better or no change ; 2=2- Questionable or occasionally worse ; 3=3- Consistently a little worse ; 4=4- Consistently much worse ; 9=9- I don’t know or not applicable Decoded Value: Yes

ECOGPT8 character 8. Planning a trip, vacation, or outing 1=1- Better or no change ; 2=2- Questionable or occasionally worse ; 3=3- Consistently a little worse ; 4=4- Consistently much worse ; 9=9- I don’t know or not applicable Decoded Value: Yes

ECOGPT9 character 9. Keeping living and work space (such as a home office, garage, or pantry) organized 1=1- Better or no change ; 2=2- Questionable or occasionally worse ; 3=3- Consistently a little worse ; 4=4- Consistently much worse ; 9=9- I don’t know or not applicable Decoded Value: Yes

ECOGPT10 character 10. Managing bill payments 1=1- Better or no change ; 2=2- Questionable or occasionally worse ; 3=3- Consistently a little worse ; 4=4- Consistently much worse ; 9=9- I don’t know or not applicable Decoded Value: Yes

ECOGPT11 character 11. Doing two things at once 1=1- Better or no change ; 2=2- Questionable or occasionally worse ; 3=3- Consistently a little worse ; 4=4- Consistently much worse ; 9=9- I don’t know or not applicable Decoded Value: Yes

ECOGPT12 character 12. Working on a particular task while talking to someone at the same time 1=1- Better or no change ; 2=2- Questionable or occasionally worse ; 3=3- Consistently a little worse ; 4=4- Consistently much worse ; 9=9- I don’t know or not applicable Decoded Value: Yes

ID numeric Record ID Range value: 1281, ..., 120520

SITEID numeric Site ID Range value: 2, ..., 941

USERDATE User Create Date Date: YYY-MM-DD

USERDATE2 User Last Modify Date Date: YYY-MM-DD

DD_CRF_VERSION_LABEL character CRF Version Label , with options: v1

LANGUAGE_CODE character Language Code e=English Decoded Value: Yes

HAS_QC_ERROR character Has Quality Check Error 0=Does not have QC error or QC error has been approved ; 1=Has QC error Decoded Value: Yes

update_stamp POSIXct: YYY-MM-DD

Author

adni-data@googlegroups.com