Everyday Cognition 12 Item - Study Partner Report

Everyday Cognition 12 Item - Study Partner Report data. More information is available at https://adni.loni.usc.edu/data-samples/adni-data/

Usage

data(ECOG12SP)

Format

A data frame with 746 observations and 29 variables.

Source

https://adni.loni.usc.edu/data-samples/adni-data/

Details

ORIGPROT character Original study protocol , with options: ADNI4, ADNI2, ADNI3, ADNIGO, ADNI1

COLPROT factor Study protocol of data collection , with factor levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4

PTID character Participant ID

RID numeric Participant roster ID Range value: 21, ..., 10618

VISCODE character Visit code , with options: 4_bl, 4_init, 4_m12

VISCODE2 character Translated visit code

COLDATE Date collected Date: YYY-MM-DD

SPID numeric Study Partner ID: Range value: 1, ..., 10

SOURCE character Information Source: 1=Participant Visit; 2=Telephone Call/Video Call; 3=Mail Decoded Value: Yes

ECOGSP1 character 1. Remembering where they have placed personal items or objects: 1=1- Better or no change; 2=2- Questionable or occasionally worse; 3=3- Consistently a little worse; 4=4- Consistently much worse; 9=9- I don’t know or not applicable Decoded Value: Yes

ECOGSP2 character 2. Remembering the current date or day of the week 1=1- Better or no change; 2=2- Questionable or occasionally worse; 3=3- Consistently a little worse; 4=4- Consistently much worse; 9=9- I don’t know or not applicable Decoded Value: Yes

ECOGSP3 character 3. Communicating or expressing ideas in a conversation 1=1- Better or no change; 2=2- Questionable or occasionally worse; 3=3- Consistently a little worse; 4=4- Consistently much worse; 9=9- I don’t know or not applicable Decoded Value: Yes

ECOGSP4 character 4. Understanding spoken directions or instructions 1=1- Better or no change; 2=2- Questionable or occasionally worse; 3=3- Consistently a little worse; 4=4- Consistently much worse; 9=9- I don’t know or not applicable Decoded Value: Yes

ECOGSP5 character 5. Difficulty calculating distances when driving or driving too close to another vehicle 1=1- Better or no change; 2=2- Questionable or occasionally worse; 3=3- Consistently a little worse; 4=4- Consistently much worse; 9=9- I don’t know or not applicable Decoded Value: Yes

ECOGSP6 character 6. Finding their way around a familiar home 1=1- Better or no change; 2=2- Questionable or occasionally worse; 3=3- Consistently a little worse; 4=4- Consistently much worse; 9=9- I don’t know or not applicable Decoded Value: Yes

ECOGSP7 character 7. Anticipating weather changes and planning accordingly 1=1- Better or no change; 2=2- Questionable or occasionally worse; 3=3- Consistently a little worse; 4=4- Consistently much worse; 9=9- I don’t know or not applicable Decoded Value: Yes

ECOGSP8 character 8. Planning a trip, vacation, or outing 1=1- Better or no change; 2=2- Questionable or occasionally worse; 3=3- Consistently a little worse; 4=4- Consistently much worse; 9=9- I don’t know or not applicable Decoded Value: Yes

ECOGSP9 character 9. Keeping living and work space (such as a home office, garage, or pantry) organized 1=1- Better or no change; 2=2- Questionable or occasionally worse; 3=3- Consistently a little worse; 4=4- Consistently much worse; 9=9- I don’t know or not applicable Decoded Value: Yes

ECOGSP10 character 10. Managing bill payments 1=1- Better or no change; 2=2- Questionable or occasionally worse; 3=3- Consistently a little worse; 4=4- Consistently much worse; 9=9- I don’t know or not applicable Decoded Value: Yes

ECOGSP11 character 11. Doing two things at once 1=1- Better or no change; 2=2- Questionable or occasionally worse; 3=3- Consistently a little worse; 4=4- Consistently much worse; 9=9- I don’t know or not applicable Decoded Value: Yes

ECOGSP12 character 12. Working on a particular task while talking to someone at the same time 1=1- Better or no change; 2=2- Questionable or occasionally worse; 3=3- Consistently a little worse; 4=4- Consistently much worse; 9=9- I don’t know or not applicable Decoded Value: Yes

ID numeric Record ID Range value: 1282, ..., 82136

SITEID numeric Site ID Range value: 2, ..., 941

USERDATE User Create Date Date: YYY-MM-DD

USERDATE2 User Last Modify Date Date: YYY-MM-DD

DD_CRF_VERSION_LABEL character CRF Version Label , with options: v1

LANGUAGE_CODE character Language Code e=English Decoded Value: Yes

HAS_QC_ERROR character Has Quality Check Error 0=Does not have QC error or QC error has been approved; 1=Has QC error Decoded Value: Yes

update_stamp POSIXct: YYY-MM-DD

Author

adni-data@googlegroups.com