Everyday Cognition - Participant Self-Report

Everyday Cognition - Participant Self-Report data. More information is available at https://adni.loni.usc.edu/data-samples/adni-data/

Usage

data(ECOGPT)

Format

A data frame with 8088 observations and 63 variables.

Source

https://adni.loni.usc.edu/data-samples/adni-data/

Details

ORIGPROT character Original study protocol , with options: ADNI1, ADNIGO, ADNI2, ADNI3

COLPROT factor Study protocol of data collection , with factor levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4

PTID character Participant ID

RID numeric Range value: 2, ..., 7125

VISCODE character

VISCODE2 character Translated visit code

VISDATE Date Assessment COLDATE when present; otherwise Registry EXAMDATE Date: YYY-MM-DD

CONCERN character Are you concerned that you have a memory or other thinking problem?

ADNIGO: 1=Yes;0=No

ADNI2: 1=Yes;0=No

ADNI3: 0=No;1=Yes

Decoded Value: Yes

MEMORY1 character 1. Remembering a few shopping items without a list.

ADNIGO: 1=1- Better or no change; 2=2- Questionable/occasionally worse; 3=3- Consistently a little worse; 4=4- Consistently much worse; 9=9- I don't know

ADNI2: 1=1- Better or no change; 2=2- Questionable/occasionally worse; 3=3- Consistently a little worse; 4=4- Consistently much worse; 9=9- I don't know

ADNI3: 1=1- Better or no change;2=2- Questionable/occasionally worse;3=3- Consistently a little worse;4=4- Consistently much worse;9=9- I don't know

Decoded Value: Yes

MEMORY2 character 2. Remembering things that happened recently (such as recent outings, events in the news).

ADNIGO: 1=1- Better or no change; 2=2- Questionable/occasionally worse; 3=3- Consistently a little worse; 4=4- Consistently much worse; 9=9- I don't know

ADNI2: 1=1- Better or no change; 2=2- Questionable/occasionally worse; 3=3- Consistently a little worse; 4=4- Consistently much worse; 9=9- I don't know

ADNI3: 1=1- Better or no change;2=2- Questionable/occasionally worse;3=3- Consistently a little worse;4=4- Consistently much worse;9=9- I don't know

Decoded Value: Yes

MEMORY3 character 3. Recalling conversations a few days later.

ADNIGO: 1=1- Better or no change; 2=2- Questionable/occasionally worse; 3=3- Consistently a little worse; 4=4- Consistently much worse; 9=9- I don't know

ADNI2: 1=1- Better or no change; 2=2- Questionable/occasionally worse; 3=3- Consistently a little worse; 4=4- Consistently much worse; 9=9- I don't know

ADNI3: 1=1- Better or no change;2=2- Questionable/occasionally worse;3=3- Consistently a little worse;4=4- Consistently much worse;9=9- I don't know

Decoded Value: Yes

MEMORY4 character 4. Remembering where I have placed objects.

ADNIGO: 1=1- Better or no change; 2=2- Questionable/occasionally worse; 3=3- Consistently a little worse; 4=4- Consistently much worse; 9=9- I don't know

ADNI2: 1=1- Better or no change; 2=2- Questionable/occasionally worse; 3=3- Consistently a little worse; 4=4- Consistently much worse; 9=9- I don't know

ADNI3: 1=1- Better or no change;2=2- Questionable/occasionally worse;3=3- Consistently a little worse;4=4- Consistently much worse;9=9- I don't know

Decoded Value: Yes

MEMORY5 character 5. Repeating stories and/or questions.

ADNIGO: 1=1- Better or no change; 2=2- Questionable/occasionally worse; 3=3- Consistently a little worse; 4=4- Consistently much worse; 9=9- I don't know

ADNI2: 1=1- Better or no change; 2=2- Questionable/occasionally worse; 3=3- Consistently a little worse; 4=4- Consistently much worse; 9=9- I don't know

ADNI3: 1=1- Better or no change;2=2- Questionable/occasionally worse;3=3- Consistently a little worse;4=4- Consistently much worse;9=9- I don't know

Decoded Value: Yes

MEMORY6 character 6. Remembering the current date or day of the week.

ADNIGO: 1=1- Better or no change; 2=2- Questionable/occasionally worse; 3=3- Consistently a little worse; 4=4- Consistently much worse; 9=9- I don't know

ADNI2: 1=1- Better or no change; 2=2- Questionable/occasionally worse; 3=3- Consistently a little worse; 4=4- Consistently much worse; 9=9- I don't know

ADNI3: 1=1- Better or no change;2=2- Questionable/occasionally worse;3=3- Consistently a little worse;4=4- Consistently much worse;9=9- I don't know

Decoded Value: Yes

MEMORY7 character 7. Remembering I have already told someone something.

ADNIGO: 1=1- Better or no change; 2=2- Questionable/occasionally worse; 3=3- Consistently a little worse; 4=4- Consistently much worse; 9=9- I don't know

ADNI2: 1=1- Better or no change; 2=2- Questionable/occasionally worse; 3=3- Consistently a little worse; 4=4- Consistently much worse; 9=9- I don't know

ADNI3: 1=1- Better or no change;2=2- Questionable/occasionally worse;3=3- Consistently a little worse;4=4- Consistently much worse;9=9- I don't know

Decoded Value: Yes

MEMORY8 character 8. Remembering appointments, meetings, or engagements.

ADNIGO: 1=1- Better or no change; 2=2- Questionable/occasionally worse; 3=3- Consistently a little worse; 4=4- Consistently much worse; 9=9- I don't know

ADNI2: 1=1- Better or no change; 2=2- Questionable/occasionally worse; 3=3- Consistently a little worse; 4=4- Consistently much worse; 9=9- I don't know

ADNI3: 1=1- Better or no change;2=2- Questionable/occasionally worse;3=3- Consistently a little worse;4=4- Consistently much worse;9=9- I don't know

Decoded Value: Yes

LANG1 character 1. Forgetting the names of objects.

ADNIGO: 1=1- Better or no change; 2=2- Questionable/occasionally worse; 3=3- Consistently a little worse; 4=4- Consistently much worse; 9=9- I don't know

ADNI2: 1=1- Better or no change; 2=2- Questionable/occasionally worse; 3=3- Consistently a little worse; 4=4- Consistently much worse; 9=9- I don't know

ADNI3: 1=1- Better or no change;2=2- Questionable/occasionally worse;3=3- Consistently a little worse;4=4- Consistently much worse;9=9- I don't know

Decoded Value: Yes

LANG2 character 2. Verbally giving instructions to others.

ADNIGO: 1=1- Better or no change; 2=2- Questionable/occasionally worse; 3=3- Consistently a little worse; 4=4- Consistently much worse; 9=9- I don't know

ADNI2: 1=1- Better or no change; 2=2- Questionable/occasionally worse; 3=3- Consistently a little worse; 4=4- Consistently much worse; 9=9- I don't know

ADNI3: 1=1- Better or no change;2=2- Questionable/occasionally worse;3=3- Consistently a little worse;4=4- Consistently much worse;9=9- I don't know

Decoded Value: Yes

LANG3 character 3. Finding the right words to use in conversations.

ADNIGO: 1=1- Better or no change; 2=2- Questionable/occasionally worse; 3=3- Consistently a little worse; 4=4- Consistently much worse; 9=9- I don't know

ADNI2: 1=1- Better or no change; 2=2- Questionable/occasionally worse; 3=3- Consistently a little worse; 4=4- Consistently much worse; 9=9- I don't know

ADNI3: 1=1- Better or no change;2=2- Questionable/occasionally worse;3=3- Consistently a little worse;4=4- Consistently much worse;9=9- I don't know

Decoded Value: Yes

LANG4 character 4. Communicating thoughts in a conversation.

ADNIGO: 1=1- Better or no change; 2=2- Questionable/occasionally worse; 3=3- Consistently a little worse; 4=4- Consistently much worse; 9=9- I don't know

ADNI2: 1=1- Better or no change; 2=2- Questionable/occasionally worse; 3=3- Consistently a little worse; 4=4- Consistently much worse; 9=9- I don't know

ADNI3: 1=1- Better or no change;2=2- Questionable/occasionally worse;3=3- Consistently a little worse;4=4- Consistently much worse;9=9- I don't know

Decoded Value: Yes

LANG5 character 5. Following a story in a book or on TV.

ADNIGO: 1=1- Better or no change; 2=2- Questionable/occasionally worse; 3=3- Consistently a little worse; 4=4- Consistently much worse; 9=9- I don't know

ADNI2: 1=1- Better or no change; 2=2- Questionable/occasionally worse; 3=3- Consistently a little worse; 4=4- Consistently much worse; 9=9- I don't know

ADNI3: 1=1- Better or no change;2=2- Questionable/occasionally worse;3=3- Consistently a little worse;4=4- Consistently much worse;9=9- I don't know

Decoded Value: Yes

LANG6 character 6. Understanding the point of what other people are trying to say.

ADNIGO: 1=1- Better or no change; 2=2- Questionable/occasionally worse; 3=3- Consistently a little worse; 4=4- Consistently much worse; 9=9- I don't know

ADNI2: 1=1- Better or no change; 2=2- Questionable/occasionally worse; 3=3- Consistently a little worse; 4=4- Consistently much worse; 9=9- I don't know

ADNI3: 1=1- Better or no change;2=2- Questionable/occasionally worse;3=3- Consistently a little worse;4=4- Consistently much worse;9=9- I don't know

Decoded Value: Yes

LANG7 character 7. Remembering the meaning of common words.

ADNIGO: 1=1- Better or no change; 2=2- Questionable/occasionally worse; 3=3- Consistently a little worse; 4=4- Consistently much worse; 9=9- I don't know

ADNI2: 1=1- Better or no change; 2=2- Questionable/occasionally worse; 3=3- Consistently a little worse; 4=4- Consistently much worse; 9=9- I don't know

ADNI3: 1=1- Better or no change;2=2- Questionable/occasionally worse;3=3- Consistently a little worse;4=4- Consistently much worse;9=9- I don't know

Decoded Value: Yes

LANG8 character 8. Describing a program I have watched on TV.

ADNIGO: 1=1- Better or no change; 2=2- Questionable/occasionally worse; 3=3- Consistently a little worse; 4=4- Consistently much worse; 9=9- I don't know

ADNI2: 1=1- Better or no change; 2=2- Questionable/occasionally worse; 3=3- Consistently a little worse; 4=4- Consistently much worse; 9=9- I don't know

ADNI3: 1=1- Better or no change;2=2- Questionable/occasionally worse;3=3- Consistently a little worse;4=4- Consistently much worse;9=9- I don't know

Decoded Value: Yes

LANG9 character 9. Understanding spoken directions or instructions.

ADNIGO: 1=1- Better or no change; 2=2- Questionable/occasionally worse; 3=3- Consistently a little worse; 4=4- Consistently much worse; 9=9- I don't know

ADNI2: 1=1- Better or no change; 2=2- Questionable/occasionally worse; 3=3- Consistently a little worse; 4=4- Consistently much worse; 9=9- I don't know

ADNI3: 1=1- Better or no change;2=2- Questionable/occasionally worse;3=3- Consistently a little worse;4=4- Consistently much worse;9=9- I don't know

Decoded Value: Yes

VISSPAT1 character 1. Following a map to find a new location.

ADNIGO: 1=1- Better or no change; 2=2- Questionable/occasionally worse; 3=3- Consistently a little worse; 4=4- Consistently much worse; 9=9- I don't know

ADNI2: 1=1- Better or no change; 2=2- Questionable/occasionally worse; 3=3- Consistently a little worse; 4=4- Consistently much worse; 9=9- I don't know

ADNI3: 1=1- Better or no change;2=2- Questionable/occasionally worse;3=3- Consistently a little worse;4=4- Consistently much worse;9=9- I don't know

Decoded Value: Yes

VISSPAT2 character 2. Reading a map and helping with directions when someone else is driving.

ADNIGO: 1=1- Better or no change; 2=2- Questionable/occasionally worse; 3=3- Consistently a little worse; 4=4- Consistently much worse; 9=9- I don't know

ADNI2: 1=1- Better or no change; 2=2- Questionable/occasionally worse; 3=3- Consistently a little worse; 4=4- Consistently much worse; 9=9- I don't know

ADNI3: 1=1- Better or no change;2=2- Questionable/occasionally worse;3=3- Consistently a little worse;4=4- Consistently much worse;9=9- I don't know

Decoded Value: Yes

VISSPAT3 character 3. Finding my car in a parking lot.

ADNIGO: 1=1- Better or no change; 2=2- Questionable/occasionally worse; 3=3- Consistently a little worse; 4=4- Consistently much worse; 9=9- I don't know

ADNI2: 1=1- Better or no change; 2=2- Questionable/occasionally worse; 3=3- Consistently a little worse; 4=4- Consistently much worse; 9=9- I don't know

ADNI3: 1=1- Better or no change;2=2- Questionable/occasionally worse;3=3- Consistently a little worse;4=4- Consistently much worse;9=9- I don't know

Decoded Value: Yes

VISSPAT4 character 4. Finding my way back to a meeting spot in the mall or other location.

ADNIGO: 1=1- Better or no change; 2=2- Questionable/occasionally worse; 3=3- Consistently a little worse; 4=4- Consistently much worse; 9=9- I don't know

ADNI2: 1=1- Better or no change; 2=2- Questionable/occasionally worse; 3=3- Consistently a little worse; 4=4- Consistently much worse; 9=9- I don't know

ADNI3: 1=1- Better or no change;2=2- Questionable/occasionally worse;3=3- Consistently a little worse;4=4- Consistently much worse;9=9- I don't know

Decoded Value: Yes

VISSPAT5 character 5. Finding my way around a familiar neighborhood.

ADNIGO: 1=1- Better or no change; 2=2- Questionable/occasionally worse; 3=3- Consistently a little worse; 4=4- Consistently much worse; 9=9- I don't know

ADNI2: 1=1- Better or no change; 2=2- Questionable/occasionally worse; 3=3- Consistently a little worse; 4=4- Consistently much worse; 9=9- I don't know

ADNI3: 1=1- Better or no change;2=2- Questionable/occasionally worse;3=3- Consistently a little worse;4=4- Consistently much worse;9=9- I don't know

Decoded Value: Yes

VISSPAT6 character 6. Finding my way around a familiar store.

ADNIGO: 1=1- Better or no change; 2=2- Questionable/occasionally worse; 3=3- Consistently a little worse; 4=4- Consistently much worse; 9=9- I don't know

ADNI2: 1=1- Better or no change; 2=2- Questionable/occasionally worse; 3=3- Consistently a little worse; 4=4- Consistently much worse; 9=9- I don't know

ADNI3: 1=1- Better or no change;2=2- Questionable/occasionally worse;3=3- Consistently a little worse;4=4- Consistently much worse;9=9- I don't know

Decoded Value: Yes

VISSPAT7 character 7. Finding my way around a house visited many times.

ADNIGO: 1=1- Better or no change; 2=2- Questionable/occasionally worse; 3=3- Consistently a little worse; 4=4- Consistently much worse; 9=9- I don't know

ADNI2: 1=1- Better or no change; 2=2- Questionable/occasionally worse; 3=3- Consistently a little worse; 4=4- Consistently much worse; 9=9- I don't know

ADNI3: 1=1- Better or no change;2=2- Questionable/occasionally worse;3=3- Consistently a little worse;4=4- Consistently much worse;9=9- I don't know

Decoded Value: Yes

VISSPAT8 character 7. Finding my way around a house visited many times. 1=1- Better or no change; 2=2- Questionable/occasionally worse; 3=3- Consistently a little worse; 4=4- Consistently much worse; 9=9- I don't know Decoded Value: Yes

PLAN1 character 1. Planning a sequence of stops on a shopping trip.

ADNIGO: 1=1- Better or no change; 2=2- Questionable/occasionally worse; 3=3- Consistently a little worse; 4=4- Consistently much worse; 9=9- I don't know

ADNI2: 1=1- Better or no change; 2=2- Questionable/occasionally worse; 3=3- Consistently a little worse; 4=4- Consistently much worse; 9=9- I don't know

ADNI3: 1=1- Better or no change;2=2- Questionable/occasionally worse;3=3- Consistently a little worse;4=4- Consistently much worse;9=9- I don't know

Decoded Value: Yes

PLAN2 character 2. The ability to anticipate weather changes and plan accordingly (i.e., bring a coat or umbrella)

ADNIGO: 1=1- Better or no change; 2=2- Questionable/occasionally worse; 3=3- Consistently a little worse; 4=4- Consistently much worse; 9=9- I don't know

ADNI2: 1=1- Better or no change; 2=2- Questionable/occasionally worse; 3=3- Consistently a little worse; 4=4- Consistently much worse; 9=9- I don't know

ADNI3: 1=1- Better or no change;2=2- Questionable/occasionally worse;3=3- Consistently a little worse;4=4- Consistently much worse;9=9- I don't know

Decoded Value: Yes

PLAN3 character 3. Developing a schedule in advance of anticipated events.

ADNIGO: 1=1- Better or no change; 2=2- Questionable/occasionally worse; 3=3- Consistently a little worse; 4=4- Consistently much worse; 9=9- I don't know

ADNI2: 1=1- Better or no change; 2=2- Questionable/occasionally worse; 3=3- Consistently a little worse; 4=4- Consistently much worse; 9=9- I don't know

ADNI3: 1=1- Better or no change;2=2- Questionable/occasionally worse;3=3- Consistently a little worse;4=4- Consistently much worse;9=9- I don't know

Decoded Value: Yes

PLAN4 character 4. Thinking things through before acting.

ADNIGO: 1=1- Better or no change; 2=2- Questionable/occasionally worse; 3=3- Consistently a little worse; 4=4- Consistently much worse; 9=9- I don't know

ADNI2: 1=1- Better or no change; 2=2- Questionable/occasionally worse; 3=3- Consistently a little worse; 4=4- Consistently much worse; 9=9- I don't know

ADNI3: 1=1- Better or no change;2=2- Questionable/occasionally worse;3=3- Consistently a little worse;4=4- Consistently much worse;9=9- I don't know

Decoded Value: Yes

PLAN5 character 5. Thinking ahead.

ADNIGO: 1=1- Better or no change; 2=2- Questionable/occasionally worse; 3=3- Consistently a little worse; 4=4- Consistently much worse; 9=9- I don't know

ADNI2: 1=1- Better or no change; 2=2- Questionable/occasionally worse; 3=3- Consistently a little worse; 4=4- Consistently much worse; 9=9- I don't know

ADNI3: 1=1- Better or no change;2=2- Questionable/occasionally worse;3=3- Consistently a little worse;4=4- Consistently much worse;9=9- I don't know

Decoded Value: Yes

ORGAN1 character 1. Keeping living and work space organized.

ADNIGO: 1=1- Better or no change; 2=2- Questionable/occasionally worse; 3=3- Consistently a little worse; 4=4- Consistently much worse; 9=9- I don't know

ADNI2: 1=1- Better or no change; 2=2- Questionable/occasionally worse; 3=3- Consistently a little worse; 4=4- Consistently much worse; 9=9- I don't know

ADNI3: 1=1- Better or no change;2=2- Questionable/occasionally worse;3=3- Consistently a little worse;4=4- Consistently much worse;9=9- I don't know

Decoded Value: Yes

ORGAN2 character 2. Balancing the checkbook without error.

ADNIGO: 1=1- Better or no change; 2=2- Questionable/occasionally worse; 3=3- Consistently a little worse; 4=4- Consistently much worse; 9=9- I don't know

ADNI2: 1=1- Better or no change; 2=2- Questionable/occasionally worse; 3=3- Consistently a little worse; 4=4- Consistently much worse; 9=9- I don't know

ADNI3: 1=1- Better or no change;2=2- Questionable/occasionally worse;3=3- Consistently a little worse;4=4- Consistently much worse;9=9- I don't know

Decoded Value: Yes

ORGAN3 character 3. Keeping financial records organized.

ADNIGO: 1=1- Better or no change; 2=2- Questionable/occasionally worse; 3=3- Consistently a little worse; 4=4- Consistently much worse; 9=9- I don't know

ADNI2: 1=1- Better or no change; 2=2- Questionable/occasionally worse; 3=3- Consistently a little worse; 4=4- Consistently much worse; 9=9- I don't know

ADNI3: 1=1- Better or no change;2=2- Questionable/occasionally worse;3=3- Consistently a little worse;4=4- Consistently much worse;9=9- I don't know

Decoded Value: Yes

ORGAN4 character 4. Prioritizing tasks by importance.

ADNIGO: 1=1- Better or no change; 2=2- Questionable/occasionally worse; 3=3- Consistently a little worse; 4=4- Consistently much worse; 9=9- I don't know

ADNI2: 1=1- Better or no change; 2=2- Questionable/occasionally worse; 3=3- Consistently a little worse; 4=4- Consistently much worse; 9=9- I don't know

ADNI3: 1=1- Better or no change;2=2- Questionable/occasionally worse;3=3- Consistently a little worse;4=4- Consistently much worse;9=9- I don't know

Decoded Value: Yes

ORGAN5 character 5. Keeping mail and papers organized.

ADNIGO: 1=1- Better or no change; 2=2- Questionable/occasionally worse; 3=3- Consistently a little worse; 4=4- Consistently much worse; 9=9- I don't know

ADNI2: 1=1- Better or no change; 2=2- Questionable/occasionally worse; 3=3- Consistently a little worse; 4=4- Consistently much worse; 9=9- I don't know

ADNI3: 1=1- Better or no change;2=2- Questionable/occasionally worse;3=3- Consistently a little worse;4=4- Consistently much worse;9=9- I don't know

Decoded Value: Yes

ORGAN6 character 6. Using an organized strategy to manage a medication schedule involving multiple medications.

ADNIGO: 1=1- Better or no change; 2=2- Questionable/occasionally worse; 3=3- Consistently a little worse; 4=4- Consistently much worse; 9=9- I don't know

ADNI2: 1=1- Better or no change; 2=2- Questionable/occasionally worse; 3=3- Consistently a little worse; 4=4- Consistently much worse; 9=9- I don't know

ADNI3: 1=1- Better or no change;2=2- Questionable/occasionally worse;3=3- Consistently a little worse;4=4- Consistently much worse;9=9- I don't know

Decoded Value: Yes

DIVATT1 character 1. The ability to do two things at once.

ADNIGO: 1=1- Better or no change; 2=2- Questionable/occasionally worse; 3=3- Consistently a little worse; 4=4- Consistently much worse; 9=9- I don't know

ADNI2: 1=1- Better or no change; 2=2- Questionable/occasionally worse; 3=3- Consistently a little worse; 4=4- Consistently much worse; 9=9- I don't know

ADNI3: 1=1- Better or no change;2=2- Questionable/occasionally worse;3=3- Consistently a little worse;4=4- Consistently much worse;9=9- I don't know

Decoded Value: Yes

DIVATT2 character 2. Returning to a task after being interrupted.

ADNIGO: 1=1- Better or no change; 2=2- Questionable/occasionally worse; 3=3- Consistently a little worse; 4=4- Consistently much worse; 9=9- I don't know

ADNI2: 1=1- Better or no change; 2=2- Questionable/occasionally worse; 3=3- Consistently a little worse; 4=4- Consistently much worse; 9=9- I don't know

ADNI3: 1=1- Better or no change;2=2- Questionable/occasionally worse;3=3- Consistently a little worse;4=4- Consistently much worse;9=9- I don't know

Decoded Value: Yes

DIVATT3 character 3. The ability to concentrate on a task without being distracted by external things in the environment.

ADNIGO: 1=1- Better or no change; 2=2- Questionable/occasionally worse; 3=3- Consistently a little worse; 4=4- Consistently much worse; 9=9- I don't know

ADNI2: 1=1- Better or no change; 2=2- Questionable/occasionally worse; 3=3- Consistently a little worse; 4=4- Consistently much worse; 9=9- I don't know

ADNI3: 1=1- Better or no change;2=2- Questionable/occasionally worse;3=3- Consistently a little worse;4=4- Consistently much worse;9=9- I don't know

Decoded Value: Yes

DIVATT4 character 4. Cooking or working and talking at the same time.

ADNIGO: 1=1- Better or no change; 2=2- Questionable/occasionally worse; 3=3- Consistently a little worse; 4=4- Consistently much worse; 9=9- I don't know

ADNI2: 1=1- Better or no change; 2=2- Questionable/occasionally worse; 3=3- Consistently a little worse; 4=4- Consistently much worse; 9=9- I don't know

ADNI3: 1=1- Better or no change;2=2- Questionable/occasionally worse;3=3- Consistently a little worse;4=4- Consistently much worse;9=9- I don't know

Decoded Value: Yes

STAFFASST character 1. How much assistance was needed from the research staff to complete this form?

ADNIGO: 1=None; 2=A little; 3=A lot; 4=Participant too cognitively impaired to complete

ADNI2: 1=None; 2=A little; 3=A lot; 4=Participant too cognitively impaired to complete

ADNI3: 1=None;2=A little;3=A lot;4=Participant too cognitively impaired to complete

Decoded Value: Yes

VALIDITY character 2. Validity of the information collected from the participant (based on research assistant's observations and interactions with the participant while they were completing the ECog):

ADNIGO: 1=1- Appeared Valid–The participant did not express any confusion or ask for any clarification in completing the ECog.;2=2- Questionable Validity–The participant needed to have the instructions repeated, and/or needed help in understanding a few of the items.; 3=3- Appeared Invalid–The participant was clearly confused about the content of many of the questions or how to complete the form in general.;4=No information collected from participant

ADNI2: 1=1- Appeared Valid–The participant did not express any confusion or ask for any clarification in completing the ECog.;2=2- Questionable Validity–The participant needed to have the instructions repeated, and/or needed help in understanding a few of the items.; 3=3- Appeared Invalid–The participant was clearly confused about the content of many of the questions or how to complete the form in general.;4=No information collected from participant

ADNI3: 1=1- Appeared Valid–The participant did not express any confusion or ask for any clarification in completing the ECog.;2=2- Questionable Validity–The participant needed to have the instructions repeated, and/or needed help in understanding a few of the items.;3=3- Appeared Invalid–The participant was clearly confused about the content of many of the questions or how to complete the form in general.;4=No information collected from participant

Decoded Value: Yes

SOURCE character Information Source: 1=Participant Visit;2=Telephone Call;3=Mail Decoded Value: Yes

EcogPtDivatt numeric Range value: 1, ..., 4

EcogPtLang numeric Range value: 1, ..., 4

EcogPtMem numeric Range value: 1, ..., 4

EcogPtOrgan numeric Range value: 1, ..., 4

EcogPtPlan numeric Range value: 1, ..., 4

EcogPtVisspat numeric Range value: 1, ..., 4

EcogPtTotal numeric Range value: 1, ..., 4

ID numeric Range value: 4, ..., 185602

SITEID numeric Range value: 1, ..., 63

USERDATE Date Date: YYY-MM-DD

USERDATE2 Date Date: YYY-MM-DD

update_stamp POSIXct POSIXct: YYY-MM-DD

Author

adni-data@googlegroups.com