ApoE Genotyping - Draw Data

ApoE Genotyping - Draw Data data. More information is available at https://adni.loni.usc.edu/data-samples/adni-data/

Usage

data(GENETIC)

Format

A data frame with 10011 observations and 58 variables.

Source

https://adni.loni.usc.edu/data-samples/adni-data/

Details

ORIGPROT character Original study protocol , with options: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4

COLPROT factor Study protocol of data collection , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4

PTID character Participant ID

RID numeric Participant roster ID Range value: 2, ..., 10895

VISCODE character Visit code

VISCODE2 character Translated visit code

VISDATE Assessment COLLDATE when present; otherwise Registry EXAMDATE Date: YYY-MM-DD

RECNO numeric Range value: 1, ..., 4

APCOLLECT character On this date, was Whole blood long read sequencing collected (1x3 mL Purple-Top EDTA tube)?

ADNIGO, ADNI2: 1=Yes ; 0=No

ADNI3, ADNI4: 0=No ; 1=Yes

Decoded Value: Yes

DNAREASON character If no, indicate reason

ADNI2, ADNI3, ADNI4: 1=Illness ; 2=Participant unavailable ; 3=Participant unwilling ; 4=Administrative problems ; 5=Withdrawn consent ; 6=DNA sample previously collected ; 7=Other (specify)

ADNI4: 1=Illness ; 2=Participant unavailable ; 3=Participant unwilling ; 4=Administrative problems ; 5=Withdrawn consent ; 7=Other (specify)

Decoded Value: Yes

APTIME Time of blood collection (24hr clock) hms: HH-MM-SS

APVOLUME numeric If Whole Blood for long read sequencing collected, volume of tube (1 x 3mL Purple-Top EDTA tube):

ADNI1, ADNIGO, ADNI2: 1..10

ADNI3, ADNI4

WBLDVOL1 numeric Volume of aliquots created 1.5mL WBLD #1: Range value: 0, ..., 1.5

WBLDVOL2 numeric Volume of aliquots created 1.5mL WBLD #2: Range value: 0, ..., 1.5

WBLDFROZT Time long read sequencing was placed in freezer(24hr clock): hms: HH-MM-SS

RNACOLL character On this date, was RNA sample collected (2 x 2.5 mL PAXgene RNA tubes)?

ADNIGO, ADNI2: 1=Yes ; 0=No

ADNI3, ADNI4: 0=No ; 1=Yes

Decoded Value: Yes

RNAREASON character If no, indicate reason 1=Illness ; 2=Participant unavailable ; 3=Participant unwilling ; 4=Administrative problems ; 5=Withdrawn consent ; 7=Other (specify) Decoded Value: Yes

RNADATE Date of RNA collection: Date: YYY-MM-DD

RNATIME Time of RNA collection (24hr clock): hms: HH-MM-SS

RNAVOL numeric Volume of blood drawn into 3 x 2.5 mL PAXgene RNA Tubes

ADNIGO, ADNI2: 1..8

ADNI3

RNAVOL1 numeric If RNA collected, volume of tubes (2.5mL PAXgene™ RNA tubes): RNA #1 Range value: 1, ..., 8

RNAVOL2 numeric If RNA collected, volume of tubes (2.5mL PAXgene™ RNA tubes): RNA #2 Range value: 0, ..., 7

RNASHIP character Was this sample shipped ambient? 0=No ; 1=Yes Decoded Value: Yes

RNAFROZT If Frozen, Time RNA placed in freezer (24hr clock): hms: HH-MM-SS

BCOAT character On this date, was Buffy Coat extracted from 2 x 10mL EDTA biomarker collection tubes?

ADNI2: 1=Yes ; 0=No

ADNI3, ADNI4: 0=No ; 1=Yes

Decoded Value: Yes

BCREASON character If No, please provide reason why the Buffy Coat was not extracted 1=Biomarker not collected ; 2=Inadequate biomarker sample ; 3=Withdrawn consent ; 4=Other (specify) Decoded Value: Yes

BCVOL numeric If Buffy Coat extracted, volume of aliquot

ADNI2: 0..99

ADNI3

BCVOL1 numeric If plasma and buffy coat collection tubes drawn, volume of tubes (10mL Purple-Top EDTA tubes): EDTA #1 Range value: 1, ..., 10

BCVOL2 numeric If plasma and buffy coat collection tubes drawn, volume of tubes (10mL Purple-Top EDTA tubes): EDTA #2 Range value: 0, ..., 10

BCEXTVOL1 numeric If Buffy Coat extracted, volume of aliquots: Buffy Coat #1: Range value: 0.1, ..., 2

BCEXTVOL2 numeric If Buffy Coat extracted, volume of aliquots: Buffy Coat #2: Range value: 0, ..., 2

BCSHIP character Was this sample shipped ambient? 0=No ; 1=Yes Decoded Value: Yes

BCFROZT If Frozen, time Buffy Coat placed in freezer (24hr clock): hms: HH-MM-SS

RBCCOLL character On this date, was RBC extracted from biomarker collection tube? 0=No ; 1=Yes Decoded Value: Yes

RBCTUBE numeric If RBC collected, total number of aliquot tubes Range value: 1, ..., 2

RBCALIQ1 numeric RBC aliquot 1 volume (2.0 mL tube) Range value: 0.1, ..., 2

RBCALIQ2 numeric RBC aliquot 2 volume (4.0 mL tube) Range value: 0, ..., 5

RBCFROZT If RBC collected, time placed in freezer hms: HH-MM-SS

PBMCCOLL character On this date, was PBMC sample collected (2 x 10mL Sodium Heparin tubes)? 0=No ; 1=Yes Decoded Value: Yes

PBMCREASON character If no, indicate reason

ADNI4: 1=Illness ; 2=Participant unavailable ; 3=Participant unwilling ; 4=Administrative problems ; 5=Withdrawn consent ; 7=Other (specify)

ADNI4: 1=Illness ; 2=Participant unavailable ; 3=Participant unwilling ; 4=Administrative problems ; 5=Withdrawn consent ; 6=Already collected in ADNI3 ; 7=Other (specify)

Decoded Value: Yes

PBMCVOL numeric If PBMC collected, volume Range value: 2, ..., 10

PBMCTIME Time of PBMC collection (24hr clock): hms: HH-MM-SS

PBMCVOL1 numeric If PBMC collected, volume of tubes: PBMC #1: Range value: 1, ..., 10

PBMCVOL2 numeric If PBMC collected, volume of tubes: PBMC #2: Range value: 0, ..., 10

CLDATE Date of cell immortalization collection Date: YYY-MM-DD

CLCOLL character On this date, was cell immortalization sample collected?

ADNI2: 1=Yes ; 0=No

ADNI3: 0=No ; 1=Yes

Decoded Value: Yes

CLREASON character If no, indicate reason 1=Illness ; 2=Participant unavailable ; 3=Participant unwilling ; 4=Administrative problems ; 5=Withdrawn consent ; 6=Other (specify) Decoded Value: Yes

CLTIME Time of cell immortalization collection (24hr clock) hms: HH-MM-SS

CLVOLUME numeric Total volume of blood drawn for cell immortalization into 2 x 8.5 mL ACD-A (yellow top) tubes Range value: -1, ..., 20

GENFEDDATE character Date FedExed

ID numeric Record ID Range value: 5, ..., 186976

SITEID numeric Site ID Range value: 1, ..., 941

USERDATE User Create Date Date: YYY-MM-DD

USERDATE2 User Last Modify Date Date: YYY-MM-DD

DD_CRF_VERSION_LABEL character CRF Version Label , with options: init, bl, annual

LANGUAGE_CODE character Language Code e=English Decoded Value: Yes

HAS_QC_ERROR character Has Quality Check Error 0=Does not have QC error or QC error has been approved ; 1=Has QC error Decoded Value: Yes

update_stamp POSIXct: YYY-MM-DD

Author

adni-data@googlegroups.com