Impact of Events Scale

Impact of Events Scale data. More information is available at https://adni.loni.usc.edu/data-samples/adni-data/

Usage

data(IES)

Format

A data frame with 190 observations and 30 variables.

Source

https://adni.loni.usc.edu/data-samples/adni-data/

Details

ORIGPROT character Original study protocol , with options: ADNI3, ADNI4, ADNIGO, ADNI2, ADNI1

COLPROT factor Study protocol of data collection , with factor levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4

PTID character Participant ID

RID numeric Participant roster ID Range value: 21, ..., 10370

VISCODE character Visit code , with options: 4_init, 4_bl, 4_m12

VISCODE2 character Translated visit code

IESDATE Date Date Collected: Date: YYY-MM-DD

IES1 character 1. I thought about it when I didn't mean to. 1=Not at All; 2=Rarely; 3=Sometimes; 4=Often Decoded Value: Yes

IES2 character 2. I avoided letting myself get upset when I thought about it or was reminded of it. 1=Not at All; 2=Rarely; 3=Sometimes; 4=Often Decoded Value: Yes

IES3 character 3. I tried to remove it from my memory. 1=Not at All; 2=Rarely; 3=Sometimes; 4=Often Decoded Value: Yes

IES4 character 4. I had trouble falling asleep, because of pictures or thoughts about it that came into my mind. 1=Not at All; 2=Rarely; 3=Sometimes; 4=Often Decoded Value: Yes

IES5 character 5. I had waves of strong feelings about it. 1=Not at All; 2=Rarely; 3=Sometimes; 4=Often Decoded Value: Yes

IES6 character 6. I had dreams about it. 1=Not at All; 2=Rarely; 3=Sometimes; 4=Often Decoded Value: Yes

IES7 character 7. I stayed away from reminders. 1=Not at All; 2=Rarely; 3=Sometimes; 4=Often Decoded Value: Yes

IES8 character 8. I felt as if it hadn't happened or wasn't real. 1=Not at All; 2=Rarely; 3=Sometimes; 4=Often Decoded Value: Yes

IES9 character 9. I tried not to talk about it. 1=Not at All; 2=Rarely; 3=Sometimes; 4=Often Decoded Value: Yes

IES10 character 10. Pictures about it popped into my mind. 1=Not at All; 2=Rarely; 3=Sometimes; 4=Often Decoded Value: Yes

IES11 character 11. Other things kept making me think about it. 1=Not at All; 2=Rarely; 3=Sometimes; 4=Often Decoded Value: Yes

IES12 character 12. I was aware that I still had a lot of feelings about it, but I didn't deal with them. 1=Not at All; 2=Rarely; 3=Sometimes; 4=Often Decoded Value: Yes

IES13 character 13. I tried not to think about it. 1=Not at All; 2=Rarely; 3=Sometimes; 4=Often Decoded Value: Yes

IES14 character 14. Any reminder brought back feelings about it. 1=Not at All; 2=Rarely; 3=Sometimes; 4=Often Decoded Value: Yes

IES15 character 15. My feelings were kind of numb. 1=Not at All; 2=Rarely; 3=Sometimes; 4=Often Decoded Value: Yes

ID numeric Record ID Range value: 4456, ..., 68933

SITEID numeric Site ID Range value: 3, ..., 941

USERDATE Date User Create Date Date: YYY-MM-DD

USERDATE2 Date User Last Modify Date Date: YYY-MM-DD

DD_CRF_VERSION_LABEL character CRF Version Label , with options: v1

LANGUAGE_CODE character Language Code e=English Decoded Value: Yes

HAS_QC_ERROR character Has Quality Check Error 0=Does not have QC error or QC error has been approved; 1=Has QC error Decoded Value: Yes

update_stamp POSIXct POSIXct: YYY-MM-DD

Author

adni-data@googlegroups.com