Inclusion Criteria

Inclusion Criteria data. More information is available at https://adni.loni.usc.edu/data-samples/adni-data/

Usage

data(INCLUSIO)

Format

A data frame with 1109 observations and 39 variables.

Source

https://adni.loni.usc.edu/data-samples/adni-data/

Details

ORIGPROT character Original study protocol , with options: ADNI1

COLPROT factor Study protocol of data collection , with factor levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4

PTID character Participant ID

RID numeric Participant roster ID Range value: 1, ..., 1435

VISCODE character Visit code , with options: f, sc

VISCODE2 character , with options: f, sc

EXAMDATE Date Criteria Confirmed Date: YYY-MM-DD

INCSIGN character 1. Have the participant and study partner signed the Informed Consent form? 1=Yes; 0=No Decoded Value: Yes

INCSGDT If Yes, date signed Date: YYY-MM-DD

INCSG15T character Check the following to indicate the participant is suitable for and consents to: 1=1.5 Tesla MRI Decoded Value: Yes

INCSGPET character 1=PET Scan Decoded Value: Yes

INCSG3T character 1=3 Tesla MRI Decoded Value: Yes

INCSGLP character 1=Lumbar Puncture Decoded Value: Yes

INCMCOMP character 2. NL - Is participant free of memory complaints, verified by an informant, aside from those normal with age?MCI - Does the subject have memory complaints and memory difficulties that are verified by an informant?AD - Does the subject have memory complaints that are verified by an informant? 1=Yes;0=No Decoded Value: Yes

INCMFXN character 3. NL - Normal memory function documented by scoring at specific cutoffs on the Logical Memory II subscale (delayed Paragraph Recall) from the Wechsler Memory Scaled - Revised (the maximum score is 25)MCI/AD - Abnormal memory function documented by scoring below the educationn adjusted cutoff on the Logical Memory II subscale (Delayed Paragraph Recall) from the Wechsler Memory Scale - Revised (the maximum score is 25) 1=Yes;0=No Decoded Value: Yes

INCMMSE character 4. NL/MCI - Does the participant have Mini-Mental State Exam score between 24 and 30 (inclusive)? (Exceptions must be made for subjects with less than 8 years of education at the discretion of the project director).AD - Does the participant have an MMSE score between 20 and 26 (inclusive)? 1=Yes; 0=No Decoded Value: Yes

INCCDR character 5. NL - Does the participant have a Clinical Dementia Rating of 0? Memory Box score must be 0.MCI - Does the participant have a Clinical Dementia Rating of 0.5? Memory Box score must be at least 0.5.AD - Does participant have a Clinical Dementia rating of 0.5 or 1.0? 1=Yes; 0=No Decoded Value: Yes

INCCOG character 6. NL - Is the participant cognitively normal based on an absence of significant impairment in cognitive functions or activities of daily living?MCI - Is the participant's general cognition and functional performance sufficiently perserved such that a diagnosis of Alzheimer's disease cannot be made by the site physician at the time of the screening visit? AD - Does the participant meet NINCDS/ADRDA criteria for probable AD? 1=Yes; 0=No Decoded Value: Yes

INCHACH character 7. Does the participant have a Modified Hachinski score less than or equal to 4? 1=Yes; 0=No Decoded Value: Yes

INCAGE character 8. Is the participant between 55 and 90 years of age inclusive? 1=Yes; 0=No Decoded Value: Yes

INCSTMED character 9. Has the participant been on stable doses of non-excluded medications for at least 4 weeks prior to screening? 1=Yes; 0=No Decoded Value: Yes

INCGDS character 10. Does the participant have a Geriatric Depression Scale score of <6? 1=Yes; 0=No Decoded Value: Yes

INCINFORM character 11. Does the participant have an informant available who they have frequent contact with (e.g. an average of 10 hours per week or more), and can accompany the participant to all clinic visits and imaging sessions for the duration of the protocol? 1=Yes; 0=No Decoded Value: Yes

INCVISUAL character 12. Does the participant have adequate visual and auditory acuity to allow neuropsychological testing? 1=Yes; 0=No Decoded Value: Yes

INCHEALTH character 13. Is the participant in good general health with no additional diseases expected to interfere with the study? 1=Yes; 0=No Decoded Value: Yes

INCFEM character 14. If female, is the participant not pregnant, lactating, or of childbearing potential (i.e. women must be two years post-menopausal or surgically sterile)? 1=Yes; 0=No;2=N/A Decoded Value: Yes

INCPART character 15. NL/MCI - Is the participant willing and able to complete all Baseline assessment and participate in a 3-year protocol?AD - Is the participant willing and able to ...2-year protocol? 1=Yes;0=No Decoded Value: Yes

INCIMAG character 16. Is the participant willing to undergo MRI 1.5 Tesla neuroimaging (PET and MRI 3 Tesla are optional) and provide DNA for ApoE assessments and banking as well as plasma samples at protocol specified time points? 1=Yes; 0=No Decoded Value: Yes

INCEDUC character 17. Has the participant completed 6 grades of education (or had a good work history sufficient to exclude mental retardation)? 1=Yes; 0=No Decoded Value: Yes

INCLANG character 18. Is the participant fluent in English or Spanish? 1=Yes; 0=No Decoded Value: Yes

INCHIST character 19a. Medical History 1=Yes;0=No Decoded Value: Yes

INCPHYS character 19b. Physical Examination? 1=Yes; 0=No Decoded Value: Yes

INCNEUR character 19c. Neurological Examination? 1=Yes; 0=No Decoded Value: Yes

INCLAB character 19d. Laboratory Tests? 1=Yes; 0=No Decoded Value: Yes

ID numeric Range value: 10, ..., 2236

SITEID numeric Site ID Range value: 1, ..., 127

USERDATE Date record created Date: YYY-MM-DD

USERDATE2 Date record last updated Date: YYY-MM-DD

update_stamp POSIXct: YYY-MM-DD

Author

adni-data@googlegroups.com