MoCA

MoCA data. More information is available at https://adni.loni.usc.edu/data-samples/adni-data/

Usage

data(MOCA)

Format

A data frame with 8962 observations and 59 variables.

Source

https://adni.loni.usc.edu/data-samples/adni-data/

Details

ORIGPROT character Original study protocol , with options: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4

COLPROT factor Study protocol of data collection , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4

PTID character Participant ID

RID numeric Participant roster ID Range value: 2, ..., 10898

VISCODE character Visit code

VISCODE2 character Translated visit code

VISDATE Assessment MCDATE when present; otherwise Registry EXAMDATE Date: YYY-MM-DD

TRAILS character Trails

ADNIGO, ADNI2: 1=Correct ; 0=Incorrect

ADNI3, ADNI4: 0=Incorrect ; 1=Correct

Decoded Value: Yes

CUBE character Copy Cube

ADNIGO, ADNI2: 1=Correct ; 0=Incorrect

ADNI3, ADNI4: 0=Incorrect ; 1=Correct

Decoded Value: Yes

CLOCKCON character Draw Clock - Contour

ADNIGO, ADNI2: 1=Correct ; 0=Incorrect

ADNI3, ADNI4: 0=Incorrect ; 1=Correct

Decoded Value: Yes

CLOCKNO character Draw Clock - Number

ADNIGO, ADNI2: 1=Correct ; 0=Incorrect

ADNI3, ADNI4: 0=Incorrect ; 1=Correct

Decoded Value: Yes

CLOCKHAN character Draw Clock - Hands

ADNIGO, ADNI2: 1=Correct ; 0=Incorrect

ADNI3, ADNI4: 0=Incorrect ; 1=Correct

Decoded Value: Yes

LION character Lion

ADNIGO, ADNI2: 1=Correct ; 0=Incorrect

ADNI3, ADNI4: 0=Incorrect ; 1=Correct

Decoded Value: Yes

RHINO character Rhinoceros

ADNIGO, ADNI2: 1=Correct ; 0=Incorrect

ADNI3, ADNI4: 0=Incorrect ; 1=Correct

Decoded Value: Yes

CAMEL character Camel

ADNIGO, ADNI2: 1=Correct ; 0=Incorrect

ADNI3, ADNI4: 0=Incorrect ; 1=Correct

Decoded Value: Yes

IMMT1W1 character Immediate (#1): Face

ADNIGO, ADNI2: 1=Correct ; 0=Incorrect

ADNI3, ADNI4: 0=Incorrect ; 1=Correct

Decoded Value: Yes

IMMT1W2 character Immediate (#1): Velvet

ADNIGO, ADNI2: 1=Correct ; 0=Incorrect

ADNI3, ADNI4: 0=Incorrect ; 1=Correct

Decoded Value: Yes

IMMT1W3 character Immediate (#1): Church

ADNIGO, ADNI2: 1=Correct ; 0=Incorrect

ADNI3, ADNI4: 0=Incorrect ; 1=Correct

Decoded Value: Yes

IMMT1W4 character Immediate (#1): Daisy

ADNIGO, ADNI2: 1=Correct ; 0=Incorrect

ADNI3, ADNI4: 0=Incorrect ; 1=Correct

Decoded Value: Yes

IMMT1W5 character Immediate (#1): Red

ADNIGO, ADNI2: 1=Correct ; 0=Incorrect

ADNI3, ADNI4: 0=Incorrect ; 1=Correct

Decoded Value: Yes

IMMT2W1 character Immediate (#2): Face

ADNIGO, ADNI2: 1=Correct ; 0=Incorrect

ADNI3, ADNI4: 0=Incorrect ; 1=Correct

Decoded Value: Yes

IMMT2W2 character Immediate (#2): Velvet

ADNIGO, ADNI2: 1=Correct ; 0=Incorrect

ADNI3, ADNI4: 0=Incorrect ; 1=Correct

Decoded Value: Yes

IMMT2W3 character Immediate (#2): Church

ADNIGO, ADNI2: 1=Correct ; 0=Incorrect

ADNI3, ADNI4: 0=Incorrect ; 1=Correct

Decoded Value: Yes

IMMT2W4 character Immediate (#2): Daisy

ADNIGO, ADNI2: 1=Correct ; 0=Incorrect

ADNI3, ADNI4: 0=Incorrect ; 1=Correct

Decoded Value: Yes

IMMT2W5 character Immediate (#2): Red

ADNIGO, ADNI2: 1=Correct ; 0=Incorrect

ADNI3, ADNI4: 0=Incorrect ; 1=Correct

Decoded Value: Yes

DIGFOR character Digits Forward

ADNIGO, ADNI2: 1=Correct ; 0=Incorrect

ADNI3, ADNI4: 0=Incorrect ; 1=Correct

Decoded Value: Yes

DIGBACK character Digits Backward

ADNIGO, ADNI2: 1=Correct ; 0=Incorrect

ADNI3, ADNI4: 0=Incorrect ; 1=Correct

Decoded Value: Yes

LETTERS numeric List of Letters/Tapping: # Errors

ADNIGO, ADNI2: 0..29

ADNI3, ADNI4

SERIAL1 character Serial 7: 1st Subtraction

ADNIGO, ADNI2: 1=Correct ; 0=Incorrect

ADNI3, ADNI4: 0=Incorrect ; 1=Correct

Decoded Value: Yes

SERIAL2 character Serial 7: 2nd Subtraction

ADNIGO, ADNI2: 1=Correct ; 0=Incorrect

ADNI3, ADNI4: 0=Incorrect ; 1=Correct

Decoded Value: Yes

SERIAL3 character Serial 7: 3rd Subtraction

ADNIGO, ADNI2: 1=Correct ; 0=Incorrect

ADNI3, ADNI4: 0=Incorrect ; 1=Correct

Decoded Value: Yes

SERIAL4 character Serial 7: 4th Subtraction

ADNIGO, ADNI2: 1=Correct ; 0=Incorrect

ADNI3, ADNI4: 0=Incorrect ; 1=Correct

Decoded Value: Yes

SERIAL5 character Serial 7: 5th Subtraction

ADNIGO, ADNI2: 1=Correct ; 0=Incorrect

ADNI3, ADNI4: 0=Incorrect ; 1=Correct

Decoded Value: Yes

REPEAT1 character Repeat Sentence: I only know that John is the one to help today.

ADNIGO, ADNI2: 1=Correct ; 0=Incorrect

ADNI3, ADNI4: 0=Incorrect ; 1=Correct

Decoded Value: Yes

REPEAT2 character Repeat Sentence: The cat always hid under the couch when dogs were in the room.

ADNIGO, ADNI2: 1=Correct ; 0=Incorrect

ADNI3, ADNI4: 0=Incorrect ; 1=Correct

Decoded Value: Yes

FFLUENCY numeric Letter Fluency - F: Total number of correct words

ADNIGO, ADNI2: 0..99

ADNI3, ADNI4

ABSTRAN character Abstraction: train-bicycle

ADNIGO, ADNI2: 1=Correct ; 0=Incorrect

ADNI3, ADNI4: 0=Incorrect ; 1=Correct

Decoded Value: Yes

ABSMEAS character Abstraction: watch-ruler

ADNIGO, ADNI2: 1=Correct ; 0=Incorrect

ADNI3, ADNI4: 0=Incorrect ; 1=Correct

Decoded Value: Yes

DELW1 character Delayed: Face

ADNIGO, ADNI2: 1=Correct with No Cue ; 2=Correct with Category Cue ; 3=Correct with Mult. Choice Cue ; 0=Incorrect

ADNI3, ADNI4: 0=Incorrect ; 1=Correct with No Cue ; 2=Correct with Category Cue ; 3=Correct with Mult. Choice Cue

Decoded Value: Yes

DELW2 character Delayed: Velvet

ADNIGO, ADNI2: 1=Correct with No Cue ; 2=Correct with Category Cue ; 3=Correct with Mult. Choice Cue ; 0=Incorrect

ADNI3, ADNI4: 0=Incorrect ; 1=Correct with No Cue ; 2=Correct with Category Cue ; 3=Correct with Mult. Choice Cue

Decoded Value: Yes

DELW3 character Delayed: Church

ADNIGO, ADNI2: 1=Correct with No Cue ; 2=Correct with Category Cue ; 3=Correct with Mult. Choice Cue ; 0=Incorrect

ADNI3, ADNI4: 0=Incorrect ; 1=Correct with No Cue ; 2=Correct with Category Cue ; 3=Correct with Mult. Choice Cue

Decoded Value: Yes

DELW4 character Delayed: Daisy

ADNIGO, ADNI2: 1=Correct with No Cue ; 2=Correct with Category Cue ; 3=Correct with Mult. Choice Cue ; 0=Incorrect

ADNI3, ADNI4: 0=Incorrect ; 1=Correct with No Cue ; 2=Correct with Category Cue ; 3=Correct with Mult. Choice Cue

Decoded Value: Yes

DELW5 character Delayed: Red

ADNIGO, ADNI2: 1=Correct with No Cue ; 2=Correct with Category Cue ; 3=Correct with Mult. Choice Cue ; 0=Incorrect

ADNI3, ADNI4: 0=Incorrect ; 1=Correct with No Cue ; 2=Correct with Category Cue ; 3=Correct with Mult. Choice Cue

Decoded Value: Yes

DATE character Date

ADNIGO, ADNI2: 1=Correct ; 0=Incorrect

ADNI3, ADNI4: 0=Incorrect ; 1=Correct

Decoded Value: Yes

MONTH character Month

ADNIGO, ADNI2: 1=Correct ; 0=Incorrect

ADNI3, ADNI4: 0=Incorrect ; 1=Correct

Decoded Value: Yes

YEAR character Year

ADNIGO, ADNI2: 1=Correct ; 0=Incorrect

ADNI3, ADNI4: 0=Incorrect ; 1=Correct

Decoded Value: Yes

DAY character Day

ADNIGO, ADNI2: 1=Correct ; 0=Incorrect

ADNI3, ADNI4: 0=Incorrect ; 1=Correct

Decoded Value: Yes

PLACE character Place

ADNIGO, ADNI2: 1=Correct ; 0=Incorrect

ADNI3, ADNI4: 0=Incorrect ; 1=Correct

Decoded Value: Yes

CITY character City

ADNIGO, ADNI2: 1=Correct ; 0=Incorrect

ADNI3, ADNI4: 0=Incorrect ; 1=Correct

Decoded Value: Yes

MOCA numeric MOCA Total Score

ADNI3: 0..30

ADNI4

SOURCE character Information Source 1=Participant Visit ; 2=Telephone Call/Video Call Decoded Value: Yes

ID numeric Record ID Range value: 4, ..., 187102

SITEID numeric Site ID Range value: 1, ..., 941

USERDATE User Create Date Date: YYY-MM-DD

USERDATE2 User Last Modify Date Date: YYY-MM-DD

DD_CRF_VERSION_LABEL character CRF Version Label , with options: v1

LANGUAGE_CODE character Language Code e=English ; s=Spanish Decoded Value: Yes

HAS_QC_ERROR character Has Quality Check Error 0=Does not have QC error or QC error has been approved ; 1=Has QC error Decoded Value: Yes

update_stamp POSIXct: YYY-MM-DD