1.5T MRI Scan Information

1.5T MRI Scan Information data. More information is available at https://adni.loni.usc.edu/data-samples/adni-data/

Usage

data(MRIMETA)

Format

A data frame with 4740 observations and 36 variables.

Source

https://adni.loni.usc.edu/data-samples/adni-data/

Details

ORIGPROT character Original study protocol , with options: ADNI1, ADNIGO, ADNI2, ADNI3

COLPROT factor Study protocol of data collection , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4

PTID character Participant ID

RID numeric Participant roster ID Range value: 2, ..., 6837

VISCODE character Visit code

VISCODE2 character Translated visit code

EXAMDATE Scan Date Date: YYY-MM-DD

FIELD_STRENGTH character Field Strength

ADNI1, ADNIGO: 1.5T=1.5 Tesla

ADNI2, ADNI3: 1.5T=1.5 Tesla ; 3T=3 Tesla

ADNI4: 3T=3 Tesla

Decoded Value: Yes

RECNO character 1=Original Scan ; 2=Rescan Decoded Value: Yes

MMCONDCT character Was the scan conducted?

ADNI1, ADNIGO, ADNI2: 1=Yes ; 0=No

ADNI3, ADNI4: 0=No ; 1=Yes

Decoded Value: Yes

MMREASON character If No, please provide reason why the scan was not conducted: 1=Illness ; 2=Participant unavailable ; 3=Participant unwilling ; 4=Administrative problems ; 5=Withdrawn consent ; 6=Other (specify) Decoded Value: Yes

MMSCOUT character 1. Tri-Planar Scout (if available, otherwise use an axial scout) 1=Yes ; 0=No Decoded Value: Yes

MMSMPRAGE character 2. Straight Sagittal MPRAGE Sequence 1=Yes ; 0=No Decoded Value: Yes

MMRMPRAGE character 3. Repeat Straight Sagittal MPRAGE Sequence 1=Yes ; 0=No Decoded Value: Yes

MMB1HEAD character 4. B1 Calibration Head Coil Scan (Only applicable for phased array head coil on GE and Siemens systems) 1=Yes ; 0=No Decoded Value: Yes

MMB1BODY character 5. B1 Calibration Body Coil Scan (Only applicable for phased array head coil on GE and Siemens systems) 1=Yes ; 0=No Decoded Value: Yes

MMECHO character 6. Straight Axial Fast or Turbo Spin Echo 1=Yes ; 0=No Decoded Value: Yes

MMPHAN character 7. In new exam; Perform ADNI QC Scan. Localizer and Straight Sagittal MPRAGE (with increased slice thickness to ensure phantom coverage) 1=Yes ; 0=No Decoded Value: Yes

MMTRNSFR character Was data transferred to LONI within 24 hours of scan?

ADNI1, ADNIGO, ADNI2: 1=Yes ; 0=No

ADNI3, ADNI4: 0=No ; 1=Yes

Decoded Value: Yes

MMTRNDATE character Transfer Date

MMTRNDATEYR_DRVD numeric Transfer Year (Derived from MMTRNDATE) Range value: 2005, ..., 2023

MMARCHIVE character Data Archived Locally

ADNI1, ADNIGO, ADNI2: 1=Yes ; 0=No

ADNI3, ADNI4: 0=No ; 1=Yes

Decoded Value: Yes

MMLPDONE character Was a Lumbar Puncture completed prior to the MRI scan?

ADNI1, ADNIGO, ADNI2: 1=Yes ; 0=No

ADNI3, ADNI4: 0=No ; 1=Yes

Decoded Value: Yes

MMLPINTER character If Yes, What was the interval between LP and MRI? 1=less than 6 hours ; 2=6-12 hours ; 3=13-24 hours ; 4=25-48 hours ; 5=49-72 hours ; 6=more than 72 hours Decoded Value: Yes

SEDATIVE character Was sedation medication consumed for use during MRI scan?

ADNI2: 1=Yes ; 0=No

ADNI3, ADNI4: 0=No ; 1=Yes

Decoded Value: Yes

OTRDEV character Other Protocol Variations

ADNI2: 1=Yes ; 0=No

ADNI3, ADNI4: 0=No ; 1=Yes

Decoded Value: Yes

MOTION character Patient Motion Problems:

ADNI2: 1=Yes ; 0=No

ADNI3, ADNI4: 0=No ; 1=Yes

Decoded Value: Yes

EYEKEEP character Did the participant keep their eyes open (MRI Tech: ask the subject right after the scan)

ADNI2: 1=Yes ; 0=No ; 2=N/A Scanner manufacturer is GE

ADNI3: 0=No ; 1=Yes ; 2=N/A Scanner manufacturer is GE

ADNI4: 0=No ; 1=Yes

Decoded Value: Yes

EYEASK character Was the participant instructed to open their eyes?

ADNI2: 1=Yes ; 0=No ; 2=N/A Scanner manufacturer is GE

ADNI3: 0=No ; 1=Yes ; 2=N/A Scanner manufacturer is GE

ADNI4: 0=No ; 1=Yes

Decoded Value: Yes

MALFUNC character Scanner Malfunction:

ADNI2: 1=Yes ; 0=No

ADNI3, ADNI4: 0=No ; 1=Yes

Decoded Value: Yes

FLDSTRNGTH character Field Strength 1=1.5 Tesla ; 2=3 Tesla Decoded Value: Yes

ID numeric Record ID Range value: 6, ..., 176409

SITEID numeric Site ID Range value: 1, ..., 127

USERDATE User Create Date Date: YYY-MM-DD

USERDATE2 User Last Modify Date Date: YYY-MM-DD

UPDATE_STAMP POSIXct: YYY-MM-DD

Author

adni-data@googlegroups.com