MAYO ADIR LAB MRI Quantitative Susceptibility Mapping

MAYO ADIR LAB MRI Quantitative Susceptibility Mapping data. More information is available at https://adni.loni.usc.edu/data-samples/adni-data/

Usage

data(MRIQSM)

Format

A data frame with 31058 observations and 39 variables.

Source

https://adni.loni.usc.edu/data-samples/adni-data/

Details

ORIGPROT character Original study protocol , with options: ADNI4, ADNI2, ADNI3, ADNIGO, ADNI1

ParticipantID character Participant ID

RID numeric

VISCODE character , with options: 4_sc, 4_m12, 4_init, 4_m24

VISCODE2 character

StudyDate Study Date YYYYMMDD

SeriesTime Time Series Started on the Scanner HHMMSS ;

SeriesNumber numeric DICOM Series Number Range value: 3, ..., 27

SeriesType character Series Type (ASL) Arterial Spin Labeling [raw data for perfusion] ; (BOLD) Blood Oxygenation Level Dependent [i.e. most functional MRI, either task-based or task-free] ; (ContrastAgent) Scans acquired after the injection of a contrast agent, typically gadolinium-based ; (FLAIR) Fluid suppressed scans used to highlight WMH ; (MRA) Angiography without a contrast agent ; (MRE) Magnetic Resonance Elastography [raw data, not derived products such as stiffness] ; (MRS) Magnetic Resonance Spectra [spectroscopy] ; (T1w) T1-weighted, usually a structural scan for anatomical volumetry and segmentation ; (T2starw) T2*-weighted [susceptibility] raw data for iron/microhemorrhages. [SWI/SWAN are labeled 'derived'] ; (MET2starw) Multiecho T2*-weighted (susceptibility) raw data. This multiecho version can be used for QSM (iron/microhemorrhages and calcium). [SWI/SWAN are labeled 'derived'] ; (T2w) T2-weighted whole head scan for volumetry and segmentation ; (T2w_HighResHippo) T2-weighted coronal high resolution hippocampus scan for hippocampal volumetry ; (calibration) Scans acquired to assist other scans ; (dMRI) Diffusion-weighted raw data [often improperly called 'DTI'] ; (derived) Anything calculated, reformatted, or annotated on-scanner and not proportional to magnetization ; (flow) Scans with velocity encoding ; (localizer) Scans used to position the field of view of subsequent target scans ; (PDT2)A two echo scan with Proton Density weighting for the first volume and T2 weighting the second ; (undetermined) The catch-all category for rare difficult to classify series or none of the above

SiteID character Site ID

PostProcessQC numeric Post Processing Series Quality Assessment (-1) Not assessed ; (1) Pass ; (4) Fail

Atlas character ADIR Lab Name for the Atlas Used for Analysis Example: ADIR122 refers to the 122 ROI atlas

Sharpening character Sharpening; tissue types included WM ; GM ; GMWM ;

RegionNumber numeric Region Number Range value: 2, ..., 49

RegionName character Region Name

RegionSide character Region Side (L) Left ; (R) Right ;

Median numeric Median of the voxel values in the given image in the given ROI Range value: -0.0375901, ..., 0.172673

TotalVoxels numeric Total number of Voxels in the given image in the given ROI Range value: 757, ..., 113537

VoxelsIgnoredZero numeric Number of Voxels Ignored due to having a Value of Zero Range value: 0, ..., 18599

VoxelsIgnoredMask numeric Number of Voxels Ignored due to being Outside Chosen Sharpening Mask Range value: 0, ..., 89504

Sum numeric Sum of the voxel values in the given image in the given ROI Range value: -1066.75, ..., 1113.09

VoxelVolume numeric Volume of a single voxel in the given image Range value: 0.999, ..., 1.001

ScannerManufacturer character DICOM Scanner Manufacturer Unknown ; GE ; Philips ; Siemens ; Canon ; UnitedImaging ; Other

ScannerModel character DICOM Scanner Model , with options: MAGNETOM Cima.X, Skyra, MAGNETOM Prisma Fit, MAGNETOM Prisma, Prisma_fit, MAGNETOM Skyra fit, MAGNETOM Vida, Prisma, MAGNETOM Skyra

SoftwareVersion character DICOM Software Version; may be extracted from vendor private tags , with options: syngo MR XA61, syngo MR E11, syngo MR XA60, syngo MR XA30

MagneticFieldStrength numeric Scanner Field Strength ; 1.5 ; 3 ; 7

Acceleration character Acceleration Unaccelerated ; Accelerated ; Ultrafast ; Undetermined

AcquisitionType numeric Acquisition Type 2D ; 3D ; Unknown

AcquisitionPlane character Acquisition Plane Axial ; Sagittal ; Coronal ; NA ; Unknown

NumberVolumes numeric Number of Volumes Range value: 5, ..., 5

SlicesPerVolume numeric Slices per Volume Range value: 80, ..., 104

SliceThickness numeric Slice Thickness Range value: 1.799999, ..., 1.8

ReceiveCoilName character Receive Coil Name , with options: Head_32, NOT SET, HeadNeck_64, HeadNeck_20_TCS, HeadNeck_20

StudyInstanceUID character DICOM StudyInstanceUID

SeriesInstanceUID character DICOM SeriesInstanceUID

LONIStudy numeric LONI Study ID Range value: 242527, ..., 396422

LONISeries numeric LONI Series ID Range value: 5162614, ..., 6273354

LONIImage numeric LONI Image ID Range value: 10248902, ..., 11680165

update_stamp POSIXct: YYYY-MM-DD HH:MM:SS