MRI Clinical Read

MRI Clinical Read data. More information is available at https://adni.loni.usc.edu/data-samples/adni-data/

Usage

data(MRIREAD)

Format

A data frame with 4106 observations and 16 variables.

Source

https://adni.loni.usc.edu/data-samples/adni-data/

Details

ORIGPROT character Original study protocol , with options: ADNI1

ID numeric Range value: 6, ..., 8254

RID numeric Participant roster ID Range value: 2, ..., 1435

SITEID numeric Site ID Range value: 1, ..., 127

VISCODE character Visit code , with options: sc, f, bl, m06, m12, uns1, m18, m24, m36, m48

USERDATE Date record created Date: YYY-MM-DD

USERDATE2 Date record last updated Date: YYY-MM-DD

EXAMDATE Examination Date Date: YYY-MM-DD

MRREAD numeric Was the 1.5T MRI done? 1=Yes ; 0=No

MRUPLD character Upload 1.5T MRI Clinical Read:

MRCOMPAT numeric Is the MRI compatible with the Inclusion/Exclusion Criteria? 1=Yes ; 0=No

MRCOMSPE character If No, explain:

MR3TREAD numeric Was a clinical read of the 3T MRI done? 1=Yes ; 0=No

MR3TUPLD character Upload 3T MRI Clinical Read :

MRSUITABL numeric Is the participant clinically suitable to remain in the study? 1=Yes ; 0=No

update_stamp POSIXct: YYY-MM-DD

Author

adni-data@googlegroups.com