UCSF MedAI - MASS and MUSE volumetrics - ADNI 1-3

UCSF MedAI - MASS and MUSE volumetrics - ADNI 1-3 data. More information is available at https://adni.loni.usc.edu/data-samples/adni-data/

Usage

data(MUSE_volume_ADNI123)

Format

A data frame with 10329 observations and 273 variables.

Source

https://adni.loni.usc.edu/data-samples/adni-data/

Details

PTID character Subject ID

EXAMDATE Date of scan Date: YYYY-MM-DD

LONI_UID numeric Image unique identifer Range value: 31863, ..., 1639219

LONI_SERIES numeric LONI Series ID Range value: 8723, ..., 1173525

LONI_STUDY numeric LONI Study ID Range value: 1945, ..., 218248

Code character Code

ProjectName character ProjectName , with options: ADNI1, ADNI2, ADNI3

3rd_Ventricle numeric 3rd Ventricle Range value: 108, ..., 5785.2

4th_Ventricle numeric 4th Ventricle Range value: 163, ..., 7720

anterior_limb_of_internal_capsule_left numeric anterior limb of internal capsule left Range value: 405.88, ..., 7186

anterior_limb_of_internal_capsule_right numeric anterior limb of internal capsule right Range value: 114, ..., 7664

Background numeric Background Range value: 9439630, ..., 17402100

BASAL_GANGLIA numeric BASAL GANGLIA Range value: 1654, ..., 45564

BASAL_GANGLIA_L numeric BASAL GANGLIA L Range value: 651, ..., 22003

BASAL_GANGLIA_R numeric BASAL GANGLIA R Range value: 369.54, ..., 23561

Brain_Stem numeric Brain Stem Range value: 1000, ..., 31144.9

Cerebellar_Vermal_Lobules_I_V numeric Cerebellar Vermal Lobules I-V Range value: 536, ..., 7639.32

Cerebellar_Vermal_Lobules_VI_VII numeric Cerebellar Vermal Lobules VI-VII Range value: 227, ..., 7059

Cerebellar_Vermal_Lobules_VIII_X numeric Cerebellar Vermal Lobules VIII-X Range value: 6, ..., 4551.47

CEREBELLUM numeric CEREBELLUM Range value: 22220, ..., 239540

CEREBELLUM_L numeric CEREBELLUM L Range value: 7956, ..., 106618

CEREBELLUM_R numeric CEREBELLUM R Range value: 3671.26, ..., 120662

corpus_callosum numeric corpus callosum Range value: 236, ..., 19811.1

CSF numeric CSF Range value: 271, ..., 6185

DEEP_GM numeric DEEP GM Range value: 2032, ..., 39022

DEEP_GM_L numeric DEEP GM L Range value: 1504.02, ..., 20942

DEEP_GM_R numeric DEEP GM R Range value: 490, ..., 18080

DEEP_WM numeric DEEP WM Range value: 1721, ..., 31408

DEEP_WM_GM numeric DEEP WM GM Range value: 9286.79, ..., 133864

DEEP_WM_GM_L numeric DEEP WM GM L Range value: 4749.15, ..., 65167

DEEP_WM_GM_R numeric DEEP WM GM R Range value: 1346.32, ..., 68697

DEEP_WM_L numeric DEEP WM L Range value: 836.41, ..., 16303

DEEP_WM_R numeric DEEP WM R Range value: 248, ..., 15105

fornix_left numeric fornix left Range value: 0, ..., 2373

fornix_right numeric fornix right Range value: 13, ..., 1956

FRONTAL numeric FRONTAL Range value: 57287, ..., 525612

FRONTAL_GM numeric FRONTAL GM Range value: 16065, ..., 248065

FRONTAL_GM_L numeric FRONTAL GM L Range value: 8499, ..., 125914

FRONTAL_GM_R numeric FRONTAL GM R Range value: 4995.95, ..., 127154

FRONTAL_INFERIOR_GM numeric FRONTAL INFERIOR GM Range value: 1098, ..., 46671.5

FRONTAL_INFERIOR_GM_L numeric FRONTAL INFERIOR GM L Range value: 179, ..., 23214.5

FRONTAL_INFERIOR_GM_R numeric FRONTAL INFERIOR GM R Range value: 315.69, ..., 23457

FRONTAL_INSULAR_GM numeric FRONTAL INSULAR GM Range value: 537, ..., 17110.8

FRONTAL_INSULAR_GM_L numeric FRONTAL INSULAR GM L Range value: 290, ..., 8568

FRONTAL_INSULAR_GM_R numeric FRONTAL INSULAR GM R Range value: 0, ..., 8925

FRONTAL_L numeric FRONTAL L Range value: 23846.3, ..., 260340

FRONTAL_LATERAL_GM numeric FRONTAL LATERAL GM Range value: 6414, ..., 141184

FRONTAL_LATERAL_GM_L numeric FRONTAL LATERAL GM L Range value: 3082, ..., 72064.5

FRONTAL_LATERAL_GM_R numeric FRONTAL LATERAL GM R Range value: 1548.73, ..., 70603.2

frontal_lobe_WM_left numeric frontal lobe WM left Range value: 11237.5, ..., 163166

frontal_lobe_WM_right numeric frontal lobe WM right Range value: 7149.4, ..., 172397

FRONTAL_MEDIAL_GM numeric FRONTAL MEDIAL GM Range value: 5840.09, ..., 50418.9

FRONTAL_MEDIAL_GM_L numeric FRONTAL MEDIAL GM L Range value: 928, ..., 25705.2

FRONTAL_MEDIAL_GM_R numeric FRONTAL MEDIAL GM R Range value: 1994, ..., 26333.4

FRONTAL_OPERCULAR_GM numeric FRONTAL OPERCULAR GM Range value: 637.53, ..., 28127

FRONTAL_OPERCULAR_GM_L numeric FRONTAL OPERCULAR GM L Range value: 109.06, ..., 11368.8

FRONTAL_OPERCULAR_GM_R numeric FRONTAL OPERCULAR GM R Range value: 0, ..., 18098

FRONTAL_R numeric FRONTAL R Range value: 12319.2, ..., 275375

FRONTAL_WM numeric FRONTAL WM Range value: 27076.2, ..., 335563

FRONTAL_WM_L numeric FRONTAL WM L Range value: 11237.5, ..., 163166

FRONTAL_WM_R numeric FRONTAL WM R Range value: 7149.4, ..., 172397

GM numeric GM Range value: 85863, ..., 834741

GM_L numeric GM L Range value: 39335.8, ..., 411209

GM_R numeric GM R Range value: 13582, ..., 429691

ICV numeric ICV Range value: 0, ..., 5283920

Left_Accumbens_Area numeric Left Accumbens Area Range value: 0, ..., 1264

Left_ACgGanterior_cingulate_gyrus numeric Left ACgG anterior cingulate gyrus Range value: 1, ..., 10569

Left_AInsanterior_insula numeric Left AIns anterior insula Range value: 232, ..., 6022.8

Left_Amygdala numeric Left Amygdala Range value: 0, ..., 1602

Left_AnG_angular_gyrus numeric Left AnG angular gyrus Range value: 291, ..., 23506

Left_AOrGanterior_orbital_gyrus numeric Left AOrG anterior orbital gyrus Range value: 0, ..., 4001.27

Left_Basal_Forebrain numeric Left Basal Forebrain Range value: 0, ..., 921.89

Left_Calccalcarine_cortex numeric Left Calc calcarine cortex Range value: 81, ..., 13316

Left_Caudate numeric Left Caudate Range value: 21, ..., 5859.6

Left_Cerebellum_Exterior numeric Left Cerebellum Exterior Range value: 4082, ..., 90466

Left_Cerebellum_White_Matter numeric Left Cerebellum White Matter Range value: 2260.12, ..., 26279.5

Left_Cerebral_Exterior numeric Left Cerebral Exterior Range value: 0, ..., 37

Left_COcentral_operculum numeric Left CO central operculum Range value: 24.44, ..., 5952

Left_Cun_cuneus numeric Left Cun cuneus Range value: 198, ..., 7998

Left_Ent_entorhinal_area numeric Left Ent entorhinal area Range value: 0, ..., 3394

Left_FOfrontal_operculum numeric Left FO frontal operculum Range value: 73.33, ..., 2907

Left_FRP_frontal_pole numeric Left FRP frontal pole Range value: 0, ..., 5932.98

Left_FuG_fusiform_gyrus numeric Left FuG fusiform gyrus Range value: 227, ..., 18872

Left_GRe_gyrus_rectus numeric Left GRe gyrus rectus Range value: 0, ..., 3229.24

Left_Hippocampus numeric Left Hippocampus Range value: 63, ..., 5385

Left_Inf_Lat_Vent numeric Left Inf Lat Vent Range value: 0, ..., 5986.29

Left_IOG_inferior_occipital_gyrus numeric Left IOG inferior occipital gyrus Range value: 35.58, ..., 8464.98

Left_ITG_inferior_temporal_gyrus numeric Left ITG inferior temporal gyrus Range value: 61.79, ..., 27773

Left_Lateral_Ventricle numeric Left Lateral Ventricle Range value: 3401, ..., 95143.2

Left_LiG_lingual_gyrus numeric Left LiG lingual gyrus Range value: 866, ..., 13046.8

Left_LOrGlateral_orbital_gyrus numeric Left LOrG lateral orbital gyrus Range value: 2, ..., 7315.09

Left_MCgGmiddle_cingulate_gyrus numeric Left MCgG middle cingulate gyrus Range value: 123, ..., 9211

Left_MFC_medial_frontal_cortex numeric Left MFC medial frontal cortex Range value: 0, ..., 8687

Left_MFG_middle_frontal_gyrus numeric Left MFG middle frontal gyrus Range value: 725, ..., 25513.4

Left_MOG_middle_occipital_gyrus numeric Left MOG middle occipital gyrus Range value: 116, ..., 10354.5

Left_MOrGmedial_orbital_gyrus numeric Left MOrG medial orbital gyrus Range value: 1, ..., 7649.92

Left_MPoGpostcentral_gyrus_medial_segment numeric Left MPoG postcentral gyrus medial segment Range value: 0, ..., 5646

Left_MPrGprecentral_gyrus_medial_segment numeric Left MPrG precentral gyrus medial segment Range value: 65, ..., 4581.6

Left_MSFGsuperior_frontal_gyrus_medial_segment numeric Left MSFG superior frontal gyrus medial segment Range value: 146, ..., 9350.44

Left_MTG_middle_temporal_gyrus numeric Left MTG middle temporal gyrus Range value: 0, ..., 20069.6

Left_OCP_occipital_pole numeric Left OCP occipital pole Range value: 2, ..., 5852

Left_OFuGoccipital_fusiform_gyrus numeric Left OFuG occipital fusiform gyrus Range value: 28, ..., 6837

Left_OpIFG_opercular_part_of_the_inferior_frontal_gyrus numeric Left OpIFG opercular part of the inferior frontal gyrus Range value: 0, ..., 5133

Left_OrIFG_orbital_part_of_the_inferior_frontal_gyrus numeric Left OrIFG orbital part of the inferior frontal gyrus Range value: 0, ..., 2668.18

Left_Pallidum numeric Left Pallidum Range value: 164, ..., 3664

Left_PCgGposterior_cingulate_gyrus numeric Left PCgG posterior cingulate gyrus Range value: 135, ..., 7769

Left_PCu_precuneus numeric Left PCu precuneus Range value: 193, ..., 17517

Left_PHG_parahippocampal_gyrus numeric Left PHG parahippocampal gyrus Range value: 21, ..., 8426

Left_PInsposterior_insula numeric Left PIns posterior insula Range value: 13, ..., 3178

Left_PoG_postcentral_gyrus numeric Left PoG postcentral gyrus Range value: 92, ..., 15704.3

Left_POparietal_operculum numeric Left PO parietal operculum Range value: 11.28, ..., 5674

Left_POrGposterior_orbital_gyrus numeric Left POrG posterior orbital gyrus Range value: 133, ..., 4952.64

Left_PPplanum_polare numeric Left PP planum polare Range value: 3.76, ..., 2705.89

Left_PrG_precentral_gyrus numeric Left PrG precentral gyrus Range value: 250, ..., 17724

Left_PTplanum_temporale numeric Left PT planum temporale Range value: 0, ..., 3166

Left_Putamen numeric Left Putamen Range value: 236, ..., 15948

Left_SCA_subcallosal_area numeric Left SCA subcallosal area Range value: 0, ..., 4091

Left_SFG_superior_frontal_gyrus numeric Left SFG superior frontal gyrus Range value: 121, ..., 18606

Left_SMC_supplementary_motor_cortex numeric Left SMC supplementary motor cortex Range value: 131, ..., 14542

Left_SMG_supramarginal_gyrus numeric Left SMG supramarginal gyrus Range value: 0, ..., 12777.1

Left_SOG_superior_occipital_gyrus numeric Left SOG superior occipital gyrus Range value: 29, ..., 4859

Left_SPL_superior_parietal_lobule numeric Left SPL superior parietal lobule Range value: 35, ..., 12336.8

Left_STG_superior_temporal_gyrus numeric Left STG superior temporal gyrus Range value: 0, ..., 11004

Left_Thalamus_Proper numeric Left Thalamus Proper Range value: 1504.02, ..., 20942

Left_TMP_temporal_pole numeric Left TMP temporal pole Range value: 4, ..., 15799

Left_TrIFG_triangular_part_of_the_inferior_frontal_gyrus numeric Left TrIFG triangular part of the inferior frontal gyrus Range value: 5.69, ..., 5700.61

Left_TTG_transverse_temporal_gyrus numeric Left TTG transverse temporal gyrus Range value: 3, ..., 2676.59

Left_Ventral_DC numeric Left Ventral DC Range value: 1128.18, ..., 7566

Left_vessel numeric Left vessel Range value: 0, ..., 203

LIMBIC numeric LIMBIC Range value: 2363, ..., 52682.4

LIMBIC_CINGULATE_GM numeric LIMBIC CINGULATE GM Range value: 1891, ..., 36253.2

LIMBIC_CINGULATE_GM_L numeric LIMBIC CINGULATE GM L Range value: 773, ..., 20667

LIMBIC_CINGULATE_GM_R numeric LIMBIC CINGULATE GM R Range value: 819, ..., 17715.6

LIMBIC_GM numeric LIMBIC GM Range value: 2363, ..., 52682.4

LIMBIC_GM_L numeric LIMBIC GM L Range value: 951, ..., 28549.2

LIMBIC_GM_R numeric LIMBIC GM R Range value: 1268, ..., 24768.2

LIMBIC_L numeric LIMBIC L Range value: 951, ..., 28549.2

LIMBIC_MEDIALTEMPORAL_GM numeric LIMBIC MEDIALTEMPORAL GM Range value: 394, ..., 16707

LIMBIC_MEDIALTEMPORAL_GM_L numeric LIMBIC MEDIALTEMPORAL GM L Range value: 178, ..., 9265

LIMBIC_MEDIALTEMPORAL_GM_R numeric LIMBIC MEDIALTEMPORAL GM R Range value: 25, ..., 8493.6

LIMBIC_R numeric LIMBIC R Range value: 1268, ..., 24768.2

OCCIPITAL numeric OCCIPITAL Range value: 17409, ..., 171976

OCCIPITAL_GM numeric OCCIPITAL GM Range value: 6559, ..., 103317

OCCIPITAL_GM_L numeric OCCIPITAL GM L Range value: 2791, ..., 51845.6

OCCIPITAL_GM_R numeric OCCIPITAL GM R Range value: 1138.33, ..., 58178

OCCIPITAL_INFERIOR_GM numeric OCCIPITAL INFERIOR GM Range value: 123, ..., 13112

OCCIPITAL_INFERIOR_GM_L numeric OCCIPITAL INFERIOR GM L Range value: 28, ..., 6837

OCCIPITAL_INFERIOR_GM_R numeric OCCIPITAL INFERIOR GM R Range value: 0, ..., 7173

OCCIPITAL_L numeric OCCIPITAL L Range value: 6257.62, ..., 87574

OCCIPITAL_LATERAL_GM numeric OCCIPITAL LATERAL GM Range value: 1160, ..., 50919.4

OCCIPITAL_LATERAL_GM_L numeric OCCIPITAL LATERAL GM L Range value: 968, ..., 25734.6

OCCIPITAL_LATERAL_GM_R numeric OCCIPITAL LATERAL GM R Range value: 76, ..., 25184.8

occipital_lobe_WM_left numeric occipital lobe WM left Range value: 1842, ..., 58205

occipital_lobe_WM_right numeric occipital lobe WM right Range value: 473.53, ..., 43135

OCCIPITAL_MEDIAL_GM numeric OCCIPITAL MEDIAL GM Range value: 2593, ..., 49534

OCCIPITAL_MEDIAL_GM_L numeric OCCIPITAL MEDIAL GM L Range value: 1171, ..., 25207.2

OCCIPITAL_MEDIAL_GM_R numeric OCCIPITAL MEDIAL GM R Range value: 695, ..., 29462

OCCIPITAL_R numeric OCCIPITAL R Range value: 1611.86, ..., 87457

OCCIPITAL_WM numeric OCCIPITAL WM Range value: 6818.61, ..., 89200

OCCIPITAL_WM_L numeric OCCIPITAL WM L Range value: 1842, ..., 58205

OCCIPITAL_WM_R numeric OCCIPITAL WM R Range value: 473.53, ..., 43135

Optic_Chiasm numeric Optic Chiasm Range value: 0, ..., 112.13

PARIETAL numeric PARIETAL Range value: 24323.5, ..., 253784

PARIETAL_GM numeric PARIETAL GM Range value: 8780, ..., 123282

PARIETAL_GM_L numeric PARIETAL GM L Range value: 2269, ..., 61649.6

PARIETAL_GM_R numeric PARIETAL GM R Range value: 1723.28, ..., 66063.9

PARIETAL_L numeric PARIETAL L Range value: 9230.37, ..., 132353

PARIETAL_LATERAL_GM numeric PARIETAL LATERAL GM Range value: 7344, ..., 95034.5

PARIETAL_LATERAL_GM_L numeric PARIETAL LATERAL GM L Range value: 1966.79, ..., 48526.5

PARIETAL_LATERAL_GM_R numeric PARIETAL LATERAL GM R Range value: 350.97, ..., 51453.7

parietal_lobe_WM_left numeric parietal lobe WM left Range value: 4881.25, ..., 80552.4

parietal_lobe_WM_right numeric parietal lobe WM right Range value: 1442.88, ..., 70180.8

PARIETAL_MEDIAL_GM numeric PARIETAL MEDIAL GM Range value: 644, ..., 33919.1

PARIETAL_MEDIAL_GM_L numeric PARIETAL MEDIAL GM L Range value: 208, ..., 18100

PARIETAL_MEDIAL_GM_R numeric PARIETAL MEDIAL GM R Range value: 436, ..., 17186.6

PARIETAL_R numeric PARIETAL R Range value: 3166.16, ..., 125142

PARIETAL_WM numeric PARIETAL WM Range value: 12208.8, ..., 150394

PARIETAL_WM_L numeric PARIETAL WM L Range value: 4881.25, ..., 80552.4

PARIETAL_WM_R numeric PARIETAL WM R Range value: 1442.88, ..., 70180.8

posterior_limb_of_internal_capsule_inc_cerebral_peduncle_left numeric posterior limb of internal capsule inc. cerebral peduncle left Range value: 190, ..., 6744

posterior_limb_of_internal_capsule_inc_cerebral_peduncle_right numeric posterior limb of internal capsule inc. cerebral peduncle right Range value: 2, ..., 7428

Right_Accumbens_Area numeric Right Accumbens Area Range value: 2, ..., 769.2

Right_ACgGanterior_cingulate_gyrus numeric Right ACgG anterior cingulate gyrus Range value: 238, ..., 7106.4

Right_AInsanterior_insula numeric Right AIns anterior insula Range value: 0, ..., 5893.2

Right_Amygdala numeric Right Amygdala Range value: 0, ..., 1758

Right_AnG_angular_gyrus numeric Right AnG angular gyrus Range value: 0, ..., 15698.2

Right_AOrGanterior_orbital_gyrus numeric Right AOrG anterior orbital gyrus Range value: 1, ..., 5168.17

Right_Basal_Forebrain numeric Right Basal Forebrain Range value: 0, ..., 1850

Right_Calccalcarine_cortex numeric Right Calc calcarine cortex Range value: 176.41, ..., 6059.14

Right_Caudate numeric Right Caudate Range value: 44, ..., 6518.51

Right_Cerebellum_Exterior numeric Right Cerebellum Exterior Range value: 2603.49, ..., 102058

Right_Cerebellum_White_Matter numeric Right Cerebellum White Matter Range value: 1067.77, ..., 24999.4

Right_Cerebral_Exterior numeric Right Cerebral Exterior Range value: 0, ..., 133

Right_COcentral_operculum numeric Right CO central operculum Range value: 0, ..., 14559

Right_Cun_cuneus numeric Right Cun cuneus Range value: 122, ..., 15861

Right_Ent_entorhinal_area numeric Right Ent entorhinal area Range value: 0, ..., 3397.58

Right_FOfrontal_operculum numeric Right FO frontal operculum Range value: 0, ..., 2672.4

Right_FRP_frontal_pole numeric Right FRP frontal pole Range value: 2, ..., 5212

Right_FuG_fusiform_gyrus numeric Right FuG fusiform gyrus Range value: 0, ..., 12541

Right_GRe_gyrus_rectus numeric Right GRe gyrus rectus Range value: 0, ..., 3131.46

Right_Hippocampus numeric Right Hippocampus Range value: 17, ..., 8343

Right_Inf_Lat_Vent numeric Right Inf Lat Vent Range value: 0, ..., 6259.54

Right_IOG_inferior_occipital_gyrus numeric Right IOG inferior occipital gyrus Range value: 0, ..., 9043.2

Right_ITG_inferior_temporal_gyrus numeric Right ITG inferior temporal gyrus Range value: 0, ..., 16604

Right_Lateral_Ventricle numeric Right Lateral Ventricle Range value: 2682, ..., 78938.4

Right_LiG_lingual_gyrus numeric Right LiG lingual gyrus Range value: 391, ..., 24919

Right_LOrGlateral_orbital_gyrus numeric Right LOrG lateral orbital gyrus Range value: 0, ..., 4709

Right_MCgGmiddle_cingulate_gyrus numeric Right MCgG middle cingulate gyrus Range value: 142, ..., 7207

Right_MFC_medial_frontal_cortex numeric Right MFC medial frontal cortex Range value: 0, ..., 4058.52

Right_MFG_middle_frontal_gyrus numeric Right MFG middle frontal gyrus Range value: 0, ..., 25613.6

Right_MOG_middle_occipital_gyrus numeric Right MOG middle occipital gyrus Range value: 0, ..., 9344

Right_MOrGmedial_orbital_gyrus numeric Right MOrG medial orbital gyrus Range value: 71, ..., 10262.9

Right_MPoGpostcentral_gyrus_medial_segment numeric Right MPoG postcentral gyrus medial segment Range value: 0, ..., 4272

Right_MPrGprecentral_gyrus_medial_segment numeric Right MPrG precentral gyrus medial segment Range value: 41, ..., 7021

Right_MSFGsuperior_frontal_gyrus_medial_segment numeric Right MSFG superior frontal gyrus medial segment Range value: 670.15, ..., 11980

Right_MTG_middle_temporal_gyrus numeric Right MTG middle temporal gyrus Range value: 0, ..., 20691

Right_OCP_occipital_pole numeric Right OCP occipital pole Range value: 0, ..., 5580.19

Right_OFuGoccipital_fusiform_gyrus numeric Right OFuG occipital fusiform gyrus Range value: 0, ..., 7173

Right_OpIFG_opercular_part_of_the_inferior_frontal_gyrus numeric Right OpIFG opercular part of the inferior frontal gyrus Range value: 0, ..., 5496

Right_OrIFG_orbital_part_of_the_inferior_frontal_gyrus numeric Right OrIFG orbital part of the inferior frontal gyrus Range value: 0, ..., 4710

Right_Pallidum numeric Right Pallidum Range value: 0, ..., 3096

Right_PCgGposterior_cingulate_gyrus numeric Right PCgG posterior cingulate gyrus Range value: 64, ..., 6173

Right_PCu_precuneus numeric Right PCu precuneus Range value: 367, ..., 16099.2

Right_PHG_parahippocampal_gyrus numeric Right PHG parahippocampal gyrus Range value: 0, ..., 5787

Right_PInsposterior_insula numeric Right PIns posterior insula Range value: 0, ..., 3529.2

Right_PoG_postcentral_gyrus numeric Right PoG postcentral gyrus Range value: 90.99, ..., 15855.3

Right_POparietal_operculum numeric Right PO parietal operculum Range value: 0, ..., 3312

Right_POrGposterior_orbital_gyrus numeric Right POrG posterior orbital gyrus Range value: 9, ..., 4669.2

Right_PPplanum_polare numeric Right PP planum polare Range value: 0, ..., 6507

Right_PrG_precentral_gyrus numeric Right PrG precentral gyrus Range value: 239.55, ..., 19960

Right_PTplanum_temporale numeric Right PT planum temporale Range value: 0, ..., 6100

Right_Putamen numeric Right Putamen Range value: 38, ..., 17285

Right_SCA_subcallosal_area numeric Right SCA subcallosal area Range value: 0, ..., 2408.58

Right_SFG_superior_frontal_gyrus numeric Right SFG superior frontal gyrus Range value: 42, ..., 18733.9

Right_SMC_supplementary_motor_cortex numeric Right SMC supplementary motor cortex Range value: 86, ..., 7913.2

Right_SMG_supramarginal_gyrus numeric Right SMG supramarginal gyrus Range value: 0, ..., 12789.5

Right_SOG_superior_occipital_gyrus numeric Right SOG superior occipital gyrus Range value: 0, ..., 5811.15

Right_SPL_superior_parietal_lobule numeric Right SPL superior parietal lobule Range value: 259.98, ..., 23139

Right_STG_superior_temporal_gyrus numeric Right STG superior temporal gyrus Range value: 0, ..., 9695

Right_Thalamus_Proper numeric Right Thalamus Proper Range value: 490, ..., 18080

Right_TMP_temporal_pole numeric Right TMP temporal pole Range value: 0, ..., 15578

Right_TrIFG_triangular_part_of_the_inferior_frontal_gyrus numeric Right TrIFG triangular part of the inferior frontal gyrus Range value: 0, ..., 5318.54

Right_TTG_transverse_temporal_gyrus numeric Right TTG transverse temporal gyrus Range value: 0, ..., 5333

Right_Ventral_DC numeric Right Ventral DC Range value: 562.67, ..., 9100

Right_vessel numeric Right vessel Range value: 0, ..., 191

TEMPORAL numeric TEMPORAL Range value: 18226.6, ..., 302471

TEMPORAL_GM numeric TEMPORAL GM Range value: 5382.53, ..., 144002

TEMPORAL_GM_L numeric TEMPORAL GM L Range value: 1565.41, ..., 70914

TEMPORAL_GM_R numeric TEMPORAL GM R Range value: 0, ..., 73088

TEMPORAL_INFERIOR_GM numeric TEMPORAL INFERIOR GM Range value: 337, ..., 23657

TEMPORAL_INFERIOR_GM_L numeric TEMPORAL INFERIOR GM L Range value: 227, ..., 18872

TEMPORAL_INFERIOR_GM_R numeric TEMPORAL INFERIOR GM R Range value: 0, ..., 12541

TEMPORAL_L numeric TEMPORAL L Range value: 6364.8, ..., 148382

TEMPORAL_LATERAL_GM numeric TEMPORAL LATERAL GM Range value: 2763.25, ..., 112467

TEMPORAL_LATERAL_GM_L numeric TEMPORAL LATERAL GM L Range value: 906.59, ..., 55262

TEMPORAL_LATERAL_GM_R numeric TEMPORAL LATERAL GM R Range value: 0, ..., 57205

temporal_lobe_WM_left numeric temporal lobe WM left Range value: 4775.95, ..., 105964

temporal_lobe_WM_right numeric temporal lobe WM right Range value: 3002.74, ..., 92761

TEMPORAL_R numeric TEMPORAL R Range value: 3671.16, ..., 154088

TEMPORAL_SUPRATEMPORAL_GM numeric TEMPORAL SUPRATEMPORAL GM Range value: 160, ..., 13666.8

TEMPORAL_SUPRATEMPORAL_GM_L numeric TEMPORAL SUPRATEMPORAL GM L Range value: 40, ..., 7451.82

TEMPORAL_SUPRATEMPORAL_GM_R numeric TEMPORAL SUPRATEMPORAL GM R Range value: 0, ..., 12022

TEMPORAL_WM numeric TEMPORAL WM Range value: 12539.7, ..., 184421

TEMPORAL_WM_L numeric TEMPORAL WM L Range value: 4775.95, ..., 105964

TEMPORAL_WM_R numeric TEMPORAL WM R Range value: 3002.74, ..., 92761

TOTALBRAIN numeric TOTALBRAIN Range value: 607187, ..., 1585380

VENTRICLE numeric VENTRICLE Range value: 8805, ..., 178896

VENTRICLE_L numeric VENTRICLE L Range value: 3439, ..., 99287

VENTRICLE_R numeric VENTRICLE R Range value: 3056, ..., 81013.2

WM numeric WM Range value: 71065.8, ..., 785128

WM_L numeric WM L Range value: 28070.9, ..., 421470

WM_R numeric WM R Range value: 21602.3, ..., 368368

update_stamp POSIXct: YYYY-MM-DD HH:MM:SS