Neuropsychological Battery

Neuropsychological Battery data. More information is available at https://adni.loni.usc.edu/data-samples/adni-data/

Usage

data(NEUROBAT)

Format

A data frame with 17622 observations and 84 variables.

Source

https://adni.loni.usc.edu/data-samples/adni-data/

Details

ORIGPROT character Original study protocol , with options: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4

COLPROT factor Study protocol of data collection , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4

PTID character Participant ID

RID numeric Participant roster ID Range value: 1, ..., 10898

VISCODE character Visit code

VISCODE2 character Translated visit code

VISDATE Assessment DATE when present; otherwise Registry EXAMDATE Date: YYY-MM-DD

SOURCE character Information Source 1=Participant Visit ; 2=Telephone Call/Video Call Decoded Value: Yes

CLOCKCIRC character 1. Approximately circular face

ADNI1, ADNI3, ADNI4: 0=Incorrect ; 1=Correct

ADNIGO, ADNI2: 1=Correct ; 0=Incorrect

Decoded Value: Yes

CLOCKSYM character 2. Symmetry of number placement

ADNI1, ADNI3, ADNI4: 0=Incorrect ; 1=Correct

ADNIGO, ADNI2: 1=Correct ; 0=Incorrect

Decoded Value: Yes

CLOCKNUM character 3. Correctness of numbers

ADNI1, ADNI3, ADNI4: 0=Incorrect ; 1=Correct

ADNIGO, ADNI2: 1=Correct ; 0=Incorrect

Decoded Value: Yes

CLOCKHAND character 4. Presence of the two hands

ADNI1, ADNI3, ADNI4: 0=Incorrect ; 1=Correct

ADNIGO, ADNI2: 1=Correct ; 0=Incorrect

Decoded Value: Yes

CLOCKTIME character 5. Presence of the two hands, set to ten after eleven

ADNI1, ADNI3, ADNI4: 0=Incorrect ; 1=Correct

ADNIGO, ADNI2: 1=Correct ; 0=Incorrect

Decoded Value: Yes

CLOCKSCOR numeric Total Score

ADNI1, ADNIGO, ADNI2: 0..5

ADNI3, ADNI4

COPYCIRC character 1. Approximately circular face

ADNI1: 1=Yes ; 0=No

ADNIGO, ADNI2: 1=Correct ; 0=Incorrect

ADNI3, ADNI4: 0=Incorrect ; 1=Correct

Decoded Value: Yes

COPYSYM character 2. Symmetry of number placement

ADNI1: 1=Yes ; 0=No

ADNIGO, ADNI2: 1=Correct ; 0=Incorrect

ADNI3, ADNI4: 0=Incorrect ; 1=Correct

Decoded Value: Yes

COPYNUM character 3. Correctness of numbers

ADNI1: 1=Yes ; 0=No

ADNIGO, ADNI2: 1=Correct ; 0=Incorrect

ADNI3, ADNI4: 0=Incorrect ; 1=Correct

Decoded Value: Yes

COPYHAND character 4. Presence of the two hands

ADNI1: 1=Yes ; 0=No

ADNIGO, ADNI2: 1=Correct ; 0=Incorrect

ADNI3, ADNI4: 0=Incorrect ; 1=Correct

Decoded Value: Yes

COPYTIME character 5. Presence of the two hands, set to ten after eleven

ADNI1: 1=Yes ; 0=No

ADNIGO, ADNI2: 1=Correct ; 0=Incorrect

ADNI3, ADNI4: 0=Incorrect ; 1=Correct

Decoded Value: Yes

COPYSCOR numeric Total Score

ADNI1, ADNIGO, ADNI2: 0..5

ADNI3, ADNI4

LMSTORY character Which Logical Memory story was used? 1=Anna Thompson ; 2=Greg Fortune Decoded Value: Yes

LIMMTOTAL numeric Total Number of Story Units Recalled

ADNI1, ADNIGO, ADNI2: 0..25

ADNI3, ADNI4

LIMMEND numeric Time Ended (24-hour clock) Range value: 33, ..., 2041

AVTOT1 numeric Trial 1 Total

ADNI1, ADNIGO, ADNI2: 0..15

ADNI3, ADNI4

AVERR1 numeric Total Intrusions Range value: 0, ..., 13

AVTOT2 numeric Trial 2 Total

ADNI1, ADNIGO, ADNI2: 0..15

ADNI3, ADNI4

AVERR2 numeric Total Intrusions Range value: 0, ..., 12

AVTOT3 numeric Trial 3 Total

ADNI1, ADNIGO, ADNI2: 0..15

ADNI3, ADNI4

AVERR3 numeric Total Intrusions Range value: 0, ..., 14

AVTOT4 numeric Trial 4 Total

ADNI1, ADNIGO, ADNI2: 0..15

ADNI3, ADNI4

AVERR4 numeric Total Intrusions Range value: 0, ..., 14

AVTOT5 numeric Trial 5 Total

ADNI1, ADNIGO, ADNI2: 0..15

ADNI3, ADNI4

AVERR5 numeric Total Intrusions Range value: 0, ..., 15

AVTOT6 numeric Trial 6 Total

ADNI1, ADNIGO, ADNI2: 0..15

ADNI3, ADNI4

AVERR6 numeric Total Intrusions Range value: 0, ..., 13

AVTOTB numeric List B Total

ADNI1, ADNIGO, ADNI2: 0..15

ADNI3, ADNI4

AVERRB numeric Total Intrusions Range value: 0, ..., 12

AVENDED numeric AVLT Time Trial B Ended Range value: 100, ..., 1925

DSPANFOR numeric Forward: Total Correct 0..12

DSPANFLTH numeric Forward: Length 3..8

DSPANBAC numeric Backward: Total Correct 0..12

DSPANBLTH numeric Backward: Length 2..7

CATANIMSC numeric Category Fluency (Animals) - Total Correct

ADNI1, ADNIGO, ADNI2: 0..77

ADNI3, ADNI4

CATANPERS numeric Category Fluency (Animals) - Perseverations

ADNI1, ADNIGO, ADNI2: 0..77

ADNI3, ADNI4

CATANINTR numeric Category Fluency (Animals) - Intrusions

ADNI1, ADNIGO, ADNI2: 0..77

ADNI3, ADNI4

CATVEGESC numeric Category Fluency (Vegetables) - Total Correct 0..77

CATVGPERS numeric Category Fluency (Vegetables) - Perseverations 0..77

CATVGINTR numeric Category Fluency (Vegetables) - Intrusions 0..77

TRAASCOR numeric Part A - Time to Complete

ADNI1, ADNI2, ADNI3, ADNI4

ADNIGO: 1..150

TRAAERRCOM numeric Errors of Commission Range value: 0, ..., 24

TRAAERROM numeric Errors of Omission

ADNI1, ADNI3, ADNI4

ADNIGO: 0..24

ADNI2: 0..77

TRABSCOR numeric Part B - Time to complete

ADNI1, ADNI2, ADNI3, ADNI4

ADNIGO: 1..300

TRABERRCOM numeric Errors of Commission Range value: 0, ..., 25

TRABERROM numeric Errors of Omission

ADNI1, ADNI2, ADNI3, ADNI4

ADNIGO: 0..24

DIGITSCOR numeric Total Correct Range value: 0, ..., 83

LDELBEGIN numeric Time Began (24-hour clock) Range value: 8, ..., 2112

LDELTOTAL numeric Total Number of Story Units Recalled

ADNI1, ADNIGO, ADNI2: 0..25

ADNI3, ADNI4

LDELCUE character Reminder given?

ADNI1, ADNIGO, ADNI2: 1=Yes ; 0=No

ADNI3, ADNI4: 0=No ; 1=Yes

Decoded Value: Yes

BNTND character Check here if: 1=Boston Naming Not Done - Spanish speaking Decoded Value: Yes

BNTSPONT numeric 1. Total correct without a cue 0..30

BNTSTIM numeric 2. Total semantic cues given 0..30

BNTCSTIM numeric 3. Total correct with a semantic cue 0..30

BNTPHON numeric 4. Total phonemic cues given 0..30

BNTCPHON numeric 5. Number of correct responses following a phonemic cue 0..30

BNTTOTAL numeric Total Number Correct (1+3) 0..30

AVDELBEGAN numeric Time Began (24-hour clock) Range value: 30, ..., 1910

AVDEL30MIN numeric 30 Minute Delay Total

ADNI1, ADNIGO, ADNI2: 0..15

ADNI3, ADNI4

AVDELERR1 numeric Total Intrusions Range value: 0, ..., 30

AVDELTOT numeric Recognition Score

ADNI1, ADNIGO, ADNI2: 0..15

ADNI3, ADNI4

AVDELERR2 numeric Total Intrusions Range value: 0, ..., 17

ANARTERR numeric ANART Total Score (Total # of errors)

ADNI1, ADNIGO, ADNI2: 0..50

ADNI3, ADNI4

ANARTND character Check here if: 1=ANART Not Done - Spanish speaking Decoded Value: Yes

ANART character Was ANART conducted? 0=No ; 1=Yes Decoded Value: Yes

MINTSEMCUE numeric MINT Total Correct - with Semantic Cue Range value: 0, ..., 32

MINTTOTAL numeric MINT Total Correct (Uncued + Correct with Semantic cue) Range value: 0, ..., 32

MINTUNCUED numeric MINT Total Uncued Correct Range value: 0, ..., 32

ID numeric Record ID Range value: 6, ..., 187101

SITEID numeric Site ID Range value: 1, ..., 941

USERDATE User Create Date Date: YYY-MM-DD

USERDATE2 User Last Modify Date Date: YYY-MM-DD

DD_CRF_VERSION_LABEL character CRF Version Label , with options: sc, annual, bl

LANGUAGE_CODE character Language Code e=English Decoded Value: Yes

HAS_QC_ERROR character Has Quality Check Error 0=Does not have QC error or QC error has been approved ; 1=Has QC error Decoded Value: Yes

update_stamp POSIXct: YYY-MM-DD