NACC Neuropathology Data Form (Version 11)

NACC Neuropathology Data Form (Version 11) data. More information is available at https://adni.loni.usc.edu/data-samples/adni-data/

Usage

data(NEUROPATH)

Format

A data frame with 121 observations and 165 variables.

Source

https://adni.loni.usc.edu/data-samples/adni-data/

Details

ORIGPROT character Original study protocol , with options: ADNI2, ADNI1, ADNI3, ADNIGO

PTID character Participant ID

RID numeric Participant roster ID Range value: 8, ..., 6891

VISCODE2 character Translated visit code , with options: AUT

NPFORMMO numeric Range value: 1, ..., 12

NPFORMDY numeric Date Form Completed - Day Range value: 1, ..., 31

NPFORMYR numeric Date Form Completed - Year Range value: 2008, ..., 2025

NPFORMDATE_DRVD 2. Date Form Completed (Derived from NPFORMMO,NPFORMDY,NPFORMYR) Date: YYY-MM-DD

NPDAGE numeric Age at death (years) Range value: 59, ..., 97

NPDODYR numeric Date of Death - Year Range value: 2007, ..., 2023

NPPMIH numeric Postmortem interval (PMI) (hours) Range value: 2, ..., 9999

NPFIX numeric 8. Fixative 1=1 Formalin ; 2=2 Paraformaldehyde ; 7=7 Other (Specify)

NPFIXX character Specify: , with options: First Paraform, then Formalin

NPWBRWT numeric a. Whole brain weight (if half brain, multiply weight by two) (9999=unknown) Range value: 850, ..., 9999

NPWBRF numeric b. Does the value in Question 9a represent fresh or fixed weight? (Check One) 1=1 Fresh ; 2=2 Fixed ; 8=8 Not applicable

NPGRCCA numeric c1. Cerebral cortex atrophy 0=0 None ; 1=1 Mild ; 2=2 Moderate ; 3=3 Severe ; 8=8 Not assessed ; 9=9 Missing/unknown

NPGRLA numeric c2. Lobar atrophy (significant frontal and/or temporal atrophy) 0=0 None ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown

NPGRHA numeric c3. Hippocampus atrophy 0=0 None ; 1=1 Mild ; 2=2 Moderate ; 3=3 Severe ; 8=8 Not assessed ; 9=9 Missing/unknown

NPGRSNH numeric c4. Substantia nigra hypopigmentation 0=0 None ; 1=1 Mild ; 2=2 Moderate ; 3=3 Severe ; 8=8 Not assessed ; 9=9 Missing/unknown

NPGRLCH numeric c5. L. ceruleus hypopigmentation 0=0 None ; 1=1 Mild ; 2=2 Moderate ; 3=3 Severe ; 8=8 Not assessed ; 9=9 Missing/unknown

NPAVAS numeric c6. Atherosclerosis (of the circle of Willis) 0=0 None ; 1=1 Mild ; 2=2 Moderate ; 3=3 Severe ; 8=8 Not assessed ; 9=9 Missing/unknown

NPTAN numeric a. Tau antibody (Check One) 1=1 Non-phospho specific ; 2=2 PHF1 ; 3=3 CP13 ; 4=4 AT8 ; 7=7 Other (Specify) ; 8=8 Not assessed

NPTANX logical Specify: A Boolean value: TRUE/ FALSE

NPABAN numeric b. Amyloid beta antibody (Check One) 1=1 4G8 ; 2=2 10D5 ; 7=7 Other (Specify) ; 8=8 Not assessed

NPABANX logical Specify: A Boolean value: TRUE/ FALSE

NPASAN numeric c. Alpha synuclein antibody 1=1 Non-phospho specific (e.g. LB509) ; 2=2 Phospho-specific (e.g. pSYN#64) ; 7=7 Other (Specify) ; 8=8 Not assessed

NPASANX logical Specify: A Boolean value: TRUE/ FALSE

NPTDPAN numeric d. TDP-43 antibody (Check One) 1=1 Non-phospho specific ; 2=2 Phospho-specific ; 7=7 Other (Specify) ; 8=8 Not assessed

NPTDPANX logical Specify: A Boolean value: TRUE/ FALSE

NPHISMB numeric e1. Modified Bielschowsky 0=0 No ; 1=1 Yes

NPHISG numeric e2. Gallyas 0=0 No ; 1=1 Yes

NPHISSS numeric e3. Other Silver stain 0=0 No ; 1=1 Yes

NPHIST numeric e4. Thioflavin 0=0 No ; 1=1 Yes

NPHISO numeric e5. Other 0=0 No ; 1=1 Yes

NPHISOX character Specify: , with options: Double IHC, AT8 + H31L21, Neurofilament IHC, Dual IHC Aß (H31L21) + Tau (AT8), LFB/PAS, p62, Neurofilament IHC

NPTHAL numeric a. Thal phase for amyloid plaques by immunohisto-chemistry (IHC) (A score - Check One) 0=0 Phase 0 (A0) ; 1=1 Phase 1 (A1) ; 2=2 Phase 2 (A1) ; 3=3 Phase 3 (A2) ; 4=4 Phase 4 (A3) ; 5=5 Phase 5 (A3) ; 8=8 Not assessed ; 9=9 Missing/unknown

NPBRAAK numeric b. Braak stage for neurofibrillary degeneration (B score - Check One) 0=0 Sage 0: AD-type neurofibrillary degeneration not present (B0) ; 1=1 Stage I (B1) ; 2=2 Stage II (B1) ; 3=3 Stage III (B2) ; 4=4 Stage IV (B2) ; 5=5 Stage V (B3) ; 6=6 Stage VI (B3) ; 7=7 The presence of a tauopathy (other than aging/AD) precludes Braak staging ; 8=8 Not assessed ; 9=9 Missing/unknown

NPNEUR numeric c. CERAD score for density of neocortical neuritic plaque (plaques with argrophilic dystrophic neurities, with or without dense amyloid cores). Score without respect to age or diagnosis. (C score - Check One) 0=0 No neuritic plaques (C0) ; 1=1 Sparse neuritic plaques (C1) ; 2=2 Moderate neuritic plaques (C2) ; 3=3 Frequent neuritic plaques (C3) ; 8=8 Not assessed ; 9=9 Missing/unknown

NPADNC numeric d. NIA-AA Alzheimer's disease neuropathologic change (ADNC) (Check One) 0=0 Not AD ; 1=1 Low ADNC ; 2=2 Intermediate ADNC ; 3=3 High ADNC ; 8=8 Not assessed ; 9=9 Missing/unknown

NPDIFF numeric e1. CERAD semi-quantitative score for diffuse plaques (plaques with non-compact amyloid and no apparent dystrophic neurites). Score from the neocortical field with the highest plaque density and without respect to age or diagnosis. (Check One) 0=0 No diffuse plaques ; 1=1 Sparse diffuse plaques ; 2=2 Moderate diffuse plaques ; 3=3 Frequent diffuse plaques ; 8=8 Not assessed ; 9=9 Missing/unknown

NPAMY numeric e2. Cerebral amyloid angiopathy (Check One) 0=0 None ; 1=1 Mild ; 2=2 Moderate ; 3=3 Severe ; 8=8 Not assessed ; 9=9 Missing/unknown

NPINF numeric a. Old infarcts observed grossly, including lacunes? (Check One) 0=0 No (SKIP TO QUESTION 12b) ; 1=1 Yes (COMPLETE QUESTIONS 12a1-12a4) ; 8=8 Not assessed (SKIP TO QUESTION 12b) ; 9=9 Missing/unknown (SKIP TO QUESTION 12b)

NPINF1A character Number of infarcts – Cerebral cortex , with options: ., 0, 1, 5

NPINF1B character Size of largest (greatest dimension in cm) , with options: ., 4, 4.5, 1

NPINF1D character Size of next (greatest dimension in cm) , with options: ., 4

NPINF1F character Size of next (greatest dimension in cm) , with options: ., 3

NPINF2A character Number of infarcts – Subcortical cerebral white matter and peri-ventricular white matter , with options: ., 1, 0

NPINF2B character Size of largest (greatest dimension in cm) , with options: ., 1.2, 0.3

NPINF2D logical Size of next (greatest dimension in cm) A Boolean value: TRUE/ FALSE

NPINF2F logical Size of next (greatest dimension in cm) A Boolean value: TRUE/ FALSE

NPINF3A character Number of infarcts – Deep cerebral gray matter or internal capsule , with options: ., 0, 1, 2

NPINF3B character Size of largest (greatest dimension in cm) , with options: ., 2, 1.5, 1, 0.3

NPINF3D character Size of next (greatest dimension in cm) , with options: ., 0.5

NPINF3F logical Size of next (greatest dimension in cm) A Boolean value: TRUE/ FALSE

NPINF4A character Number of infarcts – Brainstem or cerebellum , with options: ., 0, 1

NPINF4B character Size of largest (greatest dimension in cm) , with options: ., 0.5, 0.6

NPINF4D logical Size of next (greatest dimension in cm) A Boolean value: TRUE/ FALSE

NPINF4F logical Size of next (greatest dimension in cm) A Boolean value: TRUE/ FALSE

NPHEMO numeric b. Were single or multiple old hemorrhages observed grossly? 0=0 No (SKIP TO QUESTION 12c) ; 1=1 Yes (COMPLETE QUESTIONS 12b1-12b3) ; 8=8 Not assessed (SKIP TO QUESTION 12c) ; 9=9 Missing/unknown (SKIP TO QUESTION 12c)

NPHEMO1 character b1. Subdural or epidural hemorrhage 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown

NPHEMO2 character b2. Primary parenchymal hemorrhage Include those >5mm. If ?5mm, include as microbleed; see Question 12d. 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown

NPHEMO3 character b3. Secondary parenchymal hemorrhage (e.g., tumor, vascular malformation) 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown

NPOLD numeric c. Old microinfarcts (not observed grossly)? (Check One) 0=0 No (SKIP TO QUESTION 12d) ; 1=1 Yes (COMPLETE QUESTIONS 12c1-12c4) ; 8=8 Not assessed (SKIP TO QUESTION 12d) ; 9=9 Missing/unknown (SKIP TO QUESTION 12d)

NPOLD1 character c1. Number in screening sections of cerebral cortex (gray matter of cerebral cortex) 0=0 ; 1=1 ; 2=2 ; 3=3 or more ; 8=8 Not assessed ; 9=9 Missing/unknown

NPOLD2 character c2. Number in screening sections of subcortical white matter and periventricular white matter 0=0 ; 1=1 ; 2=2 ; 3=3 or more ; 8=8 Not assessed ; 9=9 Missing/unknown

NPOLD3 character c3. Number in screening sections of subcortical gray matter 0=0 ; 1=1 ; 2=2 ; 3=3 or more ; 8=8 Not assessed ; 9=9 Missing/unknown

NPOLD4 character c4. Number in brainstem and cerebellum 0=0 ; 1=1 ; 2=2 ; 3=3 or more ; 8=8 Not assessed ; 9=9 Missing/unknown

NPOLDD numeric d. Old cerebral microbleeds? (Check One) Include old hemorrahages that are greater than or equal to 5mn. 0=0 No (SKIP TO QUESTION 12e) ; 1=1 Yes (COMPLETE QUESTIONS 12d1-12d4) ; 8=8 Not assessed (SKIP TO QUESTION 12e) ; 9=9 Missing/unknown (SKIP TO QUESTION 12e)

NPOLDD1 character d1. Number in screening sections of cerebral cortex 0=0 ; 1=1 ; 2=2 ; 3=3 or more ; 8=8 Not assessed ; 9=9 Missing/unknown

NPOLDD2 character d2. Number in screening sections of subcortical white matter and periventricular white matter 0=0 ; 1=1 ; 2=2 ; 3=3 or more ; 8=8 Not assessed ; 9=9 Missing/unknown

NPOLDD3 character d3. Number in screening sections of subcortical gray matter 0=0 ; 1=1 ; 2=2 ; 3=3 or more ; 8=8 Not assessed ; 9=9 Missing/unknown

NPOLDD4 character d4. Number in brainstem and cerebellum 0=0 ; 1=1 ; 2=2 ; 3=3 or more ; 8=8 Not assessed ; 9=9 Missing/unknown

NPARTER numeric e. Arteriolosclerosis? (Check One) (Assessed in subcortical white or gray matter) 0=0 None ; 1=1 Mild ; 2=2 Moderate ; 3=3 Severe ; 8=8 Not assessed ; 9=9 Missing/unknown

NPWMR numeric f. White matter rarefaction? (Check One) (H&E or myelin stain may be used) 0=0 None ; 1=1 Mild ; 2=2 Moderate ; 3=3 Severe ; 8=8 Not assessed ; 9=9 Missing/unknown

NPPATH numeric g. Other pathologic changes related to ischemic or vascular disease not previously specified? 0=0 No (SKIP TO QUESTION 13) ; 1=1 Yes (COMPLETE QUESTIONS 12g1-12g12) ; 8=8 Not assessed (SKIP TO QUESTION 13) ; 9=9 Missing/unknown (SKIP TO QUESTION 13)

NPNEC character g1. Laminar necrosis 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown

NPPATH2 character g2. Acute neuronal necrosis 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown

NPPATH3 character g3. Acute/subacute gross infarcts 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown

NPPATH4 character g4. Acute/subacute microinfarcts 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown

NPPATH5 character g5. Acute/subacute gross hemorrhage 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown

NPPATH6 character g6. Acute/subacute microhemorrhage 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown

NPPATH7 character g7. Vascular malformation of any type 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown

NPPATH8 character g8. Aneurysm of any type 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown

NPPATH9 character g9. Vasculitis of any type 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown

NPPATH10 character g10. CADASIL 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown

NPPATH11 character g11. Mineralization of blood vessels 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown

NPPATHO character g12. Other 0=0 No ; 1=1 Yes

NPPATHOX character Specify: , with options: Mural/Luminal protein deposits, Coxiella Burnetii, Remote ischemic foci in CA1, subacute corticospinal tract degeneration, FOCAL SUBACUTE THROMBOSIS, R PRECENTRAL GYRUS

NPLBOD numeric Is there evidence of Lewy body pathology? (Check One) 0=0 No ; 1=1 Brainstem predominant ; 2=2 Limbic (transitional) ; 3=3 Neocortical (diffuse) ; 4=4 Amygdala predominant ; 5=5 Olfactory bulb ; 8=8 Not assessed ; 9=9 Missing/unknown

NPNLOSS numeric Neuron loss in the substantia nigra 0=0 None ; 1=1 Mild ; 2=2 Moderate ; 3=3 severe ; 8=8 Not assessed ; 9=9 Missing/unknown

NPHIPSCL numeric Hippocampal sclerosis (CA1 and/or subiculum) 0=0 None ; 1=1 Unilateral ; 2=2 Bilateral ; 3=3 Present but laterality not assessed ; 8=8 Not assessed ; 9=9 Missing/unknown

NPTDPA numeric a. Spinal cord 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown

NPTDPB numeric b. Amygdala 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown

NPTDPC numeric c. Hippocampus 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown

NPTDPD numeric d. Entorhinal/inferior temporal cortex 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown

NPTDPE numeric e. Neocortex 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown

NPFTDTAU numeric a. FTLD with tau pathology (FTLD-tau) or other tauopathy (Check One) 0=0 No (SKIP TO QUESTION 17c) ; 1=1 Yes (COMPLETE QUESTIONS 17b1-17b10) ; 8=8 Not assessed (SKIP TO QUESTION 17c) ; 9=9 Missing/unknown (SKIP TO QUESTION 17c)

NPPICK character b1. FTLD-tau (PiD) 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown

NPFTDT2 character b2. Other 3R tauopathy (Includes MAPT mutation tauopathy) 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown

NPCORT character b3. FTLD-tau (CBD) 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown

NPPROG character b4. FTLD-tau (PSP) 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown

NPFTDT5 character b5. Argyrophilic grains 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown

NPFTDT6 character b6. Other 4R tauopathy (Includes sporadic multiple systems tauopathy, globular glial tauopathy, MAPT mutation tauopathy) 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown

NPFTDT7 character b7. Chronic traumatic encephalopathy 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown

NPFTDT8 character b8. Amyotrophic lateral sclerosis (ALS)/Parkinsonism-dementia complex of Guam 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown

NPFTDT9 character b9. Tangle dominant disease 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown

NPFTDT10 character b10. Other 3R + 4F tauopathy (Includes unclassifiable, focal, glial only, MAPT mutation tauopathy, NOS) 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown

NPFTDTDP numeric c. FTLD with TDP-43 pathology (FTLD-TDP)? (Check One) 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown

NPALSMND numeric d. ALS/motor neuron disease (MND) present? (Check One) 0=0 No ; 1=1 Yes, with TDP-43 inclusions in motor neurons ; 2=2 Yes, with FUS inclusions in motor neurons ; 3=3 Yes, with SOD1 inclusions in motor neurons ; 4=4 Yes, with other inclusions ; 5=5 Yes, with no specific inclusions ; 8=8 Not assessed ; 9=9 Missing/unknown

NPOFTD numeric e. Other FTLD? (Check One) 0=0 No (SKIP TO QUESTION 18a) ; 1=1 Yes (COMPLETE QUESTIONS 17f1 - 17f5) ; 8=8 Not assessed (SKIP TO QUESTION 18a) ; 9=9 Missing/unknown (SKIP TO QUESTION 18a)

NPOFTD1 logical f1. Atypical FTLD-U (aFTLD-U) 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown

NPOFTD2 logical f2. NIFID (neuronal intermediate filament inclusions disease) 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown

NPOFTD3 logical f3. BIBD (basophilic inclusion body disease) 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown

NPOFTD4 logical f4. FTLD-UPS (ubiquitin-proteasome system [ubiquitin or p62 positive, tau/TDP-43/FUS negative inclusions]) 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown

NPOFTD5 logical f5. FTLD-NOS (includes dementia lacking distinctive histology (DLDH) and FTLD with no inclusions (FTLD-NI) detected by tau, TDP-43, or ubiquitin/p62 IHC) 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown

NPPDXA numeric a. Pigment-spheroid degeneration/NBIA 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown

NPPDXB numeric b. Multiple system atrophy 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown

NPPDXC numeric c. Prion disease 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown

NPPDXD numeric d. Trinucleotide disease (Huntington disease, SCA, other) 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown

NPPDXE numeric e. Malformation of cortical development 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown

NPPDXF numeric f. Metabolic/storage disorder of any type 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown

NPPDXG numeric g. WM disease, leukodystrophy 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown

NPPDXH numeric h. WM disease, multiple sclerosis or other demyelinating disease 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown

NPPDXI numeric i. Contusion/traumatic brain injury of any type, acute 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown

NPPDXJ numeric j. Contusion/traumatic brain injury of any type, chronic 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown

NPPDXK numeric k. Neoplasm, primary 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown

NPPDXL numeric l. Neoplasm, metastatic 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown

NPPDXM numeric m. Infectious process of any type (encephalitis, abscess, etc.) 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown

NPPDXN numeric n. Herniation, any site 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown

NPPDXO numeric o. Trisomy 21/Down syndrome 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown

NPPDXP numeric p. AD-related genes (dominantly inherited); do not include APOE or other polymorphisms or genetic risk factors. 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown

NPPDXQ numeric q. FTLD-related genes (dominantly inherited); do not include polymorphisms or genetic risk factors. 0=0 No ; 1=1 Yes ; 8=8 Not assessed ; 9=9 Missing/unknown

NPPDXR numeric r. Other (Specify) 0=0 No ; 1=1 Yes

NPPDXRX character Specify:

NPPDXS numeric s. Other (Specify) 0=0 No ; 1=1 Yes

NPPDXSX character Specify:

NPPDXT numeric t. Other (Specify) 0=0 No ; 1=1 Yes

NPPDXTX character Specify: , with options: Vascular mineralization, mild, hippo, pallidum, LATE-NC, ARTAG, Moderate, Widespread, Leptomeningeal hemosiderin bilateral anterior temporal lobes, Meningioma, transitional, WHO grade 1, 0.5cm, TDP-43 Proteinopathy, Joseph's Stage 3, LATE-NC Stage 2, ARTAG, Mild, Amygdala, Basal Forebrain, Olfactory Cortex, Parietal White Matter, LBD, very mild, substantia nigra only

NPBNKA numeric a. Banked frozen brain or half brain 0=0 No ; 1=1 Yes ; 9=9 Missing/unknown

NPBNKB numeric b. Banked frozen wedge of cerebellum or other sample for future DNA prep 0=0 No ; 1=1 Yes ; 9=9 Missing/unknown

NPBNKC numeric c. Formalin- or paraformaldehyde-fixed brain 0=0 No ; 1=1 Yes ; 9=9 Missing/unknown

NPBNKD numeric d. Paraffin-embedded blocks of brain regions 0=0 No ; 1=1 Yes ; 9=9 Missing/unknown

NPBNKE numeric e. Banked postmortem CSF 0=0 No ; 1=1 Yes ; 9=9 Missing/unknown

NPBNKF numeric f. Banked postmortem blood or serum 0=0 No ; 1=1 Yes ; 9=9 Missing/unknown

NPBNKG numeric g. Banked DNA 0=0 No ; 1=1 Yes ; 9=9 Missing/unknown

NPFAUT numeric h. Full autopsy performed? 0=0 No ; 1=1 Yes ; 9=9 Missing/unknown

NPFAUT1 character If full autopsy, first major finding , with options: Leukemia, Acute and Chronic Congestive Heart Failure

NPFAUT2 character If full autopsy, second major finding , with options: AD, Severe Atherosclerosis - Coronary, Aorta

NPFAUT3 character If full autopsy, third major finding , with options: 0, Erythroid Hyperplasia with megaloblastic changes

NPFAUT4 character If full autopsy, fourth major finding , with options: 0, .

FORMVER numeric Range value: 10, ..., 11

NPARTAG character 0=0 No ; 1=1 Yes ; 9=9 Missing/unknown

NPATGSEV character 1=Mild ; 2=Moderate ; 3=Severe ; 8=Not assesed ; 9=Missing/unknown

NPATGAMY character 0=0 No ; 1=1 Yes ; 9=9 Missing/unknown

NPATGAM1 character 0=None ; 1=Focal ; 2=Widespread ; 8=Not assessed ; 9=Missing/unknown

NPATGAM2 character 0=None ; 1=Focal ; 2=Widespread ; 8=Not assessed ; 9=Missing/unknown

NPATGAM3 character 0=None ; 1=Focal ; 2=Widespread ; 8=Not assessed ; 9=Missing/unknown

NPATGAM4 character 0=None ; 1=Focal ; 2=Widespread ; 8=Not assessed ; 9=Missing/unknown

NPATGAM5 character 0=None ; 1=Focal ; 2=Widespread ; 8=Not assessed ; 9=Missing/unknown

NPATGFRN character 0=0 No ; 1=1 Yes ; 9=9 Missing/unknown

NPATGFR1 character 0=None ; 1=Focal ; 2=Widespread ; 8=Not assessed ; 9=Missing/unknown

NPATGFR2 character 0=None ; 1=Focal ; 2=Widespread ; 8=Not assessed ; 9=Missing/unknown

NPATGFR3 character 0=None ; 1=Focal ; 2=Widespread ; 8=Not assessed ; 9=Missing/unknown

NPATGFR4 character 0=None ; 1=Focal ; 2=Widespread ; 8=Not assessed ; 9=Missing/unknown

update_stamp POSIXct: YYY-MM-DD