Neuropsychiatric Inventory

Neuropsychiatric Inventory data. More information is available at https://adni.loni.usc.edu/data-samples/adni-data/

Usage

data(NPI)

Format

A data frame with 7734 observations and 169 variables.

Source

https://adni.loni.usc.edu/data-samples/adni-data/

Details

ORIGPROT character Original study protocol , with options: ADNI1, ADNI2, ADNIGO, ADNI3, ADNI4

COLPROT factor Study protocol of data collection , with factor levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4

PTID character Participant ID

RID numeric Participant roster ID Range value: 2, ..., 10592

VISCODE character Visit code

VISCODE2 character Translated visit code

EXAMDATE Examination Date Date: YYY-MM-DD

NPISOURC character Information source

ADNI2: 1=Participant visit; 2=Telephone call

ADNI3: 1=Participant visit;2=Telephone call/Video Call

ADNI4: 1=Participant visit; 2=Telephone call/Video Call

Decoded Value: Yes

SPID numeric Study Partner ID: Range value: 1, ..., 10

NPIA character A. Delusions

ADNI2, ADNI4: 0=No; 1=Yes

ADNI3: 0=No;1=Yes

Decoded Value: Yes

NPIA1 character 1. Does P believe that he/she is in danger - that others are planning to hurt him/her?

ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPIA2 character 2. Does P believe that others are stealing from him/her?

ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPIA3 character 3. Does P believe that his/her spouse is having an affair?

ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPIA4 character 4. Does P believe that unwelcome guests are living in his/her house?

ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPIA5 character 5. Does P believe that his/her spouse or others are not who they claim to be?

ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPIA6 character 6. Does P believe that his/her house is not his/her home?

ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPIA7 character 7. Does P believe that family members plan to abandon him/her?

ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPIA8 character 8. Does P believe that television or magazine figures are actually present in the home? [Does he/she try to talk to or interact with them?]

ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPIA9 character 9. Does P believe any other unusual things that I haven't asked about?

ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPIA10A character 10a. Frequency Ratings

ADNI2: 1=1: Occasionally; 2=2: Often; 3=3: Frequently; 4=4: Very frequently

ADNI3: 1=1: Occasionally;2=2: Often;3=3: Frequently;4=4: Very frequently

ADNI4: 1=1: Occasionally; 2=2: Often; 3=3: Frequently; 4=4: Very frequently

Decoded Value: Yes

NPIA10B character 10b. Severity Ratings

ADNI2, ADNI4: 1=1: Mild; 2=2: Moderate; 3=3: Marked

ADNI3: 1=1: Mild;2=2: Moderate;3=3: Marked

Decoded Value: Yes

NPIA10C character 10c. Caregiver Distress

ADNI2, ADNI4: 0=Not at all; 1=Minimally; 2=Mildly; 3=Moderately; 4=Severely; 5=Very severely or extremely

ADNI3: 0=Not at all;1=Minimally;2=Mildly;3=Moderately;4=Severely;5=Very severely or extremely

Decoded Value: Yes

NPIATOT numeric A. Delusions: Item score

ADNI2, ADNI3: 0..12

ADNI4

NPIB character B. Hallucinations

ADNI2, ADNI4: 0=No; 1=Yes

ADNI3: 0=No;1=Yes

Decoded Value: Yes

NPIB1 character 1. Does P describe hearing voices or act as if he/she hears voices?

ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPIB2 character 2. Does P talk to people who are not there?

ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPIB3 character 3. Does P describe seeing things not seen by others or behave as if he/she is seeing things not seen by others (people, animals, lights, etc.)?

ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPIB4 character 4. Does P report smelling odors not smelled by others?

ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPIB5 character 5. Does P describe feeling things on his/her skin or otherwise appear to be feeling things crawling or touching him/her?

ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPIB6 character 6. Does P describe tastes that are without any known cause?

ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPIB7 character 7. Does P describe any other unusual sensory experiences?

ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPIB8A character 8a. Frequency Ratings

ADNI2: 1=1: Occasionally; 2=2: Often; 3=3: Frequently; 4=4: Very frequently

ADNI3: 1=1: Occasionally;2=2: Often;3=3: Frequently;4=4: Very frequently

ADNI4: 1=1: Occasionally; 2=2: Often; 3=3: Frequently; 4=4: Very frequently

Decoded Value: Yes

NPIB8B character 8b. Severity Ratings

ADNI2, ADNI4: 1=1: Mild; 2=2: Moderate; 3=3: Marked

ADNI3: 1=1: Mild;2=2: Moderate;3=3: Marked

Decoded Value: Yes

NPIB8C character 8c. Caregiver Distress

ADNI2, ADNI4: 0=Not at all; 1=Minimally; 2=Mildly; 3=Moderately; 4=Severely; 5=Very severely or extremely

ADNI3: 0=Not at all;1=Minimally;2=Mildly;3=Moderately;4=Severely;5=Very severely or extremely

Decoded Value: Yes

NPIBTOT numeric B. Hallucinations: Item score

ADNI2, ADNI3: 0..12

ADNI4

NPIC character C: Agitation/Aggression

ADNI2, ADNI4: 0=No; 1=Yes

ADNI3: 0=No;1=Yes

Decoded Value: Yes

NPIC1 character 1. Does P get upset with those trying to care for him/her or resist activities such as bathing or changing clothes?

ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPIC2 character 2. Is P stubborn, having to have things his/her way?

ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPIC3 character 3. Is P uncooperative, resistive to help from others?

ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPIC4 character 4. Does P have any other behaviors that make him/her hard to handle?

ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPIC5 character 5. Does P shout or curse angrily?

ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPIC6 character 6. Does P slam doors, kick furniture, throw things?

ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPIC7 character 7. Does P attempt to hurt or hit others?

ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPIC8 character 8. Does P have any other aggressive or agitated behaviors?

ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPIC9A character 9a. Frequency Ratings

ADNI2: 1=1: Occasionally; 2=2: Often; 3=3: Frequently; 4=4: Very frequently

ADNI3: 1=1: Occasionally;2=2: Often;3=3: Frequently;4=4: Very frequently

ADNI4: 1=1: Occasionally; 2=2: Often; 3=3: Frequently; 4=4: Very frequently

Decoded Value: Yes

NPIC9B character 9b. Severity Ratings

ADNI2, ADNI4: 1=1: Mild; 2=2: Moderate; 3=3: Marked

ADNI3: 1=1: Mild;2=2: Moderate;3=3: Marked

Decoded Value: Yes

NPIC9C character 9c. Caregiver Distress

ADNI2, ADNI4: 0=Not at all; 1=Minimally; 2=Mildly; 3=Moderately; 4=Severely; 5=Very severely or extremely

ADNI3: 0=Not at all;1=Minimally;2=Mildly;3=Moderately;4=Severely;5=Very severely or extremely

Decoded Value: Yes

NPICTOT numeric C. Agitation/Aggression: Item score

ADNI2, ADNI3: 0..12

ADNI4

NPID character D: Depression/Dysphoria

ADNI2, ADNI4: 0=No; 1=Yes

ADNI3: 0=No;1=Yes

Decoded Value: Yes

NPID1 character 1. Does P have periods of tearfulness or sobbing that seem to indicate sadness?

ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPID2 character 2. Does P say or act as if he/she is sad or low in spirits?

ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPID3 character 3. Does P put him/herself down or say that he/she feels like a failure?

ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPID4 character 4. Does P say that he/she is a bad person or deserves to be punished?

ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPID5 character 5. Does P seem very discouraged or say that he/ she has not future?

ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPID6 character 6. Does P say he/she is a burden to the family or that the family would be better off without him/her?

ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPID7 character 7. Does P express a wish for death or talk about killing him/herself?

ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPID8 character 8. Does P show any other signs of depression or sadness?

ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPID9A character 9a. Frequency Ratings

ADNI2: 1=1: Occasionally; 2=2: Often; 3=3: Frequently; 4=4: Very frequently

ADNI3: 1=1: Occasionally;2=2: Often;3=3: Frequently;4=4: Very frequently

ADNI4: 1=1: Occasionally; 2=2: Often; 3=3: Frequently; 4=4: Very frequently

Decoded Value: Yes

NPID9B character 9b. Severity Ratings

ADNI2, ADNI4: 1=1: Mild; 2=2: Moderate; 3=3: Marked

ADNI3: 1=1: Mild;2=2: Moderate;3=3: Marked

Decoded Value: Yes

NPID9C character 9c. Caregiver Distress

ADNI2, ADNI4: 0=Not at all; 1=Minimally; 2=Mildly; 3=Moderately; 4=Severely; 5=Very severely or extremely

ADNI3: 0=Not at all;1=Minimally;2=Mildly;3=Moderately;4=Severely;5=Very severely or extremely

Decoded Value: Yes

NPIDTOT numeric D. Depression/Dysphoria: Item score

ADNI2, ADNI3: 0..12

ADNI4

NPIE character E. Anxiety

ADNI2, ADNI4: 0=No; 1=Yes

ADNI3: 0=No;1=Yes

Decoded Value: Yes

NPIE1 character 1. Does P say that he/she is worried about planned events?

ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPIE2 character 2. Does P have periods of feeling shaky, unable to relax, or feeling excessively tense?

ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPIE3 character 3. Does P have periods of [or complain of] shortness of breath, gasping, or sighing for no apparent reason other than nervousness?

ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPIE4 character 4. Does P complain of butterflies in his/her stomach, or of racing or pounding of the heart in association with nervousness? [Symptoms not explained by ill health]?

ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPIE5 character 5. Does P avoid certain places or situations that make him.her more nervous such as riding in the car, meeting with friends, or being in crowds?

ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPIE6 character 6. Does P become nervous or upset when separated from you [or his/her caregiver]? [Does he/she cling to you to keep from being separated?]

ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPIE7 character 7. Does P show any other signs of anxiety?

ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPIE8A character 8a. Frequency Ratings

ADNI2: 1=1: Occasionally; 2=2: Often; 3=3: Frequently; 4=4: Very frequently

ADNI3: 1=1: Occasionally;2=2: Often;3=3: Frequently;4=4: Very frequently

ADNI4: 1=1: Occasionally; 2=2: Often; 3=3: Frequently; 4=4: Very frequently

Decoded Value: Yes

NPIE8B character 8b. Severity Ratings

ADNI2, ADNI4: 1=1: Mild; 2=2: Moderate; 3=3: Marked

ADNI3: 1=1: Mild;2=2: Moderate;3=3: Marked

Decoded Value: Yes

NPIE8C character 8c. Caregiver Distress

ADNI2, ADNI4: 0=Not at all; 1=Minimally; 2=Mildly; 3=Moderately; 4=Severely; 5=Very severely or extremely

ADNI3: 0=Not at all;1=Minimally;2=Mildly;3=Moderately;4=Severely;5=Very severely or extremely

Decoded Value: Yes

NPIETOT numeric E. Anxiety: Item score

ADNI2, ADNI3: 0..12

ADNI4

NPIF character F. Elation/Euphoria

ADNI2, ADNI4: 0=No; 1=Yes

ADNI3: 0=No;1=Yes

Decoded Value: Yes

NPIF1 character 1. Does P appear to feel too good or too happy, different from his/her usual self?

ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPIF2 character 2. Does P find humor and laugh at things that others do not find funny?

ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPIF3 character 3. Does P seem to have a childish sense of humor with a tendency to giggle or laugh inappropriately (such as when something unfortunate happens to others)?

ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPIF4 character 4. Does P tell jokes or make remarks that have little humor for others but seem funny to him/her?

ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPIF5 character 5. Does he/she play childish pranks such as pinching or playing "keep away" for the fun of it?

ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPIF6 character 6. Does P "talk big" or claim to have more abilities or wealth than is true?

ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPIF7 character 7. Does P show any other signs of feeling too good or being too happy?

ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPIF8A character 8a. Frequency Ratings

ADNI2: 1=1: Occasionally; 2=2: Often; 3=3: Frequently; 4=4: Very frequently

ADNI3: 1=1: Occasionally;2=2: Often;3=3: Frequently;4=4: Very frequently

ADNI4: 1=1: Occasionally; 2=2: Often; 3=3: Frequently; 4=4: Very frequently

Decoded Value: Yes

NPIF8B character 8b. Severity Ratings

ADNI2, ADNI4: 1=1: Mild; 2=2: Moderate; 3=3: Marked

ADNI3: 1=1: Mild;2=2: Moderate;3=3: Marked

Decoded Value: Yes

NPIF8C character 8c. Caregiver Distress

ADNI2, ADNI4: 0=Not at all; 1=Minimally; 2=Mildly; 3=Moderately; 4=Severely; 5=Very severely or extremely

ADNI3: 0=Not at all;1=Minimally;2=Mildly;3=Moderately;4=Severely;5=Very severely or extremely

Decoded Value: Yes

NPIFTOT numeric F. Elation/Euphoria: Item score

ADNI2, ADNI3: 0..12

ADNI4

NPIG character G. Apathy/Indifference

ADNI2, ADNI4: 0=No; 1=Yes

ADNI3: 0=No;1=Yes

Decoded Value: Yes

NPIG1 character 1. Does P seem less spontaneous and less active than usual?

ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPIG2 character 2. Is P less likely to initiate a conversation?

ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPIG3 character 3. Is P less affectionate or lacking in emotions when compared to his/her usual self?

ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPIG4 character 4. Does P contribute less to household chores?

ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPIG5 character 5. Does P seem less interested in the activities and plans of others?

ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPIG6 character 6. Has P lost interest in friends and family members?

ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPIG7 character 7. Is P less enthusiastic about his/her usual interests?

ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPIG8 character 8. Does P show any other signs that he/she doesn't care about doing new things?

ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPIG9A character 9a. Frequency Ratings

ADNI2: 1=1: Occasionally; 2=2: Often; 3=3: Frequently; 4=4: Very frequently

ADNI3: 1=1: Occasionally;2=2: Often;3=3: Frequently;4=4: Very frequently

ADNI4: 1=1: Occasionally; 2=2: Often; 3=3: Frequently; 4=4: Very frequently

Decoded Value: Yes

NPIG9B character 9b. Severity Ratings

ADNI2, ADNI4: 1=1: Mild; 2=2: Moderate; 3=3: Marked

ADNI3: 1=1: Mild;2=2: Moderate;3=3: Marked

Decoded Value: Yes

NPIG9C character 9c. Caregiver Distress

ADNI2, ADNI4: 0=Not at all; 1=Minimally; 2=Mildly; 3=Moderately; 4=Severely; 5=Very severely or extremely

ADNI3: 0=Not at all;1=Minimally;2=Mildly;3=Moderately;4=Severely;5=Very severely or extremely

Decoded Value: Yes

NPIGTOT numeric G. Apathy/Indifference: Item score

ADNI2, ADNI3: 0..12

ADNI4

NPIH character H. Disinhibition

ADNI2, ADNI4: 0=No; 1=Yes

ADNI3: 0=No;1=Yes

Decoded Value: Yes

NPIH1 character 1. Does P act impulsively without appearing to consider the consequences?

ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPIH2 character 2. Does P talk to total strangers as if he/she knew them?

ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPIH3 character 3. Does P say things to people that are insensitive or hurt their feelings?

ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPIH4 character 4. Does P say crude things or make sexual remarks that they would not usually have said?

ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPIH5 character 5. Does P talk openly about very personal or private matters not usually discussed in public?

ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPIH6 character 6. Does P take liberties or touch or hug others in a way that is out of character for him/her?

ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPIH7 character 7. Does P show any other signs of loss of control of his/her impulses?

ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPIH8A character 8a. Frequency Ratings

ADNI2: 1=1: Occasionally; 2=2: Often; 3=3: Frequently; 4=4: Very frequently

ADNI3: 1=1: Occasionally;2=2: Often;3=3: Frequently;4=4: Very frequently

ADNI4: 1=1: Occasionally; 2=2: Often; 3=3: Frequently; 4=4: Very frequently

Decoded Value: Yes

NPIH8B character 8b. Severity Ratings

ADNI2, ADNI4: 1=1: Mild; 2=2: Moderate; 3=3: Marked

ADNI3: 1=1: Mild;2=2: Moderate;3=3: Marked

Decoded Value: Yes

NPIH8C character 8c. Caregiver Distress

ADNI2, ADNI4: 0=Not at all; 1=Minimally; 2=Mildly; 3=Moderately; 4=Severely; 5=Very severely or extremely

ADNI3: 0=Not at all;1=Minimally;2=Mildly;3=Moderately;4=Severely;5=Very severely or extremely

Decoded Value: Yes

NPIHTOT numeric H. Disinhibition: Item score

ADNI2, ADNI3: 0..12

ADNI4

NPII character I. Irritability/Lability

ADNI2, ADNI4: 0=No; 1=Yes

ADNI3: 0=No;1=Yes

Decoded Value: Yes

NPII1 character 1. Does P have a bad temper, flying "off the handle" easily over little things?

ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPII2 character 2. Does P rapidly change moods from one to another, being fine one minute and angry the next?

ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPII3 character 3. Does P have sudden flashes of anger?

ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPII4 character 4. Is P impatient, having trouble coping with delays or waiting for planned activities?

ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPII5 character 5. Is P cranky and irritable?

ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPII6 character 6. Is P argumentative and difficult to get along with?

ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPII7 character 7. Does P show any other signs of irritability?

ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPII8A character 8a. Frequency Ratings

ADNI2: 1=1: Occasionally; 2=2: Often; 3=3: Frequently; 4=4: Very frequently

ADNI3: 1=1: Occasionally;2=2: Often;3=3: Frequently;4=4: Very frequently

ADNI4: 1=1: Occasionally; 2=2: Often; 3=3: Frequently; 4=4: Very frequently

Decoded Value: Yes

NPII8B character 8b. Severity Ratings

ADNI2, ADNI4: 1=1: Mild; 2=2: Moderate; 3=3: Marked

ADNI3: 1=1: Mild;2=2: Moderate;3=3: Marked

Decoded Value: Yes

NPII8C character 8c. Caregiver Distress

ADNI2, ADNI4: 0=Not at all; 1=Minimally; 2=Mildly; 3=Moderately; 4=Severely; 5=Very severely or extremely

ADNI3: 0=Not at all;1=Minimally;2=Mildly;3=Moderately;4=Severely;5=Very severely or extremely

Decoded Value: Yes

NPIITOT numeric I. Irritability/Lability: Item score

ADNI2, ADNI3: 0..12

ADNI4

NPIJ character J. Aberrant Motor Behavior

ADNI2, ADNI4: 0=No; 1=Yes

ADNI3: 0=No;1=Yes

Decoded Value: Yes

NPIJ1 character 1. Does P pace around the house without apparent purpose?

ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPIJ2 character 2. Does P rummage around opening and unpacking drawers or closets?

ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPIJ3 character 3. Does P repeatedly put on and take off clothing?

ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPIJ4 character 4. Does P have repetitive activities or "habits" that he/she performs over and over?

ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPIJ5 character 5. Does P engage in repetitive activities such as handling buttons, picking, wrapping string, etc?

ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPIJ6 character 6. Has P fidget excessively, seem unable to sit still, or bounce his/her feet or tap his/her fingers a lot?

ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPIJ7 character 7. Does P do any other activities over and over?

ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPIJ8A character 8a. Frequency Ratings

ADNI2: 1=1: Occasionally; 2=2: Often; 3=3: Frequently; 4=4: Very frequently

ADNI3: 1=1: Occasionally;2=2: Often;3=3: Frequently;4=4: Very frequently

ADNI4: 1=1: Occasionally; 2=2: Often; 3=3: Frequently; 4=4: Very frequently

Decoded Value: Yes

NPIJ8B character 8b. Severity Ratings

ADNI2, ADNI4: 1=1: Mild; 2=2: Moderate; 3=3: Marked

ADNI3: 1=1: Mild;2=2: Moderate;3=3: Marked

Decoded Value: Yes

NPIJ8C character 8c. Caregiver Distress

ADNI2, ADNI4: 0=Not at all; 1=Minimally; 2=Mildly; 3=Moderately; 4=Severely; 5=Very severely or extremely

ADNI3: 0=Not at all;1=Minimally;2=Mildly;3=Moderately;4=Severely;5=Very severely or extremely

Decoded Value: Yes

NPIJTOT numeric J. Aberrant Motor Behavior: Item score

ADNI2, ADNI3: 0..12

ADNI4

NPIK character K. Sleep

ADNI2, ADNI4: 0=No; 1=Yes

ADNI3: 0=No;1=Yes

Decoded Value: Yes

NPIK1 character 1. Does P have difficulty falling asleep?

ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPIK2 character 2. Does P get up during the night (do not count if the patient gets up once or twice per night only to go to the bathroom and falls back asleep immediately)?

ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPIK3 character 3. Does P wander, pace, or get involved in inappropriate activities at night?

ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPIK4 character 4. Does P awaken you during the night?

ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPIK5 character 5. Does P awaken at night, dress, and plan to go out thinking that it is morning and time to start the day?

ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPIK6 character 6. Does P awaken too early in the morning (earlier than was his/her habit)?

ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPIK7 character 7. Does P sleep excessively during the day?

ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPIK8 character 8. Does P have any other night-time behaviors that bother you that we haven't talked about?

ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPIK9A character 9a. Frequency Ratings

ADNI2: 1=1: Occasionally; 2=2: Often; 3=3: Frequently; 4=4: Very frequently

ADNI3: 1=1: Occasionally;2=2: Often;3=3: Frequently;4=4: Very frequently

ADNI4: 1=1: Occasionally; 2=2: Often; 3=3: Frequently; 4=4: Very frequently

Decoded Value: Yes

NPIK9B character 9b. Severity Ratings

ADNI2, ADNI4: 1=1: Mild; 2=2: Moderate; 3=3: Marked

ADNI3: 1=1: Mild;2=2: Moderate;3=3: Marked

Decoded Value: Yes

NPIK9C character 9c. Caregiver Distress

ADNI2, ADNI4: 0=Not at all; 1=Minimally; 2=Mildly; 3=Moderately; 4=Severely; 5=Very severely or extremely

ADNI3: 0=Not at all;1=Minimally;2=Mildly;3=Moderately;4=Severely;5=Very severely or extremely

Decoded Value: Yes

NPIKTOT numeric K. Sleep: Item score

ADNI2, ADNI3: 0..12

ADNI4

NPIL character L. Appetite and eating disorders:

ADNI2, ADNI4: 0=No; 1=Yes

ADNI3: 0=No;1=Yes

Decoded Value: Yes

NPIL1 character 1. Has he/she had a loss of appetite?

ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPIL2 character 2. Has he/she had an increase in appetite?

ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPIL3 character 3. Has he/she had a loss of weight?

ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPIL4 character 4. Has he/she gained weight?

ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPIL5 character 5. Has he/she had a change in eating behavior such as putting too much food in his/her mouth at once?

ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPIL6 character 6. Has he/she had a change in the kind of food he/she likes such as eating too many sweets or other specific types of food?

ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPIL7 character 7. Has he/she developed eating behaviors such as eating exactly the same types of food each day or eating the food in exactly the same order?

ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPIL8 character 8. Have there been any other changes in appetite or eating that I haven't asked about?

ADNI2: 1=Yes; 0=No

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPIL9A character 9a. Frequency Ratings

ADNI2: 1=1: Occasionally; 2=2: Often; 3=3: Frequently; 4=4: Very frequently

ADNI3: 1=1: Occasionally;2=2: Often;3=3: Frequently;4=4: Very frequently

ADNI4: 1=1: Occasionally; 2=2: Often; 3=3: Frequently; 4=4: Very frequently

Decoded Value: Yes

NPIL9B character 9b. Severity Ratings

ADNI2, ADNI4: 1=1: Mild; 2=2: Moderate; 3=3: Marked

ADNI3: 1=1: Mild;2=2: Moderate;3=3: Marked

Decoded Value: Yes

NPIL9C character 9c. Caregiver Distress

ADNI2, ADNI4: 0=Not at all; 1=Minimally; 2=Mildly; 3=Moderately; 4=Severely; 5=Very severely or extremely

ADNI3: 0=Not at all;1=Minimally;2=Mildly;3=Moderately;4=Severely;5=Very severely or extremely

Decoded Value: Yes

NPILTOT numeric L. Appetite and eating disorders: Item score

ADNI2, ADNI3: 0..12

ADNI4

NPITOTAL numeric NPI Total Score

ADNI2, ADNI3: 0..144

ADNI4

ID numeric Record ID Range value: 2, ..., 186208

SITEID numeric Site ID Range value: 1, ..., 941

USERDATE User Create Date Date: YYY-MM-DD

USERDATE2 User Last Modify Date Date: YYY-MM-DD

DD_CRF_VERSION_LABEL character CRF Version Label , with options: v1

LANGUAGE_CODE character Language Code e=English Decoded Value: Yes

HAS_QC_ERROR character Has Quality Check Error 0=Does not have QC error or QC error has been approved; 1=Has QC error Decoded Value: Yes

update_stamp POSIXct: YYY-MM-DD

Author

adni-data@googlegroups.com