Neuropsychiatric Inventory Q

Neuropsychiatric Inventory Q data. More information is available at https://adni.loni.usc.edu/data-samples/adni-data/

Usage

data(NPIQ)

Format

A data frame with 7181 observations and 42 variables.

Source

https://adni.loni.usc.edu/data-samples/adni-data/

Details

ORIGPROT character Original study protocol , with options: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4

COLPROT factor Study protocol of data collection , with factor levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4

PTID character Participant ID

RID numeric Participant roster ID Range value: 2, ..., 10214

VISCODE character Visit code

VISCODE2 character Translated visit code

VISDATE Assessment NPDATE when present; otherwise Registry EXAMDATE Date: YYY-MM-DD

SOURCE character Information Source

ADNIGO, ADNI2: 1=Participant Visit;2=Telephone Call

ADNI3: 1=Participant Visit;2=Telephone Call/Video Call

ADNI4: 1=Participant Visit; 2=Telephone Call/Video Call

Decoded Value: Yes

NPIA character A. Delusions: Does P believe that others are stealing from him/her, or planning to harm him/her in some way?

ADNI1, ADNIGO, ADNI2, ADNI3: 0=No;1=Yes;2=N/A

ADNI1, ADNIGO, ADNI2, ADNI3: 0=No;1=Yes;2=N/A

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPIASEV character A. Delusions: Severity Ratings

ADNI1, ADNIGO, ADNI2, ADNI4: 1=1 - Mild (noticeable, but not a significant change).; 2=2 - Moderate (significant, but not a dramatic change).; 3=3 - Severe (very marked or prominent. A dramatic change).

ADNI3: 1=1 - Mild (noticeable, but not a significant change).;2=2 - Moderate (significant, but not a dramatic change).;3=3 - Severe (very marked or prominent. A dramatic change).

ADNI1, ADNIGO, ADNI2, ADNI4: 1=1 - Mild (noticeable, but not a significant change).; 2=2 - Moderate (significant, but not a dramatic change).; 3=3 - Severe (very marked or prominent. A dramatic change).

Decoded Value: Yes

NPIB character B. Hallucinations: Does P act as if he/she hears voices? Does he/she talk to people who are not there?

ADNI1, ADNIGO, ADNI2, ADNI3: 0=No;1=Yes;2=N/A

ADNI1, ADNIGO, ADNI2, ADNI3: 0=No;1=Yes;2=N/A

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPIBSEV character B. Hallucinations: Severity Ratings

ADNI1, ADNIGO, ADNI2, ADNI4: 1=1 - Mild (noticeable, but not a significant change).; 2=2 - Moderate (significant, but not a dramatic change).; 3=3 - Severe (very marked or prominent. A dramatic change).

ADNI3: 1=1 - Mild (noticeable, but not a significant change).;2=2 - Moderate (significant, but not a dramatic change).;3=3 - Severe (very marked or prominent. A dramatic change).

ADNI1, ADNIGO, ADNI2, ADNI4: 1=1 - Mild (noticeable, but not a significant change).; 2=2 - Moderate (significant, but not a dramatic change).; 3=3 - Severe (very marked or prominent. A dramatic change).

Decoded Value: Yes

NPIC character C. Agitation/Aggression: Is P stubborn and resistive to help from others?

ADNI1, ADNIGO, ADNI2, ADNI3: 0=No;1=Yes;2=N/A

ADNI1, ADNIGO, ADNI2, ADNI3: 0=No;1=Yes;2=N/A

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPICSEV character C. Agitation/Aggression: Severity Ratings

ADNI1, ADNIGO, ADNI2, ADNI4: 1=1 - Mild (noticeable, but not a significant change).; 2=2 - Moderate (significant, but not a dramatic change).; 3=3 - Severe (very marked or prominent. A dramatic change).

ADNI3: 1=1 - Mild (noticeable, but not a significant change).;2=2 - Moderate (significant, but not a dramatic change).;3=3 - Severe (very marked or prominent. A dramatic change).

ADNI1, ADNIGO, ADNI2, ADNI4: 1=1 - Mild (noticeable, but not a significant change).; 2=2 - Moderate (significant, but not a dramatic change).; 3=3 - Severe (very marked or prominent. A dramatic change).

Decoded Value: Yes

NPID character D. Depression/Dysphoria: Does P act as if he/she is sad or in low spirits? Does he/she cry?

ADNI1, ADNIGO, ADNI2, ADNI3: 0=No;1=Yes;2=N/A

ADNI1, ADNIGO, ADNI2, ADNI3: 0=No;1=Yes;2=N/A

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPIDSEV character D. Depression/Dysphoria: Severity Ratings

ADNI1, ADNIGO, ADNI2, ADNI4: 1=1 - Mild (noticeable, but not a significant change).; 2=2 - Moderate (significant, but not a dramatic change).; 3=3 - Severe (very marked or prominent. A dramatic change).

ADNI3: 1=1 - Mild (noticeable, but not a significant change).;2=2 - Moderate (significant, but not a dramatic change).;3=3 - Severe (very marked or prominent. A dramatic change).

ADNI1, ADNIGO, ADNI2, ADNI4: 1=1 - Mild (noticeable, but not a significant change).; 2=2 - Moderate (significant, but not a dramatic change).; 3=3 - Severe (very marked or prominent. A dramatic change).

Decoded Value: Yes

NPIE character E. Anxiety: Does P become upset when separated from you? Does he/she have any other signs of nervousness, such as shortness of breath, sighing, being unable to relax, or feeling excessively tense?

ADNI1, ADNIGO, ADNI2, ADNI3: 0=No;1=Yes;2=N/A

ADNI1, ADNIGO, ADNI2, ADNI3: 0=No;1=Yes;2=N/A

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPIESEV character E. Anxiety: Severity Ratings

ADNI1, ADNIGO, ADNI2, ADNI4: 1=1 - Mild (noticeable, but not a significant change).; 2=2 - Moderate (significant, but not a dramatic change).; 3=3 - Severe (very marked or prominent. A dramatic change).

ADNI3: 1=1 - Mild (noticeable, but not a significant change).;2=2 - Moderate (significant, but not a dramatic change).;3=3 - Severe (very marked or prominent. A dramatic change).

ADNI1, ADNIGO, ADNI2, ADNI4: 1=1 - Mild (noticeable, but not a significant change).; 2=2 - Moderate (significant, but not a dramatic change).; 3=3 - Severe (very marked or prominent. A dramatic change).

Decoded Value: Yes

NPIF character F. Elation/Euphoria: Does P appear to feel too good or act excessively happy?

ADNI1, ADNIGO, ADNI2, ADNI3: 0=No;1=Yes;2=N/A

ADNI1, ADNIGO, ADNI2, ADNI3: 0=No;1=Yes;2=N/A

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPIFSEV character F. Elation/Euphoria: Severity Ratings

ADNI1, ADNIGO, ADNI2, ADNI4: 1=1 - Mild (noticeable, but not a significant change).; 2=2 - Moderate (significant, but not a dramatic change).; 3=3 - Severe (very marked or prominent. A dramatic change).

ADNI3: 1=1 - Mild (noticeable, but not a significant change).;2=2 - Moderate (significant, but not a dramatic change).;3=3 - Severe (very marked or prominent. A dramatic change).

ADNI1, ADNIGO, ADNI2, ADNI4: 1=1 - Mild (noticeable, but not a significant change).; 2=2 - Moderate (significant, but not a dramatic change).; 3=3 - Severe (very marked or prominent. A dramatic change).

Decoded Value: Yes

NPIG character G. Apathy/Indifference: Does P seem less interested in his/her usual activities and in the activities and plans of others?

ADNI1, ADNIGO, ADNI2, ADNI3: 0=No;1=Yes;2=N/A

ADNI1, ADNIGO, ADNI2, ADNI3: 0=No;1=Yes;2=N/A

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPIGSEV character G. Apathy/Indifference: Severity Ratings

ADNI1, ADNIGO, ADNI2, ADNI4: 1=1 - Mild (noticeable, but not a significant change).; 2=2 - Moderate (significant, but not a dramatic change).; 3=3 - Severe (very marked or prominent. A dramatic change).

ADNI3: 1=1 - Mild (noticeable, but not a significant change).;2=2 - Moderate (significant, but not a dramatic change).;3=3 - Severe (very marked or prominent. A dramatic change).

ADNI1, ADNIGO, ADNI2, ADNI4: 1=1 - Mild (noticeable, but not a significant change).; 2=2 - Moderate (significant, but not a dramatic change).; 3=3 - Severe (very marked or prominent. A dramatic change).

Decoded Value: Yes

NPIH character H. Disinhibition: Does P seem to act impulsively? For example, does P talk to strangers as if he/she knows them, or does P say things that may hurt people's feelings?

ADNI1, ADNIGO, ADNI2, ADNI3: 0=No;1=Yes;2=N/A

ADNI1, ADNIGO, ADNI2, ADNI3: 0=No;1=Yes;2=N/A

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPIHSEV character H. Disinhibition: Severity Ratings

ADNI1, ADNIGO, ADNI2, ADNI4: 1=1 - Mild (noticeable, but not a significant change).; 2=2 - Moderate (significant, but not a dramatic change).; 3=3 - Severe (very marked or prominent. A dramatic change).

ADNI3: 1=1 - Mild (noticeable, but not a significant change).;2=2 - Moderate (significant, but not a dramatic change).;3=3 - Severe (very marked or prominent. A dramatic change).

ADNI1, ADNIGO, ADNI2, ADNI4: 1=1 - Mild (noticeable, but not a significant change).; 2=2 - Moderate (significant, but not a dramatic change).; 3=3 - Severe (very marked or prominent. A dramatic change).

Decoded Value: Yes

NPII character I. Irritability/Lability: Is P impatient or cranky? Does he/she have difficulty coping with delays or waiting for planned activities?

ADNI1, ADNIGO, ADNI2, ADNI3: 0=No;1=Yes;2=N/A

ADNI1, ADNIGO, ADNI2, ADNI3: 0=No;1=Yes;2=N/A

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPIISEV character I. Irritability/Lability: Severity Ratings

ADNI1, ADNIGO, ADNI2, ADNI4: 1=1 - Mild (noticeable, but not a significant change).; 2=2 - Moderate (significant, but not a dramatic change).; 3=3 - Severe (very marked or prominent. A dramatic change).

ADNI3: 1=1 - Mild (noticeable, but not a significant change).;2=2 - Moderate (significant, but not a dramatic change).;3=3 - Severe (very marked or prominent. A dramatic change).

ADNI1, ADNIGO, ADNI2, ADNI4: 1=1 - Mild (noticeable, but not a significant change).; 2=2 - Moderate (significant, but not a dramatic change).; 3=3 - Severe (very marked or prominent. A dramatic change).

Decoded Value: Yes

NPIJ numeric J. Aberrant Motor Behavior: Does P engage in repetitive activities, such as pacing around the house, handling buttons, wrapping strings, or doing other things repeatedly?

ADNI1, ADNIGO, ADNI2, ADNI3: 0=No;1=Yes;2=N/A

ADNI1, ADNIGO, ADNI2, ADNI3: 0=No;1=Yes;2=N/A

ADNI4: 0=No; 1=Yes

NPIJSEV character J. Aberrant Motor Behavior: Severity Ratings

ADNI1, ADNIGO, ADNI2, ADNI4: 1=1 - Mild (noticeable, but not a significant change).; 2=2 - Moderate (significant, but not a dramatic change).; 3=3 - Severe (very marked or prominent. A dramatic change).

ADNI3: 1=1 - Mild (noticeable, but not a significant change).;2=2 - Moderate (significant, but not a dramatic change).;3=3 - Severe (very marked or prominent. A dramatic change).

ADNI1, ADNIGO, ADNI2, ADNI4: 1=1 - Mild (noticeable, but not a significant change).; 2=2 - Moderate (significant, but not a dramatic change).; 3=3 - Severe (very marked or prominent. A dramatic change).

Decoded Value: Yes

NPIK character K. Sleep: Does P awaken you during the night, rise too early in the morning, or take excessive naps during the day?

ADNI1, ADNIGO, ADNI2, ADNI3: 0=No;1=Yes;2=N/A

ADNI1, ADNIGO, ADNI2, ADNI3: 0=No;1=Yes;2=N/A

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPIKSEV character K. Sleep: Severity Ratings

ADNI1, ADNIGO, ADNI2, ADNI4: 1=1 - Mild (noticeable, but not a significant change).; 2=2 - Moderate (significant, but not a dramatic change).; 3=3 - Severe (very marked or prominent. A dramatic change).

ADNI3: 1=1 - Mild (noticeable, but not a significant change).;2=2 - Moderate (significant, but not a dramatic change).;3=3 - Severe (very marked or prominent. A dramatic change).

ADNI1, ADNIGO, ADNI2, ADNI4: 1=1 - Mild (noticeable, but not a significant change).; 2=2 - Moderate (significant, but not a dramatic change).; 3=3 - Severe (very marked or prominent. A dramatic change).

Decoded Value: Yes

NPIL character L. Appetite: Has P lost or gained weight, or had a change in the food he/she likes?

ADNI1, ADNIGO, ADNI2, ADNI3: 0=No;1=Yes;2=N/A

ADNI1, ADNIGO, ADNI2, ADNI3: 0=No;1=Yes;2=N/A

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

NPILSEV character L. Appetite: Severity Ratings

ADNI1, ADNIGO, ADNI2, ADNI4: 1=1 - Mild (noticeable, but not a significant change).; 2=2 - Moderate (significant, but not a dramatic change).; 3=3 - Severe (very marked or prominent. A dramatic change).

ADNI3: 1=1 - Mild (noticeable, but not a significant change).;2=2 - Moderate (significant, but not a dramatic change).;3=3 - Severe (very marked or prominent. A dramatic change).

ADNI1, ADNIGO, ADNI2, ADNI4: 1=1 - Mild (noticeable, but not a significant change).; 2=2 - Moderate (significant, but not a dramatic change).; 3=3 - Severe (very marked or prominent. A dramatic change).

Decoded Value: Yes

NPISCORE numeric NPI-Q Total Score

ADNI1, ADNIGO, ADNI2, ADNI3: 0..36

ADNI4

SPID numeric Study Partner ID: Range value: 1, ..., 3

ID numeric Record ID Range value: 4, ..., 186247

SITEID numeric Site ID Range value: 1, ..., 941

USERDATE User Create Date Date: YYY-MM-DD

USERDATE2 User Last Modify Date Date: YYY-MM-DD

DD_CRF_VERSION_LABEL character CRF Version Label , with options: v1

LANGUAGE_CODE character Language Code e=English Decoded Value: Yes

HAS_QC_ERROR character Has Quality Check Error 0=Does not have QC error or QC error has been approved; 1=Has QC error Decoded Value: Yes

update_stamp POSIXct: YYY-MM-DD

Author

adni-data@googlegroups.com