Neuropathology Status

Neuropathology Status data. More information is available at https://adni.loni.usc.edu/data-samples/adni-data/

Usage

data(NPSTATUS)

Format

A data frame with 11597 observations and 23 variables.

Source

https://adni.loni.usc.edu/data-samples/adni-data/

Details

ORIGPROT character Original study protocol , with options: ADNI1, ADNI2, ADNIGO, ADNI3, ADNI4

COLPROT factor Study protocol of data collection , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4

PTID character Participant ID

RID numeric Participant roster ID Range value: 2, ..., 10900

VISCODE character Visit code

VISCODE2 character Translated visit code

VISDATE Assessment NPSDATE when present; otherwise Registry EXAMDATE Date: YYY-MM-DD

NPDECIDE character Has a definitive decision about brain donation been made?

ADNI2: 1=Yes ; 0=No

ADNI3, ADNI4: 0=No ; 1=Yes

Decoded Value: Yes

NPDEC character If a decision has been made, has the participant agreed to register for brain donation?

ADNI2, ADNI3: 1=No - provide reason for nonparticipation ; 2=Yes - signed provisional consent form on file ; 3=Yes - consent form being reviewed ; 4=Yes - consent form to be completed at time of autopsy ; 5=Other

ADNI4: 1=No – provide reason for nonregistration ; 2=Yes – signed registration form on file ; 3=Yes – registration form being reviewed ; 4=Yes – registration form to be completed at time of autopsy

Decoded Value: Yes

NPDECNO character If No, select reason(s) for nonregistration.

ADNI2

ADNI3, ADNI4: 1=Not interested, personal choice, or not comfortable. ; 2=Family does not want autopsy, participant does not want to burden family after death ; 3=Religious or other beliefs ; 4=Wants body buried/cremated intact, does not want body altered after death ; 99=Other (specify)

Decoded Value: Yes

NPDECCF character If there is agreement to register, is the agreement for autopsy for a study other than ADNI?

ADNI2: 1=Yes ; 0=No

ADNI3, ADNI4: 0=No ; 1=Yes

Decoded Value: Yes

NPUNDEC character If participant is undecided, have they been given information (brochure, discussion with clinician) regarding autopsy/brain donation?

ADNI2: 1=Yes ; 0=No

ADNI3, ADNI4: 0=No ; 1=Yes

Decoded Value: Yes

NPCONS character Was registration for brain donation discussed at this visit? 0=No ; 1=Yes Decoded Value: Yes

NPCONSNO character If no, why not:

ADNI3: 1=Site error ; 2=Previous definitive no response from participant ; 3=Site not participating in brain donation ; 99=Other reason(s)

ADNI4: 1=Site error ; 2=Previous definitive "no" response from participant ; 3=Site not participating in brain donation (prior approval from ADNI leadership required) ; 99=Other reason(s)

Decoded Value: Yes

NPSOURC character Information source 1=Participant visit ; 2=Telephone call Decoded Value: Yes

ID numeric Record ID Range value: 13, ..., 187108

SITEID numeric Site ID Range value: 1, ..., 941

USERDATE User Create Date Date: YYY-MM-DD

USERDATE2 User Last Modify Date Date: YYY-MM-DD

DD_CRF_VERSION_LABEL character CRF Version Label , with options: v1, v2

LANGUAGE_CODE character Language Code e=English Decoded Value: Yes

HAS_QC_ERROR character Has Quality Check Error 0=Does not have QC error or QC error has been approved ; 1=Has QC error Decoded Value: Yes

update_stamp POSIXct: YYY-MM-DD