[DERIVED] Derived Modified Preclinical Alzheimer’s Cognitive Composite (PACC) Score

Modified preclinical alzheimer’s cognitive composite (pacc) score derived dataset.

Usage

data(PACC)

Format

A data frame with 19509 observations and 26 variables.

Source

For more details see the help vignette: vignette(topic = 'ADNIMERGE2-PACC', package = 'ADNIMERGE2')

Details

ORIGPROT character Original Study Protocol First time participation in ADNI study phase; , with character options: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4

COLPROT character Study Data Collection Protocol , with character options: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4

RID numeric Subject Identifier Range value: 1, ..., 10898

VISCODE character Study visit code

ADASQ4SCORE numeric Delayed Word Recall Score - portion of ADAS-Cog Range value: 0, ..., 10

ADASQ4SCORE_VISDATE character ADAS-Cog Collection/Visit Date

MMSE numeric Mini-Mental Examination Total Score Range value: 0, ..., 30

MMSE_VISDATE character MMSE Collection/Visit Date

LDELTOTL numeric Logical Memory - Delayed RecallScore Range value: 0, ..., 25

LDELTOTL_VISDATE character LDELTOTL Collection/Visit Date

DIGITSCR numeric Digit Symbol Substitution Score Range value: 0, ..., 83

DIGITSCR_VISDATE character DIGITSCR Collection/Visit Date

TRABSCOR numeric Trial B score Range value: 0, ..., 996

LOG.TRABSCOR numeric Trial B score - Log Scale Range value: 0, ..., 6.90475076996184

TRABSCOR_VISDATE character TRABSCOR Collection/Visit Date

PACC_VISDATE Date PACC Collection/Visit Date Either common date across non-missing PACC component collection date, the maximum relative date or the corresponding study visit date when all component collection date is not available.; Date: YYY-MM-DD

ADASQ4SCOREZ numeric ADAS Q4 Standardized Score Standardized by corresponding baseline score among cognitive normal (CN) new enrollee in ADNI study. Also adjusted for directional effect (i.e. multiplied by -1); Range value: -3.7512839877574, ..., 1.47880871040558

MMSEZ numeric Mini-Mental State Examiniation Standardized Score Standardized by corresponding baseline score among cognitive normal (CN) new enrollee in ADNI study.; Range value: -24.846889943961, ..., 0.835534887861971

LDELTOTLZ numeric Logical Memory - Delayed Recall Standardized Score Standardized by corresponding baseline score among cognitive normal (CN) new enrollee in ADNI study.; Range value: -3.6602419737235, ..., 3.51113470972132

DIGITSCRZ numeric Digit Symbol Substitution Standardized Score Standardized by corresponding baseline score among cognitive normal (CN) new enrollee in ADNI1 study phase.; Range value: -4.48438541990256, ..., 3.65180778861111

LOG.TRABSCORZ numeric Logical Memory - Delayed Recall Log Standardized Score Standardized by corresponding log transformed baseline score among cognitive normal (CN) new enrollee in ADNI study. Also adjusted for directional effect (i.e. multiplied by -1); Range value: -6.37019686085191, ..., 10.6080165632692

mPACCdigit numeric Modified PACC Digit Score Range value: -57.196348, ..., 7.622008

mPACCtrailsB numeric Modified PACC Trials B Score Range value: -57.196348, ..., 12.962368

ENRLDT Date Enrollment Date Date: YYY-MM-DD

ENRLFG character Enrollment Flag , with character options: Yes

DX character Baseline Diagnostics Status Either baseline or screening diagnostics status and require to use enrollment flag.; , with character options: CN, Dementia, MCI

Author

adni-data@googlegroups.com