PET Scan Information

PET Scan Information data. More information is available at https://adni.loni.usc.edu/data-samples/adni-data/

Usage

data(PETMETA_ADNI1)

Format

A data frame with 1957 observations and 105 variables.

Source

https://adni.loni.usc.edu/data-samples/adni-data/

Details

ORIGPROT character Original study protocol , with options: ADNI1

COLPROT factor Study protocol of data collection , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4

ID numeric Range value: 2, ..., 3958

RID numeric Participant roster ID Range value: 2, ..., 1435

PTID character Participant ID

SITEID numeric Site ID Range value: 1, ..., 127

VISCODE character Visit code , with options: bl, m06, m12, m18, m24, m36, m48

VISCODE2 character , with options: bl, m06, m12, m18, m24, m36, m48

USERDATE Date record created Date: YYY-MM-DD

USERDATE2 Date record last updated Date: YYY-MM-DD

RECNO numeric Range value: 1, ..., 2

PMCONDCT character Was the scan conducted? 1=Yes ; 0=No Decoded Value: Yes

PMREASON character Reason why the scan was not conducted: 1=Illness ; 2=Participant unavailable ; 3=Participant unwilling ; 4=Administrative problems ; 5=Withdrawn consent ; 6=Other (specify) Decoded Value: Yes

PMCONSENT character Did the participant consent to quantitative scan? 1=Yes ; 0=No Decoded Value: Yes

PMTYPE character Type of scan conducted 1=Qualitative ; 2=Quantitative Decoded Value: Yes

EXAMDATE Scan Date Date: YYY-MM-DD

PMSCANNER character GE 1=check here Decoded Value: Yes

PMGEMODEL character If GE, Scanner Model: 1=Advance ; 2=Discovery LS ; 3=Discovery ST ; 4=Discovery RX ; 5=Discovery STE Decoded Value: Yes

PMSIEMENS character Siemens/CTI 1=check here Decoded Value: Yes

PMSIEMODEL character If Siemens/CTI, Scanner Model: 1=ACCEL ; 2=Biograph ; 3=BioGraph BGO ; 4=BioGraph HiRez ; 5=EXACT ; 6=HR+ ; 7=HRRT Decoded Value: Yes

PMPHILLIPS character Phillips 1=check here Decoded Value: Yes

PMPHMODEL character If Phillips, Scanner Model: 1=Allegro ; 2=Allegro-Neuro ; 3=Gemini ; 4=Gemini GLX ; 5=Gemini-TF Decoded Value: Yes

PMQCTIME character Time of today's Scanner QC

PMBLTIME character Time of blood glucose measurement

PMBLGLUC numeric Blood Glucose (pre-FDG) Range value: 50, ..., 250

PMFDGTIME character Time of FDG dose assay

PMFDGDOS numeric FDG dose assay Range value: 3.8, ..., 54

PMFDGVOL numeric FDG Volume Range value: 0.1, ..., 10

PMINJTIME character Time of FDG injection

PMSCTIME character Time scan started (emission)

PMVARIAT character Any variations from protocol during FDG uptake? 1=Yes ; 0=No Decoded Value: Yes

PMPROTID character Predefined acquisition protocol ID

PMFRAME character Which framing rate was used? 1=6 frames, 5 min/frame (6x300s) ; 2=1 frame, 30 min (1x1800s) ; 3=Quantitative Decoded Value: Yes

PMMOTION character Subject motion problems: 1=Yes ; 0=No Decoded Value: Yes

PMMALFUN character Scanner malfunction 1=Yes ; 0=No Decoded Value: Yes

PMOTHER character Other protocol variations: 1=Yes ; 0=No Decoded Value: Yes

PMRECON character Check which of the following reconstructions was used: 2=FORE/2D-OSEM ; 4=OSEM3D-OP ; 1=3D-Ramla ; 3=3D Back-projection Decoded Value: Yes

PMSUBSET character # subsets: 14=14 ; 16=16 ; 98=N/A ; 99=Other Decoded Value: Yes

PMSUBSPE numeric If Other, specify Range value: 0, ..., 92

PMITERAT character # iterations: 2=2 ; 4=4 ; 6=6 ; 99=Other Decoded Value: Yes

PMITERSPE numeric If Other, specify: Range value: 0, ..., 16

PMRAMLA character If Ramla, Lambda=0.016? 1=Check here to confirm Decoded Value: Yes

PMFILTER character If Back Projection, Ramp filter? 1=Check here to confirm Decoded Value: Yes

PMMODEON character If FORE/2D-OSEM, Brain Mode "ON" for PET only scanners or TRIM "ON" for PET/CT scanners? 1=Check here to confirm Decoded Value: Yes

PMSMOOTH character No post-process smoothing: 1=Check here to confirm Decoded Value: Yes

PMDECAY character Decay Correction 1=Yes ; 0=No Decoded Value: Yes

PMSCATTR character Scatter Correction: 1=Yes ; 0=No Decoded Value: Yes

PMATTEN character Attenuation Correction: 1=CT ; 2=Ge-68+Segmentation ; 3=Cs-137+Segmentation Decoded Value: Yes

PMWITHDRAW character Clock for blood sample withdrawal time 1=Yes ; 0=No Decoded Value: Yes

PMWITHTIME character If No, provide the time difference , with options: offset seems to be 44 sec (16:01:11 vs 16:01:55), 39 seconds, ~50 sec off, Subject was a hard stick, Nurse had trouble drawing blood

PMCOUNT character Clock for blood sample count time 1=Yes ; 0=No Decoded Value: Yes

PMCOUNTIME character If No, provide the time difference , with options: 39 seconds, ~50 sec off

PMBG1TIME character Sample Count Time (24h) Background #1

PMBG1VOL numeric Sample Plasma Volume Counted Background #1 Range value: 200, ..., 250

PMBG1DUR numeric Sample Count Duration Background #1 Range value: 60, ..., 300

PMBG1COUNT numeric Sample Count Rate Background #1 Range value: 1, ..., 279

PMP1DTIME character Sample Draw Time (24h) Plasma 1

PMP1CTIME character Sample Count Time (24h) Plasma 1

PMP1BGL numeric Sample BGL Plasma 1 Range value: 63, ..., 168

PMP1VOL numeric Sample Plasma Volume Counted Plasma 1 Range value: 200, ..., 250

PMP1DUR numeric Sample Count Duration Plasma 1 Range value: 25, ..., 60

PMP1COUNT numeric Sample Count Rate Plasma 1 Range value: 3, ..., 9999

PMP2DTIME character Sample Draw Time (24h) Plasma 2

PMP2CTIME character Sample Count Time (24h) Plasma 2

PMP2BGL numeric Sample BGL Plasma 2 Range value: 66, ..., 164

PMP2VOL numeric Sample Plasma Volume Counted Plasma 2 Range value: 111, ..., 250

PMP2DUR numeric Sample Count Duration Plasma 2 Range value: 11, ..., 60

PMP2COUNT numeric Sample Count Rate Plasma 2 Range value: 8, ..., 9999

PMP3DTIME character Sample Draw Time (24h) Plasma 3

PMP3CTIME character Sample Count Time (24h) Plasma 3

PMP3BGL numeric Sample BGL Plasma 3 Range value: 53, ..., 165

PMP3VOL numeric Sample Plasma Volume Counted Plasma 3 Range value: 60, ..., 250

PMP3DUR numeric Sample Count Duration Plasma 3 Range value: 29, ..., 200

PMP3COUNT numeric Sample Count Rate Plasma 3 Range value: 7, ..., 9999

PMP4DTIME character Sample Draw Time (24h) Plasma 4

PMP4CTIME character Sample Count Time (24h) Plasma 4

PMP4BGL numeric Sample BGL Plasma 4 Range value: 63, ..., 168

PMP4VOL numeric Sample Plasma Volume Counted Plasma 4 Range value: 99, ..., 250

PMP4DUR numeric Sample Count Duration Plasma 4 Range value: 30, ..., 99

PMP4COUNT numeric Sample Count Rate Plasma 4 Range value: 6, ..., 9999

PMP5DTIME character Sample Draw Time (24h) Plasma 5

PMP5CTIME character Sample Count Time (24h) Plasma 5

PMP5BGL numeric Sample BGL Plasma 5 Range value: 60, ..., 177

PMP5VOL numeric Sample Plasma Volume Counted Plasma 5 Range value: 99, ..., 250

PMP5DUR numeric Sample Count Duration Plasma 5 Range value: 39, ..., 99

PMP5COUNT numeric Sample Count Rate Plasma 5 Range value: 1, ..., 9999

PMBG2CTIME character Sample Count Time (24h) Background 2

PMBG2VOL numeric Sample Plasma Volume Counted Background 2 Range value: 200, ..., 250

PMBG2DUR numeric Sample Count Duration Background 2 Range value: 1, ..., 300

PMBG2COUNT numeric Sample Count Rate Background 2 Range value: 1, ..., 283

PMPIPVOL character Was the pipetted plasma volume 200 uL? 1=Yes ; 0=No Decoded Value: Yes

PMPIPVOLSP character If No, denote volume used: , with options: 250 microL - have only fixed vol. pipettes, 250 uL fixed volume pipette, 250 micro liter (fixed volume pipette), 250 microliters, 250 microL fixed volume pipette, 250 microliter fixed volume pipette, 250 ul fixed volume pipette, 250; fixed volume pipette, 250 uL fixed volume pipette, 250 uL fixed volume pipette

PMCTIME character Was the plasma sample count time 1 minute? 1=Yes ; 0=No Decoded Value: Yes

PMCTIMESP character If No, denote count time used , with options: fixed counts 10000: 25-39 sec

PMPHACT numeric Phantom Activity at Time of Scan Range value: 0.25, ..., 3

PMPHVOL numeric Phantom Volume Range value: 5969, ..., 8538

PMPHCOUNT numeric Average Counts from Phantom Image ROI Range value: 5.7e-05, ..., 11264

PMALIQVOL numeric Aliquot Volume Range value: 50, ..., 1500

PMALIQACT numeric Aliquot Count Rate Range value: 137, ..., 6138

PMTRNSFR character Was data transferred to LONI within 24 hours of scan? 1=Yes ; 0=No Decoded Value: Yes

PMTRNDATE Transfer Date Date: YYY-MM-DD

PMARCHIVE character Data Archived Locally 1=Yes ; 0=No Decoded Value: Yes

PMLPDONE character Was a Lumbar Puncture completed prior to the PET scan? 1=Yes ; 0=No Decoded Value: Yes

PMLPINTER character If Yes, What was the interval between LP and PET? 1=less than 6 hours ; 2=6-12 hours ; 3=13-24 hours ; 4=25-48 hours ; 5=49-72 hours ; 6=more than 72 hours Decoded Value: Yes

update_stamp POSIXct: YYY-MM-DD

Author

adni-data@googlegroups.com