PIB Scan Information

PIB Scan Information data. More information is available at https://adni.loni.usc.edu/data-samples/adni-data/

Usage

data(PIBMETA)

Format

A data frame with 257 observations and 76 variables.

Source

https://adni.loni.usc.edu/data-samples/adni-data/

Details

ORIGPROT character Original study protocol , with options: ADNI1

ID numeric Range value: 2, ..., 520

RID numeric Participant roster ID Range value: 57, ..., 1430

SITEID numeric Site ID Range value: 1, ..., 124

VISCODE character Visit code , with options: m12, bl, m24, m36, m48

USERDATE Date record created Date: YYY-MM-DD

USERDATE2 Date record last updated Date: YYY-MM-DD

PBCONDCT numeric Was the scan conducted? 1=Yes ; 0=No

PBREASON numeric Reason why the scan was not conducted: 1=Illness ; 2=Participant unavailable ; 3=Participant unwilling ; 4=Administrative problems ; 5=Withdrawn consent ; 6=Other (specify)

PBOTHSPE character If Other, specify:

EXAMDATE Scan Date Date: YYY-MM-DD

PBGE numeric GE 1=check here

PBGEMODEL numeric If GE, Scanner Model: 1=Advance ; 2=Discovery LS ; 3=Discovery ST ; 4=Discovery RX ; 5=Discovery STE

PBSIEMENS numeric Siemens/CTI 1=check here

PBSIEMODEL numeric If Siemens/CTI, Scanner Model: 1=ACCEL ; 2=Biograph ; 3=BioGraph BGO ; 4=BioGraph HiRez ; 5=EXACT ; 6=HR+ ; 7=HRRT

PBPHILLIPS numeric Phillips 1=check here

PBPHMODEL numeric If Phillips, Scanner Model: 1=Allegro ; 2=Allegro-Neuro ; 3=Gemini ; 4=Gemini GLX ; 5=Gemini-TF

PBQCTIME character Time of today's Scanner QC

PBPIBTIME character Time of PIB dose assay

PBPIBDOS numeric PIB dose assay Range value: 5.8, ..., 29.8

PBRESTIME character Time of residual PIB assay

PBRESID numeric Residual left in syringe Range value: 0.01, ..., 72

PBINJDOS numeric Net injected dose of PIB Range value: 5, ..., 17.1

PBPIBVOL numeric PIB volume Range value: 3, ..., 17

PBINJTIME character Time of PIB injection

PBSCTIME character Time scan started (emission)

PBSCANCOM character Provide an explanation if start time is not within the allowable window

PBVARIAT numeric Any variations from protocol during PIB uptake? 1=Yes ; 0=No

PBVARSP character If Yes, describe: , with options: 9.4 mCi injected, Subject had to go to bathroom, Patient ate at 11:30, LOW PIB DOSE SYNTHESIZED, Participant fell asleep, scan not started in window., dose<15+/-1.5 mci due to Qc time of isotope and increased time setting upsubject in scanner.

PBPROTID character Predefined acquisition protocol ID

PBSTATDYN numeric Indicate whether scan was static or dynamic: 1=Static (1 x 20 min) ; 2=Standard Dynamic (4 x 5 min) ; 3=Dynamic (specify)

PBNMFRAM1 numeric 1. No. of Frames Range value: 4, ..., 8

PBDURATN1 numeric Duration: Range value: 4, ..., 300

PBNMFRAM2 numeric 2. No. of Frames: Range value: 0, ..., 12

PBDURATN2 numeric Duration: Range value: 0, ..., 300

PBNMFRAM3 numeric 3. No. of Frames: Range value: 0, ..., 27

PBDURATN3 numeric Duration: Range value: 0, ..., 120

PBNMFRAM4 numeric 4. No. of Frames: Range value: 0, ..., 16

PBDURATN4 numeric Duration: Range value: 0, ..., 300

PBNMFRAM5 numeric 5. No. of Frames: Range value: 0, ..., 8

PBDURATN5 numeric Duration: Range value: 0, ..., 300

PBNMFRAM6 numeric 6. No. of Frames: Range value: 0, ..., 6

PBDURATN6 numeric Duration: Range value: 0, ..., 300

PBNMFRAM7 numeric 7. No. of Frames: Range value: 0, ..., 16

PBDURATN7 numeric Duration: Range value: 0, ..., 300

PBNMFRAM8 numeric 8. No. of Frames: Range value: 0, ..., 0

PBDURATN8 numeric Duration: Range value: 0, ..., 0

PBNMFRAM9 numeric 9. No. of Frames: Range value: 0, ..., 0

PBDURATN9 numeric Duration: Range value: 0, ..., 0

PBNMFRAM10 numeric 10. No. of Frames: Range value: 0, ..., 0

PBDURATN10 numeric Duration: Range value: 0, ..., 0

PBMOTION numeric Subject motion problems: 1=Yes ; 0=No

PBMOTSP character If yes, describe:

PBMALFUN numeric Scanner malfunction 1=Yes ; 0=No

PBMALSP character If yes, describe: , with options: error with ACS creating and reading file

PBOTHER numeric Other protocol variations: 1=Yes ; 0=No

PBOTHSP character If yes, describe: , with options: Only 2 five minute frames were performed, only 2 5 minute frames were done according to the ADNI PIB protocol, A transmission scan was not performed therefore the scan was attenuation corrected using drawn (eliptical) regions of interest over the entire head (outside the skull)., The injected dose of C-11 PIB was 8.2 mCi instead of recommended dose of >10 mCi., MCH 60 min multiframe acquisition used. Tech advised to use PIB ADNI acquisition protocol, Wrong Imaging Parameters were used, subject injected with only 13.1 mCi because of delay due to scanner error, Placed a lot of towels under the neck, to make patient comfortablt, at 70min into scan - subject put hand into scanner, less than +/-1.5mci of desired dose injected

PBRECON numeric Check which of the following reconstructions was used: 2=FORE/2D-OSEM ; 4=OSEM3D-OP ; 1=3D-Ramla ; 3=3D Back-projection

PBSUBSET numeric # subsets: 14=14 ; 16=16 ; 98=N/A ; 99=Other

PBSUBSPE numeric If Other, specify Range value: Inf, ..., -Inf

PBITERAT numeric # iterations: 2=2 ; 4=4 ; 6=6 ; 99=Other

PBITERSPE numeric If Other, specify: Range value: Inf, ..., -Inf

PBRAMLA numeric If Ramla, Lambda=0.016? 1=Check here to confirm

PBFILTER numeric If Back Projection, Ramp filter? 1=Check here to confirm

PBMODEON numeric If FORE/2D-OSEM, Brain Mode "ON" for PET only scanners or TRIM "ON" for PET/CT scanners? 1=Check here to confirm

PBSMOOTH numeric No post-process smoothing: 1=Check here to confirm

PBDECAY numeric Decay Correction 1=Yes ; 0=No

PBSCATTR numeric Scatter Correction: 1=Yes ; 0=No

PBMATTEN numeric Attenuation Correction: 1=CT ; 2=Ge-68+Segmentation ; 3=Cs-137+Segmentation

PBTRSNFR numeric Was data transferred to LONI within 24 hours of scan? 1=Yes ; 0=No

PBTRNDATE character Transfer Date

PBTRNCOM character Comments

PBARCHIVE numeric Data Archived Locally 1=Yes ; 0=No

PBARCMED character Archive Medium

PBARCCOM character Comments

update_stamp POSIXct: YYY-MM-DD

Author

adni-data@googlegroups.com