ASHS volume data: 2022-02-15

ASHS volume data: 2022-02-15 data. More information is available at https://adni.loni.usc.edu/data-samples/adni-data/

Usage

data(PICSLASHS)

Format

A data frame with 1978 observations and 59 variables.

Source

https://adni.loni.usc.edu/data-samples/adni-data/

Details

ORIGPROT character Original study protocol , with options: ADNI1, ADNIGO, ADNI2, ADNI3

COLPROT factor Study protocol of data collection , with factor levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4

RID numeric roster participant ID Range value: 21, ..., 6971

VISCODE character visit code

VISCODE2 character visit code translated

EXAMDATE Date Date of MRI scan Date: YYY-MM-DD

ICV numeric Intercranial volume derived from T1 image Range value: 999281, ..., 1976566

SLICE_THICKNESS numeric Coronal slice thickness Range value: 1.99923, ..., 2.2

LEFT_CA1_VOL numeric left hemisphere ROI CA1 volume Range value: 43.03, ..., 5584.1

LEFT_CA1_NS numeric Number of coronal slices in which left hemisphere ROI CA1 appears Range value: 2, ..., 26

LEFT_CA2_VOL numeric left hemisphere ROI CA2 volume Range value: 0, ..., 7.935

LEFT_CA2_NS numeric Number of coronal slices in which left hemisphere ROI CA2 appears Range value: 0, ..., 9

LEFT_CA3_VOL numeric left hemisphere ROI CA3 volume Range value: 0, ..., 126.342

LEFT_CA3_NS numeric Number of coronal slices in which left hemisphere ROI CA3 appears Range value: 0, ..., 11

LEFT_DG_VOL numeric left hemisphere ROI dentate gyrus volume Range value: 8.85, ..., 2496.64

LEFT_DG_NS numeric Number of coronal slices in which left hemisphere ROI dentate gyrus appears Range value: 1, ..., 23

LEFT_MISC_VOL numeric left hemisphere ROI MISC volume Range value: 0, ..., 793.426

LEFT_MISC_NS numeric Number of coronal slices in which left hemisphere ROI MISC appears Range value: 0, ..., 23

LEFT_SUB_VOL numeric left hemisphere ROI Subiculum volume Range value: 0, ..., 1117.55

LEFT_SUB_NS numeric Number of coronal slices in which left hemisphere ROI Subiculum appears Range value: 0, ..., 24

LEFT_ERC_VOL numeric left hemisphere ROI Entorhinal cortex volume Range value: 0, ..., 659.79

LEFT_ERC_NS numeric Number of coronal slices in which left hemisphere ROI Entorhinal cortex appears Range value: 0, ..., 15

LEFT_BA35_VOL numeric left hemisphere ROI BA35 volume Range value: 0, ..., 890.49

LEFT_BA35_NS numeric Number of coronal slices in which left hemisphere ROI BA35 appears Range value: 0, ..., 16

LEFT_BA36_VOL numeric left hemisphere ROI BA36 volume Range value: 0, ..., 4696.96

LEFT_BA36_NS numeric Number of coronal slices in which left hemisphere ROI BA36 appears Range value: 0, ..., 16

LEFT_PHC_VOL numeric left hemisphere ROI parahippocampal cortex volume Range value: 0, ..., 5310.66

LEFT_PHC_NS numeric Number of coronal slices in which left hemisphere ROI parahippocampal cortex appears Range value: 0, ..., 11

LEFT_SULCUS_VOL numeric left hemisphere ROI sulcus volume Range value: 0, ..., 1148.15

LEFT_SULCUS_NS numeric Number of coronal slices in which left hemisphere ROI sulcus appears Range value: 0, ..., 23

LEFT_CA_VOL numeric left hemisphere ROI CA (CA1+CA2+CA3) volume Range value: 43.03, ..., 5629.27

LEFT_CA_NS numeric Number of coronal slices in which left hemisphere ROI CA (CA1+CA2+CA3) appears Range value: 2, ..., 26

RIGHT_CA1_VOL numeric right hemisphere ROI CA1 volume Range value: 0, ..., 6692.52

RIGHT_CA1_NS numeric Number of coronal slices in which right hemisphere ROI CA1 appears Range value: 0, ..., 24

RIGHT_CA2_VOL numeric right hemisphere ROI CA2 volume Range value: 0, ..., 18.616

RIGHT_CA2_NS numeric Number of coronal slices in which right hemisphere ROI CA2 appears Range value: 0, ..., 11

RIGHT_CA3_VOL numeric right hemisphere ROI CA3 volume Range value: 0, ..., 142.822

RIGHT_CA3_NS numeric Number of coronal slices in which right hemisphere ROI CA3 appears Range value: 0, ..., 10

RIGHT_DG_VOL numeric right hemisphere ROI dentate gyrus volume Range value: 0, ..., 5831.61

RIGHT_DG_NS numeric Number of coronal slices in which right hemisphere ROI dentate gyrus appears Range value: 0, ..., 22

RIGHT_MISC_VOL numeric right hemisphere ROI MISC volume Range value: 0, ..., 552.064

RIGHT_MISC_NS numeric Number of coronal slices in which right hemisphere ROI MISC appears Range value: 0, ..., 25

RIGHT_SUB_VOL numeric right hemisphere ROI Subiculum volume Range value: 0, ..., 2147.22

RIGHT_SUB_NS numeric Number of coronal slices in which right hemisphere ROI Subiculum appears Range value: 0, ..., 23

RIGHT_ERC_VOL numeric right hemisphere ROI Entorhinal cortex volume Range value: 0, ..., 2073.67

RIGHT_ERC_NS numeric Number of coronal slices in which right hemisphere ROI Entorhinal cortex appears Range value: 0, ..., 15

RIGHT_BA35_VOL numeric right hemisphere ROI BA35 volume Range value: 0, ..., 4026.19

RIGHT_BA35_NS numeric Number of coronal slices in which right hemisphere ROI BA35 appears Range value: 0, ..., 16

RIGHT_BA36_VOL numeric right hemisphere ROI BA36 volume Range value: 0, ..., 8778.09

RIGHT_BA36_NS numeric Number of coronal slices in which right hemisphere ROI BA36 appears Range value: 0, ..., 16

RIGHT_PHC_VOL numeric right hemisphere ROI parahippocampal cortex volume Range value: 0, ..., 5197.76

RIGHT_PHC_NS numeric Number of coronal slices in which right hemisphere ROI parahippocampal cortex appears Range value: 0, ..., 10

RIGHT_SULCUS_VOL numeric right hemisphere ROI sulcus volume Range value: 0, ..., 2192.08

RIGHT_SULCUS_NS numeric Number of coronal slices in which right hemisphere ROI sulcus appears Range value: 0, ..., 23

RIGHT_CA_VOL numeric right hemisphere ROI CA (CA1+CA2+CA3) volume Range value: 0, ..., 6752.63

RIGHT_CA_NS numeric Number of coronal slices in which right hemisphere ROI CA (CA1+CA2+CA3) appears Range value: 0, ..., 24

T2_QC_R character manual QC score for right hemisphere image quality -1=no image;0=Unusable (substructures and the SRLM are not identifiable) or image has motion and looks blurry/wiggly;1=Bad Quality (SRLM is hard to identify and substructures are blurry) or image has motion;2=Adequate Quality (some substructures are clear, SRLM is mostly visible);3=intermediate quality between 2 & 4;4=Good Quality (most substructures are clear, SRLM is continuously visible) Decoded Value: Yes

T2_QC_L character manual QC score for left hemisphere image quality -1=no image;0=Unusable (substructures and the SRLM are not identifiable) or image has motion and looks blurry/wiggly;1=Bad Quality (SRLM is hard to identify and substructures are blurry) or image has motion;2=Adequate Quality (some substructures are clear, SRLM is mostly visible);3=intermediate quality between 2 & 4;4=Good Quality (most substructures are clear, SRLM is continuously visible) Decoded Value: Yes

update_stamp POSIXct POSIXct: YYY-MM-DD

Author

adni-data@googlegroups.com