ASHS volume data

ASHS volume data data. More information is available at https://adni.loni.usc.edu/data-samples/adni-data/

Usage

data(PICSL_ASHS)

Format

A data frame with 2284 observations and 68 variables.

Source

https://adni.loni.usc.edu/data-samples/adni-data/

Details

ORIGPROT character Original study protocol , with options: ADNI1, ADNIGO, ADNI2, ADNI3

COLPROT factor Study protocol of data collection , with factor levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4

RID numeric roster participant ID Range value: 21, ..., 7123

PTID character Site and Unique participant ID

VISCODE character visit code , with options: init, y1, y2, y3, y4, y5, bl

VISCODE2 character visit code translated

IMAGEUID_T1 numeric Unique ID of T1 dicom Range value: 808819, ..., 1697549

IMAGEUID_T2 numeric Unique ID of T2 dicom Range value: 808822, ..., 1695114

EXAMDATE Date Date of MRI scan Date: YYY-MM-DD

LEFT_CA1_VOL numeric left hemisphere ROI CA1 volume Range value: 43.03, ..., 5584.101

LEFT_CA1_NS numeric Number of coronal slices in which left hemisphere ROI CA1 appears Range value: 2, ..., 26

LEFT_CA2_VOL numeric left hemisphere ROI CA2 volume Range value: 0.234, ..., 7.629

LEFT_CA2_NS numeric Number of coronal slices in which left hemisphere ROI CA2 appears Range value: 1, ..., 9

LEFT_CA3_VOL numeric left hemisphere ROI CA3 volume Range value: 1.526, ..., 122.287

LEFT_CA3_NS numeric Number of coronal slices in which left hemisphere ROI CA3 appears Range value: 1, ..., 11

LEFT_DG_VOL numeric left hemisphere ROI dentate gyrus volume Range value: 8.85, ..., 2496.641

LEFT_DG_NS numeric Number of coronal slices in which left hemisphere ROI dentate gyrus appears Range value: 1, ..., 23

LEFT_MISC_VOL numeric left hemisphere ROI MISC volume Range value: 28.076, ..., 793.426

LEFT_MISC_NS numeric Number of coronal slices in which left hemisphere ROI MISC appears Range value: 4, ..., 24

LEFT_SUB_VOL numeric left hemisphere ROI Subiculum volume Range value: 61.035, ..., 1117.553

LEFT_SUB_NS numeric Number of coronal slices in which left hemisphere ROI Subiculum appears Range value: 3, ..., 24

LEFT_ERC_VOL numeric left hemisphere ROI Entorhinal cortex volume Range value: 1.526, ..., 741.646

LEFT_ERC_NS numeric Number of coronal slices in which left hemisphere ROI Entorhinal cortex appears Range value: 1, ..., 15

LEFT_BA35_VOL numeric left hemisphere ROI BA35 volume Range value: 38.452, ..., 995.853

LEFT_BA35_NS numeric Number of coronal slices in which left hemisphere ROI BA35 appears Range value: 2, ..., 16

LEFT_BA36_VOL numeric left hemisphere ROI BA36 volume Range value: 631.103, ..., 4696.956

LEFT_BA36_NS numeric Number of coronal slices in which left hemisphere ROI BA36 appears Range value: 5, ..., 16

LEFT_PHC_VOL numeric left hemisphere ROI parahippocampal cortex volume Range value: 0.61, ..., 5310.664

LEFT_PHC_NS numeric Number of coronal slices in which left hemisphere ROI parahippocampal cortex appears Range value: 1, ..., 11

LEFT_SULCUS_VOL numeric left hemisphere ROI sulcus volume Range value: 43.335, ..., 1256.594

LEFT_SULCUS_NS numeric Number of coronal slices in which left hemisphere ROI sulcus appears Range value: 3, ..., 24

LEFT_CA_VOL numeric left hemisphere ROI CA (CA1+CA2+CA3) volume Range value: 43.03, ..., 5629.267

LEFT_CA_NS numeric Number of coronal slices in which left hemisphere ROI CA (CA1+CA2+CA3) appears Range value: 2, ..., 26

LEFT_HIPP_VOL numeric left hemisphere ROI CA1+CA2+CA3+DG volume Range value: 51.88, ..., 8125.907

LEFT_HIPP_NS numeric Number of coronal slices in which left hemisphere ROI CA1+CA2+CA3+DG appears Range value: 2, ..., 26

RIGHT_CA1_VOL numeric right hemisphere ROI CA1 volume Range value: 318.603, ..., 6692.515

RIGHT_CA1_NS numeric Number of coronal slices in which right hemisphere ROI CA1 appears Range value: 5, ..., 25

RIGHT_CA2_VOL numeric right hemisphere ROI CA2 volume Range value: 0.305, ..., 18.926

RIGHT_CA2_NS numeric Number of coronal slices in which right hemisphere ROI CA2 appears Range value: 1, ..., 11

RIGHT_CA3_VOL numeric right hemisphere ROI CA3 volume Range value: 0.305, ..., 165.901

RIGHT_CA3_NS numeric Number of coronal slices in which right hemisphere ROI CA3 appears Range value: 1, ..., 10

RIGHT_DG_VOL numeric right hemisphere ROI dentate gyrus volume Range value: 124.207, ..., 5831.613

RIGHT_DG_NS numeric Number of coronal slices in which right hemisphere ROI dentate gyrus appears Range value: 3, ..., 22

RIGHT_MISC_VOL numeric right hemisphere ROI MISC volume Range value: 4.578, ..., 552.064

RIGHT_MISC_NS numeric Number of coronal slices in which right hemisphere ROI MISC appears Range value: 3, ..., 25

RIGHT_SUB_VOL numeric right hemisphere ROI Subiculum volume Range value: 9.155, ..., 2147.22

RIGHT_SUB_NS numeric Number of coronal slices in which right hemisphere ROI Subiculum appears Range value: 4, ..., 24

RIGHT_ERC_VOL numeric right hemisphere ROI Entorhinal cortex volume Range value: 1.221, ..., 2073.673

RIGHT_ERC_NS numeric Number of coronal slices in which right hemisphere ROI Entorhinal cortex appears Range value: 1, ..., 15

RIGHT_BA35_VOL numeric right hemisphere ROI BA35 volume Range value: 60.73, ..., 4026.19

RIGHT_BA35_NS numeric Number of coronal slices in which right hemisphere ROI BA35 appears Range value: 3, ..., 16

RIGHT_BA36_VOL numeric right hemisphere ROI BA36 volume Range value: 346.375, ..., 8778.09

RIGHT_BA36_NS numeric Number of coronal slices in which right hemisphere ROI BA36 appears Range value: 4, ..., 16

RIGHT_PHC_VOL numeric right hemisphere ROI parahippocampal cortex volume Range value: 41.809, ..., 5197.762

RIGHT_PHC_NS numeric Number of coronal slices in which right hemisphere ROI parahippocampal cortex appears Range value: 2, ..., 10

RIGHT_SULCUS_VOL numeric right hemisphere ROI sulcus volume Range value: 15.259, ..., 2192.081

RIGHT_SULCUS_NS numeric Number of coronal slices in which right hemisphere ROI sulcus appears Range value: 3, ..., 23

RIGHT_CA_VOL numeric right hemisphere ROI CA (CA1+CA2+CA3) volume Range value: 318.909, ..., 6752.635

RIGHT_CA_NS numeric Number of coronal slices in which right hemisphere ROI CA (CA1+CA2+CA3) appears Range value: 5, ..., 25

RIGHT_HIPP_VOL numeric right hemisphere ROI CA1+CA2+CA3+DG volume Range value: 443.115, ..., 12584.248

RIGHT_HIPP_NS numeric Number of coronal slices in which right hemisphere ROI CA1+CA2+CA3+DG appears Range value: 5, ..., 25

NO_T2_DICOM character No T2 dicom found on ida.loni.usc.edu for analysis "No T2 Dicom"=no T2 dicom available;null=t2 found and analysed

ICV numeric Intercranial volume derived from T1 image Range value: 0, ..., 1922954

SLICE_THICKNESS numeric Coronal slice thickness Range value: 1.99923, ..., 2.0009

T2_QC_R character manual QC score for right hemisphere image quality -1=no image;0=Unusable (substructures and the SRLM are not identifiable) or image has motion and looks blurry/wiggly;1=Bad Quality (SRLM is hard to identify and substructures are blurry) or image has motion;2=Adequate Quality (some substructures are clear, SRLM is mostly visible);3=intermediate quality between 2 & 4;4=Good Quality (most substructures are clear, SRLM is continuously visible) Decoded Value: Yes

T2_QC_L character manual QC score for left hemisphere image quality -1=no image;0=Unusable (substructures and the SRLM are not identifiable) or image has motion and looks blurry/wiggly;1=Bad Quality (SRLM is hard to identify and substructures are blurry) or image has motion;2=Adequate Quality (some substructures are clear, SRLM is mostly visible);3=intermediate quality between 2 & 4;4=Good Quality (most substructures are clear, SRLM is continuously visible) Decoded Value: Yes

COMMENT character QC comment , with options: Motion, Coronal image was in Axial window, Heavy motion, LO-RES CORONAL, Ends in middle of hippocampus

update_stamp POSIXct POSIXct: YYY-MM-DD

Author

adni-data@googlegroups.com