Participant Demographic Information

Participant Demographic Information data. More information is available at https://adni.loni.usc.edu/data-samples/adni-data/

Usage

data(PTDEMOG)

Format

A data frame with 6141 observations and 85 variables.

Source

https://adni.loni.usc.edu/data-samples/adni-data/

Details

ORIGPROT character Original study protocol , with options: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4

COLPROT factor Study protocol of data collection , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4

PTID character Participant ID

RID numeric Participant roster ID Range value: 1, ..., 10894

VISCODE character Visit code

VISCODE2 character Translated visit code

VISDATE Registry EXAMDATE on matching VISCODE Date: YYY-MM-DD

PTSOURCE character Information Source 1=Participant Visit ; 2=Telephone Call Decoded Value: Yes

PTGENDER character 1. What sex was the participant assigned at birth, on their original birth certificate 1=Male ; 2=Female Decoded Value: Yes

PTDOB character 4. Participant Date of Birth

PTDOBYY numeric 2b. Participant Year of Birth 1900..1980

PTHAND character 5. Participant Handedness 1=Right ; 2=Left Decoded Value: Yes

PTMARRY character 6. Participant Marital Status

ADNI1, ADNIGO, ADNI2, ADNI3: 1=Married ; 2=Widowed ; 3=Divorced ; 4=Never married ; 5=Unknown

ADNI4: 1=Married ; 2=Widowed ; 3=Divorced ; 4=Never married ; 5=Unknown ; 6=Domestic Partnership

Decoded Value: Yes

PTEDUCAT numeric 7. Participant Education

ADNI1, ADNIGO, ADNI2: 0..20

ADNI3, ADNI4

PTWORKHS character 7a. Does the participant have a work history sufficient to exclude mental retardation?

ADNI1, ADNIGO, ADNI2: 1=Yes ; 0=No

ADNI3, ADNI4: 0=No ; 1=Yes

Decoded Value: Yes

PTWORK character 8a. Primary occupation during most of adult life?

ADNI1, ADNIGO, ADNI2, ADNI3

ADNI4: 1=Professional and Higher Executive Occupations, Chief Executives ; 2=Middle Professional Occupations/ Small Business Owners ; 3=Managers ; 4=Support Personnel, Drafters, Technicians ; 5=Arts, Design, Entertainment, Sports Occupations – Non-Professional ; 6=Aides / Assistants / Clerks ; 7=Laborers ; 8=Other - Never Employed ; 9=Missing/Unknown

Decoded Value: Yes

PTNOTRT character 9. Participant Retired?

ADNI1, ADNIGO: 1=Yes ; 0=No

ADNI2: 1=Yes ; 0=No ; 2=Not Applicable

ADNI3, ADNI4: 0=No ; 1=Yes ; 2=Not Applicable

Decoded Value: Yes

PTRTYR numeric Retirement Date Range value: 1926, ..., 9999

PTHOME character 10. Type of Participant residence

ADNI1, ADNIGO, ADNI2, ADNI3: 1=House (owned or rented) ; 2=Condo/Co-op (owned) ; 3=Apartment (rented) ; 4=Mobile Home ; 5=Retirement Community ; 6=Assisted Living ; 7=Skilled Nursing Facility ; 8=Other (specify)

ADNI4: 10=House (owned) ; 2=Condo/Co-op (owned) ; 3=Apartment (rented) ; 4=Mobile Home ; 5=Retirement Community ; 6=Assisted Living ; 7=Skilled Nursing Facility ; 8=Other (specify) ; 9=House (rented)

Decoded Value: Yes

PTTLANG character 11. Language to be used for testing the Participant: 1=English ; 2=Spanish Decoded Value: Yes

PTPLANG character 12. What is the participant's primary language?

ADNI1, ADNIGO, ADNI2, ADNI3: 1=English ; 2=Spanish ; 3=Other (specify)

ADNI4: 1=English ; 2=Spanish ; 3=Other

Decoded Value: Yes

PTADBEG numeric 11b. Year of onset of Alzheimer's disease symptoms (best estimate)

ADNI1: 1985..2007

ADNIGO: 1985..2012

PTCOGBEG numeric 23a. Year of onset of cognitive symptoms (best estimate)

ADNI2: 9999=NA - Cognitively Normal ; 1985=1985 ; 1986=1986 ; 1987=1987 ; 1988=1988 ; 1989=1989 ; 1990=1990 ; 1991=1991 ; 1992=1992 ; 1993=1993 ; 1994=1994 ; 1995=1995 ; 1996=1996 ; 1997=1997 ; 1998=1998 ; 1999=1999 ; 2000=2000 ; 2001=2001 ; 2002=2002 ; 2003=2003 ; 2004=2004 ; 2005=2005 ; 2006=2006 ; 2007=2007 ; 2008=2008 ; 2009=2009 ; 2010=2010 ; 2011=2011 ; 2012=2012

ADNI3: 1985=1985 ; 1986=1986 ; 1987=1987 ; 1988=1988 ; 1989=1989 ; 1990=1990 ; 1991=1991 ; 1992=1992 ; 1993=1993 ; 1994=1994 ; 1995=1995 ; 1996=1996 ; 1997=1997 ; 1998=1998 ; 1999=1999 ; 2000=2000 ; 2001=2001 ; 2002=2002 ; 2003=2003 ; 2004=2004 ; 2005=2005 ; 2006=2006 ; 2007=2007 ; 2008=2008 ; 2009=2009 ; 2010=2010 ; 2011=2011 ; 2012=2012 ; 2013=2013 ; 2014=2014 ; 2015=2015 ; 2016=2016 ; 2017=2017 ; 2018=2018 ; 2019=2019 ; 2020=2020 ; 2021=2021 ; 2022=2022 ; 9999=NA - Cognitively Normal

ADNI4: 1985=1985 ; 1986=1986 ; 1987=1987 ; 1988=1988 ; 1989=1989 ; 1990=1990 ; 1991=1991 ; 1992=1992 ; 1993=1993 ; 1994=1994 ; 1995=1995 ; 1996=1996 ; 1997=1997 ; 1998=1998 ; 1999=1999 ; 2000=2000 ; 2001=2001 ; 2002=2002 ; 2003=2003 ; 2004=2004 ; 2005=2005 ; 2006=2006 ; 2007=2007 ; 2008=2008 ; 2009=2009 ; 2010=2010 ; 2011=2011 ; 2012=2012 ; 2013=2013 ; 2014=2014 ; 2015=2015 ; 2016=2016 ; 2017=2017 ; 2018=2018 ; 2019=2019 ; 2020=2020 ; 2021=2021 ; 2022=2022 ; 2023=2023 ; 2024=2024 ; 2025=2025 ; 9999=NA - Cognitively Normal

PTADDX numeric 23b. Year of Alzheimer's Disease diagnosis

ADNI2: 9999=NA - Not Diagnosed with AD ; 1985=1985 ; 1986=1986 ; 1987=1987 ; 1988=1988 ; 1989=1989 ; 1990=1990 ; 1991=1991 ; 1992=1992 ; 1993=1993 ; 1994=1994 ; 1995=1995 ; 1996=1996 ; 1997=1997 ; 1998=1998 ; 1999=1999 ; 2000=2000 ; 2001=2001 ; 2002=2002 ; 2003=2003 ; 2004=2004 ; 2005=2005 ; 2006=2006 ; 2007=2007 ; 2008=2008 ; 2009=2009 ; 2010=2010 ; 2011=2011 ; 2012=2012

ADNI3: 1985=1985 ; 1986=1986 ; 1987=1987 ; 1988=1988 ; 1989=1989 ; 1990=1990 ; 1991=1991 ; 1992=1992 ; 1993=1993 ; 1994=1994 ; 1995=1995 ; 1996=1996 ; 1997=1997 ; 1998=1998 ; 1999=1999 ; 2000=2000 ; 2001=2001 ; 2002=2002 ; 2003=2003 ; 2004=2004 ; 2005=2005 ; 2006=2006 ; 2007=2007 ; 2008=2008 ; 2009=2009 ; 2010=2010 ; 2011=2011 ; 2012=2012 ; 2013=2013 ; 2014=2014 ; 2015=2015 ; 2016=2016 ; 2017=2017 ; 2018=2018 ; 2019=2019 ; 2020=2020 ; 2021=2021 ; 2022=2022 ; 9999=NA - Not Diagnosed with AD

ADNI4: 1985=1985 ; 1986=1986 ; 1987=1987 ; 1988=1988 ; 1989=1989 ; 1990=1990 ; 1991=1991 ; 1992=1992 ; 1993=1993 ; 1994=1994 ; 1995=1995 ; 1996=1996 ; 1997=1997 ; 1998=1998 ; 1999=1999 ; 2000=2000 ; 2001=2001 ; 2002=2002 ; 2003=2003 ; 2004=2004 ; 2005=2005 ; 2006=2006 ; 2007=2007 ; 2008=2008 ; 2009=2009 ; 2010=2010 ; 2011=2011 ; 2012=2012 ; 2013=2013 ; 2014=2014 ; 2015=2015 ; 2016=2016 ; 2017=2017 ; 2018=2018 ; 2019=2019 ; 2020=2020 ; 2021=2021 ; 2022=2022 ; 2023=2023 ; 2024=2024 ; 2025=2025 ; 9999=NA - Not Diagnosed with AD

PTETHCAT character 24. Ethnic Category 1=Hispanic or Latino ; 2=Not Hispanic or Latino ; 3=Unknown Decoded Value: Yes

PTRACCAT character 25. Racial Categories

ADNI1, ADNIGO, ADNI2, ADNI3: 1=American Indian or Alaskan Native ; 2=Asian ; 3=Native Hawaiian or Other Pacific Islander ; 4=Black or African American ; 5=White ; 6=More than one race ; 7=Unknown

ADNI4: 1=American Indian or Alaskan Native ; 2=Asian ; 3=Native Hawaiian or Other Pacific Islander ; 4=Black or African American ; 5=White ; 7=Unknown ; 8=Native Hawaiian ; 9=Other Pacific Islander

Decoded Value: Yes

PTIDENT character 2. Which term(s) best describes the participant's current gender identity? 1=Man ; 2=Woman ; 3=Transgender man ; 4=Transgender woman ; 5=Non-binary/Gender queer ; 6=Two-Spirit (if respondent is AIAN) ; 7=I use a different term ; 8=Don't know ; 9=Prefer not to say Decoded Value: Yes

PTORIENT numeric 3. Which term(s) best describes the participants current sexual orientation 1=Gay, Lesbian ; 2=Straight ; that is, not gay or lesbian, etc. ; 3=Bisexual ; 4=None of these options describe me, and I’d like to see additional options

PTORIENTOT character 3a. Are any of the following options a closer description of how you think of yourself (sexual orientation)? 1=Queer ; 10=I mean something else, please specify: ; 11=Prefer not to answer ; 2=Omnisexual, sapiosexual, or pansexual ; 3=Asexual ; 4=Two-spirit ; 5=Have not figured out or are in the process of figuring out my sexuality ; 6=Mostly straight, but sometimes attracted to people of my own sex ; 7=Do not think of myself as having sexuality ; 8=Do not use labels to identify myself ; 9=Don’t know the answer Decoded Value: Yes

PTENGSPK character 13. Do you consider yourself a monolingual English speaker? 0=No ; 1=Yes Decoded Value: Yes

PTNLANG character 14. What is your native language? 1=English ; 2=Spanish ; 3=Other Decoded Value: Yes

PTENGSPKAGE numeric 14a. For participant's whose native language is not English, how old where you when you learned to speak English? Range value: 0, ..., 46

PTCLANG character 15. What was your primary-caretaker's preferred language? 1=English ; 2=Spanish ; 3=Other Decoded Value: Yes

PTLANGSP character 16. What is the language you prefer to speak? 1=English ; 2=Spanish ; 3=Other Decoded Value: Yes

PTLANGWR character 17. What is the language you prefer to write? 1=English ; 2=Spanish ; 3=Other Decoded Value: Yes

PTSPTIM numeric 18. What percentage of the time during an average day do you speak English? Range value: 0, ..., 100

PTSPOTTIM numeric 19. What percentage of the time during an average day do you speak another language? Range value: 0, ..., 100

PTLANGPR1 character Language 1 0=No ; 1=Yes Decoded Value: Yes

PTLANGSP1 character Speaking 1=almost none ; 2=very poor ; 3=fair ; 4=functional ; 5=good ; 6=very good ; 7=like native speaker Decoded Value: Yes

PTLANGRD1 character Reading 1=almost none ; 2=very poor ; 3=fair ; 4=functional ; 5=good ; 6=very good ; 7=like native speaker Decoded Value: Yes

PTLANGWR1 character Writing 1=almost none ; 2=very poor ; 3=fair ; 4=functional ; 5=good ; 6=very good ; 7=like native speaker Decoded Value: Yes

PTLANGUN1 character Understanding 1=almost none ; 2=very poor ; 3=fair ; 4=functional ; 5=good ; 6=very good ; 7=like native speaker Decoded Value: Yes

PTLANGPR2 character Language 2 0=No ; 1=Yes Decoded Value: Yes

PTLANGSP2 character Speaking 1=almost none ; 2=very poor ; 3=fair ; 4=functional ; 5=good ; 6=very good ; 7=like native speaker Decoded Value: Yes

PTLANGRD2 character Reading 1=almost none ; 2=very poor ; 3=fair ; 4=functional ; 5=good ; 6=very good ; 7=like native speaker Decoded Value: Yes

PTLANGWR2 character Writing 1=almost none ; 2=very poor ; 3=fair ; 4=functional ; 5=good ; 6=very good ; 7=like native speaker Decoded Value: Yes

PTLANGUN2 character Understanding 1=almost none ; 2=very poor ; 3=fair ; 4=functional ; 5=good ; 6=very good ; 7=like native speaker Decoded Value: Yes

PTLANGPR3 character Language 3 0=No ; 1=Yes Decoded Value: Yes

PTLANGSP3 character Speaking 1=almost none ; 2=very poor ; 3=fair ; 4=functional ; 5=good ; 6=very good ; 7=like native speaker Decoded Value: Yes

PTLANGRD3 character Reading 1=almost none ; 2=very poor ; 3=fair ; 4=functional ; 5=good ; 6=very good ; 7=like native speaker Decoded Value: Yes

PTLANGWR3 character Writing 1=almost none ; 2=very poor ; 3=fair ; 4=functional ; 5=good ; 6=very good ; 7=like native speaker Decoded Value: Yes

PTLANGUN3 character Understanding 1=almost none ; 2=very poor ; 3=fair ; 4=functional ; 5=good ; 6=very good ; 7=like native speaker Decoded Value: Yes

PTLANGPR4 character Language 4 0=No ; 1=Yes Decoded Value: Yes

PTLANGSP4 character Speaking 1=almost none ; 2=very poor ; 3=fair ; 4=functional ; 5=good ; 6=very good ; 7=like native speaker Decoded Value: Yes

PTLANGRD4 character Reading 1=almost none ; 2=very poor ; 3=fair ; 4=functional ; 5=good ; 6=very good ; 7=like native speaker Decoded Value: Yes

PTLANGWR4 character Writing 1=almost none ; 2=very poor ; 3=fair ; 4=functional ; 5=good ; 6=very good ; 7=like native speaker Decoded Value: Yes

PTLANGUN4 character Understanding 1=almost none ; 2=very poor ; 3=fair ; 4=functional ; 5=good ; 6=very good ; 7=like native speaker Decoded Value: Yes

PTLANGPR5 character Language 5 0=No ; 1=Yes Decoded Value: Yes

PTLANGSP5 character Speaking 1=almost none ; 2=very poor ; 3=fair ; 4=functional ; 5=good ; 6=very good ; 7=like native speaker Decoded Value: Yes

PTLANGRD5 character Reading 1=almost none ; 2=very poor ; 3=fair ; 4=functional ; 5=good ; 6=very good ; 7=like native speaker Decoded Value: Yes

PTLANGWR5 character Writing 1=almost none ; 2=very poor ; 3=fair ; 4=functional ; 5=good ; 6=very good ; 7=like native speaker Decoded Value: Yes

PTLANGUN5 character Understanding 1=almost none ; 2=very poor ; 3=fair ; 4=functional ; 5=good ; 6=very good ; 7=like native speaker Decoded Value: Yes

PTLANGPR6 character Language 6 0=No ; 1=Yes Decoded Value: Yes

PTLANGSP6 character Speaking 1=almost none ; 2=very poor ; 3=fair ; 4=functional ; 5=good ; 6=very good ; 7=like native speaker Decoded Value: Yes

PTLANGRD6 character Reading 1=almost none ; 2=very poor ; 3=fair ; 4=functional ; 5=good ; 6=very good ; 7=like native speaker Decoded Value: Yes

PTLANGWR6 character Writing 1=almost none ; 2=very poor ; 3=fair ; 4=functional ; 5=good ; 6=very good ; 7=like native speaker Decoded Value: Yes

PTLANGUN6 character Understanding 1=almost none ; 2=very poor ; 3=fair ; 4=functional ; 5=good ; 6=very good ; 7=like native speaker Decoded Value: Yes

PTLANGTTL numeric 21. Total number of languages spoken (based on table above) Range value: 1, ..., 7

PTETHCATH character 24a. If Hispanic or Latino is selected, please provide country/region of origin. : 1=Mexican, Mexican Am., Chicano ; 10=Other Latin culture or origin, please specify: ; 2=Puerto Rican ; 3=Cuban ; 4=Another Hispanic, Latino, or Spanish origin ; 5=Unknown Hispanic or Latino origin ; 6=Dominican ; 7=Mexican ; 8=South American ; 9=Central American Decoded Value: Yes

PTASIAN character 25a. If Asian is selected, please provide country/region of origin. 1=Chinese ; 2=Filipino ; 3=Asian Indian ; 4=Vietnamese ; 5=Korean ; 6=Japanese ; 7=Other Asian Decoded Value: Yes

PTOPI character 25b. If Other Pacific Islander is selected, please provide further detail 1=Samoan ; 2=Chamorro ; 3=Other Pacific islander – for example, Tongan, Fijian, Marshallese, etc Decoded Value: Yes

PTBORN character 26. Were you born in the United States/Canada? 0=No ; 1=Yes ; 2=Unknown Decoded Value: Yes

PTBIRPL character 26a. If NO to 26: Where were you born? 1=Puerto Rico ; 2=Dominican Republic ; 3=Cuba ; 4=Mexico ; 5=South America ; 6=Central America ; 7=Unknown ; 8=Other Decoded Value: Yes

PTIMMAGE numeric 26b. If NO to 26: How old were you when you moved to the United States/Canada? Range value: 0, ..., 57

PTIMMWHY character 26c. What motivated you to migrate to the United States/Canada? 1=To escape our past or future persecution based on race, religion, nationality, and/or membership in a particular social group or political opinion ; 10=For jobs and business opportunities ; 11=Marriage ; 12=To meet new people from different backgrounds ; 13=To make money to send back to my family back home in my home country ; 14=Other ; 2=To escape environmental conflict or violence (i.e., war or escaping gang violence) ; 3=To escape domestic/intimate partner violence ; 4=To find refuge after being displaced due to environmental factors (i.e., natural disasters, erosion, and other environmental factors caused by climate change) ; 5=To seek better healthcare services ; 6=To seek better living conditions ; 7=To offer more opportunities for my children ; 8=Family reunification ; 9=For better educational opportunities for my family and/or me Decoded Value: Yes

PTBIRPR character 27. Were your parents born in the United States/Canada? 0=No ; 1=Yes ; 2=Unknown Decoded Value: Yes

PTBIRGR character 28. Were your grandparents born in the United States/Canada? 0=No ; 1=Yes ; 2=Unknown Decoded Value: Yes

ID numeric Record ID Range value: 12, ..., 186877

SITEID numeric Site ID Range value: 1, ..., 941

USERDATE User Create Date Date: YYY-MM-DD

USERDATE2 User Last Modify Date Date: YYY-MM-DD

DD_CRF_VERSION_LABEL character CRF Version Label , with options: v1

LANGUAGE_CODE character Language Code e=English Decoded Value: Yes

HAS_QC_ERROR character Has Quality Check Error 0=Does not have QC error or QC error has been approved ; 1=Has QC error Decoded Value: Yes

update_stamp POSIXct: YYY-MM-DD

Author

adni-data@googlegroups.com