Participant feedback questionnaire

Participant feedback questionnaire data. More information is available at https://adni.loni.usc.edu/data-samples/adni-data/

Usage

data(PTFBQNR)

Format

A data frame with 42 observations and 31 variables.

Source

https://adni.loni.usc.edu/data-samples/adni-data/

Details

ORIGPROT character Original study protocol , with options: TEAM

COLPROT factor Study protocol of data collection , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4, TEAM

PTID character Participant ID

RID numeric Participant roster ID

VISCODE character Visit code , with options: t_bl

VISCODE2 character Translated visit code , with options: bl

VISDATE Registry EXAMDATE on matching VISCODE Date: YYYY-MM-DD

MOTIVATORS character Q1: Which of the following factors were important in your decision to participate in a research study? 1=Obtaining quality medical care and access to health care professionals ; 2=Learning more about my health ; 3=I am concerned about memory health/Alzheimer’s disease ; 4=Helping others/my community ; 5=Payment ; 6=Convenience ; 7=Confidence in the research staff ; 8=Any other reasons (Please specify below) ; 99=Prefer not to say Decoded Value: Yes

ENJOY character Q2: Based on your experience so far, what do you like about participating in this research study? 1=Contributing to science ; 2=Helping others/ my community ; 3=Procedures such as the memory tests ; 4=Learning about my health ; 5=Interactions with research staff ; 6=Payment ; 7=Convenience ; 8=None of the above ; 9=Any other reasons (Please specify below): ; 99=Prefer not to say Decoded Value: Yes

DISLIKE character Q3: Based on your experience so far, what do you not like about participating in this research study? 1=The memory tests ; 2=The blood draw ; 3=The brain MRI or PET scans (if completed) ; 4=Transportation or parking issues ; 5=Interactions with research staff ; 6=Questionnaires ; 7=Inconvenience ; 8=Nothing, I am completely happy with all aspects of participation ; 9=Any other reasons ; 99=Prefer not to say Decoded Value: Yes

OVRL_EXPERENCE character Q4: Overall, how would you rate your experience in this research study so far? 1=Excellent ; 2=Very good ; 3=Good ; 4=Fair ; 5=Poor ; 99=Prefer not to say Decoded Value: Yes

PREPARED character Q5: Did the informed consent form, information, and discussions you had before participating in the study prepare you for what to expect during the study? 0=No ; 1=Yes, somewhat ; 2=Yes, mostly ; 3=Yes, completely ; 99=Prefer not to say Decoded Value: Yes

RECMMND character Q6: Based on your experience so far, would you recommend joining this research study to your family and friends? 1=Definitely Yes ; 2=Probably Yes ; 3=Probably Not ; 4=Definitely Not ; 99=Prefer not to say Decoded Value: Yes

CONTACT character Q8: When you were not at the research site and you needed to reach a member of the research team, were you able to reach her/him as soon as you wanted? 1=Never ; 2=Sometimes ; 3=Usually ; 4=Always ; 5=Did not need to reach the research team ; 99=Prefer not to say Decoded Value: Yes

DIFFICULTY character Q10: Overall, how easy or hard was it for you to participate in this research study? 1=Easy ; 2=Somewhat easy ; 3=Neither easy nor hard ; 4=Somewhat hard ; 5=Hard ; 99=Prefer not to say Decoded Value: Yes

DISLIKE2 character Q3: What did you not like (liked the least) about participating in this research study? 1=The memory tests ; 2=The blood draw ; 3=The brain MRI or PET scans (if completed) ; 4=Transportation or parking issues ; 5=Interactions with research staff ; 6=Questionnaires ; 7=Inconvenience ; 8=Nothing, I am completely happy with all aspects of participation ; 9=Any other reasons ; 99=Prefer not to say Decoded Value: Yes

ENJOY2 character Q2: What did you like the most about participating in this research study? 1=Contributing to science ; 10=None of the above ; 11=Any other reasons (Please specify below) ; 2=Helping others/contributing to my community ; 3=Procedures such as the memory tests ; 4=Learning about my health ; 5=Interactions with research staff ; 6=Learning some of my test results ; 7=Brain health recommendations/educational materials ; 8=Payment ; 9=Convenience ; 99=Prefer not to say Decoded Value: Yes

EXPOVERALL character Q11: Please rate your overall experience in the TEAM-ADNI research study, where a 1 is the worst possible experience and a 10 is the best possible experience: 1=1 = worst ; 10=10 = best ; 2=2 ; 3=3 ; 4=4 ; 5=5 ; 6=6 ; 7=7 ; 8=8 ; 9=9 ; 99=Prefer not to say Decoded Value: Yes

HEARD character Q6: Did the research team members listen carefully to you? 1=Never ; 2=Sometimes ; 3=Usually ; 4=Always ; 99=Prefer not to say Decoded Value: Yes

PREPARED2 character Q5: Did the informed consent form, information, and discussions you had before participating in the study prepare you for your experience in the study? 0=No ; 1=Yes, somewhat ; 2=Yes, mostly ; 3=Yes, completely ; 99=Prefer not to say Decoded Value: Yes

RECMMND2 character Q4: Would you recommend joining a research study like this one to your family and friends? 1=Definitely Yes ; 2=Probably Yes ; 3=Probably Not ; 4=Definitely Not ; 99=Prefer not to say Decoded Value: Yes

RESPECT character Q7: Did the research team members treat you with courtesy and respect? 1=Never ; 2=Sometimes ; 3=Usually ; 4=Always ; 99=Prefer not to say Decoded Value: Yes

VALUED character Q9: Did you feel you were a valued partner/participant in this research study? 1=Never ; 2=Sometimes ; 3=Usually ; 4=Always ; 99=Prefer not to say Decoded Value: Yes

ID numeric Record ID Range value: 7, ..., 2676

SITEID numeric Site ID Range value: 10, ..., 405

USERDATE User Create Date Date: YYYY-MM-DD

USERDATE2 User Last Modify Date Date: YYYY-MM-DD

DD_CRF_VERSION_LABEL character CRF Version Label , with options: bl

LANGUAGE_CODE character Language Code e=English Decoded Value: Yes

HAS_QC_ERROR character Has Quality Check Error 0=Does not have QC error or QC error has been approved ; 1=Has QC error Decoded Value: Yes

update_stamp POSIXct: YYYY-MM-DD HH:MM:SS