Recruitment Tracking

Recruitment Tracking data. More information is available at https://adni.loni.usc.edu/data-samples/adni-data/

Usage

data(PUBLICITY)

Format

A data frame with 1301 observations and 23 variables.

Source

https://adni.loni.usc.edu/data-samples/adni-data/

Details

ORIGPROT character Original study protocol , with options: ADNI4, ADNI3, ADNI2, ADNIGO, ADNI1

COLPROT factor Study protocol of data collection , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4

PTID character Participant ID

RID numeric Participant roster ID Range value: 21, ..., 10894

VISCODE character Visit code , with options: 4_sc, 4_init

VISCODE2 character Translated visit code

VISDATE Registry EXAMDATE on matching VISCODE Date: YYY-MM-DD

REFERRAL character Do you have a referral code for this participant? 0=No ; 1=Yes Decoded Value: Yes

HEARD character 1. How did you hear about the study?

ADNI4: 1=I received something in the mail. (Use the referral code or the phone number on the mailer as the referral code above.) ; 10=My community group newsletter/bulletin (Provide the group name below.) ; 11=Saw a flyer (Let us know where you saw the flyer below.) ; 12=Saw a printed advertisement (Share the name of the publication where you saw the ad below.) ; 13=Read a printed news article, report (Include the name of the publication and let us know if you are a subscriber below.) ; 14=Picked up a brochure (Let us know where you got the brochure below.) ; 15=Saw a video (Let us know where you saw the video below.) ; 16=Reading materials in a wait room (Let us know where you were waiting below.) ; 17=I received a phone call ; 2=I have received an email. (Include referral code above.) ; 3=I attended a research event/talk presentation. (Specify the event/talk below.) ; 4=A speaker came to one of my meetings/events. (Name speaker and meeting) ; 5=Listening to the radio. (Specify radio show below and if you are a regular listener.) ; 6=Watching TV (Specify below what you were watching and if you are a regular viewer.) ; 7=Website (Specify the website below.) ; 8=Facebook (Specify whether it was a Facebook ad or something a friend shared with you.) ; 9=Other social media (Let us know more below.) ; 99=Other (Please provide any additional information you can regarding how you heard about the study if applicable.)

ADNI4: 1=I received something in the mail. Specify below: referral code on mailer. ; 10=I saw it in my community group newsletter/bulletin. Specify below: name or type of community group and newsletter/bulletin (including if in person or virtual). ; 11=I saw a flyer. Specify below: referral code on flyer and/or name or type of place you saw it. ; 12=I saw another type of printed advertisement. Specify below: referral code on the printed advertisement and/or type of printed advertisement. ; 13=I saw a news article/report. Specify below: Online or in print. Specify name of publication and if you are a subscriber. ; 14=I picked up a brochure. Specify below: referral code on brochure and/or name or type of place you saw it. ; 15=I saw a video (non-social media). Specify below: describe video and where you saw it. ; 16=I saw it in reading materials in a waiting room. Specify below: referral code in the reading material and where you were waiting. ; 17=I received a phone call. Specify below: organization or type of organization who called. ; 18=I saw it on YouTube. Specify below: describe images/words used in the video. ; 19=I saw it on Instagram. Specify below: describe images/video/words used. ; 2=I received an email. Specify below: name of organization who emailed you and referral code in email. ; 20=I saw it on Twitter. Specify below: describe images/video/words used. ; 21=I saw it on another social media website. Specify below: name of social media site and describe images/video/words used. ; 22=I saw it on another non-social media website. Specify below: name of the website or type of website (e.g. website list of AD studies, etc.). ; 23=I attended a health fair. Specify below: name or type of event and where it was held (including if in person or virtual). ; 24=I attended a memory screening. Specify below: name or type of event and where it was held (including if in person or virtual). ; 3=I attended a research/educational event/talk presentation. Specify below: name or type of event/talk and where it was held (including if in person or virtual). ; 4=A speaker came to one of my meetings/events. Specify below: name of speaker, name or type of meeting/event, and where it was held (including if in person or virtual). ; 5=I heard about it on the radio. Specify below: name of radio show and if you are a regular listener. ; 6=I saw it on TV. Specify below: name of tv show and if you are a regular watcher. ; 8=I saw it on Facebook. Specify below: describe images/video/words used. ; 99=I heard about it some other way. Below, please provide any additional information you can regarding how you heard about the study if applicable.

Decoded Value: Yes

WHOMHRD character 2. From whom did you hear about the study?

ADNI4: 1=My general practitioner ; 10=Alzheimer’s Association or Trial Match ; 11=Alzheimer’s Prevention Initiative/Registry “API/R” ; 12=Clinical Trials.gov ; 13=NIH, NIA Alzheimer’s Disease Education and Referral Center (ADEAR) ; 14=The Center for Information and Study on Clinical Research Participation (CISCRP) ; 15=My alma mater ; 16=My trusted research institution ; 17=My community group ; 18=A trusted advisor ; 19=A news report ; 2=A family member’s general practitioner ; 20=A blogger ; 21=A trusted medical resource ; 3=My specialty doctor ; 4=A family member’s specialty doctor ; 5=An Alzheimer’s research doctor ; 6=An Alzheimer’s research team member ; 7=A friend ; 8=A family member ; 9=Brain Health Registry “BHR” ; 99=Other (please specify)

ADNI4: 1=My general practitioner ; 10=Alzheimer’s Association or Trial Match ; 11=Alzheimer’s Prevention Initiative/Registry "API/R" ; 12=Clinical Trials.gov ; 13=NIH, NIA Alzheimer’s Disease Education and Referral Center (ADEAR) ; 14=The Center for Information and Study on Clinical Research Participation (CISCRP) ; 15=My alma mater ; 16=My trusted research institution ; 17=My community group ; 18=A trusted advisor ; 19=A news report ; 2=A family member’s general practitioner ; 20=A blogger ; 21=A trusted medical resource ; 22=Community health clinic ; 23=Religious/charitable organization ; 24=Senior center ; 25=Social services organization ; 3=My specialty doctor ; 4=A family member’s specialty doctor ; 5=An Alzheimer’s research doctor ; 6=An Alzheimer’s research team member ; 7=A friend ; 8=A family member ; 99=Other (please specify)

Decoded Value: Yes

HUBSITE character 2a. If yes, is your clinical site a Hub Site/Hub City? 0=No ; 1=Yes Decoded Value: Yes

PTTYPE character 1. Is this participant a rollover? 0=No ; 1=Yes Decoded Value: Yes

RECRUIT character 2. Was this participant a direct recruit from the site? 0=No ; 1=Yes Decoded Value: Yes

REFCRL character 2b. If yes, was this participant referred to your site via the site Community Research Liaison (CRL) staff member? 0=No ; 1=Yes Decoded Value: Yes

REFEBISU character 3. Was the participant referred to your clinical site via participation in Remote (online) ADNI cohorts (such as the Remote Digital cohort or the Remote Blood cohort)? 0=No ; 1=Yes Decoded Value: Yes

ID numeric Record ID Range value: 318, ..., 106719

SITEID numeric Site ID Range value: 2, ..., 941

USERDATE User Create Date Date: YYY-MM-DD

USERDATE2 User Last Modify Date Date: YYY-MM-DD

DD_CRF_VERSION_LABEL character CRF Version Label , with options: v1, v2

LANGUAGE_CODE character Language Code e=English Decoded Value: Yes

HAS_QC_ERROR character Has Quality Check Error 0=Does not have QC error or QC error has been approved ; 1=Has QC error Decoded Value: Yes

update_stamp POSIXct: YYY-MM-DD

Author

adni-data@googlegroups.com