DIG Remotely administered Assessment of Everyday Cognition (Ecog) Questionnaire – Informant-Report Version

[DIG] Remotely administered Assessment of Everyday Cognition (Ecog) Questionnaire – Informant-Report Version data. More information is available at https://adni.loni.usc.edu/data-samples/adni-data/

Usage

data(RMT_ECOG12SP)

Format

A data frame with 3794 observations and 21 variables.

Source

https://adni.loni.usc.edu/data-samples/adni-data/

Details

RMT_PHASE character Phase of the study in which this observation was collected , with options: ADNI4

ADNIOnlineID character Participant Identifier for the Remote Cohort

PTID character Participant Identifier for the in-clinic ADNI study (if applicable)

RMT_Timepoint character Timepoint of observation , with options: m00, m06, m12, m18, m24

RMT_StatusDate Date of completion Date: YYY-MM-DD

RMT_SP_CONCERN numeric Are you concerned that they have a memory or other thinking problem? 0=No ; 1=Yes ; 3=Prefer not to say

RMT_SP_MEMREF numeric Compared to one year ago, do you feel that their memory has declined substantially? 0=No ; 1=Yes ; 3=Prefer not to say

RMT_ecogsp_Duration numeric Amount of time elapsed between beginning and completing the assessment (seconds) Range value: 21, ..., 86281

RMT_ecogsp1 numeric Remembering where they have placed personal items or objects 1=Better or no change ; 2=Questionable / Occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 9=Don't know / Not applicable

RMT_ecogsp2 numeric Remembering the current date or day of the week 1=Better or no change ; 2=Questionable / Occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 9=Don't know / Not applicable

RMT_ecogsp3 numeric Communicating or expressing ideas in a conversation 1=Better or no change ; 2=Questionable / Occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 9=Don't know / Not applicable

RMT_ecogsp4 numeric Understanding spoken directions or instructions 1=Better or no change ; 2=Questionable / Occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 9=Don't know / Not applicable

RMT_ecogsp5 numeric Difficulty calculating distances when driving or driving too close to another vehicle 1=Better or no change ; 2=Questionable / Occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 9=Don't know / Not applicable

RMT_ecogsp6 numeric Finding their way around a familiar home 1=Better or no change ; 2=Questionable / Occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 9=Don't know / Not applicable

RMT_ecogsp7 numeric Anticipating weather changes and planning accordingly 1=Better or no change ; 2=Questionable / Occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 9=Don't know / Not applicable

RMT_ecogsp8 numeric Planning a trip, vacation, or outing 1=Better or no change ; 2=Questionable / Occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 9=Don't know / Not applicable

RMT_ecogsp9 numeric Keeping a living and work space (such as a home office, garage, or pantry) organized 1=Better or no change ; 2=Questionable / Occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 9=Don't know / Not applicable

RMT_ecogsp10 numeric Managing bill payments 1=Better or no change ; 2=Questionable / Occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 9=Don't know / Not applicable

RMT_ecogsp11 numeric Doing two things at once 1=Better or no change ; 2=Questionable / Occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 9=Don't know / Not applicable

RMT_ecogsp12 numeric Working on a particular task while talking to someone at the same time 1=Better or no change ; 2=Questionable / Occasionally worse ; 3=Consistently a little worse ; 4=Consistently much worse ; 9=Don't know / Not applicable

update_stamp POSIXct: YYY-MM-DD

Author

adni-data@googlegroups.com