Novoic Storyteller Data

Novoic Storyteller Data data. More information is available at https://adni.loni.usc.edu/data-samples/adni-data/

Usage

data(RMT_STORYTELLER)

Format

A data frame with 15878 observations and 72 variables.

Source

https://adni.loni.usc.edu/data-samples/adni-data/

Details

RMT_PHASE character Phase of the study in which this observation was collected , with options: ADNI4

ADNIOnlineID character Participant Identifier for the Remote Cohort

PTID character Participant Identifier for the in-clinic ADNI study (if applicable)

RMT_Timepoint character Timepoint of observation , with options: m00, m06, m12, m18, m24, m30

EXAMDATE Date of examination Date: YYYY-MM-DD

SCREENER_SCORE numeric Overall G-match score for the screener Range value: 0, ..., 100

SCREENER_RECOMMEND character Overall recommendation for the screener 0=No ; 1=Yes Decoded Value: Yes

QC_ATTEMPTS numeric Number of attempts at the quality control task Range value: 1, ..., 11

QC_PASSED character Whether the quality control task was passed 0=No ; 1=Yes Decoded Value: Yes

USER_AGENT character User agent string

STORY1IMM_LANGUAGE character Language for story 1 immediate recall , with options: en-US, es-US

STORY1IMM_PLAYBACK_COUNT numeric Audio stimulus playback count for story 1 immediate recall Range value: 0, ..., 5

STORY1IMM_AUDIO_LENGTH numeric Participant audio length for story 1 immediate recall Range value: 0.76, ..., 305.18

STORY1IMM_NUM_PRONOUNS numeric Number of pronouns for story 1 immediate recall Range value: 0, ..., 77

STORY1IMM_NUM_NOUNS numeric Number of nouns for story 1 immediate recall Range value: 0, ..., 46

STORY1IMM_NUM_PROPER_NOUNS numeric Number of proper nouns for story 1 immediate recall Range value: 0, ..., 13

STORY1IMM_NUM_VERBS numeric Number of verbs for story 1 immediate recall Range value: 0, ..., 56

STORY1IMM_NUM_ADVERBS numeric Number of adverbs for story 1 immediate recall Range value: 0, ..., 28

STORY1IMM_NUM_ADJECTIVES numeric Number of adjectives for story 1 immediate recall Range value: 0, ..., 18

STORY1IMM_NUM_WORDS numeric Number of words for story 1 immediate recall Range value: 0, ..., 350

STORY1IMM_IDEA_DENSITY numeric Idea density for story 1 immediate recall Range value: 0, ..., 1

STORY1IMM_UNIQUE_WORDS_RATIO numeric Unique words ratio for story 1 immediate recall Range value: 0.357142857142857, ..., 1

STORY1IMM_MEAN_LETTER_PER_WORD numeric Mean number of letters per word for story 1 immediate recall Range value: 2, ..., 9

STORY1IMM_GMATCH numeric G-match score (source text vs transcript) for story 1 immediate recall Range value: 0, ..., 100

STORY1IMM_SPEECH_DURATION numeric Speech duration for story 1 immediate recall Range value: 0, ..., 112.6

STORY1IMM_NONSPEECH_PROPORTION numeric Non-speech proportion of time for story 1 immediate recall Range value: 0, ..., 0.912912912912913

STORY1IMM_SPEECH_RATE numeric Speech rate for story 1 immediate recall Range value: 0.0338983050847458, ..., 5.55555555555556

STORY1IMM_ARTICULATION_RATE numeric Articulation rate for story 1 immediate recall Range value: 0.0338983050847458, ..., 5.55555555555556

STORY1IMM_NUM_PAUSES numeric Number of pauses for story 1 immediate recall Range value: 0, ..., 35

STORY1IMM_MEAN_PAUSE_LENGTH numeric Mean length of pauses for story 1 immediate recall Range value: 0.3, ..., 39.8

STORY2IMM_LANGUAGE character Language for story 2 immediate recall , with options: en-US, es-US

STORY2IMM_PLAYBACK_COUNT numeric Audio stimulus playback count for story 2 immediate recall Range value: 0, ..., 5

STORY2IMM_AUDIO_LENGTH numeric Participant audio length for story 2 immediate recall Range value: 0.08, ..., 305.52

STORY2IMM_NUM_PRONOUNS numeric Number of pronouns for story 2 immediate recall Range value: 0, ..., 62

STORY2IMM_NUM_NOUNS numeric Number of nouns for story 2 immediate recall Range value: 0, ..., 70

STORY2IMM_NUM_PROPER_NOUNS numeric Number of proper nouns for story 2 immediate recall Range value: 0, ..., 17

STORY2IMM_NUM_VERBS numeric Number of verbs for story 2 immediate recall Range value: 0, ..., 69

STORY2IMM_NUM_ADVERBS numeric Number of adverbs for story 2 immediate recall Range value: 0, ..., 34

STORY2IMM_NUM_ADJECTIVES numeric Number of adjectives for story 2 immediate recall Range value: 0, ..., 22

STORY2IMM_NUM_WORDS numeric Number of words for story 2 immediate recall Range value: 0, ..., 363

STORY2IMM_IDEA_DENSITY numeric Idea density for story 2 immediate recall Range value: 0, ..., 1

STORY2IMM_UNIQUE_WORDS_RATIO numeric Unique words ratio for story 2 immediate recall Range value: 0.235955056179775, ..., 1

STORY2IMM_MEAN_LETTER_PER_WORD numeric Mean number of letters per word for story 2 immediate recall Range value: 2, ..., 12

STORY2IMM_GMATCH numeric G-match score (source text vs transcript) for story 2 immediate recall Range value: 0, ..., 100

STORY2IMM_SPEECH_DURATION numeric Speech duration for story 2 immediate recall Range value: 0, ..., 128.6

STORY2IMM_NONSPEECH_PROPORTION numeric Non-speech proportion of time for story 2 immediate recall Range value: 0, ..., 0.898190045248869

STORY2IMM_SPEECH_RATE numeric Speech rate for story 2 immediate recall Range value: 0.0925925925925926, ..., 6.36363636363636

STORY2IMM_ARTICULATION_RATE numeric Articulation rate for story 2 immediate recall Range value: 0.0925925925925926, ..., 6.36363636363636

STORY2IMM_NUM_PAUSES numeric Number of pauses for story 2 immediate recall Range value: 0, ..., 38

STORY2IMM_MEAN_PAUSE_LENGTH numeric Mean length of pauses for story 2 immediate recall Range value: 0.3, ..., 42.3

STORY1DEL_LANGUAGE character Language for story 1 delayed recall , with options: en-US, en, es-US

STORY1DEL_PLAYBACK_COUNT numeric Audio stimulus playback count for story 1 delayed recall Range value: 0, ..., 3

STORY1DEL_AUDIO_LENGTH numeric Participant audio length for story 1 delayed recall Range value: 0.1, ..., 305.62

STORY1DEL_NUM_PRONOUNS numeric Number of pronouns for story 1 delayed recall Range value: 0, ..., 116

STORY1DEL_NUM_NOUNS numeric Number of nouns for story 1 delayed recall Range value: 0, ..., 109

STORY1DEL_NUM_PROPER_NOUNS numeric Number of proper nouns for story 1 delayed recall Range value: 0, ..., 21

STORY1DEL_NUM_VERBS numeric Number of verbs for story 1 delayed recall Range value: 0, ..., 75

STORY1DEL_NUM_ADVERBS numeric Number of adverbs for story 1 delayed recall Range value: 0, ..., 60

STORY1DEL_NUM_ADJECTIVES numeric Number of adjectives for story 1 delayed recall Range value: 0, ..., 28

STORY1DEL_NUM_WORDS numeric Number of words for story 1 delayed recall Range value: 0, ..., 668

STORY1DEL_IDEA_DENSITY numeric Idea density for story 1 delayed recall Range value: 0, ..., 1

STORY1DEL_UNIQUE_WORDS_RATIO numeric Unique words ratio for story 1 delayed recall Range value: 0.366766467065868, ..., 1

STORY1DEL_MEAN_LETTER_PER_WORD numeric Mean number of letters per word for story 1 delayed recall Range value: 1, ..., 12

STORY1DEL_GMATCH numeric G-match score (source text vs transcript) for story 1 delayed recall Range value: 0, ..., 100

STORY1DEL_GMATCH_IMM_DEL numeric G-match score (immediate transcript vs delayed transcript) for story 1 delayed recall Range value: 0, ..., 100

STORY1DEL_SPEECH_DURATION numeric Speech duration for story 1 delayed recall Range value: 0, ..., 231.5

STORY1DEL_NONSPEECH_PROPORTION numeric Non-speech proportion of time for story 1 delayed recall Range value: 0, ..., 0.934402332361516

STORY1DEL_SPEECH_RATE numeric Speech rate for story 1 delayed recall Range value: 0.0551724137931034, ..., 7.5

STORY1DEL_ARTICULATION_RATE numeric Articulation rate for story 1 delayed recall Range value: 0.0840336134453782, ..., 7.5

STORY1DEL_NUM_PAUSES numeric Number of pauses for story 1 delayed recall Range value: 0, ..., 48

STORY1DEL_MEAN_PAUSE_LENGTH numeric Mean length of pauses for story 1 delayed recall Range value: 0.3, ..., 64.2

update_stamp POSIXct: YYYY-MM-DD HH:MM:SS