DIG Remotely administered screening questionnaire

[DIG] Remotely administered screening questionnaire data. More information is available at https://adni.loni.usc.edu/data-samples/adni-data/

Usage

data(RMT_Screening)

Format

A data frame with 12227 observations and 26 variables.

Source

https://adni.loni.usc.edu/data-samples/adni-data/

Details

RMT_PHASE character Phase of the study in which this observation was collected , with options: ADNI4

ADNIOnlineID character Participant Identifier for the Remote Cohort

PTID character Participant Identifier for the in-clinic ADNI study (if applicable)

RMT_Timepoint character Timepoint of observation , with options: m00, m06, m12, m18

RMT_StatusDate Date of completion Date: YYY-MM-DD

RMT_MCI_Dx numeric Have you been diagnosed with Mild Cognitive Impairment (MCI) or a memory problem by a healthcare provider? 0=No ; 1=Yes ; 2=I don't know ; 3=Prefer not to say

RMT_AD_Dx numeric Have you been diagnosed with Alzheimer's Disease (AD) by a healthcare provider? 0=No ; 1=Yes ; 2=I don't know ; 3=Prefer not to say

RMT_DEM_Dx numeric Have you been diagnosed with dementia by a healthcare provider? 0=No ; 1=Yes ; 2=I don't know ; 3=Prefer not to say

RMT_CImed numeric Have you ever been prescribed a medication for cognitive impairment or memory problems by a healthcare provider? 0=No ; 1=Yes ; 2=I don't know ; 3=Prefer not to say

RMT_FamAD numeric Have any of your biological parents, full sibling(s), or biological children ever been diagnosed with Alzheimer's Disease or dementia by a healthcare provider? 0=No ; 1=Yes ; 2=I don't know ; 3=Prefer not to say

RMT_NursHm numeric Do you reside in a skilled nursing facility? 0=No ; 1=Yes ; 3=Prefer not to say

RMT_MntlHlth numeric In the past year, have you been diagnosed with any of the following conditions: schizophrenia, bipolar disorder, or major depression? 0=No ; 1=Yes ; 3=Prefer not to say

RMT_BrainCndtn numeric Have you ever been diagnosed with any of the following conditions: Parkinson's disease, Huntington's disease, a brain tumor, a seizure disorder, or multiple sclerosis? 0=No ; 1=Yes ; 3=Prefer not to say

RMT_AlchlDrug numeric In the past two years, have you been diagnosed by a healthcare provider as having alcohol use disorder or drug use disorder? 0=No ; 1=Yes ; 3=Prefer not to say

RMT_Metal numeric Do you have a pacemaker, an implanted medical device (such as aneurysm clips, artificial heart valve, IUD, etc.), metal fragments (for example, due to military service or metal working), or other foreign object(s) in your body?Note: if the foreign object is an artificial knee or hip replacement, please answer no 0=No ; 1=Yes ; 3=Prefer not to say

RMT_Metal_rplc numeric Is the metal in your body an artificial knee or hip replacement? 0=No ; 1=Yes

RMT_Metal_other numeric Do you have any other metal in your body (such as pacemaker, aneurysm clips, metal fragments, etc.)? 0=No ; 1=Yes

RMT_MemStudy numeric Are you willing to participate in a 5-year study that includes memory and thinking tests? 0=No ; 1=Yes ; 3=Prefer not to say

RMT_MRIPET numeric Are you willing to participate in a 5-year study that includes MRI and PET scans (pictures of your brain, at no cost to you)? 0=No ; 1=Yes ; 3=Prefer not to say

RMT_Clausphba numeric Do you experience claustrophobia (fear of confined spaces) that would prevent you from completing a MRI or PET scan (pictures of your brain)? 0=No ; 1=Yes ; 2=I don't know ; 3=Prefer not to say

RMT_BLDwill numeric Are you willing to have a blood draw at a nearby local clinic (Quest Diagnostics Patient Service Center) at no cost to you? 0=No ; 1=Yes ; 3=Prefer not to say

RMT_Web numeric Do you regularly have access to the internet? 0=No ; 1=Yes ; 3=Prefer not to say

RMT_Device numeric Do you have a computer, smartphone, or tablet device (such as an iPad, Chromebook, etc.)? 0=No ; 1=Yes ; 3=Prefer not to say

Referral_BLD numeric Has this participant been referred to the remote blood cohort? 0=No ; 1=Yes

Referral_CLIN numeric Has this participant been referred to the in-clinic ADNI study? 0=No ; 1=Yes

update_stamp POSIXct: YYY-MM-DD