Study Visits Summary

Study Visits Summary data. More information is available at https://adni.loni.usc.edu/data-samples/adni-data/

Usage

data(STUDYSUM)

Format

A data frame with 983 observations and 30 variables.

Source

https://adni.loni.usc.edu/data-samples/adni-data/

Details

ORIGPROT character Original study protocol , with options: ADNI3, ADNI2, ADNIGO, ADNI1, ADNI4

COLPROT factor Study protocol of data collection , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4

PTID character Participant ID

RID numeric Participant roster ID Range value: 2, ..., 10893

VISCODE character Visit code , with options: sc, init, bl, y1, y2, y4, y3, y5, y6, 4_disp

VISCODE2 character Translated visit code

SDSTATUS character Final disposition

ADNI3: 1=Never enrolled ; 2=Enrolled - Early discontinuation of study visits ; 3=Enrolled - Completed ADNI3 study visits

ADNI4: 1=Never enrolled ; 2=Enrolled - Early discontinuation of study visits ; 3=Enrolled - Completed ADNI4 study visits

Decoded Value: Yes

SDDATE Final ADNI4 disposition date Date: YYY-MM-DD

INCLUSION character Never enrolled due to: Failing to meet Inclusion Criteria 0=No ; 1=Yes ; 2=Not Assessed Decoded Value: Yes

EXCLUSION character Never enrolled due to: Meeting Exclusion Criteria 0=No ; 1=Yes ; 2=Not Assessed Decoded Value: Yes

INCROLL character If Never enrolled due to failing Inclusion Criteria select criteria:

ADNI3: rollover1=Rollover Pts - 1. Must have been enrolled and followed in ADNI1, ADNIGO, or ADNI2 for at least one year. ; rollover2=Rollover Pts - 2. Willing and able to continue to participate in an ongoing longitudinal study. A reduced battery of tests is allowable if the participant is not able/willing to complete the full battery.

ADNI4: rollover1=Rollover Pts - 1. Must have been enrolled and followed in one of the following previous ADNI studies: ADNIGO, ADNI2, ADNI3 for at least one year. ; rollover2=Rollover Pts - 2. Willing and able to continue to participant in an ongoing longitudinal study. A reduced battery of tests is allowable. ; rollover3=Rollover Pts - 3. Study partner may be available who has frequent contact with the participant (i.e., minimum average of 2 hours per week), and may be able to accompany the participant to clinic visits or provide information remotely (e.g. over the phone).

Decoded Value: Yes

VERSION character Indicate version of criteria evaluated:

ADNI3: 1=Original protocol inclusion/exclusion criteria (e.g. Protocol Version 1.0 - 24 MAY 2016) ; 2=Updated inclusion/exclusion criteria (e.g. Protocol Version 2.0 - 04APR2019)

ADNI4: 1=Original protocol inclusion/exclusion criteria

Decoded Value: Yes

INCNEWPT character If Never enrolled due to failing Inclusion Criteria select criteria:

ADNI3: newpt1=1. Memory [CN] Participant with or without subjective memory complaints ; [MCI] and [AD] Memory concern expressed by participant, study partner or clinician. ; newpt10=10. Visual and auditory acuity adequate for neuropsychological testing. ; newpt11=11. Good general health with no diseases expected to interfere with the study. ; newpt12=12. Participant is not pregnant, lactating, or of childbearing potential. ; newpt13=13. Willing and able to participate in a longitudinal imaging study. ; newpt14=14. Modified Hachinski Ischemic Score less than or equal to 4. ; newpt15=15. Completed six grades of education or has a good work history (sufficient to exclude mental retardation). ; newpt16=16. Must speak English or Spanish fluently. ; newpt17=17. Willing to undergo repeated MRIs (3Tesla) and at least two PET scans ; newpt18=18. Agrees to collection of blood for genomic analysis (including GWAS sequencing and other analysis), APOE testing and biospecimen banking. ; newpt19=19. Agrees to collection of blood for biomarker testing. ; newpt2=2. Logical Memory II Delayed [CN] ≥9 for 16+ yrs of education, ≥5 for 8-15 yrs, ≥3 for 0-7 yrs ; [MCI] < 11 for 16+ yrs of education, ≤ 9 for 8-15 yrs, ≤ 6 for 0-7 yrs ; [AD] ≤ 8 for 16+ yrs of education, ≤ 4 for 8-15 yrs, ≤ 2 for 0-7 yrs. ; newpt20=20. Agrees to at least one lumbar puncture for the collection of CSF. ; newpt21=21. Agrees to share genomic data and biomarker samples. ; newpt3=3. MMSE [CN & MCI] 24-30 ; [AD] 20-24 (Protocol version 1)/ [AD] 20-26 (after protocol version 1) ; newpt4=4. CDR [CN] Global: 0 Memory: 0 ; [MCI] Global: 0.5 Memory: 0.5+ ; [AD] Global: 0.5-1.0. ; newpt5=5. [CN] Cognitively normal ; [MCI] General cognition and functional performance sufficiently preserved such that a diagnosis of AD cannot be made ; [AD] NINCDS/ADRDA criteria for probable AD ; newpt6=6. Stable Medications [CN] Stability of Permitted Meds for 4 wks ; [MCI] & [AD] Stability of Permitted Meds for 4 wks & Cholinesterase inhibitors/memantine stable for 12 weeks prior to Screen ; newpt7=7. Geriatric Depression Scale score less than 6. ; newpt8=8. Age between 55-90 years (inclusive). ; newpt9=9. Study partner with frequent contact with the participant/able to accompany participant to clinic visits.

ADNI4: newpt1=1. [CN] Participants may or may not have a significant subjective memory concern as reported by participant, study partner, or clinician ; [MCI] and [DEM] Participant must have a subjective memory concern as reported by participant, study partner, or clinician. ; newpt10=4. Visual and auditory acuity adequate for neuropsychological testing. ; newpt11=5. Good general health with no disease expected to interfere with the study. ; newpt12=6. Participant is not pregnant, lactating, or of childbearing potential (i.e., women must be two years post-menopausal or surgically sterile). ; newpt13=7. Willing and able to participate in a longitudinal imaging study. ; newpt14=8. Must be literate and speak English or Spanish fluently. ; newpt15=9. Agrees to collection of blood for GWAS, APOE testing, DNA and RNA testing ; newpt16=10. Agrees to collection of blood for biomarker testing. ; newpt17=11. Agrees to participate in the ADNI study which includes cognitive evaluation, MRI and PET scans. ; newpt18=12. Flexibility can be made to all criteria for those with at least 8 years in a low socio-economic status (SES) neighborhood. ; newpt2=2. [CN] Normal memory function documented by scoring above demographically-adjusted cutoffs on the Logical Memory II subscale (Delayed Paragraph Recall, Paragraph A only) from the Wechsler Memory Scale – Revised (the maximum score is 25): a. ≥9 for 16 or more years of education b. ≥ 5 for 8-15 years of education c. ≥ 3 for 0-7 years of education d. Note: cut-offs may be modified over time as the field evolves in this area ; [MCI] and [DEM] Abnormal memory function documented by scoring within the demographically- adjusted ranges on the Logical Memory II subscale (Delayed Paragraph Recall, Paragraph A only) from the Wechsler Memory Scale – Revised (the maximum score is 25): a. ≤11 for 16 or more years of education b. ≤9 for 8-15 years of education c. ≤6 for 0-7 years of education. d. Note: cut-offs may be modified over time as the field evolves in this area. ; newpt3=3. [CN] and [DEM] Mini-Mental State Exam score between 24 and 30 (inclusive) (Exceptions may be made for participants with less than 8 years of education at the discretion of the Project Director and/or Clinical Core) ; [DEM] Mini-Mental State Exam score between 20 and 28 (inclusive) (Exceptions may be made for participants with less than 8 years of education at the discretion of the Project Director and/or Clinical Core). ; newpt4=4. [CN] Clinical Dementia Rating = 0. Memory Box score must be 0. ; [MCI] Clinical Dementia Rating = 0.5. Memory Box score must be at least 0.5 ; [DEM] Clinical Dementia Rating = 0.5 or 1.0. ; newpt5=5. [CN] Cognitively normal, based on an absence of significant impairment in cognitive functions or activities of daily living. [MCI] General cognition and functional performance sufficiently preserved such that a diagnosis of dementia cannot be made by the site physician at the time of the screening visit. [DEM] Meets the National Institute on Aging/Alzheimer’s Association Diagnostic Guidelines for Dementia (2011). ; newpt6=6. [CN] Stability of Permitted Medications for 4 weeks. In particular, participants may: a. Take stable doses of antidepressants lacking significant anticholinergic side effects (if they are not currently depressed and do not have a history of major depression within the past 1 years) b. Estrogen replacement therapy is permissible c. Gingko biloba is permissible, but discouraged d. Washout from psychoactive medication (e.g., excluded antidepressants, neuroleptics, chronic anxiolytics or sedative hypnotics, etc.) for at least 4 weeks prior to screening. [MCI] and [DEM] Stability of Permitted Medications for 4 weeks. In particular, participants may: a. Take stable doses of antidepressants lacking significant anticholinergic side effects (if they are not currently depressed and do not have a history of major depression within the past 1 year) b. Estrogen replacement therapy is permissible c. Gingko biloba is permissible, but discouraged d. Washout from psychoactive medication (e.g., excluded antidepressants, neuroleptics, chronic anxiolytics or sedative hypnotics, etc.) for at least 4 weeks prior to screening e. Cholinesterase inhibitors and memantine are allowable if stable for 12 weeks prior to screen f. Aducanumab and any other approved treatments for the neurobiology of AD if stable for 24 weeks prior to screen ; newpt7=1. Geriatric Depression Scale score less than 10. ; newpt8=2. Age between 55-90 years (inclusive). ; newpt9=3. Study partner who has frequent contact with the participant (i.e., minimum average of 2 hours per week) and may be able to accompany the participant to clinic visits or provide information remotely (e.g. over the phone).

EXCCRIT character If Never enrolled due to meeting Exclusion Criteria select criteria:

ADNI3: newpt1=1. [CN] Any significant neurologic disease ; [MCI] Any significant neurologic disease other than suspected incipient AD ; [AD] Any significant neurologic disease other than AD. ; newpt10=10. Residence in a skilled nursing facility. ; newpt11=11. Current use of specific psychoactive medications, current use of warfarin or other anticoagulants such as dabigatran, rivaroxaban and apixaban (exclusionary for lumbar puncture). ; newpt12=12. Current use of exclusionary medications. ; newpt13=13. Investigational agents are prohibited one month prior to entry and for the duration of the trial. ; newpt14=14. Participation in clinical studies involving neuropsychological measures being collected more than one time per year. ; newpt15=15. History of risk factors for torsades de pointes or taking medications known to prolong the QT interval. (AV-1451) ; newpt16=16. Have an ECG obtained prior to the AV-1451 PET scan that is clinically significant, QTcB below 458 msec in males, or 474 msec in females. (AV-1451) ; newpt2=2. Screening/Baseline MRI brain scan with evidence of infection, infarction, focal lesions, multiple lacunes or lacunes in a critical memory structure. ; newpt3=3. Contraindications for MRI. ; newpt4=4. Major depression, bipolar disorder as described in DSM-IV within the past 1 year. Psychotic features, agitation or behavioral problems within the last 3 months. ; newpt5=5. Currently treated with medication for obsessive-compulsive disorder or attention deficit disorder. ; newpt6=6. History of schizophrenia (DSM IV criteria). ; newpt7=7. History of alcohol or substance abuse or dependence within the past 2 years (DSM IV criteria). ; newpt8=8. Any significant systemic illness or unstable medical condition. ; newpt9=9. Clinically significant abnormalities in B12 or TFTs that might interfere with the study.

ADNI4: newpt1=1. [CN] Any significant neurologic disease, such as Parkinson’s disease, vascular cognitive impairment/dementia, Huntington’s disease, normal pressure hydrocephalus, brain tumor, progressive supranuclear palsy, seizure disorder, subdural hematoma, multiple sclerosis, or history of significant head trauma followed by persistent neurologic defaults or known structural brain abnormalities ; [MCI] Any significant neurologic disease other than suspected Alzheimer’s disease, such as Parkinson’s disease (Parkinsonian symptoms complicating MCI/AD are acceptable), vascular cognitive impairment dementia (multiple lacunes less than or equal to 1.5 cm and/or extensive white matter changes are acceptable), Huntington’s disease, normal pressure hydrocephalus, brain tumor (clinically insignificant meningioma acceptable), progressive supranuclear palsy, seizure disorder, subdural hematoma, multiple sclerosis, or history of significant head trauma followed by persistent neurologic defaults or known structural brain abnormalities. ; [DEM] Same as MCI ; newpt10=9. Clinically significant abnormalities in B12, or thyroid function tests that might interfere with the study. A low B12 is exclusionary, unless follow-up labs (homocysteine (HC) and methylmalonic acid (MMA)) indicate that it is not physiologically significant. ; newpt11=10. Residence in a skilled nursing facility. ; newpt12=11. Current use of specific psychoactive medications (e.g. certain antidepressants, neuroleptics, chronic anxiolytics or sedative hypnotics, etc.), at the discretion of the clinician. ; newpt13=12. Current use of any other exclusionary medications. ; newpt14=13. Investigational agents are prohibited for five half-lives or one month, whichever time period is longer, prior to entry and for the duration of the trial. ; newpt15=14. Participation in clinical studies involving neuropsychological measures being collected more than once time per year. ; newpt16=15. Female that is pregnant, lactating, or of childbearing potential. ; newpt17=16. Flexibility can be made to all criteria for those with at least 8 years in a low socio-economic status (SES) neighborhood ; newpt2=1. Screening/Baseline MRI brain scan with evidence of infection, or other clinically significant focal lesions. Participants with cortical strokes, not large enough to distort anatomy, multiple lacunar infarctions or extensive white matter disease are allowed. ; newpt3=2. Screening/Baseline MRI brain scan with evidence of large structural abnormalities that would corrupt image analytical pipelines –e.g. large hemispheric infarcts, large areas of encephalomalacia, large arachnoid cysts. ; newpt4=3. Unable to complete MRIs for any reason (e.g. pacemaker or other implanted metal devices, severe claustrophobia, anxiety which prevents MRI scans, too large to fit, etc.). ; newpt5=4. Current major depression, bipolar disorder as described in DMS-IV within the past 1 year. Psychotic features, agitation or behavioral problems within the last 3 months which could lead to difficulty complying with the protocol. ; newpt6=5. Currently treated with medication for obsessive-compulsive disorder or attention deficit disorder. ; newpt7=6. History of schizophrenia (DSM-5 criteria). ; newpt8=7. History of alcohol or substance disorder within the past 2 years (DSM-5 criteria). ; newpt9=8. Any significant systemic illness or unstable medical condition which could lead to difficulty complying with the protocol.

MRIFIND character Never enrolled due to: MRI finding(s) other than those listed in Exclusion Criteria

ADNI3: 0=No ; 1=Yes ; 2=Not Assessed

ADNI4: 0=No ; 1=Yes

Decoded Value: Yes

NVRDISC character Never enrolled due to: Discontinuation (e.g. consent withdrawal, lost to follow-up) 0=No ; 1=Yes Decoded Value: Yes

NVROT character Never enrolled due to Other reason(s) 0=No ; 1=Yes Decoded Value: Yes

SDPRIMARY character Primary reason for early discontinuation of study visits

ADNI3: 1=Death ; 2=Safety risk ; 4=Non-compliance ; 5=Adverse event ; 6=Investigator recommendation ; 7=Other non-site clinician recommendation ; 8=Started prohibited medication ; 9=Participant unwilling or unable to participate ; 10=Study partner unwilling or unable to participate ; 11=Study terminated ; 12=Starting new trial ; 13=Lost to follow up ; 14=Coordinating Center request ; 15=Other ; 16=COVID-19 Pandemic disruption

ADNI4: 1=Death ; 10=Study partner unwilling or unable to participate ; 11=Study terminated ; 12=Starting new trial ; 13=Lost to follow up ; 14=Coordinating Center request ; 15=Other ; 16=COVID-19 Pandemic disruption ; 2=Safety risk ; 4=Non-compliance ; 5=Adverse event ; 6=Investigator recommendation ; 7=Other non-site clinician recommendation ; 8=Started prohibited medication ; 9=Participant unwilling or unable to participate

Decoded Value: Yes

SDPRIMAE numeric If Adverse Event, provide Adverse Event number Range value: 1, ..., 5

SDSECOND character Were there other contributing reason(s) for early discontinuation of study visits? 0=No ; 1=Yes Decoded Value: Yes

SDSECREAS character If Yes, select other contributing reason(s) for early discontinuation of study visits:

ADNI3: 2=Safety risk ; 4=Non-compliance ; 5=Adverse event ; 6=Investigator recommendation ; 7=Other non-site clinician recommendation ; 8=Started prohibited medication ; 9=Participant unwilling or unable to participate ; 10=Study partner unwilling or unable to participate ; 12=Starting new trial ; 14=Coordinating Center request ; 15=Other ; 16=COVID-19 Pandemic disruption

ADNI4: 10=Study partner unwilling or unable to participate ; 12=Starting new trial ; 14=Coordinating Center request ; 15=Other ; 16=COVID-19 Pandemic disruption ; 2=Safety risk ; 4=Non-compliance ; 5=Adverse event ; 6=Investigator recommendation ; 7=Other non-site clinician recommendation ; 8=Started prohibited medication ; 9=Participant unwilling or unable to participate

Decoded Value: Yes

AENUM character If Adverse Event, provide Adverse Event number , with options: 1, 4, 3, 5,6, 2, 5, 8, 10, 11, 10, 1,2

ID numeric Record ID Range value: 433, ..., 186882

SITEID numeric Site ID Range value: 1, ..., 941

USERDATE User Create Date Date: YYY-MM-DD

USERDATE2 User Last Modify Date Date: YYY-MM-DD

DD_CRF_VERSION_LABEL character CRF Version Label , with options: newpt, rollover

LANGUAGE_CODE character Language Code e=English Decoded Value: Yes

HAS_QC_ERROR character Has Quality Check Error 0=Does not have QC error or QC error has been approved ; 1=Has QC error Decoded Value: Yes

update_stamp POSIXct: YYY-MM-DD

Author

adni-data@googlegroups.com