Early Discontinuation and Withdrawal

Early Discontinuation and Withdrawal data. More information is available at https://adni.loni.usc.edu/data-samples/adni-data/

Usage

data(TREATDIS)

Format

A data frame with 1889 observations and 18 variables.

Source

https://adni.loni.usc.edu/data-samples/adni-data/

Details

ORIGPROT character Original study protocol , with options: ADNI1, ADNIGO, ADNI2

COLPROT character Study protocol of data collection , with options: ADNI1, ADNIGO, ADNI2

PTID character Participant ID

RID numeric Participant roster ID Range value: 2, ..., 5295

VISCODE character Visit code

VISCODE2 character Translated visit code

RECNO numeric Range value: 1, ..., 6

EXAMDATE character Examination Date

EXAMDATEYR_DRVD numeric Year of Discontinuation/Withdrawal/Completion oF ADNI2 (Derived from EXAMDATE) Range value: 2000, ..., 2019

WDRAWTYPE numeric 1. Is this a Full or Partial Withdrawal?

ADNI1: 1=Full;2=Partial;3=Completion of ADNI 1, not continuing to extension protocol

ADNIGO: 1=Full;2=Partial;3=Completion of ADNIGO, not continuing to extension protocol

ADNI2: 1=Full;2=Partial

WDPARTIAL character If Partial, what is the participant withdrawing from?

ADNI1: 1=1.5 T MRI;2=3.0 T MRI;3=PET;4=Lumbar Puncture;5=PIB

ADNIGO: 1=1.5 T MRI;2=3.0 T MRI;3=FDG PET;6=AV-45 PET; 4=Lumbar Puncture;7=consent to store/share samples/clinical data;8=DNA sample collection; 9=consent to store and share DNA; 10=RNA sample collection;11=consent to store and share RNA; 12=provisional autopsy consent

ADNI2: 1=1.5 T MRI;2=3.0 T MRI;3=FDG PET;6=AV-45 PET; 20 =Tau AV-1451 F 18; 4=Lumbar Puncture;7=consent to store/share samples/clinical data;8=DNA sample collection; 9=consent to store and share DNA; 10=RNA sample collection;11=consent to store and share RNA; 12=provisional autopsy consent; 13=Full Clinic Visits (agrees to only telephone follow-up); 21 = Cogstate Brief Battery

WDREASON character Reason for Withdrawal

ADNI1: 1=Adverse Event;2=Death;3=Safety Risk;4=Protocol Violation;5=Non-Compliance;6=Investigator Judgment;7=Consent Withdrawn;8=Study Terminated;9=Loss of Study Partner;10=Lost to Follow-Up;11=Coordinating Center Request

ADNIGO: 1=Death (SAE required, provide details below); 2=Withdrawn Consent;3=Perceived Safety Risk;4=Protocol Violations; 5=Non-Compliance with Protocol;6=Adverse Event (provide detailed explanation below);7=Elects to enroll in clinical trial;8=Contraindication to Procedure (e.g. pacemaker);9=Elects to start an excluded FDA Approved treatment for AD (patient/caregiver);10=Participant Burden (provide detailed explanation below);11=Loss of Study Partner with no replacement available;12=Study Partner Burden;13=Study Partner unwilling or unable to participate (provide detailed explanation below) ;14=Subject Moved: unable to transfer or to travel distance for visits;15=Subject Moved to Nursing Facility; 16=Study Partner Moved: unable to transfer or to travel distance for visits; 17=Study Terminated (Mandated by Project Director or IRB); 18=Non-responsive (Registered mail returned, not showing for appointments, not answering phone calls); 19=Did not understand consent; 20=Compensation inadequate/ not competitive; 21=Other (provide detailed explanation below);

ADNI2: 1=Death (SAE required, provide details below); 2=Withdrawn Consent;3=Perceived Safety Risk;4=Protocol Violations; 5=Non-Compliance with Protocol;6=Adverse Event (provide detailed explanation below);7=Elects to enroll in clinical trial;8=Contraindication to Procedure (e.g. pacemaker);9=Elects to start an excluded FDA Approved treatment for AD (patient/caregiver);10=Participant Burden (provide detailed explanation below);11=Loss of Study Partner with no replacement available;12=Study Partner Burden;13=Study Partner unwilling or unable to participate (provide detailed explanation below) ;14=Subject Moved: unable to transfer or to travel distance for visits;15=Subject Moved to Nursing Facility; 16=Study Partner Moved: unable to transfer or to travel distance for visits; 17=Study Terminated (Mandated by Project Director or IRB); 18=Non-responsive (Registered mail returned, not showing for appointments, not answering phone calls); 19=Did not understand consent; 20=Compensation inadequate/ not competitive; 21=Other (provide detailed explanation below);

WDFOLLOWUP character 2. Follow-up: (check all that apply)

ADNI1: 1=Agrees to return for all follow-up visits.;2=Agrees to return for final visit.;3=Agrees to unscheduled visit.;4=Refuses/unable to return for any future visits.

ADNIGO: 1=Agrees to return for all follow-up visits.;2=Agrees to return for final visit.;4=Refuses/unable to return for any future visits.

ADNI2: 1=Agrees to return for all follow-up visits.;5=Agrees to follow-up phone visits only;2=Agrees to return for a final in-clinic visit.;4=Refuses/unable to return for any future visits.

ID numeric Range value: 2, ..., 2346

SITEID numeric Site ID Range value: 1, ..., 127

USERDATE Date Date record created Date: YYY-MM-DD

USERDATE2 Date Date record last updated Date: YYY-MM-DD

update_stamp POSIXct POSIXct: YYY-MM-DD

Author

adni-data@googlegroups.com