UA (Gene Alexander) MRI SPM voxel based morphometry (VBM) analysis

UA (Gene Alexander) MRI SPM voxel based morphometry (VBM) analysis data. More information is available at https://adni.loni.usc.edu/data-samples/adni-data/

Usage

data(UASPMVBM)

Format

A data frame with 4091 observations and 128 variables.

Source

https://adni.loni.usc.edu/data-samples/adni-data/

Details

ORIGPROT character Original study protocol , with options: ADNI1

ID numeric Range value: 1, ..., 4091

RID numeric Range value: 3, ..., 1435

SITEID numeric Range value: 1, ..., 127

VISCODE character , with options: sc, m06, m12, bl, f

USERDATE Date: YYY-MM-DD

RECNO numeric 1 = 1.5T cross-sectional screen ; 2 = 1.5T MO_06 longitudinal ; 3 = 1.5T MO_12 longitudinal ; 4 = 3T cross-sectional baseline ; 5 = 3T MO_06 longitudinal ; 6 = 3T MO_12 longitudinal

EXAMDATE Date: YYY-MM-DD

IMAGEUID numeric imageUID Range value: 31084, ..., 125949

FLDSTRENG numeric Field Strength Range value: 1.5, ..., 3

PRECENTL numeric Precentral_L. 1) Mean gray matter values extracted from 1.5T ADNI MPRAGE MRI segmented gray matter maps using longitudinal voxel based morphometry (VBM) processing steps (Chetelat et al., Neuroimage, 2005, 27:934-946) adapted for SPM5 with customized tissue-specific priors derived from a representative sample of 300 ADNI subjects (ADNI300). The regions of interest (ROI) were defined by Anatomical Automatic Labeling (AAL, Tzourio-Mazoyer et al., Neuroimage, 2002, 15:273-89, http://www.cyceron.fr/freeware/) after warping the standard AAL ROI template into the customized sample-specific ADNI300 template space. The ADNI300-customized AAL ROI template was applied to spatially normalized and segmented SPM5 gray matter maps for the longitudinal VBM generated baseline and 6 month (indicated by RECNO=2) or 12 month (indicated by RECNO=3) follow up scans without Gaussian smoothing to reduce averaging across regional borders. 2) The ADNI300 template was created from AD, MCI, and Normal subjects proportional to the distribution of these subject groups enrolled in the whole ADNI study and the groups were matched in age and gender. 3) Regional values will be subject to change due to study alterations in image pre- or post-processing, inclusion of additional analysis time points, and/or modifications or updates in image analysis procedures. 4) SPM analyses were performed by ADNI Analysis Team Members: Gene E. Alexander, Ph.D., Kewei Chen, Ph.D., and Eric M. Reiman, M.D. Range value: 0.109464, ..., 0.519062

PRECENTR numeric Precentral_R Range value: 0.087301, ..., 0.519366

FRONTSUPL numeric Frontal_Sup_L Range value: 0.072548, ..., 0.572452

FRONTSUPR numeric Frontal_Sup_R Range value: 0.06832, ..., 0.495008

FRONTSORBL numeric Frontal_Sup_Orb_L Range value: 0.079956, ..., 0.664292

FRONTSORBR numeric Frontal_Sup_Orb_R Range value: 0.070114, ..., 0.646994

FRONTMIDL numeric Frontal_Mid_L Range value: 0.093176, ..., 0.60747

FRONTMIDR numeric Frontal_Mid_R Range value: 0.0698, ..., 0.626904

FRTMIDORBL numeric Frontal_Mid_Orb_L Range value: 0.073239, ..., 0.75648

FRTMIDORBR numeric Frontal_Mid_Orb_R Range value: 0.061234, ..., 0.745696

FRONTINOPL numeric Frontal_Inf_Oper_L Range value: 0.131869, ..., 0.626053

FRONTINOPR numeric Frontal_Inf_Oper_R Range value: 0.115524, ..., 0.628719

FRONTINTRL numeric Frontal_Inf_Tri_L Range value: 0.139826, ..., 0.563734

FRONTINTRR numeric Frontal_Inf_Tri_R Range value: 0.110499, ..., 0.569364

FRONTINOBL numeric Frontal_Inf_Orb_L Range value: 0.113868, ..., 0.750045

FRONTINOBR numeric Frontal_Inf_Orb_R Range value: 0.092592, ..., 0.696247

ROLANDOPL numeric Rolandic_Oper_L Range value: 0.152697, ..., 0.668278

ROLANDOPR numeric Rolandic_Oper_R Range value: 0.163112, ..., 0.689594

SUPMOTORL numeric Supp_Motor_Area_L Range value: 0.083472, ..., 0.622019

SUPMOTORR numeric Supp_Motor_Area_R Range value: 0.113827, ..., 0.607832

OLFACTL numeric Olfactory_L Range value: 0.171352, ..., 0.813431

OLFACTR numeric Olfactory_R Range value: 0.172559, ..., 0.758829

FRONTSMEDL numeric Frontal_Sup_Medial_L Range value: 0.074001, ..., 0.581061

FRONTSMEDR numeric Frontal_Sup_Medial_R Range value: 0.084855, ..., 0.660669

FRTMEDORBL numeric Frontal_Med_Orb_L Range value: 0.129765, ..., 0.845215

FRTMEDORBR numeric Frontal_Med_Orb_R Range value: 0.150819, ..., 0.810541

RECTUSL numeric Rectus_L Range value: 0.136071, ..., 0.755095

RECTUSR numeric Rectus_R Range value: 0.15509, ..., 0.742195

INSULAL numeric Insula_L Range value: 0.159247, ..., 0.76472

INSULAR numeric Insula_R Range value: 0.149596, ..., 0.760286

CINGANTL numeric Cingulum_Ant_L Range value: 0.130653, ..., 0.660412

CINGANTR numeric Cingulum_Ant_R Range value: 0.200086, ..., 0.667653

CINGMIDL numeric Cingulum_Mid_L Range value: 0.203155, ..., 0.648047

CINGMIDR numeric Cingulum_Mid_R Range value: 0.227015, ..., 0.736913

CINGPOSTL numeric Cingulum_Post_L Range value: 0.141583, ..., 0.730277

CINGPOSTR numeric Cingulum_Post_R Range value: 0.120043, ..., 0.623823

HIPPL numeric Hippocampus_L Range value: 0.195745, ..., 0.862557

HIPPR numeric Hippocampus_R Range value: 0.079445, ..., 0.838027

PARAHIPPL numeric ParaHippocampal_L Range value: 0.194895, ..., 0.785074

PARAHIPPR numeric Range value: 0.269733, ..., 0.811848

AMYGDL numeric Range value: 0.125053, ..., 0.970368

AMYGDR numeric Amygdala_R Range value: 0.157966, ..., 0.860282

CALCARINEL numeric Calcarine_L Range value: 0.135687, ..., 0.776054

CALCARINER numeric Calcarine_R Range value: 0.142091, ..., 0.721373

CUNEUSL numeric Range value: 0.106092, ..., 0.711471

CUNEUSR numeric Range value: 0.120579, ..., 0.694838

LINGUALL numeric Range value: 0.2063, ..., 0.734729

LINGUALR numeric Range value: 0.180107, ..., 0.715062

OCCSUPL numeric Range value: 0.100288, ..., 0.558346

OCCSUPR numeric Range value: 0.110552, ..., 0.547554

OCCMIDL numeric Range value: 0.149659, ..., 0.69116

OCCMIDR numeric Range value: 0.128717, ..., 0.671495

OCCINFL numeric Range value: 0.15837, ..., 0.783026

OCCINFR numeric Range value: 0.105393, ..., 0.814613

FUSIFORML numeric Range value: 0.243454, ..., 0.814792

FUSIFORMR numeric Range value: 0.214296, ..., 0.806602

POSTCENTL numeric Range value: 0.132477, ..., 0.539285

POSTCENTR numeric Range value: 0.098583, ..., 0.525414

PARIETSUPL numeric Range value: 0.106263, ..., 0.645473

PARIETSUPR numeric Range value: 0.082337, ..., 0.599451

PARIETINFL numeric Range value: 0.161056, ..., 0.660393

PARIETINFR numeric Range value: 0.1459, ..., 0.691876

SUPRAMARGL numeric Range value: 0.200084, ..., 0.687163

SUPRAMARGR numeric Range value: 0.152732, ..., 0.625606

ANGULARL numeric Range value: 0.178645, ..., 0.71849

ANGULARR numeric Range value: 0.15023, ..., 0.641721

PRECUNEUSL numeric Range value: 0.152047, ..., 0.633392

PRECUNEUSR numeric Range value: 0.164068, ..., 0.710013

PARCENTLBL numeric Range value: 0.081076, ..., 0.550229

PARCENTLBR numeric Range value: 0.095236, ..., 0.600817

CAUDATEL numeric Range value: 0.101717, ..., 0.757027

CAUDATER numeric Range value: 0.128286, ..., 0.719425

PUTAMENL numeric Range value: 0.23163, ..., 0.857368

PUTAMENR numeric Range value: 0.241973, ..., 0.752296

PALLIDUML numeric Range value: 0.025142, ..., 0.584371

PALLIDUMR numeric Range value: 0.001927, ..., 0.636137

THALAMUSL numeric Range value: 0.159741, ..., 0.583169

THALAMUSR numeric Range value: 0.13538, ..., 0.611456

HESCHLL numeric Range value: 0.123076, ..., 0.703712

HESCHLR numeric Range value: 0.094019, ..., 0.605857

TEMPSUPL numeric Range value: 0.213788, ..., 0.643675

TEMPSUPR numeric Range value: 0.152122, ..., 0.618064

TEMPPLSUPL numeric Range value: 0.130266, ..., 0.721552

TEMPPLSUPR numeric Range value: 0.103663, ..., 0.641147

TEMPMIDL numeric Range value: 0.216623, ..., 0.677134

TEMPMIDR numeric Range value: 0.166822, ..., 0.679876

TEMPPLMIDL numeric Range value: 0.088095, ..., 0.800392

TEMPPLMIDR numeric Range value: 0.071037, ..., 0.723863

TEMPINFL numeric Range value: 0.190501, ..., 0.710179

TEMPINFR numeric Range value: 0.147539, ..., 0.749946

CEREBCR1L numeric Range value: 0.227836, ..., 0.800763

CEREBCR1R numeric Range value: 0.176948, ..., 0.825526

CEREBCR2L numeric Range value: 0.181098, ..., 1.47167

CEREBCR2R numeric Range value: 0.156126, ..., 1.1756

CEREB3L numeric Range value: 0.175525, ..., 0.800511

CEREB3R numeric Range value: 0.182603, ..., 0.682015

CEREB45L numeric Range value: 0.297412, ..., 0.78897

CEREB45R numeric Range value: 0.341734, ..., 0.858938

CEREB6L numeric Range value: 0.360521, ..., 0.912022

CEREB6R numeric Range value: 0.36296, ..., 0.859137

CEREB7BL numeric Range value: 0.22627, ..., 1.84525

CEREB7BR numeric Range value: 0.189325, ..., 1.63047

CEREB8L numeric Range value: 0.243068, ..., 2.01657

CEREB8R numeric Range value: 0.218044, ..., 1.68857

CEREB9L numeric Range value: 0.260033, ..., 1.40136

CEREB9R numeric Range value: 0.126837, ..., 1.1916

CEREB10L numeric Range value: 0.075762, ..., 0.71962

CEREB10R numeric Range value: 0.10381, ..., 0.752279

VERMIS12 numeric Range value: 0.08083, ..., 0.818248

VERMIS3 numeric Range value: 0.155702, ..., 0.796297

VERMIS45 numeric Range value: 0.207042, ..., 0.710177

VERMIS6 numeric Range value: 0.187682, ..., 0.841051

VERMIS7 numeric Range value: 0.210638, ..., 0.79367

VERMIS8 numeric Range value: 0.286865, ..., 0.994217

VERMIS9 numeric Range value: 0.365639, ..., 0.956039

VERMIS10 numeric Range value: 0.042025, ..., 0.684707

ETIV numeric Range value: 1.33109, ..., 2.62237

update_stamp POSIXct: YYY-MM-DD

Author

adni-data@googlegroups.com