UC Berkeley - Amyloid PET 6mm Res analysis

UC Berkeley - Amyloid PET 6mm Res analysis data. More information is available at https://adni.loni.usc.edu/data-samples/adni-data/

Usage

data(UCBERKELEY_AMY_6MM)

Format

A data frame with 4127 observations and 344 variables.

Source

https://adni.loni.usc.edu/data-samples/adni-data/

Details

ORIGPROT character Original study protocol , with options: ADNI1, ADNI2, ADNI3, ADNIGO, ADNI4

LONIUID character LONI Image ID

RID numeric Participant roster ID (unique participant identifier that should be used to merge data) Range value: 21, ..., 10340

VISCODE character Visit code

SITEID numeric Site ID Range value: 2, ..., 941

PTID character ADNI Participant ID consisting of 3-digit "SITEID" and 4-digit "RID"

SCANDATE Date Scan acquisition date Date: YYY-MM-DD

PROCESSDATE Date Date quantification pipeline was run Date: YYY-MM-DD

IMAGE_RESOLUTION character Spatial resolution in mm^3 FWHM of Step4 Preprocessed PET image. See UCBERKELEY PET 8mm-to-6mm Transformation Methods PDF on LONI for details. , with options: 6mm, 8mm

qc_flag numeric Range value: -2, ..., 2

TRACER character PET Radiotracer , with options: FBB, FBP

TRACER_SUVR_WARNING character Reminder of the incomparability of SUVRs from different tracers. See UCBERKELEY Amyloid Processing Methods PDF. , with options: USE CENTILOIDS (NOT SUVRs) TO COMPARE ACROSS TRACERS

AMYLOID_STATUS numeric Amyloid positivity determined by applying thresholds to cortical summary SUVR normalized by whole cerebellum; See UCBERKELEY Amyloid Processing Methods PDF on LONI for details. Range value: 0, ..., 1

AMYLOID_STATUS_COMPOSITE_REF numeric Amyloid positivity determined by applying thresholds to cortical summary SUVR normalized by composite ref region; See UCBERKELEY Amyloid Processing Methods PDF on LONI for details. Range value: 0, ..., 1

CENTILOIDS numeric Summary cortical SUVR normalized by whole cerebellum and transformed to Centiloids. All SUVRs in this file are normalized to the whole cerebellum unless otherwise noted. See UCBERKELEY Amyloid Processing Methods PDF on LONI for details. Range value: -39, ..., 346

SUMMARY_SUVR numeric Summary cortical SUVR normalized by whole cerebellum. All SUVRs in this file are normalized to the whole cerebellum unless otherwise noted. See UCBERKELEY Amyloid Processing Methods PDF on LONI for details. Range value: 0.797, ..., 2.842

SUMMARY_VOLUME numeric Summary cortical ROI size in mm^3 Range value: 179914, ..., 423018

WHOLECEREBELLUM_SUVR numeric Reference region - mean SUVR of whole cerebellum. Regions defined by Freesurfer; See UCBERKELEY Amyloid Processing Methods PDF on LONI for details. Range value: 1, ..., 1

WHOLECEREBELLUM_VOLUME numeric Reference region - volume of whole cerebellum. Regions defined by Freesurfer; See UCBERKELEY Amyloid Processing Methods PDF on LONI for details. Range value: 85374, ..., 180735

COMPOSITE_REF_SUVR numeric Reference region - mean SUVR of composite ref region (avg of whole cerebellum, brainstem, eroded WM). Regions defined by Freesurfer; See UCBERKELEY Amyloid Processing Methods PDF on LONI for details. Range value: 1.078, ..., 2.362

COMPOSITE_REF_VOLUME numeric Reference region - volume of composite ref region (avg of whole cerebellum, brainstem, eroded WM). Regions defined by Freesurfer; See UCBERKELEY Amyloid Processing Methods PDF on LONI for details. Range value: 173936, ..., 502106

CEREBELLUM_CORTEX_SUVR numeric Reference region - mean SUVR of bilateral-cerebellum-cortex defined by Freesurfer (previously CEREBELLUMGREYMATTER_SUVR); See UCBERKELEY Amyloid Processing Methods PDF on LONI for details. Range value: 0.754, ..., 0.955

CEREBELLUM_CORTEX_VOLUME numeric Reference region - volume of bilateral-cerebellum-cortex defined by Freesurfer (previously CEREBELLUMGREYMATTER_VOLUME); See UCBERKELEY Amyloid Processing Methods PDF on LONI for details. Range value: 68832, ..., 142592

ERODED_SUBCORTICALWM_SUVR numeric Reference region - mean SUVR of eroded subcortical white matter. Regions defined by Freesurfer; See UCBERKELEY Amyloid Processing Methods PDF on LONI for details. Range value: 1.192, ..., 3.97

ERODED_SUBCORTICALWM_VOLUME numeric Reference region - volume of eroded subcortical white matter. Regions defined by Freesurfer; See UCBERKELEY Amyloid Processing Methods PDF on LONI for details. Range value: 55673, ..., 313526

BRAINSTEM_SUVR numeric Reference region - mean SUVR of brainstem. Regions defined by Freesurfer; See UCBERKELEY Amyloid Processing Methods PDF on LONI for details. Range value: 1.043, ..., 2.116

BRAINSTEM_VOLUME numeric brain-stem ROI size in mm^3 Range value: 13764, ..., 31506

CC_ANTERIOR_SUVR numeric cc-anterior Range value: 0.564, ..., 2.676

CC_ANTERIOR_VOLUME numeric cc-anterior ROI size in mm^3 Range value: 171, ..., 1688

CC_CENTRAL_SUVR numeric cc-central Range value: 0.361, ..., 2.497

CC_CENTRAL_VOLUME numeric cc-central ROI size in mm^3 Range value: 217, ..., 1297

CC_MID_ANTERIOR_SUVR numeric cc-mid-anterior Range value: 0.391, ..., 2.526

CC_MID_ANTERIOR_VOLUME numeric cc-mid-anterior ROI size in mm^3 Range value: 225, ..., 1170

CC_MID_POSTERIOR_SUVR numeric cc-mid-posterior Range value: 0.432, ..., 2.48

CC_MID_POSTERIOR_VOLUME numeric cc-mid-posterior ROI size in mm^3 Range value: 179, ..., 1796

CC_POSTERIOR_SUVR numeric cc-posterior Range value: 0.825, ..., 3.454

CC_POSTERIOR_VOLUME numeric cc-posterior ROI size in mm^3 Range value: 218, ..., 2196

CSF_SUVR numeric csf Range value: 0.339, ..., 1.194

CSF_VOLUME numeric csf ROI size in mm^3 Range value: 573, ..., 6061

VENTRICLE_3RD_SUVR numeric 3rd-ventricle Range value: 0.379, ..., 1.213

VENTRICLE_3RD_VOLUME numeric 3rd-ventricle ROI size in mm^3 Range value: 474, ..., 5244

VENTRICLE_4TH_SUVR numeric 4th-ventricle Range value: 0.541, ..., 1.299

VENTRICLE_4TH_VOLUME numeric 4th-ventricle ROI size in mm^3 Range value: 561, ..., 5586

VENTRICLE_5TH_SUVR numeric 5th-ventricle Range value: 0.422, ..., 1.414

VENTRICLE_5TH_VOLUME numeric 5th-ventricle ROI size in mm^3 Range value: 1, ..., 11

WM_HYPOINTENSITIES_SUVR numeric wm-hypointensities Range value: 0.742, ..., 3.046

WM_HYPOINTENSITIES_VOLUME numeric wm-hypointensities ROI size in mm^3 Range value: 268, ..., 72336

NON_WM_HYPOINTENSITIES_SUVR numeric non-wm-hypointensities Range value: 0.427, ..., 2.76

NON_WM_HYPOINTENSITIES_VOLUME numeric non-wm-hypointensities ROI size in mm^3 Range value: 1, ..., 233

CTX_BANKSSTS_SUVR numeric ctx-bilateral-bankssts Range value: 0.915, ..., 3.049

CTX_BANKSSTS_VOLUME numeric ctx-bilateral-bankssts ROI size in mm^3 Range value: 2094, ..., 8963

CTX_CAUDALANTERIORCINGULATE_SUVR numeric ctx-bilateral-caudalanteriorcingulate Range value: 0.806, ..., 2.95

CTX_CAUDALANTERIORCINGULATE_VOLUME numeric ctx-bilateral-caudalanteriorcingulate ROI size in mm^3 Range value: 1116, ..., 7847

CTX_CAUDALMIDDLEFRONTAL_SUVR numeric ctx-bilateral-caudalmiddlefrontal Range value: 0.786, ..., 2.915

CTX_CAUDALMIDDLEFRONTAL_VOLUME numeric ctx-bilateral-caudalmiddlefrontal ROI size in mm^3 Range value: 5472, ..., 21954

CTX_CUNEUS_SUVR numeric ctx-bilateral-cuneus Range value: 0.783, ..., 2.557

CTX_CUNEUS_VOLUME numeric ctx-bilateral-cuneus ROI size in mm^3 Range value: 3581, ..., 10596

CTX_ENTORHINAL_SUVR numeric ctx-bilateral-entorhinal Range value: 0.505, ..., 1.528

CTX_ENTORHINAL_VOLUME numeric ctx-bilateral-entorhinal ROI size in mm^3 Range value: 746, ..., 6911

CTX_FRONTALPOLE_SUVR numeric ctx-bilateral-frontalpole Range value: 0.421, ..., 2.557

CTX_FRONTALPOLE_VOLUME numeric ctx-bilateral-frontalpole ROI size in mm^3 Range value: 982, ..., 3825

CTX_FUSIFORM_SUVR numeric ctx-bilateral-fusiform Range value: 0.803, ..., 2.24

CTX_FUSIFORM_VOLUME numeric ctx-bilateral-fusiform ROI size in mm^3 Range value: 8002, ..., 26997

CTX_INFERIORPARIETAL_SUVR numeric ctx-bilateral-inferiorparietal Range value: 0.795, ..., 2.818

CTX_INFERIORPARIETAL_VOLUME numeric ctx-bilateral-inferiorparietal ROI size in mm^3 Range value: 11643, ..., 37678

CTX_INFERIORTEMPORAL_SUVR numeric ctx-bilateral-inferiortemporal Range value: 0.804, ..., 2.483

CTX_INFERIORTEMPORAL_VOLUME numeric ctx-bilateral-inferiortemporal ROI size in mm^3 Range value: 8623, ..., 31454

CTX_INSULA_SUVR numeric ctx-bilateral-insula Range value: 0.721, ..., 2.614

CTX_INSULA_VOLUME numeric ctx-bilateral-insula ROI size in mm^3 Range value: 7644, ..., 33551

CTX_ISTHMUSCINGULATE_SUVR numeric ctx-bilateral-isthmuscingulate Range value: 0.806, ..., 3.141

CTX_ISTHMUSCINGULATE_VOLUME numeric ctx-bilateral-isthmuscingulate ROI size in mm^3 Range value: 1985, ..., 16944

CTX_LATERALOCCIPITAL_SUVR numeric ctx-bilateral-lateraloccipital Range value: 0.802, ..., 2.764

CTX_LATERALOCCIPITAL_VOLUME numeric ctx-bilateral-lateraloccipital ROI size in mm^3 Range value: 9370, ..., 35627

CTX_LATERALORBITOFRONTAL_SUVR numeric ctx-bilateral-lateralorbitofrontal Range value: 0.801, ..., 2.682

CTX_LATERALORBITOFRONTAL_VOLUME numeric ctx-bilateral-lateralorbitofrontal ROI size in mm^3 Range value: 7937, ..., 22861

CTX_LINGUAL_SUVR numeric ctx-bilateral-lingual Range value: 0.77, ..., 2.127

CTX_LINGUAL_VOLUME numeric ctx-bilateral-lingual ROI size in mm^3 Range value: 6595, ..., 20629

CTX_MEDIALORBITOFRONTAL_SUVR numeric ctx-bilateral-medialorbitofrontal Range value: 0.764, ..., 2.676

CTX_MEDIALORBITOFRONTAL_VOLUME numeric ctx-bilateral-medialorbitofrontal ROI size in mm^3 Range value: 6400, ..., 15313

CTX_MIDDLETEMPORAL_SUVR numeric ctx-bilateral-middletemporal Range value: 0.785, ..., 2.503

CTX_MIDDLETEMPORAL_VOLUME numeric ctx-bilateral-middletemporal ROI size in mm^3 Range value: 9286, ..., 32354

CTX_PARACENTRAL_SUVR numeric ctx-bilateral-paracentral Range value: 0.835, ..., 3.277

CTX_PARACENTRAL_VOLUME numeric ctx-bilateral-paracentral ROI size in mm^3 Range value: 4084, ..., 13660

CTX_PARAHIPPOCAMPAL_SUVR numeric ctx-bilateral-parahippocampal Range value: 0.547, ..., 1.819

CTX_PARAHIPPOCAMPAL_VOLUME numeric ctx-bilateral-parahippocampal ROI size in mm^3 Range value: 1751, ..., 6474

CTX_PARSOPERCULARIS_SUVR numeric ctx-bilateral-parsopercularis Range value: 0.784, ..., 2.951

CTX_PARSOPERCULARIS_VOLUME numeric ctx-bilateral-parsopercularis ROI size in mm^3 Range value: 4303, ..., 12836

CTX_PARSORBITALIS_SUVR numeric ctx-bilateral-parsorbitalis Range value: 0.744, ..., 2.371

CTX_PARSORBITALIS_VOLUME numeric ctx-bilateral-parsorbitalis ROI size in mm^3 Range value: 2438, ..., 7089

CTX_PARSTRIANGULARIS_SUVR numeric ctx-bilateral-parstriangularis Range value: 0.778, ..., 2.972

CTX_PARSTRIANGULARIS_VOLUME numeric ctx-bilateral-parstriangularis ROI size in mm^3 Range value: 4260, ..., 11719

CTX_PERICALCARINE_SUVR numeric ctx-bilateral-pericalcarine Range value: 0.854, ..., 2.808

CTX_PERICALCARINE_VOLUME numeric ctx-bilateral-pericalcarine ROI size in mm^3 Range value: 1847, ..., 13257

CTX_POSTCENTRAL_SUVR numeric ctx-bilateral-postcentral Range value: 0.712, ..., 2.707

CTX_POSTCENTRAL_VOLUME numeric ctx-bilateral-postcentral ROI size in mm^3 Range value: 7517, ..., 28111

CTX_POSTERIORCINGULATE_SUVR numeric ctx-bilateral-posteriorcingulate Range value: 0.807, ..., 3.356

CTX_POSTERIORCINGULATE_VOLUME numeric ctx-bilateral-posteriorcingulate ROI size in mm^3 Range value: 1978, ..., 11175

CTX_PRECENTRAL_SUVR numeric ctx-bilateral-precentral Range value: 0.846, ..., 2.82

CTX_PRECENTRAL_VOLUME numeric ctx-bilateral-precentral ROI size in mm^3 Range value: 13766, ..., 43751

CTX_PRECUNEUS_SUVR numeric ctx-bilateral-precuneus Range value: 0.807, ..., 3.386

CTX_PRECUNEUS_VOLUME numeric ctx-bilateral-precuneus ROI size in mm^3 Range value: 9272, ..., 27581

CTX_ROSTRALANTERIORCINGULATE_SUVR numeric ctx-bilateral-rostralanteriorcingulate Range value: 0.77, ..., 2.933

CTX_ROSTRALANTERIORCINGULATE_VOLUME numeric ctx-bilateral-rostralanteriorcingulate ROI size in mm^3 Range value: 1558, ..., 8525

CTX_ROSTRALMIDDLEFRONTAL_SUVR numeric ctx-bilateral-rostralmiddlefrontal Range value: 0.786, ..., 3.074

CTX_ROSTRALMIDDLEFRONTAL_VOLUME numeric ctx-bilateral-rostralmiddlefrontal ROI size in mm^3 Range value: 12486, ..., 50046

CTX_SUPERIORFRONTAL_SUVR numeric ctx-bilateral-superiorfrontal Range value: 0.762, ..., 2.867

CTX_SUPERIORFRONTAL_VOLUME numeric ctx-bilateral-superiorfrontal ROI size in mm^3 Range value: 23254, ..., 60833

CTX_SUPERIORPARIETAL_SUVR numeric ctx-bilateral-superiorparietal Range value: 0.725, ..., 2.824

CTX_SUPERIORPARIETAL_VOLUME numeric ctx-bilateral-superiorparietal ROI size in mm^3 Range value: 13205, ..., 36908

CTX_SUPERIORTEMPORAL_SUVR numeric ctx-bilateral-superiortemporal Range value: 0.755, ..., 2.602

CTX_SUPERIORTEMPORAL_VOLUME numeric ctx-bilateral-superiortemporal ROI size in mm^3 Range value: 11398, ..., 32667

CTX_SUPRAMARGINAL_SUVR numeric ctx-bilateral-supramarginal Range value: 0.762, ..., 3.014

CTX_SUPRAMARGINAL_VOLUME numeric ctx-bilateral-supramarginal ROI size in mm^3 Range value: 8902, ..., 30284

CTX_TEMPORALPOLE_SUVR numeric ctx-bilateral-temporalpole Range value: 0.535, ..., 1.813

CTX_TEMPORALPOLE_VOLUME numeric ctx-bilateral-temporalpole ROI size in mm^3 Range value: 1332, ..., 9644

CTX_TRANSVERSETEMPORAL_SUVR numeric ctx-bilateral-transversetemporal Range value: 0.661, ..., 2.907

CTX_TRANSVERSETEMPORAL_VOLUME numeric ctx-bilateral-transversetemporal ROI size in mm^3 Range value: 923, ..., 3541

ACCUMBENS_AREA_SUVR numeric bilateral-accumbens-area Range value: 0.62, ..., 2.61

ACCUMBENS_AREA_VOLUME numeric bilateral-accumbens-area ROI size in mm^3 Range value: 325, ..., 1636

AMYGDALA_SUVR numeric bilateral-amygdala Range value: 0.551, ..., 1.599

AMYGDALA_VOLUME numeric bilateral-amygdala ROI size in mm^3 Range value: 934, ..., 5040

CAUDATE_SUVR numeric bilateral-caudate Range value: 0.611, ..., 2.055

CAUDATE_VOLUME numeric bilateral-caudate ROI size in mm^3 Range value: 4068, ..., 13237

CEREBELLUM_WHITE_MATTER_SUVR numeric bilateral-cerebellum-white-matter Range value: 1.196, ..., 2.079

CEREBELLUM_WHITE_MATTER_VOLUME numeric bilateral-cerebellum-white-matter ROI size in mm^3 Range value: 13671, ..., 49420

CEREBRAL_WHITE_MATTER_SUVR numeric bilateral-cerebral-white-matter Range value: 1.13, ..., 3.527

CEREBRAL_WHITE_MATTER_VOLUME numeric bilateral-cerebral-white-matter ROI size in mm^3 Range value: 238555, ..., 664133

CHOROID_PLEXUS_SUVR numeric bilateral-choroid-plexus Range value: 0.399, ..., 1.434

CHOROID_PLEXUS_VOLUME numeric bilateral-choroid-plexus ROI size in mm^3 Range value: 459, ..., 3832

HIPPOCAMPUS_SUVR numeric bilateral-hippocampus Range value: 0.653, ..., 1.538

HIPPOCAMPUS_VOLUME numeric bilateral-hippocampus ROI size in mm^3 Range value: 3255, ..., 11921

OPTIC_CHIASM_SUVR numeric optic-chiasm Range value: 0.336, ..., 1.146

OPTIC_CHIASM_VOLUME numeric optic-chiasm ROI size in mm^3 Range value: 11, ..., 356

INF_LAT_VENT_SUVR numeric bilateral-inf-lat-vent Range value: 0.479, ..., 1.805

INF_LAT_VENT_VOLUME numeric bilateral-inf-lat-vent ROI size in mm^3 Range value: 164, ..., 16102

LATERAL_VENTRICLE_SUVR numeric bilateral-lateral-ventricle Range value: 0.284, ..., 1.441

LATERAL_VENTRICLE_VOLUME numeric bilateral-lateral-ventricle ROI size in mm^3 Range value: 5800, ..., 146015

PALLIDUM_SUVR numeric bilateral-pallidum Range value: 1.007, ..., 2.579

PALLIDUM_VOLUME numeric bilateral-pallidum ROI size in mm^3 Range value: 2306, ..., 7048

PUTAMEN_SUVR numeric bilateral-putamen Range value: 0.919, ..., 2.525

PUTAMEN_VOLUME numeric bilateral-putamen ROI size in mm^3 Range value: 4023, ..., 17943

THALAMUS_PROPER_SUVR numeric bilateral-thalamus-proper Range value: 0.784, ..., 1.828

THALAMUS_PROPER_VOLUME numeric bilateral-thalamus-proper ROI size in mm^3 Range value: 8557, ..., 24225

VENTRALDC_SUVR numeric bilateral-ventraldc Range value: 0.983, ..., 2.249

VENTRALDC_VOLUME numeric bilateral-ventraldc ROI size in mm^3 Range value: 4643, ..., 11189

VESSEL_SUVR numeric bilateral-vessel Range value: 0.537, ..., 1.885

VESSEL_VOLUME numeric bilateral-vessel ROI size in mm^3 Range value: 1, ..., 604

CTX_LH_BANKSSTS_SUVR numeric ctx-lh-bankssts Range value: 0.822, ..., 3.159

CTX_LH_BANKSSTS_VOLUME numeric ctx-lh-bankssts ROI size in mm^3 Range value: 960, ..., 5268

CTX_LH_CAUDALANTERIORCINGULATE_SUVR numeric ctx-lh-caudalanteriorcingulate Range value: 0.803, ..., 3.029

CTX_LH_CAUDALANTERIORCINGULATE_VOLUME numeric ctx-lh-caudalanteriorcingulate ROI size in mm^3 Range value: 442, ..., 5165

CTX_LH_CAUDALMIDDLEFRONTAL_SUVR numeric ctx-lh-caudalmiddlefrontal Range value: 0.798, ..., 3.059

CTX_LH_CAUDALMIDDLEFRONTAL_VOLUME numeric ctx-lh-caudalmiddlefrontal ROI size in mm^3 Range value: 2037, ..., 16456

CTX_LH_CUNEUS_SUVR numeric ctx-lh-cuneus Range value: 0.805, ..., 2.504

CTX_LH_CUNEUS_VOLUME numeric ctx-lh-cuneus ROI size in mm^3 Range value: 1627, ..., 6010

CTX_LH_ENTORHINAL_SUVR numeric ctx-lh-entorhinal Range value: 0.511, ..., 1.595

CTX_LH_ENTORHINAL_VOLUME numeric ctx-lh-entorhinal ROI size in mm^3 Range value: 139, ..., 3806

CTX_LH_FRONTALPOLE_SUVR numeric ctx-lh-frontalpole Range value: 0.43, ..., 2.818

CTX_LH_FRONTALPOLE_VOLUME numeric ctx-lh-frontalpole ROI size in mm^3 Range value: 417, ..., 1953

CTX_LH_FUSIFORM_SUVR numeric ctx-lh-fusiform Range value: 0.82, ..., 2.311

CTX_LH_FUSIFORM_VOLUME numeric ctx-lh-fusiform ROI size in mm^3 Range value: 3185, ..., 17168

CTX_LH_INFERIORPARIETAL_SUVR numeric ctx-lh-inferiorparietal Range value: 0.801, ..., 2.97

CTX_LH_INFERIORPARIETAL_VOLUME numeric ctx-lh-inferiorparietal ROI size in mm^3 Range value: 4994, ..., 17983

CTX_LH_INFERIORTEMPORAL_SUVR numeric ctx-lh-inferiortemporal Range value: 0.775, ..., 2.614

CTX_LH_INFERIORTEMPORAL_VOLUME numeric ctx-lh-inferiortemporal ROI size in mm^3 Range value: 4040, ..., 16913

CTX_LH_INSULA_SUVR numeric ctx-lh-insula Range value: 0.696, ..., 2.742

CTX_LH_INSULA_VOLUME numeric ctx-lh-insula ROI size in mm^3 Range value: 3855, ..., 27008

CTX_LH_ISTHMUSCINGULATE_SUVR numeric ctx-lh-isthmuscingulate Range value: 0.83, ..., 3.278

CTX_LH_ISTHMUSCINGULATE_VOLUME numeric ctx-lh-isthmuscingulate ROI size in mm^3 Range value: 615, ..., 14857

CTX_LH_LATERALOCCIPITAL_SUVR numeric ctx-lh-lateraloccipital Range value: 0.737, ..., 2.639

CTX_LH_LATERALOCCIPITAL_VOLUME numeric ctx-lh-lateraloccipital ROI size in mm^3 Range value: 4386, ..., 20415

CTX_LH_LATERALORBITOFRONTAL_SUVR numeric ctx-lh-lateralorbitofrontal Range value: 0.813, ..., 2.776

CTX_LH_LATERALORBITOFRONTAL_VOLUME numeric ctx-lh-lateralorbitofrontal ROI size in mm^3 Range value: 3652, ..., 15394

CTX_LH_LINGUAL_SUVR numeric ctx-lh-lingual Range value: 0.727, ..., 2.184

CTX_LH_LINGUAL_VOLUME numeric ctx-lh-lingual ROI size in mm^3 Range value: 2975, ..., 10566

CTX_LH_MEDIALORBITOFRONTAL_SUVR numeric ctx-lh-medialorbitofrontal Range value: 0.743, ..., 2.792

CTX_LH_MEDIALORBITOFRONTAL_VOLUME numeric ctx-lh-medialorbitofrontal ROI size in mm^3 Range value: 2937, ..., 9820

CTX_LH_MIDDLETEMPORAL_SUVR numeric ctx-lh-middletemporal Range value: 0.747, ..., 2.603

CTX_LH_MIDDLETEMPORAL_VOLUME numeric ctx-lh-middletemporal ROI size in mm^3 Range value: 3320, ..., 15289

CTX_LH_PARACENTRAL_SUVR numeric ctx-lh-paracentral Range value: 0.774, ..., 3.554

CTX_LH_PARACENTRAL_VOLUME numeric ctx-lh-paracentral ROI size in mm^3 Range value: 1709, ..., 9133

CTX_LH_PARAHIPPOCAMPAL_SUVR numeric ctx-lh-parahippocampal Range value: 0.585, ..., 2.127

CTX_LH_PARAHIPPOCAMPAL_VOLUME numeric ctx-lh-parahippocampal ROI size in mm^3 Range value: 885, ..., 4346

CTX_LH_PARSOPERCULARIS_SUVR numeric ctx-lh-parsopercularis Range value: 0.792, ..., 3.107

CTX_LH_PARSOPERCULARIS_VOLUME numeric ctx-lh-parsopercularis ROI size in mm^3 Range value: 2298, ..., 8062

CTX_LH_PARSORBITALIS_SUVR numeric ctx-lh-parsorbitalis Range value: 0.7, ..., 2.537

CTX_LH_PARSORBITALIS_VOLUME numeric ctx-lh-parsorbitalis ROI size in mm^3 Range value: 438, ..., 3562

CTX_LH_PARSTRIANGULARIS_SUVR numeric ctx-lh-parstriangularis Range value: 0.76, ..., 3.055

CTX_LH_PARSTRIANGULARIS_VOLUME numeric ctx-lh-parstriangularis ROI size in mm^3 Range value: 1547, ..., 6687

CTX_LH_PERICALCARINE_SUVR numeric ctx-lh-pericalcarine Range value: 0.796, ..., 2.652

CTX_LH_PERICALCARINE_VOLUME numeric ctx-lh-pericalcarine ROI size in mm^3 Range value: 899, ..., 11046

CTX_LH_POSTCENTRAL_SUVR numeric ctx-lh-postcentral Range value: 0.627, ..., 2.88

CTX_LH_POSTCENTRAL_VOLUME numeric ctx-lh-postcentral ROI size in mm^3 Range value: 4591, ..., 18003

CTX_LH_POSTERIORCINGULATE_SUVR numeric ctx-lh-posteriorcingulate Range value: 0.775, ..., 3.484

CTX_LH_POSTERIORCINGULATE_VOLUME numeric ctx-lh-posteriorcingulate ROI size in mm^3 Range value: 1083, ..., 8475

CTX_LH_PRECENTRAL_SUVR numeric ctx-lh-precentral Range value: 0.83, ..., 2.977

CTX_LH_PRECENTRAL_VOLUME numeric ctx-lh-precentral ROI size in mm^3 Range value: 6606, ..., 31664

CTX_LH_PRECUNEUS_SUVR numeric ctx-lh-precuneus Range value: 0.805, ..., 3.599

CTX_LH_PRECUNEUS_VOLUME numeric ctx-lh-precuneus ROI size in mm^3 Range value: 4263, ..., 19194

CTX_LH_ROSTRALANTERIORCINGULATE_SUVR numeric ctx-lh-rostralanteriorcingulate Range value: 0.771, ..., 3.013

CTX_LH_ROSTRALANTERIORCINGULATE_VOLUME numeric ctx-lh-rostralanteriorcingulate ROI size in mm^3 Range value: 571, ..., 6345

CTX_LH_ROSTRALMIDDLEFRONTAL_SUVR numeric ctx-lh-rostralmiddlefrontal Range value: 0.784, ..., 3.144

CTX_LH_ROSTRALMIDDLEFRONTAL_VOLUME numeric ctx-lh-rostralmiddlefrontal ROI size in mm^3 Range value: 5531, ..., 24802

CTX_LH_SUPERIORFRONTAL_SUVR numeric ctx-lh-superiorfrontal Range value: 0.77, ..., 3.048

CTX_LH_SUPERIORFRONTAL_VOLUME numeric ctx-lh-superiorfrontal ROI size in mm^3 Range value: 11052, ..., 40962

CTX_LH_SUPERIORPARIETAL_SUVR numeric ctx-lh-superiorparietal Range value: 0.735, ..., 3.068

CTX_LH_SUPERIORPARIETAL_VOLUME numeric ctx-lh-superiorparietal ROI size in mm^3 Range value: 5489, ..., 24136

CTX_LH_SUPERIORTEMPORAL_SUVR numeric ctx-lh-superiortemporal Range value: 0.747, ..., 2.744

CTX_LH_SUPERIORTEMPORAL_VOLUME numeric ctx-lh-superiortemporal ROI size in mm^3 Range value: 4632, ..., 21023

CTX_LH_SUPRAMARGINAL_SUVR numeric ctx-lh-supramarginal Range value: 0.77, ..., 3.212

CTX_LH_SUPRAMARGINAL_VOLUME numeric ctx-lh-supramarginal ROI size in mm^3 Range value: 5209, ..., 21200

CTX_LH_TEMPORALPOLE_SUVR numeric ctx-lh-temporalpole Range value: 0.527, ..., 1.946

CTX_LH_TEMPORALPOLE_VOLUME numeric ctx-lh-temporalpole ROI size in mm^3 Range value: 48, ..., 5851

CTX_LH_TRANSVERSETEMPORAL_SUVR numeric ctx-lh-transversetemporal Range value: 0.664, ..., 3.141

CTX_LH_TRANSVERSETEMPORAL_VOLUME numeric ctx-lh-transversetemporal ROI size in mm^3 Range value: 577, ..., 2282

CTX_RH_BANKSSTS_SUVR numeric ctx-rh-bankssts Range value: 0.87, ..., 2.951

CTX_RH_BANKSSTS_VOLUME numeric ctx-rh-bankssts ROI size in mm^3 Range value: 873, ..., 4240

CTX_RH_CAUDALANTERIORCINGULATE_SUVR numeric ctx-rh-caudalanteriorcingulate Range value: 0.807, ..., 2.88

CTX_RH_CAUDALANTERIORCINGULATE_VOLUME numeric ctx-rh-caudalanteriorcingulate ROI size in mm^3 Range value: 454, ..., 3780

CTX_RH_CAUDALMIDDLEFRONTAL_SUVR numeric ctx-rh-caudalmiddlefrontal Range value: 0.775, ..., 2.745

CTX_RH_CAUDALMIDDLEFRONTAL_VOLUME numeric ctx-rh-caudalmiddlefrontal ROI size in mm^3 Range value: 2492, ..., 9630

CTX_RH_CUNEUS_SUVR numeric ctx-rh-cuneus Range value: 0.744, ..., 2.618

CTX_RH_CUNEUS_VOLUME numeric ctx-rh-cuneus ROI size in mm^3 Range value: 982, ..., 6284

CTX_RH_ENTORHINAL_SUVR numeric ctx-rh-entorhinal Range value: 0.495, ..., 1.576

CTX_RH_ENTORHINAL_VOLUME numeric ctx-rh-entorhinal ROI size in mm^3 Range value: 455, ..., 3722

CTX_RH_FRONTALPOLE_SUVR numeric ctx-rh-frontalpole Range value: 0.408, ..., 2.343

CTX_RH_FRONTALPOLE_VOLUME numeric ctx-rh-frontalpole ROI size in mm^3 Range value: 381, ..., 2370

CTX_RH_FUSIFORM_SUVR numeric ctx-rh-fusiform Range value: 0.777, ..., 2.242

CTX_RH_FUSIFORM_VOLUME numeric ctx-rh-fusiform ROI size in mm^3 Range value: 4093, ..., 12881

CTX_RH_INFERIORPARIETAL_SUVR numeric ctx-rh-inferiorparietal Range value: 0.759, ..., 2.739

CTX_RH_INFERIORPARIETAL_VOLUME numeric ctx-rh-inferiorparietal ROI size in mm^3 Range value: 6177, ..., 21817

CTX_RH_INFERIORTEMPORAL_SUVR numeric ctx-rh-inferiortemporal Range value: 0.782, ..., 2.501

CTX_RH_INFERIORTEMPORAL_VOLUME numeric ctx-rh-inferiortemporal ROI size in mm^3 Range value: 3238, ..., 15738

CTX_RH_INSULA_SUVR numeric ctx-rh-insula Range value: 0.748, ..., 2.48

CTX_RH_INSULA_VOLUME numeric ctx-rh-insula ROI size in mm^3 Range value: 3554, ..., 9794

CTX_RH_ISTHMUSCINGULATE_SUVR numeric ctx-rh-isthmuscingulate Range value: 0.777, ..., 3.009

CTX_RH_ISTHMUSCINGULATE_VOLUME numeric ctx-rh-isthmuscingulate ROI size in mm^3 Range value: 703, ..., 3770

CTX_RH_LATERALOCCIPITAL_SUVR numeric ctx-rh-lateraloccipital Range value: 0.748, ..., 2.873

CTX_RH_LATERALOCCIPITAL_VOLUME numeric ctx-rh-lateraloccipital ROI size in mm^3 Range value: 4984, ..., 19271

CTX_RH_LATERALORBITOFRONTAL_SUVR numeric ctx-rh-lateralorbitofrontal Range value: 0.79, ..., 2.584

CTX_RH_LATERALORBITOFRONTAL_VOLUME numeric ctx-rh-lateralorbitofrontal ROI size in mm^3 Range value: 4176, ..., 9919

CTX_RH_LINGUAL_SUVR numeric ctx-rh-lingual Range value: 0.773, ..., 2.191

CTX_RH_LINGUAL_VOLUME numeric ctx-rh-lingual ROI size in mm^3 Range value: 3263, ..., 10478

CTX_RH_MEDIALORBITOFRONTAL_SUVR numeric ctx-rh-medialorbitofrontal Range value: 0.75, ..., 2.57

CTX_RH_MEDIALORBITOFRONTAL_VOLUME numeric ctx-rh-medialorbitofrontal ROI size in mm^3 Range value: 3216, ..., 9115

CTX_RH_MIDDLETEMPORAL_SUVR numeric ctx-rh-middletemporal Range value: 0.781, ..., 2.537

CTX_RH_MIDDLETEMPORAL_VOLUME numeric ctx-rh-middletemporal ROI size in mm^3 Range value: 4756, ..., 17219

CTX_RH_PARACENTRAL_SUVR numeric ctx-rh-paracentral Range value: 0.786, ..., 3.046

CTX_RH_PARACENTRAL_VOLUME numeric ctx-rh-paracentral ROI size in mm^3 Range value: 1981, ..., 5470

CTX_RH_PARAHIPPOCAMPAL_SUVR numeric ctx-rh-parahippocampal Range value: 0.506, ..., 1.718

CTX_RH_PARAHIPPOCAMPAL_VOLUME numeric ctx-rh-parahippocampal ROI size in mm^3 Range value: 840, ..., 3420

CTX_RH_PARSOPERCULARIS_SUVR numeric ctx-rh-parsopercularis Range value: 0.769, ..., 2.801

CTX_RH_PARSOPERCULARIS_VOLUME numeric ctx-rh-parsopercularis ROI size in mm^3 Range value: 1716, ..., 7542

CTX_RH_PARSORBITALIS_SUVR numeric ctx-rh-parsorbitalis Range value: 0.749, ..., 2.391

CTX_RH_PARSORBITALIS_VOLUME numeric ctx-rh-parsorbitalis ROI size in mm^3 Range value: 810, ..., 3933

CTX_RH_PARSTRIANGULARIS_SUVR numeric ctx-rh-parstriangularis Range value: 0.774, ..., 2.905

CTX_RH_PARSTRIANGULARIS_VOLUME numeric ctx-rh-parstriangularis ROI size in mm^3 Range value: 2134, ..., 7512

CTX_RH_PERICALCARINE_SUVR numeric ctx-rh-pericalcarine Range value: 0.87, ..., 2.967

CTX_RH_PERICALCARINE_VOLUME numeric ctx-rh-pericalcarine ROI size in mm^3 Range value: 831, ..., 4363

CTX_RH_POSTCENTRAL_SUVR numeric ctx-rh-postcentral Range value: 0.73, ..., 2.52

CTX_RH_POSTCENTRAL_VOLUME numeric ctx-rh-postcentral ROI size in mm^3 Range value: 2926, ..., 13033

CTX_RH_POSTERIORCINGULATE_SUVR numeric ctx-rh-posteriorcingulate Range value: 0.809, ..., 3.228

CTX_RH_POSTERIORCINGULATE_VOLUME numeric ctx-rh-posteriorcingulate ROI size in mm^3 Range value: 552, ..., 4499

CTX_RH_PRECENTRAL_SUVR numeric ctx-rh-precentral Range value: 0.834, ..., 2.647

CTX_RH_PRECENTRAL_VOLUME numeric ctx-rh-precentral ROI size in mm^3 Range value: 6158, ..., 18058

CTX_RH_PRECUNEUS_SUVR numeric ctx-rh-precuneus Range value: 0.801, ..., 3.171

CTX_RH_PRECUNEUS_VOLUME numeric ctx-rh-precuneus ROI size in mm^3 Range value: 4969, ..., 14854

CTX_RH_ROSTRALANTERIORCINGULATE_SUVR numeric ctx-rh-rostralanteriorcingulate Range value: 0.752, ..., 2.768

CTX_RH_ROSTRALANTERIORCINGULATE_VOLUME numeric ctx-rh-rostralanteriorcingulate ROI size in mm^3 Range value: 559, ..., 3180

CTX_RH_ROSTRALMIDDLEFRONTAL_SUVR numeric ctx-rh-rostralmiddlefrontal Range value: 0.77, ..., 3.011

CTX_RH_ROSTRALMIDDLEFRONTAL_VOLUME numeric ctx-rh-rostralmiddlefrontal ROI size in mm^3 Range value: 6285, ..., 26835

CTX_RH_SUPERIORFRONTAL_SUVR numeric ctx-rh-superiorfrontal Range value: 0.746, ..., 2.677

CTX_RH_SUPERIORFRONTAL_VOLUME numeric ctx-rh-superiorfrontal ROI size in mm^3 Range value: 10909, ..., 32025

CTX_RH_SUPERIORPARIETAL_SUVR numeric ctx-rh-superiorparietal Range value: 0.714, ..., 2.574

CTX_RH_SUPERIORPARIETAL_VOLUME numeric ctx-rh-superiorparietal ROI size in mm^3 Range value: 6110, ..., 18993

CTX_RH_SUPERIORTEMPORAL_SUVR numeric ctx-rh-superiortemporal Range value: 0.748, ..., 2.447

CTX_RH_SUPERIORTEMPORAL_VOLUME numeric ctx-rh-superiortemporal ROI size in mm^3 Range value: 4975, ..., 15383

CTX_RH_SUPRAMARGINAL_SUVR numeric ctx-rh-supramarginal Range value: 0.751, ..., 2.787

CTX_RH_SUPRAMARGINAL_VOLUME numeric ctx-rh-supramarginal ROI size in mm^3 Range value: 3693, ..., 14291

CTX_RH_TEMPORALPOLE_SUVR numeric ctx-rh-temporalpole Range value: 0.503, ..., 1.79

CTX_RH_TEMPORALPOLE_VOLUME numeric ctx-rh-temporalpole ROI size in mm^3 Range value: 287, ..., 4200

CTX_RH_TRANSVERSETEMPORAL_SUVR numeric ctx-rh-transversetemporal Range value: 0.657, ..., 2.634

CTX_RH_TRANSVERSETEMPORAL_VOLUME numeric ctx-rh-transversetemporal ROI size in mm^3 Range value: 269, ..., 1566

LEFT_ACCUMBENS_AREA_SUVR numeric left-accumbens-area Range value: 0.661, ..., 2.617

LEFT_ACCUMBENS_AREA_VOLUME numeric left-accumbens-area ROI size in mm^3 Range value: 92, ..., 821

LEFT_AMYGDALA_SUVR numeric left-amygdala Range value: 0.465, ..., 1.619

LEFT_AMYGDALA_VOLUME numeric left-amygdala ROI size in mm^3 Range value: 339, ..., 2496

LEFT_CAUDATE_SUVR numeric left-caudate Range value: 0.615, ..., 2.23

LEFT_CAUDATE_VOLUME numeric left-caudate ROI size in mm^3 Range value: 1660, ..., 6903

LEFT_CEREBELLUM_CORTEX_SUVR numeric left-cerebellum-cortex Range value: 0.751, ..., 0.955

LEFT_CEREBELLUM_CORTEX_VOLUME numeric left-cerebellum-cortex ROI size in mm^3 Range value: 34655, ..., 69630

LEFT_CEREBELLUM_WHITE_MATTER_SUVR numeric left-cerebellum-white-matter Range value: 1.179, ..., 2.055

LEFT_CEREBELLUM_WHITE_MATTER_VOLUME numeric left-cerebellum-white-matter ROI size in mm^3 Range value: 6892, ..., 26253

LEFT_CEREBRAL_WHITE_MATTER_SUVR numeric left-cerebral-white-matter Range value: 1.147, ..., 3.634

LEFT_CEREBRAL_WHITE_MATTER_VOLUME numeric left-cerebral-white-matter ROI size in mm^3 Range value: 20485, ..., 332312

LEFT_CHOROID_PLEXUS_SUVR numeric left-choroid-plexus Range value: 0.332, ..., 1.465

LEFT_CHOROID_PLEXUS_VOLUME numeric left-choroid-plexus ROI size in mm^3 Range value: 130, ..., 2141

LEFT_HIPPOCAMPUS_SUVR numeric left-hippocampus Range value: 0.601, ..., 1.545

LEFT_HIPPOCAMPUS_VOLUME numeric left-hippocampus ROI size in mm^3 Range value: 1314, ..., 6080

LEFT_INF_LAT_VENT_SUVR numeric left-inf-lat-vent Range value: 0.525, ..., 1.871

LEFT_INF_LAT_VENT_VOLUME numeric left-inf-lat-vent ROI size in mm^3 Range value: 53, ..., 11328

LEFT_LATERAL_VENTRICLE_SUVR numeric left-lateral-ventricle Range value: 0.292, ..., 1.457

LEFT_LATERAL_VENTRICLE_VOLUME numeric left-lateral-ventricle ROI size in mm^3 Range value: 3358, ..., 82443

LEFT_PALLIDUM_SUVR numeric left-pallidum Range value: 0.99, ..., 2.626

LEFT_PALLIDUM_VOLUME numeric left-pallidum ROI size in mm^3 Range value: 1066, ..., 3513

LEFT_PUTAMEN_SUVR numeric left-putamen Range value: 0.941, ..., 2.533

LEFT_PUTAMEN_VOLUME numeric left-putamen ROI size in mm^3 Range value: 1788, ..., 9271

LEFT_THALAMUS_PROPER_SUVR numeric left-thalamus-proper Range value: 0.686, ..., 1.822

LEFT_THALAMUS_PROPER_VOLUME numeric left-thalamus-proper ROI size in mm^3 Range value: 4217, ..., 11791

LEFT_VENTRALDC_SUVR numeric left-ventraldc Range value: 1.004, ..., 2.346

LEFT_VENTRALDC_VOLUME numeric left-ventraldc ROI size in mm^3 Range value: 2259, ..., 5676

LEFT_VESSEL_SUVR numeric left-vessel Range value: 0.537, ..., 2.012

LEFT_VESSEL_VOLUME numeric left-vessel ROI size in mm^3 Range value: 1, ..., 537

RIGHT_ACCUMBENS_AREA_SUVR numeric right-accumbens-area Range value: 0.578, ..., 2.605

RIGHT_ACCUMBENS_AREA_VOLUME numeric right-accumbens-area ROI size in mm^3 Range value: 114, ..., 843

RIGHT_AMYGDALA_SUVR numeric right-amygdala Range value: 0.627, ..., 1.612

RIGHT_AMYGDALA_VOLUME numeric right-amygdala ROI size in mm^3 Range value: 483, ..., 2544

RIGHT_CAUDATE_SUVR numeric right-caudate Range value: 0.592, ..., 1.997

RIGHT_CAUDATE_VOLUME numeric right-caudate ROI size in mm^3 Range value: 2003, ..., 6880

RIGHT_CEREBELLUM_CORTEX_SUVR numeric right-cerebellum-cortex Range value: 0.756, ..., 0.977

RIGHT_CEREBELLUM_CORTEX_VOLUME numeric right-cerebellum-cortex ROI size in mm^3 Range value: 33950, ..., 73171

RIGHT_CEREBELLUM_WHITE_MATTER_SUVR numeric right-cerebellum-white-matter Range value: 1.203, ..., 2.105

RIGHT_CEREBELLUM_WHITE_MATTER_VOLUME numeric right-cerebellum-white-matter ROI size in mm^3 Range value: 6599, ..., 25269

RIGHT_CEREBRAL_WHITE_MATTER_SUVR numeric right-cerebral-white-matter Range value: 1.103, ..., 3.42

RIGHT_CEREBRAL_WHITE_MATTER_VOLUME numeric right-cerebral-white-matter ROI size in mm^3 Range value: 127204, ..., 331821

RIGHT_CHOROID_PLEXUS_SUVR numeric right-choroid-plexus Range value: 0.421, ..., 1.433

RIGHT_CHOROID_PLEXUS_VOLUME numeric right-choroid-plexus ROI size in mm^3 Range value: 316, ..., 2028

RIGHT_HIPPOCAMPUS_SUVR numeric right-hippocampus Range value: 0.614, ..., 1.532

RIGHT_HIPPOCAMPUS_VOLUME numeric right-hippocampus ROI size in mm^3 Range value: 1576, ..., 5928

RIGHT_INF_LAT_VENT_SUVR numeric right-inf-lat-vent Range value: 0.384, ..., 1.748

RIGHT_INF_LAT_VENT_VOLUME numeric right-inf-lat-vent ROI size in mm^3 Range value: 58, ..., 7510

RIGHT_LATERAL_VENTRICLE_SUVR numeric right-lateral-ventricle Range value: 0.263, ..., 1.456

RIGHT_LATERAL_VENTRICLE_VOLUME numeric right-lateral-ventricle ROI size in mm^3 Range value: 2442, ..., 71992

RIGHT_PALLIDUM_SUVR numeric right-pallidum Range value: 0.991, ..., 2.534

RIGHT_PALLIDUM_VOLUME numeric right-pallidum ROI size in mm^3 Range value: 999, ..., 3535

RIGHT_PUTAMEN_SUVR numeric right-putamen Range value: 0.894, ..., 2.518

RIGHT_PUTAMEN_VOLUME numeric right-putamen ROI size in mm^3 Range value: 2178, ..., 8672

RIGHT_THALAMUS_PROPER_SUVR numeric right-thalamus-proper Range value: 0.835, ..., 1.871

RIGHT_THALAMUS_PROPER_VOLUME numeric right-thalamus-proper ROI size in mm^3 Range value: 3580, ..., 12434

RIGHT_VENTRALDC_SUVR numeric right-ventraldc Range value: 0.959, ..., 2.142

RIGHT_VENTRALDC_VOLUME numeric right-ventraldc ROI size in mm^3 Range value: 2384, ..., 5630

RIGHT_VESSEL_SUVR numeric right-vessel Range value: 0.736, ..., 2.149

RIGHT_VESSEL_VOLUME numeric right-vessel ROI size in mm^3 Range value: 1, ..., 300

update_stamp POSIXct POSIXct: YYY-MM-DD

Author

adni-data@googlegroups.com