UC Berkeley - Tau PET PVC 6mm Res analysis

UC Berkeley - Tau PET PVC 6mm Res analysis data. More information is available at https://adni.loni.usc.edu/data-samples/adni-data/

Usage

data(UCBERKELEY_TAUPVC_6MM)

Format

A data frame with 1811 observations and 335 variables.

Source

https://adni.loni.usc.edu/data-samples/adni-data/

Details

ORIGPROT character Original study protocol , with options: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4

LONIUID character LONI Image ID

RID numeric Participant roster ID (unique participant identifier that should be used to merge data) Range value: 21, ..., 10340

VISCODE character Visit code

SITEID numeric Site ID Range value: 2, ..., 941

PTID character ADNI Participant ID consisting of 3-digit "SITEID" and 4-digit "RID"

SCANDATE Date SCAN acquisition date Date: YYY-MM-DD

PROCESSDATE Date Date quantification pipeline was run Date: YYY-MM-DD

TRACER character PET Radiotracer , with options: FTP, MK6240

TRACER_SUVR_WARNING character Reminder of the incomparability of SUVRs from different tracers. See UCBERKELEY Tau Processing Methods PDF. , with options: DO NOT COMPARE SUVRs ACROSS TRACERS

META_TEMPORAL_SUVR numeric Weighted mean SUVR of Jack et al temporal meta-ROI. All SUVRs in this file are normalized to the inferior cerebellar grey matter unless otherwise noted. See UCBERKELEY Tau PET Processing Methods PDF on LONI for details. Range value: 0.625, ..., 7.089

META_TEMPORAL_VOLUME numeric Jack et al temporal meta-ROI size in mm^3 Range value: 36639, ..., 97274

CTX_ENTORHINAL_SUVR numeric ctx-bilateral-entorhinal Range value: 0.729, ..., 5.592

CTX_ENTORHINAL_VOLUME numeric ctx-bilateral-entorhinal ROI size in mm^3 Range value: 746, ..., 5946

INFERIORCEREBELLUM_SUVR numeric Reference region - mean SUVR of inferior cerebellar grey matter Range value: 1, ..., 1

INFERIORCEREBELLUM_VOLUME numeric Inferior Cerebellar Grey Matter ROI size in mm^3 Range value: 40753, ..., 87660

CEREBRAL_WHITE_MATTER_SUVR numeric bilateral-cerebral-white-matter Range value: 0.631, ..., 1.859

CEREBRAL_WHITE_MATTER_VOLUME numeric bilateral-cerebral-white-matter ROI size in mm^3 Range value: 224785, ..., 578367

BRAINSTEM_SUVR numeric brain-stem Range value: 0.535, ..., 1.389

BRAINSTEM_VOLUME numeric brain-stem ROI size in mm^3 Range value: 13764, ..., 31388

CC_ANTERIOR_SUVR numeric cc-anterior Range value: 0.631, ..., 1.859

CC_ANTERIOR_VOLUME numeric cc-anterior ROI size in mm^3 Range value: 171, ..., 1763

CC_CENTRAL_SUVR numeric cc-central Range value: 0.631, ..., 1.859

CC_CENTRAL_VOLUME numeric cc-central ROI size in mm^3 Range value: 233, ..., 1380

CC_MID_ANTERIOR_SUVR numeric cc-mid-anterior Range value: 0.631, ..., 1.859

CC_MID_ANTERIOR_VOLUME numeric cc-mid-anterior ROI size in mm^3 Range value: 225, ..., 1241

CC_MID_POSTERIOR_SUVR numeric cc-mid-posterior Range value: 0.631, ..., 1.859

CC_MID_POSTERIOR_VOLUME numeric cc-mid-posterior ROI size in mm^3 Range value: 189, ..., 1796

CC_POSTERIOR_SUVR numeric cc-posterior Range value: 0.631, ..., 1.859

CC_POSTERIOR_VOLUME numeric cc-posterior ROI size in mm^3 Range value: 259, ..., 2176

CSF_SUVR numeric csf Range value: 0.017, ..., 0.731

CSF_VOLUME numeric csf ROI size in mm^3 Range value: 628, ..., 3561

VENTRICLE_3RD_SUVR numeric 3rd-ventricle Range value: 0.017, ..., 0.73

VENTRICLE_3RD_VOLUME numeric 3rd-ventricle ROI size in mm^3 Range value: 484, ..., 5244

VENTRICLE_4TH_SUVR numeric 4th-ventricle Range value: 0.004, ..., 0.727

VENTRICLE_4TH_VOLUME numeric 4th-ventricle ROI size in mm^3 Range value: 739, ..., 5081

VENTRICLE_5TH_SUVR numeric 5th-ventricle Range value: 0.185, ..., 0.655

VENTRICLE_5TH_VOLUME numeric 5th-ventricle ROI size in mm^3 Range value: 1, ..., 11

WM_HYPOINTENSITIES_SUVR numeric wm-hypointensities Range value: 0.001, ..., 2.379

WM_HYPOINTENSITIES_VOLUME numeric wm-hypointensities ROI size in mm^3 Range value: 268, ..., 72336

NON_WM_HYPOINTENSITIES_SUVR numeric non-wm-hypointensities Range value: 0.001, ..., 0.887

NON_WM_HYPOINTENSITIES_VOLUME numeric non-wm-hypointensities ROI size in mm^3 Range value: 1, ..., 233

CTX_BANKSSTS_SUVR numeric ctx-bilateral-bankssts Range value: 0.692, ..., 8.47

CTX_BANKSSTS_VOLUME numeric ctx-bilateral-bankssts ROI size in mm^3 Range value: 2121, ..., 8963

CTX_CAUDALANTERIORCINGULATE_SUVR numeric ctx-bilateral-caudalanteriorcingulate Range value: 0.036, ..., 3.894

CTX_CAUDALANTERIORCINGULATE_VOLUME numeric ctx-bilateral-caudalanteriorcingulate ROI size in mm^3 Range value: 1309, ..., 6906

CTX_CAUDALMIDDLEFRONTAL_SUVR numeric ctx-bilateral-caudalmiddlefrontal Range value: 0.724, ..., 7.984

CTX_CAUDALMIDDLEFRONTAL_VOLUME numeric ctx-bilateral-caudalmiddlefrontal ROI size in mm^3 Range value: 5775, ..., 16670

CTX_CUNEUS_SUVR numeric ctx-bilateral-cuneus Range value: 0.538, ..., 9.75

CTX_CUNEUS_VOLUME numeric ctx-bilateral-cuneus ROI size in mm^3 Range value: 3675, ..., 10467

CTX_FRONTALPOLE_SUVR numeric ctx-bilateral-frontalpole Range value: 0.098, ..., 6.29

CTX_FRONTALPOLE_VOLUME numeric ctx-bilateral-frontalpole ROI size in mm^3 Range value: 982, ..., 3825

CTX_FUSIFORM_SUVR numeric ctx-bilateral-fusiform Range value: 0.459, ..., 8.054

CTX_FUSIFORM_VOLUME numeric ctx-bilateral-fusiform ROI size in mm^3 Range value: 10371, ..., 26241

CTX_INFERIORPARIETAL_SUVR numeric ctx-bilateral-inferiorparietal Range value: 0.831, ..., 8.66

CTX_INFERIORPARIETAL_VOLUME numeric ctx-bilateral-inferiorparietal ROI size in mm^3 Range value: 12378, ..., 37678

CTX_INFERIORTEMPORAL_SUVR numeric ctx-bilateral-inferiortemporal Range value: 0.687, ..., 8.061

CTX_INFERIORTEMPORAL_VOLUME numeric ctx-bilateral-inferiortemporal ROI size in mm^3 Range value: 9660, ..., 31454

CTX_INSULA_SUVR numeric ctx-bilateral-insula Range value: 0.7, ..., 4.639

CTX_INSULA_VOLUME numeric ctx-bilateral-insula ROI size in mm^3 Range value: 7460, ..., 27196

CTX_ISTHMUSCINGULATE_SUVR numeric ctx-bilateral-isthmuscingulate Range value: 0.564, ..., 7.722

CTX_ISTHMUSCINGULATE_VOLUME numeric ctx-bilateral-isthmuscingulate ROI size in mm^3 Range value: 1985, ..., 8883

CTX_LATERALOCCIPITAL_SUVR numeric ctx-bilateral-lateraloccipital Range value: 0.454, ..., 11.414

CTX_LATERALOCCIPITAL_VOLUME numeric ctx-bilateral-lateraloccipital ROI size in mm^3 Range value: 9370, ..., 35627

CTX_LATERALORBITOFRONTAL_SUVR numeric ctx-bilateral-lateralorbitofrontal Range value: 0.752, ..., 5.529

CTX_LATERALORBITOFRONTAL_VOLUME numeric ctx-bilateral-lateralorbitofrontal ROI size in mm^3 Range value: 8842, ..., 18822

CTX_LINGUAL_SUVR numeric ctx-bilateral-lingual Range value: 0.491, ..., 8.206

CTX_LINGUAL_VOLUME numeric ctx-bilateral-lingual ROI size in mm^3 Range value: 8165, ..., 20629

CTX_MEDIALORBITOFRONTAL_SUVR numeric ctx-bilateral-medialorbitofrontal Range value: 0.695, ..., 4.693

CTX_MEDIALORBITOFRONTAL_VOLUME numeric ctx-bilateral-medialorbitofrontal ROI size in mm^3 Range value: 6803, ..., 13876

CTX_MIDDLETEMPORAL_SUVR numeric ctx-bilateral-middletemporal Range value: 0.508, ..., 6.668

CTX_MIDDLETEMPORAL_VOLUME numeric ctx-bilateral-middletemporal ROI size in mm^3 Range value: 11213, ..., 31685

CTX_PARACENTRAL_SUVR numeric ctx-bilateral-paracentral Range value: 0.722, ..., 3.298

CTX_PARACENTRAL_VOLUME numeric ctx-bilateral-paracentral ROI size in mm^3 Range value: 4734, ..., 10986

CTX_PARAHIPPOCAMPAL_SUVR numeric ctx-bilateral-parahippocampal Range value: 0.279, ..., 5.406

CTX_PARAHIPPOCAMPAL_VOLUME numeric ctx-bilateral-parahippocampal ROI size in mm^3 Range value: 1751, ..., 5501

CTX_PARSOPERCULARIS_SUVR numeric ctx-bilateral-parsopercularis Range value: 0.697, ..., 6.747

CTX_PARSOPERCULARIS_VOLUME numeric ctx-bilateral-parsopercularis ROI size in mm^3 Range value: 4928, ..., 12836

CTX_PARSORBITALIS_SUVR numeric ctx-bilateral-parsorbitalis Range value: 0.737, ..., 6.378

CTX_PARSORBITALIS_VOLUME numeric ctx-bilateral-parsorbitalis ROI size in mm^3 Range value: 2452, ..., 6892

CTX_PARSTRIANGULARIS_SUVR numeric ctx-bilateral-parstriangularis Range value: 0.654, ..., 5.871

CTX_PARSTRIANGULARIS_VOLUME numeric ctx-bilateral-parstriangularis ROI size in mm^3 Range value: 4786, ..., 11719

CTX_PERICALCARINE_SUVR numeric ctx-bilateral-pericalcarine Range value: 0.313, ..., 5.844

CTX_PERICALCARINE_VOLUME numeric ctx-bilateral-pericalcarine ROI size in mm^3 Range value: 2165, ..., 7978

CTX_POSTCENTRAL_SUVR numeric ctx-bilateral-postcentral Range value: 0.78, ..., 3.712

CTX_POSTCENTRAL_VOLUME numeric ctx-bilateral-postcentral ROI size in mm^3 Range value: 11509, ..., 26525

CTX_POSTERIORCINGULATE_SUVR numeric ctx-bilateral-posteriorcingulate Range value: 0.624, ..., 6.589

CTX_POSTERIORCINGULATE_VOLUME numeric ctx-bilateral-posteriorcingulate ROI size in mm^3 Range value: 2565, ..., 11535

CTX_PRECENTRAL_SUVR numeric ctx-bilateral-precentral Range value: 0.731, ..., 3.127

CTX_PRECENTRAL_VOLUME numeric ctx-bilateral-precentral ROI size in mm^3 Range value: 13766, ..., 37386

CTX_PRECUNEUS_SUVR numeric ctx-bilateral-precuneus Range value: 0.754, ..., 8.131

CTX_PRECUNEUS_VOLUME numeric ctx-bilateral-precuneus ROI size in mm^3 Range value: 9272, ..., 27581

CTX_ROSTRALANTERIORCINGULATE_SUVR numeric ctx-bilateral-rostralanteriorcingulate Range value: 0.497, ..., 4.24

CTX_ROSTRALANTERIORCINGULATE_VOLUME numeric ctx-bilateral-rostralanteriorcingulate ROI size in mm^3 Range value: 1558, ..., 7474

CTX_ROSTRALMIDDLEFRONTAL_SUVR numeric ctx-bilateral-rostralmiddlefrontal Range value: 0.782, ..., 9.23

CTX_ROSTRALMIDDLEFRONTAL_VOLUME numeric ctx-bilateral-rostralmiddlefrontal ROI size in mm^3 Range value: 16054, ..., 50046

CTX_SUPERIORFRONTAL_SUVR numeric ctx-bilateral-superiorfrontal Range value: 0.727, ..., 5.643

CTX_SUPERIORFRONTAL_VOLUME numeric ctx-bilateral-superiorfrontal ROI size in mm^3 Range value: 25550, ..., 57959

CTX_SUPERIORPARIETAL_SUVR numeric ctx-bilateral-superiorparietal Range value: 0.649, ..., 9.384

CTX_SUPERIORPARIETAL_VOLUME numeric ctx-bilateral-superiorparietal ROI size in mm^3 Range value: 13348, ..., 37095

CTX_SUPERIORTEMPORAL_SUVR numeric ctx-bilateral-superiortemporal Range value: 0.795, ..., 4.761

CTX_SUPERIORTEMPORAL_VOLUME numeric ctx-bilateral-superiortemporal ROI size in mm^3 Range value: 12461, ..., 32667

CTX_SUPRAMARGINAL_SUVR numeric ctx-bilateral-supramarginal Range value: 0.805, ..., 6.568

CTX_SUPRAMARGINAL_VOLUME numeric ctx-bilateral-supramarginal ROI size in mm^3 Range value: 10961, ..., 29495

CTX_TEMPORALPOLE_SUVR numeric ctx-bilateral-temporalpole Range value: 0.666, ..., 6.179

CTX_TEMPORALPOLE_VOLUME numeric ctx-bilateral-temporalpole ROI size in mm^3 Range value: 2178, ..., 9644

CTX_TRANSVERSETEMPORAL_SUVR numeric ctx-bilateral-transversetemporal Range value: 0.247, ..., 5.094

CTX_TRANSVERSETEMPORAL_VOLUME numeric ctx-bilateral-transversetemporal ROI size in mm^3 Range value: 1100, ..., 3541

ACCUMBENS_AREA_SUVR numeric bilateral-accumbens-area Range value: 0.309, ..., 4.112

ACCUMBENS_AREA_VOLUME numeric bilateral-accumbens-area ROI size in mm^3 Range value: 300, ..., 1636

AMYGDALA_SUVR numeric bilateral-amygdala Range value: 0.619, ..., 4.068

AMYGDALA_VOLUME numeric bilateral-amygdala ROI size in mm^3 Range value: 980, ..., 4740

CAUDATE_SUVR numeric bilateral-caudate Range value: 0.561, ..., 2.545

CAUDATE_VOLUME numeric bilateral-caudate ROI size in mm^3 Range value: 4068, ..., 12321

CEREBELLUM_CORTEX_SUVR numeric bilateral-cerebellum-cortex Range value: 0.489, ..., 1.357

CEREBELLUM_CORTEX_VOLUME numeric bilateral-cerebellum-cortex ROI size in mm^3 Range value: 72344, ..., 142592

CEREBELLUM_WHITE_MATTER_SUVR numeric bilateral-cerebellum-white-matter Range value: 0.492, ..., 1.685

CEREBELLUM_WHITE_MATTER_VOLUME numeric bilateral-cerebellum-white-matter ROI size in mm^3 Range value: 13671, ..., 48785

CHOROID_PLEXUS_SUVR numeric bilateral-choroid-plexus Range value: 0.605, ..., 16.369

CHOROID_PLEXUS_VOLUME numeric bilateral-choroid-plexus ROI size in mm^3 Range value: 459, ..., 3434

HIPPOCAMPUS_SUVR numeric bilateral-hippocampus Range value: 0.622, ..., 3.284

HIPPOCAMPUS_VOLUME numeric bilateral-hippocampus ROI size in mm^3 Range value: 3255, ..., 11461

OPTIC_CHIASM_SUVR numeric optic-chiasm Range value: 0.001, ..., 3.02

OPTIC_CHIASM_VOLUME numeric optic-chiasm ROI size in mm^3 Range value: 45, ..., 333

INF_LAT_VENT_SUVR numeric bilateral-inf-lat-vent Range value: 0.004, ..., 6.095

INF_LAT_VENT_VOLUME numeric bilateral-inf-lat-vent ROI size in mm^3 Range value: 164, ..., 16102

LATERAL_VENTRICLE_SUVR numeric bilateral-lateral-ventricle Range value: 0.005, ..., 1.244

LATERAL_VENTRICLE_VOLUME numeric bilateral-lateral-ventricle ROI size in mm^3 Range value: 5800, ..., 138561

PALLIDUM_SUVR numeric bilateral-pallidum Range value: 0.633, ..., 3.077

PALLIDUM_VOLUME numeric bilateral-pallidum ROI size in mm^3 Range value: 2454, ..., 6324

PUTAMEN_SUVR numeric bilateral-putamen Range value: 0.651, ..., 2.777

PUTAMEN_VOLUME numeric bilateral-putamen ROI size in mm^3 Range value: 4438, ..., 15492

THALAMUS_PROPER_SUVR numeric bilateral-thalamus-proper Range value: 0.655, ..., 2.052

THALAMUS_PROPER_VOLUME numeric bilateral-thalamus-proper ROI size in mm^3 Range value: 8557, ..., 19663

VENTRALDC_SUVR numeric bilateral-ventraldc Range value: 0.663, ..., 2.208

VENTRALDC_VOLUME numeric bilateral-ventraldc ROI size in mm^3 Range value: 4643, ..., 11189

VESSEL_SUVR logical bilateral-vessel Boolean value: TRUE/ FALSE

VESSEL_VOLUME numeric bilateral-vessel ROI size in mm^3 Range value: 1, ..., 604

CTX_LH_BANKSSTS_SUVR numeric ctx-lh-bankssts Range value: 0.638, ..., 8.571

CTX_LH_BANKSSTS_VOLUME numeric ctx-lh-bankssts ROI size in mm^3 Range value: 779, ..., 4723

CTX_LH_CAUDALANTERIORCINGULATE_SUVR numeric ctx-lh-caudalanteriorcingulate Range value: 0.083, ..., 3.875

CTX_LH_CAUDALANTERIORCINGULATE_VOLUME numeric ctx-lh-caudalanteriorcingulate ROI size in mm^3 Range value: 461, ..., 4960

CTX_LH_CAUDALMIDDLEFRONTAL_SUVR numeric ctx-lh-caudalmiddlefrontal Range value: 0.778, ..., 8.657

CTX_LH_CAUDALMIDDLEFRONTAL_VOLUME numeric ctx-lh-caudalmiddlefrontal ROI size in mm^3 Range value: 2609, ..., 12065

CTX_LH_CUNEUS_SUVR numeric ctx-lh-cuneus Range value: 0.461, ..., 7.967

CTX_LH_CUNEUS_VOLUME numeric ctx-lh-cuneus ROI size in mm^3 Range value: 1627, ..., 5342

CTX_LH_ENTORHINAL_SUVR numeric ctx-lh-entorhinal Range value: 0.71, ..., 5.736

CTX_LH_ENTORHINAL_VOLUME numeric ctx-lh-entorhinal ROI size in mm^3 Range value: 291, ..., 3307

CTX_LH_FRONTALPOLE_SUVR numeric ctx-lh-frontalpole Range value: 0.156, ..., 6.617

CTX_LH_FRONTALPOLE_VOLUME numeric ctx-lh-frontalpole ROI size in mm^3 Range value: 486, ..., 1794

CTX_LH_FUSIFORM_SUVR numeric ctx-lh-fusiform Range value: 0.506, ..., 8.781

CTX_LH_FUSIFORM_VOLUME numeric ctx-lh-fusiform ROI size in mm^3 Range value: 3944, ..., 13678

CTX_LH_INFERIORPARIETAL_SUVR numeric ctx-lh-inferiorparietal Range value: 0.763, ..., 9.371

CTX_LH_INFERIORPARIETAL_VOLUME numeric ctx-lh-inferiorparietal ROI size in mm^3 Range value: 5861, ..., 17756

CTX_LH_INFERIORTEMPORAL_SUVR numeric ctx-lh-inferiortemporal Range value: 0.692, ..., 7.794

CTX_LH_INFERIORTEMPORAL_VOLUME numeric ctx-lh-inferiortemporal ROI size in mm^3 Range value: 4547, ..., 16512

CTX_LH_INSULA_SUVR numeric ctx-lh-insula Range value: 0.687, ..., 4.563

CTX_LH_INSULA_VOLUME numeric ctx-lh-insula ROI size in mm^3 Range value: 3825, ..., 21138

CTX_LH_ISTHMUSCINGULATE_SUVR numeric ctx-lh-isthmuscingulate Range value: 0.518, ..., 7.809

CTX_LH_ISTHMUSCINGULATE_VOLUME numeric ctx-lh-isthmuscingulate ROI size in mm^3 Range value: 615, ..., 6967

CTX_LH_LATERALOCCIPITAL_SUVR numeric ctx-lh-lateraloccipital Range value: 0.616, ..., 10.54

CTX_LH_LATERALOCCIPITAL_VOLUME numeric ctx-lh-lateraloccipital ROI size in mm^3 Range value: 4386, ..., 17628

CTX_LH_LATERALORBITOFRONTAL_SUVR numeric ctx-lh-lateralorbitofrontal Range value: 0.717, ..., 6.507

CTX_LH_LATERALORBITOFRONTAL_VOLUME numeric ctx-lh-lateralorbitofrontal ROI size in mm^3 Range value: 3412, ..., 11834

CTX_LH_LINGUAL_SUVR numeric ctx-lh-lingual Range value: 0.404, ..., 7.338

CTX_LH_LINGUAL_VOLUME numeric ctx-lh-lingual ROI size in mm^3 Range value: 3248, ..., 11629

CTX_LH_MEDIALORBITOFRONTAL_SUVR numeric ctx-lh-medialorbitofrontal Range value: 0.715, ..., 4.624

CTX_LH_MEDIALORBITOFRONTAL_VOLUME numeric ctx-lh-medialorbitofrontal ROI size in mm^3 Range value: 3241, ..., 7019

CTX_LH_MIDDLETEMPORAL_SUVR numeric ctx-lh-middletemporal Range value: 0.467, ..., 7.928

CTX_LH_MIDDLETEMPORAL_VOLUME numeric ctx-lh-middletemporal ROI size in mm^3 Range value: 4242, ..., 14767

CTX_LH_PARACENTRAL_SUVR numeric ctx-lh-paracentral Range value: 0.648, ..., 3.251

CTX_LH_PARACENTRAL_VOLUME numeric ctx-lh-paracentral ROI size in mm^3 Range value: 2046, ..., 7180

CTX_LH_PARAHIPPOCAMPAL_SUVR numeric ctx-lh-parahippocampal Range value: 0.527, ..., 5.488

CTX_LH_PARAHIPPOCAMPAL_VOLUME numeric ctx-lh-parahippocampal ROI size in mm^3 Range value: 938, ..., 3513

CTX_LH_PARSOPERCULARIS_SUVR numeric ctx-lh-parsopercularis Range value: 0.703, ..., 6.923

CTX_LH_PARSOPERCULARIS_VOLUME numeric ctx-lh-parsopercularis ROI size in mm^3 Range value: 2451, ..., 7630

CTX_LH_PARSORBITALIS_SUVR numeric ctx-lh-parsorbitalis Range value: 0.596, ..., 7.216

CTX_LH_PARSORBITALIS_VOLUME numeric ctx-lh-parsorbitalis ROI size in mm^3 Range value: 438, ..., 3294

CTX_LH_PARSTRIANGULARIS_SUVR numeric ctx-lh-parstriangularis Range value: 0.616, ..., 6.51

CTX_LH_PARSTRIANGULARIS_VOLUME numeric ctx-lh-parstriangularis ROI size in mm^3 Range value: 1834, ..., 5341

CTX_LH_PERICALCARINE_SUVR numeric ctx-lh-pericalcarine Range value: 0.292, ..., 4.6

CTX_LH_PERICALCARINE_VOLUME numeric ctx-lh-pericalcarine ROI size in mm^3 Range value: 984, ..., 4678

CTX_LH_POSTCENTRAL_SUVR numeric ctx-lh-postcentral Range value: 0.682, ..., 3.436

CTX_LH_POSTCENTRAL_VOLUME numeric ctx-lh-postcentral ROI size in mm^3 Range value: 5982, ..., 14589

CTX_LH_POSTERIORCINGULATE_SUVR numeric ctx-lh-posteriorcingulate Range value: 0.452, ..., 6.441

CTX_LH_POSTERIORCINGULATE_VOLUME numeric ctx-lh-posteriorcingulate ROI size in mm^3 Range value: 1083, ..., 8715

CTX_LH_PRECENTRAL_SUVR numeric ctx-lh-precentral Range value: 0.711, ..., 3.627

CTX_LH_PRECENTRAL_VOLUME numeric ctx-lh-precentral ROI size in mm^3 Range value: 6606, ..., 25709

CTX_LH_PRECUNEUS_SUVR numeric ctx-lh-precuneus Range value: 0.772, ..., 8.728

CTX_LH_PRECUNEUS_VOLUME numeric ctx-lh-precuneus ROI size in mm^3 Range value: 4263, ..., 17038

CTX_LH_ROSTRALANTERIORCINGULATE_SUVR numeric ctx-lh-rostralanteriorcingulate Range value: 0.64, ..., 4.188

CTX_LH_ROSTRALANTERIORCINGULATE_VOLUME numeric ctx-lh-rostralanteriorcingulate ROI size in mm^3 Range value: 571, ..., 5073

CTX_LH_ROSTRALMIDDLEFRONTAL_SUVR numeric ctx-lh-rostralmiddlefrontal Range value: 0.79, ..., 9.649

CTX_LH_ROSTRALMIDDLEFRONTAL_VOLUME numeric ctx-lh-rostralmiddlefrontal ROI size in mm^3 Range value: 6861, ..., 23211

CTX_LH_SUPERIORFRONTAL_SUVR numeric ctx-lh-superiorfrontal Range value: 0.625, ..., 6.289

CTX_LH_SUPERIORFRONTAL_VOLUME numeric ctx-lh-superiorfrontal ROI size in mm^3 Range value: 13132, ..., 39158

CTX_LH_SUPERIORPARIETAL_SUVR numeric ctx-lh-superiorparietal Range value: 0.708, ..., 8.873

CTX_LH_SUPERIORPARIETAL_VOLUME numeric ctx-lh-superiorparietal ROI size in mm^3 Range value: 7205, ..., 23420

CTX_LH_SUPERIORTEMPORAL_SUVR numeric ctx-lh-superiortemporal Range value: 0.769, ..., 5.366

CTX_LH_SUPERIORTEMPORAL_VOLUME numeric ctx-lh-superiortemporal ROI size in mm^3 Range value: 5891, ..., 17284

CTX_LH_SUPRAMARGINAL_SUVR numeric ctx-lh-supramarginal Range value: 0.801, ..., 6.942

CTX_LH_SUPRAMARGINAL_VOLUME numeric ctx-lh-supramarginal ROI size in mm^3 Range value: 5543, ..., 17431

CTX_LH_TEMPORALPOLE_SUVR numeric ctx-lh-temporalpole Range value: 0.673, ..., 5.922

CTX_LH_TEMPORALPOLE_VOLUME numeric ctx-lh-temporalpole ROI size in mm^3 Range value: 48, ..., 5851

CTX_LH_TRANSVERSETEMPORAL_SUVR numeric ctx-lh-transversetemporal Range value: 0.228, ..., 4.831

CTX_LH_TRANSVERSETEMPORAL_VOLUME numeric ctx-lh-transversetemporal ROI size in mm^3 Range value: 601, ..., 2249

CTX_RH_BANKSSTS_SUVR numeric ctx-rh-bankssts Range value: 0.694, ..., 10.774

CTX_RH_BANKSSTS_VOLUME numeric ctx-rh-bankssts ROI size in mm^3 Range value: 952, ..., 4240

CTX_RH_CAUDALANTERIORCINGULATE_SUVR numeric ctx-rh-caudalanteriorcingulate Range value: 0.093, ..., 3.907

CTX_RH_CAUDALANTERIORCINGULATE_VOLUME numeric ctx-rh-caudalanteriorcingulate ROI size in mm^3 Range value: 368, ..., 3780

CTX_RH_CAUDALMIDDLEFRONTAL_SUVR numeric ctx-rh-caudalmiddlefrontal Range value: 0.599, ..., 7.329

CTX_RH_CAUDALMIDDLEFRONTAL_VOLUME numeric ctx-rh-caudalmiddlefrontal ROI size in mm^3 Range value: 2379, ..., 8501

CTX_RH_CUNEUS_SUVR numeric ctx-rh-cuneus Range value: 0.573, ..., 11.746

CTX_RH_CUNEUS_VOLUME numeric ctx-rh-cuneus ROI size in mm^3 Range value: 982, ..., 6165

CTX_RH_ENTORHINAL_SUVR numeric ctx-rh-entorhinal Range value: 0.678, ..., 5.448

CTX_RH_ENTORHINAL_VOLUME numeric ctx-rh-entorhinal ROI size in mm^3 Range value: 411, ..., 3452

CTX_RH_FRONTALPOLE_SUVR numeric ctx-rh-frontalpole Range value: 0.05, ..., 6.011

CTX_RH_FRONTALPOLE_VOLUME numeric ctx-rh-frontalpole ROI size in mm^3 Range value: 381, ..., 2370

CTX_RH_FUSIFORM_SUVR numeric ctx-rh-fusiform Range value: 0.407, ..., 7.601

CTX_RH_FUSIFORM_VOLUME numeric ctx-rh-fusiform ROI size in mm^3 Range value: 5003, ..., 13810

CTX_RH_INFERIORPARIETAL_SUVR numeric ctx-rh-inferiorparietal Range value: 0.838, ..., 8.115

CTX_RH_INFERIORPARIETAL_VOLUME numeric ctx-rh-inferiorparietal ROI size in mm^3 Range value: 6272, ..., 21817

CTX_RH_INFERIORTEMPORAL_SUVR numeric ctx-rh-inferiortemporal Range value: 0.682, ..., 8.358

CTX_RH_INFERIORTEMPORAL_VOLUME numeric ctx-rh-inferiortemporal ROI size in mm^3 Range value: 4566, ..., 15738

CTX_RH_INSULA_SUVR numeric ctx-rh-insula Range value: 0.714, ..., 4.712

CTX_RH_INSULA_VOLUME numeric ctx-rh-insula ROI size in mm^3 Range value: 3635, ..., 9212

CTX_RH_ISTHMUSCINGULATE_SUVR numeric ctx-rh-isthmuscingulate Range value: 0.615, ..., 7.641

CTX_RH_ISTHMUSCINGULATE_VOLUME numeric ctx-rh-isthmuscingulate ROI size in mm^3 Range value: 783, ..., 3770

CTX_RH_LATERALOCCIPITAL_SUVR numeric ctx-rh-lateraloccipital Range value: 0.292, ..., 12.431

CTX_RH_LATERALOCCIPITAL_VOLUME numeric ctx-rh-lateraloccipital ROI size in mm^3 Range value: 4984, ..., 18366

CTX_RH_LATERALORBITOFRONTAL_SUVR numeric ctx-rh-lateralorbitofrontal Range value: 0.752, ..., 5.269

CTX_RH_LATERALORBITOFRONTAL_VOLUME numeric ctx-rh-lateralorbitofrontal ROI size in mm^3 Range value: 4176, ..., 9667

CTX_RH_LINGUAL_SUVR numeric ctx-rh-lingual Range value: 0.639, ..., 9.147

CTX_RH_LINGUAL_VOLUME numeric ctx-rh-lingual ROI size in mm^3 Range value: 3485, ..., 10478

CTX_RH_MEDIALORBITOFRONTAL_SUVR numeric ctx-rh-medialorbitofrontal Range value: 0.657, ..., 4.993

CTX_RH_MEDIALORBITOFRONTAL_VOLUME numeric ctx-rh-medialorbitofrontal ROI size in mm^3 Range value: 3216, ..., 7241

CTX_RH_MIDDLETEMPORAL_SUVR numeric ctx-rh-middletemporal Range value: 0.548, ..., 6.166

CTX_RH_MIDDLETEMPORAL_VOLUME numeric ctx-rh-middletemporal ROI size in mm^3 Range value: 5922, ..., 17219

CTX_RH_PARACENTRAL_SUVR numeric ctx-rh-paracentral Range value: 0.711, ..., 4.662

CTX_RH_PARACENTRAL_VOLUME numeric ctx-rh-paracentral ROI size in mm^3 Range value: 2261, ..., 5526

CTX_RH_PARAHIPPOCAMPAL_SUVR numeric ctx-rh-parahippocampal Range value: 0.589, ..., 5.711

CTX_RH_PARAHIPPOCAMPAL_VOLUME numeric ctx-rh-parahippocampal ROI size in mm^3 Range value: 1096, ..., 2799

CTX_RH_PARSOPERCULARIS_SUVR numeric ctx-rh-parsopercularis Range value: 0.689, ..., 6.516

CTX_RH_PARSOPERCULARIS_VOLUME numeric ctx-rh-parsopercularis ROI size in mm^3 Range value: 2139, ..., 7542

CTX_RH_PARSORBITALIS_SUVR numeric ctx-rh-parsorbitalis Range value: 0.705, ..., 5.569

CTX_RH_PARSORBITALIS_VOLUME numeric ctx-rh-parsorbitalis ROI size in mm^3 Range value: 1261, ..., 4301

CTX_RH_PARSTRIANGULARIS_SUVR numeric ctx-rh-parstriangularis Range value: 0.631, ..., 5.286

CTX_RH_PARSTRIANGULARIS_VOLUME numeric ctx-rh-parstriangularis ROI size in mm^3 Range value: 2268, ..., 7512

CTX_RH_PERICALCARINE_SUVR numeric ctx-rh-pericalcarine Range value: 0.156, ..., 6.99

CTX_RH_PERICALCARINE_VOLUME numeric ctx-rh-pericalcarine ROI size in mm^3 Range value: 831, ..., 4363

CTX_RH_POSTCENTRAL_SUVR numeric ctx-rh-postcentral Range value: 0.74, ..., 4.752

CTX_RH_POSTCENTRAL_VOLUME numeric ctx-rh-postcentral ROI size in mm^3 Range value: 4677, ..., 13254

CTX_RH_POSTERIORCINGULATE_SUVR numeric ctx-rh-posteriorcingulate Range value: 0.721, ..., 6.73

CTX_RH_POSTERIORCINGULATE_VOLUME numeric ctx-rh-posteriorcingulate ROI size in mm^3 Range value: 994, ..., 4499

CTX_RH_PRECENTRAL_SUVR numeric ctx-rh-precentral Range value: 0.743, ..., 3.159

CTX_RH_PRECENTRAL_VOLUME numeric ctx-rh-precentral ROI size in mm^3 Range value: 6369, ..., 18058

CTX_RH_PRECUNEUS_SUVR numeric ctx-rh-precuneus Range value: 0.71, ..., 7.6

CTX_RH_PRECUNEUS_VOLUME numeric ctx-rh-precuneus ROI size in mm^3 Range value: 5009, ..., 14854

CTX_RH_ROSTRALANTERIORCINGULATE_SUVR numeric ctx-rh-rostralanteriorcingulate Range value: 0.336, ..., 4.304

CTX_RH_ROSTRALANTERIORCINGULATE_VOLUME numeric ctx-rh-rostralanteriorcingulate ROI size in mm^3 Range value: 559, ..., 3308

CTX_RH_ROSTRALMIDDLEFRONTAL_SUVR numeric ctx-rh-rostralmiddlefrontal Range value: 0.774, ..., 8.814

CTX_RH_ROSTRALMIDDLEFRONTAL_VOLUME numeric ctx-rh-rostralmiddlefrontal ROI size in mm^3 Range value: 7884, ..., 26835

CTX_RH_SUPERIORFRONTAL_SUVR numeric ctx-rh-superiorfrontal Range value: 0.793, ..., 5.001

CTX_RH_SUPERIORFRONTAL_VOLUME numeric ctx-rh-superiorfrontal ROI size in mm^3 Range value: 11678, ..., 28618

CTX_RH_SUPERIORPARIETAL_SUVR numeric ctx-rh-superiorparietal Range value: 0.587, ..., 9.934

CTX_RH_SUPERIORPARIETAL_VOLUME numeric ctx-rh-superiorparietal ROI size in mm^3 Range value: 6110, ..., 18993

CTX_RH_SUPERIORTEMPORAL_SUVR numeric ctx-rh-superiortemporal Range value: 0.717, ..., 5.169

CTX_RH_SUPERIORTEMPORAL_VOLUME numeric ctx-rh-superiortemporal ROI size in mm^3 Range value: 5963, ..., 15383

CTX_RH_SUPRAMARGINAL_SUVR numeric ctx-rh-supramarginal Range value: 0.81, ..., 7.764

CTX_RH_SUPRAMARGINAL_VOLUME numeric ctx-rh-supramarginal ROI size in mm^3 Range value: 4622, ..., 14291

CTX_RH_TEMPORALPOLE_SUVR numeric ctx-rh-temporalpole Range value: 0.577, ..., 6.325

CTX_RH_TEMPORALPOLE_VOLUME numeric ctx-rh-temporalpole ROI size in mm^3 Range value: 287, ..., 4200

CTX_RH_TRANSVERSETEMPORAL_SUVR numeric ctx-rh-transversetemporal Range value: 0.08, ..., 5.439

CTX_RH_TRANSVERSETEMPORAL_VOLUME numeric ctx-rh-transversetemporal ROI size in mm^3 Range value: 461, ..., 1566

LEFT_ACCUMBENS_AREA_SUVR numeric left-accumbens-area Range value: 0.301, ..., 4.736

LEFT_ACCUMBENS_AREA_VOLUME numeric left-accumbens-area ROI size in mm^3 Range value: 116, ..., 821

LEFT_AMYGDALA_SUVR numeric left-amygdala Range value: 0.605, ..., 5.24

LEFT_AMYGDALA_VOLUME numeric left-amygdala ROI size in mm^3 Range value: 375, ..., 2468

LEFT_CAUDATE_SUVR numeric left-caudate Range value: 0.502, ..., 2.459

LEFT_CAUDATE_VOLUME numeric left-caudate ROI size in mm^3 Range value: 1993, ..., 5745

LEFT_CEREBELLUM_CORTEX_SUVR numeric left-cerebellum-cortex Range value: 0.443, ..., 1.356

LEFT_CEREBELLUM_CORTEX_VOLUME numeric left-cerebellum-cortex ROI size in mm^3 Range value: 36265, ..., 69630

LEFT_CEREBELLUM_WHITE_MATTER_SUVR numeric left-cerebellum-white-matter Range value: 0.492, ..., 1.685

LEFT_CEREBELLUM_WHITE_MATTER_VOLUME numeric left-cerebellum-white-matter ROI size in mm^3 Range value: 6892, ..., 26253

LEFT_CEREBRAL_WHITE_MATTER_SUVR numeric left-cerebral-white-matter Range value: 0.631, ..., 1.859

LEFT_CEREBRAL_WHITE_MATTER_VOLUME numeric left-cerebral-white-matter ROI size in mm^3 Range value: 30465, ..., 288688

LEFT_CHOROID_PLEXUS_SUVR numeric left-choroid-plexus Range value: 0.275, ..., 18.556

LEFT_CHOROID_PLEXUS_VOLUME numeric left-choroid-plexus ROI size in mm^3 Range value: 130, ..., 1916

LEFT_HIPPOCAMPUS_SUVR numeric left-hippocampus Range value: 0.425, ..., 3.66

LEFT_HIPPOCAMPUS_VOLUME numeric left-hippocampus ROI size in mm^3 Range value: 1314, ..., 6080

LEFT_INF_LAT_VENT_SUVR numeric left-inf-lat-vent Range value: 0.004, ..., 6.095

LEFT_INF_LAT_VENT_VOLUME numeric left-inf-lat-vent ROI size in mm^3 Range value: 60, ..., 11328

LEFT_LATERAL_VENTRICLE_SUVR numeric left-lateral-ventricle Range value: 0.005, ..., 1.244

LEFT_LATERAL_VENTRICLE_VOLUME numeric left-lateral-ventricle ROI size in mm^3 Range value: 3358, ..., 79775

LEFT_PALLIDUM_SUVR numeric left-pallidum Range value: 0.629, ..., 3.087

LEFT_PALLIDUM_VOLUME numeric left-pallidum ROI size in mm^3 Range value: 1090, ..., 3117

LEFT_PUTAMEN_SUVR numeric left-putamen Range value: 0.612, ..., 2.86

LEFT_PUTAMEN_VOLUME numeric left-putamen ROI size in mm^3 Range value: 1945, ..., 7466

LEFT_THALAMUS_PROPER_SUVR numeric left-thalamus-proper Range value: 0.644, ..., 1.895

LEFT_THALAMUS_PROPER_VOLUME numeric left-thalamus-proper ROI size in mm^3 Range value: 3953, ..., 9616

LEFT_VENTRALDC_SUVR numeric left-ventraldc Range value: 0.663, ..., 2.208

LEFT_VENTRALDC_VOLUME numeric left-ventraldc ROI size in mm^3 Range value: 2259, ..., 5676

LEFT_VESSEL_SUVR logical left-vessel Boolean value: TRUE/ FALSE

LEFT_VESSEL_VOLUME numeric left-vessel ROI size in mm^3 Range value: 1, ..., 537

RIGHT_ACCUMBENS_AREA_SUVR numeric right-accumbens-area Range value: 0.212, ..., 4.141

RIGHT_ACCUMBENS_AREA_VOLUME numeric right-accumbens-area ROI size in mm^3 Range value: 114, ..., 815

RIGHT_AMYGDALA_SUVR numeric right-amygdala Range value: 0.574, ..., 3.994

RIGHT_AMYGDALA_VOLUME numeric right-amygdala ROI size in mm^3 Range value: 605, ..., 2453

RIGHT_CAUDATE_SUVR numeric right-caudate Range value: 0.555, ..., 2.641

RIGHT_CAUDATE_VOLUME numeric right-caudate ROI size in mm^3 Range value: 2003, ..., 6880

RIGHT_CEREBELLUM_CORTEX_SUVR numeric right-cerebellum-cortex Range value: 0.534, ..., 1.357

RIGHT_CEREBELLUM_CORTEX_VOLUME numeric right-cerebellum-cortex ROI size in mm^3 Range value: 36079, ..., 73171

RIGHT_CEREBELLUM_WHITE_MATTER_SUVR numeric right-cerebellum-white-matter Range value: 0.492, ..., 1.685

RIGHT_CEREBELLUM_WHITE_MATTER_VOLUME numeric right-cerebellum-white-matter ROI size in mm^3 Range value: 6779, ..., 22532

RIGHT_CEREBRAL_WHITE_MATTER_SUVR numeric right-cerebral-white-matter Range value: 0.631, ..., 1.859

RIGHT_CEREBRAL_WHITE_MATTER_VOLUME numeric right-cerebral-white-matter ROI size in mm^3 Range value: 129854, ..., 290343

RIGHT_CHOROID_PLEXUS_SUVR numeric right-choroid-plexus Range value: 0.537, ..., 14.78

RIGHT_CHOROID_PLEXUS_VOLUME numeric right-choroid-plexus ROI size in mm^3 Range value: 316, ..., 1912

RIGHT_HIPPOCAMPUS_SUVR numeric right-hippocampus Range value: 0.675, ..., 3.09

RIGHT_HIPPOCAMPUS_VOLUME numeric right-hippocampus ROI size in mm^3 Range value: 1941, ..., 5775

RIGHT_INF_LAT_VENT_SUVR numeric right-inf-lat-vent Range value: 0.004, ..., 6.095

RIGHT_INF_LAT_VENT_VOLUME numeric right-inf-lat-vent ROI size in mm^3 Range value: 58, ..., 7228

RIGHT_LATERAL_VENTRICLE_SUVR numeric right-lateral-ventricle Range value: 0.005, ..., 1.244

RIGHT_LATERAL_VENTRICLE_VOLUME numeric right-lateral-ventricle ROI size in mm^3 Range value: 2442, ..., 71992

RIGHT_PALLIDUM_SUVR numeric right-pallidum Range value: 0.595, ..., 3.187

RIGHT_PALLIDUM_VOLUME numeric right-pallidum ROI size in mm^3 Range value: 1197, ..., 3207

RIGHT_PUTAMEN_SUVR numeric right-putamen Range value: 0.682, ..., 2.899

RIGHT_PUTAMEN_VOLUME numeric right-putamen ROI size in mm^3 Range value: 2222, ..., 8026

RIGHT_THALAMUS_PROPER_SUVR numeric right-thalamus-proper Range value: 0.64, ..., 2.224

RIGHT_THALAMUS_PROPER_VOLUME numeric right-thalamus-proper ROI size in mm^3 Range value: 4305, ..., 10375

RIGHT_VENTRALDC_SUVR numeric right-ventraldc Range value: 0.663, ..., 2.208

RIGHT_VENTRALDC_VOLUME numeric right-ventraldc ROI size in mm^3 Range value: 2384, ..., 5634

RIGHT_VESSEL_SUVR logical right-vessel Boolean value: TRUE/ FALSE

RIGHT_VESSEL_VOLUME numeric right-vessel ROI size in mm^3 Range value: 1, ..., 236

update_stamp POSIXct POSIXct: YYY-MM-DD

Author

adni-data@googlegroups.com