UC Berkeley - Tau PET 6mm Res analysis

UC Berkeley - Tau PET 6mm Res analysis data. More information is available at https://adni.loni.usc.edu/data-samples/adni-data/

Usage

data(UCBERKELEY_TAU_6MM)

Format

A data frame with 2327 observations and 340 variables.

Source

https://adni.loni.usc.edu/data-samples/adni-data/

Details

ORIGPROT character Original study protocol , with options: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4

LONIUID numeric LONI Image ID Range value: 513671, ..., 11432364

PTID character ADNI Participant ID consisting of 3-digit "SITEID" and 4-digit "RID"

RID numeric Participant roster ID (unique participant identifier that should be used to merge data) Range value: 21, ..., 10875

VISCODE character Visit code

VISCODE2 character Translated visit code

SCANDATE SCAN acquisition date Date: YYY-MM-DD

SITEID numeric Site ID Range value: 2, ..., 941

PROCESSDATE Date quantification pipeline was run Date: YYY-MM-DD

IMAGE_RESOLUTION character Spatial resolution in mm^3 FWHM of Step4 Preprocessed PET image. See UCBERKELEY PET 8mm-to-6mm Transformation Methods PDF on LONI for details. , with options: 6mm, 8mm

qc_flag numeric Range value: -2, ..., 2

TRACER character PET Radiotracer , with options: FTP, MK6240, PI2620

TRACER_SUVR_WARNING character Reminder of the incomparability of SUVRs from different tracers. See UCBERKELEY Tau Processing Methods PDF. , with options: DO NOT COMPARE SUVRs ACROSS TRACERS

META_TEMPORAL_SUVR numeric Weighted mean SUVR of Jack et al temporal meta-ROI. All SUVRs in this file are normalized to the inferior cerebellar grey matter unless otherwise noted. See UCBERKELEY Tau PET Processing Methods PDF on LONI for details. Range value: 0.774, ..., 3.833

META_TEMPORAL_VOLUME numeric Jack et al temporal meta-ROI size in mm^3 Range value: 36639, ..., 97274

CTX_ENTORHINAL_SUVR numeric ctx-bilateral-entorhinal Range value: 0.674, ..., 3.398

CTX_ENTORHINAL_VOLUME numeric ctx-bilateral-entorhinal ROI size in mm^3 Range value: 746, ..., 5946

INFERIORCEREBELLUM_SUVR numeric Reference region - mean SUVR of inferior cerebellar grey matter Range value: 1, ..., 1

INFERIORCEREBELLUM_VOLUME numeric Inferior Cerebellar Grey Matter ROI size in mm^3 Range value: 40753, ..., 87660

ERODED_SUBCORTICALWM_SUVR numeric Reference region - mean SUVR of subcortical white matter eroded away from cortex Range value: 0.455, ..., 1.943

ERODED_SUBCORTICALWM_VOLUME numeric Subcortical white matter eroded away from cortex size in mm^3 Range value: 55673, ..., 424418

BRAINSTEM_SUVR numeric brain-stem Range value: 0.46, ..., 1.562

BRAINSTEM_VOLUME numeric brain-stem ROI size in mm^3 Range value: 13764, ..., 31388

CC_ANTERIOR_SUVR numeric cc-anterior Range value: 0.445, ..., 1.376

CC_ANTERIOR_VOLUME numeric cc-anterior ROI size in mm^3 Range value: 67, ..., 1763

CC_CENTRAL_SUVR numeric cc-central Range value: 0.429, ..., 1.581

CC_CENTRAL_VOLUME numeric cc-central ROI size in mm^3 Range value: 233, ..., 1380

CC_MID_ANTERIOR_SUVR numeric cc-mid-anterior Range value: 0.444, ..., 1.486

CC_MID_ANTERIOR_VOLUME numeric cc-mid-anterior ROI size in mm^3 Range value: 59, ..., 1241

CC_MID_POSTERIOR_SUVR numeric cc-mid-posterior Range value: 0.329, ..., 1.52

CC_MID_POSTERIOR_VOLUME numeric cc-mid-posterior ROI size in mm^3 Range value: 189, ..., 1659

CC_POSTERIOR_SUVR numeric cc-posterior Range value: 0.514, ..., 2.063

CC_POSTERIOR_VOLUME numeric cc-posterior ROI size in mm^3 Range value: 259, ..., 2196

CSF_SUVR numeric csf Range value: 0.451, ..., 1.629

CSF_VOLUME numeric csf ROI size in mm^3 Range value: 628, ..., 3763

VENTRICLE_3RD_SUVR numeric 3rd-ventricle Range value: 0.382, ..., 1.243

VENTRICLE_3RD_VOLUME numeric 3rd-ventricle ROI size in mm^3 Range value: 484, ..., 5244

VENTRICLE_4TH_SUVR numeric 4th-ventricle Range value: 0.471, ..., 1.633

VENTRICLE_4TH_VOLUME numeric 4th-ventricle ROI size in mm^3 Range value: 739, ..., 5081

VENTRICLE_5TH_SUVR numeric 5th-ventricle Range value: 0.257, ..., 0.861

VENTRICLE_5TH_VOLUME numeric 5th-ventricle ROI size in mm^3 Range value: 1, ..., 11

WM_HYPOINTENSITIES_SUVR numeric wm-hypointensities Range value: 0.454, ..., 1.534

WM_HYPOINTENSITIES_VOLUME numeric wm-hypointensities ROI size in mm^3 Range value: 268, ..., 72336

NON_WM_HYPOINTENSITIES_SUVR numeric non-wm-hypointensities Range value: 0.276, ..., 1.867

NON_WM_HYPOINTENSITIES_VOLUME numeric non-wm-hypointensities ROI size in mm^3 Range value: 1, ..., 233

CTX_BANKSSTS_SUVR numeric ctx-bilateral-bankssts Range value: 0.629, ..., 4.493

CTX_BANKSSTS_VOLUME numeric ctx-bilateral-bankssts ROI size in mm^3 Range value: 2121, ..., 8963

CTX_CAUDALANTERIORCINGULATE_SUVR numeric ctx-bilateral-caudalanteriorcingulate Range value: 0.531, ..., 2.575

CTX_CAUDALANTERIORCINGULATE_VOLUME numeric ctx-bilateral-caudalanteriorcingulate ROI size in mm^3 Range value: 1309, ..., 8378

CTX_CAUDALMIDDLEFRONTAL_SUVR numeric ctx-bilateral-caudalmiddlefrontal Range value: 0.587, ..., 3.909

CTX_CAUDALMIDDLEFRONTAL_VOLUME numeric ctx-bilateral-caudalmiddlefrontal ROI size in mm^3 Range value: 5775, ..., 16670

CTX_CUNEUS_SUVR numeric ctx-bilateral-cuneus Range value: 0.757, ..., 4.443

CTX_CUNEUS_VOLUME numeric ctx-bilateral-cuneus ROI size in mm^3 Range value: 3366, ..., 10467

CTX_FRONTALPOLE_SUVR numeric ctx-bilateral-frontalpole Range value: 0.555, ..., 3.382

CTX_FRONTALPOLE_VOLUME numeric ctx-bilateral-frontalpole ROI size in mm^3 Range value: 982, ..., 3888

CTX_FUSIFORM_SUVR numeric ctx-bilateral-fusiform Range value: 0.8, ..., 4.095

CTX_FUSIFORM_VOLUME numeric ctx-bilateral-fusiform ROI size in mm^3 Range value: 9510, ..., 26688

CTX_INFERIORPARIETAL_SUVR numeric ctx-bilateral-inferiorparietal Range value: 0.807, ..., 4.384

CTX_INFERIORPARIETAL_VOLUME numeric ctx-bilateral-inferiorparietal ROI size in mm^3 Range value: 12378, ..., 37678

CTX_INFERIORTEMPORAL_SUVR numeric ctx-bilateral-inferiortemporal Range value: 0.803, ..., 4.528

CTX_INFERIORTEMPORAL_VOLUME numeric ctx-bilateral-inferiortemporal ROI size in mm^3 Range value: 9660, ..., 31454

CTX_INSULA_SUVR numeric ctx-bilateral-insula Range value: 0.574, ..., 2.95

CTX_INSULA_VOLUME numeric ctx-bilateral-insula ROI size in mm^3 Range value: 7460, ..., 27196

CTX_ISTHMUSCINGULATE_SUVR numeric ctx-bilateral-isthmuscingulate Range value: 0.626, ..., 3.888

CTX_ISTHMUSCINGULATE_VOLUME numeric ctx-bilateral-isthmuscingulate ROI size in mm^3 Range value: 1985, ..., 9811

CTX_LATERALOCCIPITAL_SUVR numeric ctx-bilateral-lateraloccipital Range value: 0.853, ..., 5.025

CTX_LATERALOCCIPITAL_VOLUME numeric ctx-bilateral-lateraloccipital ROI size in mm^3 Range value: 9370, ..., 35627

CTX_LATERALORBITOFRONTAL_SUVR numeric ctx-bilateral-lateralorbitofrontal Range value: 0.656, ..., 3.399

CTX_LATERALORBITOFRONTAL_VOLUME numeric ctx-bilateral-lateralorbitofrontal ROI size in mm^3 Range value: 8842, ..., 18822

CTX_LINGUAL_SUVR numeric ctx-bilateral-lingual Range value: 0.796, ..., 3.919

CTX_LINGUAL_VOLUME numeric ctx-bilateral-lingual ROI size in mm^3 Range value: 5946, ..., 20629

CTX_MEDIALORBITOFRONTAL_SUVR numeric ctx-bilateral-medialorbitofrontal Range value: 0.576, ..., 3.174

CTX_MEDIALORBITOFRONTAL_VOLUME numeric ctx-bilateral-medialorbitofrontal ROI size in mm^3 Range value: 6625, ..., 13876

CTX_MIDDLETEMPORAL_SUVR numeric ctx-bilateral-middletemporal Range value: 0.793, ..., 4

CTX_MIDDLETEMPORAL_VOLUME numeric ctx-bilateral-middletemporal ROI size in mm^3 Range value: 11213, ..., 31685

CTX_PARACENTRAL_SUVR numeric ctx-bilateral-paracentral Range value: 0.552, ..., 2.363

CTX_PARACENTRAL_VOLUME numeric ctx-bilateral-paracentral ROI size in mm^3 Range value: 4500, ..., 10986

CTX_PARAHIPPOCAMPAL_SUVR numeric ctx-bilateral-parahippocampal Range value: 0.682, ..., 3.196

CTX_PARAHIPPOCAMPAL_VOLUME numeric ctx-bilateral-parahippocampal ROI size in mm^3 Range value: 1751, ..., 5685

CTX_PARSOPERCULARIS_SUVR numeric ctx-bilateral-parsopercularis Range value: 0.591, ..., 3.589

CTX_PARSOPERCULARIS_VOLUME numeric ctx-bilateral-parsopercularis ROI size in mm^3 Range value: 4801, ..., 12836

CTX_PARSORBITALIS_SUVR numeric ctx-bilateral-parsorbitalis Range value: 0.811, ..., 3.85

CTX_PARSORBITALIS_VOLUME numeric ctx-bilateral-parsorbitalis ROI size in mm^3 Range value: 2452, ..., 6892

CTX_PARSTRIANGULARIS_SUVR numeric ctx-bilateral-parstriangularis Range value: 0.715, ..., 3.512

CTX_PARSTRIANGULARIS_VOLUME numeric ctx-bilateral-parstriangularis ROI size in mm^3 Range value: 4786, ..., 11719

CTX_PERICALCARINE_SUVR numeric ctx-bilateral-pericalcarine Range value: 0.658, ..., 3.439

CTX_PERICALCARINE_VOLUME numeric ctx-bilateral-pericalcarine ROI size in mm^3 Range value: 2109, ..., 8270

CTX_POSTCENTRAL_SUVR numeric ctx-bilateral-postcentral Range value: 0.702, ..., 2.411

CTX_POSTCENTRAL_VOLUME numeric ctx-bilateral-postcentral ROI size in mm^3 Range value: 10032, ..., 26525

CTX_POSTERIORCINGULATE_SUVR numeric ctx-bilateral-posteriorcingulate Range value: 0.565, ..., 3.46

CTX_POSTERIORCINGULATE_VOLUME numeric ctx-bilateral-posteriorcingulate ROI size in mm^3 Range value: 2054, ..., 18216

CTX_PRECENTRAL_SUVR numeric ctx-bilateral-precentral Range value: 0.607, ..., 2.036

CTX_PRECENTRAL_VOLUME numeric ctx-bilateral-precentral ROI size in mm^3 Range value: 13766, ..., 37386

CTX_PRECUNEUS_SUVR numeric ctx-bilateral-precuneus Range value: 0.638, ..., 4.037

CTX_PRECUNEUS_VOLUME numeric ctx-bilateral-precuneus ROI size in mm^3 Range value: 9272, ..., 27581

CTX_ROSTRALANTERIORCINGULATE_SUVR numeric ctx-bilateral-rostralanteriorcingulate Range value: 0.576, ..., 2.972

CTX_ROSTRALANTERIORCINGULATE_VOLUME numeric ctx-bilateral-rostralanteriorcingulate ROI size in mm^3 Range value: 1558, ..., 7474

CTX_ROSTRALMIDDLEFRONTAL_SUVR numeric ctx-bilateral-rostralmiddlefrontal Range value: 0.685, ..., 4.571

CTX_ROSTRALMIDDLEFRONTAL_VOLUME numeric ctx-bilateral-rostralmiddlefrontal ROI size in mm^3 Range value: 16054, ..., 50046

CTX_SUPERIORFRONTAL_SUVR numeric ctx-bilateral-superiorfrontal Range value: 0.525, ..., 3.007

CTX_SUPERIORFRONTAL_VOLUME numeric ctx-bilateral-superiorfrontal ROI size in mm^3 Range value: 25550, ..., 58830

CTX_SUPERIORPARIETAL_SUVR numeric ctx-bilateral-superiorparietal Range value: 0.701, ..., 4.043

CTX_SUPERIORPARIETAL_VOLUME numeric ctx-bilateral-superiorparietal ROI size in mm^3 Range value: 12027, ..., 37095

CTX_SUPERIORTEMPORAL_SUVR numeric ctx-bilateral-superiortemporal Range value: 0.703, ..., 2.914

CTX_SUPERIORTEMPORAL_VOLUME numeric ctx-bilateral-superiortemporal ROI size in mm^3 Range value: 12461, ..., 32667

CTX_SUPRAMARGINAL_SUVR numeric ctx-bilateral-supramarginal Range value: 0.72, ..., 3.539

CTX_SUPRAMARGINAL_VOLUME numeric ctx-bilateral-supramarginal ROI size in mm^3 Range value: 10961, ..., 31391

CTX_TEMPORALPOLE_SUVR numeric ctx-bilateral-temporalpole Range value: 0.785, ..., 3.797

CTX_TEMPORALPOLE_VOLUME numeric ctx-bilateral-temporalpole ROI size in mm^3 Range value: 2178, ..., 9644

CTX_TRANSVERSETEMPORAL_SUVR numeric ctx-bilateral-transversetemporal Range value: 0.594, ..., 2.568

CTX_TRANSVERSETEMPORAL_VOLUME numeric ctx-bilateral-transversetemporal ROI size in mm^3 Range value: 1100, ..., 3541

ACCUMBENS_AREA_SUVR numeric bilateral-accumbens-area Range value: 0.402, ..., 2.475

ACCUMBENS_AREA_VOLUME numeric bilateral-accumbens-area ROI size in mm^3 Range value: 300, ..., 1636

AMYGDALA_SUVR numeric bilateral-amygdala Range value: 0.45, ..., 3.175

AMYGDALA_VOLUME numeric bilateral-amygdala ROI size in mm^3 Range value: 980, ..., 4740

CAUDATE_SUVR numeric bilateral-caudate Range value: 0.432, ..., 1.696

CAUDATE_VOLUME numeric bilateral-caudate ROI size in mm^3 Range value: 3588, ..., 12321

CEREBELLUM_CORTEX_SUVR numeric bilateral-cerebellum-cortex Range value: 0.947, ..., 1.277

CEREBELLUM_CORTEX_VOLUME numeric bilateral-cerebellum-cortex ROI size in mm^3 Range value: 68832, ..., 142592

CEREBELLUM_WHITE_MATTER_SUVR numeric bilateral-cerebellum-white-matter Range value: 0.598, ..., 1.485

CEREBELLUM_WHITE_MATTER_VOLUME numeric bilateral-cerebellum-white-matter ROI size in mm^3 Range value: 13671, ..., 48785

CEREBRAL_WHITE_MATTER_SUVR numeric bilateral-cerebral-white-matter Range value: 0.517, ..., 2.576

CEREBRAL_WHITE_MATTER_VOLUME numeric bilateral-cerebral-white-matter ROI size in mm^3 Range value: 224785, ..., 718037

CHOROID_PLEXUS_SUVR numeric bilateral-choroid-plexus Range value: 0.521, ..., 3.465

CHOROID_PLEXUS_VOLUME numeric bilateral-choroid-plexus ROI size in mm^3 Range value: 459, ..., 3434

HIPPOCAMPUS_SUVR numeric bilateral-hippocampus Range value: 0.512, ..., 2.761

HIPPOCAMPUS_VOLUME numeric bilateral-hippocampus ROI size in mm^3 Range value: 3255, ..., 11461

OPTIC_CHIASM_SUVR numeric optic-chiasm Range value: 0.406, ..., 1.953

OPTIC_CHIASM_VOLUME numeric optic-chiasm ROI size in mm^3 Range value: 2, ..., 361

INF_LAT_VENT_SUVR numeric bilateral-inf-lat-vent Range value: 0.504, ..., 2.291

INF_LAT_VENT_VOLUME numeric bilateral-inf-lat-vent ROI size in mm^3 Range value: 164, ..., 16102

LATERAL_VENTRICLE_SUVR numeric bilateral-lateral-ventricle Range value: 0.218, ..., 1.468

LATERAL_VENTRICLE_VOLUME numeric bilateral-lateral-ventricle ROI size in mm^3 Range value: 5800, ..., 138561

PALLIDUM_SUVR numeric bilateral-pallidum Range value: 0.475, ..., 2.529

PALLIDUM_VOLUME numeric bilateral-pallidum ROI size in mm^3 Range value: 2454, ..., 6324

PUTAMEN_SUVR numeric bilateral-putamen Range value: 0.504, ..., 2.4

PUTAMEN_VOLUME numeric bilateral-putamen ROI size in mm^3 Range value: 4438, ..., 15492

THALAMUS_PROPER_SUVR numeric bilateral-thalamus-proper Range value: 0.481, ..., 2.088

THALAMUS_PROPER_VOLUME numeric bilateral-thalamus-proper ROI size in mm^3 Range value: 8557, ..., 20466

VENTRALDC_SUVR numeric bilateral-ventraldc Range value: 0.51, ..., 1.833

VENTRALDC_VOLUME numeric bilateral-ventraldc ROI size in mm^3 Range value: 4441, ..., 11382

VESSEL_SUVR numeric bilateral-vessel Range value: 0.6, ..., 2.342

VESSEL_VOLUME numeric bilateral-vessel ROI size in mm^3 Range value: 1, ..., 604

CTX_LH_BANKSSTS_SUVR numeric ctx-lh-bankssts Range value: 0.643, ..., 4.46

CTX_LH_BANKSSTS_VOLUME numeric ctx-lh-bankssts ROI size in mm^3 Range value: 779, ..., 4723

CTX_LH_CAUDALANTERIORCINGULATE_SUVR numeric ctx-lh-caudalanteriorcingulate Range value: 0.533, ..., 2.563

CTX_LH_CAUDALANTERIORCINGULATE_VOLUME numeric ctx-lh-caudalanteriorcingulate ROI size in mm^3 Range value: 352, ..., 4960

CTX_LH_CAUDALMIDDLEFRONTAL_SUVR numeric ctx-lh-caudalmiddlefrontal Range value: 0.584, ..., 4.124

CTX_LH_CAUDALMIDDLEFRONTAL_VOLUME numeric ctx-lh-caudalmiddlefrontal ROI size in mm^3 Range value: 2609, ..., 12065

CTX_LH_CUNEUS_SUVR numeric ctx-lh-cuneus Range value: 0.777, ..., 4.044

CTX_LH_CUNEUS_VOLUME numeric ctx-lh-cuneus ROI size in mm^3 Range value: 1411, ..., 5342

CTX_LH_ENTORHINAL_SUVR numeric ctx-lh-entorhinal Range value: 0.702, ..., 3.3

CTX_LH_ENTORHINAL_VOLUME numeric ctx-lh-entorhinal ROI size in mm^3 Range value: 291, ..., 3408

CTX_LH_FRONTALPOLE_SUVR numeric ctx-lh-frontalpole Range value: 0.58, ..., 3.351

CTX_LH_FRONTALPOLE_VOLUME numeric ctx-lh-frontalpole ROI size in mm^3 Range value: 486, ..., 1796

CTX_LH_FUSIFORM_SUVR numeric ctx-lh-fusiform Range value: 0.799, ..., 4.258

CTX_LH_FUSIFORM_VOLUME numeric ctx-lh-fusiform ROI size in mm^3 Range value: 3944, ..., 13678

CTX_LH_INFERIORPARIETAL_SUVR numeric ctx-lh-inferiorparietal Range value: 0.796, ..., 4.489

CTX_LH_INFERIORPARIETAL_VOLUME numeric ctx-lh-inferiorparietal ROI size in mm^3 Range value: 5047, ..., 17756

CTX_LH_INFERIORTEMPORAL_SUVR numeric ctx-lh-inferiortemporal Range value: 0.791, ..., 4.312

CTX_LH_INFERIORTEMPORAL_VOLUME numeric ctx-lh-inferiortemporal ROI size in mm^3 Range value: 4547, ..., 16512

CTX_LH_INSULA_SUVR numeric ctx-lh-insula Range value: 0.574, ..., 2.909

CTX_LH_INSULA_VOLUME numeric ctx-lh-insula ROI size in mm^3 Range value: 3825, ..., 21138

CTX_LH_ISTHMUSCINGULATE_SUVR numeric ctx-lh-isthmuscingulate Range value: 0.641, ..., 3.946

CTX_LH_ISTHMUSCINGULATE_VOLUME numeric ctx-lh-isthmuscingulate ROI size in mm^3 Range value: 615, ..., 6967

CTX_LH_LATERALOCCIPITAL_SUVR numeric ctx-lh-lateraloccipital Range value: 0.835, ..., 4.843

CTX_LH_LATERALOCCIPITAL_VOLUME numeric ctx-lh-lateraloccipital ROI size in mm^3 Range value: 4386, ..., 17628

CTX_LH_LATERALORBITOFRONTAL_SUVR numeric ctx-lh-lateralorbitofrontal Range value: 0.657, ..., 3.759

CTX_LH_LATERALORBITOFRONTAL_VOLUME numeric ctx-lh-lateralorbitofrontal ROI size in mm^3 Range value: 3412, ..., 11834

CTX_LH_LINGUAL_SUVR numeric ctx-lh-lingual Range value: 0.81, ..., 3.763

CTX_LH_LINGUAL_VOLUME numeric ctx-lh-lingual ROI size in mm^3 Range value: 2540, ..., 11629

CTX_LH_MEDIALORBITOFRONTAL_SUVR numeric ctx-lh-medialorbitofrontal Range value: 0.574, ..., 3.151

CTX_LH_MEDIALORBITOFRONTAL_VOLUME numeric ctx-lh-medialorbitofrontal ROI size in mm^3 Range value: 3183, ..., 7019

CTX_LH_MIDDLETEMPORAL_SUVR numeric ctx-lh-middletemporal Range value: 0.788, ..., 4.237

CTX_LH_MIDDLETEMPORAL_VOLUME numeric ctx-lh-middletemporal ROI size in mm^3 Range value: 4242, ..., 14767

CTX_LH_PARACENTRAL_SUVR numeric ctx-lh-paracentral Range value: 0.578, ..., 2.094

CTX_LH_PARACENTRAL_VOLUME numeric ctx-lh-paracentral ROI size in mm^3 Range value: 1868, ..., 7180

CTX_LH_PARAHIPPOCAMPAL_SUVR numeric ctx-lh-parahippocampal Range value: 0.691, ..., 3.139

CTX_LH_PARAHIPPOCAMPAL_VOLUME numeric ctx-lh-parahippocampal ROI size in mm^3 Range value: 938, ..., 3643

CTX_LH_PARSOPERCULARIS_SUVR numeric ctx-lh-parsopercularis Range value: 0.576, ..., 3.525

CTX_LH_PARSOPERCULARIS_VOLUME numeric ctx-lh-parsopercularis ROI size in mm^3 Range value: 2451, ..., 7630

CTX_LH_PARSORBITALIS_SUVR numeric ctx-lh-parsorbitalis Range value: 0.781, ..., 4.214

CTX_LH_PARSORBITALIS_VOLUME numeric ctx-lh-parsorbitalis ROI size in mm^3 Range value: 438, ..., 3294

CTX_LH_PARSTRIANGULARIS_SUVR numeric ctx-lh-parstriangularis Range value: 0.679, ..., 3.546

CTX_LH_PARSTRIANGULARIS_VOLUME numeric ctx-lh-parstriangularis ROI size in mm^3 Range value: 1834, ..., 5545

CTX_LH_PERICALCARINE_SUVR numeric ctx-lh-pericalcarine Range value: 0.668, ..., 3.021

CTX_LH_PERICALCARINE_VOLUME numeric ctx-lh-pericalcarine ROI size in mm^3 Range value: 755, ..., 4678

CTX_LH_POSTCENTRAL_SUVR numeric ctx-lh-postcentral Range value: 0.692, ..., 2.335

CTX_LH_POSTCENTRAL_VOLUME numeric ctx-lh-postcentral ROI size in mm^3 Range value: 4696, ..., 14589

CTX_LH_POSTERIORCINGULATE_SUVR numeric ctx-lh-posteriorcingulate Range value: 0.598, ..., 3.512

CTX_LH_POSTERIORCINGULATE_VOLUME numeric ctx-lh-posteriorcingulate ROI size in mm^3 Range value: 990, ..., 11595

CTX_LH_PRECENTRAL_SUVR numeric ctx-lh-precentral Range value: 0.597, ..., 2.202

CTX_LH_PRECENTRAL_VOLUME numeric ctx-lh-precentral ROI size in mm^3 Range value: 6606, ..., 25709

CTX_LH_PRECUNEUS_SUVR numeric ctx-lh-precuneus Range value: 0.637, ..., 4.285

CTX_LH_PRECUNEUS_VOLUME numeric ctx-lh-precuneus ROI size in mm^3 Range value: 4103, ..., 17038

CTX_LH_ROSTRALANTERIORCINGULATE_SUVR numeric ctx-lh-rostralanteriorcingulate Range value: 0.59, ..., 2.997

CTX_LH_ROSTRALANTERIORCINGULATE_VOLUME numeric ctx-lh-rostralanteriorcingulate ROI size in mm^3 Range value: 571, ..., 5073

CTX_LH_ROSTRALMIDDLEFRONTAL_SUVR numeric ctx-lh-rostralmiddlefrontal Range value: 0.676, ..., 4.723

CTX_LH_ROSTRALMIDDLEFRONTAL_VOLUME numeric ctx-lh-rostralmiddlefrontal ROI size in mm^3 Range value: 6861, ..., 23265

CTX_LH_SUPERIORFRONTAL_SUVR numeric ctx-lh-superiorfrontal Range value: 0.491, ..., 3.231

CTX_LH_SUPERIORFRONTAL_VOLUME numeric ctx-lh-superiorfrontal ROI size in mm^3 Range value: 13132, ..., 39158

CTX_LH_SUPERIORPARIETAL_SUVR numeric ctx-lh-superiorparietal Range value: 0.704, ..., 4.091

CTX_LH_SUPERIORPARIETAL_VOLUME numeric ctx-lh-superiorparietal ROI size in mm^3 Range value: 5411, ..., 23420

CTX_LH_SUPERIORTEMPORAL_SUVR numeric ctx-lh-superiortemporal Range value: 0.688, ..., 3.121

CTX_LH_SUPERIORTEMPORAL_VOLUME numeric ctx-lh-superiortemporal ROI size in mm^3 Range value: 5891, ..., 18312

CTX_LH_SUPRAMARGINAL_SUVR numeric ctx-lh-supramarginal Range value: 0.714, ..., 3.696

CTX_LH_SUPRAMARGINAL_VOLUME numeric ctx-lh-supramarginal ROI size in mm^3 Range value: 5543, ..., 17431

CTX_LH_TEMPORALPOLE_SUVR numeric ctx-lh-temporalpole Range value: 0.761, ..., 3.788

CTX_LH_TEMPORALPOLE_VOLUME numeric ctx-lh-temporalpole ROI size in mm^3 Range value: 48, ..., 5851

CTX_LH_TRANSVERSETEMPORAL_SUVR numeric ctx-lh-transversetemporal Range value: 0.598, ..., 2.81

CTX_LH_TRANSVERSETEMPORAL_VOLUME numeric ctx-lh-transversetemporal ROI size in mm^3 Range value: 601, ..., 2254

CTX_RH_BANKSSTS_SUVR numeric ctx-rh-bankssts Range value: 0.614, ..., 4.997

CTX_RH_BANKSSTS_VOLUME numeric ctx-rh-bankssts ROI size in mm^3 Range value: 952, ..., 4240

CTX_RH_CAUDALANTERIORCINGULATE_SUVR numeric ctx-rh-caudalanteriorcingulate Range value: 0.526, ..., 2.584

CTX_RH_CAUDALANTERIORCINGULATE_VOLUME numeric ctx-rh-caudalanteriorcingulate ROI size in mm^3 Range value: 368, ..., 4843

CTX_RH_CAUDALMIDDLEFRONTAL_SUVR numeric ctx-rh-caudalmiddlefrontal Range value: 0.59, ..., 3.7

CTX_RH_CAUDALMIDDLEFRONTAL_VOLUME numeric ctx-rh-caudalmiddlefrontal ROI size in mm^3 Range value: 2379, ..., 9461

CTX_RH_CUNEUS_SUVR numeric ctx-rh-cuneus Range value: 0.737, ..., 4.89

CTX_RH_CUNEUS_VOLUME numeric ctx-rh-cuneus ROI size in mm^3 Range value: 982, ..., 6165

CTX_RH_ENTORHINAL_SUVR numeric ctx-rh-entorhinal Range value: 0.64, ..., 3.625

CTX_RH_ENTORHINAL_VOLUME numeric ctx-rh-entorhinal ROI size in mm^3 Range value: 411, ..., 3452

CTX_RH_FRONTALPOLE_SUVR numeric ctx-rh-frontalpole Range value: 0.535, ..., 3.409

CTX_RH_FRONTALPOLE_VOLUME numeric ctx-rh-frontalpole ROI size in mm^3 Range value: 381, ..., 2370

CTX_RH_FUSIFORM_SUVR numeric ctx-rh-fusiform Range value: 0.8, ..., 4.335

CTX_RH_FUSIFORM_VOLUME numeric ctx-rh-fusiform ROI size in mm^3 Range value: 4673, ..., 14227

CTX_RH_INFERIORPARIETAL_SUVR numeric ctx-rh-inferiorparietal Range value: 0.798, ..., 4.717

CTX_RH_INFERIORPARIETAL_VOLUME numeric ctx-rh-inferiorparietal ROI size in mm^3 Range value: 6272, ..., 21817

CTX_RH_INFERIORTEMPORAL_SUVR numeric ctx-rh-inferiortemporal Range value: 0.815, ..., 4.886

CTX_RH_INFERIORTEMPORAL_VOLUME numeric ctx-rh-inferiortemporal ROI size in mm^3 Range value: 4566, ..., 16344

CTX_RH_INSULA_SUVR numeric ctx-rh-insula Range value: 0.573, ..., 2.99

CTX_RH_INSULA_VOLUME numeric ctx-rh-insula ROI size in mm^3 Range value: 3635, ..., 9212

CTX_RH_ISTHMUSCINGULATE_SUVR numeric ctx-rh-isthmuscingulate Range value: 0.609, ..., 3.835

CTX_RH_ISTHMUSCINGULATE_VOLUME numeric ctx-rh-isthmuscingulate ROI size in mm^3 Range value: 783, ..., 4091

CTX_RH_LATERALOCCIPITAL_SUVR numeric ctx-rh-lateraloccipital Range value: 0.842, ..., 5.236

CTX_RH_LATERALOCCIPITAL_VOLUME numeric ctx-rh-lateraloccipital ROI size in mm^3 Range value: 4984, ..., 18366

CTX_RH_LATERALORBITOFRONTAL_SUVR numeric ctx-rh-lateralorbitofrontal Range value: 0.656, ..., 3.115

CTX_RH_LATERALORBITOFRONTAL_VOLUME numeric ctx-rh-lateralorbitofrontal ROI size in mm^3 Range value: 4176, ..., 9667

CTX_RH_LINGUAL_SUVR numeric ctx-rh-lingual Range value: 0.76, ..., 4.088

CTX_RH_LINGUAL_VOLUME numeric ctx-rh-lingual ROI size in mm^3 Range value: 3406, ..., 10478

CTX_RH_MEDIALORBITOFRONTAL_SUVR numeric ctx-rh-medialorbitofrontal Range value: 0.577, ..., 3.198

CTX_RH_MEDIALORBITOFRONTAL_VOLUME numeric ctx-rh-medialorbitofrontal ROI size in mm^3 Range value: 3216, ..., 7241

CTX_RH_MIDDLETEMPORAL_SUVR numeric ctx-rh-middletemporal Range value: 0.797, ..., 4.048

CTX_RH_MIDDLETEMPORAL_VOLUME numeric ctx-rh-middletemporal ROI size in mm^3 Range value: 5922, ..., 17219

CTX_RH_PARACENTRAL_SUVR numeric ctx-rh-paracentral Range value: 0.517, ..., 2.749

CTX_RH_PARACENTRAL_VOLUME numeric ctx-rh-paracentral ROI size in mm^3 Range value: 2261, ..., 5526

CTX_RH_PARAHIPPOCAMPAL_SUVR numeric ctx-rh-parahippocampal Range value: 0.641, ..., 3.329

CTX_RH_PARAHIPPOCAMPAL_VOLUME numeric ctx-rh-parahippocampal ROI size in mm^3 Range value: 1096, ..., 2800

CTX_RH_PARSOPERCULARIS_SUVR numeric ctx-rh-parsopercularis Range value: 0.612, ..., 3.672

CTX_RH_PARSOPERCULARIS_VOLUME numeric ctx-rh-parsopercularis ROI size in mm^3 Range value: 1783, ..., 7542

CTX_RH_PARSORBITALIS_SUVR numeric ctx-rh-parsorbitalis Range value: 0.792, ..., 3.499

CTX_RH_PARSORBITALIS_VOLUME numeric ctx-rh-parsorbitalis ROI size in mm^3 Range value: 1261, ..., 4301

CTX_RH_PARSTRIANGULARIS_SUVR numeric ctx-rh-parstriangularis Range value: 0.743, ..., 3.481

CTX_RH_PARSTRIANGULARIS_VOLUME numeric ctx-rh-parstriangularis ROI size in mm^3 Range value: 2268, ..., 7512

CTX_RH_PERICALCARINE_SUVR numeric ctx-rh-pericalcarine Range value: 0.649, ..., 3.823

CTX_RH_PERICALCARINE_VOLUME numeric ctx-rh-pericalcarine ROI size in mm^3 Range value: 831, ..., 4617

CTX_RH_POSTCENTRAL_SUVR numeric ctx-rh-postcentral Range value: 0.707, ..., 2.677

CTX_RH_POSTCENTRAL_VOLUME numeric ctx-rh-postcentral ROI size in mm^3 Range value: 4677, ..., 13254

CTX_RH_POSTERIORCINGULATE_SUVR numeric ctx-rh-posteriorcingulate Range value: 0.526, ..., 3.412

CTX_RH_POSTERIORCINGULATE_VOLUME numeric ctx-rh-posteriorcingulate ROI size in mm^3 Range value: 994, ..., 6621

CTX_RH_PRECENTRAL_SUVR numeric ctx-rh-precentral Range value: 0.617, ..., 2.176

CTX_RH_PRECENTRAL_VOLUME numeric ctx-rh-precentral ROI size in mm^3 Range value: 6369, ..., 18058

CTX_RH_PRECUNEUS_SUVR numeric ctx-rh-precuneus Range value: 0.638, ..., 3.817

CTX_RH_PRECUNEUS_VOLUME numeric ctx-rh-precuneus ROI size in mm^3 Range value: 5009, ..., 14854

CTX_RH_ROSTRALANTERIORCINGULATE_SUVR numeric ctx-rh-rostralanteriorcingulate Range value: 0.549, ..., 2.942

CTX_RH_ROSTRALANTERIORCINGULATE_VOLUME numeric ctx-rh-rostralanteriorcingulate ROI size in mm^3 Range value: 559, ..., 3508

CTX_RH_ROSTRALMIDDLEFRONTAL_SUVR numeric ctx-rh-rostralmiddlefrontal Range value: 0.693, ..., 4.42

CTX_RH_ROSTRALMIDDLEFRONTAL_VOLUME numeric ctx-rh-rostralmiddlefrontal ROI size in mm^3 Range value: 7884, ..., 26835

CTX_RH_SUPERIORFRONTAL_SUVR numeric ctx-rh-superiorfrontal Range value: 0.555, ..., 2.784

CTX_RH_SUPERIORFRONTAL_VOLUME numeric ctx-rh-superiorfrontal ROI size in mm^3 Range value: 10729, ..., 29394

CTX_RH_SUPERIORPARIETAL_SUVR numeric ctx-rh-superiorparietal Range value: 0.679, ..., 4.144

CTX_RH_SUPERIORPARIETAL_VOLUME numeric ctx-rh-superiorparietal ROI size in mm^3 Range value: 6110, ..., 18993

CTX_RH_SUPERIORTEMPORAL_SUVR numeric ctx-rh-superiortemporal Range value: 0.718, ..., 3.153

CTX_RH_SUPERIORTEMPORAL_VOLUME numeric ctx-rh-superiortemporal ROI size in mm^3 Range value: 5963, ..., 15383

CTX_RH_SUPRAMARGINAL_SUVR numeric ctx-rh-supramarginal Range value: 0.726, ..., 3.95

CTX_RH_SUPRAMARGINAL_VOLUME numeric ctx-rh-supramarginal ROI size in mm^3 Range value: 4622, ..., 14663

CTX_RH_TEMPORALPOLE_SUVR numeric ctx-rh-temporalpole Range value: 0.727, ..., 3.802

CTX_RH_TEMPORALPOLE_VOLUME numeric ctx-rh-temporalpole ROI size in mm^3 Range value: 287, ..., 4666

CTX_RH_TRANSVERSETEMPORAL_SUVR numeric ctx-rh-transversetemporal Range value: 0.589, ..., 2.637

CTX_RH_TRANSVERSETEMPORAL_VOLUME numeric ctx-rh-transversetemporal ROI size in mm^3 Range value: 461, ..., 1566

LEFT_ACCUMBENS_AREA_SUVR numeric left-accumbens-area Range value: 0.391, ..., 2.533

LEFT_ACCUMBENS_AREA_VOLUME numeric left-accumbens-area ROI size in mm^3 Range value: 116, ..., 821

LEFT_AMYGDALA_SUVR numeric left-amygdala Range value: 0.452, ..., 3.177

LEFT_AMYGDALA_VOLUME numeric left-amygdala ROI size in mm^3 Range value: 375, ..., 2468

LEFT_CAUDATE_SUVR numeric left-caudate Range value: 0.435, ..., 1.709

LEFT_CAUDATE_VOLUME numeric left-caudate ROI size in mm^3 Range value: 1871, ..., 5745

LEFT_CEREBELLUM_CORTEX_SUVR numeric left-cerebellum-cortex Range value: 0.939, ..., 1.232

LEFT_CEREBELLUM_CORTEX_VOLUME numeric left-cerebellum-cortex ROI size in mm^3 Range value: 34655, ..., 69630

LEFT_CEREBELLUM_WHITE_MATTER_SUVR numeric left-cerebellum-white-matter Range value: 0.597, ..., 1.502

LEFT_CEREBELLUM_WHITE_MATTER_VOLUME numeric left-cerebellum-white-matter ROI size in mm^3 Range value: 6892, ..., 26253

LEFT_CEREBRAL_WHITE_MATTER_SUVR numeric left-cerebral-white-matter Range value: 0.519, ..., 2.69

LEFT_CEREBRAL_WHITE_MATTER_VOLUME numeric left-cerebral-white-matter ROI size in mm^3 Range value: 30465, ..., 354965

LEFT_CHOROID_PLEXUS_SUVR numeric left-choroid-plexus Range value: 0.486, ..., 3.683

LEFT_CHOROID_PLEXUS_VOLUME numeric left-choroid-plexus ROI size in mm^3 Range value: 130, ..., 1916

LEFT_HIPPOCAMPUS_SUVR numeric left-hippocampus Range value: 0.516, ..., 2.73

LEFT_HIPPOCAMPUS_VOLUME numeric left-hippocampus ROI size in mm^3 Range value: 1314, ..., 6080

LEFT_INF_LAT_VENT_SUVR numeric left-inf-lat-vent Range value: 0.509, ..., 2.297

LEFT_INF_LAT_VENT_VOLUME numeric left-inf-lat-vent ROI size in mm^3 Range value: 53, ..., 11328

LEFT_LATERAL_VENTRICLE_SUVR numeric left-lateral-ventricle Range value: 0.214, ..., 1.454

LEFT_LATERAL_VENTRICLE_VOLUME numeric left-lateral-ventricle ROI size in mm^3 Range value: 3016, ..., 79775

LEFT_PALLIDUM_SUVR numeric left-pallidum Range value: 0.48, ..., 2.591

LEFT_PALLIDUM_VOLUME numeric left-pallidum ROI size in mm^3 Range value: 1090, ..., 3436

LEFT_PUTAMEN_SUVR numeric left-putamen Range value: 0.516, ..., 2.322

LEFT_PUTAMEN_VOLUME numeric left-putamen ROI size in mm^3 Range value: 1945, ..., 7466

LEFT_THALAMUS_PROPER_SUVR numeric left-thalamus-proper Range value: 0.496, ..., 2.057

LEFT_THALAMUS_PROPER_VOLUME numeric left-thalamus-proper ROI size in mm^3 Range value: 3953, ..., 10060

LEFT_VENTRALDC_SUVR numeric left-ventraldc Range value: 0.511, ..., 1.818

LEFT_VENTRALDC_VOLUME numeric left-ventraldc ROI size in mm^3 Range value: 2259, ..., 5708

LEFT_VESSEL_SUVR numeric left-vessel Range value: 0.53, ..., 2.342

LEFT_VESSEL_VOLUME numeric left-vessel ROI size in mm^3 Range value: 1, ..., 537

RIGHT_ACCUMBENS_AREA_SUVR numeric right-accumbens-area Range value: 0.411, ..., 2.439

RIGHT_ACCUMBENS_AREA_VOLUME numeric right-accumbens-area ROI size in mm^3 Range value: 114, ..., 815

RIGHT_AMYGDALA_SUVR numeric right-amygdala Range value: 0.445, ..., 3.397

RIGHT_AMYGDALA_VOLUME numeric right-amygdala ROI size in mm^3 Range value: 595, ..., 2453

RIGHT_CAUDATE_SUVR numeric right-caudate Range value: 0.423, ..., 1.721

RIGHT_CAUDATE_VOLUME numeric right-caudate ROI size in mm^3 Range value: 1717, ..., 6880

RIGHT_CEREBELLUM_CORTEX_SUVR numeric right-cerebellum-cortex Range value: 0.94, ..., 1.361

RIGHT_CEREBELLUM_CORTEX_VOLUME numeric right-cerebellum-cortex ROI size in mm^3 Range value: 34177, ..., 73171

RIGHT_CEREBELLUM_WHITE_MATTER_SUVR numeric right-cerebellum-white-matter Range value: 0.58, ..., 1.559

RIGHT_CEREBELLUM_WHITE_MATTER_VOLUME numeric right-cerebellum-white-matter ROI size in mm^3 Range value: 6779, ..., 22532

RIGHT_CEREBRAL_WHITE_MATTER_SUVR numeric right-cerebral-white-matter Range value: 0.514, ..., 2.462

RIGHT_CEREBRAL_WHITE_MATTER_VOLUME numeric right-cerebral-white-matter ROI size in mm^3 Range value: 129854, ..., 363072

RIGHT_CHOROID_PLEXUS_SUVR numeric right-choroid-plexus Range value: 0.506, ..., 3.537

RIGHT_CHOROID_PLEXUS_VOLUME numeric right-choroid-plexus ROI size in mm^3 Range value: 126, ..., 1912

RIGHT_HIPPOCAMPUS_SUVR numeric right-hippocampus Range value: 0.509, ..., 2.787

RIGHT_HIPPOCAMPUS_VOLUME numeric right-hippocampus ROI size in mm^3 Range value: 852, ..., 5775

RIGHT_INF_LAT_VENT_SUVR numeric right-inf-lat-vent Range value: 0.448, ..., 2.447

RIGHT_INF_LAT_VENT_VOLUME numeric right-inf-lat-vent ROI size in mm^3 Range value: 58, ..., 7228

RIGHT_LATERAL_VENTRICLE_SUVR numeric right-lateral-ventricle Range value: 0.221, ..., 1.502

RIGHT_LATERAL_VENTRICLE_VOLUME numeric right-lateral-ventricle ROI size in mm^3 Range value: 2442, ..., 71992

RIGHT_PALLIDUM_SUVR numeric right-pallidum Range value: 0.471, ..., 2.574

RIGHT_PALLIDUM_VOLUME numeric right-pallidum ROI size in mm^3 Range value: 688, ..., 3207

RIGHT_PUTAMEN_SUVR numeric right-putamen Range value: 0.487, ..., 2.56

RIGHT_PUTAMEN_VOLUME numeric right-putamen ROI size in mm^3 Range value: 2106, ..., 8026

RIGHT_THALAMUS_PROPER_SUVR numeric right-thalamus-proper Range value: 0.468, ..., 2.123

RIGHT_THALAMUS_PROPER_VOLUME numeric right-thalamus-proper ROI size in mm^3 Range value: 3522, ..., 10406

RIGHT_VENTRALDC_SUVR numeric right-ventraldc Range value: 0.51, ..., 1.847

RIGHT_VENTRALDC_VOLUME numeric right-ventraldc ROI size in mm^3 Range value: 1892, ..., 5674

RIGHT_VESSEL_SUVR numeric right-vessel Range value: 0.502, ..., 2.188

RIGHT_VESSEL_VOLUME numeric right-vessel ROI size in mm^3 Range value: 1, ..., 278

update_stamp POSIXct: YYY-MM-DD