Anders Dale Lab (UCSD) - Derived Volumes

Anders Dale Lab (UCSD) - Derived Volumes data. More information is available at https://adni.loni.usc.edu/data-samples/adni-data/

Usage

data(UCSDVOL)

Format

A data frame with 3128 observations and 25 variables.

Source

https://adni.loni.usc.edu/data-samples/adni-data/

Details

ORIGPROT character Original study protocol , with options: ADNI1

RID numeric Range value: 2, ..., 1435

SITEID numeric Range value: 1, ..., 127

VISCODE character , with options: sc, m06, m12, m24, m18, m36

USERDATE Date: YYY-MM-DD

EXAMDATE Examination Date Date: YYY-MM-DD

LONISID numeric LONISID Range value: 1945, ..., 28453

CDATE character Creation Date , with options: 12/09/2010

QCPASS numeric Range value: 0, ..., 1

BRAIN numeric Whole Brain. A summary measure of total brain parenchyma, including the Cerebral-Cortex, Cerebellum-Cortex, Thalamus-Proper, CaudatePutamen, Pallidum, Hippocampus, Amygdala, Accumbens-area, VentralDC, Cerebral-White-Matter, Cerebellum-White-Matter, and WM-hypointensities. As such, it does not include ventricles or other CSF spaces, and is not intended to be used as a measure of intracranial volume. Range value: 629016, ..., 1361820

EICV numeric Estimated Intracranial Volume, aka eTIV. To generate the mask, the baseline image is automatically segmented; all thus-defined brain and ventricular voxels are given the value 1 with all other voxels 0. This binary mask is then repeatedly smoothed with a Gaussian kernel to produce a simply connected uniform mask, covering all sulci, whose boundary tapers smoothly from 1 to 0 over the length of a few voxels. Ideally, the mask would end on the skull and include the brain stem down to where it starts to bend with the neck. The smoothing can be controlled fairly nicely to begin tapering at the skull, so that voxels with a mask value less than 1 can be considered outside the ICV and therefore ignored. Range value: 1059560, ..., 1951440

VENTRICLES numeric Ventricles Range value: 7552.46, ..., 168830

LHIPPOC numeric Left Hippocampus Range value: 1287.93, ..., 5514.94

RHIPPOC numeric Right Hippocampus Range value: 569.549, ..., 5750.4

LINFLATVEN numeric Left inferior lateral ventricle Range value: 232, ..., 8490.65

RINFLATVEN numeric Right inferior lateral ventricle Range value: 136.612, ..., 11628.5

LMIDTEMP numeric Left Middle Temporal Range value: 1.43278, ..., 2.951

RMIDTEMP numeric Right Middle Temporal Range value: 1.40978, ..., 3.1

LINFTEMP numeric Left Inferior Temporal Range value: 1.51919, ..., 3.1769

RINFTEMP numeric Right Inferior Temporal Range value: 1.55386, ..., 3.14129

LFUSIFORM numeric Left Fusiform Range value: 1.50766, ..., 2.928

RFUSIFORM numeric Right Fusiform Range value: 1.396, ..., 2.83648

LENTORHIN numeric Left Entorhinal Range value: 1.45156, ..., 4.08892

RENTORHIN numeric Right Entorhinal Range value: 1.4584, ..., 4.44725

update_stamp POSIXct: YYY-MM-DD

Author

adni-data@googlegroups.com