Longitudinal FreeSurfer (FreeSurfer Version 4.4): 2016-02-01

Longitudinal FreeSurfer (FreeSurfer Version 4.4): 2016-02-01 data. More information is available at https://adni.loni.usc.edu/data-samples/adni-data/

Usage

data(UCSFFSL)

Format

A data frame with 3570 observations and 376 variables.

Source

https://adni.loni.usc.edu/data-samples/adni-data/

Details

ORIGPROT character Original study protocol , with options: ADNI1

RID numeric Participant roster ID Range value: 3, ..., 1430

VISCODE character Visit code , with options: m06, m12, m24, sc, m18, m36, m48, f, bl, m60

VISCODE2 character Translated visit code , with options: m06, m12, m24, sc, m18, m36, m48, f, bl, m60

EXAMDATE The scan date for the image processed Date: YYY-MM-DD

VERSION Date: YYY-MM-DD

FLDSTRENG numeric The field strength. 1.5 or 3 Range value: 1.5, ..., 1.5

LONISID numeric The LONI study UID for the image processed Range value: 1948, ..., 44565

LONIUID numeric The LONI series UID for the image processed. Range value: 8723, ..., 146809

IMAGEUID numeric The ImageUID for the image processed. Range value: 31863, ..., 322396

RUNDATE Date: YYY-MM-DD

STATUS character , with options: complete, partial

BASETP1 numeric Whether or not the base image includes a timepoint 1 image. 0 = No, 1 = Yes

BASETP2 numeric Whether or not the base image includes a timepoint 2 image. 0 = No, 1 = Yes

BASETP3 numeric Whether or not the base image includes a timepoint 3 image. 0 = No, 1 = Yes

BASETP4 numeric Whether or not the base image includes a timepoint 4 image. 0 = No, 1 = Yes

BASETP5 numeric Whether or not the base image includes a timepoint 5 image. 0 = No, 1 = Yes

BASETP6 numeric Whether or not the base image includes a timepoint 6 image. 0 = No, 1 = Yes

BASETP7 numeric Whether or not the base image includes a timepoint 7 image. 0 = No, 1 = Yes

BASETP8 numeric Whether or not the base image includes a timepoint 8 image. 0 = No, 1 = Yes

OVERALLQC character An overall quality rating. Refer to additional QC variables for Partial rating. Pass/Fail/Partial

TEMPQC character QC rating of temporal lobe. Fail affects the following regions: LeftTemporalPole (ST60); RightTemporalPole (ST119); LeftFusiform (ST26); RightFusiform (ST85); LeftSuperiorTemporal (ST58); RightSuperiorTemporal (ST117); LeftInferiorTemporal (ST32); RightInferiorTemporal (ST91); LeftMiddleTemporal (ST40); RightMiddleTemporal (ST99) Pass/Fail

FRONTQC character QC rating of the frontal lobe. Fail affects the following regions: LeftFrontalPole (ST25); RightFrontalPole (ST84); LeftPrecentral (ST51); RightPrecentral (ST110); LeftSuperiorFrontal (ST56); RightSuperiorFrontal (ST115); LeftCaudalMiddleFrontal (ST15); RightCaudalMiddleFrontal (ST74); LeftRostralMiddleFrontal (ST55); RightRostralMiddleFrontal (ST114); LeftMedialOrbitofrontal (ST39); RightMedialOrbitofrontal (ST98) Pass/Fail

PARQC character QC rating of the parietal lobe. Fail affects the following regions: LeftPostcentral (ST49); RightPostcentral (ST108); LeftSuperiorParietal (ST57); RightSuperiorParietal (ST116); LeftSupraMarginal (ST59); RightSupraMarginal (ST118); LeftParacentral (ST43); RightParacentral (ST102); LeftInferiorParietal (ST31); RightInferiorParietal (ST90) Pass/Fail

INSULAQC character QC rating of the insula. Fail affects the following regions: LeftInsula (ST129); RightInsula (ST130) Pass/Fail

OCCQC character QC rating of the occipital lobe. Fail affects the following regions: LeftLingual (ST38); RightLingual (ST97); LeftLateralOccipital (ST35); RightLateralOccipital (ST94); LeftPericalcarine (ST48); RightPericalcarine (ST107); LeftCuneus (ST23); RightCuneus (ST82) Pass/Fail

BGQC character QC rating of the basal ganglia. Fail affects the following regions: LeftPutamen (ST53); RightPutamen (ST112); LeftCaudate (ST16); RightCaudate (ST75); LeftPallidum (ST42); RightPallidum (ST101) Pass/Fail

CWMQC character QC rating of the cerebral WM. Fail affects the following regions: LeftCerebralWM (ST20); RightCerebralWM (ST79) Pass/Fail

VENTQC character QC rating of the ventricles. Fail affects the following regions: LeftLateralVentricle (ST37); RightLateralVentricle (ST96); LeftInferiorLateralVentricle (ST30); RightInferiorLateralVentricle (ST89); ThirdVentricle (ST127); FourthVentricle (ST9); FifthVentrical (ST8); LeftChoroidPlexus (ST21); RightChoroidPlexus (ST80) , with options: Pass, Fail

ST100SV logical Volume (WM Parcellation) of RightNonWMHypoIntensities Boolean value: TRUE/ FALSE

ST101SV numeric Volume (WM Parcellation) of RightPallidum Range value: 309, ..., 2317

ST102CV numeric Volume (Cortical Parcellation) of RightParacentral Range value: 1899, ..., 5944

ST102SA numeric Surface Area of RightParacentral Range value: 981, ..., 2711

ST102TA numeric Cortical Thickness Average of RightParacentral Range value: 1.385, ..., 3.281

ST102TS numeric Cortical Thickness Standard Deviation of RightParacentral Range value: 0.363, ..., 1.458

ST103CV numeric Volume (Cortical Parcellation) of RightParahippocampal Range value: 592, ..., 3042

ST103SA numeric Surface Area of RightParahippocampal Range value: 360, ..., 1055

ST103TA numeric Cortical Thickness Average of RightParahippocampal Range value: 0.701, ..., 3.365

ST103TS numeric Cortical Thickness Standard Deviation of RightParahippocampal Range value: 0.456, ..., 1.922

ST104CV numeric Volume (Cortical Parcellation) of RightParsOpercularis Range value: 1523, ..., 7323

ST104SA numeric Surface Area of RightParsOpercularis Range value: 803, ..., 2847

ST104TA numeric Cortical Thickness Average of RightParsOpercularis Range value: 1.368, ..., 2.917

ST104TS numeric Cortical Thickness Standard Deviation of RightParsOpercularis Range value: 0.323, ..., 1.509

ST105CV numeric Volume (Cortical Parcellation) of RightParsOrbitalis Range value: 920, ..., 3744

ST105SA numeric Surface Area of RightParsOrbitalis Range value: 308, ..., 1613

ST105TA numeric Cortical Thickness Average of RightParsOrbitalis Range value: 1.508, ..., 3.416

ST105TS numeric Cortical Thickness Standard Deviation of RightParsOrbitalis Range value: 0.406, ..., 1.294

ST106CV numeric Volume (Cortical Parcellation) of RightParsTriangularis Range value: 1437, ..., 5845

ST106SA numeric Surface Area of RightParsTriangularis Range value: 763, ..., 2555

ST106TA numeric Cortical Thickness Average of RightParsTriangularis Range value: 1.307, ..., 3.247

ST106TS numeric Cortical Thickness Standard Deviation of RightParsTriangularis Range value: 0.32, ..., 1.344

ST107CV numeric Volume (Cortical Parcellation) of RightPericalcarine Range value: 695, ..., 4131

ST107SA numeric Surface Area of RightPericalcarine Range value: 523, ..., 2430

ST107TA numeric Cortical Thickness Average of RightPericalcarine Range value: 1.096, ..., 2.495

ST107TS numeric Cortical Thickness Standard Deviation of RightPericalcarine Range value: 0.186, ..., 1.361

ST108CV numeric Volume (Cortical Parcellation) of RightPostcentral Range value: 4756, ..., 12487

ST108SA numeric Surface Area of RightPostcentral Range value: 2215, ..., 6691

ST108TA numeric Cortical Thickness Average of RightPostcentral Range value: 1.31, ..., 2.753

ST108TS numeric Cortical Thickness Standard Deviation of RightPostcentral Range value: 0.398, ..., 1.224

ST109CV numeric Volume (Cortical Parcellation) of RightPosteriorCingulate Range value: 1117, ..., 4607

ST109SA numeric Surface Area of RightPosteriorCingulate Range value: 730, ..., 2130

ST109TA numeric Cortical Thickness Average of RightPosteriorCingulate Range value: 0.989, ..., 3.3

ST109TS numeric Cortical Thickness Standard Deviation of RightPosteriorCingulate Range value: 0.423, ..., 1.441

ST10CV numeric Volume (Cortical Parcellation) of Icv Range value: 1153430, ..., 2110290

ST110CV numeric Volume (Cortical Parcellation) of RightPrecentral Range value: 6097, ..., 19531

ST110SA numeric Surface Area of RightPrecentral Range value: 3247, ..., 8900

ST110TA numeric Cortical Thickness Average of RightPrecentral Range value: 1.352, ..., 3.717

ST110TS numeric Cortical Thickness Standard Deviation of RightPrecentral Range value: 0.391, ..., 1.565

ST111CV numeric Volume (Cortical Parcellation) of RightPrecuneus Range value: 4104, ..., 12036

ST111SA numeric Surface Area of RightPrecuneus Range value: 2517, ..., 5668

ST111TA numeric Cortical Thickness Average of RightPrecuneus Range value: 1.352, ..., 2.731

ST111TS numeric Cortical Thickness Standard Deviation of RightPrecuneus Range value: 0.373, ..., 1.308

ST112SV numeric Volume (WM Parcellation) of RightPutamen Range value: 1890, ..., 7198

ST113CV numeric Volume (Cortical Parcellation) of RightRostralAnteriorCingulate Range value: 0, ..., 2879

ST113SA numeric Surface Area of RightRostralAnteriorCingulate Range value: 243, ..., 1265

ST113TA numeric Cortical Thickness Average of RightRostralAnteriorCingulate Range value: 0, ..., 3.912

ST113TS numeric Cortical Thickness Standard Deviation of RightRostralAnteriorCingulate Range value: 0, ..., 1.324

ST114CV numeric Volume (Cortical Parcellation) of RightRostralMiddleFrontal Range value: 6116, ..., 20637

ST114SA numeric Surface Area of RightRostralMiddleFrontal Range value: 3287, ..., 9365

ST114TA numeric Cortical Thickness Average of RightRostralMiddleFrontal Range value: 1.376, ..., 2.997

ST114TS numeric Cortical Thickness Standard Deviation of RightRostralMiddleFrontal Range value: 0.37, ..., 1.738

ST115CV numeric Volume (Cortical Parcellation) of RightSuperiorFrontal Range value: 7447, ..., 28356

ST115SA numeric Surface Area of RightSuperiorFrontal Range value: 3528, ..., 10319

ST115TA numeric Cortical Thickness Average of RightSuperiorFrontal Range value: 1.527, ..., 3.483

ST115TS numeric Cortical Thickness Standard Deviation of RightSuperiorFrontal Range value: 0.425, ..., 1.763

ST116CV numeric Volume (Cortical Parcellation) of RightSuperiorParietal Range value: 6015, ..., 16887

ST116SA numeric Surface Area of RightSuperiorParietal Range value: 2878, ..., 7581

ST116TA numeric Cortical Thickness Average of RightSuperiorParietal Range value: 1.287, ..., 3.247

ST116TS numeric Cortical Thickness Standard Deviation of RightSuperiorParietal Range value: 0.307, ..., 1.326

ST117CV numeric Volume (Cortical Parcellation) of RightSuperiorTemporal Range value: 4362, ..., 13160

ST117SA numeric Surface Area of RightSuperiorTemporal Range value: 2449, ..., 6647

ST117TA numeric Cortical Thickness Average of RightSuperiorTemporal Range value: 1.472, ..., 3.045

ST117TS numeric Cortical Thickness Standard Deviation of RightSuperiorTemporal Range value: 0.473, ..., 1.416

ST118CV numeric Volume (Cortical Parcellation) of RightSupramarginal Range value: 4188, ..., 14020

ST118SA numeric Surface Area of RightSupramarginal Range value: 2287, ..., 5908

ST118TA numeric Cortical Thickness Average of RightSupramarginal Range value: 1.474, ..., 2.849

ST118TS numeric Cortical Thickness Standard Deviation of RightSupramarginal Range value: 0.419, ..., 1.296

ST119CV numeric Volume (Cortical Parcellation) of RightTemporalPole Range value: 660, ..., 3150

ST119SA numeric Surface Area of RightTemporalPole Range value: 148, ..., 965

ST119TA numeric Cortical Thickness Average of RightTemporalPole Range value: 1.288, ..., 4.646

ST119TS numeric Cortical Thickness Standard Deviation of RightTemporalPole Range value: 0.289, ..., 1.423

ST11SV numeric Volume (WM Parcellation) of LeftAccumbensArea Range value: 12, ..., 765

ST120SV numeric Volume (WM Parcellation) of RightThalamus Range value: 3725, ..., 10172

ST121CV numeric Volume (Cortical Parcellation) of RightTransverseTemporal Range value: 329, ..., 1617

ST121SA numeric Surface Area of RightTransverseTemporal Range value: 173, ..., 639

ST121TA numeric Cortical Thickness Average of RightTransverseTemporal Range value: 1.064, ..., 4.192

ST121TS numeric Cortical Thickness Standard Deviation of RightTransverseTemporal Range value: 0.163, ..., 1.292

ST122SV logical Volume (WM Parcellation) of RightUndetermined Boolean value: TRUE/ FALSE

ST123CV numeric Volume (Cortical Parcellation) of RightUnknown Range value: 2302, ..., 9554

ST123SA numeric Surface Area of RightUnknown Range value: 3802, ..., 8432

ST123TA numeric Cortical Thickness Average of RightUnknown Range value: 0.395, ..., 1.513

ST123TS numeric Cortical Thickness Standard Deviation of RightUnknown Range value: 0.709, ..., 1.772

ST124SV numeric Volume (WM Parcellation) of RightVentralDC Range value: 2183, ..., 5733

ST125SV numeric Volume (WM Parcellation) of RightVessel Range value: 2, ..., 521

ST126SV logical Volume (WM Parcellation) of RightWMHypoIntensities Boolean value: TRUE/ FALSE

ST127SV numeric Volume (WM Parcellation) of ThirdVentricle Range value: 625, ..., 5173

ST128SV numeric Volume (WM Parcellation) of WMHypoIntensities Range value: 1069, ..., 62310

ST129CV numeric Volume (Cortical Parcellation) of LeftInsula Range value: 2719, ..., 7888

ST129SA numeric Surface Area of LeftInsula Range value: 1499, ..., 3014

ST129TA numeric Cortical Thickness Average of LeftInsula Range value: 1.502, ..., 3.333

ST129TS numeric Cortical Thickness Standard Deviation of LeftInsula Range value: 0.455, ..., 1.466

ST12SV numeric Volume (WM Parcellation) of LeftAmygdala Range value: 343, ..., 1613

ST130CV numeric Volume (Cortical Parcellation) of RightInsula Range value: 2858, ..., 7632

ST130SA numeric Surface Area of RightInsula Range value: 1383, ..., 3069

ST130TA numeric Cortical Thickness Average of RightInsula Range value: 1.456, ..., 3.575

ST130TS numeric Cortical Thickness Standard Deviation of RightInsula Range value: 0.406, ..., 1.357

ST13CV numeric Volume (Cortical Parcellation) of LeftBankssts Range value: 874, ..., 6014

ST13SA numeric Surface Area of LeftBankssts Range value: 509, ..., 2650

ST13TA numeric Cortical Thickness Average of LeftBankssts Range value: 1.29, ..., 3.155

ST13TS numeric Cortical Thickness Standard Deviation of LeftBankssts Range value: 0.293, ..., 1.341

ST14CV numeric Volume (Cortical Parcellation) of LeftCaudalAnteriorCingulate Range value: 478, ..., 3148

ST14SA numeric Surface Area of LeftCaudalAnteriorCingulate Range value: 349, ..., 1402

ST14TA numeric Cortical Thickness Average of LeftCaudalAnteriorCingulate Range value: 1.448, ..., 3.746

ST14TS numeric Cortical Thickness Standard Deviation of LeftCaudalAnteriorCingulate Range value: 0.385, ..., 1.489

ST15CV numeric Volume (Cortical Parcellation) of LeftCaudalMiddleFrontal Range value: 2296, ..., 9379

ST15SA numeric Surface Area of LeftCaudalMiddleFrontal Range value: 1202, ..., 4094

ST15TA numeric Cortical Thickness Average of LeftCaudalMiddleFrontal Range value: 1.267, ..., 3.345

ST15TS numeric Cortical Thickness Standard Deviation of LeftCaudalMiddleFrontal Range value: 0.259, ..., 1.444

ST16SV numeric Volume (WM Parcellation) of LeftCaudate Range value: 1359, ..., 7198

ST17SV numeric Volume (WM Parcellation) of LeftCerebellumCortex Range value: 20849, ..., 62978

ST18SV numeric Volume (WM Parcellation) of LeftCerebellumWM Range value: 10791, ..., 51951

ST19SV numeric Volume (WM Parcellation) of LeftCerebralCortex Range value: 96144, ..., 236449

ST1SV numeric Volume (WM Parcellation) of Brainstem Range value: 13778, ..., 33268

ST20SV numeric Volume (WM Parcellation) of LeftCerebralWM Range value: 161817, ..., 382583

ST21SV numeric Volume (WM Parcellation) of LeftChoroidPlexus Range value: 1024, ..., 5720

ST22CV logical Volume (Cortical Parcellation) of LeftCorpusCallosum Boolean value: TRUE/ FALSE

ST22SA logical Surface Area of LeftCorpusCallosum Boolean value: TRUE/ FALSE

ST22TA logical Cortical Thickness Average of LeftCorpusCallosum Boolean value: TRUE/ FALSE

ST22TS logical Cortical Thickness Standard Deviation of LeftCorpusCallosum Boolean value: TRUE/ FALSE

ST23CV numeric Volume (Cortical Parcellation) of LeftCuneus Range value: 818, ..., 4761

ST23SA numeric Surface Area of LeftCuneus Range value: 495, ..., 2342

ST23TA numeric Cortical Thickness Average of LeftCuneus Range value: 1.049, ..., 2.669

ST23TS numeric Cortical Thickness Standard Deviation of LeftCuneus Range value: 0.284, ..., 1.321

ST24CV numeric Volume (Cortical Parcellation) of LeftEntorhinal Range value: 488, ..., 3753

ST24SA numeric Surface Area of LeftEntorhinal Range value: 138, ..., 843

ST24TA numeric Cortical Thickness Average of LeftEntorhinal Range value: 1.348, ..., 4.287

ST24TS numeric Cortical Thickness Standard Deviation of LeftEntorhinal Range value: 0.461, ..., 1.437

ST25CV numeric Volume (Cortical Parcellation) of LeftFrontalPole Range value: 156, ..., 1172

ST25SA numeric Surface Area of LeftFrontalPole Range value: 53, ..., 337

ST25TA numeric Cortical Thickness Average of LeftFrontalPole Range value: 1.254, ..., 3.785

ST25TS numeric Cortical Thickness Standard Deviation of LeftFrontalPole Range value: 0.227, ..., 1.081

ST26CV numeric Volume (Cortical Parcellation) of LeftFusiform Range value: 3402, ..., 12965

ST26SA numeric Surface Area of LeftFusiform Range value: 1632, ..., 5521

ST26TA numeric Cortical Thickness Average of LeftFusiform Range value: 1.484, ..., 3.081

ST26TS numeric Cortical Thickness Standard Deviation of LeftFusiform Range value: 0.434, ..., 1.188

ST27SA numeric Surface Area of LeftHemisphere Range value: 53766, ..., 122415

ST28CV logical Volume (Cortical Parcellation) of LeftHemisphereWM Boolean value: TRUE/ FALSE

ST29SV numeric Volume (WM Parcellation) of LeftHippocampus Range value: 1345, ..., 5330

ST2SV numeric Volume (WM Parcellation) of CorpusCallosumAnterior Range value: 29, ..., 2343

ST30SV numeric Volume (WM Parcellation) of LeftInferiorLateralVentricle Range value: 23, ..., 7994

ST31CV numeric Volume (Cortical Parcellation) of LeftInferiorParietal Range value: 4309, ..., 17500

ST31SA numeric Surface Area of LeftInferiorParietal Range value: 1936, ..., 7806

ST31TA numeric Cortical Thickness Average of LeftInferiorParietal Range value: 1.347, ..., 3.117

ST31TS numeric Cortical Thickness Standard Deviation of LeftInferiorParietal Range value: 0.296, ..., 1.343

ST32CV numeric Volume (Cortical Parcellation) of LeftInferiorTemporal Range value: 3918, ..., 14817

ST32SA numeric Surface Area of LeftInferiorTemporal Range value: 1479, ..., 5635

ST32TA numeric Cortical Thickness Average of LeftInferiorTemporal Range value: 1.556, ..., 3.255

ST32TS numeric Cortical Thickness Standard Deviation of LeftInferiorTemporal Range value: 0.453, ..., 1.377

ST33SV logical Volume (WM Parcellation) of LeftInterior Boolean value: TRUE/ FALSE

ST34CV numeric Volume (Cortical Parcellation) of LeftIsthmusCingulate Range value: 1111, ..., 3432

ST34SA numeric Surface Area of LeftIsthmusCingulate Range value: 437, ..., 1625

ST34TA numeric Cortical Thickness Average of LeftIsthmusCingulate Range value: 1.426, ..., 3.291

ST34TS numeric Cortical Thickness Standard Deviation of LeftIsthmusCingulate Range value: 0.417, ..., 1.914

ST35CV numeric Volume (Cortical Parcellation) of LeftLateralOccipital Range value: 5102, ..., 16149

ST35SA numeric Surface Area of LeftLateralOccipital Range value: 2539, ..., 7768

ST35TA numeric Cortical Thickness Average of LeftLateralOccipital Range value: 1.291, ..., 2.782

ST35TS numeric Cortical Thickness Standard Deviation of LeftLateralOccipital Range value: 0.281, ..., 1.334

ST36CV numeric Volume (Cortical Parcellation) of LeftLateralOrbitofrontal Range value: 2866, ..., 8842

ST36SA numeric Surface Area of LeftLateralOrbitofrontal Range value: 1271, ..., 3795

ST36TA numeric Cortical Thickness Average of LeftLateralOrbitofrontal Range value: 1.637, ..., 3.467

ST36TS numeric Cortical Thickness Standard Deviation of LeftLateralOrbitofrontal Range value: 0.453, ..., 1.259

ST37SV numeric Volume (WM Parcellation) of LeftLateralVentricle Range value: 3052, ..., 88479

ST38CV numeric Volume (Cortical Parcellation) of LeftLingual Range value: 2357, ..., 9036

ST38SA numeric Surface Area of LeftLingual Range value: 1567, ..., 4962

ST38TA numeric Cortical Thickness Average of LeftLingual Range value: 1.363, ..., 3.074

ST38TS numeric Cortical Thickness Standard Deviation of LeftLingual Range value: 0.373, ..., 1.134

ST39CV numeric Volume (Cortical Parcellation) of LeftMedialOrbitofrontal Range value: 1982, ..., 6219

ST39SA numeric Surface Area of LeftMedialOrbitofrontal Range value: 684, ..., 2750

ST39TA numeric Cortical Thickness Average of LeftMedialOrbitofrontal Range value: 1.408, ..., 3.12

ST39TS numeric Cortical Thickness Standard Deviation of LeftMedialOrbitofrontal Range value: 0.382, ..., 1.299

ST3SV numeric Volume (WM Parcellation) of CorpusCallosumCentral Range value: 66, ..., 3700

ST40CV numeric Volume (Cortical Parcellation) of LeftMiddleTemporal Range value: 3444, ..., 14083

ST40SA numeric Surface Area of LeftMiddleTemporal Range value: 1571, ..., 5461

ST40TA numeric Cortical Thickness Average of LeftMiddleTemporal Range value: 1.451, ..., 3.131

ST40TS numeric Cortical Thickness Standard Deviation of LeftMiddleTemporal Range value: 0.357, ..., 1.317

ST41SV logical Volume (WM Parcellation) of LeftNonWMHypoIntensities Boolean value: TRUE/ FALSE

ST42SV numeric Volume (WM Parcellation) of LeftPallidum Range value: 434, ..., 2868

ST43CV numeric Volume (Cortical Parcellation) of LeftParacentral Range value: 1262, ..., 5636

ST43SA numeric Surface Area of LeftParacentral Range value: 839, ..., 2255

ST43TA numeric Cortical Thickness Average of LeftParacentral Range value: 1.303, ..., 3.429

ST43TS numeric Cortical Thickness Standard Deviation of LeftParacentral Range value: 0.335, ..., 1.374

ST44CV numeric Volume (Cortical Parcellation) of LeftParahippocampal Range value: 806, ..., 3043

ST44SA numeric Surface Area of LeftParahippocampal Range value: 355, ..., 1259

ST44TA numeric Cortical Thickness Average of LeftParahippocampal Range value: 1.295, ..., 3.74

ST44TS numeric Cortical Thickness Standard Deviation of LeftParahippocampal Range value: 0.408, ..., 1.492

ST45CV numeric Volume (Cortical Parcellation) of LeftParsOpercularis Range value: 1739, ..., 7951

ST45SA numeric Surface Area of LeftParsOpercularis Range value: 775, ..., 3693

ST45TA numeric Cortical Thickness Average of LeftParsOpercularis Range value: 1.351, ..., 2.982

ST45TS numeric Cortical Thickness Standard Deviation of LeftParsOpercularis Range value: 0.254, ..., 1.295

ST46CV numeric Volume (Cortical Parcellation) of LeftParsOrbitalis Range value: 754, ..., 3013

ST46SA numeric Surface Area of LeftParsOrbitalis Range value: 215, ..., 1154

ST46TA numeric Cortical Thickness Average of LeftParsOrbitalis Range value: 1.571, ..., 3.826

ST46TS numeric Cortical Thickness Standard Deviation of LeftParsOrbitalis Range value: 0.371, ..., 1.3

ST47CV numeric Volume (Cortical Parcellation) of LeftParsTriangularis Range value: 1071, ..., 5452

ST47SA numeric Surface Area of LeftParsTriangularis Range value: 540, ..., 2234

ST47TA numeric Cortical Thickness Average of LeftParsTriangularis Range value: 1.346, ..., 3.477

ST47TS numeric Cortical Thickness Standard Deviation of LeftParsTriangularis Range value: 0.267, ..., 1.213

ST48CV numeric Volume (Cortical Parcellation) of LeftPericalcarine Range value: 712, ..., 3906

ST48SA numeric Surface Area of LeftPericalcarine Range value: 631, ..., 2015

ST48TA numeric Cortical Thickness Average of LeftPericalcarine Range value: 1.04, ..., 2.869

ST48TS numeric Cortical Thickness Standard Deviation of LeftPericalcarine Range value: 0.192, ..., 1.198

ST49CV numeric Volume (Cortical Parcellation) of LeftPostcentral Range value: 5033, ..., 14648

ST49SA numeric Surface Area of LeftPostcentral Range value: 2527, ..., 8112

ST49TA numeric Cortical Thickness Average of LeftPostcentral Range value: 1.268, ..., 2.814

ST49TS numeric Cortical Thickness Standard Deviation of LeftPostcentral Range value: 0.318, ..., 1.28

ST4SV numeric Volume (WM Parcellation) of CorpusCallosumMidAnterior Range value: 101, ..., 3804

ST50CV numeric Volume (Cortical Parcellation) of LeftPosteriorCingulate Range value: 1566, ..., 4828

ST50SA numeric Surface Area of LeftPosteriorCingulate Range value: 745, ..., 2095

ST50TA numeric Cortical Thickness Average of LeftPosteriorCingulate Range value: 1.414, ..., 3.233

ST50TS numeric Cortical Thickness Standard Deviation of LeftPosteriorCingulate Range value: 0.407, ..., 1.652

ST51CV numeric Volume (Cortical Parcellation) of LeftPrecentral Range value: 5693, ..., 19234

ST51SA numeric Surface Area of LeftPrecentral Range value: 2806, ..., 7554

ST51TA numeric Cortical Thickness Average of LeftPrecentral Range value: 1.373, ..., 3.171

ST51TS numeric Cortical Thickness Standard Deviation of LeftPrecentral Range value: 0.314, ..., 1.391

ST52CV numeric Volume (Cortical Parcellation) of LeftPrecuneus Range value: 4166, ..., 12683

ST52SA numeric Surface Area of LeftPrecuneus Range value: 2350, ..., 5436

ST52TA numeric Cortical Thickness Average of LeftPrecuneus Range value: 1.344, ..., 2.658

ST52TS numeric Cortical Thickness Standard Deviation of LeftPrecuneus Range value: 0.325, ..., 1.299

ST53SV numeric Volume (WM Parcellation) of LeftPutamen Range value: 1200, ..., 7116

ST54CV numeric Volume (Cortical Parcellation) of LeftRostralAnteriorCingulate Range value: 786, ..., 3771

ST54SA numeric Surface Area of LeftRostralAnteriorCingulate Range value: 302, ..., 1361

ST54TA numeric Cortical Thickness Average of LeftRostralAnteriorCingulate Range value: 1.459, ..., 3.901

ST54TS numeric Cortical Thickness Standard Deviation of LeftRostralAnteriorCingulate Range value: 0.374, ..., 1.391

ST55CV numeric Volume (Cortical Parcellation) of LeftRostralMiddleFrontal Range value: 5192, ..., 17414

ST55SA numeric Surface Area of LeftRostralMiddleFrontal Range value: 2623, ..., 8667

ST55TA numeric Cortical Thickness Average of LeftRostralMiddleFrontal Range value: 1.242, ..., 3.207

ST55TS numeric Cortical Thickness Standard Deviation of LeftRostralMiddleFrontal Range value: 0.282, ..., 1.2

ST56CV numeric Volume (Cortical Parcellation) of LeftSuperiorFrontal Range value: 7074, ..., 28034

ST56SA numeric Surface Area of LeftSuperiorFrontal Range value: 4189, ..., 10329

ST56TA numeric Cortical Thickness Average of LeftSuperiorFrontal Range value: 1.434, ..., 3.346

ST56TS numeric Cortical Thickness Standard Deviation of LeftSuperiorFrontal Range value: 0.351, ..., 1.327

ST57CV numeric Volume (Cortical Parcellation) of LeftSuperiorParietal Range value: 6009, ..., 17166

ST57SA numeric Surface Area of LeftSuperiorParietal Range value: 2802, ..., 7685

ST57TA numeric Cortical Thickness Average of LeftSuperiorParietal Range value: 1.189, ..., 3.096

ST57TS numeric Cortical Thickness Standard Deviation of LeftSuperiorParietal Range value: 0.256, ..., 1.355

ST58CV numeric Volume (Cortical Parcellation) of LeftSuperiorTemporal Range value: 4132, ..., 18177

ST58SA numeric Surface Area of LeftSuperiorTemporal Range value: 2406, ..., 6198

ST58TA numeric Cortical Thickness Average of LeftSuperiorTemporal Range value: 1.391, ..., 2.954

ST58TS numeric Cortical Thickness Standard Deviation of LeftSuperiorTemporal Range value: 0.377, ..., 1.355

ST59CV numeric Volume (Cortical Parcellation) of LeftSupramarginal Range value: 4481, ..., 13506

ST59SA numeric Surface Area of LeftSupramarginal Range value: 2020, ..., 6240

ST59TA numeric Cortical Thickness Average of LeftSupramarginal Range value: 1.38, ..., 3.176

ST59TS numeric Cortical Thickness Standard Deviation of LeftSupramarginal Range value: 0.359, ..., 1.344

ST5SV numeric Volume (WM Parcellation) of CorpusCallosumMidPosterior Range value: 43, ..., 2648

ST60CV numeric Volume (Cortical Parcellation) of LeftTemporalPole Range value: 491, ..., 3722

ST60SA numeric Surface Area of LeftTemporalPole Range value: 146, ..., 999

ST60TA numeric Cortical Thickness Average of LeftTemporalPole Range value: 1.29, ..., 4.396

ST60TS numeric Cortical Thickness Standard Deviation of LeftTemporalPole Range value: 0.269, ..., 1.278

ST61SV numeric Volume (WM Parcellation) of LeftThalamus Range value: 4053, ..., 10823

ST62CV numeric Volume (Cortical Parcellation) of LeftTransverseTemporal Range value: 352, ..., 1810

ST62SA numeric Surface Area of LeftTransverseTemporal Range value: 223, ..., 1042

ST62TA numeric Cortical Thickness Average of LeftTransverseTemporal Range value: 1.133, ..., 3.536

ST62TS numeric Cortical Thickness Standard Deviation of LeftTransverseTemporal Range value: 0.2, ..., 1.608

ST63SV logical Volume (WM Parcellation) of LeftUndetermined Boolean value: TRUE/ FALSE

ST64CV numeric Volume (Cortical Parcellation) of LeftUnknown Range value: 3389, ..., 9489

ST64SA numeric Surface Area of LeftUnknown Range value: 4207, ..., 8682

ST64TA numeric Cortical Thickness Average of LeftUnknown Range value: 0.59, ..., 1.49

ST64TS numeric Cortical Thickness Standard Deviation of LeftUnknown Range value: 0.857, ..., 1.72

ST65SV numeric Volume (WM Parcellation) of LeftVentralDC Range value: 2086, ..., 7611

ST66SV numeric Volume (WM Parcellation) of LeftVessel Range value: 2, ..., 490

ST67SV logical Volume (WM Parcellation) of LeftWMHypoIntensities Boolean value: TRUE/ FALSE

ST68SV numeric Volume (WM Parcellation) of NonWMHypoIntensities Range value: 1, ..., 1231

ST69SV numeric Volume (WM Parcellation) of OpticChiasm Range value: 78, ..., 687

ST6SV numeric Volume (WM Parcellation) of CorpusCallosumPosterior Range value: 54, ..., 1484

ST70SV numeric Volume (WM Parcellation) of RightAccumbensArea Range value: 41, ..., 776

ST71SV numeric Volume (WM Parcellation) of RightAmygdala Range value: 161, ..., 1971

ST72CV numeric Volume (Cortical Parcellation) of RightBankssts Range value: 759, ..., 4078

ST72SA numeric Surface Area of RightBankssts Range value: 512, ..., 1966

ST72TA numeric Cortical Thickness Average of RightBankssts Range value: 1.369, ..., 2.939

ST72TS numeric Cortical Thickness Standard Deviation of RightBankssts Range value: 0.265, ..., 1.338

ST73CV numeric Volume (Cortical Parcellation) of RightCaudalAnteriorCingulate Range value: 3, ..., 3553

ST73SA numeric Surface Area of RightCaudalAnteriorCingulate Range value: 402, ..., 1590

ST73TA numeric Cortical Thickness Average of RightCaudalAnteriorCingulate Range value: 0.002, ..., 3.762

ST73TS numeric Cortical Thickness Standard Deviation of RightCaudalAnteriorCingulate Range value: 0.067, ..., 1.401

ST74CV numeric Volume (Cortical Parcellation) of RightCaudalMiddleFrontal Range value: 1966, ..., 10695

ST74SA numeric Surface Area of RightCaudalMiddleFrontal Range value: 1086, ..., 4684

ST74TA numeric Cortical Thickness Average of RightCaudalMiddleFrontal Range value: 1.321, ..., 3.5

ST74TS numeric Cortical Thickness Standard Deviation of RightCaudalMiddleFrontal Range value: 0.32, ..., 1.587

ST75SV numeric Volume (WM Parcellation) of RightCaudate Range value: 1131, ..., 6378

ST76SV numeric Volume (WM Parcellation) of RightCerebellumCortex Range value: 22794, ..., 51265

ST77SV numeric Volume (WM Parcellation) of RightCerebellumWM Range value: 10755, ..., 53115

ST78SV numeric Volume (WM Parcellation) of RightCerebralCortex Range value: 103732, ..., 254274

ST79SV numeric Volume (WM Parcellation) of RightCerebralWM Range value: 156938, ..., 388359

ST7SV numeric Volume (WM Parcellation) of Csf Range value: 662, ..., 4346

ST80SV numeric Volume (WM Parcellation) of RightChoroidPlexus Range value: 1225, ..., 5050

ST81CV logical Volume (Cortical Parcellation) of RightCorpusCallosum Boolean value: TRUE/ FALSE

ST81SA logical Surface Area of RightCorpusCallosum Boolean value: TRUE/ FALSE

ST81TA logical Cortical Thickness Average of RightCorpusCallosum Boolean value: TRUE/ FALSE

ST81TS logical Cortical Thickness Standard Deviation of RightCorpusCallosum Boolean value: TRUE/ FALSE

ST82CV numeric Volume (Cortical Parcellation) of RightCuneus Range value: 734, ..., 5266

ST82SA numeric Surface Area of RightCuneus Range value: 491, ..., 2416

ST82TA numeric Cortical Thickness Average of RightCuneus Range value: 1.277, ..., 2.817

ST82TS numeric Cortical Thickness Standard Deviation of RightCuneus Range value: 0.264, ..., 1.402

ST83CV numeric Volume (Cortical Parcellation) of RightEntorhinal Range value: 238, ..., 3382

ST83SA numeric Surface Area of RightEntorhinal Range value: 135, ..., 850

ST83TA numeric Cortical Thickness Average of RightEntorhinal Range value: 1.047, ..., 4.542

ST83TS numeric Cortical Thickness Standard Deviation of RightEntorhinal Range value: 0.431, ..., 1.881

ST84CV numeric Volume (Cortical Parcellation) of RightFrontalPole Range value: 198, ..., 1443

ST84SA numeric Surface Area of RightFrontalPole Range value: 64, ..., 499

ST84TA numeric Cortical Thickness Average of RightFrontalPole Range value: 1.407, ..., 3.944

ST84TS numeric Cortical Thickness Standard Deviation of RightFrontalPole Range value: 0.228, ..., 1.244

ST85CV numeric Volume (Cortical Parcellation) of RightFusiform Range value: 3281, ..., 13506

ST85SA numeric Surface Area of RightFusiform Range value: 1717, ..., 5368

ST85TA numeric Cortical Thickness Average of RightFusiform Range value: 1.494, ..., 3.349

ST85TS numeric Cortical Thickness Standard Deviation of RightFusiform Range value: 0.434, ..., 1.335

ST86SA numeric Surface Area of RightHemisphere Range value: 61188.2, ..., 127391

ST87CV logical Volume (Cortical Parcellation) of RightHemisphereWM Boolean value: TRUE/ FALSE

ST88SV numeric Volume (WM Parcellation) of RightHippocampus Range value: 636, ..., 5338

ST89SV numeric Volume (WM Parcellation) of RightInferiorLateralVentricle Range value: 2, ..., 9831

ST8SV numeric Volume (WM Parcellation) of FifthVentricle Range value: 1, ..., 206

ST90CV numeric Volume (Cortical Parcellation) of RightInferiorParietal Range value: 5859, ..., 18632

ST90SA numeric Surface Area of RightInferiorParietal Range value: 2665, ..., 8571

ST90TA numeric Cortical Thickness Average of RightInferiorParietal Range value: 1.367, ..., 2.749

ST90TS numeric Cortical Thickness Standard Deviation of RightInferiorParietal Range value: 0.329, ..., 1.335

ST91CV numeric Volume (Cortical Parcellation) of RightInferiorTemporal Range value: 3651, ..., 15180

ST91SA numeric Surface Area of RightInferiorTemporal Range value: 1393, ..., 5767

ST91TA numeric Cortical Thickness Average of RightInferiorTemporal Range value: 1.611, ..., 3.37

ST91TS numeric Cortical Thickness Standard Deviation of RightInferiorTemporal Range value: 0.488, ..., 1.302

ST92SV logical Volume (WM Parcellation) of RightInterior Boolean value: TRUE/ FALSE

ST93CV numeric Volume (Cortical Parcellation) of RightIsthmusCingulate Range value: 968, ..., 3445

ST93SA numeric Surface Area of RightIsthmusCingulate Range value: 453, ..., 1425

ST93TA numeric Cortical Thickness Average of RightIsthmusCingulate Range value: 1.178, ..., 3.187

ST93TS numeric Cortical Thickness Standard Deviation of RightIsthmusCingulate Range value: 0.405, ..., 2.018

ST94CV numeric Volume (Cortical Parcellation) of RightLateralOccipital Range value: 4733, ..., 16735

ST94SA numeric Surface Area of RightLateralOccipital Range value: 2594, ..., 7463

ST94TA numeric Cortical Thickness Average of RightLateralOccipital Range value: 1.356, ..., 2.938

ST94TS numeric Cortical Thickness Standard Deviation of RightLateralOccipital Range value: 0.346, ..., 1.268

ST95CV numeric Volume (Cortical Parcellation) of RightLateralOrbitofrontal Range value: 2656, ..., 9057

ST95SA numeric Surface Area of RightLateralOrbitofrontal Range value: 1554, ..., 3834

ST95TA numeric Cortical Thickness Average of RightLateralOrbitofrontal Range value: 1.652, ..., 3.339

ST95TS numeric Cortical Thickness Standard Deviation of RightLateralOrbitofrontal Range value: 0.494, ..., 1.981

ST96SV numeric Volume (WM Parcellation) of RightLateralVentricle Range value: 2216, ..., 72583

ST97CV numeric Volume (Cortical Parcellation) of RightLingual Range value: 2937, ..., 9109

ST97SA numeric Surface Area of RightLingual Range value: 1644, ..., 4516

ST97TA numeric Cortical Thickness Average of RightLingual Range value: 1.244, ..., 2.824

ST97TS numeric Cortical Thickness Standard Deviation of RightLingual Range value: 0.387, ..., 1.448

ST98CV numeric Volume (Cortical Parcellation) of RightMedialOrbitofrontal Range value: 240, ..., 7096

ST98SA numeric Surface Area of RightMedialOrbitofrontal Range value: 940, ..., 2899

ST98TA numeric Cortical Thickness Average of RightMedialOrbitofrontal Range value: 0.234, ..., 2.992

ST98TS numeric Cortical Thickness Standard Deviation of RightMedialOrbitofrontal Range value: 0.395, ..., 1.209

ST99CV numeric Volume (Cortical Parcellation) of RightMiddleTemporal Range value: 4496, ..., 15374

ST99SA numeric Surface Area of RightMiddleTemporal Range value: 1635, ..., 5419

ST99TA numeric Cortical Thickness Average of RightMiddleTemporal Range value: 1.557, ..., 3.271

ST99TS numeric Cortical Thickness Standard Deviation of RightMiddleTemporal Range value: 0.439, ..., 1.305

ST9SV numeric Volume (WM Parcellation) of FourthVentricle Range value: 505, ..., 6098

update_stamp POSIXct: YYY-MM-DD

Author

adni-data@googlegroups.com