[DERIVED] Derived Vital Signs

Vital signs derived dataset.

Usage

data(VS)

Format

A data frame with 118720 observations and 20 variables.

Source

For more details see the help vignette: vignette(topic = 'ADNIMERGE2-Derived-Data', package = 'ADNIMERGE2')

Details

STUDYID character Study Identifier ; , with character options: ADNI

DOMAIN character Domain Abbrevation , with character options: VS

USUBJID character Unique Subject Identifier

VSSEQ integer Sequence Number Range value: 1, ..., 147

VSGRPID factor Group ID Study Protocol of Data Collection; , with factor levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4

VSTESTCD character Vital Signs Test Short Name , with character options: DIABP, HEIGHT, PLUSE, RESP, SYSBP, TEMP, WEIGHT

VSTEST character Vital Signs Test Name , with character options: Diastolic Blood Pressure, Height, Pluse Rate, Respiratory Rate, Temperature, Weight

VSORRES character Result or Finding in Original Units

VSORRESU character Original Units , with character options: mmHg, inch, NA, beats/min, breaths/min, F, LB, cm, kg, C

VSSTRESC character Character Result/Finding in Std Format

VSSTRESN numeric Numeric Result/Finding in Standard Units Range value: 2.3, ..., 493

VSSTRESU character Standard Units , with character options: mmHg, cm, beats/min, breaths/min, C, kg

VSLOC character Location of Vital Signs Measurment , with character options: TYMPANIC, ORAL, OTHER, NA

VSBLFL character Baseline Flag , with character options: Y

VSDRVFL character Derived Flag , with character options: Yes

VISITNUM numeric Visit Number Range value: -2, ..., 47

VISIT character Visit Name

EPOCH character Epoch , with character options: Screening, Baseline, Follow-up

VSDTC Date Date/Time of Measurements Date: YYY-MM-DD

VSDY integer Study Day of Vital Signs Range value: -363, ..., 7021

Author

adni-data@googlegroups.com