Derive PACC scores data

This is a wrapper function to generate PACC scores data based on ADNIMERGE2-PACC articles.

Usage

derive_pacc_data(
  data_source = c("ADNIMERGE2", "EXTERNAL"),
  data_list = list(),
  data_source_date = NULL,
  vignette_path = system.file("doc/ADNIMERGE2-PACC.Rmd", package = "ADNIMERGE2"),
  show_pacc_datadict = TRUE,
  envir = new.env(),
  quiet = TRUE,
  ...
)

Arguments

data_source

A character vector of PACC input data source. Either ADNIMERGE2 or EXTERNAL data source.

ADNIMERGE2: To use raw datasets from currently installed ADNIMERGE2 R data package

EXTERNAL: To use an external data source.

data_list

A named list object. By default, it is a null list object. The default value is only applicable if the data source is "ADNIMERGE2". If an external data source (i.e., data_source = "EXTERNAL") is used, then data_list must be a named list object that contains the following dataset.

ADAS: ADAS-Cognitive Behavior Data

MMSE: Mini-Mental State Exam Data

NEUROBAT: Neuropsychological Battery Data

REGISTRY: Registry Data

DXSUM: Diagnostic Summary Data

NOTE: We suggest to make sure all the above datasets were prepared with similar data preparation workflow as in ADNIMERGE2 that includes decode coded variables, convert missing values as 'NA' and add common study protocol identifier variables (i.e., ORIGPROT and COLPROT). Please see ./data-raw/data-prep.R and ./data-raw/data-prep-recode.R for more information regarding ADNIMERGE2 data preparation workflow.

data_source_date

Data source date. It must be a character/date object. By default, it is null if the data source is 'ADNIMERGE2' (i.e., data_source = "ADNIMERGE2") as the data download date from ADNIMERGE2 R data package is used.

vignette_path

File path of PACC score vignettes. By default, ADNIMERGE2-PACC vignettes from ADNIMERGE2 R package is used.

Please see vignette(topic = "ADNIMERGE2-PACC", package = "ADNIMERGE2") for more information.

NOTE: The remote GitHub version ADNIMERGE2-PACC.Rmd vignette script can be also used. Please see examples below.

show_pacc_datadict

A Boolean value to show a data dictionary for PACC scores data. By default, a data dictionary dataset is returned along with the actual PACC scores data.

envir

Inherited from render.

Render environment in which the code chunks are to be evaluated during knitting. However, the code chunks are evaluated in new environment by default.

quiet

Inherited from render. By default, it is TRUE.

...

Additional arguments that pass through render

Value

Either a listed data.frame or single data.frame.

A listed data.frame of the actual PACC scores data (PACC) and corresponding data dictionary data (i.e., 'PACC_DATADICT') if show_pacc_datadict is TRUE.

A single data.frame that only includes PACC scores data if show_pacc_datadict is FALSE.

See also

Examples

if (FALSE) { # \dontrun{
# Generate PACC scores data based on currently installed ADNIMERGE2 data package

pacc_data_list <- derive_pacc_data(
  data_source = "ADNIMERGE2",
  data_list = list(),
  data_source_date = NULL,
  vignette_path = system.file("doc/ADNIMERGE2-PACC.Rmd", package = "ADNIMERGE2"),
  show_pacc_datadict = TRUE,
  envir = new.env()
)
names(pacc_data_list)
pacc_data <- pacc_data_list$PACC
pacc_data_dict <- pacc_data_list$PACC_DATADICT

# To return only PACC data and evaluated all code chunks in global environment
pacc_data1 <- derive_pacc_data(
  data_source = "ADNIMERGE2",
  data_source_date = NULL,
  vignette_path = system.file("doc/ADNIMERGE2-PACC.Rmd", package = "ADNIMERGE2"),
  show_pacc_datadict = FALSE,
  envir = globalenv()
)

# Supposes the following datasets were generated outside ADNIMERGE2 R package
# and the data preparation workflow follows similar procedures as
# ADNIMERGE2 data package build workflow.
EXTERNAL_DATA_LIST <- list(
  ADAS = ADNIMERGE2::ADAS,
  MMSE = ADNIMERGE2::MMSE,
  NEUROBAT = ADNIMERGE2::NEUROBAT,
  REGISTRY = ADNIMERGE2::REGISTRY,
  DXSUM = ADNIMERGE2::DXSUM
)
pacc_data2 <- derive_pacc_data(
  data_source = "EXTERNAL",
  data_list = EXTERNAL_DATA_LIST,
  data_source_date = ADNIMERGE2::DATA_DOWNLOADED_DATE,
  vignette_path = system.file("doc/ADNIMERGE2-PACC.Rmd", package = "ADNIMERGE2"),
  show_pacc_datadict = FALSE
)

# Using a local or remote GitHub version vignettes script to generate PACC scores data
# NOTE: please restart R session as needed.
remotes::install_github(repo = "atri-biostats/ADNIMERGE2")
PACC_VIGNETTE_PATH <- system.file("doc/ADNIMERGE2-PACC.Rmd", package = "ADNIMERGE2")
pacc_data3 <- derive_pacc_data(
  data_source = "EXTERNAL",
  data_list = EXTERNAL_DATA_LIST,
  data_source_date = "2026-09-28",
  vignette_path = PACC_VIGNETTE_PATH,
  show_pacc_datadict = FALSE,
  envir = globalenv()
)
} # }