Prepare PACC score input raw datasets

Prepare PACC score input raw datasets

Usage

prepare_pacc_input_data(
  data_source = "ADNIMERGE2",
  external_data_list = list()
)

Arguments

data_source

Input data source

ADNIMERGE2: if the input raw datasets are obtained from ADNIMERGE2 R data package. This is the default setting.

EXTERNAL: if input raw datasets are obtained from a different data source other than the ADNIMERGE2 R data package. However, this options required to provide three dataset(s) as input. Please see external_data_list argument.

external_data_list

A named list object of data.frame. Only applicable if data_source = 'EXTERNAL' If data_source is 'EXTERNAL', then external_data_list must include ADAS, MMSE and NEUROBAT data.

Value

A named list of three data.frames:

PACC_ADAS_LONG: Question 4 (Q4) task subscore from ADAS-Cognitive Behavior assessment (ADAS)

PACC_MMSE_LONG: Mini-Mental State Exam score (MMSE)

PACC_NEUROBAT_LONG: Digit Symbol Substitution Test score, Logical Memory IIa Delayed Recall score and Trails B score from Neuropsychological Battery assessment (NEUROBAT)

Each datasets will contains the following two columns in addition to ORIGPROT, COLPROT, PTID, RID, VISCODE and `VISDATE.

SCORE_SOURCE: PACC score component name

SCORE: Raw component score

Details

This function is used to prepare PACC score component input dataset either from ADNIMERGE2 R package or external data source.

NOTE: It is still required to install ADNIMERGE2 R data package to perform internal validation regarding data format.

See also

Examples

if (FALSE) { # \dontrun{
# To use datasets from ADNIMERGE2 R data package
pacc_input_data <- prepare_pacc_input_data(
  data_source = "ADNIMERGE2",
  external_data_list = list()
)
names(pacc_input_data)
lapply(pacc_input_data, colnames)

# To use external data source
# Suppose all the three required datasets are obtained from an external source
pacc_input_data1 <- prepare_pacc_input_data(
  data_source = "EXTERNAL",
  external_data_list = list(
    "ADAS" = ADNIMERGE2::ADAS,
    "MMSE" = ADNIMERGE2::MMSE,
    "NEUROBAT" = ADNIMERGE2::NEUROBAT
  )
)
names(pacc_input_data1)
lapply(pacc_input_data1, colnames)
} # }