Adverse Events Log

Adverse Events Log data. More information is available at https://adni.loni.usc.edu/data-samples/adni-data/

Usage

data(ADVERSE)

Format

A data frame with 3405 observations and 89 variables.

Source

https://adni.loni.usc.edu/data-samples/adni-data/

Details

ORIGPROT character Original study protocol , with options: ADNIGO, ADNI3, ADNI1, ADNI2, ADNI4

COLPROT factor Study protocol of data collection , with factor levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4

PTID character Participant ID

RID numeric Participant roster ID Range value: 2, ..., 10436

VISCODE character Visit code

VISCODE2 character Translated visit code

AENUMBER numeric Adverse event number Range value: 1, ..., 34

AECHANGE character Is this event a:

ADNI3: 1=New adverse event;2=Worsening condition or symptom present at screening/initial visit

ADNI4: 1=New adverse event; 2=Worsening condition or symptom present at screening/initial visit

Decoded Value: Yes

AEPENUM numeric Initial Health Assessment Condition Number: Range value: 0, ..., 59

AECOMPOUND character Is this event part of a compound event?

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

AEHONSDT Date Event onset date Date: YYY-MM-DD

AEOUTCOME character Event outcome

ADNI3: 1=Recovered/Resolved;2=Recovering/Resolving;3=Not recovered/Not resolved;4=Recovered/Resolved with sequelae;5=Fatal;6=Unknown

ADNI4: 1=Recovered/Resolved; 2=Recovering/Resolving; 3=Not recovered/Not resolved; 4=Recovered/Resolved with sequelae; 5=Fatal; 6=Unknown

Decoded Value: Yes

AEHCSDT Date Event cease date Date: YYY-MM-DD

AEHDTHDT numeric Date of death Range value: 2017, ..., 2025

AEHDBRN character Was a brain autopsy performed? 0=No;1=Yes Decoded Value: Yes

AEHCHRON character Initial Chronicity

ADNI3: 1=Single occurrence;2=Intermittent - daily;3=Intermittent - weekly;4=Intermittent - monthly or less frequently;5=Persistent

ADNI4: 1=Single occurrence; 2=Intermittent - daily; 3=Intermittent - weekly; 4=Intermittent - monthly or less frequently; 5=Persistent

Decoded Value: Yes

AEHSEVR character Initial Severity

ADNI3: 1=Mild;2=Moderate;3=Severe

ADNI4: 1=Mild; 2=Moderate; 3=Severe

Decoded Value: Yes

AERELAD character Related to symptoms associated with the development of early stage AD or progression of AD

ADNI3: 1=Definitely;2=Probably;3=Possibly;4=Unlikely;5=Not related

ADNI4: 1=Definitely; 2=Probably; 3=Possibly; 4=Unlikely; 5=Not related

Decoded Value: Yes

AERELCM character Related to concomitant therapy

ADNI3: 1=Definitely;2=Probably;3=Possibly;4=Unlikely;5=Not related

ADNI4: 1=Definitely; 2=Probably; 3=Possibly; 4=Unlikely; 5=Not related

Decoded Value: Yes

AERELFLRBTBN character Related to Florbetaben tracer

ADNI3: 1=Definitely;2=Probably;3=Possibly;4=Unlikely;5=Not related

ADNI4: 1=Definitely; 2=Probably; 3=Possibly; 4=Unlikely; 5=Not related

Decoded Value: Yes

AERELFLRBPR character Related to Florbetapir tracer

ADNI3: 1=Definitely;2=Probably;3=Possibly;4=Unlikely;5=Not related

ADNI4: 1=Definitely; 2=Probably; 3=Possibly; 4=Unlikely; 5=Not related

Decoded Value: Yes

AEHIMG character Related to PET/MRI imaging procedure

ADNI3: 1=Definitely;2=Probably;3=Possibly;4=Unlikely;5=Not related

ADNI4: 1=Definitely; 2=Probably; 3=Possibly; 4=Unlikely; 5=Not related

Decoded Value: Yes

AERELTAU character Related to Flotaucipir tracer

ADNI3: 1=Definitely;2=Probably;3=Possibly;4=Unlikely;5=Not related

ADNI4: 1=Definitely; 2=Probably; 3=Possibly; 4=Unlikely; 5=Not related

Decoded Value: Yes

AERELNAV character Related to NAV-4694 tracer 1=Definitely; 2=Probably; 3=Possibly; 4=Unlikely; 5=Not related Decoded Value: Yes

AERELMK character Related to MK-6240 tracer 1=Definitely; 2=Probably; 3=Possibly; 4=Unlikely; 5=Not related Decoded Value: Yes

AERELPI character Related to PI-2620 tracer 1=Definitely; 2=Probably; 3=Possibly; 4=Unlikely; 5=Not related Decoded Value: Yes

AEHLUMB character Related to lumbar puncture

ADNI3: 1=Definitely;2=Probably;3=Possibly;4=Unlikely;5=Not related

ADNI4: 1=Definitely; 2=Probably; 3=Possibly; 4=Unlikely; 5=Not related

Decoded Value: Yes

AERELCOVID character Related to COVID-19 illness

ADNI3: 1=Definitely (Confirmed COVID-19 test);2=Probably;3=Possibly;4=Unlikely;5=Not related

ADNI4: 1=Definitely (Confirmed COVID-19 test); 2=Probably; 3=Possibly; 4=Unlikely; 5=Not related

Decoded Value: Yes

AERELPAN character Related to COVID-19 pandemic disruption

ADNI3: 1=Definitely;2=Probably;3=Possibly;4=Unlikely;5=Not related

ADNI4: 1=Definitely; 2=Probably; 3=Possibly; 4=Unlikely; 5=Not related

Decoded Value: Yes

AERELATESP character Related to other study procedure(s)

ADNI3: 1=Definitely;2=Probably;3=Possibly;4=Unlikely;5=Not related

ADNI4: 1=Definitely; 2=Probably; 3=Possibly; 4=Unlikely; 5=Not related

Decoded Value: Yes

AEHCMEDS character Was treatment (medication) for AE required?

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

AESERIOUS character Was event serious at any time during the trial?

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

AESERDATE Date Date event became serious: Date: YYY-MM-DD

AEDTSITE Date Date site became aware that the event is serious: Date: YYY-MM-DD

SAELIFE character Is this an SAE because event was life-threatening?

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

SAEHOSPIT character Is this an SAE because it lead to hospitalization?

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

SAEPROLONG character Is this an SAE because it lead to prolongation of hospitalization?

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

SAEDEATH character Is this an SAE because it lead to death?

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

SAECONGEN character Is this an SAE because it lead to a congenital anomaly or birth defect?

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

SAEDISAB character Is this an SAE because it caused persistent or significant disability/incapacity?

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

SAEOTHER character Is this an SAE because it lead to another medically important condition?

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

AECHANGE1 character Is there a change to report?

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

AEDATE1 Date Change 1. Date Date: YYY-MM-DD

AESEV1 character Change 1: Severity

ADNI3: 1=Mild;2=Moderate;3=Severe

ADNI4: 1=Mild; 2=Moderate; 3=Severe

Decoded Value: Yes

AECHRON1 character Change 1: Chronicity

ADNI3: 1=Single occurrence;2=Intermittent - daily;3=Intermittent - weekly;4=Intermittent - monthly or less frequently;5=Persistent

ADNI4: 1=Single occurrence; 2=Intermittent - daily; 3=Intermittent - weekly; 4=Intermittent - monthly or less frequently; 5=Persistent

Decoded Value: Yes

AECHANGE2 character Is there a 2nd change to report?

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

Decoded Value: Yes

AEDATE2 Date Change 2. Date Date: YYY-MM-DD

AESEV2 character Change 2: Severity

ADNI3: 1=Mild;2=Moderate;3=Severe

ADNI4: 1=Mild; 2=Moderate; 3=Severe

Decoded Value: Yes

AECHRON2 character Change 2: Chronicity

ADNI3: 1=Single occurrence;2=Intermittent - daily;3=Intermittent - weekly;4=Intermittent - monthly or less frequently;5=Persistent

ADNI4: 1=Single occurrence; 2=Intermittent - daily; 3=Intermittent - weekly; 4=Intermittent - monthly or less frequently; 5=Persistent

Decoded Value: Yes

AECHANGE3 numeric Is there a 3rd change to report?

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

AEDATE3 logical Change 3. Date Boolean value: TRUE/ FALSE

AESEV3 logical Change 3: Severity

ADNI3: 1=Mild;2=Moderate;3=Severe

ADNI4: 1=Mild; 2=Moderate; 3=Severe

AECHRON3 logical Change 3: Chronicity

ADNI3: 1=Single occurrence;2=Intermittent - daily;3=Intermittent - weekly;4=Intermittent - monthly or less frequently;5=Persistent

ADNI4: 1=Single occurrence; 2=Intermittent - daily; 3=Intermittent - weekly; 4=Intermittent - monthly or less frequently; 5=Persistent

AECHANGE4 logical Is there a 4th change to report?

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

AEDATE4 logical Change 4. Date Boolean value: TRUE/ FALSE

AESEV4 logical Change 4: Severity

ADNI3: 1=Mild;2=Moderate;3=Severe

ADNI4: 1=Mild; 2=Moderate; 3=Severe

AECHRON4 logical Change 4: Chronicity

ADNI3: 1=Single occurrence;2=Intermittent - daily;3=Intermittent - weekly;4=Intermittent - monthly or less frequently;5=Persistent

ADNI4: 1=Single occurrence; 2=Intermittent - daily; 3=Intermittent - weekly; 4=Intermittent - monthly or less frequently; 5=Persistent

AECHANGE5 logical Is there a 5th change to report?

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

AEDATE5 logical Change 5. Date Boolean value: TRUE/ FALSE

AESEV5 logical Change 5: Severity

ADNI3: 1=Mild;2=Moderate;3=Severe

ADNI4: 1=Mild; 2=Moderate; 3=Severe

AECHRON5 logical Change 5: Chronicity

ADNI3: 1=Single occurrence;2=Intermittent - daily;3=Intermittent - weekly;4=Intermittent - monthly or less frequently;5=Persistent

ADNI4: 1=Single occurrence; 2=Intermittent - daily; 3=Intermittent - weekly; 4=Intermittent - monthly or less frequently; 5=Persistent

AECHANGE6 logical Is there a 6th change to report?

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

AEDATE6 logical Change 6. Date Boolean value: TRUE/ FALSE

AESEV6 logical Change 6: Severity

ADNI3: 1=Mild;2=Moderate;3=Severe

ADNI4: 1=Mild; 2=Moderate; 3=Severe

AECHRON6 logical Change 6: Chronicity

ADNI3: 1=Single occurrence;2=Intermittent - daily;3=Intermittent - weekly;4=Intermittent - monthly or less frequently;5=Persistent

ADNI4: 1=Single occurrence; 2=Intermittent - daily; 3=Intermittent - weekly; 4=Intermittent - monthly or less frequently; 5=Persistent

AECHANGE7 logical Is there a 7th change to report?

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

AEDATE7 logical Change 7. Date Boolean value: TRUE/ FALSE

AESEV7 logical Change 7: Severity

ADNI3: 1=Mild;2=Moderate;3=Severe

ADNI4: 1=Mild; 2=Moderate; 3=Severe

AECHRON7 logical Change 7: Chronicity

ADNI3: 1=Single occurrence;2=Intermittent - daily;3=Intermittent - weekly;4=Intermittent - monthly or less frequently;5=Persistent

ADNI4: 1=Single occurrence; 2=Intermittent - daily; 3=Intermittent - weekly; 4=Intermittent - monthly or less frequently; 5=Persistent

AECHANGE8 logical Is there a 8th change to report?

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

AEDATE8 logical Change 8. Date Boolean value: TRUE/ FALSE

AESEV8 logical Change 8: Severity

ADNI3: 1=Mild;2=Moderate;3=Severe

ADNI4: 1=Mild; 2=Moderate; 3=Severe

AECHRON8 logical Change 8: Chronicity

ADNI3: 1=Single occurrence;2=Intermittent - daily;3=Intermittent - weekly;4=Intermittent - monthly or less frequently;5=Persistent

ADNI4: 1=Single occurrence; 2=Intermittent - daily; 3=Intermittent - weekly; 4=Intermittent - monthly or less frequently; 5=Persistent

AECHANGE9 logical Is there a 9th change to report?

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

AEDATE9 logical Change 9. Date Boolean value: TRUE/ FALSE

AESEV9 logical Change 9: Severity

ADNI3: 1=Mild;2=Moderate;3=Severe

ADNI4: 1=Mild; 2=Moderate; 3=Severe

AECHRON9 logical Change 9: Chronicity

ADNI3: 1=Single occurrence;2=Intermittent - daily;3=Intermittent - weekly;4=Intermittent - monthly or less frequently;5=Persistent

ADNI4: 1=Single occurrence; 2=Intermittent - daily; 3=Intermittent - weekly; 4=Intermittent - monthly or less frequently; 5=Persistent

AECHANGE10 logical Is there a 10th change to report?

ADNI3: 0=No;1=Yes

ADNI4: 0=No; 1=Yes

AEDATE10 logical Change 10. Date Boolean value: TRUE/ FALSE

AESEV10 logical Change 10: Severity

ADNI3: 1=Mild;2=Moderate;3=Severe

ADNI4: 1=Mild; 2=Moderate; 3=Severe

AECHRON10 logical Change 10: Chronicity

ADNI3: 1=Single occurrence;2=Intermittent - daily;3=Intermittent - weekly;4=Intermittent - monthly or less frequently;5=Persistent

ADNI4: 1=Single occurrence; 2=Intermittent - daily; 3=Intermittent - weekly; 4=Intermittent - monthly or less frequently; 5=Persistent

ID numeric Record ID Range value: 563, ..., 185565

SITEID numeric Site ID Range value: 1, ..., 941

USERDATE Date User Create Date Date: YYY-MM-DD

USERDATE2 Date User Last Modify Date Date: YYY-MM-DD

DD_CRF_VERSION_LABEL character CRF Version Label , with options: v1

LANGUAGE_CODE character Language Code e=English Decoded Value: Yes

HAS_QC_ERROR character Has Quality Check Error 0=Does not have QC error or QC error has been approved; 1=Has QC error Decoded Value: Yes

update_stamp POSIXct POSIXct: YYY-MM-DD

Author

adni-data@googlegroups.com