ASHS volume data

ASHS volume data data. More information is available at https://adni.loni.usc.edu/data-samples/adni-data/

Usage

data(PICSLASHS)

Format

A data frame with 3916 observations and 68 variables.

Source

https://adni.loni.usc.edu/data-samples/adni-data/

Details

ORIGPROT character Original study protocol , with options: ADNIGO, ADNI1, ADNI2, ADNI3, ADNI4

COLPROT factor Study protocol of data collection , with levels: ADNI1, ADNIGO, ADNI2, ADNI3, ADNI4

PTID character Site and Unique participant ID

RID numeric roster participant ID Range value: 21, ..., 10796

VISCODE character visit code

VISCODE2 character visit code translated

EXAMDATE Date of MRI scan Date: YYY-MM-DD

PROCESSDATE Date image was procesed for data in this row Date: YYY-MM-DD

IMAGEUID_T1 numeric Unique ID of T1 dicom Range value: 239888, ..., 11309435

IMAGEUID_T2 numeric Unique ID of T2 dicom Range value: 239883, ..., 11306907

LEFT_CA1_VOL numeric left hemisphere ROI CA1 volume Range value: 43.03, ..., 5584.101

LEFT_CA1_NS numeric Number of coronal slices in which left hemisphere ROI CA1 appears Range value: 2, ..., 46

LEFT_CA2_VOL numeric left hemisphere ROI CA2 volume Range value: 0.234, ..., 9.46

LEFT_CA2_NS numeric Number of coronal slices in which left hemisphere ROI CA2 appears Range value: 1, ..., 10

LEFT_CA3_VOL numeric left hemisphere ROI CA3 volume Range value: 0.467, ..., 195.923

LEFT_CA3_NS numeric Number of coronal slices in which left hemisphere ROI CA3 appears Range value: 1, ..., 11

LEFT_DG_VOL numeric left hemisphere ROI dentate gyrus volume Range value: 8.85, ..., 2731.324

LEFT_DG_NS numeric Number of coronal slices in which left hemisphere ROI dentate gyrus appears Range value: 1, ..., 38

LEFT_MISC_VOL numeric left hemisphere ROI MISC volume Range value: 7.629, ..., 793.426

LEFT_MISC_NS numeric Number of coronal slices in which left hemisphere ROI MISC appears Range value: 3, ..., 35

LEFT_SUB_VOL numeric left hemisphere ROI Subiculum volume Range value: 24.585, ..., 1701.356

LEFT_SUB_NS numeric Number of coronal slices in which left hemisphere ROI Subiculum appears Range value: 1, ..., 43

LEFT_ERC_VOL numeric left hemisphere ROI Entorhinal cortex volume Range value: 0.916, ..., 785.523

LEFT_ERC_NS numeric Number of coronal slices in which left hemisphere ROI Entorhinal cortex appears Range value: 1, ..., 19

LEFT_BA35_VOL numeric left hemisphere ROI BA35 volume Range value: 38.452, ..., 995.853

LEFT_BA35_NS numeric Number of coronal slices in which left hemisphere ROI BA35 appears Range value: 2, ..., 25

LEFT_BA36_VOL numeric left hemisphere ROI BA36 volume Range value: 538.94, ..., 4696.956

LEFT_BA36_NS numeric Number of coronal slices in which left hemisphere ROI BA36 appears Range value: 5, ..., 27

LEFT_PHC_VOL numeric left hemisphere ROI parahippocampal cortex volume Range value: 0.305, ..., 5310.664

LEFT_PHC_NS numeric Number of coronal slices in which left hemisphere ROI parahippocampal cortex appears Range value: 1, ..., 18

LEFT_SULCUS_VOL numeric left hemisphere ROI sulcus volume Range value: 13.41, ..., 1256.594

LEFT_SULCUS_NS numeric Number of coronal slices in which left hemisphere ROI sulcus appears Range value: 1, ..., 41

LEFT_CA_VOL numeric left hemisphere ROI CA (CA1+CA2+CA3) volume Range value: 43.03, ..., 5629.267

LEFT_CA_NS numeric Number of coronal slices in which left hemisphere ROI CA (CA1+CA2+CA3) appears Range value: 2, ..., 46

LEFT_HIPP_VOL numeric left hemisphere ROI CA1+CA2+CA3+DG volume Range value: 51.88, ..., 8125.907

LEFT_HIPP_NS numeric Number of coronal slices in which left hemisphere ROI CA1+CA2+CA3+DG appears Range value: 2, ..., 46

RIGHT_CA1_VOL numeric right hemisphere ROI CA1 volume Range value: 188.599, ..., 6692.515

RIGHT_CA1_NS numeric Number of coronal slices in which right hemisphere ROI CA1 appears Range value: 5, ..., 50

RIGHT_CA2_VOL numeric right hemisphere ROI CA2 volume Range value: 0.305, ..., 19.531

RIGHT_CA2_NS numeric Number of coronal slices in which right hemisphere ROI CA2 appears Range value: 1, ..., 11

RIGHT_CA3_VOL numeric right hemisphere ROI CA3 volume Range value: 0.234, ..., 171.509

RIGHT_CA3_NS numeric Number of coronal slices in which right hemisphere ROI CA3 appears Range value: 1, ..., 10

RIGHT_DG_VOL numeric right hemisphere ROI dentate gyrus volume Range value: 67.293, ..., 5831.613

RIGHT_DG_NS numeric Number of coronal slices in which right hemisphere ROI dentate gyrus appears Range value: 3, ..., 41

RIGHT_MISC_VOL numeric right hemisphere ROI MISC volume Range value: 2.136, ..., 552.064

RIGHT_MISC_NS numeric Number of coronal slices in which right hemisphere ROI MISC appears Range value: 2, ..., 44

RIGHT_SUB_VOL numeric right hemisphere ROI Subiculum volume Range value: 9.155, ..., 2147.22

RIGHT_SUB_NS numeric Number of coronal slices in which right hemisphere ROI Subiculum appears Range value: 2, ..., 46

RIGHT_ERC_VOL numeric right hemisphere ROI Entorhinal cortex volume Range value: 1.221, ..., 2073.673

RIGHT_ERC_NS numeric Number of coronal slices in which right hemisphere ROI Entorhinal cortex appears Range value: 1, ..., 24

RIGHT_BA35_VOL numeric right hemisphere ROI BA35 volume Range value: 2.747, ..., 4026.19

RIGHT_BA35_NS numeric Number of coronal slices in which right hemisphere ROI BA35 appears Range value: 2, ..., 28

RIGHT_BA36_VOL numeric right hemisphere ROI BA36 volume Range value: 4.883, ..., 8778.09

RIGHT_BA36_NS numeric Number of coronal slices in which right hemisphere ROI BA36 appears Range value: 1, ..., 29

RIGHT_PHC_VOL numeric right hemisphere ROI parahippocampal cortex volume Range value: 0.61, ..., 5197.762

RIGHT_PHC_NS numeric Number of coronal slices in which right hemisphere ROI parahippocampal cortex appears Range value: 1, ..., 18

RIGHT_SULCUS_VOL numeric right hemisphere ROI sulcus volume Range value: 3.505, ..., 2192.081

RIGHT_SULCUS_NS numeric Number of coronal slices in which right hemisphere ROI sulcus appears Range value: 2, ..., 44

RIGHT_CA_VOL numeric right hemisphere ROI CA (CA1+CA2+CA3) volume Range value: 188.599, ..., 6752.635

RIGHT_CA_NS numeric Number of coronal slices in which right hemisphere ROI CA (CA1+CA2+CA3) appears Range value: 5, ..., 50

RIGHT_HIPP_VOL numeric right hemisphere ROI CA1+CA2+CA3+DG volume Range value: 276.794, ..., 12584.248

RIGHT_HIPP_NS numeric Number of coronal slices in which right hemisphere ROI CA1+CA2+CA3+DG appears Range value: 5, ..., 50

ICV numeric Intercranial volume derived from T1 image Range value: 541618.623, ..., 1944066.216

SLICE_THICKNESS numeric Coronal slice thickness Range value: 0.983051, ..., 14

T2_QC_R character manual QC score for right hemisphere image quality -1=no image ; 0=Unusable (substructures and the SRLM are not identifiable) or image has motion and looks blurry/wiggly ; 1=Bad Quality (SRLM is hard to identify and substructures are blurry) or image has motion ; 2=Adequate Quality (some substructures are clear, SRLM is mostly visible) ; 3=intermediate quality between 2 & 4 ; 4=Good Quality (most substructures are clear, SRLM is continuously visible) Decoded Value: Yes

T2_QC_L character manual QC score for left hemisphere image quality -1=no image ; 0=Unusable (substructures and the SRLM are not identifiable) or image has motion and looks blurry/wiggly ; 1=Bad Quality (SRLM is hard to identify and substructures are blurry) or image has motion ; 2=Adequate Quality (some substructures are clear, SRLM is mostly visible) ; 3=intermediate quality between 2 & 4 ; 4=Good Quality (most substructures are clear, SRLM is continuously visible) Decoded Value: Yes

T2_QC_COMMENT character QC comment If value is "QC from MAYOADIRL_MRI_IMAGEQC", QC rating mapped from MAYOADIRL scale. MAYOADIRL scale is 1: Excellent, 2: Good, 3: Fair, 4: Unusable, -1: Not evaluated. MAYOADIRL scale mapped to PICSLASHS scale is -1:-1, 4:0, 3:2, 2:3, 1:4. Any comments from sheet MAYOADIRL_MRI_IMAGEQC included in comment.

update_stamp POSIXct: YYY-MM-DD